% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADAE}
\alias{ADAE}
\title{[DERIVED] Analysis Dataset of Adverse Events}
\format{
A data frame with 4001 observations and 62 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Analysis-Data', package = 'ADNIMERGE2')}
}
\usage{
data(ADAE)
}
\description{
Analysis dataset of adverse events derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier , with character options: ADNI
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{SUBJID} \emph{character} Subject Identifier for the Study  
\item \strong{SITEID} \emph{numeric} Study Site Identifier Range value: 1, ..., 941
\item \strong{TRTA} \emph{character} Actual Arm  
\item \strong{TRTP} \emph{character} Planned Arm  
\item \strong{AGE} \emph{numeric} Age (in Years) Range value: 50.5, ..., 90.3
\item \strong{AGEU} \emph{character} Age Units , with character options: Years
\item \strong{AGEGR1} \emph{character} Pooled Age Group 1 , with character options: >=65, <65
\item \strong{AGEGR1N} \emph{numeric} Pooled Age Group 1 (N) Range value: 1, ..., 2
\item \strong{SEX} \emph{character} Sex , with character options: Male, Female
\item \strong{RACE} \emph{factor} Race , with factor levels: American Indian or Alaskan Native, Asian, Black or African American, Native Hawaiian or Other Pacific Islander, Other Pacific Islander, White, More than one race, Unknown
\item \strong{RACEN} \emph{numeric} Race (N) Range value: 1, ..., 999
\item \strong{ETHNIC} \emph{character} Ethnicity , with character options: Not Hispanic or Latino, Unknown, Hispanic or Latino
\item \strong{DX} \emph{factor} Baseline Diagnostics Status , with factor levels: CN, MCI, DEM
\item \strong{ENRLFL} \emph{character} Enrolled Population Flag , with character options: Y
\item \strong{ENRLDT} \emph{Date} Date of Enrollment Date: YYY-MM-DD
\item \strong{EOSDT} \emph{Date} End of Study Date Date: YYY-MM-DD
\item \strong{EOSDTF} \emph{character} End of Study Date Imputation Flag , with character options: M
\item \strong{ASTDT} \emph{Date} Analysis Start Date Date: YYY-MM-DD
\item \strong{ASTDTF} \emph{character} Analysis Start Date Imputation Flag  
\item \strong{ASTDY} \emph{numeric} Analysis Start Relative Day Range value: -4327, ..., 7176
\item \strong{AENDT} \emph{Date} Analysis End Date Date: YYY-MM-DD
\item \strong{AENDTF} \emph{character} Analysis End Date Imputation Flag  
\item \strong{AENDY} \emph{numeric} Analysis End Relative Day Range value: -146, ..., 337328
\item \strong{ADURN} \emph{numeric} Adverse Events Duration (N) Range value: -363, ..., 332062
\item \strong{ADURU} \emph{character} Analysis Duration Unit , with character options: days
\item \strong{AETERM} \emph{character} Reported Term for the Adverse Event  
\item \strong{AELLT} \emph{character} Lowest Level Term  
\item \strong{AELLTCD} \emph{character} Lowest Level Term Code  
\item \strong{AEDECOD} \emph{character} Dictionary-Derived Term  
\item \strong{AEPTCD} \emph{character} Preferred Term Code  
\item \strong{AEHLT} \emph{character} High Level Term  
\item \strong{AEHLTCD} \emph{character} High Level Term Code  
\item \strong{AEHLGT} \emph{character} High Level Group Term  
\item \strong{AEHLGTCD} \emph{character} High Level Group Term Code  
\item \strong{AESOC} \emph{character} Primary Sytem Organ Class  
\item \strong{AESOCCD} \emph{character} Primary Sytem Organ Class Code  
\item \strong{AESEV} \emph{character} Severity/Intensity , with character options: Severe, Mild, Moderate
\item \strong{AESER} \emph{character} Serious Event , with character options: Yes, No
\item \strong{AESCONG} \emph{character} Congenital Anomaly or Birth Defect , with character options: No
\item \strong{AESDISAB} \emph{character} Persist or Signif Disability/Incapacity , with character options: No, Yes
\item \strong{AESDTH} \emph{character} Results in Death , with character options: Yes, No
\item \strong{AESHOSP} \emph{character} Requires or Prologns Hospitalization , with character options: No, Yes
\item \strong{AESLIFE} \emph{character} Is Life Threating , with character options: Yes, No
\item \strong{AESMIE} \emph{character} Other Medically Important Serious Event , with character options: No, Yes
\item \strong{AERELCM} \emph{character} Related to Concomitant Therapy , with character options: Not related, Possibly, Unlikely, Definitely, Probably
\item \strong{AERELFLRBTBN} \emph{character} Related to Florbetaben Tracer , with character options: Not related, Unlikely, Definitely, Possibly, Probably
\item \strong{AERELFLRBPR} \emph{character} Related to Florbetapir Tracer , with character options: Not related, Unlikely, Possibly, Probably
\item \strong{AEHIMG} \emph{character} Related to PET/MRI Imaging Procedure , with character options: Not related, Possibly, Unlikely, Definitely, Probably
\item \strong{AERELTAU} \emph{character} Related to Flotaucipir Tracer , with character options: Not related, Probably, Unlikely, Possibly
\item \strong{AERELNAV} \emph{character} Related to NAV-4694 Tracer , with character options: Not related, Possibly, Unlikely
\item \strong{AERELMK} \emph{character} Related to MK-6240 Tracer , with character options: Not related, Unlikely
\item \strong{AERELPI} \emph{character} Related to PI-2620 Tracer , with character options: Not related, Unlikely
\item \strong{AEHLUMB} \emph{character} Related to Lumbar Puncture , with character options: Not related, Definitely, Probably, Unlikely, Possibly
\item \strong{AERELCOVID} \emph{character} Related to COVID-19 Illness , with character options: Not related, Unlikely, Definitely (Confirmed COVID-19 test), Possibly, Probably
\item \strong{AERELPAN} \emph{character} Related to COVID-19 Pandemic Disruption , with character options: Not related, Definitely, Probably, Unlikely, Possibly
\item \strong{AERELAD} \emph{character} Related to Early Stage/Progression of AD , with character options: Not related, Unlikely, Possibly, Probably, Definitely
\item \strong{AERELATESP} \emph{character} Related to Other Study Procedure(s) , with character options: Not related, Unlikely, Probably, Definitely, Possibly
\item \strong{AECONTRT} \emph{character} Concomitant or Additional Trtmnt Given , with character options: No, Yes
\item \strong{AEOUT} \emph{character} Outcome of Adverse Event , with character options: Fatal, Not recovered/Not resolved, Recovered/Resolved, Recovering/Resolving, Recovered/Resolved with sequelae, Unknown
\item \strong{AESEQ} \emph{integer} Sequence Number Range value: 1, ..., 32
}
}
\keyword{derived_dataset}
