% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADAS}
\alias{ADAS}
\title{ADAS-Cognitive Behavior}
\format{
A data frame with 12868 observations and 17 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ADAS)
}
\description{
ADAS-Cognitive Behavior data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment EXAMDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{TOTSCORE} \emph{numeric} Total Score (ADAS 11) 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2} 
\item \emph{ADNI3:} 0..70
}
\item \strong{TOTAL13} \emph{numeric} Total Score (ADAS 13) 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2} 
\item \emph{ADNI3:} 0..85
}
\item \strong{ID} \emph{numeric} Record ID Range value: 4, ..., 16500223
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: list1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English ; s=Spanish
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
