% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADI}
\alias{ADI}
\title{Area Deprivation Index Score}
\format{
A data frame with 1515 observations and 18 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ADI)
}
\description{
Area Deprivation Index Score data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNI2, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{ADIDATE} \emph{ } Date of collection: Date: YYY-MM-DD
\item \strong{ADISTATE} \emph{numeric} State Decile Score: Range value: 0, ..., 10
\item \strong{ADINATIONAL} \emph{character} National Percentile:  
\item \strong{ADIREV} \emph{character} Has the participant moved since their last visit? 0=No ; 1=Yes ; 2=Not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 49, ..., 128244
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1, v2
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{subject_characteristics}
