% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{adjust_scbl_record}
\alias{adjust_scbl_record}
\title{Carrying Forward Screening Record as Baseline Record}
\usage{
adjust_scbl_record(
  .data,
  wide_format = TRUE,
  extra_id_cols = NULL,
  adjust_date_col = NULL,
  check_col = NULL
)
}
\arguments{
\item{.data}{Data.frame}

\item{wide_format}{A Boolean indicator of wide data format, Default: TRUE}

\item{extra_id_cols}{ID variable names in addition to \code{RID}, \code{COLPROT} and \code{VISCODE}.
Only applicable for a long format data.}

\item{adjust_date_col}{Adjusting date column values.
Only applicable if \code{adjust_date_col} is non-missing}

\item{check_col}{Column name which values used to determine date column adjustment.
Only applicable if \code{adjust_date_col} is non-missing}
}
\value{
A data.frame that contains adjusted baseline record and screening record.
}
\description{
This function is used to carry forward screening record as baseline record per
subject id (\code{RID}) and study data collection phase (\code{COLPROT}).
}
\examples{
\dontrun{
pacc_mmse_long_file <- system.file(
  "/extradata/pacc-raw-input/pacc_mmse_long.csv",
  package = "ADNIMERGE2"
)
pacc_mmse_long <- readr::read_csv(
  file = pacc_mmse_long_file,
  guess_max = Inf
)
# For a wide format data
bl_sc_mmse_record <- adjust_scbl_record(
  .data = pacc_mmse_long \%>\%
    pivot_wider(names_from = SCORE_SOURCE, values_from = SCORE),
  wide_format = TRUE
)
head(bl_sc_mmse_record)

# For a long format data without adjusting for visit date
bl_sc_mmse_record_long <- adjust_scbl_record(
  .data = pacc_mmse_long,
  wide_format = FALSE,
  extra_id_cols = "SCORE_SOURCE"
)
head(bl_sc_mmse_record_long)

# For a long format data with adjusting for visit date
bl_sc_mmse_record_long <- adjust_scbl_record(
  .data = pacc_mmse_long,
  wide_format = FALSE,
  extra_id_cols = "SCORE_SOURCE",
  adjust_date_col = "VISDATE",
  check_col = "SCORE"
)
}
}
\keyword{adni_enroll_fun}
