% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{adjust_scbl_visitdate}
\alias{adjust_scbl_visitdate}
\title{Adjust Baseline Visit/Assessment Collection Date}
\usage{
adjust_scbl_visitdate(
  .data,
  .adj_scbl_data,
  date_col,
  check_col,
  extra_id_cols = NULL
)
}
\arguments{
\item{.data}{Data.frame}

\item{.adj_scbl_data}{A data.frame generated using \code{\link{adjust_scbl_record}()} function}

\item{date_col}{Date column name}

\item{check_col}{Column name which values used to determine date column adjustment.
Only applicable if \code{adjust_date_col} is non-missing}

\item{extra_id_cols}{ID variable names in addition to \code{RID}, \code{COLPROT} and \code{VISCODE}.
Only applicable for a long format data.}
}
\value{
A data.frame the same as \code{.adj_scbl_data} input data with adjusted values.
}
\description{
This function is used to adjust the visit/assessment collection date
 of a baseline record that was carried forward from a screening visit.
}
\examples{
\dontrun{
pacc_neurobat_long_file <- system.file(
  "/extradata/pacc-raw-input/pacc_neurobat_long.csv",
  package = "ADNIMERGE2"
)
pacc_neurobat_long <- readr::read_csv(
  file = pacc_neurobat_long_file,
  guess_max = Inf
)
bl_sc_neurobat_record_long <- adjust_scbl_record(
  .data = pacc_neurobat_long,
  wide_format = FALSE,
  extra_id_cols = "SCORE_SOURCE",
  adjust_date_col = NULL
)
# Only for baseline visit records
adjust_scbl_visitdate(
  .data = pacc_neurobat_long,
  .adj_scbl_data = bl_sc_neurobat_record_long,
  date_col = "VISDATE",
  check_col = "SCORE",
  id_cols = "SCORE_SOURCE"
)
}
}
\seealso{
\code{\link{adjust_scbl_record}()}
}
\keyword{internal}
