% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADNI2_ECG}
\alias{ADNI2_ECG}
\title{ECG}
\format{
A data frame with 14 observations and 23 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ADNI2_ECG)
}
\description{
ECG data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI1, ADNIGO
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 254
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 55, ..., 4889
\item \strong{SITEID} \emph{numeric} Site ID Range value: 13, ..., 58
\item \strong{VISCODE} \emph{character} Visit code , with options: v31, v41, tau
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: m36, m108, m60, m48
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{logical} Date record last updated A Boolean value: TRUE/ FALSE
\item \strong{DONE} \emph{numeric} Was assessment/procedure done? 1=Yes ; 0=No
\item \strong{NDREASON} \emph{numeric} If No, reason not done: 1=Participant Unable for Cognitive Reasons ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 4=Equipment related issue ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\item \strong{ECOMPDT} \emph{logical} Date ECG Completed A Boolean value: TRUE/ FALSE
\item \strong{ECOMPTM} \emph{logical} Time ECG Completed A Boolean value: TRUE/ FALSE
\item \strong{ECGPR} \emph{logical} Was there a PR interval? 1=Yes ; 0=No
\item \strong{ECGPRI} \emph{logical} PR Interval 20..500
\item \strong{ECGQRS} \emph{logical} QRS Duration 20..300
\item \strong{ECGQTC} \emph{logical} QT Interval 100..600
\item \strong{ECGHRT} \emph{logical} Heart Rate 30..180
\item \strong{ECGNORM} \emph{logical} ECG Status: 1=Normal ; 0=Abnormal
\item \strong{ECGABNORM} \emph{numeric} If abnormal, please describe: Range value: -1, ..., -1
\item \strong{ECGREV} \emph{logical} Was the result reviewed prior to the AV-1451 dose? 1=Yes ; 0=No
\item \strong{ECGLIMIT} \emph{logical} Was the result within limits specified in the addendum? 1=Yes ; 0=No
\item \strong{ECGCOMM} \emph{numeric} Comments: Range value: -1, ..., -1
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{medical_history}
