% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADNI2_OTELGTAU}
\alias{ADNI2_OTELGTAU}
\title{Tau AV-1451 PET Eligibility}
\format{
A data frame with 148 observations and 22 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ADNI2_OTELGTAU)
}
\description{
Tau AV-1451 PET Eligibility data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI1, ADNIGO
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 308
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 55, ..., 5277
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 60
\item \strong{VISCODE} \emph{character} Visit code , with options: v31, v21, v41, v51, tau
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: m36, m108, m24, m96, m48, m60, m120, m72
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 2
\item \strong{TAUELG} \emph{numeric} Please confirm if participant meets eligibility criteria: 1=Participant is eligible ; 0=Participant should be excluded
\item \strong{EXCLUSION} \emph{numeric} If participant should be excluded, select reason: 1=Exclusion Criteria
\item \strong{NVROT} \emph{numeric}   1=Other (specify)
\item \strong{NVRSPECIFY} \emph{character} If Other, specify: , with options: Participant did not consent to scan, Exluded from Past EKG results
\item \strong{EXC1} \emph{logical} If Exclusion Criteria, which criteria does the participant meet: 1=[1] Has any condition that, in the investigator's opinion, could increase risk to the participant, limit the participant's ability to tolerate the experimental procedures, or interfere with collection/analysis of the data (for example, participants with severe chronic back pain might not be able to lie still during the scanning procedures).
\item \strong{EXC2} \emph{logical}   1=[2] Has abnormal findings on physical examination, or laboratory screening tests that suggest the patient might have a condition that could, in the opinion of the investigator, affect his or her response to the radiopharmaceutical and related testing procedures.
\item \strong{EXC3} \emph{logical}   1=[3] Is deemed likely to be unable to perform all of the addendum imaging procedures for any reason.
\item \strong{EXC4} \emph{logical}   1=[4] Female that is pregnant, lactating, or of childbearing potential
\item \strong{EXC5} \emph{numeric}   1=[5] Has a history of risk factors for Torsades de Pointes or is taking medications known to prolong QT interval (a list of prohibited and discouraged medications is provided by the Sponsor).
\item \strong{EXC6} \emph{numeric}   1=[6] Has ECG obtained prior to the initial 18F-AV-1451 scan that in the opinion of the investigator is clinically significant with regard to the subject's participation in the study. Bazett's corrected QT (QTcB) interval must be evaluated and must not exceed 458 msec in males or 474 msec in females.
\item \strong{EXC7} \emph{logical}   1=[7] Has hypersensitivity to 18F-AV-1451 or any of its excipients.
\item \strong{TAUCOMM} \emph{character} Comments: , with options: QT interval 502, Pt originally scheduled 1/26/16 but avid had transportation delays that would'vd had scan ending @ 1830 and pt didn't want end that late.  Soonest could reschedule is 2/23/16., Abnormal EKG, but no evidence of conduct delay. Communicated with ADNI medical director and Bill Jasvist who felt subject okay if electrically stable, which he is., Patient taking cyclobenzaprine, which can cause prolongation of QT interval, Participant eligible to continue, Since consent obtained and record/EKG's initially reviewed, subject has been started on flecainide, so is no longer eligible for tau scan.
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{imaging}
