% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADNI2_VISITID}
\alias{ADNI2_VISITID}
\title{ADNI2 Visit Code Mapping List}
\format{
A data frame with 24147 observations and 6 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ADNI2_VISITID)
}
\description{
ADNI2 Visit Code Mapping List . More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 5296
\item \strong{PTID} \emph{character}    
\item \strong{VISCODE} \emph{character}    
\item \strong{VISCODE2} \emph{character}    
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{enrollment}
