% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adni_phase}
\alias{adni_phase}
\title{List ADNI Study Protocol/Phases}
\usage{
adni_phase()
}
\value{
A character vector with list of ADNI study phases.
}
\description{
This function is used to generate all the ADNI study phases.
}
\examples{
\dontrun{
adni_study_phase <- adni_phase()
adni_study_phase
}
}
\seealso{
Other ADNI study protocol/phase: 
\code{\link{adni_phase_order_num}()},
\code{\link{adni_pttype}()},
\code{\link{adni_study_track}()},
\code{\link{convert_adni_phase_order_num}()},
\code{\link{create_col_protocol}()},
\code{\link{create_orig_protocol}()},
\code{\link{original_study_protocol}()}
}
\concept{ADNI study protocol/phase}
\keyword{adni_procotol_fun}
