% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adni_phase_order_num}
\alias{adni_phase_order_num}
\title{ADNI Study Protocol Order}
\usage{
adni_phase_order_num(phase)
}
\arguments{
\item{phase}{ADNI study protocol phases: \code{ADNI1}, \code{ADNIGO}, \code{ADNI2},
\code{ADNI3} or \code{ADNI4}}
}
\value{
A numeric vector with the same size as \code{phase} with study protocol order number.
}
\description{
This function is used to extract the ADNI study protocol order number.
}
\examples{
\dontrun{
adni_phase_order_num(phase = c("ADNI1", "ADNI3", "ADNI4", "ADNIGO"))
}
}
\seealso{
Other ADNI study protocol/phase: 
\code{\link{adni_phase}()},
\code{\link{adni_pttype}()},
\code{\link{adni_study_track}()},
\code{\link{convert_adni_phase_order_num}()},
\code{\link{create_col_protocol}()},
\code{\link{create_orig_protocol}()},
\code{\link{original_study_protocol}()}
}
\concept{ADNI study protocol/phase}
\keyword{adni_procotol_fun}
