% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adni_pttype}
\alias{adni_pttype}
\title{ADNI PTYPE}
\usage{
adni_pttype()
}
\value{
A character vector of ADNI PTYPE.
}
\description{
ADNI PTYPE
}
\examples{
\dontrun{
adni_pttype()
}
}
\seealso{
Other ADNI study protocol/phase: 
\code{\link{adni_phase}()},
\code{\link{adni_phase_order_num}()},
\code{\link{adni_study_track}()},
\code{\link{convert_adni_phase_order_num}()},
\code{\link{create_col_protocol}()},
\code{\link{create_orig_protocol}()},
\code{\link{original_study_protocol}()}
}
\concept{ADNI study protocol/phase}
\keyword{adni_procotol_fun}
