% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adni_study_track}
\alias{adni_study_track}
\title{Get ADNI Study Track}
\usage{
adni_study_track(cur_study_phase, orig_study_phase)
}
\arguments{
\item{cur_study_phase}{Current ADNI study phase}

\item{orig_study_phase}{Original study protocol that subject enrolled newly for the
first time in the ADNI study}
}
\value{
A character vector with the same size as \code{cur_study_phase} with study track.
}
\description{
This function is used to identify study track (i.e. subject type) based
 on the provided ADNI phase.
}
\seealso{
Other ADNI study protocol/phase: 
\code{\link{adni_phase}()},
\code{\link{adni_phase_order_num}()},
\code{\link{adni_pttype}()},
\code{\link{convert_adni_phase_order_num}()},
\code{\link{create_col_protocol}()},
\code{\link{create_orig_protocol}()},
\code{\link{original_study_protocol}()}
}
\concept{ADNI study protocol/phase}
\keyword{adni_procotol_fun}
