% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADNIMERGE2-package.R
\docType{package}
\name{ADNIMERGE2-package}
\alias{ADNIMERGE2}
\alias{ADNIMERGE2-package}
\title{ADNIMERGE2: ADNI Study Data Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This is an R data package for the Alzheimer's Disease Neuroimaging Initiative (ADNI) study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://atri-biostats.github.io/ADNIMERGE2}
  \item Report bugs at \url{https://github.com/atri-biostats/ADNIMERGE2/issues}
}

}
\author{
\strong{Maintainer}: ATRI Biostatistics \email{biostat_request@atrihub.io}

}
\keyword{internal}
