% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADQS}
\alias{ADQS}
\title{[DERIVED] Analysis Dataset of Questionnaires}
\format{
A data frame with 320021 observations and 36 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Analysis-Data', package = 'ADNIMERGE2')}
}
\usage{
data(ADQS)
}
\description{
Analysis dataset of questionnaires derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier , with character options: ADNI
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{SUBJID} \emph{character} Subject Identifier for the Study  
\item \strong{SITEID} \emph{numeric} Study Site Identifier Range value: 1, ..., 941
\item \strong{AGE} \emph{numeric} Age (in Years) Range value: 50.5, ..., 91.4
\item \strong{AGEU} \emph{character} Age Units , with character options: Years
\item \strong{AGEGR1} \emph{character} Pooled Age Group 1 , with character options: >=65, <65
\item \strong{AGEGR1N} \emph{numeric} Pooled Age Group 1 (N) Range value: 1, ..., 2
\item \strong{SEX} \emph{character} Sex , with character options: Male, Female
\item \strong{RACE} \emph{factor} Race , with factor levels: American Indian or Alaskan Native, Asian, Black or African American, Native Hawaiian or Other Pacific Islander, Other Pacific Islander, White, More than one race, Unknown
\item \strong{BMI} \emph{numeric} Body Mass Index Range value: 18.64, ..., 47.87
\item \strong{BMIBLU} \emph{character} Body Mass Index Unit , with character options: kg/m^2
\item \strong{DX} \emph{factor} Baseline Diagnostics Status , with factor levels: CN, MCI, DEM
\item \strong{AMYSTAT} \emph{character} Baseline Amyloid Status , with character options: Elevated, Non Elevated
\item \strong{ENRLFL} \emph{character} Enrolled Population Flag , with character options: Y
\item \strong{ENRLDT} \emph{Date} Date of Enrollment Date: YYY-MM-DD
\item \strong{EOSDT} \emph{Date} End of Study Date Date: YYY-MM-DD
\item \strong{EOSDTF} \emph{character} End of Study Date Imputation Flag , with character options: M
\item \strong{TRTA} \emph{character} Actual Arm  
\item \strong{TRTP} \emph{character} Planned Arm  
\item \strong{AVISIT} \emph{character} Analysis Visit  
\item \strong{AVISITN} \emph{numeric} Analysis Visit (N) Range value: 0, ..., 35075
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -2, ..., 49
\item \strong{ADY} \emph{numeric} Analysis Relative Day Range value: -365625, ..., 35075
\item \strong{ADT} \emph{Date} Analysis Date Date: YYY-MM-DD
\item \strong{PARAM} \emph{character} Parameter  
\item \strong{PARAMCD} \emph{character} Parameter Code  
\item \strong{PARAMN} \emph{numeric} Parameter (N) Range value: 1, ..., 23
\item \strong{AVAL} \emph{numeric} Analysis Value Range value: -500, ..., 996
\item \strong{AVALC} \emph{character} Analysis Value (C)  
\item \strong{BASE} \emph{numeric} Baseline Value Range value: -100, ..., 300
\item \strong{CHG} \emph{numeric} Change from Baseline Range value: -525, ..., 754
\item \strong{PCHG} \emph{numeric} Percent Change from Baseline Range value: -37931.6333654082, ..., 19900.483702932
\item \strong{ABLFL} \emph{character} Baseline Record Flag , with character options: Y
\item \strong{QSSEQ} \emph{integer} Sequence Number Range value: 1, ..., 414
}
}
\keyword{derived_dataset}
