% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADRS}
\alias{ADRS}
\title{[DERIVED] Response Analysis Dataset}
\format{
A data frame with 16140 observations and 34 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Analysis-Data', package = 'ADNIMERGE2')}
}
\usage{
data(ADRS)
}
\description{
Response analysis dataset derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier , with character options: ADNI
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{SUBJID} \emph{character} Subject Identifier for the Study  
\item \strong{COLPROT} \emph{factor} Study Protocol of Data Collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PARAMCD} \emph{character} Parameter Code , with character options: DX, DEATH
\item \strong{PARAM} \emph{character} Parameter , with character options: Clinical Diagnostics Status, Death
\item \strong{PARAMN} \emph{numeric} Parameter (N) Range value: 1, ..., 2
\item \strong{AVISIT} \emph{character} Analysis Visit  
\item \strong{AVISITN} \emph{numeric} Analysis Visit (N) Range value: -1849, ..., 7208
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -2, ..., 999
\item \strong{ADY} \emph{numeric} Analysis Relative Day Range value: -1849, ..., 7208
\item \strong{ADT} \emph{Date} Analysis Date Date: YYY-MM-DD
\item \strong{AVAL} \emph{numeric} Analysis Value Range value: 1, ..., 4
\item \strong{AVALC} \emph{character} Analysis Value (C) , with character options: DEM, MCI, CN, DEATH
\item \strong{BASE} \emph{character} Baseline Value , with character options: DEM, MCI, CN
\item \strong{ABLFL} \emph{character} Baseline Record Flag , with character options: Y
\item \strong{ENRLFL} \emph{character} Enrolled Population Flag , with character options: Y
\item \strong{ENRLDT} \emph{Date} Date of Enrollment Date: YYY-MM-DD
\item \strong{EOSDT} \emph{Date} End of Study Date Date: YYY-MM-DD
\item \strong{EOSDTF} \emph{character} End of Study Date Imputation Flag , with character options: M
\item \strong{BLBFL} \emph{character} Baseline Records Population Flag , with character options: Y
\item \strong{FOLLOWPFL} \emph{character} Follow-up Records Population Flag , with character options: Y, M
\item \strong{ANL01FL} \emph{character} Analysis Flag , with character options: Y
\item \strong{AGE} \emph{numeric} Age (in Years) Range value: 50.5, ..., 91.4
\item \strong{AGEU} \emph{character} Age Units , with character options: Years
\item \strong{AGEGR1} \emph{character} Pooled Age Group 1 , with character options: >=65, <65
\item \strong{AGEGR1N} \emph{numeric} Pooled Age Group 1 (N) Range value: 1, ..., 2
\item \strong{SEX} \emph{character} Sex , with character options: Male, Female
\item \strong{RACE} \emph{factor} Race , with factor levels: American Indian or Alaskan Native, Asian, Black or African American, Native Hawaiian or Other Pacific Islander, Other Pacific Islander, White, More than one race, Unknown
\item \strong{BMI} \emph{numeric} Body Mass Index Range value: 18.64, ..., 47.87
\item \strong{BMIBLU} \emph{character} Body Mass Index Unit , with character options: kg/m^2
\item \strong{DX} \emph{factor} Baseline Diagnostics Status , with factor levels: CN, MCI, DEM
\item \strong{AMYSTAT} \emph{character} Baseline Amyloid Status , with character options: Elevated, Non Elevated
}
}
\keyword{derived_dataset}
