% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADSL}
\alias{ADSL}
\title{[DERIVED] Subject-Level Analysis Dataset}
\format{
A data frame with 5146 observations and 55 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Analysis-Data', package = 'ADNIMERGE2')}
}
\usage{
data(ADSL)
}
\description{
Subject-level analysis dataset derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier , with character options: ADNI
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{SUBJID} \emph{character} Subject Identifier for the Study  
\item \strong{SITEID} \emph{numeric} Study Site Identifier Range value: 1, ..., 941
\item \strong{ORIGPROT} \emph{factor} Original Study Protocol , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{AGE} \emph{numeric} Age (in Years) Range value: 50.5, ..., 91.4
\item \strong{AGEU} \emph{character} Age Units , with character options: Years
\item \strong{AGEGR1N} \emph{numeric} Pooled Age Group 1 (N) Range value: 1, ..., 2
\item \strong{AGEGR1} \emph{character} Pooled Age Group 1 , with character options: >=65, <65
\item \strong{SEX} \emph{character} Sex , with character options: Male, Female
\item \strong{RACE} \emph{factor} Race , with factor levels: American Indian or Alaskan Native, Asian, Black or African American, Native Hawaiian or Other Pacific Islander, Other Pacific Islander, White, More than one race, Unknown
\item \strong{RACEN} \emph{numeric} Race (N) Range value: 1, ..., 999
\item \strong{ETHNIC} \emph{character} Ethnicity , with character options: Not Hispanic or Latino, Unknown, Hispanic or Latino
\item \strong{ENRLFL} \emph{character} Enrolled Population Flag , with character options: Y
\item \strong{DTHFL} \emph{character} Death Flag , with character options: Yes
\item \strong{ARM} \emph{character} Description of Planned Arm  
\item \strong{ACTARM} \emph{character} Description of Actual Arm  
\item \strong{ENRLDT} \emph{Date} Date of Enrollment Date: YYY-MM-DD
\item \strong{EOSDT} \emph{Date} End of Study Date Date: YYY-MM-DD
\item \strong{EOSDTF} \emph{character} End of Study Date Imputation Flag , with character options: M
\item \strong{DTHDT} \emph{Date} Date of Death Date: YYY-MM-DD
\item \strong{DTHDTF} \emph{character} Date of Death Imputation Flag , with character options: M
\item \strong{EDUC} \emph{numeric} Education Range value: 4, ..., 20
\item \strong{MARISTAT} \emph{character} Marital Status , with character options: Married, Widowed, Divorced, Never married, Domestic Partnership, Unknown
\item \strong{BMI} \emph{numeric} Body Mass Index Range value: 18.64, ..., 47.87
\item \strong{BMIBLU} \emph{character} Body Mass Index Unit , with character options: kg/m^2
\item \strong{DX} \emph{factor} Baseline Diagnostics Status , with factor levels: CN, MCI, DEM
\item \strong{AMYSTAT} \emph{character} Baseline Amyloid Status , with character options: Elevated, Non Elevated
\item \strong{AMYSTATC} \emph{character} Baseline Amyloid Status - Composite Ref , with character options: Non Elevated, Elevated
\item \strong{TRACER} \emph{character} Image Tracer Type , with character options: FBP, FBB
\item \strong{IMAGERES} \emph{character} Image Resolution , with character options: 6MM
\item \strong{ADASTT11} \emph{numeric} Baseline ADAS-Cog Item 11 Total Score Range value: 0, ..., 42.67
\item \strong{ADASTT13} \emph{numeric} Baseline ADAS-Cog Item 13 Total Score Range value: 0, ..., 54.67
\item \strong{CDGLOBAL} \emph{numeric} Baseline CDR Global Score Range value: 0, ..., 2
\item \strong{CDRSB} \emph{numeric} Baseline CDR Sum of Boxes Score Range value: 0, ..., 15
\item \strong{ECOGPTTT} \emph{numeric} Baseline Everyday Cognition - Pt Score Range value: 1, ..., 3.85714285714286
\item \strong{ECOGSPTT} \emph{numeric} Baseline Everyday Cognition - Sp Score Range value: 1, ..., 3.92307692307692
\item \strong{FAQTOTAL} \emph{numeric} Baseline FAQ Total Score Range value: 0, ..., 30
\item \strong{FCISCORE} \emph{numeric} Baseline FCI Score Range value: 4, ..., 74
\item \strong{GDTOTAL} \emph{numeric} Baseline GDS Score Range value: 0, ..., 7
\item \strong{MMSCORE} \emph{numeric} Baseline MMSE Score Range value: 12, ..., 30
\item \strong{MOCA} \emph{numeric} Baseline MOCA Score Range value: 4, ..., 30
\item \strong{NPITOTAL} \emph{numeric} Baseline NPI Total Score Range value: 0, ..., 81
\item \strong{NPIQTOTL} \emph{numeric} Baseline NPIQ Total Score Range value: 0, ..., 24
\item \strong{DIGITSCR} \emph{numeric} Baseline DIGIT Score Range value: 0, ..., 80
\item \strong{LDELTOTL} \emph{numeric} Baseline LDEL Total Score Range value: 0, ..., 23
\item \strong{LIMMTOTL} \emph{numeric} Baseline LIMM Total Score Range value: 0, ..., 23
\item \strong{RAVLTFG} \emph{numeric} Baseline RAVLT Forgetting Score Range value: -5, ..., 15
\item \strong{RAVLTFGP} \emph{numeric} Baseline RAVLT Percent Forgetting Score Range value: -100, ..., 100
\item \strong{RAVLTIMM} \emph{numeric} Baseline RVAL Immediate Score Range value: 0, ..., 72
\item \strong{RAVLTLRN} \emph{numeric} Baseline RVAL Learning Score Range value: -4, ..., 13
\item \strong{TRABSCOR} \emph{numeric} Baseline Trails B Score Range value: 0, ..., 300
\item \strong{APOE} \emph{character} APOE Genotype , with character options: ε4/ε4, ε3/ε3, ε3/ε4, ε2/ε3, ε2/ε2, ε2/ε4
\item \strong{MPACCDIGIT} \emph{numeric} Baseline mPACCdigit Range value: -19.523912, ..., 6.327653
\item \strong{MPACCTRAILSB} \emph{numeric} Baseline mPACCtrialsB Range value: -25.342335, ..., 7.673357
}
}
\keyword{derived_dataset}
