% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADSXLIST}
\alias{ADSXLIST}
\title{Diagnosis and Symptoms Checklist}
\format{
A data frame with 4884 observations and 40 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ADSXLIST)
}
\description{
Diagnosis and Symptoms Checklist data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 2407
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{AXNAUSEA} \emph{character} 1. Nausea 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXVOMIT} \emph{character} 2. Vomiting 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXDIARRH} \emph{character} 3. Diarrhea 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXCONSTP} \emph{character} 4. Constipation 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXABDOMN} \emph{character} 5. Abdominal discomfort 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXSWEATN} \emph{character} 6. Sweating 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXDIZZY} \emph{character} 7. Dizziness 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXENERGY} \emph{character} 8. Low energy 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXDROWSY} \emph{character} 9. Drowsiness 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXVISION} \emph{character} 10. Blurred vision 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXHDACHE} \emph{character} 11. Headache 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXDRYMTH} \emph{character} 12. Dry mouth 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXBREATH} \emph{character} 13. Shortness of breath 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXCOUGH} \emph{character} 14. Coughing 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXPALPIT} \emph{character} 15. Palpitations 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXCHEST} \emph{character} 16. Chest pain 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXURNDIS} \emph{character} 17. Urinary discomfort (e.g., burning) 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXURNFRQ} \emph{character} 18. Urinary frequency 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXANKLE} \emph{character} 19. Ankle swelling 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXMUSCLE} \emph{character} 20. Muscloskeletal pain 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXRASH} \emph{character} 21. Rash 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXINSOMN} \emph{character} 22. Insomnia 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXDPMOOD} \emph{character} 23. Depressed mood 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXCRYING} \emph{character} 24. Crying 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXELMOOD} \emph{character} 25. Elevated mood 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXWANDER} \emph{character} 26. Wandering 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXFALL} \emph{character} 27. Fall 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{AXOTHER} \emph{character} 28. Other/Diagnosis 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 8410
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adnigo}
\keyword{assessments}
