% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADVERSE}
\alias{ADVERSE}
\title{Adverse Events Log}
\format{
A data frame with 4001 observations and 89 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ADVERSE)
}
\description{
Adverse Events Log data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNIGO, ADNI3, ADNI1, ADNI2, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10887
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{AENUMBER} \emph{numeric} Adverse event number Range value: 1, ..., 34
\item \strong{AECHANGE} \emph{character} Is this event a: 1=New adverse event ; 2=Worsening condition or symptom present at screening/initial visit
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEPENUM} \emph{numeric} Initial Health Assessment Condition Number: Range value: 0, ..., 59
\item \strong{AECOMPOUND} \emph{character} Is this event part of a compound event? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHONSDT} \emph{ } Event onset date Date: YYY-MM-DD
\item \strong{AEOUTCOME} \emph{character} Event outcome 1=Recovered/Resolved ; 2=Recovering/Resolving ; 3=Not recovered/Not resolved ; 4=Recovered/Resolved with sequelae ; 5=Fatal ; 6=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHCSDT} \emph{ } Event cease date Date: YYY-MM-DD
\item \strong{AEHDTHDT} \emph{numeric} Date of death Range value: 2017, ..., 2025
\item \strong{AEHDBRN} \emph{character} Was a brain autopsy performed? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHCHRON} \emph{character} Initial Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHSEVR} \emph{character} Initial Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELAD} \emph{character} Related to symptoms associated with the development of early stage AD or progression of AD 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELCM} \emph{character} Related to concomitant therapy 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELFLRBTBN} \emph{character} Related to Florbetaben tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELFLRBPR} \emph{character} Related to Florbetapir tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHIMG} \emph{character} Related to PET/MRI imaging procedure 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELTAU} \emph{character} Related to Flotaucipir tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELNAV} \emph{character} Related to NAV-4694 tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELMK} \emph{character} Related to MK-6240 tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELPI} \emph{character} Related to PI-2620 tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHLUMB} \emph{character} Related to lumbar puncture 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELCOVID} \emph{character} Related to COVID-19 illness 1=Definitely (Confirmed COVID-19 test) ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELPAN} \emph{character} Related to COVID-19 pandemic disruption 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AERELATESP} \emph{character} Related to other study procedure(s) 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHCMEDS} \emph{character} Was treatment (medication) for AE required? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AESERIOUS} \emph{character} Was event serious at any time during the trial? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AESERDATE} \emph{ } Date event became serious: Date: YYY-MM-DD
\item \strong{AEDTSITE} \emph{ } Date site became aware that the event is serious: Date: YYY-MM-DD
\item \strong{SAELIFE} \emph{character} Is this an SAE because event was life-threatening? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{SAEHOSPIT} \emph{character} Is this an SAE because it lead to  hospitalization? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{SAEPROLONG} \emph{character} Is this an SAE because it lead to  prolongation of hospitalization? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{SAEDEATH} \emph{character} Is this an SAE because it lead to  death? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{SAECONGEN} \emph{character} Is this an SAE because it lead to a congenital anomaly or birth defect? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{SAEDISAB} \emph{character} Is this an SAE because it caused  persistent or significant disability/incapacity? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{SAEOTHER} \emph{character} Is this an SAE because it lead to another medically important condition? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE1} \emph{character} Is there a change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE1} \emph{ } Change 1.  Date Date: YYY-MM-DD
\item \strong{AESEV1} \emph{character} Change 1: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON1} \emph{character} Change 1: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE2} \emph{character} Is there a 2nd change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE2} \emph{ } Change 2.  Date Date: YYY-MM-DD
\item \strong{AESEV2} \emph{character} Change 2: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON2} \emph{character} Change 2: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE3} \emph{character} Is there a 3rd change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE3} \emph{ } Change 3.  Date Date: YYY-MM-DD
\item \strong{AESEV3} \emph{character} Change 3: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON3} \emph{character} Change 3: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE4} \emph{character} Is there a 4th change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE4} \emph{ } Change 4.  Date Date: YYY-MM-DD
\item \strong{AESEV4} \emph{character} Change 4: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON4} \emph{character} Change 4: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE5} \emph{character} Is there a 5th change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE5} \emph{logical} Change 5.  Date A Boolean value: TRUE/ FALSE
\item \strong{AESEV5} \emph{character} Change 5: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON5} \emph{character} Change 5: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE6} \emph{character} Is there a 6th change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE6} \emph{logical} Change 6.  Date A Boolean value: TRUE/ FALSE
\item \strong{AESEV6} \emph{character} Change 6: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON6} \emph{character} Change 6: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE7} \emph{character} Is there a 7th change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE7} \emph{logical} Change 7.  Date A Boolean value: TRUE/ FALSE
\item \strong{AESEV7} \emph{character} Change 7: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON7} \emph{character} Change 7: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE8} \emph{character} Is there a 8th change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE8} \emph{logical} Change 8.  Date A Boolean value: TRUE/ FALSE
\item \strong{AESEV8} \emph{character} Change 8: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON8} \emph{character} Change 8: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE9} \emph{character} Is there a 9th change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE9} \emph{logical} Change 9.  Date A Boolean value: TRUE/ FALSE
\item \strong{AESEV9} \emph{character} Change 9: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON9} \emph{character} Change 9: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHANGE10} \emph{character} Is there a 10th change to report? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEDATE10} \emph{logical} Change 10.  Date A Boolean value: TRUE/ FALSE
\item \strong{AESEV10} \emph{character} Change 10: Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AECHRON10} \emph{character} Change 10: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 563, ..., 186890
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{adni4}
\keyword{medical_history}
