% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AE}
\alias{AE}
\title{[DERIVED]  Derived  Adverse Events}
\format{
A data frame with 4001 observations and 44 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(AE)
}
\description{
Adverse events derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbreviation , with character options: AE
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{AEGRPID} \emph{character} Group ID , with character options: ADNI3, ADNI4
\item \strong{AESEQ} \emph{integer} Sequence Number Range value: 1, ..., 32
\item \strong{AETERM} \emph{character} Reported Term for the Adverse Event  
\item \strong{AELLT} \emph{character} Lowest Level Term  
\item \strong{AELLTCD} \emph{character} Lowest Level Term Code  
\item \strong{AEDECOD} \emph{character} Dictionary-Derived Term  
\item \strong{AEPTCD} \emph{character} Preferred Term Code  
\item \strong{AEHLT} \emph{character} High Level Term  
\item \strong{AEHLTCD} \emph{character} High Level Term Code  
\item \strong{AEHLGT} \emph{character} High Level Group Term  
\item \strong{AEHLGTCD} \emph{character} High Level Group Term Code  
\item \strong{AESOC} \emph{character} Primary Sytem Organ Class  
\item \strong{AESOCCD} \emph{character} Primary Sytem Organ Class Code  
\item \strong{AESEV} \emph{character} Severity/Intensity , with character options: Mild, Moderate, Severe
\item \strong{AESER} \emph{character} Serious Event , with character options: No, Yes
\item \strong{AEOUT} \emph{character} Outcome of Adverse Event , with character options: Not recovered/Not resolved, Recovered/Resolved, Recovering/Resolving, Fatal, Recovered/Resolved with sequelae, Unknown
\item \strong{AESCONG} \emph{character} Congential Anomaly or Birth Defect , with character options: No
\item \strong{AESDISAB} \emph{character} Persist or Signif Disability/Incapacity , with character options: No, Yes
\item \strong{AESDTH} \emph{character} Results in Death , with character options: No, Yes
\item \strong{AESHOSP} \emph{character} Requires or Prologns Hospitalization , with character options: Yes, No
\item \strong{AESLIFE} \emph{character} Is Life Threating , with character options: Yes, No
\item \strong{AESMIE} \emph{character} Other Medically Important Serious Event , with character options: No, Yes
\item \strong{AECONTRT} \emph{character} Concomitant or Additional Trtmnt Given , with character options: No, Yes
\item \strong{AERELAD} \emph{character} Related to Early Stage or Progression of AD , with character options: Not related, Unlikely, Probably, Possibly, Definitely
\item \strong{AERELCM} \emph{character} Related to Concomitant Therapy , with character options: Not related, Possibly, Unlikely, Probably, Definitely
\item \strong{AERELFLRBTBN} \emph{character} Related to Florbetaben Tracer , with character options: Not related, Unlikely, Possibly, Definitely, Probably
\item \strong{AERELFLRBPR} \emph{character} Related to Florbetapir Tracer , with character options: Not related, Unlikely, Possibly, Probably
\item \strong{AEHIMG} \emph{character} Related to PET/MRI Imaging Procedure , with character options: Not related, Unlikely, Possibly, Probably, Definitely
\item \strong{AERELTAU} \emph{character} Related to Flotaucipir Tracer  , with character options: Not related, Possibly, Probably, Unlikely
\item \strong{AERELNAV} \emph{character} Related to NAV-4694 Tracer , with character options: Not related, Unlikely, Possibly
\item \strong{AERELMK} \emph{character} Related to MK-6240 Tracer , with character options: Not related, Unlikely
\item \strong{AERELPI} \emph{character} Related to PI-2620 Tracer , with character options: Not related, Unlikely
\item \strong{AEHLUMB} \emph{character} Related to Lumbar Puncture , with character options: Not related, Definitely, Probably, Possibly, Unlikely
\item \strong{AERELCOVID} \emph{character} Related to COVID-19 Illness , with character options: Not related, Definitely (Confirmed COVID-19 test), Unlikely, Probably, Possibly
\item \strong{AERELPAN} \emph{character} Related to COVID-19 Pandemic Disruption , with character options: Not related, Unlikely, Definitely, Possibly, Probably
\item \strong{AERELATESP} \emph{character} Related to Other Study Procedure(s) , with character options: Not related, Unlikely, Probably, Possibly, Definitely
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -1, ..., 49
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{EPOCH} \emph{character} Epoch , with character options: Follow-up, Baseline, Screening
\item \strong{AESTDTC} \emph{character} Start Date/Time of Adverse Event  
\item \strong{AEENDTC} \emph{character} End Date/Time of Adverse Event  
}
}
\keyword{derived_dataset}
