% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ALIQUOT_COUNT}
\alias{ALIQUOT_COUNT}
\title{Aliquot Count Table}
\format{
A data frame with 28506 observations and 15 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ALIQUOT_COUNT)
}
\description{
Aliquot Count Table data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10788
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{Primary} \emph{character} Sample Primary Type BLD=Blood ; CSF=Cerebrospinal fluid ; URN=Urine
\item \strong{Additive} \emph{character} Sample Additive Type EDT=Ethylenediaminetetraacetic acid ; NON=None
\item \strong{Derivative} \emph{character} Sample Derivative Type PLA=Plasma ; SER=Serum ; CSF=Cerebrospinal fluid ; URN=Urine
\item \strong{Comment} \emph{character} Additional Comments , with options: Received, not aliquoted
\item \strong{Total} \emph{numeric} Number of aliquots immediately after being aliquoted Range value: 1, ..., 64
\item \strong{Remain} \emph{numeric} Number of aliquots at the time this dataset is created Range value: 1, ..., 50
\item \strong{Volume} \emph{numeric} Vol of aliquots at the time this dataset is created. Range value: 0.5, ..., 25
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{biospecimen}
