% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMAS}
\alias{AMAS}
\title{Abbreviated Multidimensional Acculturation Scale}
\format{
A data frame with 1061 observations and 59 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AMAS)
}
\description{
Abbreviated Multidimensional Acculturation Scale data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNI2, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{DONE} \emph{character} Was assessment/procedure done? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 1=Participant Unable for Cognitive Reasons ; 10=Not called for by protocol ; 11=Participant in a non-USA Site ; 12=Participant is not Latinx ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS1} \emph{character} 1. I think of myself as being U.S. American. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS2} \emph{character} 2. I feel good about being U.S. American. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS3} \emph{character} 3. Being U.S. American plays an important part in my life. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS4} \emph{character} 4. I feel that I am part of the U.S. American culture. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS5} \emph{character} 5. I have a strong sense of being U.S. American. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS6} \emph{character} 6. I am proud of being U.S. American. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS7} \emph{character} 7. I think of myself as being___(a member of my culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS8} \emph{character} 8. I feel good about being____(a member of my culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS9} \emph{character} 9. Being___(a member of my culture of origin) plays an important part in my life. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS10} \emph{character} 10. I feel that I am part of___culture (culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS11} \emph{character} 11. I have a strong sense of being___(culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS12} \emph{character} 12. I am proud of being___(culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS13} \emph{character} 13. at school or work 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS14} \emph{character} 14. with American friends 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS15} \emph{character} 15. on the phone 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS16} \emph{character} 16. with strangers 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS17} \emph{character} 17. in general 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS18} \emph{character} 18. on television or in movies 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS19} \emph{character} 19. in newspapers and magazines 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS20} \emph{character} 20. words in songs 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS21} \emph{character} 21. in general 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS22} \emph{character} 22. with family 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS23} \emph{character} 23. with friends from the same country as you 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS24} \emph{character} 24. on the phone 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS25} \emph{character} 25. with strangers 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS26} \emph{character} 26. in general 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS27} \emph{character} 27. on television or in movies 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS28} \emph{character} 28. in newspapers and magazines 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS29} \emph{character} 29. words in songs 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS30} \emph{character} 30. in general 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS31} \emph{character} 31. American national heroes 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS32} \emph{character} 32. popular American television 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS33} \emph{character} 33. popular American newspapers 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS34} \emph{character} 34. popular American actors and actresses 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS35} \emph{character} 35. American history 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS36} \emph{character} 36. American political leaders 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS37} \emph{character} 37. national heroes from your native culture 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS38} \emph{character} 38. popular television shows in your native language 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS39} \emph{character} 39. popular newspapers and magazines in your native language 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS40} \emph{character} 40. popular actors and actresses from your native culture 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS41} \emph{character} 41. history of your native culture 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMAS42} \emph{character} 42. political leaders from your native culture 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 71, ..., 126855
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{subject_characteristics}
