% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMNART}
\alias{AMNART}
\title{American National Adult Reading Test}
\format{
A data frame with 1087 observations and 68 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AMNART)
}
\description{
American National Adult Reading Test data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNI2, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{DONE} \emph{character} Was assessment/procedure done? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 1=Participant Unable for Cognitive Reasons ; 10=Not called for by protocol ; 11=Participant in a non-USA Site ; 12=Not an English-speaking participant ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{SOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call/Video Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART1} \emph{character} 1. ache 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART2} \emph{character} 2. debt 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART3} \emph{character} 3. pint 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART4} \emph{character} 4. depot 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART5} \emph{character} 5. chord 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART6} \emph{character} 6. bouquet 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART7} \emph{character} 7. deny 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART8} \emph{character} 8. capon 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART9} \emph{character} 9. heir 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART10} \emph{character} 10. aisle 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART11} \emph{character} 11. subtle 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART12} \emph{character} 12. nausea 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART13} \emph{character} 13. gauge 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART14} \emph{character} 14. naive 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART15} \emph{character} 15. thyme 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART16} \emph{character} 16. courteous 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART17} \emph{character} 17. algae 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART18} \emph{character} 18. fetal 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART19} \emph{character} 19. quadruped 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART20} \emph{character} 20. epitome 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART21} \emph{character} 21. superfluous 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART22} \emph{character} 22. chamois 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART23} \emph{character} 23. papyrus 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART24} \emph{character} 24. asthma 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART25} \emph{character} 25. hiatus 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART26} \emph{character} 26. simile 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART27} \emph{character} 27. blatant 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART28} \emph{character} 28. cellist 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART29} \emph{character} 29. zealot 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART30} \emph{character} 30. abstemious 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART31} \emph{character} 31. meringue 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART32} \emph{character} 32. placebo 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART33} \emph{character} 33. facade 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART34} \emph{character} 34. pugilist 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART35} \emph{character} 35. virulent 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART36} \emph{character} 36. worsted 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART37} \emph{character} 37. detente 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART38} \emph{character} 38. anise 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART39} \emph{character} 39. sieve 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART40} \emph{character} 40. chassis 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART41} \emph{character} 41. beatify 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART42} \emph{character} 42. scion 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART43} \emph{character} 43. cabal 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART44} \emph{character} 44. apropos 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART45} \emph{character} 45. caprice 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART46} \emph{character} 46. demesne 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART47} \emph{character} 47. imbroglio 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART48} \emph{character} 48. hyperbole 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART49} \emph{character} 49. syncope 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMNART50} \emph{character} 50. prelate 1=Correct ; 2=Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 95, ..., 128242
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
