% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMYDISC}
\alias{AMYDISC}
\title{Amyloid Disclosure Visit}
\format{
A data frame with 651 observations and 34 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AMYDISC)
}
\description{
Amyloid Disclosure Visit data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNIGO, ADNI1, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10878
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment EXAMDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{DONE} \emph{character} Was an amyloid PET disclosure event conducted? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{DISC} \emph{character} Was amyloid PET result disclosed to the participant? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{DISCND} \emph{character} If no, please explain why: 1=Participant not psychologically well ; 2=Participant's study partner not available (Study partners are required to attend for MCI and DEM participants) ; 3=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{OUTCOME} \emph{character} What was the amyloid PET scan result? 0=Non-Elevated ; 1=Elevated
\emph{Decoded Value: }\strong{Yes}
\item \strong{SOURCE} \emph{character} In what setting was the disclosure conducted? 1=Remotely ; 2=In-Person
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTAGE} \emph{numeric} How long did the disclosure visit last? Range value: 3, ..., 90
\item \strong{JOINDISC} \emph{character} Did someone join the participant at the visit? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{JOINNUM} \emph{numeric} If yes, how many others joined the participant at the visit? Range value: 1, ..., 3
\item \strong{JOINSPCY} \emph{character} If yes, was their relationship to the participant? 1=Partner/spouse ; 2=Child/Child-In-Law ; 3=Sibling ; 4=Other relative ; 5=Friend ; 6=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTGRASP} \emph{character} Please rate this participant's understanding of their amyloid PET scan result. 1=Inadequate understanding ; 2=Marginal understanding ; 3=Adequate understanding
\emph{Decoded Value: }\strong{Yes}
\item \strong{SPGRASP} \emph{character} If there was an ADNI study partner present during this visit, please rate their understanding of the amyloid result. 1=Inadequate understanding ; 2=Marginal understanding ; 3=Adequate understanding ; 4=Not applicable – no study partner was present
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTREACT} \emph{character} Which of the following terms best describes this participant's overall emotional reaction to this visit? 1=Overall positive ; 2=Overall neutral ; 3=Overall mixed (both positive and negative) ; 4=Overall negative
\emph{Decoded Value: }\strong{Yes}
\item \strong{SPREACT} \emph{character} If there was an ADNI study partner present, which of the following terms best describes this study partner's overall emotional reaction to this visit? 1=Overall positive ; 2=Overall neutral ; 3=Overall mixed (both positive and negative) ; 4=Overall negative ; 5=Not applicable – no study partner was present
\emph{Decoded Value: }\strong{Yes}
\item \strong{TOPICS} \emph{character} Which of the following topics did you discuss with this participant? 1=Other risk factors (e.g., age, sex, family history) ; 10=Other ; 2=Meaning of the result ; 3=PET scan image (i.e., showed participant their image) ; 4=Mean cortical SUVR ; 5=PET read CRF ; 6=Emotional impact of learning this information ; 7=Potential stigma when sharing this result with others ; 8=Potential discrimination (e.g., employment, insurance, housing) ; 9=Novel therapies for AD
\emph{Decoded Value: }\strong{Yes}
\item \strong{CHAL} \emph{character} Did you experience any challenges during this visit, including in your discussion of any of the above? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTREAD} \emph{character} How confident did you feel assessing this participant's readiness to receive their result? 1=Not at all confident ; 2=A little confident ; 3=Somewhat confident ; 4=Confident ; 5=Very confident
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTDISCL} \emph{character} How confident did you feel disclosing to this participant? 1=Not at all confident ; 2=A little confident ; 3=Somewhat confident ; 4=Confident ; 5=Very confident
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTQSTNS} \emph{character} How confident did you feel answering this participant's questions? 1=Not at all confident ; 2=A little confident ; 3=Somewhat confident ; 4=Confident ; 5=Very confident
\emph{Decoded Value: }\strong{Yes}
\item \strong{VSTRESULT} \emph{character} Overall, how did this visit go? 1=Poor ; 2=Fair ; 3=Good ; 4=Very good ; 5=Excellent
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 3900, ..., 128118
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{enrollment}
