% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMYMETA}
\alias{AMYMETA}
\title{Amyloid PET Scan Information}
\format{
A data frame with 2969 observations and 39 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AMYMETA)
}
\description{
Amyloid PET Scan Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{DONE} \emph{character} Was the scan conducted? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 
\itemize{
\item \emph{ADNI3:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Not called for by protocol ; 8=Software/scanner error ; 99=Other (specify)
\item \emph{ADNI4:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Not called for by protocol ; 8=Software/scanner error ; 9=Tracer issues ; 99=Other (specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{TRACERISS} \emph{character} Tracer issues 1=Dose failure ; 2=Late delivery or other issues with delivery ; 3=Incorrect tracer delivered by radiopharmacy ; 4=Incorrect tracer ordered by site ; 99=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{RADTRACER} \emph{character} If scan not done, was radiotracer administered? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{TRACERTYPE} \emph{character} Amyloid Imaging radiotracer used 
\itemize{
\item \emph{ADNI3:} 1=Florbetaben (Neuraceq) ; 2=Florbetapir (Amyvid)
\item \emph{ADNI4:} 1=Florbetaben (Neuraceq) ; 2=Florbetapir (Amyvid) ; 3=NAV4694
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SCANDATE} \emph{ } Scan Date/Dose Date Date: YYY-MM-DD
\item \strong{QCTIME} \emph{ } Time of today's Scanner QC hms: HH-MM-SS
\item \strong{ASSAYTIME} \emph{ } Time of dose assay hms: HH-MM-SS
\item \strong{DOSEFLRBTBN} \emph{numeric} Florbetaben (Neuraceq):  Injected dose assay Range value: 5, ..., 10.5
\item \strong{DOSEFLRBTPR} \emph{numeric} Florbetapir (Amyvid):  Injected dose assay Range value: 5.1, ..., 13.6
\item \strong{DOSENAV} \emph{numeric} NAV4694: Injected dose assay Range value: 6.2, ..., 8.9
\item \strong{INJTIME} \emph{ } Time of injection hms: HH-MM-SS
\item \strong{SCANTIME} \emph{ } Emission Scan Start Time hms: HH-MM-SS
\item \strong{EARLYADD} \emph{character} Does the study include an 'early add-on' scan (0-20) min? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{EARLYNO} \emph{character} If no, why? 1=Not a participating site in the early add on sequence study ; 2=Subject did not consent to the early add on sequence study ; 3=Administrative problems ; 4=Software/scanner error ; 5=Withdrawn consent ; 99=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{EARLYTIME} \emph{ } If yes, Early add on sequence start time: [H H M M] hms: HH-MM-SS
\item \strong{VARIAT} \emph{character} Any variations from protocol during uptake? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PROTID} \emph{character} Predefined Acquisition Protocol Name  
\item \strong{MOTION} \emph{character} Participant motion problems 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{OTHERVAR} \emph{character} Other protocol variations 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{RECON} \emph{character} Select reconstruction used 
\itemize{
\item \emph{ADNI3:} 1=3D or LOR Ramla (Philips) ; 2=FORE/2D-OSEM (Siemens) ; 4=3D-OSEM (Siemens) ; 5=3D Iterative (VPHD / GE) ; 6=FORE Iterative (GE)
\item \emph{ADNI4:} 1=3D or LOR Ramla (for Philips Gemini scanners) ; 2=FORE/2D-OSEM (Siemens) ; 4=3D-OSEM (Siemens) ; 5=3D Iterative (VPHD / GE) ; 6=FORE Iterative (GE) ; 7=3D-OSEM (Philips Vereos and Ingenuity Scanners)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SHARP} \emph{character} For Philips Gemini scanners: was "smooth" parameter set to "sharp"? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{ARCHIVE} \emph{character} Data archived locally? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{LONI} \emph{character} Data transferred to LONI? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{TRANDATE} \emph{ } Transfer Date Date: YYY-MM-DD
\item \strong{ID} \emph{numeric} Record ID Range value: 225, ..., 187097
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{adni4}
\keyword{imaging}
