% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMYREAD}
\alias{AMYREAD}
\title{Amyloid PET Read}
\format{
A data frame with 883 observations and 24 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AMYREAD)
}
\description{
Amyloid PET Read data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNIGO, ADNI2, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10888
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{SCANDATE} \emph{ } Scan acquisition date: Date: YYY-MM-DD
\item \strong{TRACERTYPE} \emph{character} Tracer type 1=18F-Florbetaben/NeuraCeqTM (Threshold SUVR 1.12) ; 2=18F-AV-45/Florbetapir/AmyvidTM (Threshold SUVR 1.17) ; 3=18F-NAV4694 (Threshold SUVR 1.14)
\emph{Decoded Value: }\strong{Yes}
\item \strong{READDATE} \emph{ } Visual Read Date Date: YYY-MM-DD
\item \strong{OUTCOME} \emph{character} Based on tracer-specific guidelines, the scan was read as: 0=Non Elevated (no evidence for cortical tracer binding) ; 1=Elevated (evidence for cortical tracer binding)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CORTREGION} \emph{character} Indicate regions where cortical tracer retention is observed in one or both hemispheres: 1=Frontal ; 2=Parietal ; 3=Post Cingulate/Precuneus ; 4=Temporal ; 5=Occipital ; 6=None
\emph{Decoded Value: }\strong{Yes}
\item \strong{CONSENS} \emph{character} Is consensus meeting needed for this scan? 1=No, visual read and quantification are congruent ; 2=Yes, this scan should be reviewed in consensus session
\emph{Decoded Value: }\strong{Yes}
\item \strong{CONGRU} \emph{character} No, visual read and quantification are congruent, and scan is 0=Non Elevated scan ; 1=Elevated scan
\emph{Decoded Value: }\strong{Yes}
\item \strong{CONSRSN} \emph{character} Yes, this scan should be reviewed in consensus session because: 1=Incongruency between initial visual read and quantification ; 2=Other reason
\emph{Decoded Value: }\strong{Yes}
\item \strong{CONSENSDATE} \emph{ } Date of consensus meeting Date: YYY-MM-DD
\item \strong{CONSENSRES} \emph{character} Consensus Result? 0=Non Elevated scan ; 1=Elevated scan
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 2718, ..., 128293
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{enrollment}
