% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANTIAMYTX}
\alias{ANTIAMYTX}
\title{Anti-Amyloid Treatment Information}
\format{
A data frame with 20 observations and 21 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ANTIAMYTX)
}
\description{
Anti-Amyloid Treatment Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI3
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 6702, ..., 10871
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_sc, 4_m12, 4_bl
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: sc, m12, m66, bl
\item \strong{TXREC} \emph{character} This Participant received an Anti-Amyloid treatment. 1=Yes - I confirm this participant has received an Anti-Amyloid Treatment.
\emph{Decoded Value: }\strong{Yes}
\item \strong{TXNAME} \emph{character} Treatment Name 
\itemize{
\item \emph{ADNI3:} 1=Aducanumab
\item \emph{ADNI4:} 1=Aducanumab ; 2=Lecanumab ; 3=Donanemab
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{STARTDATE} \emph{ } Treatment Start Date Date: YYY-MM-DD
\item \strong{ENDDATE} \emph{ } Treatment End Date Date: YYY-MM-DD
\item \strong{TXUNITS} \emph{character} Units 1=milligram (mg) / milliliter (mL) ; 99=other (OT)
\emph{Decoded Value: }\strong{Yes}
\item \strong{TXFREQNC} \emph{character} Frequency 1=every 4 weeks (q4w) ; 99=other (OT)
\emph{Decoded Value: }\strong{Yes}
\item \strong{TXROUTE} \emph{character} Route 1=intravenous (iv) ; 99=other (OT)
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 6236, ..., 124763
\item \strong{SITEID} \emph{numeric} Site ID Range value: 10, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{medical_history}
