% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APOERES}
\alias{APOERES}
\title{ApoE Genotyping - Results}
\format{
A data frame with 3008 observations and 17 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(APOERES)
}
\description{
ApoE Genotyping - Results data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10787
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, f
\item \strong{GENOTYPE} \emph{character} Apolipoprotein E (Apo-E) Genotype , with options: 3/3, 3/4, 2/3, 4/4, 2/4, 2/2
\item \strong{APTESTDT} \emph{ } Date Test Performed Date: YYY-MM-DD
\item \strong{APVOLUME} \emph{numeric} Volume of Blood Shipped in Lavendar Top Tube 1..10
\item \strong{APRECEIVE} \emph{character} Sample recieved within 24 hours of blood draw? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{APAMBTEMP} \emph{character} Sample shipped at ambient temperature? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{APRESAMP} \emph{character} Request Resample? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{APUSABLE} \emph{character} Sample Useable? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 2330
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{logical} Date record last updated A Boolean value: TRUE/ FALSE
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{biospecimen}
