% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ARM}
\alias{ARM}
\title{ARM}
\format{
A data frame with 4705 observations and 15 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ARM)
}
\description{
ARM data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 5296
\item \strong{PTNO} \emph{numeric} 4 digit Participant Number Range value: 1, ..., 1435
\item \strong{TYPE} \emph{character} ID Type S=Study Participant ; P=Phantom ; V=Volunteer
\emph{Decoded Value: }\strong{Yes}
\item \strong{ARM} \emph{character} Diagnosis/Scan Category Assignment 
\itemize{
\item \emph{ADNI1:} 1=NL - 1.5T only ; 2=MCI - 1.5T only ; 3=AD - 1.5T only ; 4=NL - PET+1.5T ; 5=MCI - PET+1.5T ; 6=AD - PET+1.5T ; 7=NL - 3T+1.5T ; 8=MCI - 3T+1.5T ; 9=AD - 3T+1.5T ;
\item \emph{ADNIGO:} 1=NL - 1.5T only ; 2=MCI - 1.5T only ; 3=AD - 1.5T only ; 4=NL - PET+1.5T ; 5=MCI - PET+1.5T ; 6=AD - PET+1.5T ; 7=NL - 3T+1.5T ; 8=MCI - 3T+1.5T ; 9=AD - 3T+1.5T ; 10=EMCI
\item \emph{ADNI2:} 1=NL - (ADNI1 1.5T only) ; 2=LMCI - (ADNI1 1.5T only) ; 3=AD - (ADNI1 1.5T only) ; 4=NL - (ADNI1 PET+1.5T) ; 5=LMCI - (ADNI1 PET+1.5T) ; 6=AD - (ADNI1 PET+1.5T) ; 7=NL - (ADNI1 3T+1.5T) ; 8=LMCI - (ADNI1 3T+1.5T) ; 9=AD - (ADNI1 3T+1.5T) ; 10=EMCI ; 11=SMC - (Significant Memory Concern)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{ENROLLED} \emph{character} Enrollment Status 
\itemize{
\item \emph{ADNI1:} null=pending enrollment ; 1=randomized - assigned a scan category ; -1=screen failed ; -2=screen failed after randomization
\item \emph{ADNIGO:} null=pending enrollment ; 1=randomized - assigned a scan category ; -1=screen failed ; -2=screen failed after randomization ; 2=Enrolled
\item \emph{ADNI2:} null=pending enrollment ; 1=randomized - assigned a scan category ; -1=screen failed ; -2=phantom screen ; 2=Approved for enrollment in ADNIGO ; 3=Approved for enrollment in ADNI2
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RANDDATE} \emph{ } Randomization Date Date: YYY-MM-DD
\item \strong{MAPPDATE} \emph{ } Monitor Approval Date Date: YYY-MM-DD
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 1715
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{enrollment}
