% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-assert.R
\name{assert_non_missing}
\alias{assert_non_missing}
\title{Checks Missing Values}
\usage{
assert_non_missing(...)
}
\arguments{
\item{...}{\code{\link[assertr]{assert}} arguments}
}
\value{
An assertr error message if there is record with missing values.
}
\description{
This function is used to check any missing values within a provided column.
}
\examples{
\dontrun{
library(assertr)
ADNIMERGE2::REGISTRY \%>\%
  assert_non_missing(RID)
}
}
\seealso{
\code{\link[assertr]{assert}}
 \code{\link[assertr]{not_na}}

Other assert checks: 
\code{\link{assert_uniq}()}
}
\concept{assert checks}
\keyword{utils_fun}
