% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-assert.R
\name{assert_uniq}
\alias{assert_uniq}
\title{Checks Duplicated Values/Records}
\usage{
assert_uniq(...)
}
\arguments{
\item{...}{\code{\link[assertr]{assert_rows}} arguments}
}
\value{
An assertr error message if there is any duplicated records.
}
\description{
This function is used to check any duplicated records within provided column(s).
 Either a single column or across multiple columns.
}
\examples{
\dontrun{
library(assertr)
ADNIMERGE2::REGISTRY \%>\%
  assert_uniq(RID, VISCODE)
}
}
\seealso{
\code{\link[assertr]{assert_rows}}
  \code{\link[assertr]{col_concat}}
  \code{\link[assertr]{is_uniq}}

Other assert checks: 
\code{\link{assert_non_missing}()}
}
\concept{assert checks}
\keyword{utils_fun}
