% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AV45FOLLOW}
\alias{AV45FOLLOW}
\title{AV-45 24-48 Hour Follow-Up}
\format{
A data frame with 1745 observations and 16 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AV45FOLLOW)
}
\description{
AV-45 24-48 Hour Follow-Up data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 8, ..., 2978
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5285
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 61
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{CONTACT} \emph{character} Was 24-48 hours post imaging follow-up telephone contact made? 
\itemize{
\item \emph{ADNIGO:} 1=Yes ; 0=No ; 2=N/A - no AV-45 scan conducted
\item \emph{ADNI2:} 1=Yes ; 0=No ; 2=N/A - no Florbetapir F18 scan conducted
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CALLDATE} \emph{ } Date of telephone contact: Date: YYY-MM-DD
\item \strong{CALLTIME} \emph{ } Time of telephone contact: hms: HH-MM-SS
\item \strong{CONTACTED} \emph{character} Person who was contacted: 1=Participant ; 2=Study Partner
\emph{Decoded Value: }\strong{Yes}
\item \strong{CALLAE} \emph{character} Were any Adverse Events reported? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adnigo}
\keyword{medical_history}
