% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AV45META}
\alias{AV45META}
\title{AV-45 PET Scan Information}
\format{
A data frame with 3400 observations and 52 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AV45META)
}
\description{
AV-45 PET Scan Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 8, ..., 6410
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5296
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 61
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 3
\item \strong{PMCONDCT} \emph{character} Was the scan conducted? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMREASON} \emph{character} Reason why the scan was not conducted: 
\itemize{
\item \emph{ADNIGO:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify)
\item \emph{ADNI2:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 8=Software/scanner error ; 5=Withdrawn consent ; 7=Not called for by the protocol ; 6=Other (specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RADTRACER} \emph{character} If no, was radiotracer administered? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{PMSCANNER} \emph{character} GE 1=check here
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMGEMODEL} \emph{character} If GE, Scanner Model: 
\itemize{
\item \emph{ADNIGO:} 1=Advance ; 2=Discovery LS ; 3=Discovery ST ; 4=Discovery RX ; 5=Discovery STE/VCT
\item \emph{ADNI2:} 1=Advance ; 2=Discovery LS ; 3=Discovery ST ; 4=Discovery RX ; 5=Discovery STE/VCT ; 6=Discovery 600 ; 7=Discovery 690
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSIEMENS} \emph{character} Siemens 1=check here
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSIEMODEL} \emph{character} If Siemens, Scanner Model: 1=ACCEL/EXACT ; 2=Biograph (Model 1023/1024) ; 7=Biograph HiRes (Model 1080) ; 3=BioGraph TruePoint (Model 1093/1094) ; 4=BioGraph mCT ; 5=HR+ ; 6=HRRT
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMPHILLIPS} \emph{character} Phillips 1=check here
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMPHMODEL} \emph{character} If Phillips, Scanner Model: 1=Allegro ; 2=Gemini ; 3=Gemini-GXL ; 4=Gemini-TF
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMQCTIME} \emph{character} Time of today's Scanner QC  
\item \strong{PMFDGTIME} \emph{character} Time of Florbetapir F18 dose assay  
\item \strong{PMFDGDOS} \emph{numeric} Florbetapir F18 dose assay Range value: 0.4, ..., 98
\item \strong{PMFDGVOL} \emph{numeric} Florbetapir F18 Volume Range value: 0, ..., 97
\item \strong{PMINJTIME} \emph{character} Time of Florbetapir F18 injection  
\item \strong{PMSCTIME} \emph{character} Standard Emission Scan Start Time  
\item \strong{PMVARIAT} \emph{character} Any variations from protocol during Florbetapir F18 uptake? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMPROTID} \emph{character} Predefined Acquisition Protocol ID  
\item \strong{PMFRAME} \emph{character} For the standard late acquisition, which framing rate was used? 
\itemize{
\item \emph{ADNIGO:} 1=4 frames, 5 min/frame (4x300s) ; 2=2 scans, 10 min each (2x600s) (only for BioGraph scanners without list-mode)
\item \emph{ADNI2:} 1=4 frames, 5 min/frame (4x300s) ; 2=2 scans, 10 min each (2x600s) (only for BioGraph scanners without list-mode) ; 3=1 frame, 20 min (1x1200s) (only for the oldest BioGraph scanners:  Models 1023 or 1024)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMDEVIAT} \emph{character} If any deviations, describe:  
\item \strong{PMMOTION} \emph{character} Subject motion problems: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMMALFUN} \emph{character} Scanner malfunction 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMOTHER} \emph{character} Other protocol variations: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMRECON} \emph{character} Check which of the following reconstructions was used: 
\itemize{
\item \emph{ADNIGO:} 2=FORE/2D-OSEM (Siemens) ; 4=OSEM3D (Siemens) (If HHRT scanners using OP, please select OSEM3D) ; 5=3D Iterative (GE) ; 1=3D-Ramla (Philips) ; 3=3D Back-projection (GE)
\item \emph{ADNI2:} 2=FORE/2D-OSEM (Siemens) ; 4=OSEM3D (Siemens) (If HHRT scanners using OP, please select OSEM3D) ; 5=3D Iterative (GE) ; 6=FORE Iterative (GE) ; 1=3D or LOR Ramla (Philips) ; 3=3D Back-projection (GE)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSUBSET} \emph{character} # subsets: 
\itemize{
\item \emph{ADNIGO:} 14=14 ; 16=16 ; 20=20 ; 99=Other
\item \emph{ADNI2:} 12=12 ; 14=14 ; 16=16 ; 20=20 ; 21=21 ; 24=24 ; 32=32 ; 99=Other
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSUBSPE} \emph{numeric} If Other, specify: Range value: 12, ..., 32
\item \strong{PMITERAT} \emph{character} # iterations: 4=4 ; 6=6 ; 99=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMITERSPE} \emph{numeric} If Other, specify: Range value: 2, ..., 2
\item \strong{PMRAMLA} \emph{numeric} If 3D or LOR Ramla, Lambda= Range value: 0.008, ..., 1
\item \strong{PMSHARP} \emph{character} Was 'Smooth' parameter set to 'Sharp'? 1=Check here to confirm
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMFILTER} \emph{character} If 3D Back-Projection, Ramp filter? 1=Check here to confirm
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMMODEON} \emph{character} If FORE/2D-OSEM select one of the following: 
\itemize{
\item \emph{ADNIGO:} 1=Brain Mode "ON" for PET-only Siemens scanners ; 2=TRIM "ON" for PET/CT Siemens scanners (older software versions) ; 3=TRIM not available for PET/CT scanners (new software versions)
\item \emph{ADNI2:} 1=Brain Mode 'ON' for PET-only Siemens scanners ; 2=TRIM 'ON' for PET/CT Siemens scanners (older software versions) ; 3=TRIM not available for PET/CT scanners (new software versions)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSMOOTH} \emph{character} No post-process smoothing: 1=Check here to confirm
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMATTEN} \emph{character} Attenuation Correction: 1=CT ; 2=Ge-68+Segmentation ; 3=Cs-137+Segmentation
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMTRNSFR} \emph{character} Was data transferred to LONI within 24 hours of scan? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMTRNDATE} \emph{character} Transfer Date  
\item \strong{PMTRNDATEYR_DRVD} \emph{numeric} Transfer Year (Derived from PMTRNDATE) Range value: 2001, ..., 2017
\item \strong{PMARCHIVE} \emph{character} Was all raw PET data archived locally to be able to do complete reconstruction of PET Scan if needed? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMLPDONE} \emph{character} Was a Lumbar Puncture completed prior to the Florbetapir F18 scan? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMLPINTER} \emph{character} If Yes, What was the interval between LP and Florbetapir F18? 1=less than 6 hours ; 2=6-12 hours ; 3=13-24 hours ; 4=25-48 hours ; 5=49-72 hours ; 6=more than 72 hours
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adnigo}
\keyword{imaging}
