% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AV45QC}
\alias{AV45QC}
\title{AV-45 PET QC Tracking}
\format{
A data frame with 2707 observations and 39 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AV45QC)
}
\description{
AV-45 PET QC Tracking data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 4754
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 5296
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 61
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 6
\item \strong{SCAN} \emph{character} Scan 1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{PQFILENO} \emph{numeric} Upload Number Range value: 1, ..., 6
\item \strong{PQDATTRA} \emph{ } Date images uploaded to LONI Date: YYY-MM-DD
\item \strong{LONIUID} \emph{character} LONI Unique Series ID  
\item \strong{PQDATE} \emph{ } Date reviewed Date: YYY-MM-DD
\item \strong{PQACQFR} \emph{character} Standard Sequence Acquisition 
\itemize{
\item \emph{ADNIGO:} 1=4 frames, 5 min/frame ; 2=2 scans, 10 min each
\item \emph{ADNI2:} 1=4 frames, 5 min/frame ; 2=2 scans, 10 min each ; 3=1 scan, 20min ; 4=Not applicable - early add on sequence
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQCORALG} \emph{character} Correct Reconstruction Algorithm? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQRECON} \emph{character} Check which of the following reconstruction algorithms was used 
\itemize{
\item \emph{ADNIGO:} 2=FORE/2D-OSEM (Siemens) ; 4=OSEM3D (Siemens) ; 5=3D Iterative (GE) ; 1=3D-Ramla (Philips) ; 3=3D Back-projection (GE)
\item \emph{ADNI2:} 2=FORE/2D-OSEM (Siemens) ; 4=OSEM3D (Siemens) ; 5=3D Iterative (GE) ; 6=FORE Iterative (GE) ; 1=3D or LOR Ramla (Philips) ; 3=3D Back-projection (GE)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQCORNUM} \emph{character} If FORE/2D-OSEM or OSEM3D or 3D Iterative or FORE Iterative, correct # subsets and iterations? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQCORFIL} \emph{character} Correct Filters/TRIM? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQCORTHI} \emph{character} Correct Slice Thickness (PET/CT systems)? 1=Yes ; 0=No ; 2=N/A (PET only system)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQVOXEL} \emph{character} Acceptable Voxel Size? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQALLFRAME} \emph{character} Were all frames acceptable? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQUNUSABLE} \emph{character} Indicate which frames were unacceptable:  
\item \strong{PQREASON} \emph{character} Indicate why frames were unacceptable: 1=Subject motion ; 2=Other (Please Specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQREASSP} \emph{character} If Other, Specify:  
\item \strong{PASS} \emph{character} Pass QC? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQPROERR} \emph{character} A. Processing Error(s) 1=Incorrect reconstruction algorithm ; 2=Incorrect filtering ; 3=Incorrect slice thickness ; 4=Incorrect voxel size ; 5=Attenuation Error ; 6=Normalization Error ; 7=Inadequate counts ; 8=Inadequate field of view ; 9=Incorrect acquisition protocol ; 10=Unfixable image artifact (e.g. attenuation, normalization, etc.) ; 99=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQERRSP} \emph{character} Specify  
\item \strong{PQACTION} \emph{character} B. QC outcome 1=Request Reprocessing ; 2=Request Rescan ; 3=Unuseable - Don't request rescan (provide details) ; 4=Allow Rescan (rescan not requested)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQNONESP} \emph{character} If unusable, reason for not rescanning  
\item \strong{PQISSUES} \emph{character} Imaging issues 1=Motion - minimal ; 2=Motion - mild ; 3=Motion - mild-to-moderate ; 4=Motion - moderate ; 5=Motion - moderate-to-severe ; 6=Top Cutoff - none ; 7=Top Cutoff - 0-3mm ; 8=Top Cutoff - 3-6mm ; 9=Top Cutoff - 6-9mm ; 10=Top Cutoff - >9mm ; 11=Bottom Cutoff - none ; 12=Bottom Cutoff - 0-3mm ; 13=Bottom Cutoff - 3-6mm ; 14=Bottom Cutoff - 6-9mm ; 15=Bottom Cutoff - >9mm ; 99=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQISSOTH} \emph{character} Other, specify , with options: Good Images 3DRP
\item \strong{COMMENTS} \emph{character} Additional QC Comments  
\item \strong{QUARANTINE} \emph{character} Release from quarantine? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PAYSITE} \emph{character} Pay Site? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adnigo}
\keyword{imaging}
