% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AV45VITALS}
\alias{AV45VITALS}
\title{AV-45 Pre and Post Injection Vitals}
\format{
A data frame with 1802 observations and 28 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(AV45VITALS)
}
\description{
AV-45 Pre and Post Injection Vitals data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 6, ..., 3156
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5288
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 61
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 2
\item \strong{SCANDONE} \emph{character} Was scan conducted? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{SCANDATE} \emph{ } Florbetapir F18 Scan Date: Date: YYY-MM-DD
\item \strong{PREHEART} \emph{numeric} Heart Rate (bpm): 40..130
\item \strong{PRERESP} \emph{numeric} Respiration (per minute): 6..40
\item \strong{PRESYSTBP} \emph{numeric} Systolic Blood Pressure (mmHg): 76..250
\item \strong{PREDIABP} \emph{numeric} Diastolic Blood Pressure (mmHg): 36..130
\item \strong{PRETEMP} \emph{numeric} Temperature: Range value: -1, ..., 99.9
\item \strong{PRETEMPS} \emph{character} Temperature Source: 1=Oral ; 2=Tympanic ; 3=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PRETEMPU} \emph{character} Temperature Units: 1=Fahrenheit ; 2=Celsius
\emph{Decoded Value: }\strong{Yes}
\item \strong{POSTHEART} \emph{numeric} Heart Rate (bpm): 40..130
\item \strong{POSTRESP} \emph{numeric} Respiration (per minute): 6..40
\item \strong{POSTSYSTBP} \emph{numeric} Systolic Blood Pressure (mmHg): 76..250
\item \strong{POSTDIABP} \emph{numeric} Diastolic Blood Pressure (mmHg): 36..130
\item \strong{POSTTEMP} \emph{numeric} Temperature: Range value: -1, ..., 99.5
\item \strong{POSTTEMPS} \emph{character} Temperature Source: 1=Oral ; 2=Tympanic ; 3=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{POSTTEMPU} \emph{character} Temperature Units: 1=Fahrenheit ; 2=Celsius
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adnigo}
\keyword{medical_history}
