% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAIMRINMRC}
\alias{BAIMRINMRC}
\title{BAI -  MRI NMRC Summaries: 2016-09-12}
\format{
A data frame with 726 observations and 12 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BAIMRINMRC)
}
\description{
BAI -  MRI NMRC Summaries: 2016-09-12 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 3, ..., 5290
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, m12, m24, v02, v21, v11
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: sc, m12, m24, scmri
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{VERSION} \emph{numeric} Version of Table Range value: 1, ..., 2
\item \strong{LONIUID} \emph{numeric}   Range value: 31933, ..., 557537
\item \strong{RUNDATE} \emph{ } DATE analysis performed Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Complete/Incomplete , with options: Complete
\item \strong{IPCA} \emph{numeric} percent annualized whole brain atrophy detected by Iterative Principal component Analysis (IPCA) Range value: -0.114104, ..., 0.869081
\item \strong{IPCA2} \emph{numeric} whole brain atrophy measured from T1 weighted MRI using iterative pincipal component analysis Range value: -0.0093138, ..., 1.16992
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
