% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAIPETNMRC}
\alias{BAIPETNMRC}
\title{BAI -  PET NMRC Summaries: 2018-04-12}
\format{
A data frame with 5598 observations and 345 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BAIPETNMRC)
}
\description{
BAI -  PET NMRC Summaries: 2018-04-12 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 6009
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{numeric} Version of Table Range value: 18, ..., 22
\item \strong{LONIUID} \emph{numeric} LONI Unique Image Identifier Range value: 25392, ..., 880516
\item \strong{RUNDATE} \emph{ } DATE analysis performed Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Complete/Incomplete , with options: Complete
\item \strong{MODALITY} \emph{character} AV45/FDG/TAU , with options: FDG, AV45
\item \strong{HIPPL01} \emph{numeric} Globally normalized CMRgl from Hippocampus_L Range value: 0.845947, ..., 1.57295
\item \strong{FRTSUPL01} \emph{numeric} Globally normalized CMRgl from Frontal_sup_L Range value: 1.15639, ..., 2.1298
\item \strong{FRTMIDL01} \emph{numeric} Globally normalized CMRgl from Frontal_mid_L Range value: 1.11277, ..., 2.12475
\item \strong{PARAHIPL01} \emph{numeric} Globally normalized CMRgl from ParaHippocampal_L Range value: 0.912239, ..., 1.5574
\item \strong{FUSFRML01} \emph{numeric} Globally normalized CMRgl from Fusiform_L Range value: 0.833843, ..., 1.61013
\item \strong{OCCMIDL01} \emph{numeric} Globally normalized CMRgl from Occipital_mid_L Range value: 1.08684, ..., 1.97188
\item \strong{ANGULL01} \emph{numeric} Globally normalized CMRgl from Angular_L Range value: 0.815963, ..., 2.18265
\item \strong{PARIINFL01} \emph{numeric} Globally normalized CMRgl from Parietal_Inf_L Range value: 0.802001, ..., 1.88925
\item \strong{SUPMRGL01} \emph{numeric} Globally normalized CMRgl from SupraMarginal_L Range value: 1.2772, ..., 1.92524
\item \strong{TMPMIDL01} \emph{numeric} Globally normalized CMRgl from Temporal _mid_L Range value: 0.995531, ..., 1.7407
\item \strong{PRECUNL01} \emph{numeric} Globally normalized CMRgl from Precuneus_L Range value: 1.29607, ..., 2.24714
\item \strong{CINGPSTL01} \emph{numeric} Globally normalized CMRgl from Cingulum_Post_L Range value: 0.972679, ..., 2.25141
\item \strong{PARAHIPR01} \emph{numeric} Globally normalized CMRgl from ParaHippocampal_R Range value: 0.806904, ..., 1.61033
\item \strong{FRTSUPR01} \emph{numeric} Globally normalized CMRgl from Frontal_Sup_R Range value: 1.34728, ..., 2.17756
\item \strong{OCCMIDR01} \emph{numeric} Globally normalized CMRgl from Occipital_mid_R Range value: 1.05336, ..., 1.97827
\item \strong{FUSFRMR01} \emph{numeric} Globally normalized CMRgl from Fusiform_R Range value: 0.829141, ..., 1.71375
\item \strong{FRTMIDR01} \emph{numeric} Globally normalized CMRgl from Frontal_mid_R Range value: 1.11686, ..., 1.79081
\item \strong{ANGULR01} \emph{numeric} Globally normalized CMRgl from Angular_R Range value: 1.15124, ..., 2.26394
\item \strong{PARIINFR01} \emph{numeric} Globally normalized CMRgl from Parietal_Inf_R Range value: 1.13354, ..., 2.0973
\item \strong{TMPMIDR01} \emph{numeric} Globally normalized CMRgl from Temporal_mid_R Range value: 1.18349, ..., 2.17798
\item \strong{SUPMRGR01} \emph{numeric} Globally normalized CMRgl from SupraMarginal_R Range value: 1.01858, ..., 1.62387
\item \strong{PRCUNSR01} \emph{numeric} Globally normalized CMRgl from Prcuneus_R Range value: 1.32524, ..., 2.38231
\item \strong{HIPPR01} \emph{numeric} Globally normalized CMRgl from Hippocampus_R Range value: Inf, ..., -Inf
\item \strong{LINGUALL01} \emph{numeric} Globally normalized CMRgl from Lingual_L Range value: Inf, ..., -Inf
\item \strong{LINGUALR01} \emph{numeric} Globally normalized CMRgl from Lingual_R Range value: Inf, ..., -Inf
\item \strong{CINGPSTR01} \emph{numeric} Globally normalized CMRgl from Cingulum_Post_R Range value: Inf, ..., -Inf
\item \strong{FRTINFL01} \emph{numeric} Globally normalized CMRgl from Frontal_Inf_L Range value: Inf, ..., -Inf
\item \strong{FRTINFR01} \emph{numeric} Globally normalized CMRgl from Frontal_Inf_R Range value: Inf, ..., -Inf
\item \strong{PARISUPL01} \emph{numeric} Globally normalized CMRgl from Parietal_Sup_L Range value: Inf, ..., -Inf
\item \strong{PARISUPR01} \emph{numeric} Globally normalized CMRgl from Parietal_Sup_R Range value: Inf, ..., -Inf
\item \strong{INSULAL01} \emph{numeric} Globally normalized CMRgl from Insula_L Range value: Inf, ..., -Inf
\item \strong{INSULAR01} \emph{numeric} Globally normalized CMRgl from Insula_R Range value: Inf, ..., -Inf
\item \strong{CINGANTL01} \emph{numeric} Globally normalized CMRgl from Cingulum_Ant_L Range value: Inf, ..., -Inf
\item \strong{CINGANTR01} \emph{numeric} Globally normalized CMRgl from Cingulum_Ant_R Range value: Inf, ..., -Inf
\item \strong{CINGMIDL01} \emph{numeric} Globally normalized CMRgl from Cingulum_Mid_L Range value: Inf, ..., -Inf
\item \strong{CINGMIDR01} \emph{numeric} Globally normalized CMRgl from Cingulum_Mid_R Range value: Inf, ..., -Inf
\item \strong{TMPSUPL01} \emph{numeric} Globally normalized CMRgl from Temporal _Sup_L Range value: Inf, ..., -Inf
\item \strong{TMPSUPR01} \emph{numeric} Globally normalized CMRgl from Temporal _Sup_R Range value: Inf, ..., -Inf
\item \strong{TMPINFL01} \emph{numeric} Globally normalized CMRgl from Temporal_Inf_L Range value: Inf, ..., -Inf
\item \strong{TMPINFR01} \emph{numeric} Globally normalized CMRgl from Temporal_Inf_R Range value: 1.07602, ..., 1.93136
\item \strong{FRTSUPL02} \emph{numeric} Globally normalized CMRgl from Frontal_Sup_L Range value: 1.04989, ..., 1.88073
\item \strong{FRTMIDL02} \emph{numeric} Globally normalized CMRgl from Frontal_Mid_L Range value: 1.09853, ..., 2.05721
\item \strong{CINGPSTL02} \emph{numeric} Globally normalized CMRgl from Cingulum_Post_L Range value: 1.44932, ..., 2.52221
\item \strong{CINGMIDL02} \emph{numeric} Globally normalized CMRgl from Cingulum_Mid_L Range value: 1.46622, ..., 2.48963
\item \strong{FRTSMEDL02} \emph{numeric} Globally normalized CMRgl from Frontal_Sup_Medial_L Range value: 1.09455, ..., 2.19089
\item \strong{FUSFRML02} \emph{numeric} Globally normalized CMRgl from Fusiform_L Range value: 0.808137, ..., 1.655
\item \strong{TMPMIDL02} \emph{numeric} Globally normalized CMRgl from Temporal_Mid_L Range value: 0.953942, ..., 1.80762
\item \strong{OCCMIDL02} \emph{numeric} Globally normalized CMRgl from Occipital_Mid_L Range value: 1.09092, ..., 2.03661
\item \strong{TMPINFL02} \emph{numeric} Globally normalized CMRgl from Temporal_Inf_L Range value: 0.821633, ..., 1.55902
\item \strong{PRECUNL02} \emph{numeric} Globally normalized CMRgl from Precuneus_L Range value: 1.36488, ..., 2.25017
\item \strong{CINGANTL02} \emph{numeric} Globally normalized CMRgl from Cingulum_Ant_L Range value: 1.12117, ..., 2.20225
\item \strong{FRTMIDOR02} \emph{numeric} Globally normalized CMRgl from Frontal_Mid_Orb_R Range value: 1.21638, ..., 2.2145
\item \strong{FRTMIDR02} \emph{numeric} Globally normalized CMRgl from Frontal_Mid_R Range value: 1.27595, ..., 2.08287
\item \strong{FRTSUPR02} \emph{numeric} Globally normalized CMRgl from Frontal_Sup_R Range value: 1.32622, ..., 2.19736
\item \strong{ANGULR02} \emph{numeric} Globally normalized CMRgl from Angular_R Range value: 1.12105, ..., 2.09255
\item \strong{ANGULL02} \emph{numeric} Globally normalized CMRgl from Angular_L Range value: 0.874483, ..., 2.12218
\item \strong{PARIINFR02} \emph{numeric} Globally normalized CMRgl from Parietal_Inf_R Range value: 1.17451, ..., 2.25359
\item \strong{TMPMIDR02} \emph{numeric} Globally normalized CMRgl from Temporal_Mid_R Range value: 1.11721, ..., 2.02806
\item \strong{INSULAR02} \emph{numeric} Globally normalized CMRgl from Insula_R Range value: 0.908476, ..., 1.75238
\item \strong{TMPPOSR02} \emph{numeric} Globally normalized CMRgl from Temporal_Pole_Sup_R Range value: 0.831152, ..., 1.67344
\item \strong{CINGANTR02} \emph{numeric} Globally normalized CMRgl from Cingulum_Ant_R Range value: 1.07998, ..., 1.99467
\item \strong{CINGMIDR02} \emph{numeric} Globally normalized CMRgl from Cingulum_Mid_R Range value: 1.36282, ..., 2.47688
\item \strong{PRECUNR02} \emph{numeric} Globally normalized CMRgl from Precuneus_R Range value: 1.362, ..., 2.48783
\item \strong{RECTUSR02} \emph{numeric} Globally normalized CMRgl from Rectus_R Range value: 0.990824, ..., 1.71868
\item \strong{CINGPSTR02} \emph{numeric} Globally normalized CMRgl from Cingulum_Post_R Range value: Inf, ..., -Inf
\item \strong{PARIINFL02} \emph{numeric} Globally normalized CMRgl from Parietal_Inf_L Range value: Inf, ..., -Inf
\item \strong{INSULAL02} \emph{numeric} Globally normalized CMRgl from Insula_L Range value: Inf, ..., -Inf
\item \strong{FRTINFL02} \emph{numeric} Globally normalized CMRgl from Frontal_Inf_L Range value: Inf, ..., -Inf
\item \strong{FRTINFR02} \emph{numeric} Globally normalized CMRgl from Frontal_Inf_R Range value: Inf, ..., -Inf
\item \strong{OCCMIDR02} \emph{numeric} Globally normalized CMRgl from Occipital_Mid_R Range value: Inf, ..., -Inf
\item \strong{FRTMIDOL02} \emph{numeric} Globally normalized CMRgl from Frontal_Mid_Orb_Right Range value: Inf, ..., -Inf
\item \strong{FRTSMEDR02} \emph{numeric} Globally normalized CMRgl from Frontal_Sup_Medial_R Range value: Inf, ..., -Inf
\item \strong{FUSFMR02} \emph{numeric} Globally normalized CMRgl from Fusiform_R Range value: Inf, ..., -Inf
\item \strong{RECTUSL02} \emph{numeric} Globally normalized CMRgl from Rectus_L Range value: Inf, ..., -Inf
\item \strong{TMPPOSL02} \emph{numeric} Globally normalized CMRgl from Temporal_Pole_Sup_L Range value: Inf, ..., -Inf
\item \strong{PARISUPL02} \emph{numeric} Globally normalized CMRgl from Parietal_Sup_L Range value: Inf, ..., -Inf
\item \strong{PARISUPR02} \emph{numeric} Globally normalized CMRgl from Parietal_Sup_L Range value: Inf, ..., -Inf
\item \strong{SUPMRGL02} \emph{numeric} Globally normalized CMRgl from SupraMarginal_L Range value: Inf, ..., -Inf
\item \strong{SUPMRGR02} \emph{numeric} Globally normalized CMRgl from SupraMarginal_R Range value: Inf, ..., -Inf
\item \strong{TMPINFR02} \emph{numeric} Globally normalized CMRgl from Temporal_Inf_L Range value: Inf, ..., -Inf
\item \strong{FRTMIDL03} \emph{numeric} Frontal_Mid_L Range value: Inf, ..., -Inf
\item \strong{FRTMIDR03} \emph{numeric} Frontal_Mid_R Range value: Inf, ..., -Inf
\item \strong{FRTSUPL03} \emph{numeric} Frontal_Sup_L Range value: Inf, ..., -Inf
\item \strong{FRTSUPR03} \emph{numeric} Frontal_Sup_R Range value: Inf, ..., -Inf
\item \strong{FRTSMEDL03} \emph{numeric} Frontal_Sup_Medial_L Range value: Inf, ..., -Inf
\item \strong{FRTSMEDR03} \emph{numeric} Frontal_Sup_Medial_R Range value: Inf, ..., -Inf
\item \strong{TMPSUPL03} \emph{numeric} Temporal_Sup_L Range value: Inf, ..., -Inf
\item \strong{TMPSUPR03} \emph{numeric} Temporal_Sup_R Range value: 0.906624, ..., 1.63651
\item \strong{TMPMIDL03} \emph{numeric} Temporal_Mid_L Range value: 0.908312, ..., 1.45835
\item \strong{TMPMIDR03} \emph{numeric} Temporal_Mid_R Range value: 0.870199, ..., 1.58523
\item \strong{TMPINFL03} \emph{numeric} Temporal_Inf_L Range value: 0.868289, ..., 1.42746
\item \strong{TMPINFR03} \emph{numeric} Temporal_Inf_R Range value: 0.826913, ..., 1.52987
\item \strong{FUSFRML03} \emph{numeric} Fusiform_L Range value: Inf, ..., -Inf
\item \strong{FUSFRMR03} \emph{numeric} Fusiform_R Range value: Inf, ..., -Inf
\item \strong{LINGUALL03} \emph{numeric} Lingual_L Range value: Inf, ..., -Inf
\item \strong{LINGUALR03} \emph{numeric} Lingual_R Range value: Inf, ..., -Inf
\item \strong{PARAHIPL03} \emph{numeric} ParaHippocampal_L Range value: Inf, ..., -Inf
\item \strong{PARAHIPR03} \emph{numeric} ParaHippocampal_R Range value: Inf, ..., -Inf
\item \strong{PARISUPL03} \emph{numeric} Parietal_Sup_L Range value: Inf, ..., -Inf
\item \strong{PARISUPR03} \emph{numeric} Parietal_Sup_R Range value: Inf, ..., -Inf
\item \strong{PARIINFL03} \emph{numeric} Parietal_Inf_L Range value: 1.14395, ..., 1.82124
\item \strong{PARIINFR03} \emph{numeric} Parietal_Inf_R Range value: 1.11841, ..., 1.94766
\item \strong{PRECUNL03} \emph{numeric} Precuneus_L Range value: 1.30438, ..., 2.46674
\item \strong{PRECUNR03} \emph{numeric} Precuneus_R Range value: 1.5906, ..., 2.42236
\item \strong{CINGANT03} \emph{numeric} Cingulum_Ant Range value: 0.967814, ..., 1.73057
\item \strong{CINGMID03} \emph{numeric} Cingulum_Mid Range value: 1.16246, ..., 2.08253
\item \strong{CINGPST03} \emph{numeric} Cingulum_Post Range value: 0.832101, ..., 2.00208
\item \strong{ANGULR03} \emph{numeric} Angular_R Range value: 0.86501, ..., 1.59786
\item \strong{OCCMIDL03} \emph{numeric} Occipital_Mid_L Range value: Inf, ..., -Inf
\item \strong{OCCMIDR03} \emph{numeric} Occipital_Mid_R Range value: Inf, ..., -Inf
\item \strong{OCCSUPL03} \emph{numeric} Occipital_Sup_L Range value: Inf, ..., -Inf
\item \strong{OCCSUPR03} \emph{numeric} Occipital_Sup_R Range value: Inf, ..., -Inf
\item \strong{OCCINFL03} \emph{numeric} Occipital_Inf_L Range value: Inf, ..., -Inf
\item \strong{OCCINFR03} \emph{numeric} Occipital_Inf_R Range value: Inf, ..., -Inf
\item \strong{THALAML03} \emph{numeric} Thalamus_L Range value: Inf, ..., -Inf
\item \strong{THALAMR03} \emph{numeric} Thalamus_R Range value: Inf, ..., -Inf
\item \strong{FRTMEDL03} \emph{numeric} Frontal_Medial_Left Range value: Inf, ..., -Inf
\item \strong{FRTMEDR03} \emph{numeric} Frontal_Medial_Right Range value: Inf, ..., -Inf
\item \strong{SUPMRGL03} \emph{numeric} SupraMarginal_L Range value: 0.672312, ..., 1.30923
\item \strong{SUPMRGR03} \emph{numeric} SupraMarginal_R Range value: Inf, ..., -Inf
\item \strong{FRTMIDL04} \emph{numeric} Frontal_Mid_L Range value: Inf, ..., -Inf
\item \strong{FRTMIDR04} \emph{numeric} Frontal_Mid_R Range value: Inf, ..., -Inf
\item \strong{FRTSUPL04} \emph{numeric} Frontal_Sup_L Range value: Inf, ..., -Inf
\item \strong{FRTSUPR04} \emph{numeric} Frontal_Sup_R Range value: Inf, ..., -Inf
\item \strong{FRTSMEDL04} \emph{numeric} Frontal_Sup_Medial_L Range value: Inf, ..., -Inf
\item \strong{FRTSMEDR04} \emph{numeric} Frontal_Sup_Medial_R Range value: Inf, ..., -Inf
\item \strong{TMPSUPL04} \emph{numeric} Temporal_Sup_L Range value: Inf, ..., -Inf
\item \strong{TMPSUPR04} \emph{numeric} Temporal_Sup_R Range value: 1.03792, ..., 1.76477
\item \strong{TMPMIDL04} \emph{numeric} Temporal_Mid_L Range value: 0.835183, ..., 1.68137
\item \strong{TMPMIDR04} \emph{numeric} Temporal_Mid_R Range value: 0.891837, ..., 1.68217
\item \strong{TMPINFL04} \emph{numeric} Temporal_Inf_L Range value: 0.80388, ..., 1.84787
\item \strong{TMPINFR04} \emph{numeric} Temporal_Inf_R Range value: 0.927987, ..., 1.68568
\item \strong{FUSFRML04} \emph{numeric} Fusiform_L Range value: Inf, ..., -Inf
\item \strong{FUSFRMR04} \emph{numeric} Fusiform_R Range value: Inf, ..., -Inf
\item \strong{LINGUALL04} \emph{numeric} Lingual_L Range value: Inf, ..., -Inf
\item \strong{LINGUALR04} \emph{numeric} Lingual_R Range value: Inf, ..., -Inf
\item \strong{PARAHIPL04} \emph{numeric} ParaHippocampal_L Range value: Inf, ..., -Inf
\item \strong{PARAHIPR04} \emph{numeric} ParaHippocampal_R Range value: Inf, ..., -Inf
\item \strong{PARISUPL04} \emph{numeric} Parietal_Sup_L Range value: Inf, ..., -Inf
\item \strong{PARISUPR04} \emph{numeric} Parietal_Sup_R Range value: Inf, ..., -Inf
\item \strong{PARIINFL04} \emph{numeric} Parietal_Inf_L Range value: Inf, ..., -Inf
\item \strong{PARIINFR04} \emph{numeric} Parietal_Inf_R Range value: Inf, ..., -Inf
\item \strong{PRECUNL04} \emph{numeric} Precuneus_L Range value: Inf, ..., -Inf
\item \strong{PRECUNR04} \emph{numeric} Precuneus_R Range value: Inf, ..., -Inf
\item \strong{CINGANT04} \emph{numeric} Cingulum_Ant Range value: Inf, ..., -Inf
\item \strong{CINGMID04} \emph{numeric} Cingulum_Mid Range value: Inf, ..., -Inf
\item \strong{CINGPST04} \emph{numeric} Cingulum_Post Range value: 1.06925, ..., 2.12651
\item \strong{OCCMIDL04} \emph{numeric} Occipital_Mid_L Range value: Inf, ..., -Inf
\item \strong{OCCMIDR04} \emph{numeric} Occipital_Mid_R Range value: 0.897243, ..., 1.78806
\item \strong{OCCSUPL04} \emph{numeric} Occipital_Sup_L Range value: Inf, ..., -Inf
\item \strong{OCCSUPR04} \emph{numeric} Occipital_Sup_R Range value: Inf, ..., -Inf
\item \strong{OCCINFL04} \emph{numeric} Occipital_Inf_L Range value: Inf, ..., -Inf
\item \strong{OCCINFR04} \emph{numeric} Occipital_Inf_R Range value: Inf, ..., -Inf
\item \strong{THALAML04} \emph{numeric} Thalamus_L Range value: Inf, ..., -Inf
\item \strong{THALAMR04} \emph{numeric} Thalamus_R Range value: Inf, ..., -Inf
\item \strong{FRTMEDL04} \emph{numeric} Frontal_Medial_Left Range value: Inf, ..., -Inf
\item \strong{FRTMEDR04} \emph{numeric} Frontal_Medial_Right Range value: Inf, ..., -Inf
\item \strong{SUPMRGL04} \emph{numeric} SupraMarginal_L Range value: 1.12521, ..., 1.92945
\item \strong{SUPMRGR04} \emph{numeric} SupraMarginal_R Range value: Inf, ..., -Inf
\item \strong{ANGULL04} \emph{numeric} Angular_L Range value: Inf, ..., -Inf
\item \strong{ANGULR04} \emph{numeric} Angular_R Range value: 0.843005, ..., 1.73865
\item \strong{FRTMIDL05} \emph{numeric} Frontal_Mid_L Range value: Inf, ..., -Inf
\item \strong{FRTMIDR05} \emph{numeric} Frontal_Mid_R Range value: Inf, ..., -Inf
\item \strong{FRTSUPL05} \emph{numeric} Frontal_Sup_L Range value: Inf, ..., -Inf
\item \strong{FRTSUPR05} \emph{numeric} Frontal_Sup_R Range value: Inf, ..., -Inf
\item \strong{FRTSMEDL05} \emph{numeric} Frontal_Sup_Medial_L Range value: Inf, ..., -Inf
\item \strong{FRTSMEDR05} \emph{numeric} Frontal_Sup_Medial_R Range value: Inf, ..., -Inf
\item \strong{TMPSUPL05} \emph{numeric} Temporal_Sup_L Range value: Inf, ..., -Inf
\item \strong{TMPSUPR05} \emph{numeric} Temporal_Sup_R Range value: 0.87772, ..., 1.34958
\item \strong{TMPMIDL05} \emph{numeric} Temporal_Mid_L Range value: 1.23457, ..., 1.96356
\item \strong{TMPMIDR05} \emph{numeric} Temporal_Mid_R Range value: 1.03621, ..., 1.64083
\item \strong{TMPINFL05} \emph{numeric} Temporal_Inf_L Range value: Inf, ..., -Inf
\item \strong{TMPINFR05} \emph{numeric} Temporal_Inf_R Range value: 1.03917, ..., 1.76632
\item \strong{FUSFRML05} \emph{numeric} Fusiform_L Range value: Inf, ..., -Inf
\item \strong{FUSFRMR05} \emph{numeric} Fusiform_R Range value: Inf, ..., -Inf
\item \strong{PARAHIPL05} \emph{numeric} ParaHippocampal_L Range value: Inf, ..., -Inf
\item \strong{PARAHIPR05} \emph{numeric} ParaHippocampal_R Range value: Inf, ..., -Inf
\item \strong{PARISUPL05} \emph{numeric} Parietal_Sup_L Range value: Inf, ..., -Inf
\item \strong{PARISUPR05} \emph{numeric} Parietal_Sup_R Range value: Inf, ..., -Inf
\item \strong{PARIINFL05} \emph{numeric} Parietal_Inf_L Range value: 0.859358, ..., 1.60546
\item \strong{PARIINFR05} \emph{numeric} Parietal_Inf_R Range value: 0.884295, ..., 1.61886
\item \strong{PRECUNL05} \emph{numeric} Precuneus_L Range value: Inf, ..., -Inf
\item \strong{PRECUNR05} \emph{numeric} Precuneus_R Range value: 1.13373, ..., 2.44158
\item \strong{CINGANT05} \emph{numeric} Cingulum_Ant Range value: Inf, ..., -Inf
\item \strong{CINGMID05} \emph{numeric} Cingulum_Mid Range value: Inf, ..., -Inf
\item \strong{CINGPST05} \emph{numeric} Cingulum_Post Range value: 1.05241, ..., 2.07561
\item \strong{OCCMIDL05} \emph{numeric} Occipital_Mid_L Range value: Inf, ..., -Inf
\item \strong{OCCMIDR05} \emph{numeric} Occipital_Mid_R Range value: Inf, ..., -Inf
\item \strong{OCCSUPL05} \emph{numeric} Occipital_Sup_L Range value: Inf, ..., -Inf
\item \strong{OCCSUPR05} \emph{numeric} Occipital_Sup_R Range value: Inf, ..., -Inf
\item \strong{OCCINFL05} \emph{numeric} Occipital_Inf_L Range value: Inf, ..., -Inf
\item \strong{OCCINFR05} \emph{numeric} Occipital_Inf_R Range value: Inf, ..., -Inf
\item \strong{THALAML05} \emph{numeric} Thalamus_L Range value: Inf, ..., -Inf
\item \strong{THALAMR05} \emph{numeric} Thalamus_R Range value: Inf, ..., -Inf
\item \strong{FRTMEDL05} \emph{numeric} Frontal_Medial_Left Range value: Inf, ..., -Inf
\item \strong{FRTMEDR05} \emph{numeric} Frontal_Medial_Right Range value: Inf, ..., -Inf
\item \strong{ANGULR05} \emph{numeric} Angular_Right Range value: 1.04344, ..., 1.56019
\item \strong{ANGULL05} \emph{numeric} Angular_Left Range value: Inf, ..., -Inf
\item \strong{CALCARL05} \emph{numeric} Calcarine_L Range value: 0.819858, ..., 1.8764
\item \strong{CALCARR05} \emph{numeric} Calcarine_R Range value: Inf, ..., -Inf
\item \strong{LINGUALL05} \emph{numeric} Lingual_L Range value: 0.807419, ..., 1.81036
\item \strong{LINGUALR05} \emph{numeric} Lingual_R Range value: Inf, ..., -Inf
\item \strong{CEREB8L05} \emph{numeric} Cerebelum_8_L Range value: 1.02268, ..., 1.69108
\item \strong{CEREB8R05} \emph{numeric} Cerebelum_8_R Range value: Inf, ..., -Inf
\item \strong{CERBCR2L05} \emph{numeric} Cerebelum_Crus2_L Range value: 1.02973, ..., 1.68651
\item \strong{CERBCR2R05} \emph{numeric} Cerebelum_Crus2_R Range value: Inf, ..., -Inf
\item \strong{SUPMRGL05} \emph{numeric} SupraMarginal_L Range value: 0.862619, ..., 1.29025
\item \strong{SUPMRGR05} \emph{numeric} SupraMarginal_R Range value: Inf, ..., -Inf
\item \strong{FRTMIDL06} \emph{numeric} Frontal_Mid_L Range value: Inf, ..., -Inf
\item \strong{FRTMIDR06} \emph{numeric} Frontal_Mid_R Range value: Inf, ..., -Inf
\item \strong{FRTSUPL06} \emph{numeric} Frontal_Sup_L Range value: Inf, ..., -Inf
\item \strong{FRTSUPR06} \emph{numeric} Frontal_Sup_R Range value: Inf, ..., -Inf
\item \strong{FRTSMEDL06} \emph{numeric} Frontal_Sup_Medial_L Range value: Inf, ..., -Inf
\item \strong{FRTSMEDR06} \emph{numeric} Frontal_Sup_Medial_R Range value: Inf, ..., -Inf
\item \strong{TMPSUPL06} \emph{numeric} Temporal_Sup_L Range value: Inf, ..., -Inf
\item \strong{TMPSUPR06} \emph{numeric} Temporal_Sup_R Range value: 1.18306, ..., 1.75266
\item \strong{TMPMIDL06} \emph{numeric} Temporal_Mid_L Range value: 0.969484, ..., 1.60007
\item \strong{TMPMIDR06} \emph{numeric} Temporal_Mid_R Range value: 1.1327, ..., 1.65291
\item \strong{TMPINFL06} \emph{numeric} Temporal_Inf_L Range value: 1.00196, ..., 1.86324
\item \strong{TMPINFR06} \emph{numeric} Temporal_Inf_R Range value: 1.19813, ..., 1.97752
\item \strong{FUSFRML06} \emph{numeric} Fusiform_L Range value: 1.07881, ..., 1.71197
\item \strong{FUSFRMR06} \emph{numeric} Fusiform_R Range value: 0.88559, ..., 1.60234
\item \strong{LINGUALL06} \emph{numeric} Lingual_L Range value: Inf, ..., -Inf
\item \strong{LINGUALR06} \emph{numeric} Lingual_R Range value: Inf, ..., -Inf
\item \strong{PARAHIPL06} \emph{numeric} ParaHippocampal_L Range value: Inf, ..., -Inf
\item \strong{PARAHIPR06} \emph{numeric} ParaHippocampal_R Range value: Inf, ..., -Inf
\item \strong{PARISUPL06} \emph{numeric} Parietal_Sup_L Range value: Inf, ..., -Inf
\item \strong{PARISUPR06} \emph{numeric} Parietal_Sup_R Range value: Inf, ..., -Inf
\item \strong{PARIINFL06} \emph{numeric} Parietal_Inf_L Range value: Inf, ..., -Inf
\item \strong{PARIINFR06} \emph{numeric} Parietal_Inf_R Range value: Inf, ..., -Inf
\item \strong{PRECUNL06} \emph{numeric} Precuneus_L Range value: Inf, ..., -Inf
\item \strong{PRECUNR06} \emph{numeric} Precuneus_R Range value: Inf, ..., -Inf
\item \strong{CINGANT06} \emph{numeric} Cingulum_Ant Range value: Inf, ..., -Inf
\item \strong{CINGMID06} \emph{numeric} Cingulum_Mid Range value: Inf, ..., -Inf
\item \strong{CINGPST06} \emph{numeric} Cingulum_Post Range value: Inf, ..., -Inf
\item \strong{OCCMIDL06} \emph{numeric} Occipital_Mid_L Range value: Inf, ..., -Inf
\item \strong{OCCMIDR06} \emph{numeric} Occipital_Mid_R Range value: 0.873268, ..., 1.60524
\item \strong{OCCSUPL06} \emph{numeric} Occipital_Sup_L Range value: Inf, ..., -Inf
\item \strong{OCCSUPR06} \emph{numeric} Occipital_Sup_R Range value: Inf, ..., -Inf
\item \strong{OCCINFL06} \emph{numeric} Occipital_Inf_L Range value: Inf, ..., -Inf
\item \strong{OCCINFR06} \emph{numeric} Occipital_Inf_R Range value: Inf, ..., -Inf
\item \strong{THALAML06} \emph{numeric} Thalamus_L Range value: Inf, ..., -Inf
\item \strong{THALAMR06} \emph{numeric} Thalamus_R Range value: Inf, ..., -Inf
\item \strong{FRTMEDL06} \emph{numeric} Frontal_Medial_Left Range value: Inf, ..., -Inf
\item \strong{FRTMEDR06} \emph{numeric} Frontal_Medial_Right Range value: Inf, ..., -Inf
\item \strong{HIPPL06} \emph{numeric} Hippocampus_L Range value: Inf, ..., -Inf
\item \strong{HIPPR06} \emph{numeric} Hippocampus_R Range value: 1.01451, ..., 1.59095
\item \strong{INDGLBNDX} \emph{numeric} Individualized Global Index Range value: Inf, ..., -Inf
\item \strong{INDTEPNDX} \emph{numeric} Individualized Temporal index Range value: Inf, ..., -Inf
\item \strong{PRECUNL07} \emph{numeric} Precuneus_L Range value: 1.15795, ..., 2.14675
\item \strong{CINGPST07} \emph{numeric} Cingulum_Post Range value: 0.912872, ..., 1.86358
\item \strong{PARIINFR07} \emph{numeric} Parietal_Inf_R Range value: 0.86859, ..., 1.58379
\item \strong{TMPMIDL07} \emph{numeric} Temporal_Mid_L Range value: 0.798215, ..., 1.408
\item \strong{CUNEUSL08} \emph{numeric} Cuneus_L Range value: 1.17065, ..., 2.03573
\item \strong{TMPMIDR08} \emph{numeric} Temporal_Mid_R Range value: 0.958192, ..., 1.76537
\item \strong{TMPINFL09} \emph{numeric} Temporal_Inf_L Range value: 0.92968, ..., 1.66584
\item \strong{CINGPST09} \emph{numeric} Cingulum_Post Range value: 1.03065, ..., 2.02593
\item \strong{PRECUNL09} \emph{numeric} Precuneus_L Range value: 1.54287, ..., 2.68315
\item \strong{PARIINFR09} \emph{numeric} Parietal_Inf_R Range value: 1.0283, ..., 1.90475
\item \strong{TMPINFL10} \emph{numeric} Temporal_Inf_L Range value: 0.908405, ..., 1.65062
\item \strong{PRECUNL10} \emph{numeric} Precuneus_L Range value: 1.31553, ..., 2.24931
\item \strong{TMPINFL11} \emph{numeric} Temporal_Inf_L Range value: 1.1063, ..., 1.55387
\item \strong{OCCMIDL11} \emph{numeric} Occipital_Mid_L Range value: 0.973173, ..., 1.55247
\item \strong{CINGPST11} \emph{numeric} Cingulum_Post_L Range value: 0.857508, ..., 1.87147
\item \strong{CINGMID11} \emph{numeric} Cingulum_Post_R/Cingulum_Mid_R Range value: 1.03255, ..., 1.87679
\item \strong{TMPPOSR12} \emph{numeric} Temporal_Pole_Sup_R Range value: 0.45133, ..., 0.936241
\item \strong{CINGPSTR12} \emph{numeric} Cingulum_Post_R/Cingulum_Mid_R Range value: 0.424733, ..., 0.949642
\item \strong{FROITRN001} \emph{numeric} Training group, all subjects Range value: 0.74025, ..., 1.24215
\item \strong{FROITRN002} \emph{numeric} Training group, exclude HRRT and BioGraph HiRez Range value: 0.796567, ..., 1.23335
\item \strong{FROITST001} \emph{numeric} Testing group, all scans Range value: 0.636083, ..., 1.14834
\item \strong{FROITST002} \emph{numeric} Testing group, exclude HRRT and BioGraph HiRez Range value: 0.813588, ..., 1.10135
\item \strong{FROIWHL001} \emph{numeric} All subjects Range value: 0.700387, ..., 1.80342
\item \strong{FROIWHL002} \emph{numeric} Exclude HRRT and BioGraph HiRez Range value: 0.751776, ..., 1.63939
\item \strong{FRPMSK001} \emph{numeric} Exclude HRRT and BioGraph HiRez Range value: 0.788669, ..., 1.40005
\item \strong{FRPMSK002} \emph{numeric} All subjects Range value: 0.788669, ..., 1.40442
\item \strong{FRTMSK003} \emph{numeric} Exclude HRRT and BioGraph HiRez Range value: 0.848437, ..., 1.46291
\item \strong{FRTMSK004} \emph{numeric} All subjects Range value: 0.816065, ..., 1.51455
\item \strong{FRTMSK005} \emph{numeric} All subjects Range value: 0.781659, ..., 1.30399
\item \strong{FRTMSK006} \emph{numeric} All subjects Range value: 0.626235, ..., 1.11443
\item \strong{FRTMSK007} \emph{numeric} All subjects Range value: 0.74963, ..., 1.43373
\item \strong{FRTMSK008} \emph{numeric} Exclude HRRT and BioGraph HiRez Range value: 0.715832, ..., 1.39802
\item \strong{FRTMSK009} \emph{numeric} All subjects Range value: 0.597849, ..., 1.17415
\item \strong{FRTMSK010} \emph{numeric} Exclude HRRT and BioGraph HiRez Range value: 0.558626, ..., 1.13881
\item \strong{FRTMSK011} \emph{numeric} All subjects Range value: 0.868071, ..., 1.43818
\item \strong{HCISUB1} \emph{numeric} hypometabolic convergence index for 74 AD subjects Range value: 4.03, ..., 28.16
\item \strong{HCI99} \emph{numeric} hypometabolic convergence index for normal controls, stable MCI, MCI converter and ADs Range value: 2.59, ..., 24.09
\item \strong{HCI} \emph{numeric} hypometabolic convergence index for ADNI1/ADNI2/ADNIGO subjects Range value: 0.314205, ..., 34.6
\item \strong{NL1HIPP} \emph{numeric} difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from Hippocampus for 58 AD patients Range value: -0.149295, ..., 0.201397
\item \strong{NL1PTEMP} \emph{numeric} difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from the parietotemporal region for 58 AD patients Range value: -0.148037, ..., 0.24458
\item \strong{NL1PSTCG} \emph{numeric} difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from the posterior cingulate for for 58 AD patients Range value: -0.172309, ..., 0.339278
\item \strong{NL2PSTCG} \emph{numeric} difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from Posterior Cingulate in 129 MCIs. Range value: -0.208048, ..., 0.243036
\item \strong{NL2PARI} \emph{numeric} difference of globally normalized CMRgl from 12 months to 24 months from the inferior parietal region in 129 MCI subjects Range value: -0.118032, ..., 0.438594
\item \strong{NL3FRONT} \emph{numeric} difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from superior frontal region of 74 normal controls. Range value: -0.169811, ..., 0.228805
\item \strong{DECLADFR} \emph{numeric} difference of globally normalized CMRgl from Baseline to 12 months from the superior frontal region of 74 AD patients Range value: -0.132305, ..., 0.216164
\item \strong{DECLADTMP} \emph{numeric} CMRgl from Baseline from the inferior temporal region of 74 AD patients Range value: 0.905927, ..., 1.77742
\item \strong{LONGHCI} \emph{numeric} Extension of HCI to longitudinal data Range value: Inf, ..., -Inf
\item \strong{BSLVBM} \emph{numeric} VBManalysis of structural MRI data Range value: Inf, ..., -Inf
\item \strong{ACI} \emph{numeric} amyloid convergence index for AV-45 subjects Range value: 132.86, ..., 39557.4
\item \strong{MCSUVR} \emph{numeric} mean cortical standard uptake value ratio Range value: 0.5793, ..., 1.4485
\item \strong{FNMPREC} \emph{numeric} precuneus FDG measure Range value: 0.944265, ..., 1.50352
\item \strong{FNAPREC} \emph{numeric} precuneus FDG measure Range value: 1.68434, ..., 2.64043
\item \strong{FNAPC} \emph{numeric} posterior cingulate FDG measure Range value: 1.49469, ..., 2.42776
\item \strong{FEAPREC} \emph{numeric} precuneus FDG measure Range value: 1.72042, ..., 2.91048
\item \strong{FMAPREC} \emph{numeric} precuneus FDG measure Range value: 1.76315, ..., 2.68583
\item \strong{FMAPC} \emph{numeric} posterior cingulate FDG measure Range value: 1.76229, ..., 2.71944
\item \strong{AVNEFRONT} \emph{numeric} frontal AV-45 measure Range value: 0.797, ..., 1.7925
\item \strong{AVNMFRONT} \emph{numeric} frontal AV-45 measure Range value: 0.6615, ..., 1.6667
\item \strong{AVNAFRONT} \emph{numeric} frontal AV-45 measure Range value: 0.669215, ..., 1.57765
\item \strong{AVEMTEMP} \emph{numeric} temporal AV-45 measure Range value: 0.406411, ..., 1.20786
\item \strong{AVEATEMP} \emph{numeric} temporal AV-45 measure Range value: 0.564236, ..., 1.82057
\item \strong{AVMAOCCIP} \emph{numeric} occipital AV-45 measure Range value: 0.736488, ..., 1.73013
\item \strong{FEAPC2} \emph{numeric} posterior cingulate FDG measure Range value: 1.27542, ..., 2.5255
\item \strong{FEMPC2} \emph{numeric} posterior cingulate FDG measure Range value: 1.37597, ..., 2.34252
\item \strong{FMAPREC2} \emph{numeric} precuneus FDG measure Range value: 1.13538, ..., 2.40924
\item \strong{FNAPARI2} \emph{numeric} parietal FDG measure Range value: 0.910594, ..., 2.16165
\item \strong{FNEHC2} \emph{numeric} hippocampal FDG measure Range value: 0.899809, ..., 1.3592
\item \strong{FNMTEMP2} \emph{numeric} temporal FDG measure Range value: 1.06895, ..., 1.83971
\item \strong{AVEATEMP2} \emph{numeric} temporal AV-45 measure Range value: 0.542087, ..., 1.88432
\item \strong{AVEMTEMP2} \emph{numeric} temporal AV-45 measure Range value: 0.844229, ..., 2.01578
\item \strong{AVMAFRNT2} \emph{numeric} frontal AV-45 measure Range value: 0.69846, ..., 1.94358
\item \strong{AVNATEMP2} \emph{numeric} temporal AV-45 measure Range value: 0.562108, ..., 1.92533
\item \strong{AVNEFRNT2} \emph{numeric} frontal AV-45 measure Range value: 0.617324, ..., 1.6891
\item \strong{AVNMPREC2} \emph{numeric} precuneus AV-45 measure Range value: 0.612387, ..., 55.2062
\item \strong{MMPLS} \emph{numeric} AVfdgPLS Range value: -3.28788, ..., 2.11634
\item \strong{ICA} \emph{numeric} AV_ICA Range value: -2.5446, ..., 2.9716
\item \strong{PCA2} \emph{numeric} AV_SSM Range value: -4.60124, ..., 5.44689
\item \strong{PLSFDG01} \emph{numeric} PLS for FDG PET Range value: -28.021, ..., 39.222
\item \strong{HCI001} \emph{numeric} Hypometabolic Convergence Index Range value: 0.314205, ..., 29.6843
\item \strong{MCSUVRWM} \emph{numeric} Mean Cortical SUVR Corpus Callosum and Centrum Semiovale Combined Range value: 0.4857, ..., 1.6982
\item \strong{MCSUVRWM2} \emph{numeric} Mean Cortical SUVR Corpus Callosum and Centrum Semiovale Combined Range value: 0.48568, ..., 1.6982
\item \strong{HCI_2014} \emph{numeric} Hypometabolic Convergence Index Range value: 2.0805, ..., 55.2062
\item \strong{SROI} \emph{numeric} statistical region of interest (sROI) developed and validated for optimal longitudinal FDG-PET CMRgl changes Range value: 0.8591, ..., 1.3888
\item \strong{MCSUVRCERE} \emph{numeric} Mean Cortical SUVR with Cerebellum as reference region Range value: 0.765881, ..., 2.42487
\item \strong{MCSUVRWM3} \emph{numeric} Mean Cortical SUVR with Corpus Callosum and Centrum Semiovale Combined as reference region Range value: 0.438463, ..., 1.96134
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
