% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAIPETNMRCAV45}
\alias{BAIPETNMRCAV45}
\title{BAI -  PET NMRC AV45 Summaries: 2020-10-23}
\format{
A data frame with 2815 observations and 13 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BAIPETNMRCAV45)
}
\description{
BAI -  PET NMRC AV45 Summaries: 2020-10-23 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character} Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI2, ADNIGO, ADNI3
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 6837
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{LONIUID} \emph{numeric} LONI Unique Image Identifier Range value: 177679, ..., 1280777
\item \strong{RUNDATE} \emph{ } DATE analysis performed Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Complete/Incomplete , with options: Complete
\item \strong{MODALITY} \emph{character} The type of pet scan used for analysis , with options: AV45
\item \strong{MCSUVRWM} \emph{numeric} Mean Cortical SUVR with Corpus Callosum and Centrum Semiovale Combined as reference region Range value: 0.438463, ..., 1.96134
\item \strong{MCSUVRCERE} \emph{numeric} Mean Cortical SUVR with Cerebellum as reference region Range value: 0.76563, ..., 2.42487
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{imaging}
