% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAIPETNMRCFDG}
\alias{BAIPETNMRCFDG}
\title{BAI -  PET NMRC FDG Summaries: 2020-12-11}
\format{
A data frame with 3684 observations and 14 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BAIPETNMRCFDG)
}
\description{
BAI -  PET NMRC FDG Summaries: 2020-12-11 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character} Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 6840
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{SCANDATE} \emph{character}    
\item \strong{LONIUID} \emph{numeric} LONI Unique Image Identifier Range value: 25392, ..., 1281394
\item \strong{RUNDATE} \emph{character} DATE analysis performed , with options: 06/25/2014, 10/16/2019, 05/17/2016, 11/19/2020
\item \strong{STATUS} \emph{character} Complete/Incomplete , with options: Complete
\item \strong{MODALITY} \emph{character} The type of pet scan used for analysis , with options: FDG
\item \strong{HCI} \emph{numeric} Hypometabolic Convergence Index Range value: 2.0805, ..., 55.2062
\item \strong{SROI.AD} \emph{numeric}   Range value: 0.8591, ..., 1.3888
\item \strong{SROI.MCI} \emph{numeric}   Range value: 0.7861, ..., 1.3525
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
