% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAIPETNMRCFTP}
\alias{BAIPETNMRCFTP}
\title{BAI -  PET NMRC Flortaucipir (F-AV1451) Summaries: 2022-08-17}
\format{
A data frame with 1157 observations and 14 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BAIPETNMRCFTP)
}
\description{
BAI -  PET NMRC Flortaucipir (F-AV1451) Summaries: 2022-08-17 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character} Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI3, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 6901
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{LONIUID} \emph{numeric} LONI Unique Image Identifier Range value: 513655, ..., 1481670
\item \strong{RUNDATE} \emph{ } DATE analysis performed Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Complete/Incomplete , with options: Complete
\item \strong{MODALITY} \emph{character} The type of pet scan used for analysis , with options: Flortaucipir
\item \strong{ENTORHINAL_SUVR} \emph{numeric} Entorhinal SUVR with Cerebellum Crus I as reference region Range value: 0.7951, ..., 2.4264
\item \strong{INFERIOR_TEMPORAL_SUVR} \emph{numeric} Inferior temporal SUVR with Cerebellum Crus I as reference region Range value: 0.9373, ..., 2.6767
\item \strong{TAU_METAROI} \emph{numeric} Tau metaROI with Cerebellum Crus as reference region Range value: 0.902, ..., 2.8179
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{imaging}
