% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batemanlab_20190621}
\alias{batemanlab_20190621}
\title{BATEMANLAB_20190621}
\format{
A data frame with 622 observations and 74 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(batemanlab_20190621)
}
\description{
BATEMANLAB_20190621 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNIGO, ADNI2, ADNI4
\item \strong{RID} \emph{numeric}   Range value: 2010, ..., 999999
\item \strong{VISCODE} \emph{character}   , with options: bl, v03, nv
\item \strong{VISCODE2} \emph{character}   , with options: bl, nv
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{VID} \emph{numeric}   Range value: 2, ..., 2
\item \strong{global_spec_id} \emph{character}    
\item \strong{SUBJECT} \emph{character}    
\item \strong{Box_Number} \emph{numeric}   Range value: 1, ..., 3
\item \strong{Position} \emph{numeric}   Range value: 2, ..., 100
\item \strong{ASSAY} \emph{character}    
\item \strong{ASSAY_FLUID} \emph{character}   , with options: Plasma
\item \strong{ASSAY_IP_PROTOCOL} \emph{character}   , with options: IP: 0.5mL KF
\item \strong{BATCH_N} \emph{character}   , with options: Batch #01, Batch #00 Trial, Batch #02, Batch #03, Batch #04, Batch #05, Batch #04 Extra, Batch #03 Extra
\item \strong{FILE_NAME_NE} \emph{character}    
\item \strong{INJECTION} \emph{character}   , with options: a, b, c
\item \strong{INSTRUMENT} \emph{character}   , with options: Lumos, Altis
\item \strong{LC_MS_SAMPLE_ID} \emph{character}    
\item \strong{MS_RUN_DATE} \emph{ }   Date: YYY-MM-DD
\item \strong{QC_FLAGS} \emph{character}    
\item \strong{QC_STATUS} \emph{character}   , with options: Passed, FAIL
\item \strong{SAMPLE_TYPE} \emph{character}   , with options: Clinical Plasma, QC Calibrator, External calibrator
\item \strong{PEAK_AREA_ABETA38_N14_B07} \emph{numeric}   Range value: 594, ..., 9997016
\item \strong{PEAK_AREA_ABETA38_N14_B08} \emph{numeric}   Range value: 242, ..., 28009306
\item \strong{PEAK_AREA_ABETA38_N14_B09} \emph{numeric}   Range value: 279, ..., 12765293
\item \strong{PEAK_AREA_ABETA38_N15_B07} \emph{numeric}   Range value: 70, ..., 2353988
\item \strong{PEAK_AREA_ABETA38_N15_B08} \emph{numeric}   Range value: 609, ..., 7357057
\item \strong{PEAK_AREA_ABETA38_N15_B09} \emph{numeric}   Range value: 580, ..., 3232914
\item \strong{PEAK_AREA_ABETA40_N14_B05} \emph{numeric}   Range value: 1326, ..., 3065134
\item \strong{PEAK_AREA_ABETA40_N14_B07} \emph{numeric}   Range value: 67, ..., 3572894
\item \strong{PEAK_AREA_ABETA40_N14_B08} \emph{numeric}   Range value: 314, ..., 15211655
\item \strong{PEAK_AREA_ABETA40_N14_B09} \emph{numeric}   Range value: 596, ..., 16807608
\item \strong{PEAK_AREA_ABETA40_N14_B10} \emph{numeric}   Range value: 93, ..., 7719440
\item \strong{PEAK_AREA_ABETA40_N14_B11} \emph{numeric}   Range value: 726, ..., 78428504
\item \strong{PEAK_AREA_ABETA40_N15_B05} \emph{numeric}   Range value: 1570, ..., 1437812
\item \strong{PEAK_AREA_ABETA40_N15_B07} \emph{numeric}   Range value: 754, ..., 1595695
\item \strong{PEAK_AREA_ABETA40_N15_B08} \emph{numeric}   Range value: 1785, ..., 5149927
\item \strong{PEAK_AREA_ABETA40_N15_B09} \emph{numeric}   Range value: 783, ..., 5971145
\item \strong{PEAK_AREA_ABETA40_N15_B10} \emph{numeric}   Range value: 641, ..., 2669460
\item \strong{PEAK_AREA_ABETA40_N15_B11} \emph{numeric}   Range value: 97, ..., 26842686
\item \strong{PEAK_AREA_ABETA42_N14_B08} \emph{numeric}   Range value: 153, ..., 189311
\item \strong{PEAK_AREA_ABETA42_N14_B09} \emph{numeric}   Range value: 132, ..., 144604
\item \strong{PEAK_AREA_ABETA42_N14_B10} \emph{numeric}   Range value: 14, ..., 48526
\item \strong{PEAK_AREA_ABETA42_N14_B11} \emph{numeric}   Range value: 95, ..., 2089320
\item \strong{PEAK_AREA_ABETA42_N14_B12} \emph{numeric}   Range value: 224, ..., 4360189
\item \strong{PEAK_AREA_ABETA42_N14_B13} \emph{numeric}   Range value: 77, ..., 4863396
\item \strong{PEAK_AREA_ABETA42_N15_B08} \emph{numeric}   Range value: 217, ..., 67454
\item \strong{PEAK_AREA_ABETA42_N15_B09} \emph{numeric}   Range value: 58, ..., 85718
\item \strong{PEAK_AREA_ABETA42_N15_B10} \emph{numeric}   Range value: 67, ..., 64685
\item \strong{PEAK_AREA_ABETA42_N15_B11} \emph{numeric}   Range value: 14, ..., 329991
\item \strong{PEAK_AREA_ABETA42_N15_B12} \emph{numeric}   Range value: 87, ..., 1055875
\item \strong{PEAK_AREA_ABETA42_N15_B13} \emph{numeric}   Range value: 228, ..., 1308241
\item \strong{PEAK_AREA_ABETAMD_N14_B04} \emph{numeric}   Range value: 1692, ..., 5979009
\item \strong{PEAK_AREA_ABETAMD_N14_B05} \emph{numeric}   Range value: 7249, ..., 6730840
\item \strong{PEAK_AREA_ABETAMD_N14_B06} \emph{numeric}   Range value: 270, ..., 5810
\item \strong{PEAK_AREA_ABETAMD_N14_B07} \emph{numeric}   Range value: 120, ..., 3897928
\item \strong{PEAK_AREA_ABETAMD_N14_B08} \emph{numeric}   Range value: 283, ..., 36622056
\item \strong{PEAK_AREA_ABETAMD_N14_B09} \emph{numeric}   Range value: 389, ..., 65911060
\item \strong{PEAK_AREA_ABETAMD_N14_B10} \emph{numeric}   Range value: 287, ..., 38483080
\item \strong{PEAK_AREA_ABETAMD_N14_B11} \emph{numeric}   Range value: 242, ..., 6843525
\item \strong{PEAK_AREA_ABETAMD_N15_B04} \emph{numeric}   Range value: 4746, ..., 2108805
\item \strong{PEAK_AREA_ABETAMD_N15_B05} \emph{numeric}   Range value: 8492, ..., 2662645
\item \strong{PEAK_AREA_ABETAMD_N15_B06} \emph{numeric}   Range value: 81, ..., 2198
\item \strong{PEAK_AREA_ABETAMD_N15_B07} \emph{numeric}   Range value: 17, ..., 1400012
\item \strong{PEAK_AREA_ABETAMD_N15_B08} \emph{numeric}   Range value: 591, ..., 12295365
\item \strong{PEAK_AREA_ABETAMD_N15_B09} \emph{numeric}   Range value: 251, ..., 18503856
\item \strong{PEAK_AREA_ABETAMD_N15_B10} \emph{numeric}   Range value: 424, ..., 11493864
\item \strong{PEAK_AREA_ABETAMD_N15_B11} \emph{numeric}   Range value: 30, ..., 1873868
\item \strong{PEAK_AREA_ABETA40_N14_TOUSE} \emph{numeric}   Range value: 726, ..., 118167207
\item \strong{PEAK_AREA_ABETA40_N15_TOUSE} \emph{numeric}   Range value: 2056, ..., 40321772
\item \strong{PEAK_AREA_ABETA42_N14_TOUSE} \emph{numeric}   Range value: 0, ..., 9223585
\item \strong{PEAK_AREA_ABETA42_N15_TOUSE} \emph{numeric}   Range value: 0, ..., 2348416
\item \strong{RATIO_ABETA42_40_BY_ISTD_TOUSE} \emph{numeric}   Range value: 0, ..., 2.45081
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
