% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batemanlab_20221118}
\alias{batemanlab_20221118}
\title{BATEMANLAB_20221118}
\format{
A data frame with 742 observations and 36 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(batemanlab_20221118)
}
\description{
BATEMANLAB_20221118 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{RID} \emph{numeric}   Range value: 31, ..., 5296
\item \strong{VISCODE} \emph{character}   , with options: v11, v03, v21, v41, init, nv, v06, v31, v51
\item \strong{VISCODE2} \emph{character}    
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{PROT} \emph{character}   , with options: ADNI1, ADNI2, ADNIGO
\item \strong{Sample_ID} \emph{character}    
\item \strong{File_Name_NE} \emph{character}    
\item \strong{Sample_Type} \emph{character}   , with options: Clinical Plasma, Calibration standard
\item \strong{Assay_fluid} \emph{character}   , with options: Plasma
\item \strong{LC_MS_Sample_ID} \emph{character}    
\item \strong{MS_Run_Date} \emph{character}    
\item \strong{Injection} \emph{character}   , with options: a, b, c
\item \strong{Instrument} \emph{character}   , with options: OTF Lumos
\item \strong{Batch_N} \emph{character}   , with options: ADNI-02, ADNI-03, ADNI-04, ADNI-05, ADNI-06, ADNI-07, ADNI-08, ADNI-01A
\item \strong{Assay} \emph{character}   , with options: 20220119_ADNI-2_hPlasma_Dyna_0^5mL_OTL_NO-HM_VO_JB-SL-For-Export, 20220121_ADNI-3_hPlasma_Dyna_0^5mL_OTL_NO-HM_VO_JB-SL-For-Export, 20220128_ADNI-4_hPlasma_Dyna_0^5mL_OTL_NO-HM_VO_JB-SL-For-Export, 20220131_ADNI-5_hPlasma_Dyna_0^5mL_OTL_NO-HM_VO_JB-SL-For-Export, 20220203_ADNI-6_hPlasma_Dyna_0^5mL_OTL_NO-HM_VO_JB-SL-For-Export, 20220209_ADNI-7_hPlasma_Dyna_0^5mL_OTL_NO-HM_VO_JB-SL-For-Export, 20220215_ADNI-8_hPlasma_Dyna_0^5mL_OTL_NO-HM_VO_JB-SL-For-Export, 20220219_ADNI-1B_hPlasma_Dyna_0^5mL_OTL_NO-HM_VO_JB-SL-For-Export
\item \strong{Assay_IP_protocol} \emph{character}   , with options: IP: 0.5mL Hamilton/MCX/Oxi Ver 2.0
\item \strong{QC_Status} \emph{character}   , with options: Passed, Passed with warnings, Fail
\item \strong{QC_Flags} \emph{character}   , with options: (09) Re-injected;, (00) No MS signal; (09) Re-injected;, (10) Misc.;, (02) Backpressure;
\item \strong{QC_Comment} \emph{character}    
\item \strong{Abeta_4240_Standardized} \emph{numeric}   Range value: 0.0841398, ..., 0.1732365
\item \strong{Intertcept_Standardization} \emph{numeric}   Range value: -0.032107476, ..., 0.032289702
\item \strong{Slope_Standardization} \emph{numeric}   Range value: 0.724008049, ..., 1.262250177
\item \strong{Sample_volume} \emph{numeric}   Range value: 0.43, ..., 0.45
\item \strong{Sample_volume_UNITS} \emph{character}   , with options: mL
\item \strong{Abeta_42_conc} \emph{numeric}   Range value: 0.012, ..., 0.06
\item \strong{Abeta_42_conc_UNITS} \emph{character}   , with options: ng/mL
\item \strong{Abeta_42_N14N15} \emph{numeric}   Range value: 0.3235476, ..., 1.6201517
\item \strong{Abeta_42_N15_ISTD_amount} \emph{numeric}   Range value: 0.0167, ..., 0.0167
\item \strong{Abeta_42_N15_ISTD_amount_UNITS} \emph{character}   , with options: ng
\item \strong{Abeta_40_conc} \emph{numeric}   Range value: 0.0699, ..., 0.4614
\item \strong{Abeta_40_conc_UNITS} \emph{character}   , with options: ng/mL
\item \strong{Abeta_40_N14N15} \emph{numeric}   Range value: 0.188604, ..., 1.245787
\item \strong{Abeta_40_N15_ISTD_amount} \emph{numeric}   Range value: 0.1667, ..., 0.1667
\item \strong{Abeta_40_N15_ISTD_amount_UNITS} \emph{character}   , with options: ng
\item \strong{Abeta_4240} \emph{numeric}   Range value: 0.082207847, ..., 0.171548661
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
