% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR}
\alias{BHR}
\title{Brain Health Registry (BHR)}
\format{
A data frame with 1104 observations and 14 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR)
}
\description{
Brain Health Registry (BHR) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNIGO, ADNI2, ADNI3, ADNI1
\item \strong{ID} \emph{numeric}   Range value: 522, ..., 186552
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 7125
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 63
\item \strong{VISCODE} \emph{character}   , with options: init, bl
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{DONE} \emph{numeric} Was BHR brochure given to participant and BHR Study Partner Flyer given to the study partner? 0=No ; 1=Yes
\item \strong{NDREASON} \emph{numeric} If No, reason not done: 1=Participant Unable for Cognitive Reasons ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 4=Equipment/Technical related issue ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\item \strong{EXAMDATE} \emph{ } If Yes, date brochure provided Date: YYY-MM-DD
\item \strong{WRITTEN} \emph{numeric} If Yes, was the ADNI Online ID written down on the participant's brochure? 0=No ; 1=Yes
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
