% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_BASELINE_QUESTIONNAIRE}
\alias{BHR_BASELINE_QUESTIONNAIRE}
\title{Brain Health Registry (BHR) BASELINE QUESTIONNAIRE}
\format{
A data frame with 95 observations and 73 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_BASELINE_QUESTIONNAIRE)
}
\description{
Brain Health Registry (BHR) BASELINE QUESTIONNAIRE data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNIGO, ADNI2, ADNI3, ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 69, ..., 6688
\item \strong{PTID} \emph{character} Subject code obtained from the CogState Brief Battery CRF report  
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{Latest} \emph{numeric} Is this the most recent collection of this procedure for this subject? 1=Yes ; 0=No
\item \strong{QID49} \emph{numeric} How often do you use a computer? 1=Daily ; 2=Weekly ; 3=Monthly ; 4=Almost never
\item \strong{QID50} \emph{numeric} How confident are you with computers? 1=Not confident at all ; 2=I usually need help ; 3=It depends on the task ; 4=Confident
\item \strong{QID4} \emph{numeric} Have your children, your sibling(s), or parent(s) ever been diagnosed with Alzheimer's Disease? 1=Yes ; 2=No
\item \strong{QID2} \emph{numeric} Are you concerned that you have a memory problem? 1=Yes ; 2=No
\item \strong{QID3} \emph{numeric} Do you have trouble remembering if you have already told someone something? 1=Yes ; 2=No
\item \strong{QID5} \emph{numeric} Compared to 10 years ago, would you say there has been a change in your memory? 1=Yes ; 2=No
\item \strong{QID20} \emph{numeric} In general, would you say your health is: 1=Excellent ; 2=Very Good ; 3=Good ; 4=Fair ; 5=Poor
\item \strong{QID7} \emph{numeric} Compared to one year ago, how would you rate your health in general now? 1=Much Better ; 2=Better ; 3=Same ; 4=Worse ; 5=Much Worse
\item \strong{QID8} \emph{numeric} Are you in good spirits most of the time? 1=Yes ; 2=No
\item \strong{QID21} \emph{numeric} Have you dropped many of your activities and interests? 1=Yes ; 2=No
\item \strong{QID38} \emph{numeric} Stress means a situation in which a person feels tense, restless, nervous or anxious or is unable to sleep at night because his/her mind is troubled all the time. During the past 12 months, would you say you experienced: 1=A lot of stress ; 2=A moderate amount of stress ; 3=Relatively little stress ; 4=Almost no stress at all
\item \strong{QID22} \emph{numeric} Please indicate whether you currently have or had experienced alcohol abuse in the past. 1=Yes ; 3=No
\item \strong{QID23_TEXT} \emph{numeric} How long has it been in years since you stopped your alcohol abuse? If you still abuse alcohol please write 0. Range value: 10, ..., 11
\item \strong{QID41} \emph{numeric} Please indicate whether you currently have or had experienced drug abuse in the past. 1=Yes ; 2=No
\item \strong{QID42_TEXT} \emph{logical} How long has it been in years since you stopped the drug abuse? If you still abuse drugs please write 0. A Boolean value: TRUE/ FALSE
\item \strong{QID15_1} \emph{numeric} Are you currently taking any of the following types of medications? (Select all that apply)... Anti-depressant medication 1=Yes ; =No
\item \strong{QID15_2} \emph{numeric} Are you currently taking any of the following types of medications? (Select all that apply)... Anti-anxiety medication 1=Yes ; =No
\item \strong{QID26_1} \emph{numeric} How long has your current anti-depressant prescription been stable (same medication, frequency, dose)?  ... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID26_2} \emph{numeric} How long has your current anti-depressant prescription been stable (same medication, frequency, dose)? ... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID27_1} \emph{numeric} How long has your current anti-anxiety prescription been stable (same medication, frequency, dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID27_3} \emph{numeric} How long has your current anti-anxiety prescription been stable (same medication, frequency, dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID16_1} \emph{logical} Are you currently taking any of the following medications? (Select all that apply)... Donepezil (Aricept) 1=Yes ; =No
\item \strong{QID16_2} \emph{logical} Are you currently taking any of the following medications? (Select all that apply)... Tacrine (Cognex) 1=Yes ; =No
\item \strong{QID16_3} \emph{numeric} Are you currently taking any of the following medications? (Select all that apply)... Rivastigmine (Exelon) 1=Yes ; =No
\item \strong{QID16_4} \emph{logical} Are you currently taking any of the following medications? (Select all that apply)... Memantine (Namenda) 1=Yes ; =No
\item \strong{QID16_5} \emph{numeric} Are you currently taking any of the following medications? (Select all that apply)... Galantamine (Razadyne) 1=Yes ; =No
\item \strong{QID16_6} \emph{logical} Are you currently taking any of the following medications? (Select all that apply)... Risperidone (Risperdal) 1=Yes ; =No
\item \strong{QID16_7} \emph{logical} Are you currently taking any of the following medications? (Select all that apply)... Paliperidone (Invega) 1=Yes ; =No
\item \strong{QID16_8} \emph{logical} Are you currently taking any of the following medications? (Select all that apply)... Olanzapine (Zyprexa) 1=Yes ; =No
\item \strong{QID16_9} \emph{logical} Are you currently taking any of the following medications? (Select all that apply)... Quetiapine (Seroquel) 1=Yes ; =No
\item \strong{QID16_10} \emph{logical} Are you currently taking any of the following medications? (Select all that apply)... Aripiprazole (Abilify) 1=Yes ; =No
\item \strong{QID28_1} \emph{logical} How long has your current Donepezil (Aricept) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID28_2} \emph{logical} How long has your current Donepezil (Aricept) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID29_1} \emph{logical} How long has your current Tacrine (Cognex) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID29_2} \emph{logical} How long has your current Tacrine (Cognex) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID30_1} \emph{numeric} How long has your current Rivastigmine (Exelon) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID30_2} \emph{numeric} How long has your current Rivastigmine (Exelon) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID31_1} \emph{logical} How long has your current Memantine (Namenda) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID31_2} \emph{logical} How long has your current Memantine (Namenda) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID32_1} \emph{numeric} How long has your current Galantamine (Razadyne) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID32_2} \emph{numeric} How long has your current Galantamine (Razadyne) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID33_1} \emph{logical} How long has your current Risperidone (Risperdal) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID33_2} \emph{logical} How long has your current Risperidone (Risperdal) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID34_1} \emph{logical} How long has your current Paliperidone (Invega) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID34_2} \emph{logical} How long has your current Paliperidone (Invega) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID35_1} \emph{logical} How long has your current Olanzapine (Zyprexa) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID35_2} \emph{logical} How long has your current Olanzapine (Zyprexa) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID36_1} \emph{logical} How long has your current Quetiapine (Seroquel) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID36_2} \emph{logical} How long has your current Quetiapine (Seroquel) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID37_1} \emph{logical} How long has your current Aripiprazole (Abilify) prescription been stable (same frequency and dose)?... Years 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID37_2} \emph{logical} How long has your current Aripiprazole (Abilify) prescription been stable (same frequency and dose)?... Months 1=0 ; 2=0 ~ 0 ; 3=0 ~ 1 ; 4=0 ~ 2 ; 5=0 ~ 3 ; 6=0 ~ 4 ; 7=0 ~ 5 ; 8=0 ~ 6 ; 9=0 ~ 7 ; 10=0 ~ 8 ; 11=0 ~ 9 ; 12=0 ~ 10 ; 13=0 ~ 11 ; 14=1 ; 15=1 ~ 0 ; 16=1 ~ 1 ; 17=1 ~ 2 ; 18=1 ~ 3 ; 19=1 ~ 4 ; 20=1 ~ 5 ; 21=1 ~ 6 ; 22=1 ~ 7 ; 23=1 ~ 8 ; 24=1 ~ 9 ; 25=1 ~ 10 ; 26=1 ~ 11 ; 27=2 ; 28=2 ~ 0 ; 29=2 ~ 1 ; 30=2 ~ 2 ; 31=2 ~ 3 ; 32=2 ~ 4 ; 33=2 ~ 5 ; 34=2 ~ 6 ; 35=2 ~ 7 ; 36=2 ~ 8 ; 37=2 ~ 9 ; 38=2 ~ 10 ; 39=2 ~ 11 ; 40=3 ; 41=3 ~ 0 ; 42=3 ~ 1 ; 43=3 ~ 2 ; 44=3 ~ 3 ; 45=3 ~ 4 ; 46=3 ~ 5 ; 47=3 ~ 6 ; 48=3 ~ 7 ; 49=3 ~ 8 ; 50=3 ~ 9 ; 51=3 ~ 10 ; 52=3 ~ 11 ; 53=4 ; 54=4 ~ 0 ; 55=4 ~ 1 ; 56=4 ~ 2 ; 57=4 ~ 3 ; 58=4 ~ 4 ; 59=4 ~ 5 ; 60=4 ~ 6 ; 61=4 ~ 7 ; 62=4 ~ 8 ; 63=4 ~ 9 ; 64=4 ~ 10 ; 65=4 ~ 11 ; 66=5 ; 67=5 ~ 0 ; 68=5 ~ 1 ; 69=5 ~ 2 ; 70=5 ~ 3 ; 71=5 ~ 4 ; 72=5 ~ 5 ; 73=5 ~ 6 ; 74=5 ~ 7 ; 75=5 ~ 8 ; 76=5 ~ 9 ; 77=5 ~ 10 ; 78=5 ~ 11 ; 79=6 ; 80=6 ~ 0 ; 81=6 ~ 1 ; 82=6 ~ 2 ; 83=6 ~ 3 ; 84=6 ~ 4 ; 85=6 ~ 5 ; 86=6 ~ 6 ; 87=6 ~ 7 ; 88=6 ~ 8 ; 89=6 ~ 9 ; 90=6 ~ 10 ; 91=6 ~ 11 ; 92=7 ; 93=7 ~ 0 ; 94=7 ~ 1 ; 95=7 ~ 2 ; 96=7 ~ 3 ; 97=7 ~ 4 ; 98=7 ~ 5 ; 99=7 ~ 6 ; 100=7 ~ 7 ; 101=7 ~ 8 ; 102=7 ~ 9 ; 103=7 ~ 10 ; 104=7 ~ 11 ; 105=8 ; 106=8 ~ 0 ; 107=8 ~ 1 ; 108=8 ~ 2 ; 109=8 ~ 3 ; 110=8 ~ 4 ; 111=8 ~ 5 ; 112=8 ~ 6 ; 113=8 ~ 7 ; 114=8 ~ 8 ; 115=8 ~ 9 ; 116=8 ~ 10 ; 117=8 ~ 11 ; 118=9 ; 119=9 ~ 0 ; 120=9 ~ 1 ; 121=9 ~ 2 ; 122=9 ~ 3 ; 123=9 ~ 4 ; 124=9 ~ 5 ; 125=9 ~ 6 ; 126=9 ~ 7 ; 127=9 ~ 8 ; 128=9 ~ 9 ; 129=9 ~ 10 ; 130=9 ~ 11 ; 131=10 ; 132=10 ~ 0 ; 133=10 ~ 1 ; 134=10 ~ 2 ; 135=10 ~ 3 ; 136=10 ~ 4 ; 137=10 ~ 5 ; 138=10 ~ 6 ; 139=10 ~ 7 ; 140=10 ~ 8 ; 141=10 ~ 9 ; 142=10 ~ 10 ; 143=10 ~ 11 ; 144=11 ; 145=11 ~ 0 ; 146=11 ~ 1 ; 147=11 ~ 2 ; 148=11 ~ 3 ; 149=11 ~ 4 ; 150=11 ~ 5 ; 151=11 ~ 6 ; 152=11 ~ 7 ; 153=11 ~ 8 ; 154=11 ~ 9 ; 155=11 ~ 10 ; 156=11 ~ 11 ; 157=12 ; 158=12 ~ 0 ; 159=12 ~ 1 ; 160=12 ~ 2 ; 161=12 ~ 3 ; 162=12 ~ 4 ; 163=12 ~ 5 ; 164=12 ~ 6 ; 165=12 ~ 7 ; 166=12 ~ 8 ; 167=12 ~ 9 ; 168=12 ~ 10 ; 169=12 ~ 11 ; 170=13 ; 171=13 ~ 0 ; 172=13 ~ 1 ; 173=13 ~ 2 ; 174=13 ~ 3 ; 175=13 ~ 4 ; 176=13 ~ 5 ; 177=13 ~ 6 ; 178=13 ~ 7 ; 179=13 ~ 8 ; 180=13 ~ 9 ; 181=13 ~ 10 ; 182=13 ~ 11 ; 183=14 ; 184=14 ~ 0 ; 185=14 ~ 1 ; 186=14 ~ 2 ; 187=14 ~ 3 ; 188=14 ~ 4 ; 189=14 ~ 5 ; 190=14 ~ 6 ; 191=14 ~ 7 ; 192=14 ~ 8 ; 193=14 ~ 9 ; 194=14 ~ 10 ; 195=14 ~ 11 ; 196=15 ; 197=15 ~ 0 ; 198=15 ~ 1 ; 199=15 ~ 2 ; 200=15 ~ 3 ; 201=15 ~ 4 ; 202=15 ~ 5 ; 203=15 ~ 6 ; 204=15 ~ 7 ; 205=15 ~ 8 ; 206=15 ~ 9 ; 207=15 ~ 10 ; 208=15 ~ 11 ; 209=16 ; 210=16 ~ 0 ; 211=16 ~ 1 ; 212=16 ~ 2 ; 213=16 ~ 3 ; 214=16 ~ 4 ; 215=16 ~ 5 ; 216=16 ~ 6 ; 217=16 ~ 7 ; 218=16 ~ 8 ; 219=16 ~ 9 ; 220=16 ~ 10 ; 221=16 ~ 11 ; 222=17 ; 223=17 ~ 0 ; 224=17 ~ 1 ; 225=17 ~ 2 ; 226=17 ~ 3 ; 227=17 ~ 4 ; 228=17 ~ 5 ; 229=17 ~ 6 ; 230=17 ~ 7 ; 231=17 ~ 8 ; 232=17 ~ 9 ; 233=17 ~ 10 ; 234=17 ~ 11 ; 235=18 ; 236=18 ~ 0 ; 237=18 ~ 1 ; 238=18 ~ 2 ; 239=18 ~ 3 ; 240=18 ~ 4 ; 241=18 ~ 5 ; 242=18 ~ 6 ; 243=18 ~ 7 ; 244=18 ~ 8 ; 245=18 ~ 9 ; 246=18 ~ 10 ; 247=18 ~ 11 ; 248=19 ; 249=19 ~ 0 ; 250=19 ~ 1 ; 251=19 ~ 2 ; 252=19 ~ 3 ; 253=19 ~ 4 ; 254=19 ~ 5 ; 255=19 ~ 6 ; 256=19 ~ 7 ; 257=19 ~ 8 ; 258=19 ~ 9 ; 259=19 ~ 10 ; 260=19 ~ 11 ; 261=20+ ; 262=20+ ~
\item \strong{QID39_1} \emph{numeric} Compared to one year ago ...... Do you feel that your memory has declined substantially? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID39_2} \emph{numeric} Compared to one year ago ...... Do others tell you that you tend to repeat questions over and over? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID39_3} \emph{numeric} Compared to one year ago ...... Have you been misplacing things more often? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID39_4} \emph{numeric} Compared to one year ago ...... Do you find that lately you are relying more on written reminders (e.g., shopping lists, calendars)? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID39_5} \emph{numeric} Compared to one year ago ...... Do you need more help from others to remember appointments, family occasions or holidays? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID46_6} \emph{numeric} Compared to one year ago ...... Do you have more trouble recalling names, finding the right word, or completing sentences? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID46_7} \emph{numeric} Compared to one year ago ...... Do you have more trouble driving (e.g., do you drive more slowly, have more trouble at night, tend to get lost, have accidents)? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID46_8} \emph{numeric} Compared to one year ago ...... Compared to one year ago, do you have more difficulty managing money (e.g., paying bills, calculating change, completing tax forms)? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID46_9} \emph{numeric} Compared to one year ago ...... Are you less involved in social activities? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID46_13} \emph{numeric} Compared to one year ago ...... Has your work performance (paid or volunteer) declined significantly compared to one year ago? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID47_11} \emph{numeric} Compared to one year ago ...... Do you have more trouble following the news, or the plots of books, movies or TV shows, compared to one year ago? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID47_12} \emph{numeric} Compared to one year ago ...... Are there any activities (e.g., hobbies, such as card games, crafts) that are substantially more difficult for you now compared to one year ago? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID47_13} \emph{numeric} Compared to one year ago ...... Are you more likely to become disoriented, or get lost, for example when traveling to another city? 1=Yes ; 2=No ; 3=Maybe
\item \strong{QID47_14} \emph{numeric} Compared to one year ago ...... Do you have more difficulty using household appliances (such as the washing machine, VCR or computer)? 1=Yes ; 2=No ; 3=Maybe
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
