% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_EVERYDAY_COGNITION}
\alias{BHR_EVERYDAY_COGNITION}
\title{Brain Health Registry (BHR) EVERYDAY COGNITION}
\format{
A data frame with 346 observations and 47 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_EVERYDAY_COGNITION)
}
\description{
Brain Health Registry (BHR) EVERYDAY COGNITION data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNIGO, ADNI3, ADNI1
\item \strong{RID} \emph{character} Participant roster ID  
\item \strong{PTID} \emph{character} Subject code obtained from the CogState Brief Battery CRF report  
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{latest} \emph{numeric}   Range value: 0, ..., 1
\item \strong{QID49_1} \emph{numeric} Compared to 10 years ago, has there been any change in...MEMORY... Remembering a few shopping items without a list. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID49_2} \emph{numeric} Compared to 10 years ago, has there been any change in...MEMORY... Remembering things that happened recently (such as recent outings, events in the news). 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID49_3} \emph{numeric} Compared to 10 years ago, has there been any change in...MEMORY... Recalling conversations a few days later. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID49_4} \emph{numeric} Compared to 10 years ago, has there been any change in...MEMORY... Remembering where I have placed objects. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID49_5} \emph{numeric} Compared to 10 years ago, has there been any change in...MEMORY... Repeating stories and/or questions. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID49_6} \emph{numeric} Compared to 10 years ago, has there been any change in...MEMORY... Remembering the current date or day of the week. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID49_7} \emph{numeric} Compared to 10 years ago, has there been any change in...MEMORY... Remembering I have already told someone something. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID49_8} \emph{numeric} Compared to 10 years ago, has there been any change in...MEMORY... Remembering appointments, meetings, or engagements. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_1} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Forgetting the names of objects. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_2} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Verbally giving instructions to others. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_3} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Finding the right words to use in a conversation. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_4} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Communicating thoughts in a conversation. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_5} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Following a story in a book or on TV. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_6} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Understanding the point of what other people are trying to say. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_7} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Remembering the meaning of common words. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_8} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Describing a program I have watched on TV. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID50_9} \emph{numeric} Compared to 10 years ago, has there been any change in...LANGUAGE... Understanding spoken directions or instructions. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID51_1} \emph{numeric} Compared to 10 years ago, has there been any change in...VISUAL-SPATIAL AND PERCEPTUAL ABILITIES... Following a map to find a new location. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID51_2} \emph{numeric} Compared to 10 years ago, has there been any change in...VISUAL-SPATIAL AND PERCEPTUAL ABILITIES... Reading a map and helping with directions when someone else is driving. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID51_3} \emph{numeric} Compared to 10 years ago, has there been any change in...VISUAL-SPATIAL AND PERCEPTUAL ABILITIES... Finding my car in a parking lot. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID51_4} \emph{numeric} Compared to 10 years ago, has there been any change in...VISUAL-SPATIAL AND PERCEPTUAL ABILITIES... Finding my way back to a meeting spot in the mall or other location. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID51_5} \emph{numeric} Compared to 10 years ago, has there been any change in...VISUAL-SPATIAL AND PERCEPTUAL ABILITIES... Finding my way around a familiar neighborhood. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID51_6} \emph{numeric} Compared to 10 years ago, has there been any change in...VISUAL-SPATIAL AND PERCEPTUAL ABILITIES... Finding my way around a familiar store. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID51_7} \emph{numeric} Compared to 10 years ago, has there been any change in...VISUAL-SPATIAL AND PERCEPTUAL ABILITIES... Finding my way around a house visited many times. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID52_1} \emph{numeric} Compared to 10 years ago, has there been any change in... PLANNING... Planning a sequence of stops on a shopping trip. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID52_2} \emph{numeric} Compared to 10 years ago, has there been any change in... PLANNING... The ability to anticipate weather changes and plan accordingly (i.e., bring a coat or umbrella). 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID52_3} \emph{numeric} Compared to 10 years ago, has there been any change in... PLANNING... Developing a schedule in advance of anticipated events. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID52_4} \emph{numeric} Compared to 10 years ago, has there been any change in... PLANNING... Thinking things through before acting. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID52_5} \emph{numeric} Compared to 10 years ago, has there been any change in... PLANNING... Thinking ahead. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID53_1} \emph{numeric} Compared to 10 years ago, has there been any change in... ORGANIZATION... Keeping living and work space organized. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID53_2} \emph{numeric} Compared to 10 years ago, has there been any change in... ORGANIZATION... Balancing the checkbook without error. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID53_3} \emph{numeric} Compared to 10 years ago, has there been any change in... ORGANIZATION... Keeping financial records organized. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID53_4} \emph{numeric} Compared to 10 years ago, has there been any change in... ORGANIZATION... Prioritizing tasks by importance. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID53_5} \emph{numeric} Compared to 10 years ago, has there been any change in... ORGANIZATION... Keeping mail and papers organized. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID53_6} \emph{numeric} Compared to 10 years ago, has there been any change in... ORGANIZATION... Using an organized strategy to manage a medication schedule involving multiple medications. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID54_1} \emph{numeric} Compared to 10 years ago, has there been any change in... DIVIDED ATTENTION... The ability to do two things at once. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID54_2} \emph{numeric} Compared to 10 years ago, has there been any change in... DIVIDED ATTENTION... Returning to a task after being interrupted. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID54_3} \emph{numeric} Compared to 10 years ago, has there been any change in... DIVIDED ATTENTION... The ability to concentrate on a task without being distracted by external things in the environment. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{QID54_4} \emph{numeric} Compared to 10 years ago, has there been any change in... DIVIDED ATTENTION... Cooking or working and talking at the same time. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
