% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_MEMTRAX}
\alias{BHR_MEMTRAX}
\title{Brain Health Registry (BHR) MEMTRAX}
\format{
A data frame with 1422 observations and 32 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_MEMTRAX)
}
\description{
Brain Health Registry (BHR) MEMTRAX data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNIGO, ADNI2, ADNI3, ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 69, ..., 7071
\item \strong{PTID} \emph{character} Subject code obtained from the CogState Brief Battery CRF report  
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{latest} \emph{numeric}   Range value: 0, ..., 1
\item \strong{ResultDate} \emph{ } Date the result was collected on the MemTrax website POSIXct: YYY-MM-DD
\item \strong{Result} \emph{numeric} Number of the result, in case there were multiple for the same procedure. Range value: 1, ..., 3
\item \strong{OrderSet} \emph{numeric} One of four orderings of unique/repeat Range value: 1, ..., 4
\item \strong{TotalImages} \emph{numeric} The total number of images shown to the user Range value: 49, ..., 50
\item \strong{UniqueImages} \emph{numeric} The number of unique images shown to the user Range value: 24, ..., 25
\item \strong{RepeatImages} \emph{numeric} Number of images the user had seen at least once before Range value: 25, ..., 25
\item \strong{CorrectN} \emph{numeric} Total the user answered correctly Range value: 7, ..., 50
\item \strong{CorrectPCT} \emph{numeric} Percent of images the user answered correctly Range value: 0.14, ..., 1
\item \strong{IncorrectN} \emph{numeric} Total Number of images that got the wrong response Range value: 0, ..., 42
\item \strong{IncorrectPCT} \emph{numeric} Percent of images with an incorrect response Range value: 0, ..., 0.86
\item \strong{CorrectRejectionsN} \emph{numeric} Number of first-time images that were not clicked on Range value: 1, ..., 25
\item \strong{CorrectRejectionsPCT} \emph{numeric} Percent of first-time images that were not clicked on Range value: 0.08, ..., 1
\item \strong{IncorrectRejectionsN} \emph{numeric} Number of Repeat images that were not clicked on Range value: 0, ..., 25
\item \strong{IncorrectRejectionsPCT} \emph{numeric} Percent of Repeat images that were not clicked on Range value: 0, ..., 1
\item \strong{TotalRejectionsN} \emph{numeric} Number of images that were not clicked on Range value: 4, ..., 50
\item \strong{TotalRejectionsPCT} \emph{numeric} Percent of images that were not clicked on Range value: 0.08, ..., 1
\item \strong{CorrectResponsesN} \emph{numeric} Number of Repeat images that were clicked on Range value: 0, ..., 25
\item \strong{CorrectResponsesPCT} \emph{numeric} Percent or Repeat images that were clicked on Range value: 0, ..., 1
\item \strong{CorrectResponsesRT} \emph{numeric} Mean Reaction time in seconds, when clicking on Repeat images Range value: 0, ..., 2.207
\item \strong{IncorrectResponsesN} \emph{numeric} Number of first-time images that were clicked on Range value: 0, ..., 21
\item \strong{IncorrectResponsesPCT} \emph{numeric} Percent of first-time images that were clicked on Range value: 0, ..., 0.9
\item \strong{IncorrectResponsesRT} \emph{numeric} Mean reaction time in seconds, when clicking on first-time images Range value: 0, ..., 2.989
\item \strong{TotalResponsesN} \emph{numeric} Number of images clicked on Range value: 0, ..., 46
\item \strong{TotalResponsesPCT} \emph{numeric} Percent of images clicked on Range value: 0, ..., 0.92
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
