% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_SP_ADL}
\alias{BHR_SP_ADL}
\title{Brain Health Registry (BHR) SP ADL}
\format{
A data frame with 4 observations and 16 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_SP_ADL)
}
\description{
Brain Health Registry (BHR) SP ADL data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{AboutRID} \emph{numeric} Participant roster ID Range value: 2245, ..., 5222
\item \strong{StudyPartnerID} \emph{numeric} Unique identifier for study partners Range value: 2509, ..., 3420
\item \strong{AboutPTID} \emph{character} Code obtained from CBB CRF report of subject about whom the study partner provided responses , with options: 027_s_4919, 027_s_2245, 002_s_4799, 037_s_5222
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{latest} \emph{numeric}   Range value: 1, ..., 1
\item \strong{QID2_1} \emph{numeric} Please answer the following questions using the responses provided.... The ability to make decisions. 1=Normal ; 2=Never did, but could do now ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID2_2} \emph{numeric} Please answer the following questions using the responses provided.... Learning how to use a tool, appliance, or gadget. 1=Normal ; 2=Never did, but could do now ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID2_3} \emph{numeric} Please answer the following questions using the responses provided.... Shopping alone for clothes, household necessities, or groceries. 1=Normal ; 2=Never did, but could do now ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID2_4} \emph{numeric} Please answer the following questions using the responses provided.... Playing a game of skill such as bridge or chess, or working on a hobby. 1=Normal ; 2=Never did, but could do now ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID2_5} \emph{numeric} Please answer the following questions using the responses provided.... Heating water, making a cup of coffee, turning off the stove. 1=Normal ; 2=Never did, but could do now ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID2_6} \emph{numeric} Please answer the following questions using the responses provided.... Preparing a balanced meal. 1=Normal ; 2=Never did, but could do now ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID2_7} \emph{numeric} Please answer the following questions using the responses provided.... Driving (e.g., drive more slowly, have more trouble at night, tend to get lost, have accidents)? 1=Normal ; 2=Never did, but could do now ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID2_8} \emph{numeric} Please answer the following questions using the responses provided.... Performance at work (paid or volunteer). This may be based on your own observations, or comments from your study partner's coworkers. 1=Normal ; 2=Never did, but could do now ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
