% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_SP_CAREGIVER_BURDEN}
\alias{BHR_SP_CAREGIVER_BURDEN}
\title{Brain Health Registry (BHR) SP CAREGIVER BURDEN}
\format{
A data frame with 37 observations and 133 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_SP_CAREGIVER_BURDEN)
}
\description{
Brain Health Registry (BHR) SP CAREGIVER BURDEN data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{AboutRID} \emph{numeric} Participant roster ID Range value: 2201, ..., 6644
\item \strong{StudyPartnerID} \emph{numeric} Unique identifier for study partners Range value: 2509, ..., 6051
\item \strong{AboutPTID} \emph{character} Code obtained from CBB CRF report of subject about whom the study partner provided responses  
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{latest} \emph{numeric}   Range value: 0, ..., 1
\item \strong{QID13} \emph{numeric} Are you the primary caregiver for YOUR STUDY PARTNER, who needs extra assistance due to a disability, condition, or illness? We define a primary caregiver as the person who is primarily responsible for helping make medical care decisions, is emotionally invested in the person's care, and provides help with daily activities (almost daily). Daily caregiving activities may include helping with medical or therapeutic treatment, preparing meal, and/or helping with dressing and/or bathing. Caregivers may be caring for a loved with dementia, or be the parent to a child with development delays or burdensome health conditions, for example. 1=Yes, I am a caregiver ; 2=No, I am not a caregiver ; 3=I am not a caregiver based on this definition, but I do help someone with age-related issues
\item \strong{QID53} \emph{numeric} Are you the primary caregiver for YOUR STUDY PARTNER, who needs extra assistance due to a disability, condition, or illness? We define a primary caregiver as the person who is primarily responsible for helping make medical care decisions, is emotionally invested in the person's care, and provides help with daily activities (or almost daily). Daily caregiving activities may include helping with medical or therapeutic treatments, preparing meals, and/or helping with dressing and/or bathing. Caregivers may be caring for a loved one with dementia, or be the parents to a child with developmental delays or burdensome health conditions, for example. 1=Yes, I am a caregiver for MY STUDY PARTNER ; 2=No, I am not a caregiver for MY STUDY PARTNER ; 3=I am not a caregiver based on this definition, but I do help MY STUDY PARTNER with some health issues
\item \strong{QID31} \emph{numeric} Because you do not identify as a primary caregiver for YOUR STUDY PARTNER, we have no more questions for you.  If you become a primary caregiver at a later time, you will have the opportunity to let us know at a future visit. Range value: 1, ..., 1
\item \strong{QID32} \emph{numeric} Are you the primary caregiver for more than one person? 1=Yes ; 2=No
\item \strong{QID55} \emph{numeric} Are you caregiving for more than one person? 1=Yes ; 3=No
\item \strong{QID24_13} \emph{numeric} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Alzheimer's disease 1=Yes ; =No
\item \strong{QID24_14} \emph{numeric} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Mild Cognitive Impairment (MCI) 1=Yes ; =No
\item \strong{QID24_15} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Frontotemporal Dementia 1=Yes ; =No
\item \strong{QID24_16} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Other dementia (e.g. Vascular dementia, Dementia with Lewy Bodies) 1=Yes ; =No
\item \strong{QID24_7} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Dementia 1=Yes ; =No
\item \strong{QID24_17} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Brain disease that affects movement (e.g. Parkinson's, Huntington's, ALS, Motor Neuron Disease, Multiple Sclerosis) 1=Yes ; =No
\item \strong{QID24_18} \emph{numeric} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Epilepsy 1=Yes ; =No
\item \strong{QID24_19} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Autism 1=Yes ; =No
\item \strong{QID24_20} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Psychiatric illness (e.g. schizophrenia, depression, bipolar disorder, anxiety disorder) 1=Yes ; =No
\item \strong{QID24_9} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Mental or Psychiatric Illness 1=Yes ; =No
\item \strong{QID24_21} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Intellectual Disability 1=Yes ; =No
\item \strong{QID24_6} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Mental Disability or Impairment 1=Yes ; =No
\item \strong{QID24_8} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Brain-related Injury (e.g. tumor, stroke, traumatic brain injury) 1=Yes ; =No
\item \strong{QID24_1} \emph{numeric} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Cancer 1=Yes ; =No
\item \strong{QID24_2} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Diabetes 1=Yes ; =No
\item \strong{QID24_3} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Heart Disease 1=Yes ; =No
\item \strong{QID24_4} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Other Chronic Illness 1=Yes ; =No
\item \strong{QID24_5} \emph{numeric} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Physical Disability or Injury 1=Yes ; =No
\item \strong{QID24_11} \emph{numeric} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Age-Related Issues 1=Yes ; =No
\item \strong{QID24_10} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Other, please specify: 1=Yes ; =No
\item \strong{QID24_10_TEXT} \emph{logical} What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Other, please specify: 1=Yes ; =No
\item \strong{QID25_1_TEXT} \emph{numeric} For how long have you been caregiving?    For example: 1 year, 3 months, 0 years, 7 months, 9 years, 0 months... Years Range value: 0, ..., 10
\item \strong{QID25_2_TEXT} \emph{numeric} For how long have you been caregiving?    For example: 1 year, 3 months, 0 years, 7 months, 9 years, 0 months... Months Range value: 0, ..., 8
\item \strong{QID33} \emph{numeric} Do you live with YOUR STUDY PARTNER? 1=Yes ; 2=No
\item \strong{QID34} \emph{numeric} What is your employment status? 1=Employed full-time ; 2=Employed part-time ; 3=Full-time student ; 4=Part-time student ; 5=Unemployed ; 6=Retired
\item \strong{QID35} \emph{numeric} Did you have to leave your paid job, cut down on your hours, or drop out of school due to your caregiving responsibilities? 1=Yes ; 2=No
\item \strong{QID26_TEXT} \emph{numeric} In addition to your role as caregiver, how many hours per week do you work at a paid job or attend school? If you do not have a job or go to school, please write 0. Range value: 0, ..., 50
\item \strong{QID27_TEXT} \emph{numeric} How many hours per week do you spend caring for YOUR STUDY PARTNER? We know this may be difficult to estimate, just give your best guess. If this question is not relevant to you, please enter 0. Range value: 0, ..., 8
\item \strong{QID38_1_1} \emph{numeric} Has it occurred in the past week?... Asking the same question over and over 2=No ; 1=Yes
\item \strong{QID38_1_2} \emph{numeric} Has it occurred in the past week?... Trouble remembering recent events (i.e. items in newspaper or TV) 2=No ; 1=Yes
\item \strong{QID38_1_3} \emph{numeric} Has it occurred in the past week?... Trouble remembering significant past events 2=No ; 1=Yes
\item \strong{QID38_1_4} \emph{numeric} Has it occurred in the past week?... Losing or misplacing things 2=No ; 1=Yes
\item \strong{QID38_1_5} \emph{numeric} Has it occurred in the past week?... Forgetting what day it is 2=No ; 1=Yes
\item \strong{QID38_1_6} \emph{numeric} Has it occurred in the past week?... Starting, but not finishing, things 2=No ; 1=Yes
\item \strong{QID38_2_1} \emph{numeric} How much has this bothered or upset you when it happened?... Asking the same question over and over 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID38_2_2} \emph{numeric} How much has this bothered or upset you when it happened?... Trouble remembering recent events (i.e. items in newspaper or TV) 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID38_2_3} \emph{numeric} How much has this bothered or upset you when it happened?... Trouble remembering significant past events 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID38_2_4} \emph{numeric} How much has this bothered or upset you when it happened?... Losing or misplacing things 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID38_2_5} \emph{numeric} How much has this bothered or upset you when it happened?... Forgetting what day it is 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID38_2_6} \emph{numeric} How much has this bothered or upset you when it happened?... Starting, but not finishing, things 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID42_1_1} \emph{numeric} Has it occurred in the past week?... Difficulty concentrating on a task 2=No ; 1=Yes
\item \strong{QID42_1_2} \emph{numeric} Has it occurred in the past week?... Destroying property 2=No ; 1=Yes
\item \strong{QID42_1_3} \emph{numeric} Has it occurred in the past week?... Doing things that embarrass you 2=No ; 1=Yes
\item \strong{QID42_1_4} \emph{numeric} Has it occurred in the past week?... Waking you or other family members up at night 2=No ; 1=Yes
\item \strong{QID42_1_5} \emph{numeric} Has it occurred in the past week?... Talking loudly and rapidly 2=No ; 1=Yes
\item \strong{QID42_1_25} \emph{numeric} Has it occurred in the past week?... Appeared anxious or worried 2=No ; 1=Yes
\item \strong{QID42_2_1} \emph{numeric} How much has this bothered or upset you when it happened?... Difficulty concentrating on a task 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID42_2_2} \emph{numeric} How much has this bothered or upset you when it happened?... Destroying property 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID42_2_3} \emph{numeric} How much has this bothered or upset you when it happened?... Doing things that embarrass you 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID42_2_4} \emph{numeric} How much has this bothered or upset you when it happened?... Waking you or other family members up at night 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID42_2_5} \emph{numeric} How much has this bothered or upset you when it happened?... Talking loudly and rapidly 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID42_2_25} \emph{numeric} How much has this bothered or upset you when it happened?... Appeared anxious or worried 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID43_1_1} \emph{numeric} Has it occurred in the past week?... Engaging in behavior that is potentially dangerous to self or others 2=No ; 1=Yes
\item \strong{QID43_1_2} \emph{numeric} Has it occurred in the past week?... Threatened to hurt oneself 2=No ; 1=Yes
\item \strong{QID43_1_3} \emph{numeric} Has it occurred in the past week?... Threatened to hurt others 2=No ; 1=Yes
\item \strong{QID43_1_4} \emph{numeric} Has it occurred in the past week?... Aggressive to others verbally 2=No ; 1=Yes
\item \strong{QID43_1_5} \emph{numeric} Has it occurred in the past week?... Appeared sad or depressed 2=No ; 1=Yes
\item \strong{QID43_1_6} \emph{numeric} Has it occurred in the past week?... Expressing feelings of hopelessness or sadness about the future 2=No ; 1=Yes
\item \strong{QID43_2_1} \emph{numeric} How much has this bothered or upset you when it happened?... Engaging in behavior that is potentially dangerous to self or others 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID43_2_2} \emph{numeric} How much has this bothered or upset you when it happened?... Threatened to hurt oneself 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID43_2_3} \emph{numeric} How much has this bothered or upset you when it happened?... Threatened to hurt others 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID43_2_4} \emph{numeric} How much has this bothered or upset you when it happened?... Aggressive to others verbally 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID43_2_5} \emph{numeric} How much has this bothered or upset you when it happened?... Appeared sad or depressed 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID43_2_6} \emph{numeric} How much has this bothered or upset you when it happened?... Expressing feelings of hopelessness or sadness about the future 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID44_1_1} \emph{numeric} Has it occurred in the past week?... Crying and tearfulness 2=No ; 1=Yes
\item \strong{QID44_1_2} \emph{numeric} Has it occurred in the past week?... Commenting about death of self or others 2=No ; 1=Yes
\item \strong{QID44_1_3} \emph{numeric} Has it occurred in the past week?... Talking about feeling lonely 2=No ; 1=Yes
\item \strong{QID44_1_4} \emph{numeric} Has it occurred in the past week?... Comments about feeling worthless or being a burden to others 2=No ; 1=Yes
\item \strong{QID44_1_5} \emph{numeric} Has it occurred in the past week?... Comments about feeling like a failure, or about not having any worthwhile accomplishments in life 2=No ; 1=Yes
\item \strong{QID44_1_6} \emph{numeric} Has it occurred in the past week?... Arguing, irritability, and/or complaining 2=No ; 1=Yes
\item \strong{QID44_2_1} \emph{numeric} How much has this bothered or upset you when it happened?... Crying and tearfulness 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID44_2_2} \emph{numeric} How much has this bothered or upset you when it happened?... Commenting about death of self or others 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID44_2_3} \emph{numeric} How much has this bothered or upset you when it happened?... Talking about feeling lonely 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID44_2_4} \emph{numeric} How much has this bothered or upset you when it happened?... Comments about feeling worthless or being a burden to others 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID44_2_5} \emph{numeric} How much has this bothered or upset you when it happened?... Comments about feeling like a failure, or about not having any worthwhile accomplishments in life 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID44_2_6} \emph{numeric} How much has this bothered or upset you when it happened?... Arguing, irritability, and/or complaining 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week
\item \strong{QID19_1} \emph{numeric} As a caregiver do YOU feel...... That because of the time you spend caregiving that you don't have enough time for yourself? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID19_2} \emph{numeric} As a caregiver do YOU feel...... Stressed between caregiving and trying to meet other responsibilities (work/ family)? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID19_3} \emph{numeric} As a caregiver do YOU feel...... Angry when you are around YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID19_4} \emph{numeric} As a caregiver do YOU feel...... That YOUR STUDY PARTNER currently affects your relationship with family members in a negative way? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID19_5} \emph{numeric} As a caregiver do YOU feel...... Strained when you are around YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID19_6} \emph{numeric} As a caregiver do YOU feel...... That your health has suffered because of caregiving? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID45_1} \emph{numeric} As a caregiver do YOU feel...... That you don't have as much privacy as you would like because of caregiving? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID45_2} \emph{numeric} As a caregiver do YOU feel...... That your social life has suffered because you are caring for YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID45_3} \emph{numeric} As a caregiver do YOU feel...... That you have lost control of your life since you began providing care for YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID45_4} \emph{numeric} As a caregiver do YOU feel...... Uncertain about what to do about YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID45_5} \emph{numeric} As a caregiver do YOU feel...... You should be doing more for YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID45_6} \emph{numeric} As a caregiver do YOU feel...... You could do a better job caring for YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID20_1} \emph{numeric} As a caregiver, do YOU feel...... I'm not getting enough sleep 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID20_2} \emph{numeric} As a caregiver, do YOU feel...... My health has suffered 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID20_3} \emph{numeric} As a caregiver, do YOU feel...... Caregiving has made me physically sick 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID20_4} \emph{numeric} As a caregiver, do YOU feel...... I'm physically tired 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID20_5} \emph{numeric} As a caregiver, do YOU feel...... I have headaches or other pains 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)
\item \strong{QID21_1} \emph{numeric} Select the response that best describes YOUR life.... I wish I were free to lead a life of my own 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly Agree (5)
\item \strong{QID21_2} \emph{numeric} Select the response that best describes YOUR life.... I feel trapped by MY STUDY PARTNER'S condition 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly Agree (5)
\item \strong{QID21_3} \emph{numeric} Select the response that best describes YOUR life.... I wish I could just run away 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly Agree (5)
\item \strong{QID29_1} \emph{numeric} Select the response that best describes YOUR life.... Caring for MY STUDY PARTNER is important to me 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID29_2} \emph{numeric} Select the response that best describes YOUR life.... I enjoy caring for MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID29_3} \emph{numeric} Select the response that best describes YOUR life.... Caring for MY STUDY PARTNER makes me feel good 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID29_4} \emph{numeric} Select the response that best describes YOUR life.... I feel privileged to care for MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID29_5} \emph{numeric} Select the response that best describes YOUR life.... I resent having to care for MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID29_6} \emph{numeric} Select the response that best describes YOUR life.... I really want to care for MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID29_7} \emph{logical} Select the response that best describes YOUR life.... MY STUDY PARTNER shows real appreciation of my care 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID29_8} \emph{logical} Select the response that best describes YOUR life.... Nothing I do seems to please MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID29_9} \emph{numeric} Select the response that best describes YOUR life.... MY STUDY PARTNER'S pleasure over little things gives me pleasure 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)
\item \strong{QID47_1} \emph{numeric} Select the response that best describes your life.... The things I do to take care of MY STUDY PARTNER are worthwhile 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID47_2} \emph{numeric} Select the response that best describes your life.... Most of my caregiving activities seem trivial and unimportant 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID47_3} \emph{numeric} Select the response that best describes your life.... There is great value in what I do everyday for MY STUDY PARTNER 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID47_4} \emph{numeric} Select the response that best describes your life.... Being a caregiver is an important role to me 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID47_5} \emph{numeric} Select the response that best describes your life.... I criticize myself for having negative emotions about having to be a caregiver 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID48_10} \emph{numeric} Select the response that best describes your life.... I tell myself I shouldn't be thinking or feeling the way I am about being a caregiver 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID48_6} \emph{numeric} Select the response that best describes your life.... I accept my negative feelings about caregiving, instead of pushing them away 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID48_7} \emph{numeric} Select the response that best describes your life.... Being a caregiver prevents me from having a fulfilling life 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID48_8} \emph{numeric} Select the response that best describes your life.... I am able to be happy and satisfied with my life, and be a caregiver 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID48_9} \emph{numeric} Select the response that best describes your life.... I am at peace about leading a life as a caregiver 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID49_10} \emph{numeric} As a result of being a caregiver, I have experienced the following changes in my life:... I changed my priorities about what is important in life 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID49_6} \emph{numeric} As a result of being a caregiver, I have experienced the following changes in my life:... I have a greater sense of closeness with others 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID49_7} \emph{numeric} As a result of being a caregiver, I have experienced the following changes in my life:... I established a new path for my life 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID49_8} \emph{numeric} As a result of being a caregiver, I have experienced the following changes in my life:... I discovered that I was stronger than I thought I was 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{QID49_9} \emph{numeric} As a result of being a caregiver, I have experienced the following changes in my life:... I have a stronger religious faith 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
