% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_SP_FAQ}
\alias{BHR_SP_FAQ}
\title{Brain Health Registry (BHR) SP FAQ}
\format{
A data frame with 35 observations and 18 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_SP_FAQ)
}
\description{
Brain Health Registry (BHR) SP FAQ data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{AboutRID} \emph{numeric} Participant roster ID Range value: 2201, ..., 6644
\item \strong{StudyPartnerID} \emph{numeric} Unique identifier for study partners Range value: 3240, ..., 6051
\item \strong{AboutPTID} \emph{character} Code obtained from CBB CRF report of subject about whom the study partner provided responses  
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{latest} \emph{numeric}   Range value: 0, ..., 1
\item \strong{QID3_1} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks.... Writing checks, paying bills, or balancing checkbook. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID3_2} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks.... Assembling tax records, business affairs, or other papers. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID3_3} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks.... Shopping alone for clothes, household necessities, or groceries. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID3_4} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks.... Playing a game of skill such as bridge or chess, working on a hobby. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID3_5} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks.... Heating water, making a cup of coffee, turning off the stove. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID4_1} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks... Preparing a balanced meal. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID4_2} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks... Keeping track of current events. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID4_3} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks... Paying attention to and understanding a TV program, book, or magazine. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID4_4} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks... Remembering appointments, family occasions, holidays, medications. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{QID4_5} \emph{numeric} Select the most accurate representation of YOUR STUDY PARTNER'S level of ability to perform each activity over the preceding four weeks... Traveling out of the neighborhood, driving, or arranging to take public transportation. 1=Normal ; 2=Never did, but could now do ; 3=Never did, would have difficulty now ; 4=Has difficulty, but does by self ; 5=Requires assistance ; 6=Dependent
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
