% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_SP_INITIAL}
\alias{BHR_SP_INITIAL}
\title{Brain Health Registry (BHR) SP INITIAL}
\format{
A data frame with 46 observations and 38 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_SP_INITIAL)
}
\description{
Brain Health Registry (BHR) SP INITIAL data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{AboutRID} \emph{numeric} Participant roster ID Range value: 2201, ..., 6644
\item \strong{StudyPartnerID} \emph{numeric} Unique identifier for study partners Range value: 2509, ..., 6051
\item \strong{AboutPTID} \emph{character} Code obtained from CBB CRF report of subject about whom the study partner provided responses  
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{latest} \emph{numeric}   Range value: 0, ..., 1
\item \strong{QID2} \emph{numeric} Are you concerned that YOUR STUDY PARTNER has a memory problem? 1=Yes ; 2=No ; 3=I don't know
\item \strong{QID3} \emph{numeric} Has there been a change in YOUR STUDY PARTNER'S memory in the past 10 years? 1=Yes ; 2=No ; 3=I don't know
\item \strong{QID4} \emph{numeric} Does YOUR STUDY PARTNER have trouble remembering whether he/she told someone something? 1=Yes ; 2=No ; 3=I don't know
\item \strong{QID22} \emph{numeric}   Range value: 1, ..., 3
\item \strong{QID23} \emph{numeric}   Range value: 1, ..., 3
\item \strong{QID24} \emph{numeric}   Range value: 1, ..., 3
\item \strong{QID5} \emph{numeric} Compared to one year ago, does it seem to you that lately YOUR STUDY PARTNER is relying more on written reminders (e.g., shopping lists, calendars)? 1=Yes ; 2=Maybe ; 3=No
\item \strong{QID6} \emph{numeric} Compared to one year ago, does YOUR STUDY PARTNER need more help from others to remember appointments, family occasions or holidays? 1=Yes ; 2=Maybe ; 3=No
\item \strong{QID27} \emph{numeric}   Range value: 1, ..., 3
\item \strong{QID28} \emph{numeric}   Range value: 1, ..., 3
\item \strong{QID7} \emph{numeric} Compared to one year ago, is YOUR STUDY PARTNER having more difficulty managing money (e.g., paying bills, calculating change, completing tax forms)? 1=Yes ; 2=Maybe ; 3=No
\item \strong{QID30} \emph{numeric}   Range value: 2, ..., 3
\item \strong{QID31} \emph{numeric}   Range value: 2, ..., 3
\item \strong{QID32} \emph{numeric}   Range value: 1, ..., 3
\item \strong{QID33} \emph{numeric}   Range value: 2, ..., 3
\item \strong{QID34} \emph{numeric}   Range value: 1, ..., 3
\item \strong{QID35} \emph{numeric}   Range value: 2, ..., 3
\item \strong{QID8} \emph{numeric} Compared to one year ago, does YOUR STUDY PARTNER show less interest in hobbies and social activities? 1=Yes ; 2=Maybe ; 3=No
\item \strong{QID9} \emph{numeric} Has YOUR STUDY PARTNER ever been diagnosed with Mild Cognitive Impairment (MCI)? 1=Yes ; 2=No ; 3=I don't know
\item \strong{QID10} \emph{numeric} Has YOUR STUDY PARTNER ever been diagnosed with Alzheimer's disease? 1=Yes ; 2=No ; 3=I don't know
\item \strong{QID11} \emph{numeric} Are you concerned that YOU have a memory problem? 1=Yes ; 2=No
\item \strong{QID12} \emph{numeric} Do YOU have trouble remembering if YOU have already told someone something? 1=Yes ; 2=No
\item \strong{QID13} \emph{numeric} Compared to 10 years ago, would you say that there has been a change in YOUR memory? 1=Yes ; 2=No
\item \strong{QID14} \emph{numeric} In general, would you say YOUR health is: 1=Excellent ; 2=Very Good ; 3=Good ; 4=Fair ; 5=Poor
\item \strong{QID15} \emph{numeric} Compared to one year ago, how would you rate YOUR health in general now? 1=Much Better ; 2=Better ; 3=Same ; 4=Worse ; 5=Much Worse
\item \strong{QID16} \emph{numeric} Over the last two months, how often have YOU been bothered by having little interest or pleasure in doing things? 1=Not at all ; 2=Several days ; 3=More than half the days ; 4=Nearly every day
\item \strong{QID17} \emph{numeric} Over the last two months, how often have YOU been bothered by feeling down, depressed, or hopeless? 1=Not at all ; 2=Several days ; 3=More than half the days ; 4=Nearly every day
\item \strong{QID18} \emph{numeric} Stress means a situation in which a person feels tense, restless, nervous or anxious or is unable to sleep at night because his/her mind is troubled all the time. During the past 12 months, would you say YOU experienced: 1=A lot of stress ; 2=A moderate amount of stress ; 3=Relatively little stress ; 4=Almost not stress at all
\item \strong{QID19} \emph{numeric} Have YOU ever been diagnosed with Mild Cognitive Impairment (MCI). 1=Yes ; 2=No ; 3=I don't know
\item \strong{QID20} \emph{numeric} Have YOU ever been diagnosed with Alzheimer's disease? 1=Yes ; 2=No ; 3=I don't know
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
