% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_SP_RELATIONSHIP}
\alias{BHR_SP_RELATIONSHIP}
\title{Brain Health Registry (BHR) SP RELATIONSHIP}
\format{
A data frame with 50 observations and 12 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_SP_RELATIONSHIP)
}
\description{
Brain Health Registry (BHR) SP RELATIONSHIP data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{AboutRID} \emph{numeric} Participant roster ID Range value: 2201, ..., 6644
\item \strong{StudyPartnerID} \emph{numeric} Unique identifier for study partners Range value: 2509, ..., 6051
\item \strong{AboutPTID} \emph{character} Code obtained from CBB CRF report of subject about whom the study partner provided responses  
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{latest} \emph{numeric}   Range value: 0, ..., 1
\item \strong{Relationship} \emph{character} What is your relationship to {{StudyPartnerFirstName}}? Friend ; Grandchild ; Grandparent ; Other Relative ; Paid Caregiver ; Parent ; Roommate ; Sibling ; Son or Daughter ; Son or Daughter In-Law ; Spouse
\item \strong{YearsKnown} \emph{numeric} How long have you known {{StudyPartnerFirstName}}? Range value: 6, ..., 50
\item \strong{LivesTogether} \emph{numeric} Do you and {{StudyPartnerFirstName}} currently live together? Range value: 0, ..., 1
\item \strong{HoursPerWeek} \emph{numeric} How many hours a week do you spend with {{StudyPartnerFirstName}}? Range value: 1, ..., 168
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
