% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BHR_SP_STUDY_CONFIRMATION}
\alias{BHR_SP_STUDY_CONFIRMATION}
\title{Brain Health Registry (BHR) SP STUDY CONFIRMATION}
\format{
A data frame with 40 observations and 9 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BHR_SP_STUDY_CONFIRMATION)
}
\description{
Brain Health Registry (BHR) SP STUDY CONFIRMATION data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{AboutRID} \emph{numeric} Participant roster ID Range value: 2201, ..., 6644
\item \strong{StudyPartnerID} \emph{numeric} Unique identifier for study partners Range value: 3240, ..., 6051
\item \strong{AboutPTID} \emph{character} Code obtained from CBB CRF report of subject about whom the study partner provided responses  
\item \strong{Timepoint} \emph{character} Timepoint in the default ADNI-BHR protocol m00=Baseline ; m03=Month 3 ; m06=Month 6 ; m09=Month 9
\item \strong{CollectedDateTime} \emph{ } UTC DateTime of data collection POSIXct: YYY-MM-DD
\item \strong{CollectedDate_DRVD} \emph{ }   Date: YYY-MM-DD
\item \strong{latest} \emph{numeric}   Range value: 0, ..., 1
\item \strong{QID1} \emph{numeric} Do you currently serve as the study partner in the Alzheimer's Disease Neuroimaging Initiative (ADNI) study? 1=Yes ; 2=No
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
