% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BIOMARK}
\alias{BIOMARK}
\title{Biomarker Samples}
\format{
A data frame with 13988 observations and 66 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BIOMARK)
}
\description{
Biomarker Samples data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Date of Collection Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 3
\item \strong{BIBLOOD} \emph{character} On this date, was blood for biomarkers collected? 
\itemize{
\item \emph{ADNI1, ADNIGO:} 1=Blood
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BIURINE} \emph{character}   1=Urine
\emph{Decoded Value: }\strong{Yes}
\item \strong{BLREASON} \emph{character} If no, please provide reason why blood was not collected 
\itemize{
\item \emph{ADNI2:} 1=Illness ; 2=Participant unavailable ; 3=Participant Unwilling ; 7=Administrative Problems ; 4=Withdrawn consent ; 5=Collection done on a different date for this visit ; 6=Other (specify)
\item \emph{ADNI3:} 1=Illness ; 2=Participant unavailable ; 3=Participant Unwilling ; 4=Withdrawn consent ; 5=Collection done on a different date for this visit ; 6=Other (specify) ; 7=Administrative Problems
\item \emph{ADNI4:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Withdrawn consent ; 5=Collection done on a different date for this visit ; 6=Other (specify) ; 7=Administrative Problems
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BIFAST} \emph{character} Was participant fasting for at least 6 hours? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BITIME} \emph{character} Time of Collection  
\item \strong{BIREDTIME} \emph{character} Serum:  Time Collected  
\item \strong{BIREDAMT} \emph{numeric} Serum:  Amount Collected 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..20
\item \emph{ADNI3, ADNI4} 
}
\item \strong{BIREDCENT} \emph{character} Serum:  Centrifuged Time  
\item \strong{BIREDTRNS} \emph{character} Serum:  Transfer Time  
\item \strong{BIREDVOL} \emph{numeric} Serum: Volume of Serum Transferred 
\itemize{
\item \emph{ADNI1:} 0..13
\item \emph{ADNIGO, ADNI2:} 0..20
\item \emph{ADNI3, ADNI4} 
}
\item \strong{BIREDFROZ} \emph{character} Serum: Time Frozen  
\item \strong{BILAVTIME} \emph{character} Plasma: Time Collected  
\item \strong{BINEEDLE} \emph{character} Needle Used: 1=Sprotte ; 2=Sharp
\emph{Decoded Value: }\strong{Yes}
\item \strong{BILAVAMT} \emph{numeric} Plasma: Amount Collected 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..20
\item \emph{ADNI3, ADNI4} 
}
\item \strong{BILAVCENT} \emph{character} Plasma: Centrifuged Time  
\item \strong{BILAVTRNS} \emph{character} Plasma: Transfer Time  
\item \strong{BILAVVOL} \emph{numeric} Plasma: Volume of Plasma Transferred 
\itemize{
\item \emph{ADNI1:} 0..13
\item \emph{ADNIGO, ADNI2:} 0..20
\item \emph{ADNI3, ADNI4} 
}
\item \strong{BILAVFROZ} \emph{character} Plasma: Time Frozen  
\item \strong{BIURITIME} \emph{character} Time Collected  Urine  
\item \strong{BIURIAMT} \emph{numeric} Amount Collected  Urine 1..100
\item \strong{BIURITRNS} \emph{character} Transfer Time  Urine  
\item \strong{BIURIVOL} \emph{numeric} Volume of Urine Transfered  Urine 1..13
\item \strong{BIURIFROZ} \emph{character} Time Frozen  Urine  
\item \strong{BICSF} \emph{character} On this date, was LP conducted? 
\itemize{
\item \emph{ADNI1, ADNIGO:} 1=CSF
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BINONE} \emph{character}  comment  
\itemize{
\item \emph{ADNI1:} 1=None
\item \emph{ADNIGO:} 1=No samples were collected this visit
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{REASON} \emph{character} If no, please provide reason why the LP was not conducted 
\itemize{
\item \emph{ADNI2:} 1=Illness ; 2=Participant unavailable ; 3=Participant Unwilling ; 8=Administrative Problems ; 4=Withdrawn consent ; 6=Not called for by protocol at this visit ; 7=Collection done on a different date for this visit ; 5=Other (specify)
\item \emph{ADNI3:} 1=Illness ; 2=Participant unavailable ; 3=Participant Unwilling ; 4=Withdrawn consent ; 5=Other (specify) ; 6=Not called for by protocol at this visit ; 7=Collection done on a different date for this visit ; 8=Administrative Problems
\item \emph{ADNI4:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Withdrawn consent ; 5=Other (specify) ; 7=Collection done on a different date for this visit ; 8=Administrative Problems
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BICSFFAST} \emph{character} Was participant fasting for at least 6 hours? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BICSFTIME} \emph{character} Time of LP  
\item \strong{BIMETHOD} \emph{character} Method of Collection 1=gravity ; 2=syringe suction
\emph{Decoded Value: }\strong{Yes}
\item \strong{NEEDLESIZE} \emph{character} Needle Gauge 
\itemize{
\item \emph{ADNI2:} 15=18g Quincke (sharp beveled) needle ; 7=19g Quincke (sharp beveled) needle ; 1=20g Quincke (sharp beveled) needle ; 8=21g Quincke (sharp beveled) needle ; 2=22g Quincke (sharp beveled) needle ; 9=23g Quincke (sharp beveled) needle ; 10=24g Quincke (sharp beveled) needle ; 3=25g Quincke (sharp beveled) needle ; 11=18g Sprotte (atraumatic) needle ; 12=19g Sprotte (atraumatic) needle ; 13=20g Sprotte (atraumatic) needle ; 14=21g Sprotte (atraumatic) needle ; 4=22g Sprotte (atraumatic) needle ; 16=23g Sprotte (atraumatic) needle ; 5=24g Sprotte (atraumatic) needle ; 17=25g Sprotte (atraumatic) needle ; 18=Other (specify)
\item \emph{ADNI3:} 1=20g Quincke (sharp beveled) needle ; 2=22g Quincke (sharp beveled) needle ; 3=25g Quincke (sharp beveled) needle ; 4=22g Sprotte (atraumatic) needle ; 5=24g Sprotte (atraumatic) needle ; 7=19g Quincke (sharp beveled) needle ; 8=21g Quincke (sharp beveled) needle ; 9=23g Quincke (sharp beveled) needle ; 10=24g Quincke (sharp beveled) needle ; 11=18g Sprotte (atraumatic) needle ; 12=19g Sprotte (atraumatic) needle ; 13=20g Sprotte (atraumatic) needle ; 14=21g Sprotte (atraumatic) needle ; 15=18g Quincke (sharp beveled) needle ; 16=23g Sprotte (atraumatic) needle ; 17=25g Sprotte (atraumatic) needle ; 18=Other (specify)
\item \emph{ADNI4:} 1=20g Quincke (sharp beveled) needle ; 10=24g Quincke (sharp beveled) needle ; 11=18g Sprotte (atraumatic) needle ; 12=19g Sprotte (atraumatic) needle ; 13=20g Sprotte (atraumatic) needle ; 14=21g Sprotte (atraumatic) needle ; 15=18g Quincke (sharp beveled) needle ; 16=23g Sprotte (atraumatic) needle ; 17=25g Sprotte (atraumatic) needle ; 18=Other (specify) ; 2=22g Quincke (sharp beveled) needle ; 3=25g Quincke (sharp beveled) needle ; 4=22g Sprotte (atraumatic) needle ; 5=24g Sprotte (atraumatic) needle ; 7=19g Quincke (sharp beveled) needle ; 8=21g Quincke (sharp beveled) needle ; 9=23g Quincke (sharp beveled) needle
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{INTERSPACE} \emph{character} LP performed at the: 
\itemize{
\item \emph{ADNI2:} 4=L1-L2 inter space ; 2=L2-L3 inter space ; 1=L3-L4 inter space ; 5=L4-L5 inter space
\item \emph{ADNI3, ADNI4:} 1=L3-L4 inter space ; 2=L2-L3 inter space ; 4=L1-L2 inter space ; 5=L4-L5 inter space
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTPOSITION} \emph{character} Patient Position 1=Sitting, leaned over (preferred) ; 2=Lying, curled up on side ; 3=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{COLTUBETYP} \emph{character} Type of collection tube used 1=Polypropylene ; 2=Polystyrene
\emph{Decoded Value: }\strong{Yes}
\item \strong{SHPTUBETYP} \emph{character} Type of tube used for shipping 1=Polypropylene ; 2=Polystyrene
\emph{Decoded Value: }\strong{Yes}
\item \strong{TUBEMIN} \emph{numeric} If collected in polystyrene and shipped in polypropylene, please provide estimated amount of time CSF remained in polystyrene collection tube 
\itemize{
\item \emph{ADNI2:} 0..1440
\item \emph{ADNI3, ADNI4} 
}
\item \strong{BICSFAMT} \emph{numeric} CSF:  Amount Collected 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..40
\item \emph{ADNI3, ADNI4} 
}
\item \strong{BICSFTRNS} \emph{character} CSF:  Transfer Time  
\item \strong{BICSFVOL} \emph{numeric} CSF:  Volume of CSF Transferred 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..40
\item \emph{ADNI3, ADNI4} 
}
\item \strong{BICSFFROZ} \emph{character} CSF:  Time Frozen  
\item \strong{BILPPATCH} \emph{character} Check if any of the following was performed 1=Lumbar Puncture Blood Patch
\emph{Decoded Value: }\strong{Yes}
\item \strong{BILPFLURO} \emph{character}   
\itemize{
\item \emph{ADNI1:} 1=Fluroscopy
\item \emph{ADNIGO, ADNI2:} 1=Fluoroscopy
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BILPSPFILM} \emph{character}   1=Lumbar Spine Film
\emph{Decoded Value: }\strong{Yes}
\item \strong{BILPPADATE} \emph{character} Date of Blood Patch  
\item \strong{BILPFLDATE} \emph{character} Date of Fluoroscopy  
\item \strong{BILPSPDATE} \emph{character} Date of Spine Film , with options: 10/10/2006, 10/30/2012, 03/01/2013, 2022-xx-xx, 2020-10-01, 2019-10-29, 2018-xx-xx
\item \strong{BILPPADATEYR_DRVD} \emph{numeric} Year of Blood Patch (Derived from BILPPADATE) Range value: 2006, ..., 2025
\item \strong{BILPFLDATEYR_DRVD} \emph{numeric} Year of Fluoroscopy (Derived from BILPFLDATE) Range value: 2005, ..., 2025
\item \strong{BILPSPDATEYR_DRVD} \emph{numeric} Year of Spine Film (Derived from BILPSPDATE) Range value: 2006, ..., 2022
\item \strong{BILPOTPROC} \emph{character} Check if any of the following was performed 1=Lumbar Puncture Blood Patch ; 2=Fluoroscopy ; 3=Lumbar Spine Film
\emph{Decoded Value: }\strong{Yes}
\item \strong{BIFEDDATE} \emph{ } Date FedExed Date: YYY-MM-DD
\item \strong{ID} \emph{numeric} Record ID Range value: 3, ..., 186725
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{biospecimen}
