% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLCHANGE}
\alias{BLCHANGE}
\title{Diagnostic Summary - Baseline Changes}
\format{
A data frame with 15872 observations and 30 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BLCHANGE)
}
\description{
Diagnostic Summary - Baseline Changes data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Date Form Completed Date: YYY-MM-DD
\item \strong{BCPREDX} \emph{character} Pre-visit Diagnosis 1=NL ; 2=MCI ; 3=AD
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCADAS} \emph{character} 1. Clinically relevant worsening on ADAS? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCMMSE} \emph{character} 2. Clinically relevant worsening on MMSE? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCMMSREC} \emph{character} 3. Clinically relevant worsening on MMSE recall? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCNMMMS} \emph{character} 4. Clinically relevant worsening on non-memory MMSE items? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCNEUPSY} \emph{character} 5. Clinically relevant worsening in memory on neuropsych testing? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCNONMEM} \emph{character} 6. Clinically relevant impairment/worsening in non-memory cognitive domains on neuropsych testing? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCFAQ} \emph{character} 7. Clinically relevant worsening in activities of daily living (FAQ)? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCCDR} \emph{character} 8. Clinically relevant deterioration on CDR Sum of Boxes or Overall CDR rating? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCDEPRES} \emph{character} 9. Clinically relevant depression based on clinical judgement or GDS? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCSTROKE} \emph{character} 10. Did subject have a stroke? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCDELIR} \emph{character} 11. Is there evidence of a delirium (medication effect, toxic or metabolic encephalopathy)? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCEXTCIR} \emph{character} 12. Has extenuating circumstance (such as a physical health problem, change in residence, charge in support network, death of a family member, etc.) contributed to a change in the subject's cognitive or functional performance? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCCORADL} \emph{character} 13. Is the change in clinical status corroborated by informant report of changes in ADL? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No ; -1=NA/No change in clinical status
\item \emph{ADNI3, ADNI4:} -1=NA/No change in clinical status ; 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCCORCOG} \emph{character} 14. Is the change in clinical status corroborated by informant report of changes in cognition? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No ; -1=NA/No change in clinical status
\item \emph{ADNI3, ADNI4:} -1=NA/No change in clinical status ; 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 2, ..., 187119
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: annual, sc, bl
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
