% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLENNOWCSFNFL}
\alias{BLENNOWCSFNFL}
\title{Blennow Lab ADNI1 CSF NFL}
\format{
A data frame with 415 observations and 9 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BLENNOWCSFNFL)
}
\description{
Blennow Lab ADNI1 CSF NFL data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 3, ..., 1423
\item \strong{VISCODE} \emph{character} Visit code , with options: bl
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{VOLUME} \emph{character} Volume , with options: 0.50 ML
\item \strong{CSFNFL} \emph{numeric}   Range value: 405, ..., 12647
\item \strong{COMMENTS} \emph{character} Comments , with options: reddish CSF, samples received thawed
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
