% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLENNOWCSFNG}
\alias{BLENNOWCSFNG}
\title{Blennow Lab ADNI1 CSF Ng}
\format{
A data frame with 415 observations and 9 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BLENNOWCSFNG)
}
\description{
Blennow Lab ADNI1 CSF Ng data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 3, ..., 1423
\item \strong{VISCODE} \emph{character} Visit code , with options: bl
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{CSFNG} \emph{numeric} CSFNG Range value: 0, ..., 2793.11
\item \strong{COMMENTS} \emph{character} Comments , with options: reddish CSF, samples received thawed
\item \strong{GOTCOMMENT} \emph{character}   , with options: Very low level, below detection, Above highest calibrator
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
