% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLENNOWPLASMANFL}
\alias{BLENNOWPLASMANFL}
\title{Blennow Lab ADNI1 Plasma neurofilament light (NFL): 2018-10-03}
\format{
A data frame with 598 observations and 10 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BLENNOWPLASMANFL)
}
\description{
Blennow Lab ADNI1 Plasma neurofilament light (NFL): 2018-10-03 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI4
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 999999
\item \strong{VISCODE} \emph{character} Visit code , with options: bl, sc
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: bl, sc
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{BOX_POS} \emph{character} Box position  
\item \strong{VOL} \emph{character} Sample volume , with options: 0.50 ML
\item \strong{PLASMA_NFL} \emph{numeric} Plasma NFL Range value: 5.2, ..., 250.8
\item \strong{COMMENTS} \emph{character} Comments  
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
