% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLENNOWPLASMANFLLONG}
\alias{BLENNOWPLASMANFLLONG}
\title{Blennow Lab ADNI1-2 Plasma neurofilament light (NFL) longitudinal: 2018-10-03}
\format{
A data frame with 3762 observations and 12 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BLENNOWPLASMANFLLONG)
}
\description{
Blennow Lab ADNI1-2 Plasma neurofilament light (NFL) longitudinal: 2018-10-03 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 2
\item \strong{DRAW_DATE} \emph{ }   Date: YYY-MM-DD
\item \strong{DRAW_TIME} \emph{ }   hms: HH-MM-SS
\item \strong{PLASMA_NFL} \emph{numeric} Plasma neurofilament light (NFL) Range value: 6.4, ..., 451
\item \strong{COMMENT} \emph{character}   , with options: Plasma tube cracked
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
