% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLENNOWPLASMATAU}
\alias{BLENNOWPLASMATAU}
\title{Blennow Lab ADNI1 Plasma Tau}
\format{
A data frame with 581 observations and 9 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BLENNOWPLASMATAU)
}
\description{
Blennow Lab ADNI1 Plasma Tau data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1435
\item \strong{VISCODE} \emph{character} Visit code , with options: bl
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{VOL} \emph{character} Volume , with options: 0.50 ML
\item \strong{PLASMATAU} \emph{numeric}   Range value: 0.08, ..., 17.55
\item \strong{COMMENT} \emph{character} Comments , with options: <lloq, <LOD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
