% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLSCHECK}
\alias{BLSCHECK}
\title{Baseline Symptoms Checklist}
\format{
A data frame with 2550 observations and 40 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BLSCHECK)
}
\description{
Baseline Symptoms Checklist data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code , with options: bl, sc, v01, v06, v03, v31, v11, v41, v51
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: bl, sc, m48, m60, m12, m72, m84, m24, m36, m96
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{BCNAUSEA} \emph{character} 1. Nausea 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCVOMIT} \emph{character} 2. Vomiting 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCDIARRH} \emph{character} 3. Diarrhea 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCCONSTP} \emph{character} 4. Constipation 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCABDOMN} \emph{character} 5. Abdominal discomfort 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCSWEATN} \emph{character} 6. Sweating 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCDIZZY} \emph{character} 7. Dizziness 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCENERGY} \emph{character} 8. Low energy 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCDROWSY} \emph{character} 9. Drowsiness 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCVISION} \emph{character} 10. Blurred vision 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCHDACHE} \emph{character} 11. Headache 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCDRYMTH} \emph{character} 12. Dry mouth 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCBREATH} \emph{character} 13. Shortness of breath 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCCOUGH} \emph{character} 14. Coughing 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCPALPIT} \emph{character} 15. Palpitations 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCCHEST} \emph{character} 16. Chest pain 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCURNDIS} \emph{character} 17. Urinary discomfort (e.g., burning) 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCURNFRQ} \emph{character} 18. Urinary frequency 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCANKLE} \emph{character} 19. Ankle swelling 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCMUSCLE} \emph{character} 20. Musculoskeletal pain 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCRASH} \emph{character} 21. Rash 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCINSOMN} \emph{character} 22. Insomnia 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCDPMOOD} \emph{character} 23. Depressed mood 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCCRYING} \emph{character} 24. Crying 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCELMOOD} \emph{character} 25. Elevated mood 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCWANDER} \emph{character} 26. Wandering 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCFALL} \emph{character} 27. Fall 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCOTHER} \emph{character} 28. Other 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 2982
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{medical_history}
