% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSI}
\alias{BSI}
\title{Boundary Shift Integral Summaries (Nick Fox's lab)}
\format{
A data frame with 3594 observations and 21 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(BSI)
}
\description{
Boundary Shift Integral Summaries (Nick Fox's lab) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{ID} \emph{numeric}   Range value: 1, ..., 3602
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 1435
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character}   , with options: sc, m06, m12, m24, m18, m36, bl, f
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 2
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{IMAGEUID} \emph{numeric} imageUID Range value: 30968, ..., 143906
\item \strong{FLDSTRENG} \emph{numeric} Field Strength Range value: 1.5, ..., 3
\item \strong{BBSI} \emph{numeric} Boundary Shift Integral from baseline Range value: -108.14, ..., 97.315
\item \strong{DBCBBSI} \emph{numeric} Differential Bias Corrected Boundary Shift Integral from baseline Range value: -50.143, ..., 82.892
\item \strong{VBSI} \emph{numeric} Ventricular BSI Range value: -18.088, ..., 29.725
\item \strong{BRAINVOL} \emph{numeric} Segmented brain volume. Not to be used as a measure of intracranial volume. Range value: 680.31, ..., 1400.91
\item \strong{VENTVOL} \emph{numeric} Segmented ventricular volume Range value: 6.715, ..., 179.123
\item \strong{REGCOMMENT} \emph{character}    
\item \strong{REGRATING} \emph{numeric}   1=OK ; 2=less reliable, still a reasonable measure ; 3=Borderline / unreliable, measure affected to some degree, still of some value ; 4=Rejected <ef><bf><bd> measure not reliable at all.  Suggest is not used for analysis ; 5=Non disease related pathology affecting measure, suggest is removed from analysis
\item \strong{VENTACCEPT} \emph{numeric} A boolean to indicate whether or not we have confidence in the VBSI value. Similar purpose to the REGRATING. 0=Vent not accepted ; 1=Vent accepted
\item \strong{KMNDBCBBSI} \emph{numeric} An improved version of the Differential Bias Corrected Boundary Shift Integral from baseline, which: a) Has a new intensity normalisation step that normalises scan intensity by matching the intensities by tissue type (cerebrospinal fluid, grey matter, white mater) instead of being just based on the average whole brain intensity. This helps reduce artefactual change in the BSI that might be caused by change in contrasts between scans (e.g scanner changes) b) Automatically calculates the BSI window based on k-means clustering of tissue intensities for each scan pair Range value: -24.785, ..., 83.473
\item \strong{KMNREGCOMMENT} \emph{character}    
\item \strong{KMNREGRATING} \emph{numeric}   Range value: 1, ..., 8
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
