% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C2N_PRECIVITYAD2_PLASMA}
\alias{C2N_PRECIVITYAD2_PLASMA}
\title{C2N Diagnostics - PrecivityAD2 Blood Test Results for Aβ42/40 and p-tau217/np-tau217 ratios ADNI4}
\format{
A data frame with 676 observations and 19 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(C2N_PRECIVITYAD2_PLASMA)
}
\description{
C2N Diagnostics - PrecivityAD2 Blood Test Results for Aβ42/40 and p-tau217/np-tau217 ratios [ADNI4] data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{character}   , with options: ADNI4
\item \strong{PTID} \emph{character} Patient ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10470
\item \strong{VISCODE} \emph{character} Visit code NA = Not Available
\item \strong{VISCODE2} \emph{character} Translated visit code NA = Not Available
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{Primary} \emph{character} Sample Primary Type BLD = Blood
\item \strong{Additive} \emph{character} Sample Additive Type EDT = Ethylenediaminetetraacetic acid
\item \strong{pT217_C2N} \emph{numeric} C2N Phosphorylated tau217 concentration -4 = Not Available
\item \strong{npT217_C2N} \emph{numeric} C2N Non phosphorylated tau217 concentration -4 = Not Available
\item \strong{AB42_C2N} \emph{numeric} C2N Abeta 42 concentration -4 = Not Available
\item \strong{AB40_C2N} \emph{numeric} C2N Abeta 40 concentration -4 = Not Available
\item \strong{AB42_AB40_C2N} \emph{numeric} C2N Abeta 42/40 ratio -4 = Not Available
\item \strong{pT217_npT217_C2N} \emph{numeric} C2N \% phosphorylated tau217 -4 = Not Available
\item \strong{APS2_C2N} \emph{numeric} Amyloid Probability Score 2 (APS2) result for PrecivityAD2™ -4 = Not Available
\item \strong{APOE_C2N} \emph{character} APOE proteotype , with options: E3/E3, E3/E4, E2/E3, E4/E4, E2/E4, E2/E2, N/A
\item \strong{Comments} \emph{character} Comments LOD = Value is below the limit of quantitation for indicated analyte ; ALQ = Value is above the limit of quantitation for indicated analyte ; QNS = Quantity (sample volume) not sufficient for sample analysis ; Dilution = Sample was tested with dilution, and the test values have been corrected for the dilution factor
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{biospecimen}
