% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{calculate_zscore}
\alias{calculate_zscore}
\title{Calculate Standardized Z-Score}
\usage{
calculate_zscore(x, mean, sd)
}
\arguments{
\item{x}{Numeric value}

\item{mean}{Mean value}

\item{sd}{Standard deviation value}
}
\value{
A numeric vector
}
\description{
Calculate Standardized Z-Score
}
\examples{
\dontrun{
mean <- 10
sd <- 2
x <- rnorm(n = 100, mean = mean, sd = sd)
calculate_zscore(x = x, mean = mean, sd = sd)
calculate_zscore(x = x, mean = 5, sd = 1.5)
calculate_zscore(x = c(1, 0.5, NA, 3), mean = 5, sd = 1.5)
}
}
\seealso{
Other utility functions: 
\code{\link{normalize_var_by_baseline_score}()}
}
\concept{utility functions}
\keyword{utils_fun}
