% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCI}
\alias{CCI}
\title{Cognitive Change Index}
\format{
A data frame with 1228 observations and 34 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(CCI)
}
\description{
Cognitive Change Index data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI3
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric}   Range value: 4691, ..., 7126
\item \strong{VISCODE} \emph{character}   , with options: v01, sc
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: sc
\item \strong{VISDATE} \emph{ } Assessment DATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{CCI1} \emph{character} 1. Recalling information when I really try 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI2} \emph{character} 2. Remembering names and faces of new people I meet 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI3} \emph{character} 3. Remembering things that have happened recently 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI4} \emph{character} 4. Recalling conversations a few days later 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI5} \emph{character} 5. Remembering where things are usually kept 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI6} \emph{character} 6. Remembering new information told to me 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI7} \emph{character} 7. Remembering where I placed familiar objects 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI8} \emph{character} 8. Remembering what I intended to do 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI9} \emph{character} 9. Remembering names of family members and friends 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI10} \emph{character} 10. Remembering without notes and reminders 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI11} \emph{character} 11. People who know me would find that my memory is 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI12} \emph{character} 12. Remembering things compared to my age group 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI13} \emph{character} 13. Making decisions about everyday matters 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI14} \emph{character} 14. Reasoning through a complicated problem 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI15} \emph{character} 15. Focusing on goals and carrying out a plan 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI16} \emph{character} 16. Shifting easily from one activity to the next 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI17} \emph{character} 17. Organizing my daily activities 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI18} \emph{character} 18. Understanding conversations 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI19} \emph{character} 19. Expressing myself when speaking 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI20} \emph{character} 20. Following a story in a book, movie or TV 1=1. Normal Ability (No Change or Better than 5 years ago) ; 2=2. Slight/Occasional Problem (Minimal Change from 5 years ago) ; 3=3. Mild Problem (Some Change from 5 years ago) ; 4=4. Moderate Problem (Clearly Noticeable Change from 5 years ago) ; 5=5. Severe Problem (Much Worse than 5 years ago)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CCI12TOT} \emph{numeric} CCI Index:  Sum of first 12 items 12..60
\item \strong{SOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call/Video Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 185331
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 63
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{assessments}
