% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDR}
\alias{CDR}
\title{Clinical Dementia Rating}
\format{
A data frame with 14617 observations and 26 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(CDR)
}
\description{
Clinical Dementia Rating data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment CDDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{CDSOURCE} \emph{character} Information Source 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Participant Visit ; 2=Telephone Call
\item \emph{ADNI3, ADNI4:} 1=Participant Visit ; 2=Telephone Call/Video Call
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CDVERSION} \emph{character} Version 
\itemize{
\item \emph{ADNI2:} 1=CDR version 1- Full interview with informant and participant ; 3=CDR version 2- Full interview with only informant ; 2=CDR version 3- Abbreviated interview with only informant
\item \emph{ADNI3:} 1=CDR version 1- Full interview with informant and participant ; 2=CDR version 3- Abbreviated interview with only informant ; 3=CDR version 2- Full interview with only informant
\item \emph{ADNI4:} 1=CDR version 1- Full interview with informant and participant ; 2=CDR version 3- Abbreviated (Only to be used if study partner refuses full interview) ; 3=CDR version 2- Full interview with only informant
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SPID} \emph{numeric} Study Partner ID: Range value: 1, ..., 10
\item \strong{CDMEMORY} \emph{numeric} Memory Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI4:} 0.0=0.0 ; 0.5=0.5 ; 1.0=1.0 ; 2.0=2.0 ; 3.0=3.0
\item \emph{ADNI3:} 0=0.0 ; 0.5=0.5 ; 1=1.0 ; 2=2.0 ; 3=3.0
}
\item \strong{CDORIENT} \emph{numeric} Orientation Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI4:} 0.0=0.0 ; 0.5=0.5 ; 1.0=1.0 ; 2.0=2.0 ; 3.0=3.0
\item \emph{ADNI3:} 0=0.0 ; 0.5=0.5 ; 1=1.0 ; 2=2.0 ; 3=3.0
}
\item \strong{CDJUDGE} \emph{numeric} Judgment and Problem Solving Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI4:} 0.0=0.0 ; 0.5=0.5 ; 1.0=1.0 ; 2.0=2.0 ; 3.0=3.0
\item \emph{ADNI3:} 0=0.0 ; 0.5=0.5 ; 1=1.0 ; 2=2.0 ; 3=3.0
}
\item \strong{CDCOMMUN} \emph{numeric} Community Affairs Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI4:} 0.0=0.0 ; 0.5=0.5 ; 1.0=1.0 ; 2.0=2.0 ; 3.0=3.0
\item \emph{ADNI3:} 0=0.0 ; 0.5=0.5 ; 1=1.0 ; 2=2.0 ; 3=3.0
}
\item \strong{CDHOME} \emph{numeric} Home and Hobbies Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI4:} 0.0=0.0 ; 0.5=0.5 ; 1.0=1.0 ; 2.0=2.0 ; 3.0=3.0
\item \emph{ADNI3:} 0=0.0 ; 0.5=0.5 ; 1=1.0 ; 2=2.0 ; 3=3.0
}
\item \strong{CDCARE} \emph{numeric} Personal Care Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI4:} 0.0=0.0 ; 1.0=1.0 ; 2.0=2.0 ; 3.0=3.0
\item \emph{ADNI3:} 0=0.0 ; 1=1.0 ; 2=2.0 ; 3=3.0
}
\item \strong{CDGLOBAL} \emph{numeric} Global CDR 
\itemize{
\item \emph{ADNI1, ADNI4:} 0.0=0.0 ; 0.5=0.5 ; 1.0=1.0 ; 2.0=2.0 ; 3.0=3.0
\item \emph{ADNIGO, ADNI2} 
\item \emph{ADNI3:} 0=0.0 ; 0.5=0.5 ; 1=1.0 ; 2=2.0 ; 3=3.0
}
\item \strong{CDRSB} \emph{numeric} Sum of Boxes 
\itemize{
\item \emph{ADNI3:} 0..18
\item \emph{ADNI4} 
}
\item \strong{ID} \emph{numeric} Record ID Range value: 3, ..., 187113
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
