% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Changed_Study_ID_Listing}
\alias{Changed_Study_ID_Listing}
\title{CHANGED_STUDY_ID_LISTING}
\format{
A data frame with 5 observations and 4 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(Changed_Study_ID_Listing)
}
\description{
CHANGED_STUDY_ID_LISTING data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{SUBJECT} \emph{character}   , with options: 128_S_2009, 012_S_2117, 012_S_2118, 073_S_2154, 100_S_2281
\item \strong{STUDYDATE} \emph{ }   POSIXct: YYY-MM-DD
\item \strong{OLD_STUDYID} \emph{numeric}   Range value: 27482, ..., 31527
\item \strong{NEW_STUDYID} \emph{numeric}   Range value: 27481, ..., 31526
}
}
\keyword{imaging}
\keyword{undefined_phase}
