% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_colnames}
\alias{check_colnames}
\title{Checks Column Name Exist}
\usage{
check_colnames(.data, col_names, strict = FALSE, stop_message = TRUE)
}
\arguments{
\item{.data}{Data.frame}

\item{col_names}{Column names}

\item{strict}{A Boolean value to apply strict checking.}

\item{stop_message}{A Boolean value to return a stop message if the criteria does not met.}
}
\value{
\itemize{
 \item \code{TRUE} if the provided column names are existed in the dataset based on the \code{strict} argument.
 \item Otherwise list of column names that are not existed in the dataset or a stop message if \code{stop_message} is \code{TRUE}.
 }
}
\description{
This function is used to check if the provided column names are existed in the dataset.
}
\examples{
\dontrun{
check_colnames(
  .data = ADNIMERGE2::CDR,
  col_names = c("Phase", "VISCODE"),
  strict = FALSE
)
check_colnames(
  .data = ADNIMERGE2::CDR,
  col_names = c("RID", "VISCODE"),
  strict = TRUE
)
}
}
\seealso{
Other checks function: 
\code{\link{check_duplicate_records}()},
\code{\link{check_non_missing_value}()},
\code{\link{check_object_type}()},
\code{\link{check_value_match}()},
\code{\link{detect_decimal_value}()},
\code{\link{detect_numeric_value}()},
\code{\link{is_datadict_tbl}()}
}
\concept{checks function}
\keyword{utils_fun}
