% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_duplicate_records}
\alias{check_duplicate_records}
\title{Checks Duplicated Records - Internal}
\usage{
check_duplicate_records(
  .data,
  col_names,
  stop_message = TRUE,
  return_duplicate_record = FALSE,
  extra_cols = NULL
)
}
\arguments{
\item{.data}{Data.frame}

\item{col_names}{Character vector of column names}

\item{stop_message}{A Boolean value to return a stop message if there is any duplicated records, Default: \code{TRUE}}

\item{return_duplicate_record}{A Boolean value to return any duplicated records, Default: \code{FALSE}}

\item{extra_cols}{Additional columns that will be returned if \code{stop_message} is \code{FALSE}}
}
\value{
The same data.frame  as input dataset if there is not any duplicated records.
 Otherwise either an assert stop message or duplicated records.
}
\description{
This function is used to check for any duplicated records in a dataset
 based on the combination of provided columns.
}
\examples{
\dontrun{
check_duplicate_records(
  dd = ADNIMERGE2::QS,
  col_names = c("USUBJID", "QSDTC", "QSTESTCD"),
  stop_message = TRUE,
  add_cols = NULL
)
}
}
\seealso{
Other checks function: 
\code{\link{check_colnames}()},
\code{\link{check_non_missing_value}()},
\code{\link{check_object_type}()},
\code{\link{check_value_match}()},
\code{\link{detect_decimal_value}()},
\code{\link{detect_numeric_value}()},
\code{\link{is_datadict_tbl}()}
}
\concept{checks function}
\keyword{utils_fun}
