% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{check_non_missing_value}
\alias{check_non_missing_value}
\title{Check for non-missing value}
\usage{
check_non_missing_value(x)
}
\arguments{
\item{x}{Input value}
}
\value{
A stop error if the value is missing value \code{'NULL'}
}
\description{
Check for non-missing value
}
\examples{
\dontrun{
check_non_missing_value(x = LETTERS[1:10])
check_non_missing_value(x = NULL)
}
}
\seealso{
Other checks function: 
\code{\link{check_colnames}()},
\code{\link{check_duplicate_records}()},
\code{\link{check_object_type}()},
\code{\link{check_value_match}()},
\code{\link{detect_decimal_value}()},
\code{\link{detect_numeric_value}()},
\code{\link{is_datadict_tbl}()}
}
\concept{checks function}
\keyword{utils_fun}
