% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-assert.R
\name{check_object_type}
\alias{check_object_type}
\title{Check for Object Type Mismatch}
\usage{
check_object_type(x, type)
}
\arguments{
\item{x}{Input Object}

\item{type}{Object type}
}
\value{
An error message if the input object and type does not match.
}
\description{
Check for Object Type Mismatch
}
\details{
This function is used to check the object type of an input value based on
\code{base} R package. In \code{base} R package, object type functions are
specified in \code{is.ObjectType} format. For instance, \code{base::is.logical}
is used for checking a logical/Boolean value. Please see the examples how it
is translated in this function.
}
\examples{
\dontrun{
check_object_type(x = letters[1:12], type = "character")
check_object_type(x = ADNIMERGE2::CDR, type = "data.frame")
check_object_type(x = TRUE, type = "logical")
# Error message
check_object_type(x = letters[1:12], type = "numeric")
check_object_type(x = letters[1:12], type = "factor")
}
}
\seealso{
Other checks function: 
\code{\link{check_colnames}()},
\code{\link{check_duplicate_records}()},
\code{\link{check_non_missing_value}()},
\code{\link{check_value_match}()},
\code{\link{detect_decimal_value}()},
\code{\link{detect_numeric_value}()},
\code{\link{is_datadict_tbl}()}
}
\concept{checks function}
\keyword{utils_fun}
