% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment-utils.R
\name{check_overall_phase}
\alias{check_overall_phase}
\title{Checks for Overall and PHASE List Combination}
\usage{
check_overall_phase(phase)
}
\arguments{
\item{phase}{Phase argument}
}
\value{
An error if a combination of 'Overall' and phase list character vector is provided.
}
\description{
Checks for Overall and PHASE List Combination
}
\keyword{internal}
