% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_value_match}
\alias{check_value_match}
\title{Checks Variable Value Matching}
\usage{
check_value_match(
  values,
  check_list,
  excluded.na = TRUE,
  stop_message = FALSE,
  add_stop_message = NULL,
  value_split = FALSE,
  split_pattern = "\\\\||:|;"
)
}
\arguments{
\item{values}{Existed variable values}

\item{check_list}{Vector of input values}

\item{excluded.na}{A Boolean to skip \code{NA} from the existed variable values \code{values},
Default: \code{TRUE}}

\item{stop_message}{A Boolean value to return a stop message if one of the existed values
does not match with the check list, Default: \code{FALSE}}

\item{add_stop_message}{Additional text message that will be added in the stop message.}

\item{value_split}{A Boolean value whether to split the values with specified split pattern \code{split_pattern}}

\item{split_pattern}{Split string pattern. Only applicable if \code{value_split = TRUE}}
}
\value{
\itemize{
  \item {\code{TRUE}}: If all the existed variable values are matched with the input values
  \item {\code{FALSE}}: Otherwise and with a stop message if \code{stop_message} is \code{TRUE}
}
}
\description{
This function is used to check if the variable values are the same as
 the input values.
}
\examples{
\dontrun{
check_value_match(
  values = c("-2", "2"),
  check_list = c("-2"),
  stop_message = FALSE
)
check_value_match(
  values = c("-2", "2", NA),
  check_list = c("-2", "2"),
  excluded.na = TRUE,
  stop_message = FALSE
)
}
}
\seealso{
Other checks function: 
\code{\link{check_colnames}()},
\code{\link{check_duplicate_records}()},
\code{\link{check_non_missing_value}()},
\code{\link{check_object_type}()},
\code{\link{detect_decimal_value}()},
\code{\link{detect_numeric_value}()},
\code{\link{is_datadict_tbl}()}
}
\concept{checks function}
\keyword{internal}
