% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_value_mapping}
\alias{collect_value_mapping}
\title{Collecting Coded Values for Multiple Variables}
\usage{
collect_value_mapping(.datadic, tbl_name, all_fld_name)
}
\arguments{
\item{.datadic}{Data dictionary dataset created using \code{\link{get_factor_levels_datadict}()} function}

\item{tbl_name}{Dataset name}

\item{all_fld_name}{List of column names that contains re-coded values}
}
\value{
List value that contains the \code{coded} and \code{decoded} values with corresponding
 column names that obtained using \code{\link{collect_value_mapping_single_var}} function.
}
\description{
This function is used to collect the coded values for a given set of variables.
}
\examples{
\dontrun{
data_dict_dd <- get_factor_levels_datadict(
  .datadic = ADNIMERGE2::DATADIC,
  tbl_name = "CDR",
  nested_value = TRUE
)
tbl_unique_fldname <- get_factor_fldname(
  .datadic = data_dict_dd,
  tbl_name = "CDR",
  dd_fldnames = colnames(ADNIMERGE2::CDR)
)
collect_value_mapping(
  .datadic = data_dict_dd,
  tbl_name = "CDR",
  all_fld_name = tbl_unique_fldname
)
}
}
\seealso{
\code{\link{get_factor_levels_datadict}()} \code{\link{get_factor_fldname}()}

Other data dictionary codelist related functions: 
\code{\link{collect_value_mapping_single_var}()},
\code{\link{convert_value_mapping}()},
\code{\link{convert_value_mapping_phase_specific}()}
}
\concept{data dictionary codelist related functions}
\keyword{adni_datadic_fun}
