% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_value_mapping_single_var}
\alias{collect_value_mapping_single_var}
\title{Collecting Variable Coded Values}
\usage{
collect_value_mapping_single_var(.datadic, tbl_name, fld_name)
}
\arguments{
\item{.datadic}{Data dictionary dataset created using \code{\link{get_factor_levels_datadict}()} function}

\item{tbl_name}{Dataset name}

\item{fld_name}{Field/column name}
}
\value{
A list value:
\itemize{
 \item \emph{code}: Coded values
 \item \emph{decode}: Values that will replace the coded values (\code{code})
 }
}
\description{
This function is used to collect the coded and decoded values for a given
 variable based on a data dictionary \code{\link{DATADIC}()}.
}
\examples{
\dontrun{
data_dict_dd <- get_factor_levels_datadict(
  .datadic = ADNIMERGE2::DATADIC,
  tbl_name = "ADAS",
  nested_value = TRUE
)
collect_value_mapping_single_var(
  .datadic = data_dict_dd,
  tbl_name = "ADAS",
  fld_name = "COCOMND"
)
}
}
\seealso{
\code{\link{get_factor_levels_datadict}()}

Other data dictionary codelist related functions: 
\code{\link{collect_value_mapping}()},
\code{\link{convert_value_mapping}()},
\code{\link{convert_value_mapping_phase_specific}()}
}
\concept{data dictionary codelist related functions}
\keyword{adni_datadic_fun}
