% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{compute_baseline_score_summary}
\alias{compute_baseline_score_summary}
\title{Compute Baseline Grouped Score/Numeric Summary Statistic}
\usage{
compute_baseline_score_summary(
  .data,
  filterBy,
  filterValue = c("Y", "Yes", "bl"),
  ...
)
}
\arguments{
\item{.data}{Data.frame}

\item{filterBy}{Character vector of baseline record identifier variable}

\item{filterValue}{Baseline record identifier values, Default: c("Y", "Yes", "bl")}

\item{...}{\code{\link{compute_score_summary}()} arguments}
}
\value{
Similar to \code{\link{compute_score_summary}()} result
}
\description{
A wrapper function to calculate baseline grouped summary statistic of
numeric variable(s).
}
\examples{
\dontrun{
# For long format data
# Suppose we wanted to compute the baseline summary score of available
# assessment score in 'ADNIMERGE2::ADQS' by baseline diagnosis status.
library(tidyverse)
library(ADNIMERGE2)

compute_baseline_score_summary(
  .data = ADNIMERGE2::ADQS,
  filterBy = "ABLFL",
  filterValue = "Y",
  wideFormat = FALSE,
  scoreVar = "AVAL",
  groupVar1 = "PARAMCD",
  groupVar = "DX",
  filterGroup = NULL
)

compute_baseline_score_summary(
  .data = ADNIMERGE2::ADQS,
  filterBy = "ABLFL",
  filterValue = "Y",
  wideFormat = FALSE,
  scoreVar = "AVAL",
  groupVar1 = "PARAMCD",
  groupVar = "DX",
  filterGroup = "CN"
)
}
}
\seealso{
\code{\link{compute_score_summary}()}
}
\keyword{pacc_score_utils_fun}
\keyword{utils_fun}
