% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-function.R
\name{compute_neurobat_subscore}
\alias{compute_neurobat_subscore}
\title{Scoring Function for NEUROBAT Sub-items}
\usage{
compute_neurobat_subscore(.neurobat)
}
\arguments{
\item{.neurobat}{Data.frame of \code{\link{NEUROBAT}()} eCRF}
}
\value{
A data frame the same as \code{.neurobat} with the following appended columns
  \item{LIMMTOTL }{Logical Memory - Immediate Recall Score: Between 0 and 25}
  \item{LDELTOTL }{Logical Memory - Delayed Recall Score: Between 0 and 25}
  \item{DIGITSCR }{?? Digit Symbol Substitution}
  \item{TRABSCOR }{Time to Complete Trail B Making Test Score}
  \item{RAVLTIMM }{Rey Auditory Verbal Learning Test - Immediate Score: Sum of all five trials result}
  \item{RAVLTLRN }{Rey Auditory Verbal Learning Test - Learning Score: Difference between the fifth and the first trial result}
  \item{RAVLTFG }{Rey Auditory Verbal Learning Test - Forgetting Score: Difference between the fifth trial total score and delayed time}
  \item{RAVLTFGP }{Rey Auditory Verbal Learning Test - Forgetting Percentage Score: Relative percentage of forgetting score from the fifth trial result}
}
\description{
This function is used to compute item-level subscore in \code{\link{NEUROBAT}()} eCRF.
 Please see the \code{Value} section for item-level score description.
}
\references{
[Powell, J.B., Cripe, L.I. and Dodrill, C.B., 1991. Assessment of brain impairment with the Rey Auditory Verbal Learning Test: A comparison with other neuropsychological measures. Archives of Clinical Neuropsychology, 6(4), pp.241-249.](https://doi.org/10.1016/0887-6177(91)90001-P)
}
\concept{scoring functions}
\keyword{adni_scoring_fun}
