% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{compute_pacc_score}
\alias{compute_pacc_score}
\title{Compute ADNI modified versions of the Preclinical Alzheimer's Cognitive Composite (PACC)}
\usage{
compute_pacc_score(
  .data,
  bl.summary,
  componentVars,
  rescale_trialsB = FALSE,
  keepComponents = FALSE,
  wideFormat = TRUE,
  varName = NULL,
  scoreCol = NULL,
  idCols = NULL
)
}
\arguments{
\item{.data}{A data.frame either in wide or long format. Please see the other arguments.}

\item{bl.summary}{Baseline component score summary
 It can be created either using \code{\link{compute_score_summary}()} function.
 Or a data.frame of component score summary that contains the following variables:
\itemize{
  \item {\code{VAR}}: Contains PACC component variable names
  \item {\code{MEAN}}: Mean score
  \item {\code{SD}}: Standard deviation value
}
 Often recommended to use the baseline component score that summarized by
 baseline diagnostics status.}

\item{componentVars}{Character vector of component score variable names.
The component score variable names should be arranged based the following order.
Otherwise, \strong{invalid} composite score will be calculated.

\itemize{
  \item Delayed Recall portion from ADAS Cognitive Behavior assessment (ADAS-Cog), see \code{\link{ADAS}()}
  \item Mini-Mental State Examination Score, see \code{\link{MMSE}()}
  \item Logical Memory IIa Delayed Recall Score, see \code{LDELTOTAL} score in \code{\link{NEUROBAT}()}
  \item Digit Symbol Substitution Test Score, see \code{DIGITSCOR} score in \code{\link{NEUROBAT}()}
  \item Trails B Score, see see \code{TRABSCOR} score in \code{\link{NEUROBAT}()}
}}

\item{rescale_trialsB}{A Boolean value to change the \code{Trails B} score in log scale, Default: TRUE}

\item{keepComponents}{A Boolean to keep component score, Default: FALSE}

\item{wideFormat}{A Boolean value whether the data.frame is in \code{wide} or \code{long} format, Default: TRUE}

\item{varName}{Column name that contain the component score names for long format data, Default = NULL.
Only applicable for a \code{long} format data and \code{varName} must not be missing if \code{wideFormat} is \code{FALSE}.}

\item{scoreCol}{Variable names that contains component score/numeric value for long format data, Default = NULL.
Only applicable for a \code{long} format data and \code{scoreCol} must not be missing if \code{wideFormat} is \code{FALSE}.}

\item{idCols}{Character vector of ID columns for long format data, Default: NULL.
Only applicable for a \code{long} format data and \code{idCols} must not be missing if \code{wideFormat} is \code{FALSE}.}
}
\value{
\itemize{
\item For a {\code{wide}} format input data: A \code{data.frame} with appended columns for \code{mPACCdigit} and \code{mPACCtrailsB}.
\item For a {\code{long}} format input data: A \code{data.frame} with additional rows of \code{mPACCdigit} and \code{mPACCtrailsB}.
}
}
\description{
The Preclinical Alzheimer's Cognitive Composite (\href{https://doi.org/10.1001/jamaneurol.2014.803}{PACC})
is a baseline standardized composite of

\itemize{
  \item Free and Cued Selective Reminding Test (FCSRT)
  \item Logical Memory IIa Delayed Recall (LM)
  \item Digit Symbol Substitution Test (DSST)
  \item Mini-Mental State Examination (MMSE)
}

See \code{Details} section for more information.
}
\details{
This function generates two modified versions of PACC scores based on ADNI data. FCSRT is not used in ADNI, so we use the
Delayed Recall portion of the Alzheimer's Disease Assessment Scale (ADAS) as a proxy. Score \code{mPACCdigit} uses the
DSST and only computed for \code{ADNI1} study phase. \code{mPACCtrailsB} uses (log transformed) Trails B as a proxy for DSST. Raw component scores
standardized according to the mean and standard deviation of baseline scores of ADNI subjects with normal cognition
to create \code{Z} scores for each component \code{(Z=(raw - mean(raw.bl))/sd(raw.bl))}.
The Z scores are reoriented if necessary so that greater scores reflect better performance.
The composite is the sum of these Z scores.

\strong{Missing components:} At least two components must be present to produce a score.
If more than two components are missing, the PACC will be \code{NA}.

\strong{mPACCdigit} only will be calculated for the study phase when DSST was collected, which is in ADNI1 study phase.
 Otherwise the \code{mPACCdigit} score will be missing (i.e., \code{NA}) even though the remaining PACC component scores are non-missing.

Please see \code{vignette(topic = "ADNIMERGE2-PACC-SCORE", package = "ADNIMERGE2")}
how \code{\link{compute_pacc_score}()} function can be used.
}
\examples{
\dontrun{
# Please see 'Details' section or 'ADNIMERGE2-PACC-SCORE' vignette
vignette(topic = "ADNIMERGE2-PACC-SCORE", package = "ADNIMERGE2")

# Additional examples about PACC score -----
library(nlme)
library(dplyr)
library(multcomp)
library(Hmisc)
library(ADNIMERGE)

csf2numeric <- function(x) {
  as.numeric(gsub("<", "", gsub(">", "", x)))
}

dd <- subset(adnimerge, DX.bl \%in\% c("CN", "SMC") & !is.na(mPACCtrailsB)) \%>\%
  mutate(
    ABETA = csf2numeric(ABETA)
  )

# identify those with elevated PIB PET at ANY visit OR
# elevated BASELINE AV45 PET OR elevated BASELINE FBB PET OR
# low BASELINE CSF Abeta
# AV45 ~ PIB regression from Landau et al 2012:
elevatedAmyloid <- unique(c(
  subset(dd, 0.67 * PIB + 0.15 > 1.11)$RID,
  subset(dd, VISCODE == "bl" & AV45 > 1.11)$RID,
  subset(dd, VISCODE == "bl" & FBB > 1.08)$RID,
  subset(dd, VISCODE == "bl" & ABETA < 900)$RID
))
anyAmyloid <- unique(subset(dd, !is.na(AV45.bl) | !is.na(PIB) |
  !is.na(FBB.bl) | !is.na(ABETA.bl))$RID)

dd <- dd \%>\%
  mutate(
    ElevatedAmyloid = ifelse(RID \%in\% elevatedAmyloid, 1,
      ifelse(RID \%in\% anyAmyloid, 0, NA)
    ),
    m = Month.bl,
    m2 = Month.bl^2,
    APOEe4 = APOE4 > 0
  )

summary(
  ElevatedAmyloid ~ APOEe4 + AGE + PTEDUCAT,
  data = dd,
  subset = VISCODE == "bl",
  method = "reverse",
  overall = TRUE
)

# Quadratic time model:
fit <- lme(
  fixed = mPACCtrailsB ~ mPACCtrailsB.bl + APOEe4 + AGE +
    PTEDUCAT + m + m2 + (m + m2):ElevatedAmyloid,
  random = ~ m | RID,
  data = dd,
  na.action = na.omit
)

Months <- seq(12, 96, 12)
elevated.design <- model.matrix(
  mPACCtrailsB ~ mPACCtrailsB.bl + APOEe4 + AGE + PTEDUCAT +
    m + m2 + (m + m2):ElevatedAmyloid,
  data = data.frame(
    mPACCtrailsB = 0, mPACCtrailsB.bl = 0, APOEe4 = TRUE, AGE = 75,
    PTEDUCAT = 12, ElevatedAmyloid = 1, m = Months, m2 = Months^2
  )
)

normal.design <- model.matrix(
  mPACCtrailsB ~ mPACCtrailsB.bl + APOEe4 + AGE + PTEDUCAT +
    m + m2 + (m + m2):ElevatedAmyloid,
  data = data.frame(
    mPACCtrailsB = 0, mPACCtrailsB.bl = 0, APOEe4 = TRUE, AGE = 75,
    PTEDUCAT = 12, ElevatedAmyloid = 0, m = Months, m2 = Months^2
  )
)

contrast.data <- elevated.design - normal.design
summary(multcomp::glht(fit, linfct = contrast.data))
}
}
\references{
\itemize{
  \item Donohue MC, et al. The Preclinical Alzheimer Cognitive Composite: Measuring Amyloid-Related Decline. \emph{JAMA Neurol}. 2014;71(8):961–970. doi:10.1001/jamaneurol.2014.803 \url{http://dx.doi.org/10.1001/jamaneurol.2014.803}
  \item Donohue MC, Sperling RA, Petersen R, Sun C, Weiner MW, Aisen PS, for the Alzheimer’s Disease Neuroimaging Initiative. Association Between Elevated Brain Amyloid and Subsequent Cognitive Decline Among Cognitively Normal Persons. \emph{JAMA}. 2017;317(22):2305–2316. \url{http://dx.doi.org/10.1001/jama.2017.6669}
}
}
\seealso{
\code{vignette(topic = "ADNIMERGE2-PACC-SCORE", package = "ADNIMERGE2")}
}
\author{
Michael Donohue \email{mdonohue@usc.edu}
}
\keyword{adni_scoring_fun}
\keyword{pacc_score_utils_fun}
