% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{compute_score_summary}
\alias{compute_score_summary}
\title{Compute Grouped Numeric Summary Statistic}
\usage{
compute_score_summary(
  .data,
  wideFormat = TRUE,
  scoreVar,
  groupVar = "DX",
  filterGroup = NULL,
  groupVar1 = NULL
)
}
\arguments{
\item{.data}{Data.frame}

\item{wideFormat}{A Boolean value whether the input data.frame is in a \code{wide} or \code{long} format, Default: TRUE}

\item{scoreVar}{Character vector of variable(s) that contain the actual score/numeric values

When \code{wideFormat} is \code{FALSE} (i.e., for a long format data), \code{scoreVar}
must be a length of one character vector of variable name that contains
the score/numeric values.}

\item{groupVar}{Group variable, Default: 'DX'}

\item{filterGroup}{Filter value of group variable \code{groupVar}, Default: NULL
Only applicable if the \code{groupVar} is a length of one character vector.}

\item{groupVar1}{Additional grouping variable, only applicable for \code{long} format data.frame.}
}
\value{
A data.frame with the following columns:

\itemize{
  \item \code{groupVar}: Grouping variable
  \item \code{VAR}: Score/numeric variable name
  \item \code{N}: Number of non-missing observation
  \item \code{MEAN}: Mean score
  \item \code{SD}: Standard deviation value
}
}
\description{
This function is used to compute the numeric variable summary statistic grouped
 by categorical variable.
}
\details{
All computed summary statistic are based on non-missing observation.
 The result summary will be filter by the corresponding \code{filterGroup} value(s).
}
\examples{
\dontrun{
# For long format data
# Suppose we wanted to compute the baseline summary score of
# all available assessments in \code{ADNIMERGE2::ADQS}

# By baseline diagnosis status.
library(tidyverse)
library(ADNIMERGE2)

long_format_example <- ADNIMERGE2::ADQS \%>\%
  filter(ABLFL \%in\% "Y") \%>\%
  #  Check there is only one baseline record per assessment type per subject
  ADNIMERGE2::assert_uniq(USUBJID, PARAMCD)

compute_score_summary(
  .data = long_format_example,
  wideFormat = FALSE,
  scoreVar = "AVAL",
  groupVar1 = "PARAMCD",
  groupVar = "DX",
  filterGroup = NULL
)

# For only cognitive normal (CN) subjects
compute_score_summary(
  .data = long_format_example,
  wideFormat = FALSE,
  scoreVar = "AVAL",
  groupVar1 = "PARAMCD",
  groupVar = "DX",
  filterGroup = "CN"
)

# For wide format data
# Suppose we wanted to compute the baseline summary statistic of
# \code{AGE}, \code{BMI}, \code{ADASTT11} and \code{ADASTT13}
wide_format_example <- ADNIMERGE2::ADSL \%>\%
  filter(ENRLFL \%in\% "Y")

# By baseline diagnostics status
compute_score_summary(
  .data = wide_format_example,
  wideFormat = TRUE,
  scoreVar = c("AGE", "BMI", "ADASTT11", "ADASTT13"),
  groupVar1 = NULL,
  groupVar = "DX",
  filterGroup = NULL
)


compute_score_summary(
  .data = wide_format_example,
  wideFormat = TRUE,
  scoreVar = c("AGE", "BMI", "ADASTT11", "ADASTT13"),
  groupVar1 = NULL,
  groupVar = "DX",
  filterGroup = "CN"
)

# By SEX
compute_score_summary(
  .data = wide_format_example,
  wideFormat = TRUE,
  scoreVar = c("AGE", "BMI", "ADASTT11", "ADASTT13"),
  groupVar1 = NULL,
  groupVar = "SEX",
  filterGroup = NULL
)
}
}
\seealso{
\code{\link{compute_baseline_score_summary}()}
 \code{vignette(topic = "ADNIMERGE2-PACC-SCORE", package = "ADNIMERGE2")}
}
\keyword{pacc_score_utils_fun}
\keyword{utils_fun}
