% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONSENTS}
\alias{CONSENTS}
\title{Consent Tracking}
\format{
A data frame with 1936 observations and 24 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(CONSENTS)
}
\description{
Consent Tracking data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI3, ADNI2, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10900
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_sc, 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{CONEVNT} \emph{character} Was consent obtained? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{CONTYPE} \emph{character} Indicate consent type 1=New Enrollee ; 2=Rollover ; 3=Telephone-only ; 4=Early Frames Addendum Study
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCDATE} \emph{ } Date consent(s) signed: Date: YYY-MM-DD
\item \strong{SITEVERS} \emph{ } Site revised version: Date: YYY-MM-DD
\item \strong{FUTURE} \emph{character} Consent to contact for Future Studies 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes ; 2=n/a
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MRI} \emph{character} Consent to MRI Scan procedures 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes ; 2=n/a
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMY} \emph{character} Consent to Amyloid PET Scan procedures 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes ; 2=n/a
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{AMYREAD} \emph{character} Consent to Amyloid PET Scan Result 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{TAU} \emph{character} Consent to Tau PET Scan procedures 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes ; 2=n/a
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{LPPROC} \emph{character} Consent to Lumbar Puncture procedures 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes ; 2=n/a
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 9, ..., 128024
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{enrollment}
