% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_adni_phase_order_num}
\alias{convert_adni_phase_order_num}
\title{Convert ADNI Study Phase Order Number}
\usage{
convert_adni_phase_order_num(phase_num)
}
\arguments{
\item{phase_num}{Variable that represents the ADNI study phase order number}
}
\value{
A character vector with list of ADNI study phases.
}
\description{
This function is used to convert/replace the order number of ADNI study phase
with the actual study phase name.
}
\examples{
\dontrun{
convert_adni_phase_order_num(phase_num = c(1, 3, 5))
}
}
\seealso{
Other ADNI study protocol/phase: 
\code{\link{adni_phase}()},
\code{\link{adni_phase_order_num}()},
\code{\link{adni_pttype}()},
\code{\link{adni_study_track}()},
\code{\link{create_col_protocol}()},
\code{\link{create_orig_protocol}()},
\code{\link{original_study_protocol}()}
}
\concept{ADNI study protocol/phase}
\keyword{adni_procotol_fun}
