% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-data-prep.R
\name{convert_number_days}
\alias{convert_number_days}
\title{Convert Number of Days}
\usage{
convert_number_days(
  x,
  unit = "month",
  bin = 1,
  digits = 2L,
  adjust_negative_value = FALSE
)
}
\arguments{
\item{x}{Numeric vector}

\item{unit}{Unit of measurements: either \code{week}, \code{month} or \code{year}, Default: 'month'}

\item{bin}{Bin value, Default: 1}

\item{digits}{Rounding decimal digits, Default: 52}

\item{adjust_negative_value}{Indicator to adjust for negative value}
}
\value{
A numeric vector
}
\description{
A function to convert number of days into either number of \code{weeks},
 \code{months} or \code{years} with specific bin values.
}
\details{
A small positive values will be added to zero numeric value, specially to
 adjust time interval for baseline visits.
 Furthermore, all negative value will be ceiling to zero if \code{adjust_negative_value} is set to \code{TRUE}
}
\examples{
\dontrun{
set.seed(123456)
x <- rnorm(100, mean = 100, sd = 10)
# Convert number of days into number of weeks
convert_number_days(x = x, unit = "week", bin = 1, digits = 1L)
# Convert number of days into number of months binned into 3 months
convert_number_days(x = x, unit = "month", bin = 3, digits = 2L)
# Convert number of days into number of years
convert_number_days(x = x, unit = "year", bin = 1, digits = 5L)
}
}
\seealso{
Other Utility functions used for preparing analysis data: 
\code{\link{get_numeric_dx_status}()}
}
\concept{Utility functions used for preparing analysis data}
\keyword{utils_fun}
