% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_missing_value}
\alias{convert_to_missing_value}
\title{Convert Value into Missing Values}
\usage{
convert_to_missing_value(
  .data,
  col_name = NULL,
  value = "-4",
  missing_char = NA,
  phase = NULL
)
}
\arguments{
\item{.data}{Data.frame}

\item{col_name}{Variable name}

\item{value}{Specific value that will be considered as missing value, Default: '-4'}

\item{missing_char}{Character for missing value, Default: NA}

\item{phase}{Phase-specific value replacement, Default: NULL for all ADNI phases}
}
\value{
A data frame with replaced value.
}
\description{
This function is used to convert specific value into missing values.
}
\examples{
\dontrun{
convert_to_missing_value(.data = ADNIMERGE2::ADAS)
}
}
\seealso{
Other functions to replace values: 
\code{\link{replace_values_dataset}()},
\code{\link{replace_values_phase_specific}()},
\code{\link{replace_values_single_var}()}
}
\concept{functions to replace values}
\keyword{adni_replace_fun}
\keyword{utils_fun}
