% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_usubjid_to_rid}
\alias{convert_usubjid_to_rid}
\title{Convert USUBJID to RID Format}
\usage{
convert_usubjid_to_rid(x, studyID = "ADNI")
}
\arguments{
\item{x}{Character vector of ID in \code{STUDYID-SITEID-RID} format}

\item{studyID}{Study ID, Default: 'ADNI'}
}
\value{
A numeric vector of \code{RID}
}
\description{
Convert USUBJID to RID Format
}
\examples{
\dontrun{
random_ids <- sample(1:10000, size = 10)
random_siteid <- sample(1:60, size = 1)
example_usubjid <- create_usubjid(x = random_ids, y = create_usubjid)
example_rid <- convert_usubjid_to_rid(x = example_usubjid)
# Should be similar to random_ids values
example_rid
}
}
\seealso{
Other ID format functions: 
\code{\link{create_usubjid}()}
}
\concept{ID format functions}
\keyword{adni_id_fun}
