% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_value_mapping}
\alias{convert_value_mapping}
\title{Convert List Collected Coded Values}
\usage{
convert_value_mapping(coded_values, tbl_name = NULL)
}
\arguments{
\item{coded_values}{Variable name that contains coded values}

\item{tbl_name}{Dataset name, Default: NULL}
}
\value{
A data frame that contains the following columns:
 \item{TBLNAME }{Dataset name}
 \item{FLDNAME }{Variable name}
 \item{PHASE }{ADNI study phase/protocol}
 \item{CODES }{Nested dataset that contains code and decode values}
}
\description{
This function is used to convert list coded values that is created using
 \code{\link{collect_value_mapping}()} function.
}
\examples{
\dontrun{
data_dict_dd <- get_factor_levels_datadict(
  .datadic = ADNIMERGE2::DATADIC,
  tbl_name = "CDR",
  nested_value = TRUE
)
tbl_unique_fldname <- get_factor_fldname(
  .datadic = data_dict_dd,
  tbl_name = "CDR",
  dd_fldnames = colnames(ADNIMERGE2::CDR)
)
coded_value <- collect_value_mapping(
  .datadic = data_dict_dd,
  tbl_name = "CDR",
  all_fld_name = tbl_unique_fldname
)
convert_value_mapping(coded_value = coded_value)
}
}
\seealso{
Other data dictionary codelist related functions: 
\code{\link{collect_value_mapping}()},
\code{\link{collect_value_mapping_single_var}()},
\code{\link{convert_value_mapping_phase_specific}()}
}
\concept{data dictionary codelist related functions}
\keyword{adni_datadic_fun}
