% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_value_mapping_phase_specific}
\alias{convert_value_mapping_phase_specific}
\title{Convert List Collected Coded Values - PHASE Levels}
\usage{
convert_value_mapping_phase_specific(coded_values)
}
\arguments{
\item{coded_values}{List object of coded values with study phase name}
}
\value{
A data.frame that contains the following columns:
\itemize{
  \item \emph{FLDNAME:} Variable name
  \item \emph{PHASE:} ADNI study phase/protocol
  \item \emph{CODED_VALUE:} Nested dataset that contains code and decode values
 }
}
\description{
This function is used to convert list coded values at phase levels.
}
\seealso{
Other data dictionary codelist related functions: 
\code{\link{collect_value_mapping}()},
\code{\link{collect_value_mapping_single_var}()},
\code{\link{convert_value_mapping}()}
}
\concept{data dictionary codelist related functions}
\keyword{adni_datadic_fun}
