% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_col_protocol}
\alias{create_col_protocol}
\title{Create COLPROT Column}
\usage{
create_col_protocol(.data, phaseVar = NULL, .strict_check = TRUE)
}
\arguments{
\item{.data}{Data.frame}

\item{phaseVar}{Phase column}

\item{.strict_check}{A Boolean value to apply strict check-in for missing or known values in \code{COLPOROT} variables}
}
\value{
A data.frame the same as \code{data} with appended column of \code{COLPROT}.
}
\description{
This function is used to create COLPROT column in the dataset by renaming any
specified columns. It also checks whether the renamed columns contains only ADNI study phase.
}
\examples{
\dontrun{
create_col_protocol(
  data = ADNIMERGE2::REGISTRY,
  phaseVar = "Phase",
  .strict_check = TRUE
)
create_col_protocol(
  data = ADNIMERGE2::VISITS,
  phaseVar = "Phase",
  .strict_check = FALSE
)
}
}
\seealso{
Other ADNI study protocol/phase: 
\code{\link{adni_phase}()},
\code{\link{adni_phase_order_num}()},
\code{\link{adni_pttype}()},
\code{\link{adni_study_track}()},
\code{\link{convert_adni_phase_order_num}()},
\code{\link{create_orig_protocol}()},
\code{\link{original_study_protocol}()}
}
\concept{ADNI study protocol/phase}
\keyword{adni_procotol_fun}
