% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_orig_protocol}
\alias{create_orig_protocol}
\title{Create ORIGPROT Column}
\usage{
create_orig_protocol(.data)
}
\arguments{
\item{.data}{Data.frame that contains \code{RID} variable}
}
\value{
A data frame the same as \code{.data} with appended column of \code{ORIGPROT}.
}
\description{
This function is used to create ORIGPROT in the dataset based subject RID.
}
\examples{
\dontrun{
example_data <- tibble(RID = c(1, 1000, 3500, 6645, 1000))
create_orig_protocol(data = example_data)
}
}
\seealso{
Other ADNI study protocol/phase: 
\code{\link{adni_phase}()},
\code{\link{adni_phase_order_num}()},
\code{\link{adni_pttype}()},
\code{\link{adni_study_track}()},
\code{\link{convert_adni_phase_order_num}()},
\code{\link{create_col_protocol}()},
\code{\link{original_study_protocol}()}
}
\concept{ADNI study protocol/phase}
\keyword{adni_procotol_fun}
