% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_usubjid}
\alias{create_usubjid}
\title{Create USUBJID from RID and SITEID}
\usage{
create_usubjid(x, y)
}
\arguments{
\item{x}{RID variable}

\item{y}{SITEID variable}
}
\value{
A character vector of the same size as the input args with \code{STUDYID-} prefix.
}
\description{
This function is used to create USUBJID from RID and SITEID.
}
\seealso{
Other ID format functions: 
\code{\link{convert_usubjid_to_rid}()}
}
\concept{ID format functions}
\keyword{adni_id_fun}
