% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CROSSVAL}
\alias{CROSSVAL}
\title{CROSSVAL}
\format{
A data frame with 822 observations and 6 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(CROSSVAL)
}
\description{
CROSSVAL data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 1435
\item \strong{GROUP_NAME} \emph{character}    
\item \strong{SET_ID} \emph{character}   , with options: c, e, g, d, j, b, a, f, h, i
\item \strong{TRAINING} \emph{numeric}   Range value: 0, ..., 1
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
