% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CSSRSAD}
\alias{CSSRSAD}
\title{Columbia- Suicide Severity Rating Scale (C-SSRS) + AD Specific Suicide}
\format{
A data frame with 359 observations and 19 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(CSSRSAD)
}
\description{
Columbia- Suicide Severity Rating Scale (C-SSRS) + AD Specific Suicide data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI2, ADNI3, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 69, ..., 10875
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_bl, 4_init, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{WISH} \emph{character} In the last 6 months, have you wished you were dead or wished you could go to sleep and not wake up? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{ACT} \emph{character} In the last 6 months, have you actually had thoughts of killing yourself? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{HRMACQAD} \emph{character} Would you harm yourself if you learned you had Alzheimer's Disease? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{HRMRISKAD} \emph{character} Would you harm yourself if you learned that you had an increased risk of Alzheimer's disease based on your PET scan result? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 1366, ..., 127911
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
