% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATA_DOWNLOADED_DATE}
\alias{DATA_DOWNLOADED_DATE}
\title{ADNI Study Data Downloaded Date}
\format{
A `Date` class object.
}
\usage{
data(DATA_DOWNLOADED_DATE)
}
\description{
The date when data in this package were downloaded from \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}.
}
\examples{
\dontrun{
ADNIMERGE2::DATA_DOWNLOADED_DATE
}
}
\keyword{datasets}
\keyword{other_raw_dataset}
