% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATADIC}
\alias{DATADIC}
\title{DATADIC}
\format{
A data frame with 36898 observations and 14 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(DATADIC)
}
\description{
DATADIC . More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{DATADIC_SOURCE} \emph{character}    
\item \strong{PHASE} \emph{character}    
\item \strong{CRFNAME} \emph{character}    
\item \strong{TBLNAME} \emph{character}    
\item \strong{FLDNAME} \emph{character}    
\item \strong{TEXT} \emph{character}    
\item \strong{TYPE} \emph{character}    
\item \strong{LENGTH} \emph{character}    
\item \strong{DD_CRF_VERSION} \emph{character}    
\item \strong{CODE} \emph{character}    
\item \strong{UNITS} \emph{character}    
\item \strong{STATUS} \emph{character}    
\item \strong{CODE_CHANGES} \emph{character}    
\item \strong{MAPPING_NOTES} \emph{character}    
}
}
\keyword{data_dict}
