% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{datadict_as_tibble}
\alias{datadict_as_tibble}
\title{Convert \code{datadict_tbl} class object into \code{tibble} object}
\usage{
datadict_as_tibble(.data)
}
\arguments{
\item{.data}{Input data}
}
\value{
A tibble object
}
\description{
Convert \code{datadict_tbl} class object into \code{tibble} object
}
\keyword{adni_datadic_fun}
