% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DERIVED_DATADIC}
\alias{DERIVED_DATADIC}
\title{[DERIVED] Data dictionary for derived dataset}
\format{
A data frame with 469 observations and 5 variables.
}
\source{
For more details see the help vignettes
}
\usage{
data(DERIVED_DATADIC)
}
\description{
Data dictionary for derived dataset derived dataset.
}
\details{
\itemize{
\item \strong{TBLNAME} \emph{character}    
\item \strong{CRFNAME} \emph{character}    
\item \strong{FLDNAME} \emph{character}    
\item \strong{LABEL} \emph{character}    
\item \strong{TEXT} \emph{character}    
}
}
\keyword{derived_dataset}
