% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-data-prep.R
\name{detect_baseline_score}
\alias{detect_baseline_score}
\title{Detect Closest Baseline Score}
\usage{
detect_baseline_score(cur_record_date, enroll_date, time_interval = 30)
}
\arguments{
\item{cur_record_date}{Date of current assessment record collected}

\item{enroll_date}{Enrollment date (i.e. baseline visit date)}

\item{time_interval}{Minimum window period (in days) from baseline visit date, Default: 30}
}
\value{
A character vector with the same length as \code{cur_record_date} and
 contain a flag value, \code{Yes}, for the closest record
 within the specified window period. Otherwise, missing value.
}
\description{
This function is used to flag the closest assessment record score to the
 baseline visit date (i.e., enrollment date) within a certain window period.
}
\concept{ADNI enrollment}
\keyword{utils_fun}
