% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_decimal_value}
\alias{detect_decimal_value}
\title{Detect Decimal Values}
\usage{
detect_decimal_value(value)
}
\arguments{
\item{value}{Input string}
}
\value{
A Boolean value
}
\description{
This function is used to detect decimal values.
}
\examples{
\dontrun{
# No decimal value is presented: return FALSE
detect_decimal_value(1:5)
detect_decimal_value(NA)

# Detect a decimal value: return TRUE
detect_decimal_value(seq(1, 5, by = 0.5))

# Detect a decimal value in a mixed string: return TRUE
detect_decimal_value(c("1", "2.0", "3", "999"))
}
}
\seealso{
Other checks function: 
\code{\link{check_colnames}()},
\code{\link{check_duplicate_records}()},
\code{\link{check_non_missing_value}()},
\code{\link{check_object_type}()},
\code{\link{check_value_match}()},
\code{\link{detect_numeric_value}()},
\code{\link{is_datadict_tbl}()}
}
\concept{checks function}
\keyword{utils_fun}
