% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_numeric_value}
\alias{detect_numeric_value}
\title{Detect Numeric Values}
\usage{
detect_numeric_value(value, num_type = "any", stop_message = FALSE)
}
\arguments{
\item{value}{Input string}

\item{num_type}{Numeric value type: \code{"any"}, \code{"positive"} or \code{"negative"}, Default: 'any'}

\item{stop_message}{A Boolean value to return a stop message when the specified numeric
value type is detected \code{num_type}, Default: \code{FALSE}}
}
\value{
A Boolean value or a stop message if \code{stop_message} is \code{TRUE}:
  \item{TRUE }{If any of the specified numeric type value is presented.}
  \item{FALSE }{Otherwise}
}
\description{
This function is used to detect any numeric values from a string.
}
\examples{
\dontrun{
string1 <- c("-2", "1", "1", "-1")
string2 <- c("ADNI1", "ADNI2", "ADNIGO")
detect_numeric_value(value = string1, num_type = "any")
detect_numeric_value(value = string2, num_type = "any")
detect_numeric_value(value = string1, num_type = "negative")
detect_numeric_value(value = string1, num_type = "positive", stop_message = TRUE)
}
}
\seealso{
Other checks function: 
\code{\link{check_colnames}()},
\code{\link{check_duplicate_records}()},
\code{\link{check_non_missing_value}()},
\code{\link{check_object_type}()},
\code{\link{check_value_match}()},
\code{\link{detect_decimal_value}()},
\code{\link{is_datadict_tbl}()}
}
\concept{checks function}
\keyword{utils_fun}
