% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DM}
\alias{DM}
\title{[DERIVED]  Derived  Demographic Characteristics}
\format{
A data frame with 5146 observations and 24 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(DM)
}
\description{
Demographic characteristics derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbreviation , with character options: DM
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{SUBJID} \emph{character} Subject Identifier for the Study  
\item \strong{ORIGPROT} \emph{character} Original study protocol First time participation in ADNI study phase; , with character options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{SESTDTC} \emph{Date} Screening Start Date/Time First screening start date/time; Date: YYY-MM-DD
\item \strong{RFSTDTC} \emph{Date} Subject Enrolled Date/Time * Date: YYY-MM-DD
\item \strong{RFPENDTC} \emph{character} Date/Time of End of Participation* Last known date/time when participant ended thier participation or follow-up in ADNI study. As character format for unknown MM-DD values.;  
\item \strong{DTHDTC} \emph{numeric} Date/Time of Death Range value: 2006, ..., 2025
\item \strong{DTHFL} \emph{character} Subject Death Flag , with character options: Yes
\item \strong{SITEID} \emph{numeric} Study Site Identifier Range value: 1, ..., 941
\item \strong{AGE} \emph{numeric} Age Age at first screening visit (in years); Range value: 50.5, ..., 91.4
\item \strong{AGEU} \emph{character} Age Units , with character options: Years
\item \strong{SEX} \emph{character} Sex at Birth , with character options: Female, Male
\item \strong{RACE} \emph{character} Race  
\item \strong{ETHNIC} \emph{character} Ethnicity , with character options: Not Hispanic or Latino, Hispanic or Latino, Unknown
\item \strong{ARMCD} \emph{character} Planned Arm Code  
\item \strong{ARM} \emph{character} Description of Planned Arm  
\item \strong{ACTARMCD} \emph{character} Actual Arm Code  
\item \strong{ACTARM} \emph{character} Actual Treatment Description  
\item \strong{ARMNRS} \emph{character} Reason Arm and/or Actual Arm is Null Non-Interventional Study; , with character options: Non-Interventional Study
\item \strong{COUNTRY} \emph{character} Country Country of the investigational site located;  
\item \strong{DMDTC} \emph{Date} Date/Time of Collection Date: YYY-MM-DD
\item \strong{DMDY} \emph{integer} Study Day of Collection Range value: -363, ..., 256
}
}
\keyword{derived_dataset}
