% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DTIROI_MEAN}
\alias{DTIROI_MEAN}
\title{USC - DTI ROI Summary Measures v2 + v3 (Mean)}
\format{
A data frame with 1750 observations and 310 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(DTIROI_MEAN)
}
\description{
USC - DTI ROI Summary Measures v2 + v3 (Mean) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI2, ADNI3, ADNIGO, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 11, ..., 10793
\item \strong{VISCODE} \emph{character} Visit code , with options: init, y2, y1, sc, y4, y3, 4_sc, 4_init, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{character} Preprocessing pipeline version 2 = preprocessing pipeline as outlined in methods PDF: (USC - DTI ROI Summary Measures Methods v2 [ADNI3], date: 12/01/2021) ; 3 = preprocessing pipeline as outlined in methods PDF: (USC - DTI ROI Summary Measures Methods v3 [ADNI4], date: 04/02/2025) ; -4 = image was not processed as noted in QC column.
\emph{Decoded Value: }\strong{Yes}
\item \strong{MANUFACTURER} \emph{character} Scanner Manufacturer , with options: Siemens, GE, Philips, SIEMENS, GE MEDICAL SYSTEMS
\item \strong{VOLUMES} \emph{numeric} Number of volumes prescribed by acquisition protocol Range value: 31, ..., 127
\item \strong{IMAGEUID} \emph{numeric} IMAGE ID of the raw diffusion scan Range value: 808826, ..., 11303224
\item \strong{DISTORTION_CORRECTION} \emph{character} In V3, EPI susceptibility distortion correction was run using a synthetic b0 (Synb0), b0s acquired with opposite phase encoding directions (APPA), or was completed in scanner (scanner), in addition to warping DWIs to T1w. In V2, susceptibility distortion was only corrected by warping DWIs to T1w (empty cells). Synb0=synthetic b0 ; APPA=b0s acquired with opposite phase encoding directions ; scanner=completed in scanner.
\item \strong{VOLUMES_AP} \emph{numeric} Number of volumes in the opposite phase encoding direction raw scan. Empty cell indicates there was no opposite phase encoding image provided to users on the IDA. Range value: 7, ..., 13
\item \strong{IMAGEUID_AP} \emph{numeric} IMAGE ID of the raw diffusion scan with an opposite phase encoding direction. Empty cell indicates there was no opposite phase encoding image provided to users on the IDA. Range value: 10248884, ..., 11297753
\item \strong{RUNDATE} \emph{ } Processing Date Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Processing Completed , with options: complete, not processed
\item \strong{QC} \emph{character} Preprocessed DWI and scalar map quality check notes include: (1) a note if raw DWI were not processed due to incomplete acquisition (e.g., fewer volumes acquired than protocol), failed DICOM conversion, poor field of view (i.e., image cropped), or severe artifacts (e.g., motion, distortion); or (2) notes if preprocessed DWI had any remaining artifacts (e.g., motion, distortion, ghosting, hyperintensities in FA, etc.), pathology or anatomical anomalies, atlas registration failure, or if the voxel size differed from the prescribed acquisition protocol. These images should be used at own discretion.  
\item \strong{AD_MCP} \emph{numeric} Middle cerebellar peduncle (mean) Range value: 0.000864373, ..., 0.00140574
\item \strong{AD_PCT} \emph{numeric} Pontine crossing tract (mean) Range value: 0.000635765, ..., 0.00120459
\item \strong{AD_GCC} \emph{numeric} Genu of corpus callosum (mean) Range value: 0.00108743, ..., 0.00174352
\item \strong{AD_BCC} \emph{numeric} Body of corpus callosum (mean) Range value: 0.00117643, ..., 0.00183443
\item \strong{AD_SCC} \emph{numeric} Splenium of corpus callosum (mean) Range value: 0.000952219, ..., 0.00187696
\item \strong{AD_FX} \emph{numeric} Fornix (mean) Range value: 0.00136849, ..., 0.00336679
\item \strong{AD_CST_R} \emph{numeric} Corticospinal tract right (mean) Range value: 0.000697005, ..., 0.00124421
\item \strong{AD_CST_L} \emph{numeric} Corticospinal tract left (mean) Range value: 0.000743352, ..., 0.00128856
\item \strong{AD_ML_R} \emph{numeric} Medial lemniscus right (mean) Range value: 0.000853256, ..., 0.00153265
\item \strong{AD_ML_L} \emph{numeric} Medial lemniscus left (mean) Range value: 0.00091114, ..., 0.00144778
\item \strong{AD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right (mean) Range value: 0.000773077, ..., 0.00137231
\item \strong{AD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left (mean) Range value: 0.000799225, ..., 0.00135735
\item \strong{AD_SCP_R} \emph{numeric} Superior cerebellar peduncle right (mean) Range value: 0.00109985, ..., 0.00219184
\item \strong{AD_SCP_L} \emph{numeric} Superior cerebellar peduncle left (mean) Range value: 0.000948561, ..., 0.00211851
\item \strong{AD_CP_R} \emph{numeric} Cerebral peduncle right (mean) Range value: 0.000725959, ..., 0.00176159
\item \strong{AD_CP_L} \emph{numeric} Cerebral peduncle left (mean) Range value: 0.000786702, ..., 0.0016027
\item \strong{AD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right (mean) Range value: 0.00087333, ..., 0.00160752
\item \strong{AD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left (mean) Range value: 0.000899174, ..., 0.00162506
\item \strong{AD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right (mean) Range value: 0.000959815, ..., 0.00153946
\item \strong{AD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left (mean) Range value: 0.000917365, ..., 0.00151523
\item \strong{AD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right (mean) Range value: 0.000959267, ..., 0.00154893
\item \strong{AD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left (mean) Range value: 0.000914247, ..., 0.0015178
\item \strong{AD_ACR_R} \emph{numeric} Anterior corona radiata right (mean) Range value: 0.000875194, ..., 0.00150636
\item \strong{AD_ACR_L} \emph{numeric} Anterior corona radiata left (mean) Range value: 0.000922371, ..., 0.00151156
\item \strong{AD_SCR_R} \emph{numeric} Superior corona radiata right (mean) Range value: 0.00090299, ..., 0.00159924
\item \strong{AD_SCR_L} \emph{numeric} Superior corona radiata left (mean) Range value: 0.000925883, ..., 0.0015824
\item \strong{AD_PCR_R} \emph{numeric} Posterior corona radiata right (mean) Range value: 0.00100159, ..., 0.00176965
\item \strong{AD_PCR_L} \emph{numeric} Posterior corona radiata left (mean) Range value: 0.000922813, ..., 0.00169638
\item \strong{AD_PTR_R} \emph{numeric} Posterior thalamic radiation right (mean) Range value: 0.00105155, ..., 0.00156374
\item \strong{AD_PTR_L} \emph{numeric} Posterior thalamic radiation left (mean) Range value: 0.00106455, ..., 0.00167979
\item \strong{AD_SS_R} \emph{numeric} Sagittal stratum right (mean) Range value: 0.00102514, ..., 0.00157003
\item \strong{AD_SS_L} \emph{numeric} Sagittal stratum left (mean) Range value: 0.000949998, ..., 0.00172715
\item \strong{AD_EC_R} \emph{numeric} External capsule right (mean) Range value: 0.000869933, ..., 0.00139585
\item \strong{AD_EC_L} \emph{numeric} External capsule left (mean) Range value: 0.000911379, ..., 0.00139066
\item \strong{AD_CGC_R} \emph{numeric} Cingulum right (mean) Range value: 0.000803926, ..., 0.00125218
\item \strong{AD_CGC_L} \emph{numeric} Cingulum left (mean) Range value: 0.000773778, ..., 0.00126339
\item \strong{AD_CGH_R} \emph{numeric} Cingulum right (mean) Range value: 0.000789796, ..., 0.00136406
\item \strong{AD_CGH_L} \emph{numeric} Cingulum left (mean) Range value: 0.000674585, ..., 0.00148286
\item \strong{AD_FXST_R} \emph{numeric} Fornix (cres) / Stria terminalis right (mean) Range value: 0.000902871, ..., 0.00181092
\item \strong{AD_FXST_L} \emph{numeric} Fornix (cres) / Stria terminalis left (mean) Range value: 0.000844339, ..., 0.00146839
\item \strong{AD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right (mean) Range value: 0.000917611, ..., 0.00136946
\item \strong{AD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left (mean) Range value: 0.000934935, ..., 0.00134069
\item \strong{AD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right (mean) Range value: 0.000864122, ..., 0.00185215
\item \strong{AD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left (mean) Range value: 0.000872557, ..., 0.00200136
\item \strong{AD_UNC_R} \emph{numeric} Uncinate fasciculus right (mean) Range value: 0.000828013, ..., 0.00164766
\item \strong{AD_UNC_L} \emph{numeric} Uncinate fasciculus left (mean) Range value: 0.000783775, ..., 0.00156054
\item \strong{AD_TAP_R} \emph{numeric} Tapetum right (mean) Range value: 0.00118081, ..., 0.00264749
\item \strong{AD_TAP_L} \emph{numeric} Tapetum left (mean) Range value: 0.00124293, ..., 0.00283334
\item \strong{AD_CST} \emph{numeric} Corticospinal tract Bilateral (mean) Range value: 0.000746304, ..., 0.0012439
\item \strong{AD_ML} \emph{numeric} Medial lemniscus Bilateral (mean) Range value: 0.000892504, ..., 0.00147262
\item \strong{AD_ICP} \emph{numeric} Inferior cerebellar peduncle Bilateral (mean) Range value: 0.000810984, ..., 0.00136443
\item \strong{AD_SCP} \emph{numeric} Superior cerebellar peduncle Bilateral (mean) Range value: 0.00104909, ..., 0.00209897
\item \strong{AD_CP} \emph{numeric} Cerebral peduncle Bilateral (mean) Range value: 0.000789979, ..., 0.00168134
\item \strong{AD_ALIC} \emph{numeric} Anterior limb of internal capsule Bilateral (mean) Range value: 0.000916326, ..., 0.00161495
\item \strong{AD_PLIC} \emph{numeric} Posterior limb of internal capsule Bilateral (mean) Range value: 0.000939453, ..., 0.00152027
\item \strong{AD_RLIC} \emph{numeric} Retrolenticular part of internal capsule Bilateral (mean) Range value: 0.0010485, ..., 0.00150684
\item \strong{AD_ACR} \emph{numeric} Anterior corona radiata Bilateral (mean) Range value: 0.000911011, ..., 0.00148156
\item \strong{AD_SCR} \emph{numeric} Superior corona radiata Bilateral (mean) Range value: 0.000926379, ..., 0.00159025
\item \strong{AD_PCR} \emph{numeric} Posterior corona radiata Bilateral (mean) Range value: 0.00101048, ..., 0.00173321
\item \strong{AD_PTR} \emph{numeric} Posterior thalamic radiation Bilateral (mean) Range value: 0.0010662, ..., 0.00155257
\item \strong{AD_SS} \emph{numeric} Sagittal stratum Bilateral (mean) Range value: 0.00102692, ..., 0.0015771
\item \strong{AD_EC} \emph{numeric} External capsule Bilateral (mean) Range value: 0.000901249, ..., 0.00137195
\item \strong{AD_CGC} \emph{numeric} Cingulum Bilateral (mean) Range value: 0.000787912, ..., 0.00125077
\item \strong{AD_CGH} \emph{numeric} Cingulum Bilateral (mean) Range value: 0.00073583, ..., 0.00137865
\item \strong{AD_FXST} \emph{numeric} Fornix (cres) / Stria terminalis Bilateral (mean) Range value: 0.000897928, ..., 0.00164401
\item \strong{AD_SLF} \emph{numeric} Superior longitudinal fasciculus Bilateral (mean) Range value: 0.000926227, ..., 0.00135512
\item \strong{AD_SFO} \emph{numeric} Superior fronto-occipital fasciculus Bilateral (mean) Range value: 0.000884636, ..., 0.00186439
\item \strong{AD_UNC} \emph{numeric} Uncinate fasciculus Bilateral (mean) Range value: 0.000812439, ..., 0.00161605
\item \strong{AD_TAP} \emph{numeric} Tapetum Bilateral (mean) Range value: 0.00131883, ..., 0.00267621
\item \strong{AD_IC} \emph{numeric} Full internal capsule (mean) Range value: 0.000962232, ..., 0.00150003
\item \strong{AD_CC} \emph{numeric} Full corpus callosum (mean) Range value: 0.00109255, ..., 0.00180409
\item \strong{AD_CR} \emph{numeric} Full corona radiata (mean) Range value: 0.000973624, ..., 0.00153466
\item \strong{AD_FULLWM} \emph{numeric} Full white matter (mean) Range value: 0.00101086, ..., 0.00149077
\item \strong{FA_MCP} \emph{numeric} Middle cerebellar peduncle (mean) Range value: 0.252345, ..., 0.550161
\item \strong{FA_PCT} \emph{numeric} Pontine crossing tract (mean) Range value: 0.210801, ..., 0.516748
\item \strong{FA_GCC} \emph{numeric} Genu of corpus callosum (mean) Range value: 0.294705, ..., 0.647485
\item \strong{FA_BCC} \emph{numeric} Body of corpus callosum (mean) Range value: 0.275639, ..., 0.672372
\item \strong{FA_SCC} \emph{numeric} Splenium of corpus callosum (mean) Range value: 0.311704, ..., 0.744461
\item \strong{FA_FX} \emph{numeric} Fornix (mean) Range value: 0.0611964, ..., 0.526047
\item \strong{FA_CST_R} \emph{numeric} Corticospinal tract right (mean) Range value: 0.229988, ..., 0.65296
\item \strong{FA_CST_L} \emph{numeric} Corticospinal tract left (mean) Range value: 0.22369, ..., 0.64481
\item \strong{FA_ML_R} \emph{numeric} Medial lemniscus right (mean) Range value: 0.281006, ..., 0.728448
\item \strong{FA_ML_L} \emph{numeric} Medial lemniscus left (mean) Range value: 0.286499, ..., 0.780731
\item \strong{FA_ICP_R} \emph{numeric} Inferior cerebellar peduncle right (mean) Range value: 0.199086, ..., 0.592228
\item \strong{FA_ICP_L} \emph{numeric} Inferior cerebellar peduncle left (mean) Range value: 0.202831, ..., 0.611582
\item \strong{FA_SCP_R} \emph{numeric} Superior cerebellar peduncle right (mean) Range value: 0.261885, ..., 0.718028
\item \strong{FA_SCP_L} \emph{numeric} Superior cerebellar peduncle left (mean) Range value: 0.270221, ..., 0.776648
\item \strong{FA_CP_R} \emph{numeric} Cerebral peduncle right (mean) Range value: 0.357331, ..., 0.786855
\item \strong{FA_CP_L} \emph{numeric} Cerebral peduncle left (mean) Range value: 0.364529, ..., 0.78557
\item \strong{FA_ALIC_R} \emph{numeric} Anterior limb of internal capsule right (mean) Range value: 0.317535, ..., 0.615578
\item \strong{FA_ALIC_L} \emph{numeric} Anterior limb of internal capsule left (mean) Range value: 0.256225, ..., 0.607676
\item \strong{FA_PLIC_R} \emph{numeric} Posterior limb of internal capsule right (mean) Range value: 0.504614, ..., 0.71658
\item \strong{FA_PLIC_L} \emph{numeric} Posterior limb of internal capsule left (mean) Range value: 0.420518, ..., 0.742459
\item \strong{FA_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right (mean) Range value: 0.367615, ..., 0.750679
\item \strong{FA_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left (mean) Range value: 0.296874, ..., 0.665738
\item \strong{FA_ACR_R} \emph{numeric} Anterior corona radiata right (mean) Range value: 0.21618, ..., 0.475171
\item \strong{FA_ACR_L} \emph{numeric} Anterior corona radiata left (mean) Range value: 0.204649, ..., 0.460778
\item \strong{FA_SCR_R} \emph{numeric} Superior corona radiata right (mean) Range value: 0.281018, ..., 0.561446
\item \strong{FA_SCR_L} \emph{numeric} Superior corona radiata left (mean) Range value: 0.293459, ..., 0.581761
\item \strong{FA_PCR_R} \emph{numeric} Posterior corona radiata right (mean) Range value: 0.264524, ..., 0.518704
\item \strong{FA_PCR_L} \emph{numeric} Posterior corona radiata left (mean) Range value: 0.247135, ..., 0.522745
\item \strong{FA_PTR_R} \emph{numeric} Posterior thalamic radiation right (mean) Range value: 0.291212, ..., 0.628348
\item \strong{FA_PTR_L} \emph{numeric} Posterior thalamic radiation left (mean) Range value: 0.294805, ..., 0.589033
\item \strong{FA_SS_R} \emph{numeric} Sagittal stratum right (mean) Range value: 0.258145, ..., 0.609787
\item \strong{FA_SS_L} \emph{numeric} Sagittal stratum left (mean) Range value: 0.293284, ..., 0.615108
\item \strong{FA_EC_R} \emph{numeric} External capsule right (mean) Range value: 0.214441, ..., 0.465525
\item \strong{FA_EC_L} \emph{numeric} External capsule left (mean) Range value: 0.202631, ..., 0.486719
\item \strong{FA_CGC_R} \emph{numeric} Cingulum right (mean) Range value: 0.199259, ..., 0.536063
\item \strong{FA_CGC_L} \emph{numeric} Cingulum left (mean) Range value: 0.184967, ..., 0.547625
\item \strong{FA_CGH_R} \emph{numeric} Cingulum right (mean) Range value: 0.118283, ..., 0.481906
\item \strong{FA_CGH_L} \emph{numeric} Cingulum left (mean) Range value: 0.10786, ..., 0.466886
\item \strong{FA_FXST_R} \emph{numeric} Fornix (cres) / Stria terminalis right (mean) Range value: 0.210284, ..., 0.544872
\item \strong{FA_FXST_L} \emph{numeric} Fornix (cres) / Stria terminalis left (mean) Range value: 0.235789, ..., 0.592987
\item \strong{FA_SLF_R} \emph{numeric} Superior longitudinal fasciculus right (mean) Range value: 0.301397, ..., 0.538221
\item \strong{FA_SLF_L} \emph{numeric} Superior longitudinal fasciculus left (mean) Range value: 0.274124, ..., 0.525935
\item \strong{FA_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right (mean) Range value: 0.258072, ..., 0.505397
\item \strong{FA_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left (mean) Range value: 0.213813, ..., 0.480173
\item \strong{FA_UNC_R} \emph{numeric} Uncinate fasciculus right (mean) Range value: 0.142727, ..., 0.558439
\item \strong{FA_UNC_L} \emph{numeric} Uncinate fasciculus left (mean) Range value: 0.162396, ..., 0.563396
\item \strong{FA_TAP_R} \emph{numeric} Tapetum right (mean) Range value: 0.197727, ..., 0.573132
\item \strong{FA_TAP_L} \emph{numeric} Tapetum left (mean) Range value: 0.134764, ..., 0.503668
\item \strong{FA_CST} \emph{numeric} Corticospinal tract Bilateral (mean) Range value: 0.226778, ..., 0.648897
\item \strong{FA_ML} \emph{numeric} Medial lemniscus Bilateral (mean) Range value: 0.283881, ..., 0.745101
\item \strong{FA_ICP} \emph{numeric} Inferior cerebellar peduncle Bilateral (mean) Range value: 0.201763, ..., 0.585665
\item \strong{FA_SCP} \emph{numeric} Superior cerebellar peduncle Bilateral (mean) Range value: 0.268691, ..., 0.718652
\item \strong{FA_CP} \emph{numeric} Cerebral peduncle Bilateral (mean) Range value: 0.366346, ..., 0.786166
\item \strong{FA_ALIC} \emph{numeric} Anterior limb of internal capsule Bilateral (mean) Range value: 0.28777, ..., 0.606564
\item \strong{FA_PLIC} \emph{numeric} Posterior limb of internal capsule Bilateral (mean) Range value: 0.473246, ..., 0.708232
\item \strong{FA_RLIC} \emph{numeric} Retrolenticular part of internal capsule Bilateral (mean) Range value: 0.362365, ..., 0.658326
\item \strong{FA_ACR} \emph{numeric} Anterior corona radiata Bilateral (mean) Range value: 0.216347, ..., 0.454851
\item \strong{FA_SCR} \emph{numeric} Superior corona radiata Bilateral (mean) Range value: 0.287231, ..., 0.571456
\item \strong{FA_PCR} \emph{numeric} Posterior corona radiata Bilateral (mean) Range value: 0.25717, ..., 0.516042
\item \strong{FA_PTR} \emph{numeric} Posterior thalamic radiation Bilateral (mean) Range value: 0.329049, ..., 0.609791
\item \strong{FA_SS} \emph{numeric} Sagittal stratum Bilateral (mean) Range value: 0.277118, ..., 0.594567
\item \strong{FA_EC} \emph{numeric} External capsule Bilateral (mean) Range value: 0.210151, ..., 0.464564
\item \strong{FA_CGC} \emph{numeric} Cingulum Bilateral (mean) Range value: 0.191668, ..., 0.536745
\item \strong{FA_CGH} \emph{numeric} Cingulum Bilateral (mean) Range value: 0.113071, ..., 0.460723
\item \strong{FA_FXST} \emph{numeric} Fornix (cres) / Stria terminalis Bilateral (mean) Range value: 0.222603, ..., 0.557311
\item \strong{FA_SLF} \emph{numeric} Superior longitudinal fasciculus Bilateral (mean) Range value: 0.297422, ..., 0.530435
\item \strong{FA_SFO} \emph{numeric} Superior fronto-occipital fasciculus Bilateral (mean) Range value: 0.240942, ..., 0.472103
\item \strong{FA_UNC} \emph{numeric} Uncinate fasciculus Bilateral (mean) Range value: 0.152682, ..., 0.549497
\item \strong{FA_TAP} \emph{numeric} Tapetum Bilateral (mean) Range value: 0.187994, ..., 0.531757
\item \strong{FA_IC} \emph{numeric} Full internal capsule (mean) Range value: 0.418186, ..., 0.646093
\item \strong{FA_CC} \emph{numeric} Full corpus callosum (mean) Range value: 0.303971, ..., 0.68485
\item \strong{FA_CR} \emph{numeric} Full corona radiata (mean) Range value: 0.25626, ..., 0.508249
\item \strong{FA_FULLWM} \emph{numeric} Full white matter (mean) Range value: 0.339006, ..., 0.552456
\item \strong{MD_MCP} \emph{numeric} Middle cerebellar peduncle (mean) Range value: 0.000538402, ..., 0.000922075
\item \strong{MD_PCT} \emph{numeric} Pontine crossing tract (mean) Range value: 0.000438231, ..., 0.000876774
\item \strong{MD_GCC} \emph{numeric} Genu of corpus callosum (mean) Range value: 0.000698504, ..., 0.00117093
\item \strong{MD_BCC} \emph{numeric} Body of corpus callosum (mean) Range value: 0.000736046, ..., 0.00124782
\item \strong{MD_SCC} \emph{numeric} Splenium of corpus callosum (mean) Range value: 0.00062599, ..., 0.00125812
\item \strong{MD_FX} \emph{numeric} Fornix (mean) Range value: 0.00104399, ..., 0.00298852
\item \strong{MD_CST_R} \emph{numeric} Corticospinal tract right (mean) Range value: 0.000408575, ..., 0.000918554
\item \strong{MD_CST_L} \emph{numeric} Corticospinal tract left (mean) Range value: 0.000444655, ..., 0.000899429
\item \strong{MD_ML_R} \emph{numeric} Medial lemniscus right (mean) Range value: 0.000532252, ..., 0.00100212
\item \strong{MD_ML_L} \emph{numeric} Medial lemniscus left (mean) Range value: 0.000508878, ..., 0.0010474
\item \strong{MD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right (mean) Range value: 0.000494412, ..., 0.000925809
\item \strong{MD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left (mean) Range value: 0.000524807, ..., 0.000944346
\item \strong{MD_SCP_R} \emph{numeric} Superior cerebellar peduncle right (mean) Range value: 0.000700557, ..., 0.00161914
\item \strong{MD_SCP_L} \emph{numeric} Superior cerebellar peduncle left (mean) Range value: 0.000546938, ..., 0.0014295
\item \strong{MD_CP_R} \emph{numeric} Cerebral peduncle right (mean) Range value: 0.000368035, ..., 0.00106217
\item \strong{MD_CP_L} \emph{numeric} Cerebral peduncle left (mean) Range value: 0.000410472, ..., 0.00104219
\item \strong{MD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right (mean) Range value: 0.000578151, ..., 0.00117585
\item \strong{MD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left (mean) Range value: 0.000583517, ..., 0.00116933
\item \strong{MD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right (mean) Range value: 0.000541061, ..., 0.000899276
\item \strong{MD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left (mean) Range value: 0.000542291, ..., 0.000896734
\item \strong{MD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right (mean) Range value: 0.00050776, ..., 0.000950347
\item \strong{MD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left (mean) Range value: 0.000636994, ..., 0.000962304
\item \strong{MD_ACR_R} \emph{numeric} Anterior corona radiata right (mean) Range value: 0.000651486, ..., 0.0012336
\item \strong{MD_ACR_L} \emph{numeric} Anterior corona radiata left (mean) Range value: 0.000654708, ..., 0.00123377
\item \strong{MD_SCR_R} \emph{numeric} Superior corona radiata right (mean) Range value: 0.000611519, ..., 0.00113424
\item \strong{MD_SCR_L} \emph{numeric} Superior corona radiata left (mean) Range value: 0.000609465, ..., 0.00110716
\item \strong{MD_PCR_R} \emph{numeric} Posterior corona radiata right (mean) Range value: 0.000666802, ..., 0.00136957
\item \strong{MD_PCR_L} \emph{numeric} Posterior corona radiata left (mean) Range value: 0.000669869, ..., 0.00134033
\item \strong{MD_PTR_R} \emph{numeric} Posterior thalamic radiation right (mean) Range value: 0.000642956, ..., 0.00110866
\item \strong{MD_PTR_L} \emph{numeric} Posterior thalamic radiation left (mean) Range value: 0.000670468, ..., 0.00124786
\item \strong{MD_SS_R} \emph{numeric} Sagittal stratum right (mean) Range value: 0.00064897, ..., 0.00106108
\item \strong{MD_SS_L} \emph{numeric} Sagittal stratum left (mean) Range value: 0.000650695, ..., 0.00128965
\item \strong{MD_EC_R} \emph{numeric} External capsule right (mean) Range value: 0.000634019, ..., 0.00106651
\item \strong{MD_EC_L} \emph{numeric} External capsule left (mean) Range value: 0.000631123, ..., 0.0010586
\item \strong{MD_CGC_R} \emph{numeric} Cingulum right (mean) Range value: 0.00061873, ..., 0.000888201
\item \strong{MD_CGC_L} \emph{numeric} Cingulum left (mean) Range value: 0.000642707, ..., 0.000916112
\item \strong{MD_CGH_R} \emph{numeric} Cingulum right (mean) Range value: 0.000611251, ..., 0.00105117
\item \strong{MD_CGH_L} \emph{numeric} Cingulum left (mean) Range value: 0.000513676, ..., 0.00120918
\item \strong{MD_FXST_R} \emph{numeric} Fornix (cres) / Stria terminalis right (mean) Range value: 0.000658871, ..., 0.00127149
\item \strong{MD_FXST_L} \emph{numeric} Fornix (cres) / Stria terminalis left (mean) Range value: 0.000579136, ..., 0.00105693
\item \strong{MD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right (mean) Range value: 0.000623397, ..., 0.00103359
\item \strong{MD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left (mean) Range value: 0.000621985, ..., 0.0010204
\item \strong{MD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right (mean) Range value: 0.000607355, ..., 0.00144193
\item \strong{MD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left (mean) Range value: 0.000609021, ..., 0.00164983
\item \strong{MD_UNC_R} \emph{numeric} Uncinate fasciculus right (mean) Range value: 0.000560042, ..., 0.00130304
\item \strong{MD_UNC_L} \emph{numeric} Uncinate fasciculus left (mean) Range value: 0.000497959, ..., 0.00120965
\item \strong{MD_TAP_R} \emph{numeric} Tapetum right (mean) Range value: 0.000890469, ..., 0.00210118
\item \strong{MD_TAP_L} \emph{numeric} Tapetum left (mean) Range value: 0.000879502, ..., 0.00237742
\item \strong{MD_CST} \emph{numeric} Corticospinal tract Bilateral (mean) Range value: 0.000425304, ..., 0.000902124
\item \strong{MD_ML} \emph{numeric} Medial lemniscus Bilateral (mean) Range value: 0.000531991, ..., 0.00102379
\item \strong{MD_ICP} \emph{numeric} Inferior cerebellar peduncle Bilateral (mean) Range value: 0.000511224, ..., 0.00088883
\item \strong{MD_SCP} \emph{numeric} Superior cerebellar peduncle Bilateral (mean) Range value: 0.000625826, ..., 0.00151651
\item \strong{MD_CP} \emph{numeric} Cerebral peduncle Bilateral (mean) Range value: 0.000391221, ..., 0.00104791
\item \strong{MD_ALIC} \emph{numeric} Anterior limb of internal capsule Bilateral (mean) Range value: 0.000583594, ..., 0.00117308
\item \strong{MD_PLIC} \emph{numeric} Posterior limb of internal capsule Bilateral (mean) Range value: 0.000548241, ..., 0.000890865
\item \strong{MD_RLIC} \emph{numeric} Retrolenticular part of internal capsule Bilateral (mean) Range value: 0.000607829, ..., 0.00095053
\item \strong{MD_ACR} \emph{numeric} Anterior corona radiata Bilateral (mean) Range value: 0.000653176, ..., 0.00118316
\item \strong{MD_SCR} \emph{numeric} Superior corona radiata Bilateral (mean) Range value: 0.000614531, ..., 0.00111121
\item \strong{MD_PCR} \emph{numeric} Posterior corona radiata Bilateral (mean) Range value: 0.000674202, ..., 0.00135503
\item \strong{MD_PTR} \emph{numeric} Posterior thalamic radiation Bilateral (mean) Range value: 0.000657147, ..., 0.00112437
\item \strong{MD_SS} \emph{numeric} Sagittal stratum Bilateral (mean) Range value: 0.000680688, ..., 0.00111788
\item \strong{MD_EC} \emph{numeric} External capsule Bilateral (mean) Range value: 0.000632461, ..., 0.00105231
\item \strong{MD_CGC} \emph{numeric} Cingulum Bilateral (mean) Range value: 0.000635599, ..., 0.000887752
\item \strong{MD_CGH} \emph{numeric} Cingulum Bilateral (mean) Range value: 0.000558496, ..., 0.00111457
\item \strong{MD_FXST} \emph{numeric} Fornix (cres) / Stria terminalis Bilateral (mean) Range value: 0.000658591, ..., 0.0011443
\item \strong{MD_SLF} \emph{numeric} Superior longitudinal fasciculus Bilateral (mean) Range value: 0.000622668, ..., 0.00102702
\item \strong{MD_SFO} \emph{numeric} Superior fronto-occipital fasciculus Bilateral (mean) Range value: 0.000612006, ..., 0.00154059
\item \strong{MD_UNC} \emph{numeric} Uncinate fasciculus Bilateral (mean) Range value: 0.000575353, ..., 0.00126271
\item \strong{MD_TAP} \emph{numeric} Tapetum Bilateral (mean) Range value: 0.000931168, ..., 0.00224993
\item \strong{MD_IC} \emph{numeric} Full internal capsule (mean) Range value: 0.000581847, ..., 0.000970623
\item \strong{MD_CC} \emph{numeric} Full corpus callosum (mean) Range value: 0.000702356, ..., 0.00120624
\item \strong{MD_CR} \emph{numeric} Full corona radiata (mean) Range value: 0.000658942, ..., 0.00117606
\item \strong{MD_FULLWM} \emph{numeric} Full white matter (mean) Range value: 0.000641979, ..., 0.0010227
\item \strong{RD_MCP} \emph{numeric} Middle cerebellar peduncle (mean) Range value: 0.000360952, ..., 0.000707273
\item \strong{RD_PCT} \emph{numeric} Pontine crossing tract (mean) Range value: 0.000339464, ..., 0.000738328
\item \strong{RD_GCC} \emph{numeric} Genu of corpus callosum (mean) Range value: 0.000407506, ..., 0.000937897
\item \strong{RD_BCC} \emph{numeric} Body of corpus callosum (mean) Range value: 0.000417446, ..., 0.00100016
\item \strong{RD_SCC} \emph{numeric} Splenium of corpus callosum (mean) Range value: 0.000299174, ..., 0.00101665
\item \strong{RD_FX} \emph{numeric} Fornix (mean) Range value: 0.000747619, ..., 0.00279938
\item \strong{RD_CST_R} \emph{numeric} Corticospinal tract right (mean) Range value: 0.000261961, ..., 0.000764234
\item \strong{RD_CST_L} \emph{numeric} Corticospinal tract left (mean) Range value: 0.000273211, ..., 0.000717467
\item \strong{RD_ML_R} \emph{numeric} Medial lemniscus right (mean) Range value: 0.000248836, ..., 0.000786482
\item \strong{RD_ML_L} \emph{numeric} Medial lemniscus left (mean) Range value: 0.000205269, ..., 0.000847218
\item \strong{RD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right (mean) Range value: 0.000326104, ..., 0.00073094
\item \strong{RD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left (mean) Range value: 0.000339415, ..., 0.000771752
\item \strong{RD_SCP_R} \emph{numeric} Superior cerebellar peduncle right (mean) Range value: 0.000400842, ..., 0.00133279
\item \strong{RD_SCP_L} \emph{numeric} Superior cerebellar peduncle left (mean) Range value: 0.00036689, ..., 0.00113412
\item \strong{RD_CP_R} \emph{numeric} Cerebral peduncle right (mean) Range value: 0.000240145, ..., 0.000780001
\item \strong{RD_CP_L} \emph{numeric} Cerebral peduncle left (mean) Range value: 0.000225811, ..., 0.000797323
\item \strong{RD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right (mean) Range value: 0.000387114, ..., 0.000960008
\item \strong{RD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left (mean) Range value: 0.000346503, ..., 0.000941463
\item \strong{RD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right (mean) Range value: 0.00027324, ..., 0.00058138
\item \strong{RD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left (mean) Range value: 0.000294144, ..., 0.00063928
\item \strong{RD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right (mean) Range value: 0.00023902, ..., 0.000717964
\item \strong{RD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left (mean) Range value: 0.00036453, ..., 0.000717901
\item \strong{RD_ACR_R} \emph{numeric} Anterior corona radiata right (mean) Range value: 0.000508525, ..., 0.0010987
\item \strong{RD_ACR_L} \emph{numeric} Anterior corona radiata left (mean) Range value: 0.000519783, ..., 0.00109487
\item \strong{RD_SCR_R} \emph{numeric} Superior corona radiata right (mean) Range value: 0.000435616, ..., 0.000965589
\item \strong{RD_SCR_L} \emph{numeric} Superior corona radiata left (mean) Range value: 0.000435086, ..., 0.000919441
\item \strong{RD_PCR_R} \emph{numeric} Posterior corona radiata right (mean) Range value: 0.000481327, ..., 0.00116953
\item \strong{RD_PCR_L} \emph{numeric} Posterior corona radiata left (mean) Range value: 0.000497635, ..., 0.00116231
\item \strong{RD_PTR_R} \emph{numeric} Posterior thalamic radiation right (mean) Range value: 0.000391309, ..., 0.000911663
\item \strong{RD_PTR_L} \emph{numeric} Posterior thalamic radiation left (mean) Range value: 0.000457335, ..., 0.0010319
\item \strong{RD_SS_R} \emph{numeric} Sagittal stratum right (mean) Range value: 0.000434705, ..., 0.000844208
\item \strong{RD_SS_L} \emph{numeric} Sagittal stratum left (mean) Range value: 0.000416646, ..., 0.0010709
\item \strong{RD_EC_R} \emph{numeric} External capsule right (mean) Range value: 0.000495253, ..., 0.000945899
\item \strong{RD_EC_L} \emph{numeric} External capsule left (mean) Range value: 0.000464835, ..., 0.000929433
\item \strong{RD_CGC_R} \emph{numeric} Cingulum right (mean) Range value: 0.00045234, ..., 0.000766282
\item \strong{RD_CGC_L} \emph{numeric} Cingulum left (mean) Range value: 0.000438164, ..., 0.000765136
\item \strong{RD_CGH_R} \emph{numeric} Cingulum right (mean) Range value: 0.000473599, ..., 0.000932208
\item \strong{RD_CGH_L} \emph{numeric} Cingulum left (mean) Range value: 0.000433814, ..., 0.00107233
\item \strong{RD_FXST_R} \emph{numeric} Fornix (cres) / Stria terminalis right (mean) Range value: 0.000504701, ..., 0.00107425
\item \strong{RD_FXST_L} \emph{numeric} Fornix (cres) / Stria terminalis left (mean) Range value: 0.000379256, ..., 0.000892421
\item \strong{RD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right (mean) Range value: 0.0004248, ..., 0.00086565
\item \strong{RD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left (mean) Range value: 0.000451681, ..., 0.000860261
\item \strong{RD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right (mean) Range value: 0.000462817, ..., 0.00126797
\item \strong{RD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left (mean) Range value: 0.000457079, ..., 0.00149614
\item \strong{RD_UNC_R} \emph{numeric} Uncinate fasciculus right (mean) Range value: 0.000417975, ..., 0.00113286
\item \strong{RD_UNC_L} \emph{numeric} Uncinate fasciculus left (mean) Range value: 0.000345537, ..., 0.00103479
\item \strong{RD_TAP_R} \emph{numeric} Tapetum right (mean) Range value: 0.000600145, ..., 0.00185332
\item \strong{RD_TAP_L} \emph{numeric} Tapetum left (mean) Range value: 0.000697786, ..., 0.00219405
\item \strong{RD_CST} \emph{numeric} Corticospinal tract Bilateral (mean) Range value: 0.000271487, ..., 0.00073592
\item \strong{RD_ML} \emph{numeric} Medial lemniscus Bilateral (mean) Range value: 0.000242604, ..., 0.000815548
\item \strong{RD_ICP} \emph{numeric} Inferior cerebellar peduncle Bilateral (mean) Range value: 0.000339417, ..., 0.000727756
\item \strong{RD_SCP} \emph{numeric} Superior cerebellar peduncle Bilateral (mean) Range value: 0.000399202, ..., 0.00122528
\item \strong{RD_CP} \emph{numeric} Cerebral peduncle Bilateral (mean) Range value: 0.000232078, ..., 0.000787083
\item \strong{RD_ALIC} \emph{numeric} Anterior limb of internal capsule Bilateral (mean) Range value: 0.000393487, ..., 0.000952149
\item \strong{RD_PLIC} \emph{numeric} Posterior limb of internal capsule Bilateral (mean) Range value: 0.000291553, ..., 0.000586358
\item \strong{RD_RLIC} \emph{numeric} Retrolenticular part of internal capsule Bilateral (mean) Range value: 0.000360929, ..., 0.000708689
\item \strong{RD_ACR} \emph{numeric} Anterior corona radiata Bilateral (mean) Range value: 0.000515003, ..., 0.00104115
\item \strong{RD_SCR} \emph{numeric} Superior corona radiata Bilateral (mean) Range value: 0.000446576, ..., 0.000942543
\item \strong{RD_PCR} \emph{numeric} Posterior corona radiata Bilateral (mean) Range value: 0.000488885, ..., 0.00116594
\item \strong{RD_PTR} \emph{numeric} Posterior thalamic radiation Bilateral (mean) Range value: 0.000438526, ..., 0.000914865
\item \strong{RD_SS} \emph{numeric} Sagittal stratum Bilateral (mean) Range value: 0.000425496, ..., 0.000888264
\item \strong{RD_EC} \emph{numeric} External capsule Bilateral (mean) Range value: 0.000492962, ..., 0.000928367
\item \strong{RD_CGC} \emph{numeric} Cingulum Bilateral (mean) Range value: 0.000450064, ..., 0.00076197
\item \strong{RD_CGH} \emph{numeric} Cingulum Bilateral (mean) Range value: 0.000470176, ..., 0.000982526
\item \strong{RD_FXST} \emph{numeric} Fornix (cres) / Stria terminalis Bilateral (mean) Range value: 0.000454056, ..., 0.000935865
\item \strong{RD_SLF} \emph{numeric} Superior longitudinal fasciculus Bilateral (mean) Range value: 0.000444042, ..., 0.000862964
\item \strong{RD_SFO} \emph{numeric} Superior fronto-occipital fasciculus Bilateral (mean) Range value: 0.000471453, ..., 0.00137869
\item \strong{RD_UNC} \emph{numeric} Uncinate fasciculus Bilateral (mean) Range value: 0.000434886, ..., 0.00110189
\item \strong{RD_TAP} \emph{numeric} Tapetum Bilateral (mean) Range value: 0.000710458, ..., 0.00203679
\item \strong{RD_IC} \emph{numeric} Full internal capsule (mean) Range value: 0.000353446, ..., 0.000705921
\item \strong{RD_CC} \emph{numeric} Full corpus callosum (mean) Range value: 0.000387306, ..., 0.000926626
\item \strong{RD_CR} \emph{numeric} Full corona radiata (mean) Range value: 0.000500982, ..., 0.001016
\item \strong{RD_FULLWM} \emph{numeric} Full white matter (mean) Range value: 0.000434768, ..., 0.000821996
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{adni4}
\keyword{imaging}
