% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DTIROI_ROBUSTMEAN}
\alias{DTIROI_ROBUSTMEAN}
\title{USC - DTI ROI Summary Measures v2 + v3 (Robust Mean)}
\format{
A data frame with 1750 observations and 310 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(DTIROI_ROBUSTMEAN)
}
\description{
USC - DTI ROI Summary Measures v2 + v3 (Robust Mean) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI2, ADNI3, ADNIGO, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 11, ..., 10793
\item \strong{VISCODE} \emph{character} Visit code , with options: init, y2, y1, sc, y4, y3, 4_sc, 4_init, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{character} Preprocessing pipeline version 2 = preprocessing pipeline as outlined in methods PDF: (USC - DTI ROI Summary Measures Methods v2 [ADNI3], date: 12/01/2021) ; 3 = preprocessing pipeline as outlined in methods PDF: (USC - DTI ROI Summary Measures Methods v3 [ADNI4], date: 04/02/2025) ; -4 = image was not processed as noted in QC column.
\emph{Decoded Value: }\strong{Yes}
\item \strong{MANUFACTURER} \emph{character} Scanner Manufacturer , with options: Siemens, GE, Philips, SIEMENS, GE MEDICAL SYSTEMS
\item \strong{VOLUMES} \emph{numeric} Number of volumes prescribed by acquisition protocol Range value: 31, ..., 127
\item \strong{IMAGEUID} \emph{numeric} IMAGE ID of the raw diffusion scan Range value: 808826, ..., 11303224
\item \strong{DISTORTION_CORRECTION} \emph{character} In V3, EPI susceptibility distortion correction was run using a synthetic b0 (Synb0), b0s acquired with opposite phase encoding directions (APPA), or was completed in scanner (scanner), in addition to warping DWIs to T1w. In V2, susceptibility distortion was only corrected by warping DWIs to T1w (empty cells). Synb0=synthetic b0 ; APPA=b0s acquired with opposite phase encoding directions ; scanner=completed in scanner.
\item \strong{VOLUMES_AP} \emph{numeric} Number of volumes in the opposite phase encoding direction raw scan. Empty cell indicates there was no opposite phase encoding image provided to users on the IDA. Range value: 7, ..., 13
\item \strong{IMAGEUID_AP} \emph{numeric} IMAGE ID of the raw diffusion scan with an opposite phase encoding direction. Empty cell indicates there was no opposite phase encoding image provided to users on the IDA. Range value: 10248884, ..., 11297753
\item \strong{RUNDATE} \emph{ } Processing Date Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Processing Completed , with options: complete, not processed
\item \strong{QC} \emph{character} Preprocessed DWI and scalar map quality check notes include: (1) a note if raw DWI were not processed due to incomplete acquisition (e.g., fewer volumes acquired than protocol), failed DICOM conversion, poor field of view (i.e., image cropped), or severe artifacts (e.g., motion, distortion); or (2) notes if preprocessed DWI had any remaining artifacts (e.g., motion, distortion, ghosting, hyperintensities in FA, etc.), pathology or anatomical anomalies, atlas registration failure, or if the voxel size differed from the prescribed acquisition protocol. These images should be used at own discretion.  
\item \strong{AD_MCP} \emph{numeric} Middle cerebellar peduncle (robust mean) Range value: 0.000829046, ..., 0.001317
\item \strong{AD_PCT} \emph{numeric} Pontine crossing tract (robust mean) Range value: 0.000635654, ..., 0.0012143
\item \strong{AD_GCC} \emph{numeric} Genu of corpus callosum (robust mean) Range value: 0.00107923, ..., 0.0017148
\item \strong{AD_BCC} \emph{numeric} Body of corpus callosum (robust mean) Range value: 0.00114831, ..., 0.00179691
\item \strong{AD_SCC} \emph{numeric} Splenium of corpus callosum (robust mean) Range value: 0.000931843, ..., 0.00181219
\item \strong{AD_FX} \emph{numeric} Fornix (robust mean) Range value: 0.00136773, ..., 0.00343132
\item \strong{AD_CST_R} \emph{numeric} Corticospinal tract right (robust mean) Range value: 0.000700309, ..., 0.00123863
\item \strong{AD_CST_L} \emph{numeric} Corticospinal tract left (robust mean) Range value: 0.000727272, ..., 0.00127933
\item \strong{AD_ML_R} \emph{numeric} Medial lemniscus right (robust mean) Range value: 0.000871556, ..., 0.00148612
\item \strong{AD_ML_L} \emph{numeric} Medial lemniscus left (robust mean) Range value: 0.000904022, ..., 0.00145631
\item \strong{AD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right (robust mean) Range value: 0.000769043, ..., 0.00134749
\item \strong{AD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left (robust mean) Range value: 0.000790416, ..., 0.00132972
\item \strong{AD_SCP_R} \emph{numeric} Superior cerebellar peduncle right (robust mean) Range value: 0.00107794, ..., 0.00216635
\item \strong{AD_SCP_L} \emph{numeric} Superior cerebellar peduncle left (robust mean) Range value: 0.000933815, ..., 0.00206616
\item \strong{AD_CP_R} \emph{numeric} Cerebral peduncle right (robust mean) Range value: 0.000734282, ..., 0.00172317
\item \strong{AD_CP_L} \emph{numeric} Cerebral peduncle left (robust mean) Range value: 0.000790256, ..., 0.00156891
\item \strong{AD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right (robust mean) Range value: 0.000877232, ..., 0.00160427
\item \strong{AD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left (robust mean) Range value: 0.000899512, ..., 0.00163145
\item \strong{AD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right (robust mean) Range value: 0.000964611, ..., 0.0015394
\item \strong{AD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left (robust mean) Range value: 0.000920428, ..., 0.00150803
\item \strong{AD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right (robust mean) Range value: 0.000967808, ..., 0.00151841
\item \strong{AD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left (robust mean) Range value: 0.000913532, ..., 0.0015243
\item \strong{AD_ACR_R} \emph{numeric} Anterior corona radiata right (robust mean) Range value: 0.000870891, ..., 0.00149696
\item \strong{AD_ACR_L} \emph{numeric} Anterior corona radiata left (robust mean) Range value: 0.000917782, ..., 0.00150297
\item \strong{AD_SCR_R} \emph{numeric} Superior corona radiata right (robust mean) Range value: 0.000897053, ..., 0.00159583
\item \strong{AD_SCR_L} \emph{numeric} Superior corona radiata left (robust mean) Range value: 0.000922094, ..., 0.00156574
\item \strong{AD_PCR_R} \emph{numeric} Posterior corona radiata right (robust mean) Range value: 0.000991979, ..., 0.00178614
\item \strong{AD_PCR_L} \emph{numeric} Posterior corona radiata left (robust mean) Range value: 0.000910448, ..., 0.00172159
\item \strong{AD_PTR_R} \emph{numeric} Posterior thalamic radiation right (robust mean) Range value: 0.00101088, ..., 0.00155547
\item \strong{AD_PTR_L} \emph{numeric} Posterior thalamic radiation left (robust mean) Range value: 0.00105047, ..., 0.00163995
\item \strong{AD_SS_R} \emph{numeric} Sagittal stratum right (robust mean) Range value: 0.00102135, ..., 0.00155689
\item \strong{AD_SS_L} \emph{numeric} Sagittal stratum left (robust mean) Range value: 0.000930755, ..., 0.00173607
\item \strong{AD_EC_R} \emph{numeric} External capsule right (robust mean) Range value: 0.000870701, ..., 0.00138704
\item \strong{AD_EC_L} \emph{numeric} External capsule left (robust mean) Range value: 0.00090765, ..., 0.00139369
\item \strong{AD_CGC_R} \emph{numeric} Cingulum right (robust mean) Range value: 0.00079713, ..., 0.00124808
\item \strong{AD_CGC_L} \emph{numeric} Cingulum left (robust mean) Range value: 0.00077632, ..., 0.00125187
\item \strong{AD_CGH_R} \emph{numeric} Cingulum right (robust mean) Range value: 0.000783423, ..., 0.00129411
\item \strong{AD_CGH_L} \emph{numeric} Cingulum left (robust mean) Range value: 0.000656679, ..., 0.0014759
\item \strong{AD_FXST_R} \emph{numeric} Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000898062, ..., 0.00179612
\item \strong{AD_FXST_L} \emph{numeric} Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.000832679, ..., 0.00145359
\item \strong{AD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right (robust mean) Range value: 0.00091249, ..., 0.00136921
\item \strong{AD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left (robust mean) Range value: 0.000929409, ..., 0.00134061
\item \strong{AD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right (robust mean) Range value: 0.000863616, ..., 0.00183968
\item \strong{AD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left (robust mean) Range value: 0.000868013, ..., 0.00190102
\item \strong{AD_UNC_R} \emph{numeric} Uncinate fasciculus right (robust mean) Range value: 0.00082834, ..., 0.00157334
\item \strong{AD_UNC_L} \emph{numeric} Uncinate fasciculus left (robust mean) Range value: 0.000779406, ..., 0.00151033
\item \strong{AD_TAP_R} \emph{numeric} Tapetum right (robust mean) Range value: 0.00116991, ..., 0.00270459
\item \strong{AD_TAP_L} \emph{numeric} Tapetum left (robust mean) Range value: 0.00120224, ..., 0.0028861
\item \strong{AD_CST} \emph{numeric} Corticospinal tract Bilateral (robust mean) Range value: 0.000747601, ..., 0.00123127
\item \strong{AD_ML} \emph{numeric} Medial lemniscus Bilateral (robust mean) Range value: 0.000906207, ..., 0.00145263
\item \strong{AD_ICP} \emph{numeric} Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000804694, ..., 0.00133767
\item \strong{AD_SCP} \emph{numeric} Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.00102481, ..., 0.00207276
\item \strong{AD_CP} \emph{numeric} Cerebral peduncle Bilateral (robust mean) Range value: 0.000802573, ..., 0.00164113
\item \strong{AD_ALIC} \emph{numeric} Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.00091851, ..., 0.00161555
\item \strong{AD_PLIC} \emph{numeric} Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.000942687, ..., 0.00151642
\item \strong{AD_RLIC} \emph{numeric} Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.00104101, ..., 0.00150999
\item \strong{AD_ACR} \emph{numeric} Anterior corona radiata Bilateral (robust mean) Range value: 0.000905175, ..., 0.00147056
\item \strong{AD_SCR} \emph{numeric} Superior corona radiata Bilateral (robust mean) Range value: 0.000922737, ..., 0.0015798
\item \strong{AD_PCR} \emph{numeric} Posterior corona radiata Bilateral (robust mean) Range value: 0.00100198, ..., 0.00175337
\item \strong{AD_PTR} \emph{numeric} Posterior thalamic radiation Bilateral (robust mean) Range value: 0.00105728, ..., 0.00152739
\item \strong{AD_SS} \emph{numeric} Sagittal stratum Bilateral (robust mean) Range value: 0.00101621, ..., 0.00155588
\item \strong{AD_EC} \emph{numeric} External capsule Bilateral (robust mean) Range value: 0.000898161, ..., 0.00136933
\item \strong{AD_CGC} \emph{numeric} Cingulum Bilateral (robust mean) Range value: 0.00078856, ..., 0.0012405
\item \strong{AD_CGH} \emph{numeric} Cingulum Bilateral (robust mean) Range value: 0.000731739, ..., 0.00135534
\item \strong{AD_FXST} \emph{numeric} Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.000869624, ..., 0.00162273
\item \strong{AD_SLF} \emph{numeric} Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000920892, ..., 0.00135425
\item \strong{AD_SFO} \emph{numeric} Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000869218, ..., 0.00179869
\item \strong{AD_UNC} \emph{numeric} Uncinate fasciculus Bilateral (robust mean) Range value: 0.000802693, ..., 0.00154848
\item \strong{AD_TAP} \emph{numeric} Tapetum Bilateral (robust mean) Range value: 0.00124986, ..., 0.00280773
\item \strong{AD_IC} \emph{numeric} Full internal capsule (robust mean) Range value: 0.000965563, ..., 0.00149749
\item \strong{AD_CC} \emph{numeric} Full corpus callosum (robust mean) Range value: 0.00106926, ..., 0.00175088
\item \strong{AD_CR} \emph{numeric} Full corona radiata (robust mean) Range value: 0.000969116, ..., 0.00152303
\item \strong{AD_FULLWM} \emph{numeric} Full white matter (robust mean) Range value: 0.000996245, ..., 0.00142226
\item \strong{FA_MCP} \emph{numeric} Middle cerebellar peduncle (robust mean) Range value: 0.248813, ..., 0.549816
\item \strong{FA_PCT} \emph{numeric} Pontine crossing tract (robust mean) Range value: 0.200079, ..., 0.518204
\item \strong{FA_GCC} \emph{numeric} Genu of corpus callosum (robust mean) Range value: 0.294355, ..., 0.652468
\item \strong{FA_BCC} \emph{numeric} Body of corpus callosum (robust mean) Range value: 0.266139, ..., 0.676548
\item \strong{FA_SCC} \emph{numeric} Splenium of corpus callosum (robust mean) Range value: 0.302237, ..., 0.755416
\item \strong{FA_FX} \emph{numeric} Fornix (robust mean) Range value: 0.0586734, ..., 0.532278
\item \strong{FA_CST_R} \emph{numeric} Corticospinal tract right (robust mean) Range value: 0.230267, ..., 0.656646
\item \strong{FA_CST_L} \emph{numeric} Corticospinal tract left (robust mean) Range value: 0.226838, ..., 0.650365
\item \strong{FA_ML_R} \emph{numeric} Medial lemniscus right (robust mean) Range value: 0.279958, ..., 0.7301
\item \strong{FA_ML_L} \emph{numeric} Medial lemniscus left (robust mean) Range value: 0.287144, ..., 0.787983
\item \strong{FA_ICP_R} \emph{numeric} Inferior cerebellar peduncle right (robust mean) Range value: 0.193219, ..., 0.596592
\item \strong{FA_ICP_L} \emph{numeric} Inferior cerebellar peduncle left (robust mean) Range value: 0.205008, ..., 0.595435
\item \strong{FA_SCP_R} \emph{numeric} Superior cerebellar peduncle right (robust mean) Range value: 0.257702, ..., 0.731428
\item \strong{FA_SCP_L} \emph{numeric} Superior cerebellar peduncle left (robust mean) Range value: 0.268533, ..., 0.786693
\item \strong{FA_CP_R} \emph{numeric} Cerebral peduncle right (robust mean) Range value: 0.32235, ..., 0.788253
\item \strong{FA_CP_L} \emph{numeric} Cerebral peduncle left (robust mean) Range value: 0.350584, ..., 0.783827
\item \strong{FA_ALIC_R} \emph{numeric} Anterior limb of internal capsule right (robust mean) Range value: 0.305473, ..., 0.628841
\item \strong{FA_ALIC_L} \emph{numeric} Anterior limb of internal capsule left (robust mean) Range value: 0.239185, ..., 0.617397
\item \strong{FA_PLIC_R} \emph{numeric} Posterior limb of internal capsule right (robust mean) Range value: 0.498038, ..., 0.724186
\item \strong{FA_PLIC_L} \emph{numeric} Posterior limb of internal capsule left (robust mean) Range value: 0.411263, ..., 0.751875
\item \strong{FA_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right (robust mean) Range value: 0.36305, ..., 0.738754
\item \strong{FA_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left (robust mean) Range value: 0.295967, ..., 0.665867
\item \strong{FA_ACR_R} \emph{numeric} Anterior corona radiata right (robust mean) Range value: 0.206771, ..., 0.475391
\item \strong{FA_ACR_L} \emph{numeric} Anterior corona radiata left (robust mean) Range value: 0.203699, ..., 0.464372
\item \strong{FA_SCR_R} \emph{numeric} Superior corona radiata right (robust mean) Range value: 0.259513, ..., 0.559231
\item \strong{FA_SCR_L} \emph{numeric} Superior corona radiata left (robust mean) Range value: 0.282371, ..., 0.588552
\item \strong{FA_PCR_R} \emph{numeric} Posterior corona radiata right (robust mean) Range value: 0.249689, ..., 0.519696
\item \strong{FA_PCR_L} \emph{numeric} Posterior corona radiata left (robust mean) Range value: 0.232454, ..., 0.519685
\item \strong{FA_PTR_R} \emph{numeric} Posterior thalamic radiation right (robust mean) Range value: 0.281566, ..., 0.627603
\item \strong{FA_PTR_L} \emph{numeric} Posterior thalamic radiation left (robust mean) Range value: 0.285254, ..., 0.595256
\item \strong{FA_SS_R} \emph{numeric} Sagittal stratum right (robust mean) Range value: 0.25673, ..., 0.610652
\item \strong{FA_SS_L} \emph{numeric} Sagittal stratum left (robust mean) Range value: 0.290514, ..., 0.618269
\item \strong{FA_EC_R} \emph{numeric} External capsule right (robust mean) Range value: 0.206798, ..., 0.461742
\item \strong{FA_EC_L} \emph{numeric} External capsule left (robust mean) Range value: 0.190149, ..., 0.481229
\item \strong{FA_CGC_R} \emph{numeric} Cingulum right (robust mean) Range value: 0.195094, ..., 0.539067
\item \strong{FA_CGC_L} \emph{numeric} Cingulum left (robust mean) Range value: 0.182076, ..., 0.544045
\item \strong{FA_CGH_R} \emph{numeric} Cingulum right (robust mean) Range value: 0.113434, ..., 0.47185
\item \strong{FA_CGH_L} \emph{numeric} Cingulum left (robust mean) Range value: 0.10481, ..., 0.47028
\item \strong{FA_FXST_R} \emph{numeric} Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.207085, ..., 0.550102
\item \strong{FA_FXST_L} \emph{numeric} Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.229574, ..., 0.594808
\item \strong{FA_SLF_R} \emph{numeric} Superior longitudinal fasciculus right (robust mean) Range value: 0.290606, ..., 0.537435
\item \strong{FA_SLF_L} \emph{numeric} Superior longitudinal fasciculus left (robust mean) Range value: 0.269563, ..., 0.527995
\item \strong{FA_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right (robust mean) Range value: 0.221473, ..., 0.514343
\item \strong{FA_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left (robust mean) Range value: 0.195763, ..., 0.490461
\item \strong{FA_UNC_R} \emph{numeric} Uncinate fasciculus right (robust mean) Range value: 0.142319, ..., 0.560214
\item \strong{FA_UNC_L} \emph{numeric} Uncinate fasciculus left (robust mean) Range value: 0.158052, ..., 0.567147
\item \strong{FA_TAP_R} \emph{numeric} Tapetum right (robust mean) Range value: 0.18739, ..., 0.5784
\item \strong{FA_TAP_L} \emph{numeric} Tapetum left (robust mean) Range value: 0.133857, ..., 0.510728
\item \strong{FA_CST} \emph{numeric} Corticospinal tract Bilateral (robust mean) Range value: 0.227816, ..., 0.653541
\item \strong{FA_ML} \emph{numeric} Medial lemniscus Bilateral (robust mean) Range value: 0.283328, ..., 0.753179
\item \strong{FA_ICP} \emph{numeric} Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.204613, ..., 0.574223
\item \strong{FA_SCP} \emph{numeric} Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.263951, ..., 0.730009
\item \strong{FA_CP} \emph{numeric} Cerebral peduncle Bilateral (robust mean) Range value: 0.35189, ..., 0.784614
\item \strong{FA_ALIC} \emph{numeric} Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.279054, ..., 0.616763
\item \strong{FA_PLIC} \emph{numeric} Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.461686, ..., 0.719953
\item \strong{FA_RLIC} \emph{numeric} Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.357634, ..., 0.658915
\item \strong{FA_ACR} \emph{numeric} Anterior corona radiata Bilateral (robust mean) Range value: 0.21453, ..., 0.454608
\item \strong{FA_SCR} \emph{numeric} Superior corona radiata Bilateral (robust mean) Range value: 0.271405, ..., 0.57357
\item \strong{FA_PCR} \emph{numeric} Posterior corona radiata Bilateral (robust mean) Range value: 0.244007, ..., 0.517341
\item \strong{FA_PTR} \emph{numeric} Posterior thalamic radiation Bilateral (robust mean) Range value: 0.322733, ..., 0.611
\item \strong{FA_SS} \emph{numeric} Sagittal stratum Bilateral (robust mean) Range value: 0.272583, ..., 0.596217
\item \strong{FA_EC} \emph{numeric} External capsule Bilateral (robust mean) Range value: 0.198309, ..., 0.462175
\item \strong{FA_CGC} \emph{numeric} Cingulum Bilateral (robust mean) Range value: 0.188919, ..., 0.536274
\item \strong{FA_CGH} \emph{numeric} Cingulum Bilateral (robust mean) Range value: 0.109208, ..., 0.461466
\item \strong{FA_FXST} \emph{numeric} Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.217253, ..., 0.561457
\item \strong{FA_SLF} \emph{numeric} Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.288276, ..., 0.531831
\item \strong{FA_SFO} \emph{numeric} Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.224116, ..., 0.48538
\item \strong{FA_UNC} \emph{numeric} Uncinate fasciculus Bilateral (robust mean) Range value: 0.151583, ..., 0.55104
\item \strong{FA_TAP} \emph{numeric} Tapetum Bilateral (robust mean) Range value: 0.184844, ..., 0.539506
\item \strong{FA_IC} \emph{numeric} Full internal capsule (robust mean) Range value: 0.420054, ..., 0.661411
\item \strong{FA_CC} \emph{numeric} Full corpus callosum (robust mean) Range value: 0.293799, ..., 0.691702
\item \strong{FA_CR} \emph{numeric} Full corona radiata (robust mean) Range value: 0.242039, ..., 0.505409
\item \strong{FA_FULLWM} \emph{numeric} Full white matter (robust mean) Range value: 0.328698, ..., 0.549097
\item \strong{MD_MCP} \emph{numeric} Middle cerebellar peduncle (robust mean) Range value: 0.000520727, ..., 0.000821757
\item \strong{MD_PCT} \emph{numeric} Pontine crossing tract (robust mean) Range value: 0.000442809, ..., 0.000871865
\item \strong{MD_GCC} \emph{numeric} Genu of corpus callosum (robust mean) Range value: 0.000683782, ..., 0.00108986
\item \strong{MD_BCC} \emph{numeric} Body of corpus callosum (robust mean) Range value: 0.000673059, ..., 0.00117509
\item \strong{MD_SCC} \emph{numeric} Splenium of corpus callosum (robust mean) Range value: 0.000619261, ..., 0.00119253
\item \strong{MD_FX} \emph{numeric} Fornix (robust mean) Range value: 0.0010403, ..., 0.00308202
\item \strong{MD_CST_R} \emph{numeric} Corticospinal tract right (robust mean) Range value: 0.000408661, ..., 0.000911616
\item \strong{MD_CST_L} \emph{numeric} Corticospinal tract left (robust mean) Range value: 0.000442175, ..., 0.000880286
\item \strong{MD_ML_R} \emph{numeric} Medial lemniscus right (robust mean) Range value: 0.000533072, ..., 0.00100121
\item \strong{MD_ML_L} \emph{numeric} Medial lemniscus left (robust mean) Range value: 0.000501333, ..., 0.00104549
\item \strong{MD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right (robust mean) Range value: 0.000493465, ..., 0.000881453
\item \strong{MD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left (robust mean) Range value: 0.000516165, ..., 0.000926643
\item \strong{MD_SCP_R} \emph{numeric} Superior cerebellar peduncle right (robust mean) Range value: 0.000656447, ..., 0.0015982
\item \strong{MD_SCP_L} \emph{numeric} Superior cerebellar peduncle left (robust mean) Range value: 0.00052145, ..., 0.00141151
\item \strong{MD_CP_R} \emph{numeric} Cerebral peduncle right (robust mean) Range value: 0.000363368, ..., 0.00100136
\item \strong{MD_CP_L} \emph{numeric} Cerebral peduncle left (robust mean) Range value: 0.00040876, ..., 0.000943709
\item \strong{MD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right (robust mean) Range value: 0.000573952, ..., 0.00117478
\item \strong{MD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left (robust mean) Range value: 0.000577341, ..., 0.00118213
\item \strong{MD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right (robust mean) Range value: 0.000541839, ..., 0.000884915
\item \strong{MD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left (robust mean) Range value: 0.000552822, ..., 0.000894506
\item \strong{MD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right (robust mean) Range value: 0.000535372, ..., 0.000943997
\item \strong{MD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left (robust mean) Range value: 0.000640883, ..., 0.000949934
\item \strong{MD_ACR_R} \emph{numeric} Anterior corona radiata right (robust mean) Range value: 0.00064264, ..., 0.00121927
\item \strong{MD_ACR_L} \emph{numeric} Anterior corona radiata left (robust mean) Range value: 0.000650969, ..., 0.00122298
\item \strong{MD_SCR_R} \emph{numeric} Superior corona radiata right (robust mean) Range value: 0.000611339, ..., 0.00114902
\item \strong{MD_SCR_L} \emph{numeric} Superior corona radiata left (robust mean) Range value: 0.000611529, ..., 0.00109089
\item \strong{MD_PCR_R} \emph{numeric} Posterior corona radiata right (robust mean) Range value: 0.000667923, ..., 0.00138183
\item \strong{MD_PCR_L} \emph{numeric} Posterior corona radiata left (robust mean) Range value: 0.000669616, ..., 0.0013664
\item \strong{MD_PTR_R} \emph{numeric} Posterior thalamic radiation right (robust mean) Range value: 0.000643095, ..., 0.00109604
\item \strong{MD_PTR_L} \emph{numeric} Posterior thalamic radiation left (robust mean) Range value: 0.000661464, ..., 0.00118057
\item \strong{MD_SS_R} \emph{numeric} Sagittal stratum right (robust mean) Range value: 0.00063461, ..., 0.00104085
\item \strong{MD_SS_L} \emph{numeric} Sagittal stratum left (robust mean) Range value: 0.000637773, ..., 0.00126493
\item \strong{MD_EC_R} \emph{numeric} External capsule right (robust mean) Range value: 0.000627293, ..., 0.00105366
\item \strong{MD_EC_L} \emph{numeric} External capsule left (robust mean) Range value: 0.000628879, ..., 0.00104725
\item \strong{MD_CGC_R} \emph{numeric} Cingulum right (robust mean) Range value: 0.000624982, ..., 0.000878465
\item \strong{MD_CGC_L} \emph{numeric} Cingulum left (robust mean) Range value: 0.000639246, ..., 0.000897592
\item \strong{MD_CGH_R} \emph{numeric} Cingulum right (robust mean) Range value: 0.000608728, ..., 0.00100507
\item \strong{MD_CGH_L} \emph{numeric} Cingulum left (robust mean) Range value: 0.000496251, ..., 0.00119783
\item \strong{MD_FXST_R} \emph{numeric} Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000630743, ..., 0.00123439
\item \strong{MD_FXST_L} \emph{numeric} Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.000570996, ..., 0.00101879
\item \strong{MD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right (robust mean) Range value: 0.000623097, ..., 0.00102493
\item \strong{MD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left (robust mean) Range value: 0.00062094, ..., 0.00101563
\item \strong{MD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right (robust mean) Range value: 0.00060823, ..., 0.00133526
\item \strong{MD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left (robust mean) Range value: 0.00060715, ..., 0.00162845
\item \strong{MD_UNC_R} \emph{numeric} Uncinate fasciculus right (robust mean) Range value: 0.000555553, ..., 0.00123852
\item \strong{MD_UNC_L} \emph{numeric} Uncinate fasciculus left (robust mean) Range value: 0.000500539, ..., 0.00105048
\item \strong{MD_TAP_R} \emph{numeric} Tapetum right (robust mean) Range value: 0.000806239, ..., 0.00208755
\item \strong{MD_TAP_L} \emph{numeric} Tapetum left (robust mean) Range value: 0.00083136, ..., 0.00247057
\item \strong{MD_CST} \emph{numeric} Corticospinal tract Bilateral (robust mean) Range value: 0.000424935, ..., 0.00089689
\item \strong{MD_ML} \emph{numeric} Medial lemniscus Bilateral (robust mean) Range value: 0.000531306, ..., 0.00102067
\item \strong{MD_ICP} \emph{numeric} Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000505265, ..., 0.000874894
\item \strong{MD_SCP} \emph{numeric} Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.000599782, ..., 0.00149563
\item \strong{MD_CP} \emph{numeric} Cerebral peduncle Bilateral (robust mean) Range value: 0.000388826, ..., 0.000963922
\item \strong{MD_ALIC} \emph{numeric} Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.000584356, ..., 0.0011772
\item \strong{MD_PLIC} \emph{numeric} Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.000548557, ..., 0.000872902
\item \strong{MD_RLIC} \emph{numeric} Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.000612252, ..., 0.000933875
\item \strong{MD_ACR} \emph{numeric} Anterior corona radiata Bilateral (robust mean) Range value: 0.000647226, ..., 0.00116482
\item \strong{MD_SCR} \emph{numeric} Superior corona radiata Bilateral (robust mean) Range value: 0.000615078, ..., 0.00111782
\item \strong{MD_PCR} \emph{numeric} Posterior corona radiata Bilateral (robust mean) Range value: 0.000674359, ..., 0.00137404
\item \strong{MD_PTR} \emph{numeric} Posterior thalamic radiation Bilateral (robust mean) Range value: 0.000652886, ..., 0.00109555
\item \strong{MD_SS} \emph{numeric} Sagittal stratum Bilateral (robust mean) Range value: 0.000654656, ..., 0.00107304
\item \strong{MD_EC} \emph{numeric} External capsule Bilateral (robust mean) Range value: 0.000628181, ..., 0.00103973
\item \strong{MD_CGC} \emph{numeric} Cingulum Bilateral (robust mean) Range value: 0.000633, ..., 0.000874815
\item \strong{MD_CGH} \emph{numeric} Cingulum Bilateral (robust mean) Range value: 0.000553957, ..., 0.00109127
\item \strong{MD_FXST} \emph{numeric} Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.000637934, ..., 0.00107901
\item \strong{MD_SLF} \emph{numeric} Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000621994, ..., 0.00102013
\item \strong{MD_SFO} \emph{numeric} Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000609919, ..., 0.00147151
\item \strong{MD_UNC} \emph{numeric} Uncinate fasciculus Bilateral (robust mean) Range value: 0.000571964, ..., 0.00113194
\item \strong{MD_TAP} \emph{numeric} Tapetum Bilateral (robust mean) Range value: 0.000852474, ..., 0.00225774
\item \strong{MD_IC} \emph{numeric} Full internal capsule (robust mean) Range value: 0.000576563, ..., 0.000953014
\item \strong{MD_CC} \emph{numeric} Full corpus callosum (robust mean) Range value: 0.000658579, ..., 0.00114916
\item \strong{MD_CR} \emph{numeric} Full corona radiata (robust mean) Range value: 0.000656279, ..., 0.00117547
\item \strong{MD_FULLWM} \emph{numeric} Full white matter (robust mean) Range value: 0.000626379, ..., 0.000953609
\item \strong{RD_MCP} \emph{numeric} Middle cerebellar peduncle (robust mean) Range value: 0.000351503, ..., 0.000644613
\item \strong{RD_PCT} \emph{numeric} Pontine crossing tract (robust mean) Range value: 0.000344147, ..., 0.000735081
\item \strong{RD_GCC} \emph{numeric} Genu of corpus callosum (robust mean) Range value: 0.000397507, ..., 0.00086999
\item \strong{RD_BCC} \emph{numeric} Body of corpus callosum (robust mean) Range value: 0.000389497, ..., 0.000888864
\item \strong{RD_SCC} \emph{numeric} Splenium of corpus callosum (robust mean) Range value: 0.000278346, ..., 0.000951786
\item \strong{RD_FX} \emph{numeric} Fornix (robust mean) Range value: 0.000726104, ..., 0.00289282
\item \strong{RD_CST_R} \emph{numeric} Corticospinal tract right (robust mean) Range value: 0.000260598, ..., 0.000768236
\item \strong{RD_CST_L} \emph{numeric} Corticospinal tract left (robust mean) Range value: 0.000270151, ..., 0.00071023
\item \strong{RD_ML_R} \emph{numeric} Medial lemniscus right (robust mean) Range value: 0.0002496, ..., 0.000782451
\item \strong{RD_ML_L} \emph{numeric} Medial lemniscus left (robust mean) Range value: 0.000196513, ..., 0.00085232
\item \strong{RD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right (robust mean) Range value: 0.000333829, ..., 0.00069156
\item \strong{RD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left (robust mean) Range value: 0.000327272, ..., 0.00076003
\item \strong{RD_SCP_R} \emph{numeric} Superior cerebellar peduncle right (robust mean) Range value: 0.000335188, ..., 0.00131033
\item \strong{RD_SCP_L} \emph{numeric} Superior cerebellar peduncle left (robust mean) Range value: 0.000326724, ..., 0.00111467
\item \strong{RD_CP_R} \emph{numeric} Cerebral peduncle right (robust mean) Range value: 0.000227326, ..., 0.00070077
\item \strong{RD_CP_L} \emph{numeric} Cerebral peduncle left (robust mean) Range value: 0.000208343, ..., 0.000675923
\item \strong{RD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right (robust mean) Range value: 0.000377413, ..., 0.000963219
\item \strong{RD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left (robust mean) Range value: 0.000336252, ..., 0.000951971
\item \strong{RD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right (robust mean) Range value: 0.000275269, ..., 0.000567746
\item \strong{RD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left (robust mean) Range value: 0.000291784, ..., 0.000632814
\item \strong{RD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right (robust mean) Range value: 0.000267253, ..., 0.000709532
\item \strong{RD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left (robust mean) Range value: 0.00036554, ..., 0.000698399
\item \strong{RD_ACR_R} \emph{numeric} Anterior corona radiata right (robust mean) Range value: 0.000509126, ..., 0.00107978
\item \strong{RD_ACR_L} \emph{numeric} Anterior corona radiata left (robust mean) Range value: 0.000520726, ..., 0.0010841
\item \strong{RD_SCR_R} \emph{numeric} Superior corona radiata right (robust mean) Range value: 0.000435126, ..., 0.000981095
\item \strong{RD_SCR_L} \emph{numeric} Superior corona radiata left (robust mean) Range value: 0.000418909, ..., 0.000922873
\item \strong{RD_PCR_R} \emph{numeric} Posterior corona radiata right (robust mean) Range value: 0.000484331, ..., 0.00118131
\item \strong{RD_PCR_L} \emph{numeric} Posterior corona radiata left (robust mean) Range value: 0.000499739, ..., 0.00118477
\item \strong{RD_PTR_R} \emph{numeric} Posterior thalamic radiation right (robust mean) Range value: 0.000406437, ..., 0.000872955
\item \strong{RD_PTR_L} \emph{numeric} Posterior thalamic radiation left (robust mean) Range value: 0.000445894, ..., 0.000963322
\item \strong{RD_SS_R} \emph{numeric} Sagittal stratum right (robust mean) Range value: 0.000428182, ..., 0.000827868
\item \strong{RD_SS_L} \emph{numeric} Sagittal stratum left (robust mean) Range value: 0.000395118, ..., 0.00104558
\item \strong{RD_EC_R} \emph{numeric} External capsule right (robust mean) Range value: 0.000495243, ..., 0.000939434
\item \strong{RD_EC_L} \emph{numeric} External capsule left (robust mean) Range value: 0.000463876, ..., 0.000916377
\item \strong{RD_CGC_R} \emph{numeric} Cingulum right (robust mean) Range value: 0.000446773, ..., 0.0007558
\item \strong{RD_CGC_L} \emph{numeric} Cingulum left (robust mean) Range value: 0.00043719, ..., 0.000746585
\item \strong{RD_CGH_R} \emph{numeric} Cingulum right (robust mean) Range value: 0.00047238, ..., 0.000876399
\item \strong{RD_CGH_L} \emph{numeric} Cingulum left (robust mean) Range value: 0.000415405, ..., 0.0010558
\item \strong{RD_FXST_R} \emph{numeric} Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000486854, ..., 0.00103208
\item \strong{RD_FXST_L} \emph{numeric} Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.000372604, ..., 0.000820852
\item \strong{RD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right (robust mean) Range value: 0.000423997, ..., 0.000855702
\item \strong{RD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left (robust mean) Range value: 0.000452811, ..., 0.000853233
\item \strong{RD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right (robust mean) Range value: 0.000457733, ..., 0.00116335
\item \strong{RD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left (robust mean) Range value: 0.000437973, ..., 0.00149681
\item \strong{RD_UNC_R} \emph{numeric} Uncinate fasciculus right (robust mean) Range value: 0.000418055, ..., 0.00107663
\item \strong{RD_UNC_L} \emph{numeric} Uncinate fasciculus left (robust mean) Range value: 0.000354935, ..., 0.000881413
\item \strong{RD_TAP_R} \emph{numeric} Tapetum right (robust mean) Range value: 0.000534338, ..., 0.00185332
\item \strong{RD_TAP_L} \emph{numeric} Tapetum left (robust mean) Range value: 0.000638001, ..., 0.00226463
\item \strong{RD_CST} \emph{numeric} Corticospinal tract Bilateral (robust mean) Range value: 0.000267712, ..., 0.000738526
\item \strong{RD_ML} \emph{numeric} Medial lemniscus Bilateral (robust mean) Range value: 0.000235951, ..., 0.000813956
\item \strong{RD_ICP} \emph{numeric} Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000331192, ..., 0.000717804
\item \strong{RD_SCP} \emph{numeric} Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.000341295, ..., 0.00120448
\item \strong{RD_CP} \emph{numeric} Cerebral peduncle Bilateral (robust mean) Range value: 0.000216174, ..., 0.000669671
\item \strong{RD_ALIC} \emph{numeric} Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.000390841, ..., 0.000958499
\item \strong{RD_PLIC} \emph{numeric} Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.000289528, ..., 0.000589173
\item \strong{RD_RLIC} \emph{numeric} Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.000362817, ..., 0.000687459
\item \strong{RD_ACR} \emph{numeric} Anterior corona radiata Bilateral (robust mean) Range value: 0.000515998, ..., 0.00101013
\item \strong{RD_SCR} \emph{numeric} Superior corona radiata Bilateral (robust mean) Range value: 0.00044216, ..., 0.000950696
\item \strong{RD_PCR} \emph{numeric} Posterior corona radiata Bilateral (robust mean) Range value: 0.00049104, ..., 0.00118257
\item \strong{RD_PTR} \emph{numeric} Posterior thalamic radiation Bilateral (robust mean) Range value: 0.000432238, ..., 0.000883545
\item \strong{RD_SS} \emph{numeric} Sagittal stratum Bilateral (robust mean) Range value: 0.000412026, ..., 0.000844668
\item \strong{RD_EC} \emph{numeric} External capsule Bilateral (robust mean) Range value: 0.000494361, ..., 0.000923274
\item \strong{RD_CGC} \emph{numeric} Cingulum Bilateral (robust mean) Range value: 0.000446508, ..., 0.00075128
\item \strong{RD_CGH} \emph{numeric} Cingulum Bilateral (robust mean) Range value: 0.000463489, ..., 0.000947194
\item \strong{RD_FXST} \emph{numeric} Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.00044311, ..., 0.000884391
\item \strong{RD_SLF} \emph{numeric} Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000442808, ..., 0.00085473
\item \strong{RD_SFO} \emph{numeric} Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000459779, ..., 0.00132708
\item \strong{RD_UNC} \emph{numeric} Uncinate fasciculus Bilateral (robust mean) Range value: 0.000437758, ..., 0.000960378
\item \strong{RD_TAP} \emph{numeric} Tapetum Bilateral (robust mean) Range value: 0.000594297, ..., 0.00203638
\item \strong{RD_IC} \emph{numeric} Full internal capsule (robust mean) Range value: 0.000349624, ..., 0.000680266
\item \strong{RD_CC} \emph{numeric} Full corpus callosum (robust mean) Range value: 0.000372826, ..., 0.000870901
\item \strong{RD_CR} \emph{numeric} Full corona radiata (robust mean) Range value: 0.000499836, ..., 0.00101819
\item \strong{RD_FULLWM} \emph{numeric} Full white matter (robust mean) Range value: 0.000429675, ..., 0.000771754
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{adni4}
\keyword{imaging}
