% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DTIROI_V1}
\alias{DTIROI_V1}
\title{USC - DTI ROI Summary Measures v1}
\format{
A data frame with 976 observations and 243 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(DTIROI_V1)
}
\description{
USC - DTI ROI Summary Measures v1 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 123, ..., 5295
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{numeric} Version Range value: 1, ..., 1
\item \strong{VISNAME} \emph{character}    
\item \strong{VISITNO} \emph{numeric} Visit Range value: 3, ..., 50
\item \strong{LONIUID_1} \emph{numeric} LONI Image ID for FA Range value: 299595, ..., 848440
\item \strong{LONIUID_2} \emph{numeric} LONI Image ID for MD Range value: 299814, ..., 848443
\item \strong{LONIUID_3} \emph{numeric} LONI Image ID for RD Range value: 359669, ..., 848449
\item \strong{LONIUID_4} \emph{numeric} LONI Image ID for AD Range value: 297755, ..., 848446
\item \strong{RUNDATE} \emph{ } Date Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Status , with options: Complete
\item \strong{AD_CST_L} \emph{numeric} Corticospinal tract left Range value: 0.000993997, ..., 0.0023359
\item \strong{AD_CST_R} \emph{numeric} Corticospinal tract right Range value: 0.000963274, ..., 0.00228733
\item \strong{AD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left Range value: 0.000976535, ..., 0.0022723
\item \strong{AD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right Range value: 0.000859119, ..., 0.00228807
\item \strong{AD_ML_L} \emph{numeric} Medial lemniscus left Range value: 0.00100278, ..., 0.00183072
\item \strong{AD_ML_R} \emph{numeric} Medial lemniscus right Range value: 0.0010317, ..., 0.00158208
\item \strong{AD_SCP_L} \emph{numeric} Superior cerebellar peduncle left Range value: 0.00115878, ..., 0.00207185
\item \strong{AD_SCP_R} \emph{numeric} Superior cerebellar peduncle right Range value: 0.00120185, ..., 0.00205651
\item \strong{AD_CP_L} \emph{numeric} Cerebral peduncle left Range value: 0.00129964, ..., 0.00184099
\item \strong{AD_CP_R} \emph{numeric} Cerebral peduncle right Range value: 0.00132518, ..., 0.00185971
\item \strong{AD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left Range value: 0.00100718, ..., 0.00164792
\item \strong{AD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right Range value: 0.00101425, ..., 0.00156695
\item \strong{AD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left Range value: 0.0010437, ..., 0.00142589
\item \strong{AD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right Range value: 0.00105059, ..., 0.00145844
\item \strong{AD_PTR_L} \emph{numeric} Posterior thalamic radiation left Range value: 0.00114353, ..., 0.0021751
\item \strong{AD_PTR_R} \emph{numeric} Posterior thalamic radiation right Range value: 0.00110727, ..., 0.00209163
\item \strong{AD_ACR_L} \emph{numeric} Anterior corona radiata left Range value: 0.000980971, ..., 0.00177484
\item \strong{AD_ACR_R} \emph{numeric} Anterior corona radiata right Range value: 0.00100739, ..., 0.00171728
\item \strong{AD_SCR_L} \emph{numeric} Superior corona radiata left Range value: 0.000923883, ..., 0.00178072
\item \strong{AD_SCR_R} \emph{numeric} Superior corona radiata right Range value: 0.000920291, ..., 0.00164695
\item \strong{AD_PCR_L} \emph{numeric} Posterior corona radiata left Range value: 0.000997919, ..., 0.00256759
\item \strong{AD_PCR_R} \emph{numeric} Posterior corona radiata right Range value: 0.000970576, ..., 0.00233851
\item \strong{AD_CGC_L} \emph{numeric} Cingulum left Range value: 0.000973597, ..., 0.00165317
\item \strong{AD_CGC_R} \emph{numeric} Cingulum right Range value: 0.000938041, ..., 0.00181551
\item \strong{AD_CGH_L} \emph{numeric} Cingulum (hippocampus) left Range value: 0.000940835, ..., 0.00171008
\item \strong{AD_CGH_R} \emph{numeric} Cingulum (hippocampus) right Range value: 0.000964048, ..., 0.00169439
\item \strong{AD_FX_ST_L} \emph{numeric} Fornix (cres) / Stria terminalis left Range value: 0.00117625, ..., 0.00258322
\item \strong{AD_FX_ST_R} \emph{numeric} Fornix (cres) / Stria terminalis right Range value: 0.00116545, ..., 0.00268004
\item \strong{AD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left Range value: 0.000921446, ..., 0.00132317
\item \strong{AD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right Range value: 0.000909681, ..., 0.00145328
\item \strong{AD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left Range value: 0.000879117, ..., 0.00295549
\item \strong{AD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right Range value: 0.00090537, ..., 0.00264967
\item \strong{AD_IFO_L} \emph{numeric} Inferior fronto-occipital fasciculus left Range value: 0.00103934, ..., 0.00174342
\item \strong{AD_IFO_R} \emph{numeric} Inferior fronto-occipital fasciculus right Range value: 0.00101582, ..., 0.00175759
\item \strong{AD_SS_L} \emph{numeric} Sagittal stratum left Range value: 0.00100127, ..., 0.00177597
\item \strong{AD_SS_R} \emph{numeric} Sagittal stratum right Range value: 0.00101029, ..., 0.00206854
\item \strong{AD_EC_L} \emph{numeric} External capsule left Range value: 0.000979624, ..., 0.00140867
\item \strong{AD_EC_R} \emph{numeric} External capsule right Range value: 0.000962904, ..., 0.00140934
\item \strong{AD_UNC_L} \emph{numeric} Uncinate fasciculus left Range value: 0.000914262, ..., 0.00240358
\item \strong{AD_UNC_R} \emph{numeric} Uncinate fasciculus right Range value: 0.000837411, ..., 0.00212123
\item \strong{AD_FX_L} \emph{numeric} Fornix  left Range value: 0.00158885, ..., 0.00422541
\item \strong{AD_FX_R} \emph{numeric} Fornix  right Range value: 0.00174225, ..., 0.00367729
\item \strong{AD_GCC_L} \emph{numeric} Genu of corpus callosum left Range value: 0.00138464, ..., 0.00235767
\item \strong{AD_GCC_R} \emph{numeric} Genu of corpus callosum right Range value: 0.00134543, ..., 0.00213844
\item \strong{AD_BCC_L} \emph{numeric} Body of corpus callosum  left Range value: 0.00148177, ..., 0.00312935
\item \strong{AD_BCC_R} \emph{numeric} Body of corpus callosum  right Range value: 0.0014899, ..., 0.0029777
\item \strong{AD_SCC_L} \emph{numeric} Splenium of corpus callosum left Range value: 0.00142734, ..., 0.00292163
\item \strong{AD_SCC_R} \emph{numeric} Splenium of corpus callosum right Range value: 0.00139539, ..., 0.00281836
\item \strong{AD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left Range value: 0.00106141, ..., 0.00152381
\item \strong{AD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right Range value: 0.00102322, ..., 0.00152645
\item \strong{AD_TAP_L} \emph{numeric} Tapatum left Range value: 0.00125543, ..., 0.00315392
\item \strong{AD_TAP_R} \emph{numeric} Tapatum right Range value: 0.0011004, ..., 0.00296885
\item \strong{AD_SUMGCC} \emph{numeric} Bilateral genu of the corpus callosum Range value: 0.00136834, ..., 0.00222859
\item \strong{AD_SUMBCC} \emph{numeric} Bilateral body of  the corpus callosum Range value: 0.00153413, ..., 0.003054
\item \strong{AD_SUMSCC} \emph{numeric} Bilateral splenium of  the corpus callosum Range value: 0.00143118, ..., 0.00277783
\item \strong{AD_SUMCC} \emph{numeric} Bilateral full corpus callosum Range value: 0.00148458, ..., 0.00261532
\item \strong{AD_SUMFX} \emph{numeric} Bilateral fornix Range value: 0.00166384, ..., 0.00377617
\item \strong{FA_CST_L} \emph{numeric} Corticospinal tract left Range value: 0.21429, ..., 0.478769
\item \strong{FA_CST_R} \emph{numeric} Corticospinal tract right Range value: 0.219521, ..., 0.50649
\item \strong{FA_ICP_L} \emph{numeric} Inferior cerebellar peduncle left Range value: 0.200001, ..., 0.540077
\item \strong{FA_ICP_R} \emph{numeric} Inferior cerebellar peduncle right Range value: 0.190979, ..., 0.543427
\item \strong{FA_ML_L} \emph{numeric} Medial lemniscus left Range value: 0.312138, ..., 0.600448
\item \strong{FA_ML_R} \emph{numeric} Medial lemniscus right Range value: 0.286779, ..., 0.58931
\item \strong{FA_SCP_L} \emph{numeric} Superior cerebellar peduncle left Range value: 0.29042, ..., 0.493868
\item \strong{FA_SCP_R} \emph{numeric} Superior cerebellar peduncle right Range value: 0.313576, ..., 0.528718
\item \strong{FA_CP_L} \emph{numeric} Cerebral peduncle left Range value: 0.348713, ..., 0.556213
\item \strong{FA_CP_R} \emph{numeric} Cerebral peduncle right Range value: 0.356891, ..., 0.548463
\item \strong{FA_ALIC_L} \emph{numeric} Anterior limb of internal capsule left Range value: 0.195564, ..., 0.467866
\item \strong{FA_ALIC_R} \emph{numeric} Anterior limb of internal capsule right Range value: 0.237606, ..., 0.492184
\item \strong{FA_PLIC_L} \emph{numeric} Posterior limb of internal capsule left Range value: 0.410757, ..., 0.634323
\item \strong{FA_PLIC_R} \emph{numeric} Posterior limb of internal capsule right Range value: 0.42251, ..., 0.619324
\item \strong{FA_PTR_L} \emph{numeric} Posterior thalamic radiation left Range value: 0.218698, ..., 0.464181
\item \strong{FA_PTR_R} \emph{numeric} Posterior thalamic radiation right Range value: 0.199405, ..., 0.479922
\item \strong{FA_ACR_L} \emph{numeric} Anterior corona radiata left Range value: 0.159149, ..., 0.401258
\item \strong{FA_ACR_R} \emph{numeric} Anterior corona radiata right Range value: 0.168502, ..., 0.422831
\item \strong{FA_SCR_L} \emph{numeric} Superior corona radiata left Range value: 0.243831, ..., 0.460448
\item \strong{FA_SCR_R} \emph{numeric} Superior corona radiata right Range value: 0.249534, ..., 0.417498
\item \strong{FA_PCR_L} \emph{numeric} Posterior corona radiata left Range value: 0.224325, ..., 0.462149
\item \strong{FA_PCR_R} \emph{numeric} Posterior corona radiata right Range value: 0.226174, ..., 0.476579
\item \strong{FA_CGC_L} \emph{numeric} Cingulum left Range value: 0.179103, ..., 0.375623
\item \strong{FA_CGC_R} \emph{numeric} Cingulum right Range value: 0.148849, ..., 0.35148
\item \strong{FA_CGH_L} \emph{numeric} Cingulum (hippocampus) left Range value: 0.141236, ..., 0.314793
\item \strong{FA_CGH_R} \emph{numeric} Cingulum (hippocampus) right Range value: 0.161659, ..., 0.362394
\item \strong{FA_FX_ST_L} \emph{numeric} Fornix (cres) / Stria terminalis left Range value: 0.14481, ..., 0.402883
\item \strong{FA_FX_ST_R} \emph{numeric} Fornix (cres) / Stria terminalis right Range value: 0.123952, ..., 0.378265
\item \strong{FA_SLF_L} \emph{numeric} Superior longitudinal fasciculus left Range value: 0.225005, ..., 0.413362
\item \strong{FA_SLF_R} \emph{numeric} Superior longitudinal fasciculus right Range value: 0.257282, ..., 0.42343
\item \strong{FA_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left Range value: 0.0821821, ..., 0.363496
\item \strong{FA_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right Range value: 0.0922206, ..., 0.398694
\item \strong{FA_IFO_L} \emph{numeric} Inferior fronto-occipital fasciculus left Range value: 0.181494, ..., 0.397715
\item \strong{FA_IFO_R} \emph{numeric} Inferior fronto-occipital fasciculus right Range value: 0.153711, ..., 0.396123
\item \strong{FA_SS_L} \emph{numeric} Sagittal stratum left Range value: 0.220469, ..., 0.396485
\item \strong{FA_SS_R} \emph{numeric} Sagittal stratum right Range value: 0.175595, ..., 0.404134
\item \strong{FA_EC_L} \emph{numeric} External capsule left Range value: 0.158739, ..., 0.339948
\item \strong{FA_EC_R} \emph{numeric} External capsule right Range value: 0.166009, ..., 0.346483
\item \strong{FA_UNC_L} \emph{numeric} Uncinate fasciculus left Range value: 0.0929704, ..., 0.310998
\item \strong{FA_UNC_R} \emph{numeric} Uncinate fasciculus right Range value: 0.111383, ..., 0.31548
\item \strong{FA_FX_L} \emph{numeric} Fornix  left Range value: 0.039708, ..., 0.335833
\item \strong{FA_FX_R} \emph{numeric} Fornix  right Range value: 0.0640862, ..., 0.352706
\item \strong{FA_GCC_L} \emph{numeric} Genu of corpus callosum left Range value: 0.118239, ..., 0.585558
\item \strong{FA_GCC_R} \emph{numeric} Genu of corpus callosum right Range value: 0.128932, ..., 0.602886
\item \strong{FA_BCC_L} \emph{numeric} Body of corpus callosum  left Range value: 0.110401, ..., 0.530601
\item \strong{FA_BCC_R} \emph{numeric} Body of corpus callosum  right Range value: 0.118927, ..., 0.530068
\item \strong{FA_SCC_L} \emph{numeric} Splenium of corpus callosum left Range value: 0.146024, ..., 0.603467
\item \strong{FA_SCC_R} \emph{numeric} Splenium of corpus callosum right Range value: 0.126243, ..., 0.604542
\item \strong{FA_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left Range value: 0.313875, ..., 0.501661
\item \strong{FA_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right Range value: 0.334446, ..., 0.520078
\item \strong{FA_TAP_L} \emph{numeric} Tapatum left Range value: 0.0289957, ..., 0.409767
\item \strong{FA_TAP_R} \emph{numeric} Tapatum right Range value: 0.0587773, ..., 0.47323
\item \strong{FA_SUMGCC} \emph{numeric} Bilateral genu of the corpus callosum Range value: 0.12678, ..., 0.587242
\item \strong{FA_SUMBCC} \emph{numeric} Bilateral body of  the corpus callosum Range value: 0.11513, ..., 0.53032
\item \strong{FA_SUMSCC} \emph{numeric} Bilateral splenium of  the corpus callosum Range value: 0.148758, ..., 0.602003
\item \strong{FA_SUMCC} \emph{numeric} Bilateral full corpus callosum Range value: 0.19629, ..., 0.570519
\item \strong{FA_SUMFX} \emph{numeric} Bilateral fornix Range value: 0.065717, ..., 0.329724
\item \strong{MD_CST_L} \emph{numeric} Corticospinal tract left Range value: 0.000680956, ..., 0.00204066
\item \strong{MD_CST_R} \emph{numeric} Corticospinal tract right Range value: 0.000626507, ..., 0.00197098
\item \strong{MD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left Range value: 0.000700378, ..., 0.00187846
\item \strong{MD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right Range value: 0.000662808, ..., 0.00193251
\item \strong{MD_ML_L} \emph{numeric} Medial lemniscus left Range value: 0.000666465, ..., 0.00145528
\item \strong{MD_ML_R} \emph{numeric} Medial lemniscus right Range value: 0.000642953, ..., 0.00119683
\item \strong{MD_SCP_L} \emph{numeric} Superior cerebellar peduncle left Range value: 0.000793968, ..., 0.00146032
\item \strong{MD_SCP_R} \emph{numeric} Superior cerebellar peduncle right Range value: 0.000758536, ..., 0.00147208
\item \strong{MD_CP_L} \emph{numeric} Cerebral peduncle left Range value: 0.000794933, ..., 0.00127232
\item \strong{MD_CP_R} \emph{numeric} Cerebral peduncle right Range value: 0.000848849, ..., 0.00128908
\item \strong{MD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left Range value: 0.000684958, ..., 0.0014214
\item \strong{MD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right Range value: 0.000683233, ..., 0.00117625
\item \strong{MD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left Range value: 0.000592406, ..., 0.000932953
\item \strong{MD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right Range value: 0.000592934, ..., 0.000897941
\item \strong{MD_PTR_L} \emph{numeric} Posterior thalamic radiation left Range value: 0.000776136, ..., 0.00182813
\item \strong{MD_PTR_R} \emph{numeric} Posterior thalamic radiation right Range value: 0.000746882, ..., 0.00180862
\item \strong{MD_ACR_L} \emph{numeric} Anterior corona radiata left Range value: 0.000724209, ..., 0.00152839
\item \strong{MD_ACR_R} \emph{numeric} Anterior corona radiata right Range value: 0.000742919, ..., 0.0014598
\item \strong{MD_SCR_L} \emph{numeric} Superior corona radiata left Range value: 0.000667851, ..., 0.00138808
\item \strong{MD_SCR_R} \emph{numeric} Superior corona radiata right Range value: 0.000650919, ..., 0.00123191
\item \strong{MD_PCR_L} \emph{numeric} Posterior corona radiata left Range value: 0.000701259, ..., 0.00220364
\item \strong{MD_PCR_R} \emph{numeric} Posterior corona radiata right Range value: 0.000698319, ..., 0.00187262
\item \strong{MD_CGC_L} \emph{numeric} Cingulum left Range value: 0.000733628, ..., 0.00129201
\item \strong{MD_CGC_R} \emph{numeric} Cingulum right Range value: 0.000719108, ..., 0.00150079
\item \strong{MD_CGH_L} \emph{numeric} Cingulum (hippocampus) left Range value: 0.000741311, ..., 0.00148218
\item \strong{MD_CGH_R} \emph{numeric} Cingulum (hippocampus) right Range value: 0.000752325, ..., 0.0014515
\item \strong{MD_FX_ST_L} \emph{numeric} Fornix (cres) / Stria terminalis left Range value: 0.000859886, ..., 0.00229299
\item \strong{MD_FX_ST_R} \emph{numeric} Fornix (cres) / Stria terminalis right Range value: 0.000864476, ..., 0.00240514
\item \strong{MD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left Range value: 0.000669879, ..., 0.00101601
\item \strong{MD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right Range value: 0.000640319, ..., 0.00102618
\item \strong{MD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left Range value: 0.000639704, ..., 0.00257162
\item \strong{MD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right Range value: 0.000626051, ..., 0.00234641
\item \strong{MD_IFO_L} \emph{numeric} Inferior fronto-occipital fasciculus left Range value: 0.000766934, ..., 0.00147858
\item \strong{MD_IFO_R} \emph{numeric} Inferior fronto-occipital fasciculus right Range value: 0.000752035, ..., 0.00151579
\item \strong{MD_SS_L} \emph{numeric} Sagittal stratum left Range value: 0.000743559, ..., 0.00140108
\item \strong{MD_SS_R} \emph{numeric} Sagittal stratum right Range value: 0.000746493, ..., 0.00182045
\item \strong{MD_EC_L} \emph{numeric} External capsule left Range value: 0.000745804, ..., 0.00114929
\item \strong{MD_EC_R} \emph{numeric} External capsule right Range value: 0.000718573, ..., 0.00113225
\item \strong{MD_UNC_L} \emph{numeric} Uncinate fasciculus left Range value: 0.000748275, ..., 0.00222833
\item \strong{MD_UNC_R} \emph{numeric} Uncinate fasciculus right Range value: 0.000672916, ..., 0.00188058
\item \strong{MD_FX_L} \emph{numeric} Fornix  left Range value: 0.00122076, ..., 0.00367397
\item \strong{MD_FX_R} \emph{numeric} Fornix  right Range value: 0.00127621, ..., 0.00315862
\item \strong{MD_GCC_L} \emph{numeric} Genu of corpus callosum left Range value: 0.000841326, ..., 0.00201157
\item \strong{MD_GCC_R} \emph{numeric} Genu of corpus callosum right Range value: 0.000833173, ..., 0.00174139
\item \strong{MD_BCC_L} \emph{numeric} Body of corpus callosum  left Range value: 0.00094121, ..., 0.00280511
\item \strong{MD_BCC_R} \emph{numeric} Body of corpus callosum  right Range value: 0.000917638, ..., 0.00266126
\item \strong{MD_SCC_L} \emph{numeric} Splenium of corpus callosum left Range value: 0.000850511, ..., 0.00260368
\item \strong{MD_SCC_R} \emph{numeric} Splenium of corpus callosum right Range value: 0.000798887, ..., 0.00257251
\item \strong{MD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left Range value: 0.000683271, ..., 0.0010464
\item \strong{MD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right Range value: 0.000641464, ..., 0.00112352
\item \strong{MD_TAP_L} \emph{numeric} Tapatum left Range value: 0.00089972, ..., 0.00293113
\item \strong{MD_TAP_R} \emph{numeric} Tapatum right Range value: 0.000732027, ..., 0.00277181
\item \strong{MD_SUMGCC} \emph{numeric} Bilateral genu of the corpus callosum Range value: 0.000846759, ..., 0.00186883
\item \strong{MD_SUMBCC} \emph{numeric} Bilateral body of  the corpus callosum Range value: 0.000943302, ..., 0.00271543
\item \strong{MD_SUMSCC} \emph{numeric} Bilateral splenium of  the corpus callosum Range value: 0.000830337, ..., 0.00249503
\item \strong{MD_SUMCC} \emph{numeric} Bilateral full corpus callosum Range value: 0.000884068, ..., 0.00226628
\item \strong{MD_SUMFX} \emph{numeric} Bilateral fornix Range value: 0.00126744, ..., 0.00325246
\item \strong{RD_CST_L} \emph{numeric} Corticospinal tract left Range value: 0.000513187, ..., 0.00189304
\item \strong{RD_CST_R} \emph{numeric} Corticospinal tract right Range value: 0.000458123, ..., 0.0018128
\item \strong{RD_ICP_L} \emph{numeric} Inferior cerebellar peduncle left Range value: 0.0005623, ..., 0.00168154
\item \strong{RD_ICP_R} \emph{numeric} Inferior cerebellar peduncle right Range value: 0.000502488, ..., 0.00175473
\item \strong{RD_ML_L} \emph{numeric} Medial lemniscus left Range value: 0.000407372, ..., 0.00126756
\item \strong{RD_ML_R} \emph{numeric} Medial lemniscus right Range value: 0.000423634, ..., 0.0010048
\item \strong{RD_SCP_L} \emph{numeric} Superior cerebellar peduncle left Range value: 0.000606437, ..., 0.00122119
\item \strong{RD_SCP_R} \emph{numeric} Superior cerebellar peduncle right Range value: 0.000536879, ..., 0.00120509
\item \strong{RD_CP_L} \emph{numeric} Cerebral peduncle left Range value: 0.000542581, ..., 0.00104121
\item \strong{RD_CP_R} \emph{numeric} Cerebral peduncle right Range value: 0.000587045, ..., 0.00105528
\item \strong{RD_ALIC_L} \emph{numeric} Anterior limb of internal capsule left Range value: 0.000508444, ..., 0.00130813
\item \strong{RD_ALIC_R} \emph{numeric} Anterior limb of internal capsule right Range value: 0.000483792, ..., 0.00104695
\item \strong{RD_PLIC_L} \emph{numeric} Posterior limb of internal capsule left Range value: 0.000366757, ..., 0.00069468
\item \strong{RD_PLIC_R} \emph{numeric} Posterior limb of internal capsule right Range value: 0.000352689, ..., 0.000649413
\item \strong{RD_PTR_L} \emph{numeric} Posterior thalamic radiation left Range value: 0.000573759, ..., 0.00165465
\item \strong{RD_PTR_R} \emph{numeric} Posterior thalamic radiation right Range value: 0.000539348, ..., 0.00166712
\item \strong{RD_ACR_L} \emph{numeric} Anterior corona radiata left Range value: 0.00058294, ..., 0.00140517
\item \strong{RD_ACR_R} \emph{numeric} Anterior corona radiata right Range value: 0.000586911, ..., 0.00133106
\item \strong{RD_SCR_L} \emph{numeric} Superior corona radiata left Range value: 0.000524569, ..., 0.00119176
\item \strong{RD_SCR_R} \emph{numeric} Superior corona radiata right Range value: 0.000509481, ..., 0.00102439
\item \strong{RD_PCR_L} \emph{numeric} Posterior corona radiata left Range value: 0.000536286, ..., 0.00202166
\item \strong{RD_PCR_R} \emph{numeric} Posterior corona radiata right Range value: 0.000536925, ..., 0.00163967
\item \strong{RD_CGC_L} \emph{numeric} Cingulum left Range value: 0.000597966, ..., 0.00111142
\item \strong{RD_CGC_R} \emph{numeric} Cingulum right Range value: 0.000598527, ..., 0.00134343
\item \strong{RD_CGH_L} \emph{numeric} Cingulum (hippocampus) left Range value: 0.00063963, ..., 0.00136824
\item \strong{RD_CGH_R} \emph{numeric} Cingulum (hippocampus) right Range value: 0.000625422, ..., 0.00133005
\item \strong{RD_FX_ST_L} \emph{numeric} Fornix (cres) / Stria terminalis left Range value: 0.000701229, ..., 0.00214788
\item \strong{RD_FX_ST_R} \emph{numeric} Fornix (cres) / Stria terminalis right Range value: 0.000711395, ..., 0.00226769
\item \strong{RD_SLF_L} \emph{numeric} Superior longitudinal fasciculus left Range value: 0.000529939, ..., 0.000892315
\item \strong{RD_SLF_R} \emph{numeric} Superior longitudinal fasciculus right Range value: 0.000503521, ..., 0.000872657
\item \strong{RD_SFO_L} \emph{numeric} Superior fronto-occipital fasciculus left Range value: 0.000519998, ..., 0.002456
\item \strong{RD_SFO_R} \emph{numeric} Superior fronto-occipital fasciculus right Range value: 0.000486392, ..., 0.00219478
\item \strong{RD_IFO_L} \emph{numeric} Inferior fronto-occipital fasciculus left Range value: 0.000605482, ..., 0.00134616
\item \strong{RD_IFO_R} \emph{numeric} Inferior fronto-occipital fasciculus right Range value: 0.000599599, ..., 0.00139489
\item \strong{RD_SS_L} \emph{numeric} Sagittal stratum left Range value: 0.000580973, ..., 0.00126481
\item \strong{RD_SS_R} \emph{numeric} Sagittal stratum right Range value: 0.000593587, ..., 0.00169641
\item \strong{RD_EC_L} \emph{numeric} External capsule left Range value: 0.000628894, ..., 0.00103389
\item \strong{RD_EC_R} \emph{numeric} External capsule right Range value: 0.000592118, ..., 0.000993699
\item \strong{RD_UNC_L} \emph{numeric} Uncinate fasciculus left Range value: 0.000642883, ..., 0.0021407
\item \strong{RD_UNC_R} \emph{numeric} Uncinate fasciculus right Range value: 0.000590668, ..., 0.00176026
\item \strong{RD_FX_L} \emph{numeric} Fornix  left Range value: 0.00100233, ..., 0.00339825
\item \strong{RD_FX_R} \emph{numeric} Fornix  right Range value: 0.00104319, ..., 0.00296152
\item \strong{RD_GCC_L} \emph{numeric} Genu of corpus callosum left Range value: 0.000518407, ..., 0.00183852
\item \strong{RD_GCC_R} \emph{numeric} Genu of corpus callosum right Range value: 0.000520721, ..., 0.00158473
\item \strong{RD_BCC_L} \emph{numeric} Body of corpus callosum  left Range value: 0.000638754, ..., 0.00264299
\item \strong{RD_BCC_R} \emph{numeric} Body of corpus callosum  right Range value: 0.000610442, ..., 0.00252815
\item \strong{RD_SCC_L} \emph{numeric} Splenium of corpus callosum left Range value: 0.000524191, ..., 0.0024447
\item \strong{RD_SCC_R} \emph{numeric} Splenium of corpus callosum right Range value: 0.000500637, ..., 0.00244959
\item \strong{RD_RLIC_L} \emph{numeric} Retrolenticular part of internal capsule left Range value: 0.000491442, ..., 0.000837773
\item \strong{RD_RLIC_R} \emph{numeric} Retrolenticular part of internal capsule right Range value: 0.000450588, ..., 0.000922053
\item \strong{RD_TAP_L} \emph{numeric} Tapatum left Range value: 0.000721867, ..., 0.00288428
\item \strong{RD_TAP_R} \emph{numeric} Tapatum right Range value: 0.00054784, ..., 0.00269725
\item \strong{RD_SUMGCC} \emph{numeric} Bilateral genu of the corpus callosum Range value: 0.00053028, ..., 0.00168895
\item \strong{RD_SUMBCC} \emph{numeric} Bilateral body of  the corpus callosum Range value: 0.000637537, ..., 0.00257343
\item \strong{RD_SUMSCC} \emph{numeric} Bilateral splenium of  the corpus callosum Range value: 0.000523085, ..., 0.00235363
\item \strong{RD_SUMCC} \emph{numeric} Bilateral full corpus callosum Range value: 0.000569854, ..., 0.00209175
\item \strong{RD_SUMFX} \emph{numeric} Bilateral fornix Range value: 0.00105877, ..., 0.00299716
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
