% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-data-prep.R
\name{get_numeric_dx_status}
\alias{get_numeric_dx_status}
\alias{convert_numeric_dx_status}
\title{Get/ Convert DX Status}
\usage{
get_numeric_dx_status(x)

convert_numeric_dx_status(x)
}
\arguments{
\item{x}{String}
}
\value{
Returns a numeric vector for `get_numeric_dx_status` and a character vector for
 `convert_numeric_dx_status`.
}
\description{
`get_numeric_dx_status` and `convert_numeric_dx_status` functions are used to
 convert either character diagnostics status into numeric values or
 numeric values into the corresponding diagnostics status, respectively.
}
\details{
The input value, `x`, must be a string value for `get_numeric_dx_status` and
 numeric value for `convert_numeric_dx_status`.
}
\examples{
\dontrun{
get_numeric_dx_status(x = c("CN", "MCI", "DEM"))
convert_numeric_dx_status(x = c(1, 2, 3, 5))
}
}
\seealso{
Other Utility functions used for preparing analysis data: 
\code{\link{convert_number_days}()}
}
\concept{Utility functions used for preparing analysis data}
\keyword{utils_fun}
