% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DXSUM}
\alias{DXSUM}
\title{Diagnostic Summary}
\format{
A data frame with 15881 observations and 42 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(DXSUM)
}
\description{
Diagnostic Summary data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Date Form Completed Date: YYY-MM-DD
\item \strong{DIAGNOSIS} \emph{character} 1.  Which best describes the participant's current diagnosis? 1=CN ; 2=MCI ; 3=Dementia
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXNORM} \emph{character} Normal 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXNODEP} \emph{character} Mild Depression 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMCI} \emph{character} Mild Cognitive Impairment 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMDES} \emph{character} 2a.  MCI features (select all that apply): 
\itemize{
\item \emph{ADNI1:} 1=MCI (Memory features) ; 2=MCI (Non-memory features)
\item \emph{ADNIGO, ADNI2, ADNI3, ADNI4:} 1=MCI - Memory features (amnestic) ; 2=MCI - Non-memory features (non-amnestic)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMPTR1} \emph{character} i. Subjective memory complaint 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMPTR2} \emph{character} ii. Informant memory complaint 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMPTR3} \emph{character} iii. Normal general cognitive function 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No ; 2=Marginal
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes ; 2=Marginal
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMPTR4} \emph{character} iv. Normal activities of daily living 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No ; 2=Marginal
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes ; 2=Marginal
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMPTR5} \emph{character} v. Objective memory impairment for age and education 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMPTR6} \emph{character} vi. Not demented by diagnostic criteria 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMDUE} \emph{character} 2b. Suspected cause of MCI: 1=MCI due to Alzheimer's Disease ; 2=MCI due to other etiology
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXMOTHET} \emph{character} If MCI due to other etiology, select box(es) to indicate reason: 
\itemize{
\item \emph{ADNI1:} 1=Fronto-temporal Dementia ; 2=Parkinson's Disease ; 3=Huntington's Disease ; 4=Progressive Supranuclear Palsy ; 8=Corticobasal Degeneration ; 9=Vascular Dementia ; 10=Prion-Associated Dementia ; 14=Other (specify)
\item \emph{ADNIGO, ADNI2, ADNI3:} 1=Fronto-temporal Dementia ; 2=Parkinson's Disease ; 3=Huntington's Disease ; 4=Progressive Supranuclear Palsy ; 5=Alcohol-related Dementia ; 6=NPH ; 7=Major Depression ; 8=Corticobasal Degeneration ; 9=Vascular Dementia ; 10=Prion-Associated Dementia ; 11=HIV ; 12=Primary Progressive Aphasia ; 13=Posterior Cortical Dysfunction ; 14=Other (specify)
\item \emph{ADNI4:} 1=Fronto-temporal Dementia ; 10=Prion-Associated Dementia ; 11=HIV ; 12=Primary Progressive Aphasia ; 13=Posterior Cortical Dysfunction ; 14=Other (specify) ; 2=Parkinson's Disease ; 3=Huntington's Disease ; 4=Progressive Supranuclear Palsy ; 5=Alcohol-related Dementia ; 6=NPH ; 7=Major Depression ; 8=Corticobasal Degeneration ; 9=Vascular Dementia
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXDSEV} \emph{character} 3a.  Dementia Severity - Clinician's Impression 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXDDUE} \emph{character} 3b.  Suspected cause of dementia 1=Dementia due to Alzheimer's Disease ; 2=Dementia due to other etiology
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXAD} \emph{character} Alzheimer's Disease 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXAPP} \emph{character} If Dementia due to Alzheimer's Disease, indicate likelihood: 1=Probable ; 2=Possible
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXAPROB} \emph{character} If Probable AD, select box(es) for other symptoms present: 1=None ; 2=Stroke(s) ; 3=Depression ; 4=Delirium ; 5=Parkinsonism ; 6=Metabolic/Toxic Disorder (specify) ; 7=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXAPOSS} \emph{character} If Possible AD, select box(es) to indicate reason: 1=Atypical clinical course or features (specify) ; 2=Stroke(s) ; 3=Depression ; 4=Delirium ; 5=Parkinsonism ; 6=Metabolic / Toxic Disorder (specify) ; 7=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXPARK} \emph{character} 4b.  Parkinsonism symptoms present? 
\itemize{
\item \emph{ADNI1:} 1=Yes
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXPDES} \emph{character} If Parkinsonism, select box which indicates best diagnosis: 1=Parkinsonism without cognitive impairment ; 2=Parkinsonism with cognitive impairment, not demented ; 3=Parkinsonism with cognitive impairment, demented ; 4=Atypical Parkinsonism
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXPCOG} \emph{character} If Parkinsonism with cognitive impairment, demented 1=PD ; 2=PDD ; 3=DLB ; 4=PDAD
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXPATYP} \emph{character} If Atypical Parkinsonism 1=PSP ; 2=CBGD ; 3=OPCA ; 4=SND ; 5=Shy Drager ; 6=Vascular ; 7=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXDEP} \emph{character} 4a. Depressive symptoms present? 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXOTHDEM} \emph{character} Other Dementia (not Alzheimer's Disease) 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXODES} \emph{character} If dementia due to other etiology, select best diagnosis: 
\itemize{
\item \emph{ADNI1:} 1=Frontal ; 3=Huntington's Disease ; 5=Alcohol-related Dementia ; 6=NPH ; 7=Major Depression ; 9=Vascular Dementia ; 10=Prion-Associated Dementia ; 11=HIV ; 12=Primary Progressive Aphasia ; 13=Posterior Cortical Dysfunction ; 14=Other (specify)
\item \emph{ADNIGO, ADNI2, ADNI3:} 1=Fronto-temporal Dementia ; 2=Parkinson's Disease ; 3=Huntington's Disease ; 4=Progressive Supranuclear Palsy ; 5=Alcohol-related Dementia ; 6=NPH ; 7=Major Depression ; 8=Corticobasal Degeneration ; 9=Vascular Dementia ; 10=Prion-Associated Dementia ; 11=HIV ; 12=Primary Progressive Aphasia ; 13=Posterior Cortical Dysfunction ; 14=Other (specify)
\item \emph{ADNI4:} 1=Fronto-temporal Dementia ; 10=Prion-Associated Dementia ; 11=HIV ; 12=Primary Progressive Aphasia ; 13=Posterior Cortical Dysfunction ; 14=Other (specify) ; 2=Parkinson's Disease ; 3=Huntington's Disease ; 4=Progressive Supranuclear Palsy ; 5=Alcohol-related Dementia ; 6=NPH ; 7=Major Depression ; 8=Corticobasal Degeneration ; 9=Vascular Dementia
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DXCONFID} \emph{character} Physician Confidence in Diagnosis: 1=Uncertain ; 2=Mildly Confident ; 3=Moderately Confident ; 4=Highly Confident
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 2, ..., 187120
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
