% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ECOG12SP}
\alias{ECOG12SP}
\title{Everyday Cognition 12 Item - Study Partner Report}
\format{
A data frame with 1230 observations and 29 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ECOG12SP)
}
\description{
Everyday Cognition 12 Item - Study Partner Report data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI2, ADNI3, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_bl, 4_init, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{COLDATE} \emph{ } Date collected Date: YYY-MM-DD
\item \strong{SPID} \emph{numeric} Study Partner ID: Range value: 1, ..., 10
\item \strong{SOURCE} \emph{character} Information Source: 1=Participant Visit ; 2=Telephone Call/Video Call ; 3=Mail
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP1} \emph{character} 1. Remembering where they have placed personal items or objects: 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP2} \emph{character} 2. Remembering the current date or day of the week 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP3} \emph{character} 3. Communicating or expressing ideas in a conversation 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP4} \emph{character} 4. Understanding spoken directions or instructions 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP5} \emph{character} 5. Difficulty calculating distances when driving or driving too close to another vehicle 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP6} \emph{character} 6. Finding their way around a familiar home 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP7} \emph{character} 7. Anticipating weather changes and planning accordingly 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP8} \emph{character} 8. Planning a trip, vacation, or outing 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP9} \emph{character} 9. Keeping living and work space (such as a home office, garage, or pantry) organized 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP10} \emph{character} 10. Managing bill payments 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP11} \emph{character} 11. Doing two things at once 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ECOGSP12} \emph{character} 12. Working on a particular task while talking to someone at the same time 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 1282, ..., 128274
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
