% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ECOGPT}
\alias{ECOGPT}
\title{Everyday Cognition - Participant Self-Report}
\format{
A data frame with 8093 observations and 63 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(ECOGPT)
}
\description{
Everyday Cognition - Participant Self-Report data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 7125
\item \strong{VISCODE} \emph{character}    
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment COLDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{CONCERN} \emph{character} Are you concerned that you have a memory or other thinking problem? 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEMORY1} \emph{character} 1. Remembering a few shopping items without a list. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEMORY2} \emph{character} 2. Remembering things that happened recently (such as recent outings, events in the news). 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEMORY3} \emph{character} 3. Recalling conversations a few days later. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEMORY4} \emph{character} 4. Remembering where I have placed objects. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEMORY5} \emph{character} 5. Repeating stories and/or questions. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEMORY6} \emph{character} 6. Remembering the current date or day of the week. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEMORY7} \emph{character} 7. Remembering I have already told someone something. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEMORY8} \emph{character} 8. Remembering appointments, meetings, or engagements. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG1} \emph{character} 1. Forgetting the names of objects. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG2} \emph{character} 2. Verbally giving instructions to others. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG3} \emph{character} 3. Finding the right words to use in conversations. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG4} \emph{character} 4. Communicating thoughts in a conversation. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG5} \emph{character} 5. Following a story in a book or on TV. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG6} \emph{character} 6. Understanding the point of what other people are trying to say. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG7} \emph{character} 7. Remembering the meaning of common words. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG8} \emph{character} 8. Describing a program I have watched on TV. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{LANG9} \emph{character} 9. Understanding spoken directions or instructions. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISSPAT1} \emph{character} 1. Following a map to find a new location. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISSPAT2} \emph{character} 2. Reading a map and helping with directions when someone else is driving. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISSPAT3} \emph{character} 3. Finding my car in a parking lot. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISSPAT4} \emph{character} 4. Finding my way back to a meeting spot in the mall or other location. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISSPAT5} \emph{character} 5. Finding my way around a familiar neighborhood. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISSPAT6} \emph{character} 6. Finding my way around a familiar store. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISSPAT7} \emph{character} 7. Finding my way around a house visited many times. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISSPAT8} \emph{character} 7. Finding my way around a house visited many times. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{PLAN1} \emph{character} 1. Planning a sequence of stops on a shopping trip. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{PLAN2} \emph{character} 2. The ability to anticipate weather changes and plan accordingly (i.e., bring a coat or umbrella) 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{PLAN3} \emph{character} 3. Developing a schedule in advance of anticipated events. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{PLAN4} \emph{character} 4. Thinking things through before acting. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{PLAN5} \emph{character} 5. Thinking ahead. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{ORGAN1} \emph{character} 1. Keeping living and work space organized. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{ORGAN2} \emph{character} 2. Balancing the checkbook without error. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{ORGAN3} \emph{character} 3. Keeping financial records organized. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{ORGAN4} \emph{character} 4. Prioritizing tasks by importance. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{ORGAN5} \emph{character} 5. Keeping mail and papers organized. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{ORGAN6} \emph{character} 6. Using an organized strategy to manage a medication schedule involving multiple medications. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{DIVATT1} \emph{character} 1. The ability to do two things at once. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{DIVATT2} \emph{character} 2. Returning to a task after being interrupted. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{DIVATT3} \emph{character} 3. The ability to concentrate on a task without being distracted by external things in the environment. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{DIVATT4} \emph{character} 4. Cooking or working and talking at the same time. 1=1- Better or no change ; 2=2- Questionable/occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{STAFFASST} \emph{character} 1. How much assistance was needed from the research staff to complete this form? 1=None ; 2=A little ; 3=A lot ; 4=Participant too cognitively impaired to complete
\emph{Decoded Value: }\strong{Yes}
\item \strong{VALIDITY} \emph{character} 2. Validity of the information collected from the participant (based on research assistant's observations and interactions with the participant while they were completing the ECog): 1=1- Appeared Valid--The participant did not express any confusion or ask for any clarification in completing the ECog. ; 2=2- Questionable Validity--The participant needed to have the instructions repeated, and/or needed help in understanding a few of the items. ; 3=3- Appeared Invalid--The participant was clearly confused about the content of many of the questions or how to complete the form in general. ; 4=No information collected from participant
\emph{Decoded Value: }\strong{Yes}
\item \strong{SOURCE} \emph{character} Information Source: 1=Participant Visit ; 2=Telephone Call ; 3=Mail
\emph{Decoded Value: }\strong{Yes}
\item \strong{EcogPtDivatt} \emph{numeric} Participant ECog - Div atten 1..4
\item \strong{EcogPtLang} \emph{numeric} Participant ECog - Lang 1..4
\item \strong{EcogPtMem} \emph{numeric} Participant ECog - Mem 1..4
\item \strong{EcogPtOrgan} \emph{numeric} Participant ECog - Organ 1..4
\item \strong{EcogPtPlan} \emph{numeric} Participant ECog - Plan 1..4
\item \strong{EcogPtVisspat} \emph{numeric} Participant ECog - Vis//Spat 1..4
\item \strong{EcogPtTotal} \emph{numeric} Participant ECog - Total 1..4
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 187109
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 63
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{assessments}
