% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMBICDCB}
\alias{EMBICDCB}
\title{Embic Corporation – Digital Cognitive Biomarkers}
\format{
A data frame with 10398 observations and 14 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(EMBICDCB)
}
\description{
Embic Corporation – Digital Cognitive Biomarkers data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 7016
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{N1} \emph{numeric} Probability of encoding into the DURABLY LEARNED State in(0,1)
\item \strong{N2} \emph{numeric} Probability of encoding into the TRANSIENTLY LEARNED State in(0,1)
\item \strong{N3} \emph{numeric} Probability of encoding into the DURABLY LEARNED State, following previous TRANSIENT LEARNING (N2) in(0,1)
\item \strong{N4} \emph{numeric} Probability of encoding into the DURABLY LEARNED State, due to successful retrieval (R1) from the TRANSIENTLY LEARNED State in(0,1)
\item \strong{R1} \emph{numeric} Probability of retrieving from the TRANSIENTLY LEARNED State in(0,1)
\item \strong{R2} \emph{numeric} Probability of retrieving from the DURABLY LEARNED State in(0,1)
\item \strong{R3} \emph{numeric} Probability of retrieving from the DURABLY LEARNED State after a 5-minute delay with distraction in(0,1)
\item \strong{M1} \emph{numeric} Probability of immediate recall of a non-durably stored episodic memory in(0,1)
\item \strong{M2} \emph{numeric} Probability of immediate recall of a durably stored episodic memory in(0,1)
\item \strong{M3} \emph{numeric} Probability of delayed recall of a durably stored episodic memory in(0,1)
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
