% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMBICqCP}
\alias{EMBICqCP}
\title{Embic Corporation Quantified Cognitive Processes}
\format{
A data frame with 10868 observations and 32 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(EMBICqCP)
}
\description{
Embic Corporation Quantified Cognitive Processes data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 7121
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{N1M} \emph{numeric} Latent probability of encoding into the DURABLY LEARNED State, mean of posterior distribution Range value: 0.247471986, ..., 0.887025596
\item \strong{N1SD} \emph{numeric} Latent probability of encoding into the DURABLY LEARNED State, standard deviation of posterior distribution Range value: 0.090024432, ..., 0.216397064
\item \strong{N2M} \emph{numeric} Latent probability of encoding into the TRANSIENTLY LEARNED State, mean of posterior distribution Range value: 0.526028145, ..., 0.881808601
\item \strong{N2SD} \emph{numeric} Latent probability of encoding into the TRANSIENTLY LEARNED State, standard deviation of posterior distribution Range value: 0.095430324, ..., 0.280335425
\item \strong{N3M} \emph{numeric} Latent probability of encoding into the DURABLY LEARNED State, following previous TRANSIENT LEARNING (N2), mean of posterior distribution Range value: 0.102276577, ..., 0.60547497
\item \strong{N3SD} \emph{numeric} Latent probability of encoding into the DURABLY LEARNED State, following previous TRANSIENT LEARNING (N2), standard deviation of posterior distribution Range value: 0.094818984, ..., 0.197688929
\item \strong{N4M} \emph{numeric} Latent probability of encoding into the DURABLY LEARNED State, due to successful retrieval (R1) from the TRANSIENTLY LEARNED State, mean of posterior distribution Range value: 0.255882063, ..., 0.600975588
\item \strong{N4SD} \emph{numeric} Latent probability of encoding into the DURABLY LEARNED State, due to successful retrieval (R1) from the TRANSIENTLY LEARNED State, standard deviation of posterior distribution Range value: 0.152613408, ..., 0.205755652
\item \strong{R1M} \emph{numeric} Latent probability of retrieving from the TRANSIENTLY LEARNED State, mean of posterior distribution Range value: 0.152002992, ..., 0.852401624
\item \strong{R1SD} \emph{numeric} Latent probability of retrieving from the TRANSIENTLY LEARNED State, standard deviation of posterior distribution Range value: 0.110312216, ..., 0.200643535
\item \strong{R2M} \emph{numeric} Latent probability of retrieving from the DURABLY LEARNED State, mean of posterior distribution Range value: 0.353427651, ..., 0.966844561
\item \strong{R2SD} \emph{numeric} Latent probability of retrieving from the DURABLY LEARNED State, standard deviation of posterior distribution Range value: 0.030812429, ..., 0.230961137
\item \strong{R3M} \emph{numeric} Latent probability of retrieving from the DURABLY LEARNED State after a 5-minute delay with distraction, mean of posterior distribution Range value: 0.338101647, ..., 0.930136192
\item \strong{R3SD} \emph{numeric} Latent probability of retrieving from the DURABLY LEARNED State after a 5-minute delay with distraction, standard deviation of posterior distribution Range value: 0.060749158, ..., 0.206419116
\item \strong{MIFR1M} \emph{numeric} Latent probability of recall of an episodic memory on immediate recall task 1, mean of posterior distribution Range value: 0.263038273, ..., 0.922307206
\item \strong{MIFR1SD} \emph{numeric} Latent probability of recall of an episodic memory on immediate recall task 1, standard deviation of posterior distribution Range value: 0.048716683, ..., 0.142291254
\item \strong{MIFR2M} \emph{numeric} Latent probability of recall of an episodic memory on immediate recall task 2, mean of posterior distribution Range value: 0.341925484, ..., 0.957671907
\item \strong{MIFR2SD} \emph{numeric} Latent probability of recall of an episodic memory on immediate recall task 2, standard deviation of posterior distribution Range value: 0.031952712, ..., 0.145817863
\item \strong{MIFR3M} \emph{numeric} Latent probability of recall of an episodic memory on immediate recall task 3, mean of posterior distribution Range value: 0.346259673, ..., 0.96372281
\item \strong{MIFR3SD} \emph{numeric} Latent probability of recall of an episodic memory on immediate recall task 3, standard deviation of posterior distribution Range value: 0.030881024, ..., 0.166425722
\item \strong{MDFR1M} \emph{numeric} Latent probability of recall of a durably stored episodic memory on delayed recall task 1, mean of posterior distribution Range value: 0.332168013, ..., 0.920247727
\item \strong{MDFR1SD} \emph{numeric} Latent probability of recall of a durably stored episodic memory on delayed recall task 1, standard deviation of posterior distribution Range value: 0.061971477, ..., 0.152114523
\item \strong{MTranM} \emph{numeric} Latent probability of immediate recall of a non-durably stored episodic memory on any task, mean of posterior distribution Range value: 0.061089955, ..., 0.481603733
\item \strong{MTranSD} \emph{numeric} Latent probability of immediate recall of a non-durably stored episodic memory on any task, standard deviation of posterior distribution Range value: 0.0592292, ..., 0.195199311
\item \strong{MDuraM} \emph{numeric} Latent probability of immediate recall of a durably stored episodic memory on any task, mean of posterior distribution Range value: 0.451639125, ..., 0.984213771
\item \strong{MDuraSD} \emph{numeric} Latent probability of immediate recall of a durably stored episodic memory on any task, standard deviation of posterior distribution Range value: 0.018741947, ..., 0.167210736
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{assessments}
\keyword{undefined_phase}
