% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EXCLUSIO}
\alias{EXCLUSIO}
\title{Exclusion Criteria}
\format{
A data frame with 1087 observations and 23 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(EXCLUSIO)
}
\description{
Exclusion Criteria data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1435
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, f
\item \strong{VISCODE2} \emph{character}   , with options: sc, f
\item \strong{EXAMDATE} \emph{ } Date Criteria Confirmed Date: YYY-MM-DD
\item \strong{EXNEURO} \emph{character} 1. NL - Does the participant have a significant neurologic disease such as Parkinson's disease, multi-infarct dementia, Huntington's disease, normal pressure hydrocephalus, brain tumor, progressive supranuclear palsy, seizure disorder, subdural hematoma, multiple sclerosis, or history of significant head trauma followed by persistent neurologic defaults or known structural brain abnormalities.MCI - Does the participant have a significant neurologic disease other than suspected incipient Alzheimer's disease such as...AD - Does the participant have a significant neurologic disease other than Alzheimer's disease including... 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXNIMAG} \emph{character} 2. Does the participant's screening/baseline MRI scans have evidence of infection, infarction, or other focal lesions?  Participants with multiple lacunes or lacunes in a critical memory structure are excluded. 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXMRI} \emph{character} 3. Does the participant have a pacemaker, aneurysm clips, artificial heart valves, ear implants, metal fragments or foreign objects in the eyes, skin or body. 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXPSYCH} \emph{character} 4. NL - Has the participant had major depression or bipolar disorder as described in DSM-IV within the past year or a history of schizophrenia (DSM IV criteria)?MCI/AD - Does the participant have a history of major depression...or a history of psychotic features, agitation, or behavioral problems within the last 3 months which could lead to difficulty complying with the protocol? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXABUSE} \emph{character} 5. Does the participant have a history of alcohol or substance abuse or dependence within the past 2 years (DSM IV criteria)? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXILLNESS} \emph{character} 6. Does the participant have a significant systemic illness or unstable medical condition which could lead to difficulty complying with the protocol? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXLAB} \emph{character} 7. Does the participant have any clinically significant abnormalities in B12, RPR, or TFTs that might interfere with the study. 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXRESID} \emph{character} 8. Does the participant reside in a skilled nursing facility? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXMEDS} \emph{character} 9. Is the participant currently taking, or has he/she taken in the last 4 weeks, any excluded medication(s) as described in the Procedures Manual? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXINVEST} \emph{character} 10. Has the participant used another investigational agent within one month prior to screening? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXPARTICIP} \emph{character} 11. Is the participant participating in a clinical study involving neuropsychological measures being collected more than one time per year? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 6, ..., 2188
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{enrollment}
