% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_codelist_datadict}
\alias{extract_codelist_datadict}
\title{Extract Codedlist from DATADIC}
\usage{
extract_codelist_datadict(.datadic)
}
\arguments{
\item{.datadic}{Data dictionary dataset}
}
\value{
A data.frame the same as \code{.datadic} with only coded list.
}
\description{
This function is used to extract codelist from a data dictionary \code{\link{DATADIC}()}.
}
\examples{
\dontrun{
extract_codelist_datadict(.datadic = ADNIMERGE2::DATADIC)
}
}
\seealso{
Other data dictionary related functions: 
\code{\link{get_factor_fldname}()},
\code{\link{get_factor_levels_datadict}()}
}
\concept{data dictionary related functions}
\keyword{adni_datadic_fun}
