% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FAMHXPAR}
\alias{FAMHXPAR}
\title{Family History - Parents}
\format{
A data frame with 2381 observations and 25 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FAMHXPAR)
}
\description{
Family History - Parents data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI1, ADNI2, ADNI4, ADNIGO
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, init, 4_sc, 4_init, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{MOTHALIVE} \emph{character} Is the participant's mother living? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MOTHAGE} \emph{numeric} Current age, or if deceased, age at death? Range value: 17, ..., 105
\item \strong{MOTHDEM} \emph{character} Did/Does the biological mother have dementia? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MOTHAD} \emph{character} If yes, did/does biological mother have Alzheimer's Disease? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MOTHSXAGE} \emph{numeric} At what approximate age did mother's symptoms start? Range value: 16, ..., 102
\item \strong{FATHALIVE} \emph{character} Is the participant's father living? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FATHAGE} \emph{numeric} Current age, or if deceased, age at death? Range value: 24, ..., 110
\item \strong{FATHDEM} \emph{character} Did/Does the biological father have dementia? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FATHAD} \emph{character} If yes, did/does biological father have Alzheimer's Disease? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FATHSXAGE} \emph{numeric} At what approximate age did father's symptoms start? Range value: 45, ..., 99
\item \strong{ID} \emph{numeric} Record ID Range value: 15, ..., 185322
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{adni4}
\keyword{subject_characteristics}
