% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FAMHXSIB}
\alias{FAMHXSIB}
\title{Family History - Sibling Log}
\format{
A data frame with 6879 observations and 23 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FAMHXSIB)
}
\description{
Family History - Sibling Log data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI1, ADNI2, ADNI4, ADNIGO
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, init, 4_sc, 4_init, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{SIBYOB} \emph{numeric} Biological sibling year of birth Range value: 1912, ..., 1996
\item \strong{SIBRELAT} \emph{character} Specify relationship 1=Same biological mother and father ; 2=Same biological mother ; 3=Same biological father
\emph{Decoded Value: }\strong{Yes}
\item \strong{SIBGENDER} \emph{character} What sex was originally listed at birth? 
\itemize{
\item \emph{ADNI3:} 1=Male ; 2=Female
\item \emph{ADNI4:} 1=Male ; 2=Female ; 3=Decline to answer
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SIBALIVE} \emph{character} Is sibling living? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SIBAGE} \emph{numeric} Current age or if deceased age of death? Range value: 0, ..., 99
\item \strong{SIBDEMENT} \emph{character} Did/Does this sibling have dementia? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SIBAD} \emph{character} If yes, did/does sibling have Alzheimer's Disease? 
\itemize{
\item \emph{ADNI3:} 0=No ; 1=Yes
\item \emph{ADNI4:} 0=No ; 1=Yes ; 2=I don't know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SIBSXAGE} \emph{numeric} At what approximate age did symptoms start? Range value: 0, ..., 96
\item \strong{ID} \emph{numeric} Record ID Range value: 16, ..., 185324
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{adni4}
\keyword{subject_characteristics}
