% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FAQ}
\alias{FAQ}
\title{Functional Assessment Questionnaire}
\format{
A data frame with 13272 observations and 28 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FAQ)
}
\description{
Functional Assessment Questionnaire data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment FAQDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{SOURCE} \emph{character} Information Source 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Participant Visit ; 2=Telephone Call
\item \emph{ADNI3, ADNI4:} 1=Participant Visit ; 2=Telephone Call/Video Call
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQFINAN} \emph{character} 1. Writing checks, paying bills, or balancing checkbook. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQFORM} \emph{character} 2. Assembling tax records, business affairs, or other papers. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQSHOP} \emph{character} 3. Shopping alone for clothes, household necessities, or groceries. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQGAME} \emph{character} 4. Playing a game of skill such as bridge or chess, working on a hobby. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQBEVG} \emph{character} 5. Heating water, making a cup of coffee, turning off the stove. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQMEAL} \emph{character} 6. Preparing a balanced meal. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQEVENT} \emph{character} 7. Keeping track of current events. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQTV} \emph{character} 8. Paying attention to and understanding a TV program, book, or magazine. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQREM} \emph{character} 9. Remembering appointments, family occasions, holidays, medications. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQTRAVL} \emph{character} 10. Traveling out of the neighborhood, driving, or arranging to take public transportation. 0=Normal (0) ; 1=Never did, but could do now (0) ; 2=Never did, would have difficulty now (1) ; 3=Has difficulty, but does by self (1) ; 4=Requires assistance (2) ; 5=Dependent (3)
\emph{Decoded Value: }\strong{Yes}
\item \strong{FAQTOTAL} \emph{numeric} FAQ Total Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0..30
\item \emph{ADNI4} 
}
\item \strong{SPID} \emph{numeric} Study Partner ID: Range value: 1, ..., 10
\item \strong{ID} \emph{numeric} Record ID Range value: 4, ..., 187117
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
