% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FCI}
\alias{FCI}
\title{Financial Capacity Instrument Short Form (FCI-SF)}
\format{
A data frame with 2626 observations and 70 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FCI)
}
\description{
Financial Capacity Instrument Short Form (FCI-SF) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 7125
\item \strong{VISCODE} \emph{character}   , with options: init, bl, y1, y2, y4, y3, y5
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{MC_SCORE} \emph{numeric} Mental Calculation Score (Items 1-2) 0..4
\item \strong{FC_SCORE} \emph{numeric} Financial Conceptual Knowledge Score (Items 3-6) 0..8
\item \strong{SNGLCHKSCORE} \emph{numeric} Single/Checkbook Register Score (Items 7-16): 0..20
\item \strong{COMPLXCHKSCORE} \emph{numeric} Complex Checkbook/Register Score (Items 17-30): 0..28
\item \strong{CHKCOMPOSTIME} \emph{numeric} Checkbook/Register Composite Time (Items 7-16 & 17-30): 0..540
\item \strong{CHKCOMPOSSCORE} \emph{numeric} Combined Checkbook/Register Score (Items 7-16 + 17-30 ): 0..48
\item \strong{BANKSTATSCORE} \emph{numeric} Bank Statement Management Score (Items 31-37): 0..14
\item \strong{TOTCOMPTIME} \emph{numeric} TOTAL Composite Time (Items 4, 5, 7-16, 17-30): 0..720
\item \strong{FCISCORE} \emph{numeric} TOTAL FCI-SF SCORE: 0..74
\item \strong{DONE} \emph{character} Was assessment/procedure done? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 1=Participant Unable for Cognitive Reasons ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 4=Equipment related issue ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{NICKELS} \emph{character} 1.  Nickels 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{QUARTERS} \emph{character} 2.  Quarters 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{BUDGET} \emph{character} 3. Budget 0=0 ; 1=1 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXPMEDINS} \emph{character} Experience with Medical Insurance Plan? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{INSURANCE} \emph{character} 4. Health Care Insurance Problem 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEDINSTIME} \emph{numeric} Time to complete Item 4 Range value: 0, ..., 90
\item \strong{EXPINCTAX} \emph{character} Experience with Income Taxes? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{TAXCREDA} \emph{character} 5. a. Tax Credit 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{QUES5TIME} \emph{numeric} Time to complete Item 5.a. Range value: 0, ..., 40
\item \strong{TAXCREDB} \emph{character} 5. b. Tax Credit 0=0 ; 1=1 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXPCHECKREG} \emph{character} Experience with Checkbook/Check Register? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PAYEE} \emph{character} 6. Payee 0=0 ; 1=1 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK7} \emph{character} 7. Payee section 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK8} \emph{character} 8.  $ amount section (numerical) 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK9} \emph{character} 9.  $ amount section (written) 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK10} \emph{character} 10. Signature of payer 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK11} \emph{character} 11. Date properly written 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK12} \emph{character} 12. Number of check 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK13} \emph{character} 13. Date properly entered 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK14} \emph{character} 14. Payee section 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK15} \emph{character} 15. Amount of check 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHK16} \emph{character} 16. New account balance 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{SNGLCHKTIME} \emph{numeric} Time to complete items 7-16 Range value: 1, ..., 240
\item \strong{COMPLXCHK17} \emph{character} 17. Number of check 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK18} \emph{character} 18. Date properly entered 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK19} \emph{character} 19. Payee section 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK20} \emph{character} 20. Amount of check 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK21} \emph{character} 21. New Account Balance 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK22} \emph{character} 22. Number of check 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK23} \emph{character} 23. Date properly entered 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK24} \emph{character} 24. Payee section 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK25} \emph{character} 25. Amount of check 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK26} \emph{character} 26. New Account Balance 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK27} \emph{character} 27. Date properly entered 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK28} \emph{character} 28. Description of Transaction 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK29} \emph{character} 29. Amount of deposit 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHK30} \emph{character} 30. New account balance 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMPLXCHKTIME} \emph{numeric} Time to complete Items 17-30 Range value: 1, ..., 300
\item \strong{BANKSTATA} \emph{character} 31. a. Bank Statement 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{BANKSTATB} \emph{character} 31. b. Bank Statement 0=0 ; 1=1 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXPBANKSTATE} \emph{character} Experience with Bank Statement? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{BANKSTAT32} \emph{character} 32. Interest Rate 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{BANKSTAT33} \emph{character} 33. Time Period 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{BANKSTAT34} \emph{character} 34. Checks Cleared 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{BANKSTAT35} \emph{character} 35. Quarterly Interest 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{BANKSTAT36} \emph{character} 36. Gaps in Check Sequence 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{BANKSTAT37} \emph{character} 37. Date of the Withdrawal Payment 0=0 ; 2=2 ; 99=OT
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 12, ..., 187106
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 63
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{assessments}
