% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FHQ}
\alias{FHQ}
\title{Family History Questionnaire}
\format{
A data frame with 2952 observations and 19 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FHQ)
}
\description{
Family History Questionnaire data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, f, v01, v06, v07, v21, v03
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: sc, f, m60, m48, m12, m72, m24, m18, m84, bl
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{FHQSOURCE} \emph{character} Information Souce 1=Participant Visit ; 2=Telephone Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{FHQPROV} \emph{character} Indicate below who provided the information collected for this questionnaire: 1=Participant only ; 2=Study Partner only ; 3=Both Participant and Study Partner
\emph{Decoded Value: }\strong{Yes}
\item \strong{FHQMOM} \emph{character} Dementia 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No ; 2=Don't Know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FHQMOMAD} \emph{character} Alzheimer's Disease 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No ; 2=Don't Know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FHQDAD} \emph{character} Dementia 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No ; 2=Don't Know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FHQDADAD} \emph{character} Alzheimer's Disease 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No ; 2=Don't Know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FHQSIB} \emph{character} 3. Does the participant have any siblings? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 2886
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{subject_characteristics}
