% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FNIHBC_BLOOD_BIOMARKER_TRAJECTORIES}
\alias{FNIHBC_BLOOD_BIOMARKER_TRAJECTORIES}
\title{Clinical Utility of Blood Biomarker Trajectories Study}
\format{
A data frame with 24154 observations and 21 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FNIHBC_BLOOD_BIOMARKER_TRAJECTORIES)
}
\description{
Clinical Utility of Blood Biomarker Trajectories Study data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{STUDY_ID} \emph{character} Assigned by ADNI Biomarker Core - FNIH BC ADNI 00002 , with options: FNIH BC ADNI 00002
\item \strong{SAMPLEID} \emph{character} Tube ID assigned by the ADNI Biomarker Core. This column used for participant unblinding. The endogenous QC samples confirmed to have low or high concentrations of p-tau217 are labled as EQC Low or EQC High.  
\item \strong{PTID} \emph{character}    
\item \strong{RID} \emph{numeric} Patient ID Range value: 21, ..., 6309
\item \strong{EXAMDATE} \emph{ } Date when the sample was collected Date: YYY-MM-DD
\item \strong{VISCODE2} \emph{character} Months from baseline rounded to nearest 6 months  
\item \strong{ASSAYPLATFORM} \emph{character} Name of the assay platform utilized in the study , with options: Precivity, SIMOA, Lumipulse G1200, Elecsys
\item \strong{ASSAYVERSION} \emph{character} Version of assay used to perform the sample analysis  
\item \strong{PERFORMINGLAB} \emph{character} Name of the laboratory performing the sample analysis , with options: C2N Diagnostics, QuanterixAcceleratorLab, IndianaUniversity_BAL, University of Gothenburg
\item \strong{INSTRUMENTID} \emph{character} Identification number assigned to the analyzer or instrument utilized to generate the study data , with options: LC-MS/MS, HDX-298, HDX-240, HDX-123, FL_1, HDX-297, HDX-214, HDX-411, HDX-230, Cobas 801 42E5-06
\item \strong{MATRIX} \emph{character} Sample or matrix type, K2 EDTA plasma for all core study samples , with options: K2 EDTA plasma
\item \strong{TESTNAME} \emph{character} Name of the analyte test/analysis  
\item \strong{PLASMA_BIOMARKER} \emph{character} Description of assay developer, matrix, and analyte of biomarker measurement. This variable is intended to assist with data analyses.  
\item \strong{TESTVALUE} \emph{character} Value generated by sample analysis. This should be the absolute quantification value, and if not, prior approval of methods and data to be presented are needed. If there are replicates, this should be the average concentration of the replicate results.  
\item \strong{UNITS} \emph{character} Units of measure , with options: pg/mL, \%, NG/ML, PG/ML
\item \strong{CV} \emph{character} Coefficient of variance if there are replicate measures per sample  
\item \strong{RUN_NO} \emph{numeric} Analytical run, batch, or plate number Range value: 1, ..., 40
\item \strong{SEQUENCE_NO} \emph{numeric} Run order sequence number within an analytical run or over the course of the study if using the same instrument throughout Range value: 1, ..., 393
\item \strong{RUNDATE} \emph{ } Date the sample analysis was performed Date: YYY-MM-DD
\item \strong{COMMENTS} \emph{character} Data annotations where applicable  
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
