% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FOXLABBSI}
\alias{FOXLABBSI}
\title{Fox Lab BSI measures}
\format{
A data frame with 21692 observations and 29 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FOXLABBSI)
}
\description{
Fox Lab BSI measures data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character} Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor} Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10796
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{LONIUID} \emph{numeric} UID of fully pre-processed scan for which results are provided Range value: 30968, ..., 11309435
\item \strong{LONIUID_BASE} \emph{numeric} UID of fully pre-processed baseline scan for longitudinal results Range value: 30968, ..., 1586128
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{RUNDATE} \emph{ } Analysis run date Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Analysis completeness 0=Partial ; 1=Complete
\emph{Decoded Value: }\strong{Yes}
\item \strong{VERSION} \emph{ } Version date of results Date: YYY-MM-DD
\item \strong{BRAINVOL} \emph{numeric} Whole brain volume Range value: 680.31, ..., 1501.51
\item \strong{VENTVOL} \emph{numeric} Ventricular volume Range value: 5.511, ..., 189.373
\item \strong{HIPPOVOL_R} \emph{numeric} Right side hippocampal volume Range value: 1.37, ..., 4.337
\item \strong{HIPPOVOL_L} \emph{numeric} Left side hippocampal volume Range value: 1.119, ..., 4.008
\item \strong{DBCBBSI} \emph{numeric} Whole brain classic BSI Range value: -27.842, ..., 97.557
\item \strong{KMNDBCBBSI} \emph{numeric} Whole brain KN-BSI Range value: -26.663, ..., 109.811
\item \strong{VBSI} \emph{numeric} Ventricular BSI Range value: -6.092, ..., 44.03063
\item \strong{HBSI_R} \emph{numeric} Right side hippocampal BSI Range value: -0.243328, ..., 0.876119
\item \strong{HBSI_L} \emph{numeric} Left side hippocampal BSI Range value: -0.246401, ..., 0.743503
\item \strong{REGRATING} \emph{character} Longitudinal QC rating 1=Good ; 2=Acceptable ; 3=Borderline acceptable ; 4=Fail/unacceptable ; 5=Significant non-disease related pathology affects measure ; 7=Not rated
\emph{Decoded Value: }\strong{Yes}
\item \strong{VENTACCEPT} \emph{character} Ventricular QC rating 0=FAIL ; 1=PASS
\emph{Decoded Value: }\strong{Yes}
\item \strong{HPACCEPT_R} \emph{character} Right side hippocampal QC rating 0=FAIL ; 1=PASS
\emph{Decoded Value: }\strong{Yes}
\item \strong{HPACCEPT_L} \emph{character} Left side hippocampal QC rating 0=FAIL ; 1=PASS
\emph{Decoded Value: }\strong{Yes}
\item \strong{MRSEQUENCE} \emph{character} MRI sequence 
\itemize{
\item \emph{[ADNI1,GO,2,3,4]:} ADNI1=ADNI1 T1 ; Non-Acc=ADNIGO and ADNI2 non-accelerated T1 ; Acc=Accelerated T1 from ADNIGO, ADNI2, ADNI3, ADNI4
\item \emph{NA:} ADNI-1=ADNI-1 T1 ; Non-Acc=ADNI-GO and ADNI-2 non-accelerated T1 ; Acc=ADNI-GO and ADNI-2 Accelerated T1 or ADNI-3 Accelerated T1
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MRFIELD} \emph{numeric} MRI field strength Range value: 1.5, ..., 3
\item \strong{QC_PASS} \emph{character} Selected for cross-sectional analysis 0=FAIL/UNUSED ; 1=PASS
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{imaging}
