% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FUJIREBIOABETA}
\alias{FUJIREBIOABETA}
\title{Fujirebio Beta-Amyloid Ratio}
\format{
A data frame with 442 observations and 20 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FUJIREBIOABETA)
}
\description{
Fujirebio Beta-Amyloid Ratio data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI4
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 42, ..., 999999
\item \strong{VISCODE} \emph{character} Visit code , with options: v06, m48, m60, nv, m36, bl, v11, v03, v21
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: m60, m72, m48, nv, m36, bl, m24
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{GUSPECID} \emph{character} Specimen ID  
\item \strong{VID} \emph{numeric} Visit ID Range value: 2, ..., 10
\item \strong{DRAWDTE} \emph{ }   Date: YYY-MM-DD
\item \strong{DER} \emph{character}   , with options: CSF
\item \strong{SITE} \emph{character} Site  
\item \strong{VOL} \emph{numeric} Volume Range value: 0.15, ..., 0.45
\item \strong{RECDTE} \emph{ }   Date: YYY-MM-DD
\item \strong{STORDTE} \emph{ }   Date: YYY-MM-DD
\item \strong{RUN} \emph{numeric} Run Range value: 1, ..., 10
\item \strong{RUNDATE} \emph{ } Run date Date: YYY-MM-DD
\item \strong{ABETA42} \emph{numeric} Beta-Amyloid 1-42 Range value: 42, ..., 2267
\item \strong{ABETA40} \emph{numeric} Beta-Amyloid 1-40 Range value: 874, ..., 27373
\item \strong{ABETA42_40} \emph{numeric} Beta-Amyloid 1-42/1-40 ratio Range value: 0.022, ..., 0.827
\item \strong{COMMENTS} \emph{character} Comments , with options: Not enough volume for 1-40
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
