% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FUJIREBIOABETAPLASMA}
\alias{FUJIREBIOABETAPLASMA}
\title{Plasma Amyloid Beta (1-42/1-40) Ratio Biomarker Feasibility Study - Fujirebio Lumipulse}
\format{
A data frame with 514 observations and 11 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(FUJIREBIOABETAPLASMA)
}
\description{
Plasma Amyloid Beta (1-42/1-40) Ratio Biomarker Feasibility Study - Fujirebio Lumipulse data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI2, ADNIGO
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 31, ..., 5263
\item \strong{VISCODE} \emph{character} Record time of visit , with options: v41, v51, init, v21, v52
\item \strong{VISCODE2} \emph{character} Record time of visit by month  
\item \strong{EXAMDATE} \emph{ } Record time of visit by date Date: YYY-MM-DD
\item \strong{SPECIMEN_TYPE} \emph{character} Specimen sample type - plasma , with options: Plasma
\item \strong{ANALYTE} \emph{character} Biomarker analyte tested , with options: AB142P, AB140P
\item \strong{COUNT} \emph{numeric} Sample count Range value: 8339, ..., 204336
\item \strong{CONCENTRATION} \emph{numeric} Concentration of biomarker in pg/ml Range value: 14.57, ..., 555.7
\item \strong{MM_NUMBER} \emph{numeric} For samples tested in duplicate sample test 1 or 2 Range value: 1, ..., 2
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
