% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDSCALE}
\alias{GDSCALE}
\title{Geriatric Depression Scale}
\format{
A data frame with 13694 observations and 33 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(GDSCALE)
}
\description{
Geriatric Depression Scale data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment GDDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{SOURCE} \emph{character} Information Source 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Participant Visit ; 2=Telephone Call
\item \emph{ADNI3, ADNI4:} 1=Participant Visit ; 2=Telephone Call/Video Call
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDUNABL} \emph{character} Participant is unable to complete the GDS based on the clinician's best judgement. 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Participant is unable to complete the GDS based on the clinician's best judgement.
\item \emph{ADNI3, ADNI4:} 0=Box not checked ; 1=Box checked
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDSATIS} \emph{character} 1. Are you basically satisfied with your life? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(0) ; 0=No(1)
\item \emph{ADNI3, ADNI4:} 0=No (1) ; 1=Yes (0)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDDROP} \emph{character} 2. Have you dropped many of your activities and interests? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDEMPTY} \emph{character} 3. Do you feel that your life is empty? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDBORED} \emph{character} 4. Do you often get bored? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDSPIRIT} \emph{character} 5. Are you in good spirits most of the time? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(0) ; 0=No(1)
\item \emph{ADNI3, ADNI4:} 0=No (1) ; 1=Yes (0)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDAFRAID} \emph{character} 6. Are you afraid that something bad is going to happen to you? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDHAPPY} \emph{character} 7. Do you feel happy most of the time? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(0) ; 0=No(1)
\item \emph{ADNI3, ADNI4:} 0=No (1) ; 1=Yes (0)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDHELP} \emph{character} 8. Do you often feel helpless? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDHOME} \emph{character} 9. Do you prefer to stay at home, rather than going out and doing new things? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDMEMORY} \emph{character} 10. Do you feel you have more problems with memory than most? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDALIVE} \emph{character} 11. Do you think its wonderful to be alive now? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(0) ; 0=No(1)
\item \emph{ADNI3, ADNI4:} 0=No (1) ; 1=Yes (0)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDWORTH} \emph{character} 12. Do you feel pretty worthless the way you are now? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDENERGY} \emph{character} 13. Do you feel full of energy? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(0) ; 0=No(1)
\item \emph{ADNI3, ADNI4:} 0=No (1) ; 1=Yes (0)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDHOPE} \emph{character} 14. Do you feel that your situation is hopeless? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDBETTER} \emph{character} 15. Do you think that most people are better off than you are? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes(1) ; 0=No(0)
\item \emph{ADNI3, ADNI4:} 0=No (0) ; 1=Yes (1)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{GDTOTAL} \emph{numeric} Total Score 
\itemize{
\item \emph{ADNI1:} 0..15
\item \emph{ADNIGO, ADNI2, ADNI3, ADNI4} 
}
\item \strong{ID} \emph{numeric} Record ID Range value: 6, ..., 187103
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
