% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GENETIC}
\alias{GENETIC}
\title{ApoE Genotyping - Draw Data}
\format{
A data frame with 10109 observations and 58 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(GENETIC)
}
\description{
ApoE Genotyping - Draw Data data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment COLLDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 4
\item \strong{APCOLLECT} \emph{character} On this date, was Whole blood long read sequencing collected (1x3 mL Purple-Top EDTA tube)? 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DNAREASON} \emph{character} If no, indicate reason 
\itemize{
\item \emph{ADNI2, ADNI3, ADNI4:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=DNA sample previously collected ; 7=Other (specify)
\item \emph{ADNI4:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Other (specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{APTIME} \emph{ } Time of blood collection (24hr clock) hms: HH-MM-SS
\item \strong{APVOLUME} \emph{numeric} If Whole Blood for long read sequencing collected, volume of tube (1 x 3mL Purple-Top EDTA tube): 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1..10
\item \emph{ADNI3, ADNI4} 
}
\item \strong{WBLDVOL1} \emph{numeric} Volume of aliquots created 1.5mL WBLD #1: Range value: 0, ..., 1.5
\item \strong{WBLDVOL2} \emph{numeric} Volume of aliquots created 1.5mL WBLD #2: Range value: 0, ..., 1.5
\item \strong{WBLDFROZT} \emph{ } Time long read sequencing was placed in freezer(24hr clock): hms: HH-MM-SS
\item \strong{RNACOLL} \emph{character} On this date, was RNA sample collected (2 x 2.5 mL PAXgene RNA tubes)? 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RNAREASON} \emph{character} If no, indicate reason 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{RNADATE} \emph{ } Date of RNA collection: Date: YYY-MM-DD
\item \strong{RNATIME} \emph{ } Time of RNA collection (24hr clock): hms: HH-MM-SS
\item \strong{RNAVOL} \emph{numeric} Volume of blood drawn into 3 x 2.5 mL PAXgene RNA Tubes 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1..8
\item \emph{ADNI3} 
}
\item \strong{RNAVOL1} \emph{numeric} If RNA collected, volume of tubes (2.5mL PAXgene™ RNA tubes): RNA #1 Range value: 1, ..., 2.5
\item \strong{RNAVOL2} \emph{numeric} If RNA collected, volume of tubes (2.5mL PAXgene™ RNA tubes): RNA #2 Range value: 0, ..., 2.5
\item \strong{RNASHIP} \emph{character} Was this sample shipped ambient? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{RNAFROZT} \emph{ } If Frozen, Time RNA placed in freezer (24hr clock): hms: HH-MM-SS
\item \strong{BCOAT} \emph{character} On this date, was Buffy Coat extracted from 2 x 10mL EDTA biomarker collection tubes? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCREASON} \emph{character} If No, please provide reason why the Buffy Coat was not extracted 1=Biomarker not collected ; 2=Inadequate biomarker sample ; 3=Withdrawn consent ; 4=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCVOL} \emph{numeric} If Buffy Coat extracted, volume of aliquot 
\itemize{
\item \emph{ADNI2:} 0..99
\item \emph{ADNI3} 
}
\item \strong{BCVOL1} \emph{numeric} If plasma and buffy coat collection tubes drawn, volume of tubes (10mL Purple-Top EDTA tubes): EDTA #1 Range value: 1, ..., 10
\item \strong{BCVOL2} \emph{numeric} If plasma and buffy coat collection tubes drawn, volume of tubes (10mL Purple-Top EDTA tubes): EDTA #2 Range value: 0, ..., 10
\item \strong{BCEXTVOL1} \emph{numeric} If Buffy Coat extracted, volume of aliquots: Buffy Coat #1: Range value: 0.1, ..., 2
\item \strong{BCEXTVOL2} \emph{numeric} If Buffy Coat extracted, volume of aliquots: Buffy Coat #2: Range value: 0, ..., 2
\item \strong{BCSHIP} \emph{character} Was this sample shipped ambient? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{BCFROZT} \emph{ } If Frozen, time Buffy Coat placed in freezer (24hr clock): hms: HH-MM-SS
\item \strong{RBCCOLL} \emph{character} On this date, was RBC extracted from biomarker collection tube? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{RBCTUBE} \emph{numeric} If RBC collected, total number of aliquot tubes Range value: 1, ..., 2
\item \strong{RBCALIQ1} \emph{numeric} RBC aliquot 1 volume (2.0 mL tube) Range value: 0.1, ..., 2
\item \strong{RBCALIQ2} \emph{numeric} RBC aliquot 2 volume (4.0 mL tube) Range value: 0, ..., 5
\item \strong{RBCFROZT} \emph{ } If RBC collected, time placed in freezer hms: HH-MM-SS
\item \strong{PBMCCOLL} \emph{character} On this date, was PBMC sample collected (2 x 10mL Sodium Heparin tubes)? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PBMCREASON} \emph{character} If no, indicate reason 
\itemize{
\item \emph{ADNI4:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Other (specify)
\item \emph{ADNI4:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Already collected in ADNI3 ; 7=Other (specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PBMCVOL} \emph{numeric} If PBMC collected, volume Range value: 2, ..., 10
\item \strong{PBMCTIME} \emph{ } Time of PBMC collection (24hr clock): hms: HH-MM-SS
\item \strong{PBMCVOL1} \emph{numeric} If PBMC collected, volume of tubes: PBMC #1: Range value: 1, ..., 10
\item \strong{PBMCVOL2} \emph{numeric} If PBMC collected, volume of tubes: PBMC #2: Range value: 0, ..., 10
\item \strong{CLDATE} \emph{ } Date of cell immortalization collection Date: YYY-MM-DD
\item \strong{CLCOLL} \emph{character} On this date, was cell immortalization sample collected? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLREASON} \emph{character} If no, indicate reason 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLTIME} \emph{ } Time of cell immortalization collection (24hr clock) hms: HH-MM-SS
\item \strong{CLVOLUME} \emph{numeric} Total volume of blood drawn for cell immortalization into 2 x 8.5 mL ACD-A (yellow top) tubes Range value: -1, ..., 20
\item \strong{GENFEDDATE} \emph{character} Date FedExed  
\item \strong{ID} \emph{numeric} Record ID Range value: 5, ..., 186976
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: init, bl, annual
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{biospecimen}
