% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment-utils.R
\name{get_adni_blscreen_dxsum}
\alias{get_adni_blscreen_dxsum}
\title{Get ADNI Baseline/Screening Diagnostics Summary}
\usage{
get_adni_blscreen_dxsum(.dxsum, visit_type = "baseline", phase = "Overall")
}
\arguments{
\item{.dxsum}{Data.frame of \code{\link{DXSUM}()} eCRF}

\item{visit_type}{Either \code{baseline} or \code{screen} diagnostic status, Default: 'baseline'}

\item{phase}{Either \code{Overall} or ADNI study phase specific, Default: 'Overall'}
}
\value{
A data.frame with the following variables:
\itemize{
 \item {\code{RID}}: Subject ID
 \item {\code{ORIGPROT}}: Original study protocols
 \item {\code{COLPROT}}: Study data collection protocols
 \item {\code{EXAMDATE}}: Baseline/screening diagnostics assessment collection date
 \item {\code{DIAGNOSIS}}: Diagnostics status.
}
}
\description{
Get ADNI Baseline/Screening Diagnostics Summary
}
\details{
This function is used to extract either the baseline or screening diagnostics status
 subjects when they participate for the first time in ADNI study.

\strong{Screening Diagnostics Status}: The diagnostics status collected at screening visit.

\strong{Baseline Diagnostics Status}:

\itemize{
 \item The diagnostics status collected at baseline visit.

 \item The diagnostics status collected at screening status for subjects that
 did not have a baseline collected diagnostics status.
 The corresponding screening visit/exam date will be used as a baseline record.

 \strong{NOTE}: The result data will contains the screening diagnostics status
  of those who did not enrolled in the study (i.e. with screening failure) as their
  baseline diagnostics status. Required to adjust baseline diagnostics with
  corresponding subject enrollment date/enrollment flag.
  See examples for more information.
}
}
\examples{
\dontrun{
library(tidyverse)
library(ADNIMERGE2)

# Enrollment flag
enrollment_flag <- ADNIMERGE2::REGISTRY \%>\%
  get_adni_enrollment(.registry = .) \%>\%
  mutate(ENRLDT = as.Date(EXAMDATE)) \%>\%
  select(RID, ENRLDT)

# Baseline diagnostics status for newly enrolled subject in ADNI study
overall_baseline_dx <- get_adni_blscreen_dxsum(
  .dxsum = ADNIMERGE2::DXSUM,
  phase = "Overall",
  visit_type = "baseline"
) \%>\%
  left_join(
    enrollment_flag,
    by = "RID"
  ) \%>\%
  # Only enrolled subjects
  filter(ENRLFG \%in\% "Yes")

# Phase-specific baseline diagnostic status: -----
# When subject enrolled in ADNI3 study phase
adni3_baseline_dx <- get_adni_blscreen_dxsum(
  .dxsum = ADNIMERGE2::DXSUM,
  phase = "ADNI3",
  visit_type = "baseline"
) \%>\%
  left_join(
    enrollment_flag,
    by = "RID"
  ) \%>\%
  # Only enrolled subjects
  filter(ENRLFG \%in\% "Yes")

# Screening Diagnostics Status: -----
# When subject were screened for first time in ADNI study.
first_screen_dx <- get_adni_blscreen_dxsum(
  .dxsum = ADNIMERGE2::DXSUM,
  phase = "Overall",
  visit_type = "screen"
)
# Phase-specific screening diagnostic status: ------
# When subject were screened for ADNI4 study phase
adni3_screen_dx <- get_adni_blscreen_dxsum(
  .dxsum = ADNIMERGE2::DXSUM,
  phase = "ADNI4",
  visit_type = "screen"
)
}
}
\seealso{
Other ADNI specific functions: 
\code{\link{get_adni_enrollment}()},
\code{\link{get_adni_screen_date}()}
}
\concept{ADNI specific functions}
\keyword{adni_enroll_fun}
