% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment-utils.R
\name{get_adni_enrollment}
\alias{get_adni_enrollment}
\title{ADNI Enrollment Date/First Baseline Visit Date}
\usage{
get_adni_enrollment(.registry)
}
\arguments{
\item{.registry}{Data.frame of \code{\link{REGISTRY}()} eCRF}
}
\value{
A data.frame of overall enrollment in ADNI study with the following variables:
\itemize{
 \item {\code{RID}}: Subject ID
 \item {\code{ORIGPROT}}: Original study protocols
 \item {\code{COLPROT}}: Study data collection protocols
 \item {\code{EXAMDATE}}: Enrollment date (i.e. first baseline visit date)
}
}
\description{
This function is used to extract enrollment date (baseline visit date)
 when subjects are enrolled in ADNI study for the first time.
}
\examples{
\dontrun{
overall_enroll_registry <- get_adni_enrollment(
  .registry = ADNIMERGE2::REGISTRY
)
}
}
\seealso{
\code{\link{get_adni_screen_date}()}

Other ADNI specific functions: 
\code{\link{get_adni_blscreen_dxsum}()},
\code{\link{get_adni_screen_date}()}
}
\concept{ADNI specific functions}
\keyword{adni_enroll_fun}
