% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment-utils.R
\name{get_adni_screen_date}
\alias{get_adni_screen_date}
\title{Get ADNI Screening Date}
\usage{
get_adni_screen_date(
  .registry,
  phase = "Overall",
  multiple_screen_visit = FALSE
)
}
\arguments{
\item{.registry}{Data.frame of \code{\link{REGISTRY}()} eCRF}

\item{phase}{Either \code{Overall} or phase-specific screening date, Default: 'Overall'}

\item{multiple_screen_visit}{A Boolean value to include multiple screen visits in ADNIGO and ADNI2 phases, Default: FALSE}
}
\value{
A data.frame contains the following variables:
 \itemize{
  \item {\code{RID}}: Subject ID
  \item {\code{ORIGPROT}}: Original study protocols
  \item {\code{COLPROT}}: Study data collection protocols
  \item {\code{SCREENDATE}}: Screening visit date
  \item {\code{VISCODE}}: Screening visit code only when \code{multiple_screen_visit} is \code{TRUE}
}
}
\description{
Get ADNI Screening Date
}
\details{
This function is used to extract subject's screening date in ADNI study when
they participate in the study for the first time.

\strong{Overall Screening}: Across the whole ADNI study phases
\strong{Phase Specific Screening}: Based on a particular set of ADNI study phases
\strong{Multiple Screening Visit}: Accounts for multiple screening stage,
 particularly in \code{ADNIGO} and \code{ADNI2} study phases.
}
\examples{
\dontrun{
library(tidyverse)
library(ADNIMERGE2)

# Overall screening: ----
# When subjects were screened for the first time in ADNI study.
overall_screen_registry <- get_adni_screen_date(
  .registry = ADNIMERGE2::REGISTRY,
  phase = "Overall",
  multiple_screen_visit = FALSE
)

# Phase-specific screening:
# When subject screened for the first time in ADNI3 study phase.
adni3_screen_registry <- get_adni_screen_date(
  .registry = ADNIMERGE2::REGISTRY,
  phase = "ADNI3",
  multiple_screen_visit = FALSE
)

# Multiple screen visits in ADNIGO and ADNI2 study phases.
adnigo2_screen_registry <- get_adni_screen_date(
  .registry = ADNIMERGE2::REGISTRY,
  phase = c("ADNIGO", "ADNI2"),
  multiple_screen_visit = TRUE
)

# Screening across each ADNI study phases
phase_screen_registry <- get_adni_screen_date(
  .registry = ADNIMERGE2::REGISTRY,
  phase = ADNIMERGE2::adni_phase(),
  multiple_screen_visit = FALSE
)
}
}
\seealso{
\code{\link{get_adni_enrollment}()}

Other ADNI specific functions: 
\code{\link{get_adni_blscreen_dxsum}()},
\code{\link{get_adni_enrollment}()}
}
\concept{ADNI specific functions}
\keyword{adni_enroll_fun}
