% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment-utils.R
\name{get_baseline_vistcode}
\alias{get_baseline_vistcode}
\title{Get Baseline Visit Code}
\usage{
get_baseline_vistcode()
}
\value{
A character vector
}
\description{
Get Baseline Visit Code
}
\examples{
\dontrun{
get_baseline_vistcode()
}
}
\seealso{
\code{\link{get_screen_vistcode}()}
 \code{\link{VISITS}()}

Other ADNI visit codes: 
\code{\link{get_screen_vistcode}()}
}
\concept{ADNI visit codes}
\keyword{adni_utils}
