% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cols_name}
\alias{get_cols_name}
\title{Extract Column Name}
\usage{
get_cols_name(.data, col_name)
}
\arguments{
\item{.data}{Data.frame}

\item{col_name}{Column names}
}
\value{
A character vector of column names that are existed in the dataset.
 Otherwise return \code{NA}.
}
\description{
This function is used to get column names if the provided column name
  is existed in the dataset.
}
\examples{
\dontrun{
get_cols_name(.data = ADNIMERGE2::CDR, col_name = c("Phase", "VISCODE"))
get_cols_name(.data = ADNIMERGE2::CDR, col_name = c("RID", "VISCODE"))
}
}
\keyword{utils_fun}
