% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cols_value}
\alias{get_cols_value}
\title{Get Columns With Specific Values}
\usage{
get_cols_value(.data, value, col_name = NULL)
}
\arguments{
\item{.data}{Data.frame}

\item{value}{Specific value}

\item{col_name}{Character vector of columns, Default: \code{NULL}. By default, all available columns will be checked.}
}
\value{
A character vector of column names that contains the provided specific value.
 Otherwise return \code{NA}.
}
\description{
This function is used to list columns that contains a specific value.
}
\examples{
\dontrun{
get_cols_value(data = ADNIMERGE2::CDR, value = "Telephone Call")
}
}
\keyword{utils_fun}
