% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment-utils.R
\name{get_death_flag}
\alias{get_death_flag}
\title{Death Flag}
\usage{
get_death_flag(.studysum, .adverse, .recadv)
}
\arguments{
\item{.studysum}{Dispositions record for ADNI3-4, see \code{\link{STUDYSUM}()}}

\item{.adverse}{Adverse events record for ADNI3-4 study phase, see \code{\link{ADVERSE}()}}

\item{.recadv}{Adverse events record for ADNI1-GO-2, see \code{\link{RECADV}()}}
}
\value{
A data.frame with the following columns:
\itemize{
 \item {\code{RID}}: Subject ID
 \item {\code{ORIGPROT}}: Original study protocols
 \item {\code{COLPROT}}: Study data collection protocols which the event was recorded
 \item {\code{DTHDTC}}: Death date
 \item {\code{DTHFL}}: Death flag, \code{Yes}
}
}
\description{
This function is used to extract death records in the study based on the
adverse events and disposition records.
}
\examples{
\dontrun{
library(tidyverse)
library(ADNIMERGE2)
get_death_flag(
  .studysum = ADNIMERGE2::STUDYSUM,
  .adverse = ADNIMERGE2::ADVERSE,
  .recadv = ADNIMERGE2::RECADV
)
}
}
\seealso{
Other ADNI flag functions: 
\code{\link{get_disposition_flag}()}
}
\concept{ADNI flag functions}
\keyword{adni_enroll_fun}
