% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment-utils.R
\name{get_disposition_flag}
\alias{get_disposition_flag}
\title{Study Discontinuation Flag}
\usage{
get_disposition_flag(.registry, .studysum)
}
\arguments{
\item{.registry}{Registry record for ADNI1-GO-2, see \code{\link{REGISTRY}()}}

\item{.studysum}{Disposition record for ADNI3-4, see \code{\link{STUDYSUM}()}}
}
\value{
A data.frame with the following columns:
\itemize{
 \item {\code{RID}}: Subject ID
 \item {\code{ORIGPROT}}: Original study protocols
 \item {\code{COLPROT}}: Study data collection protocols which the event was recorded
 \item {\code{SDSTATUS}}: Disposition status
 \item {\code{SDDATE}}: Disposition date
}
}
\description{
This function is used to get early discontinuation list in ADNI study.
  Based on the \code{\link{REGISTRY}()} eCRF for ADNI1-GO-2 and
  \code{\link{STUDYSUM}} eCRF for ADNI3-4.
}
\examples{
\dontrun{
get_disposition_flag(
  .registry = ADNIMERGE2::REGISTRY,
  .studysum = ADNIMERGE2::STUDYSUM
)
}
}
\seealso{
Other ADNI flag functions: 
\code{\link{get_death_flag}()}
}
\concept{ADNI flag functions}
\keyword{adni_enroll_fun}
