% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_factor_fldname}
\alias{get_factor_fldname}
\title{Gets Factor Field Name (FLDNAME)}
\usage{
get_factor_fldname(.datadic, tbl_name, dd_fldnames = NULL)
}
\arguments{
\item{.datadic}{Data dictionary dataset that generated using \code{\link{get_factor_levels_datadict}()}}

\item{tbl_name}{Table name}

\item{dd_fldnames}{Variable names in the actual dataset, Default = NULL.}
}
\value{
A character vector of variable names.
}
\description{
This function is used to identify factor field name (FLDNAME) based on a
  data dictionary dataset \code{\link{DATADIC}()}
  from <https://adni.loni.usc.edu/data-samples/adni-data/>.
}
\examples{
\dontrun{
data_dict_dd <- get_factor_levels_datadict(
  .datadic = ADNIMERGE2::DATADIC,
  tbl_name = NULL,
  nested_value = TRUE
)
# List of available factor columns in data dictionary
# \code{\link{DATADIC}()} for the \code{\link{CDR}()} dataset
get_factor_fldname(
  .datadic = data_dict_dd,
  tbl_name = "CDR",
  dd_fldnames = NULL
)
# List of factor columns that available in \code{\link{CDR}()} and data dictionary DATADIC
get_factor_fldname(
  .datadic = data_dict_dd,
  tbl_name = "CDR",
  dd_fldnames = colnames(ADNIMERGE2::CDR)
)
}
}
\seealso{
\code{\link{get_factor_levels_datadict}()} \code{\link{DATADIC}()}

Other data dictionary related functions: 
\code{\link{extract_codelist_datadict}()},
\code{\link{get_factor_levels_datadict}()}
}
\concept{data dictionary related functions}
\keyword{adni_datadic_fun}
