% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_factor_levels_datadict}
\alias{get_factor_levels_datadict}
\title{Gets Factor Levels from DATADIC Dataset}
\usage{
get_factor_levels_datadict(
  .datadic,
  tbl_name = NULL,
  nested_value = FALSE,
  spliter1 = ";| ;| ; ",
  spliter2 = "=| =| = "
)
}
\arguments{
\item{.datadic}{Data dictionary dataset}

\item{tbl_name}{Table name, Default: NULL that generate for all available TBLNAMEs in the \code{data_dict}}

\item{nested_value}{Unnest the factor level variables in long format if \code{TRUE}
otherwise nested with \code{CODES} variable.}

\item{spliter1}{First split parameter/pattern, Default:";| ;| ; "}

\item{spliter2}{Second split parameter/pattern, Default: "=| =| = "}
}
\value{
A data.frame that appended with the following variables:
\itemize{
  \item \emph{prefix}: Actual value in the dataset
  \item \emph{suffix}: Coded values from \code{\link{DATADIC}()}
  \item \emph{class_type}: Class type, "factor"
  }
}
\description{
This function is used to generate the coded levels of FLDNAME using
 a data dictionary dataset \code{\link{DATADIC}()} downloaded from
 <https://adni.loni.usc.edu/data-samples/adni-data/>.
}
\examples{
\dontrun{
get_factor_levels_datadict(
  .datadic = ADNIMERGE2::DATADIC,
  tbl_name = NULL,
  nested_value = TRUE
)
}
}
\seealso{
\code{\link{split_strings}()}

Other data dictionary related functions: 
\code{\link{extract_codelist_datadict}()},
\code{\link{get_factor_fldname}()}
}
\concept{data dictionary related functions}
\keyword{adni_datadic_fun}
