% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment-utils.R
\name{get_screen_vistcode}
\alias{get_screen_vistcode}
\title{Get Screening Visit Code}
\usage{
get_screen_vistcode(type = "all")
}
\arguments{
\item{type}{Type either all screening visit codes or the first screening visit per
study phase, Default: 'all'}
}
\value{
A character vector
}
\description{
Get Screening Visit Code
}
\examples{
\dontrun{
get_screen_vistcode(type = "all")
get_screen_vistcode(type = "first")
}
}
\seealso{
\code{\link{get_baseline_vistcode}()}
 \code{\link{VISITS}()}

Other ADNI visit codes: 
\code{\link{get_baseline_vistcode}()}
}
\concept{ADNI visit codes}
\keyword{adni_utils}
