% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{get_vars_common_date}
\alias{get_vars_common_date}
\title{Get Common Date Across Variables}
\usage{
get_vars_common_date(
  .data,
  date_cols,
  select_method = "min",
  compared_ref_date = FALSE,
  ref_date_col = NULL,
  preferred_date_col = NULL
)
}
\arguments{
\item{.data}{A wide format data.frame}

\item{date_cols}{Character vector of date column names}

\item{select_method}{Selection method if there is more than one unique non-missing date, Default: 'min'
Either the minimum date (\code{'min'}), or the maximum date (\code{'max'}).}

\item{compared_ref_date}{A Boolean to compared the common date with the reference date if it is provided, Default: FALSE}

\item{ref_date_col}{Reference date column name, Default: NULL}

\item{preferred_date_col}{Preferred date when common date and reference date are different.
Only applicable if \code{compared_ref_date} is \code{TRUE}.}
}
\value{
A data.frame with the appended columns:
\itemize{
 \item COMMON_DATE Common date among the provided date columns
 \item FINAL_DATE Final date after comparing with reference date if \code{compared_ref_date} is \code{TRUE}. Otherwise the same as \code{COMMON_DATE}.
 \item DATE_RECORD_TYPE Record type to indicate whether the date columns are the same rowwise or not.
 }
}
\description{
Get Common Date Across Variables
}
\details{
This function is used to get a common date value across multiple date columns and
 compared with reference date column if it is provided. The comparison algorithm
 is based on rowwise operation and presented as follow:

 For records that have at least one non-missing date columns based on the list of
 provided date columns:
 \itemize{
   \item Select one unique date value if all date columns are the same/equal.
   \item Select either the minimum or maximum date based on the selection method (\code{select_method}) if at least one date column is differ from the remaining provided date columns.
 }

The reference date column must be present for any comparison with reference date. Then, the reference date will be used for any missing common date.

Otherwise, the date will considered as missing \code{NA}.
}
\examples{
\dontrun{
# Please see the \code{ADNIMERGE2-PACC-SCORE} vignette
vignette(topic = "ADNIMERGE2-PACC-SCORE", package = "ADNIMERGE2")
}

}
\keyword{utils_fun}
