% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GF}
\alias{GF}
\title{[DERIVED]  Derived  Genomics Findings}
\format{
A data frame with 3008 observations and 14 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(GF)
}
\description{
Genomics findings derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier ; , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbrevation , with character options: GF
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{GFSEQ} \emph{integer} Sequence Number Range value: 1, ..., 1
\item \strong{GFGRPID} \emph{factor} Group ID Study Protocol of Data Collection; , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{GFTESTCD} \emph{character} Short Name of Genomic Measurement , with character options: APOE
\item \strong{GFTEST} \emph{character} Name of Genomic Measurement , with character options: Apolipoprotein E
\item \strong{GFSTDTL} \emph{character} Measurement, Test, or Examination Detail , with character options: GENOTYPE
\item \strong{GFORRES} \emph{character} Result or Finding in Original Units , with character options: 4/4, 3/3, 3/4, 2/3, 2/2, 2/4
\item \strong{GFSTRESC} \emph{character} Character Result/Finding in Std Format , with character options: ε4/ε4, ε3/ε3, ε3/ε4, ε2/ε3, ε2/ε2, ε2/ε4
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -1, ..., -1
\item \strong{VISIT} \emph{character} Visit Name , with character options: ADNI1 Screening, ADNI1 Screen Fail
\item \strong{GFDTC} \emph{Date} Date/Time of Specimen Collection Date: YYY-MM-DD
\item \strong{GFDY} \emph{integer} Study Day of Specimen Collection Range value: -351, ..., 5
}
}
\keyword{derived_dataset}
