% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IES}
\alias{IES}
\title{Impact of Events Scale}
\format{
A data frame with 614 observations and 30 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(IES)
}
\description{
Impact of Events Scale data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNIGO, ADNI2, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10878
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{IESDATE} \emph{ } Date Collected: Date: YYY-MM-DD
\item \strong{IES1} \emph{character} 1. I thought about it when I didn't mean to. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES2} \emph{character} 2. I avoided letting myself get upset when I thought about it or was reminded of it. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES3} \emph{character} 3. I tried to remove it from my memory. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES4} \emph{character} 4. I had trouble falling asleep, because of pictures or thoughts about it that came into my mind. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES5} \emph{character} 5. I had waves of strong feelings about it. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES6} \emph{character} 6. I had dreams about it. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES7} \emph{character} 7. I stayed away from reminders. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES8} \emph{character} 8. I felt as if it hadn't happened or wasn't real. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES9} \emph{character} 9. I tried not to talk about it. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES10} \emph{character} 10. Pictures about it popped into my mind. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES11} \emph{character} 11. Other things kept making me think about it. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES12} \emph{character} 12. I was aware that I still had a lot of feelings about it, but I didn't deal with them. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES13} \emph{character} 13. I tried not to think about it. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES14} \emph{character} 14. Any reminder brought back feelings about it. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{IES15} \emph{character} 15. My feelings were kind of numb. 1=Not at All ; 2=Rarely ; 3=Sometimes ; 4=Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 4456, ..., 128219
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
