% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INCLUSIO}
\alias{INCLUSIO}
\title{Inclusion Criteria}
\format{
A data frame with 1109 observations and 39 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(INCLUSIO)
}
\description{
Inclusion Criteria data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 1435
\item \strong{VISCODE} \emph{character} Visit code , with options: f, sc
\item \strong{VISCODE2} \emph{character}   , with options: f, sc
\item \strong{EXAMDATE} \emph{ } Date Criteria Confirmed Date: YYY-MM-DD
\item \strong{INCSIGN} \emph{character} 1. Have the participant and study partner signed the Informed Consent form? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCSGDT} \emph{ } If Yes, date signed Date: YYY-MM-DD
\item \strong{INCSG15T} \emph{character} Check the following to indicate the participant is suitable for and consents to: 1=1.5 Tesla MRI
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCSGPET} \emph{character}   1=PET Scan
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCSG3T} \emph{character}   1=3 Tesla MRI
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCSGLP} \emph{character}   1=Lumbar Puncture
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCMCOMP} \emph{character} 2. NL - Is participant free of memory complaints, verified by an informant, aside from those normal with age?MCI - Does the subject have memory complaints and memory difficulties that are verified by an informant?AD -  Does the subject have memory complaints that are verified by an informant? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCMFXN} \emph{character} 3. NL - Normal memory function documented by scoring at specific cutoffs on the Logical Memory II subscale (delayed Paragraph Recall) from the Wechsler Memory Scaled - Revised (the maximum score is 25)MCI/AD - Abnormal memory function documented by scoring below the educationn adjusted cutoff on the Logical Memory II subscale (Delayed  Paragraph Recall) from the Wechsler Memory Scale - Revised (the maximum score is 25) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCMMSE} \emph{character} 4. NL/MCI - Does the participant have Mini-Mental State Exam score between 24 and 30 (inclusive)?  (Exceptions must be made for subjects with less than 8 years of education at the discretion of the project director).AD - Does the participant have an MMSE score between 20 and 26 (inclusive)? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCCDR} \emph{character} 5. NL - Does the participant have a Clinical Dementia Rating of 0?  Memory Box score must be 0.MCI - Does the participant have a Clinical Dementia Rating of 0.5?  Memory Box score must be at least 0.5.AD - Does participant have a Clinical Dementia rating of 0.5 or 1.0? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCCOG} \emph{character} 6. NL - Is the participant cognitively normal based on an absence of significant impairment in cognitive functions or activities of daily living?MCI - Is the participant's general cognition and functional performance sufficiently perserved such that a diagnosis of Alzheimer's disease cannot be made by the site physician at the time of the screening visit? AD - Does the participant meet NINCDS/ADRDA criteria for probable AD? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCHACH} \emph{character} 7. Does the participant have a Modified Hachinski score less than or equal to 4? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCAGE} \emph{character} 8. Is the participant between 55 and 90 years of age inclusive? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCSTMED} \emph{character} 9. Has the participant been on stable doses of non-excluded medications for at least 4 weeks prior to screening? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCGDS} \emph{character} 10. Does the participant have a Geriatric Depression Scale score of <6? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCINFORM} \emph{character} 11. Does the participant have an informant available who they have frequent contact with (e.g. an average of 10 hours per week or more), and can accompany the participant to all clinic visits and imaging sessions for the duration of the protocol? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCVISUAL} \emph{character} 12. Does the participant have adequate visual and auditory acuity to allow neuropsychological testing? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCHEALTH} \emph{character} 13. Is the participant in good general health with no additional diseases expected to interfere with the study? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCFEM} \emph{character} 14. If female, is the participant not pregnant, lactating, or of childbearing potential (i.e. women must be two years post-menopausal or surgically sterile)? 1=Yes ; 0=No ; 2=N/A
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCPART} \emph{character} 15. NL/MCI - Is the participant willing and able to complete all Baseline assessment and participate in a 3-year protocol?AD - Is the participant willing and able to ...2-year protocol? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCIMAG} \emph{character} 16. Is the participant willing to undergo MRI 1.5 Tesla neuroimaging (PET and MRI 3 Tesla are optional) and provide DNA for ApoE assessments and banking as well as plasma samples at protocol specified time points? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCEDUC} \emph{character} 17. Has the participant completed 6 grades of education (or had a good work history sufficient to exclude mental retardation)? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCLANG} \emph{character} 18. Is the participant fluent in English or Spanish? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCHIST} \emph{character} 19a. Medical History 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCPHYS} \emph{character} 19b. Physical Examination? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCNEUR} \emph{character} 19c. Neurological Examination? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCLAB} \emph{character} 19d. Laboratory Tests? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 10, ..., 2236
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{enrollment}
