% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INITHEALTH}
\alias{INITHEALTH}
\title{Initial Health Assessment}
\format{
A data frame with 31733 observations and 27 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(INITHEALTH)
}
\description{
Initial Health Assessment data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI3, ADNI2, ADNIGO, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: init, sc, 4_sc, 4_init, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{IHNUM} \emph{numeric} Condition Number Range value: 1, ..., 122
\item \strong{IHSYMPTOM} \emph{character} System/Category 
\itemize{
\item \emph{ADNI3:} 1=Psychiatric ; 2=Neurologic (other than Cognitive Disorder) ; 3=Head, Eyes, Ears, Nose, Throat ; 4=Cardiovascular ; 5=Respiratory ; 6=Hepatic ; 7=Dermatologic-Connective Tissue ; 8=Musculoskeletal ; 9=Endocrine-Metabolic ; 10=Gastrointestinal ; 11=Hematopoietic-Lymphatic ; 12=Renal-Genitourinary ; 13=Allergies or Drug Sensitivities ; 14=Smoking, Alcohol Use, and/or Drug Use ; 17=Malignancy ; 18=Other ; 19=Cognitive Disorder/Symptom
\item \emph{ADNI4:} 1=Psychiatric ; 10=Gastrointestinal ; 11=Hematopoietic-Lymphatic ; 12=Renal-Genitourinary ; 13=Allergies or Drug Sensitivities ; 14=Smoking, Alcohol Use, and/or Drug Use ; 17=Malignancy ; 18=Other ; 19=Cognitive Disorder/Symptom ; 2=Neurologic (other than Cognitive Disorder) ; 3=Head, Eyes, Ears, Nose, Throat ; 4=Cardiovascular ; 5=Respiratory ; 6=Hepatic ; 7=Dermatologic-Connective Tissue ; 8=Musculoskeletal ; 9=Endocrine-Metabolic
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IHDESC} \emph{character} Description/Condition  
\item \strong{IHSURG} \emph{character} Is this a surgery? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{IHSURGDATE} \emph{ } If Yes, surgery date Date: YYY-MM-DD
\item \strong{IHPRESENT} \emph{character} Is this a condition or symptom present in the three months prior to screening? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{IHCHRON} \emph{character} Chronicity 1=Single occurrence ; 2=Intermittent ; 3=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{IHSEVER} \emph{character} Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{IHDTONSET} \emph{ } Onset Date Date: YYY-MM-DD
\item \strong{IHONGOING} \emph{character} Is symptom/condition ongoing? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{IHCEASE} \emph{ } Cease Date Date: YYY-MM-DD
\item \strong{IHCOMM} \emph{character} Comments:  
\item \strong{ID} \emph{numeric} Record ID Range value: 14, ..., 184577
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{adni4}
\keyword{medical_history}
