% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-assert.R
\name{is_datadict_tbl}
\alias{is_datadict_tbl}
\title{Check \code{is_datadict_tbl} Object Class Type}
\usage{
is_datadict_tbl(x)
}
\arguments{
\item{x}{Object}
}
\value{
An error message if the class object is not \code{is_datadict_tbl}.
}
\description{
Check \code{is_datadict_tbl} Object Class Type
}
\seealso{
Other checks function: 
\code{\link{check_colnames}()},
\code{\link{check_duplicate_records}()},
\code{\link{check_non_missing_value}()},
\code{\link{check_object_type}()},
\code{\link{check_value_match}()},
\code{\link{detect_decimal_value}()},
\code{\link{detect_numeric_value}()}
}
\concept{checks function}
\keyword{adni_datadic_fun}
