% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LABDATA}
\alias{LABDATA}
\title{Laboratory Data}
\format{
A data frame with 2463 observations and 133 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(LABDATA)
}
\description{
Laboratory Data data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 2524
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5133
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, nv, f, bl, m18, v01
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: sc, nv, f, bl, m18
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 3
\item \strong{ACCNO} \emph{character}    
\item \strong{COVVIS} \emph{character}   , with options: sc, nv, f, bl, m18, ADNISCR
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{AXT117} \emph{character} Test AXT117; Thyroid Stim. Hormone  
\item \strong{BAT126} \emph{character} Test BAT126; Vitamin B12  
\item \strong{CMT1} \emph{character} Test CMT1; Color , with options: CMC01, SCC09, SCC106, SCC10, SCC311, -1
\item \strong{CMT10} \emph{character} Test CMT10; Urine Nitrite , with options: CMC04, CMC16, SCC09, SCC106, SCC10, SCC311, -1
\item \strong{CMT11} \emph{character} Test CMT11; Leukocyte Esterase , with options: CMC04, CMC07, SCC09, CMC08, CMC06, CMC05, SCC106, SCC10, SCC311, -1
\item \strong{CMT2} \emph{character} Test CMT2; Specific Gravity  
\item \strong{CMT3} \emph{character} Test CMT3; pH , with options: 6, 5, 7, SCC09, 8, SCC106, SCC10, SCC311, -1
\item \strong{CMT43} \emph{character} Test CMT43; Blood (+) , with options: CMC04, CMC07, SCC09, CMC06, CMC05, CMC08, SCC106, SCC10, SCC311, -1
\item \strong{CMT49} \emph{character} Test CMT49; Urine Protein (3+) , with options: CMC04, CMC05, SCC09, CMC06, CMC07, SCC106, SCC10, SCC311, CMC08, -1
\item \strong{CMT5} \emph{character} Test CMT5; Urine Urobilinogen (new) , with options: CMC04, CMC07, CMC06, SCC09, CMC05, SCC106, CMC08, SCC10, SCC311, -1
\item \strong{CMT6} \emph{character} Test CMT6; Urine Ketones , with options: CMC04, SCC09, CMC06, CMC05, CMC08, SCC106, SCC10, SCC311, -1
\item \strong{CMT7} \emph{character} Test CMT7; Urine Bilirubin , with options: CMC04, SCC09, SCC106, SCC10, SCC311, -1
\item \strong{HMT97} \emph{numeric} Test HMT97; Blasts Range value: -1, ..., -1
\item \strong{HMT55} \emph{numeric} Test HMT55; Lymphoma Cell Range value: -1, ..., -1
\item \strong{HMT56} \emph{numeric} Test HMT56; Prolymphocyte Range value: -1, ..., -1
\item \strong{HMT57} \emph{numeric} Test HMT57; Sezary Cell Range value: -1, ..., -1
\item \strong{HMT59} \emph{numeric} Test HMT59; Lymphocyte, Immunoblastic Range value: -1, ..., -1
\item \strong{HMT60} \emph{numeric} Test HMT60; Lymphocyte, Plasmacytoid Range value: -1, ..., -1
\item \strong{HMT61} \emph{numeric} Test HMT61; Lymphocyte, Reactive Range value: -1, ..., -1
\item \strong{HMT62} \emph{numeric} Test HMT62; Lymphoblast Range value: -1, ..., -1
\item \strong{HMT96} \emph{numeric} Test HMT96; Atypical Lymphocytes Range value: -1, ..., 8
\item \strong{HMT63} \emph{numeric} Test HMT63; Promonocyte Range value: -1, ..., -1
\item \strong{HMT64} \emph{numeric} Test HMT64; Monoblast Range value: -1, ..., -1
\item \strong{HMT65} \emph{numeric} Test HMT65; Monocyte, Immature Range value: -1, ..., -1
\item \strong{HMT49} \emph{numeric} Test HMT49; Eosinophil, Immature Range value: -1, ..., -1
\item \strong{HMT48} \emph{numeric} Test HMT48; Basophil, Immature Range value: -1, ..., -1
\item \strong{HMT50} \emph{numeric} Test HMT50; Malighant Cell, NOS Range value: -1, ..., -1
\item \strong{HMT51} \emph{character} Test HMT51; WBC Comment: , with options: HMC61, -1
\item \strong{HMT52} \emph{numeric} Test HMT52; Plasma Cell Precursor Range value: -1, ..., -1
\item \strong{HMT53} \emph{numeric} Test HMT53; Plasma Cell, Immature Range value: -1, ..., -1
\item \strong{HMT54} \emph{numeric} Test HMT54; Plasma Cell, Mature Range value: -1, ..., -1
\item \strong{HMT58} \emph{numeric} Test HMT58; Hairy Cell Range value: -1, ..., -1
\item \strong{HMT70} \emph{numeric} Test HMT70; Nucleated Red Blood Cells Range value: -1, ..., 1
\item \strong{HMT94} \emph{numeric} Test HMT94; Blasts Range value: -1, ..., -1
\item \strong{HMT72} \emph{numeric} Test HMT72; Sezary Cell (Absolute) Range value: -1, ..., -1
\item \strong{HMT80} \emph{numeric} Test HMT80; Lymphoma Cell Range value: -1, ..., -1
\item \strong{HMT81} \emph{numeric} Test HMT81; Prolymphocyte Range value: -1, ..., -1
\item \strong{HMT83} \emph{numeric} Test HMT83; Lymphocyte, Immunoblastic Range value: -1, ..., -1
\item \strong{HMT84} \emph{numeric} Test HMT84; Lymphocyte, Plasmacytoid Range value: -1, ..., -1
\item \strong{HMT85} \emph{numeric} Test HMT85; Lymphocyte, Reactive Range value: -1, ..., -1
\item \strong{HMT86} \emph{numeric} Test HMT86; Lymphoblast Range value: -1, ..., -1
\item \strong{HMT95} \emph{numeric} Test HMT95; Atypical Lymphocytes Range value: -1, ..., 0.48
\item \strong{HMT87} \emph{numeric} Test HMT87; Promonocyte Range value: -1, ..., -1
\item \strong{HMT88} \emph{numeric} Test HMT88; Monoblast Range value: -1, ..., -1
\item \strong{HMT89} \emph{numeric} Test HMT89; Monocyte, Immature Range value: -1, ..., -1
\item \strong{HMT75} \emph{numeric} Test HMT75; Eosinophil, Immature Range value: -1, ..., -1
\item \strong{HMT74} \emph{numeric} Test HMT74; Basophil, Immature Range value: -1, ..., -1
\item \strong{HMT76} \emph{numeric} Test HMT76; Malignant Cell, NOS Range value: -1, ..., -1
\item \strong{HMT77} \emph{numeric} Test HMT77; Plasma Cell Precursor Range value: -1, ..., -1
\item \strong{HMT78} \emph{numeric} Test HMT78; Plasma Cell, Immature Range value: -1, ..., -1
\item \strong{HMT79} \emph{numeric} Test HMT79; Plasma Cell, Mature Range value: -1, ..., -1
\item \strong{HMT82} \emph{numeric} Test HMT82; Hairy Cell Range value: -1, ..., -1
\item \strong{HMT98} \emph{numeric} Test HMT98; Nucleated Red Blood Cells Range value: -1, ..., 1
\item \strong{HMT99} \emph{numeric} Test HMT99; WBC Comment: Range value: -1, ..., -1
\item \strong{HMT21} \emph{numeric} Test HMT21; Bands (\%) Range value: -1, ..., 3
\item \strong{HMT66} \emph{numeric} Test HMT66; Myeloblast Range value: -1, ..., -1
\item \strong{HMT67} \emph{numeric} Test HMT67; Promyelocyte Range value: -1, ..., -1
\item \strong{HMT68} \emph{numeric} Test HMT68; Myelocyte Range value: -1, ..., -1
\item \strong{HMT69} \emph{numeric} Test HMT69; Metamyelocyte Range value: -1, ..., 1
\item \strong{HMT20} \emph{numeric} Test HMT20; Bands Range value: -1, ..., 0.3
\item \strong{HMT90} \emph{numeric} Test HMT90; Myeloblast Range value: -1, ..., -1
\item \strong{HMT91} \emph{numeric} Test HMT91; Promyelocyte Range value: -1, ..., -1
\item \strong{HMT92} \emph{numeric} Test HMT92; Myelocyte Range value: -1, ..., -1
\item \strong{HMT93} \emph{numeric} Test HMT93; Metamyelocyte Range value: -1, ..., 0.1
\item \strong{HMT10} \emph{character} Test HMT10; Monocytes  
\item \strong{HMT100} \emph{character} Test HMT100; MCH  
\item \strong{HMT102} \emph{character} Test HMT102; MCHC  
\item \strong{HMT11} \emph{character} Test HMT11; Eosinophils  
\item \strong{HMT12} \emph{character} Test HMT12; Basophils  
\item \strong{HMT13} \emph{character} Test HMT13; Platelets  
\item \strong{HMT15} \emph{character} Test HMT15; Neutrophils (\%)  
\item \strong{HMT16} \emph{character} Test HMT16; Lymphocytes (\%)  
\item \strong{HMT17} \emph{character} Test HMT17; Monocytes (\%)  
\item \strong{HMT18} \emph{character} Test HMT18; Eosinophils (\%)  
\item \strong{HMT19} \emph{character} Test HMT19; Basophils (\%)  
\item \strong{HMT2} \emph{character} Test HMT2; Hematocrit  
\item \strong{HMT3} \emph{character} Test HMT3; RBC  
\item \strong{HMT4} \emph{character} Test HMT4; MCV  
\item \strong{HMT40} \emph{character} Test HMT40; Hemoglobin  
\item \strong{HMT7} \emph{character} Test HMT7; WBC  
\item \strong{HMT71} \emph{character} Test HMT71; RBC Morphology  
\item \strong{HMT8} \emph{character} Test HMT8; Neutrophils  
\item \strong{HMT9} \emph{character} Test HMT9; Lymphocytes  
\item \strong{ORT1859} \emph{character} Test ORT1859; Trep.Pall.Ab (FTA-ABS) (-70) -38 , with options: ORC402, ORC403, ORC399, SCC293, -1
\item \strong{RCT1} \emph{character} Test RCT1; Total Bilirubin  
\item \strong{RCT11} \emph{character} Test RCT11; Serum Glucose  
\item \strong{RCT12} \emph{character} Test RCT12; Total Protein  
\item \strong{RCT13} \emph{character} Test RCT13; Albumin  
\item \strong{RCT14} \emph{character} Test RCT14; Creatine Kinase  
\item \strong{RCT1407} \emph{character} Test RCT1407; Alkaline Phosphatase  
\item \strong{RCT1408} \emph{character} Test RCT1408; LDH  
\item \strong{RCT183} \emph{character} Test RCT183; Calcium (EDTA)  
\item \strong{RCT19} \emph{character} Test RCT19; Triglycerides (GPO)  
\item \strong{RCT20} \emph{character} Test RCT20; Cholesterol (High Performance)  
\item \strong{RCT29} \emph{character} Test RCT29; Direct Bilirubin  
\item \strong{RCT3} \emph{character} Test RCT3; GGT  
\item \strong{RCT392} \emph{character} Test RCT392; Creatinine (Rate Blanked)  
\item \strong{RCT4} \emph{character} Test RCT4; ALT (SGPT)  
\item \strong{RCT5} \emph{character} Test RCT5; AST (SGOT)  
\item \strong{RCT6} \emph{character} Test RCT6; Urea Nitrogen  
\item \strong{RCT8} \emph{character} Test RCT8; Serum Uric Acid  
\item \strong{RCT9} \emph{character} Test RCT9; Phosphorus  
\item \strong{SLT3} \emph{character} Test SLT3; Rapid Plasma Reagin , with options: SLC03, SLC04, SCC09, ADC05, SCC18, -1
\item \strong{CMT53} \emph{character}   , with options: CMC01, SCC09, CMC77, SCC106, SCC10, SCC311, -1
\item \strong{BAT324} \emph{character}    
\item \strong{UAT13} \emph{character}   , with options: UAC38, UAC39, UAC37, SCC09, SCC18, -1, SCC10, SCC282
\item \strong{UAT10} \emph{character}   , with options: UAC12, UAC4, SCC09, SCC18, -1, UAC3, SCC10, SCC282
\item \strong{UAT11} \emph{character}   , with options: UAC4, UAC12, UAC5, UAC2, UAC3, SCC09, SCC18, -1, SCC10, SCC282
\item \strong{UAT2} \emph{character}    
\item \strong{UAT3} \emph{character}    
\item \strong{UAT43} \emph{character}   , with options: UAC3, UAC12, SCC09, UAC4, UAC2, SCC18, -1, UAC5, SCC10, SCC282
\item \strong{UAT49} \emph{character}   , with options: UAC12, UAC2, UAC3, SCC09, SCC18, -1, UAC5, UAC4, SCC10, SCC282
\item \strong{UAT59} \emph{character}   , with options: UAC1, SCC09, SCC18, UAC18, -1, SCC10, SCC282
\item \strong{UAT65} \emph{numeric}   Range value: -1, ..., -1
\item \strong{UAT7} \emph{character}   , with options: UAC12, SCC09, SCC18, -1, SCC10, SCC282
\item \strong{UAT1} \emph{character}   , with options: UAC41, UAC40, SCC09, SCC18, UAC42, -1, SCC10, SCC282
\item \strong{UAT5} \emph{character}    
\item \strong{UAT6} \emph{character}   , with options: UAC12, SCC09, UAC3, SCC18, -1, UAC2, SCC10, SCC282
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{biospecimen}
