% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LABTESTS}
\alias{LABTESTS}
\title{Clinical Laboratory Tests}
\format{
A data frame with 5474 observations and 22 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(LABTESTS)
}
\description{
Clinical Laboratory Tests data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, f, v01, 4_sc, 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Test Review Date Date: YYY-MM-DD
\item \strong{LABLOOD} \emph{character} Was blood drawn for clinical labs? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{LAURINE} \emph{character} Was urine sample obtained? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{LAABNORM} \emph{character} Are there any clinically significant laboratory abnormalities that would exclude the participant from the study? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{COLLDATE} \emph{ } Sample collection date: Date: YYY-MM-DD
\item \strong{LABLOODML} \emph{numeric} If Yes, how much blood was collected: Range value: 0.5, ..., 80
\item \strong{LAURINEML} \emph{numeric} If Yes, how much urine was collected: Range value: 0.5, ..., 90
\item \strong{LABFAST} \emph{character} Was the participant fasting for at least 6 hours? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 6, ..., 186881
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: sc, init, v1, bl
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{biospecimen}
