% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LB}
\alias{LB}
\title{[DERIVED]  Derived  Laboratory Test Results}
\format{
A data frame with 326837 observations and 25 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(LB)
}
\description{
Laboratory test results derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier ; , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbrevation , with character options: LB
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{ORIGPROT} \emph{character} Original Study Protocol , with character options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{LBGRPID} \emph{character} Group ID Study Protocol of Data Collection; , with character options: ADNI1, ADNI2, ADNI3, ADNIGO, ADNI4
\item \strong{LBSEQ} \emph{integer} Sequence Number Range value: 1, ..., 295
\item \strong{LBTESTCD} \emph{character} Lab Test or Examination Short Name  
\item \strong{LBTEST} \emph{character} Lab Test or Examination Name  
\item \strong{LBORRES} \emph{character} Result or Finding in Original Units  
\item \strong{LBORRESU} \emph{character} Original Units  
\item \strong{LBSTRESC} \emph{character} Character Result/Finding in Std Format  
\item \strong{LBSTRESN} \emph{numeric} Numeric Result/Finding in Standard Units Range value: -5, ..., 37480
\item \strong{LBSTRESU} \emph{character} Standard Units  
\item \strong{LBSTAT} \emph{character} Completion Status , with character options: NOT DONE
\item \strong{LBREASND} \emph{character} Reason Test Not Done  
\item \strong{LBSPEC} \emph{character} Specimen Type , with character options: PLASMA
\item \strong{LBSPCCND} \emph{character} Specimen Condition  
\item \strong{LBNAM} \emph{character} Vendor Name , with character options: UPENN, C2N, URMC
\item \strong{LBBLFL} \emph{character} Baseline Flag , with character options: Y
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -2, ..., Inf
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{VISCODE} \emph{character} Visit Code  
\item \strong{EPOCH} \emph{character} Epoch , with character options: Baseline, Follow-up, Screening
\item \strong{LBDTC} \emph{Date} Date/Time of Specimen Collection Date: YYY-MM-DD
\item \strong{LBDY} \emph{integer} Study Day of Specimen Collection Range value: -363, ..., 7035
}
}
\keyword{derived_dataset}
