% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_csv_files}
\alias{load_csv_files}
\title{Load Single/Multiple CSV File(s) to Specific Environment}
\usage{
load_csv_files(input_dir = NULL, full_file_path = NULL, .envr = NULL)
}
\arguments{
\item{input_dir}{Directory path, Default: NULL}

\item{full_file_path}{Full file path, Default: NULL.}

\item{.envr}{Environment for loading the data, Default: NULL.
 By default, it loads the data into global environment,
please see \code{\link[rlang]{caller_env}}}
}
\value{
Load imported CSV data.
}
\description{
This function is used to load single/multiple CSV files after importing
the file(s) from a specific directory or providing the full file path.
}
\examples{
\dontrun{
input_dir <- system.file("extradata/pacc-raw-input", package = "ADNIMERGE2")
# Load all CSV files to global environment
load_csv_files(
  input_dir = input_dir,
  full_file_path = NULL,
  .envr = globalenv()
)
}
}
\keyword{utils_fun}
