% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LOCLAB}
\alias{LOCLAB}
\title{CSF - Local Lab Results}
\format{
A data frame with 8516 observations and 23 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(LOCLAB)
}
\description{
CSF - Local Lab Results data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment SAMPLEDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{SAMPLECOLL} \emph{character} Was LP conducted? 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{LABLOOD} \emph{character} If Yes, were safety labs collected before the LP? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{CTWHITE} \emph{numeric} White Blood Cell Count 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..100
\item \emph{ADNI3, ADNI4} 
}
\item \strong{CTRED} \emph{numeric} Red Blood Cell Count Range value: 0, ..., 9999
\item \strong{PROTEIN} \emph{numeric} Protein Results Range value: 0, ..., 761
\item \strong{GLUCOSE} \emph{numeric} Glucose Results Range value: 0, ..., 360
\item \strong{SAMPLETIME} \emph{ } Time of Sample Collection hms: HH-MM-SS
\item \strong{SENTTIME} \emph{character} Time sent to Local Lab  
\item \strong{ID} \emph{numeric} Record ID Range value: 2, ..., 187123
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{biospecimen}
