% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAYOADIRL_MRI_FMRI}
\alias{MAYOADIRL_MRI_FMRI}
\title{Jack Lab - Default Mode Network connectivity: 2017-11-07}
\format{
A data frame with 892 observations and 34 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MAYOADIRL_MRI_FMRI)
}
\description{
Jack Lab - Default Mode Network connectivity: 2017-11-07 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 69, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{SCANDATE} \emph{ } scan date Date: YYY-MM-DD
\item \strong{VERSION} \emph{character}   , with options: 10/31/2017
\item \strong{RUNDATE} \emph{ } Run Date Date: YYY-MM-DD
\item \strong{IMAGEUID} \emph{numeric} LONI Image ID Range value: 180734, ..., 836285
\item \strong{STATUS} \emph{character} analysis status , with options: final
\item \strong{ADMNRV} \emph{numeric} Anterior DMN RV Range value: 0.29117, ..., 1.0735
\item \strong{PDMNRV} \emph{numeric} Posterior DMN RV Range value: 0.224, ..., 1.238
\item \strong{DMNRVR} \emph{numeric} DMN RV-ratio Range value: 0.42999, ..., 1.8383
\item \strong{MAGSTRENG} \emph{numeric} Mag Field Range value: 3, ..., 3
\item \strong{SERDESC} \emph{character} Series Description , with options: Extended Resting State fMRI, Resting State fMRI, Extended Resting State fMRI CLEAR
\item \strong{PITCH} \emph{numeric} Max Pitch Change Range value: 0.08206, ..., 6.538
\item \strong{ROLL} \emph{numeric} Max Roll Change Range value: 0.05287, ..., 8.99299
\item \strong{YAW} \emph{numeric} Max Yaw Change Range value: 0.046934, ..., 4.1313
\item \strong{MMX} \emph{numeric} Max X change Range value: 0.04264, ..., 4.3603
\item \strong{MMY} \emph{numeric} Max Y change Range value: 0.05614, ..., 4.9939
\item \strong{MMZ} \emph{numeric} Max Z change Range value: 0.06255, ..., 6.1477
\item \strong{SLICEORD} \emph{character} temporal slice order , with options: 1   3   5   7   9  11  13  15  17  19  21  23  25  27  29  31  33  35  37  39  41  43  45  47   2   4   6   8  10  12  14  16  18  20  22  24  26  28  30  32  34  36  38  40  42  44  46  48, 1   8  15  22  29  36  43   2   9  16  23  30  37  44   3  10  17  24  31  38  45   4  11  18  25  32  39  46   5  12  19  26  33  40  47   6  13  20  27  34  41  48   7  14  21  28  35  42, 1   7  13  19  25  31   2   8  14  20  26  32   3   9  15  21  27  33   4  10  16  22  28  34   5  11  17  23  29  35   6  12  18  24  30  36, 1   9  17  25  33  41  49   2  10  18  26  34  42  50   3  11  19  27  35  43  51   4  12  20  28  36  44  52   5  13  21  29  37  45   6  14  22  30  38  46   7  15  23  31  39  47   8  16  24  32  40  48, 36  30  24  18  12   6  35  29  23  17  11   5  34  28  22  16  10   4  33  27  21  15   9   3  32  26  20  14   8   2  31  25  19  13   7   1, 42  35  28  21  14   7  48  41  34  27  20  13   6  47  40  33  26  19  12   5  46  39  32  25  18  11   4  45  38  31  24  17  10   3  44  37  30  23  16   9   2  43  36  29  22  15   8   1, 48  46  44  42  40  38  36  34  32  30  28  26  24  22  20  18  16  14  12  10   8   6   4   2  47  45  43  41  39  37  35  33  31  29  27  25  23  21  19  17  15  13  11   9   7   5   3   1, 1   3   5   7   9  11  13  15  17  19  21  23  25  27  29  31  33  35  37  39  41  43  45   2   4   6   8  10  12  14  16  18  20  22  24  26  28  30  32  34  36  38  40  42  44, 1   3   5   7   9  11  13  15  17  19  21  23  25  27  29  31  33  35  37  39  41  43  45   2   4   6   8  10  12  14  16  18  20  22  24  26  28  30  32  34  36  38  40  42  44  46
\item \strong{COVVERMIS} \emph{numeric} vermis coverage Range value: 0, ..., 1
\item \strong{COVCEREBEL} \emph{numeric} cerebellum coverage Range value: 0, ..., 1
\item \strong{COVOCCIP} \emph{numeric} occip lobe coverage Range value: 0, ..., 1
\item \strong{COVSUPER} \emph{numeric} superior aspect coverage Range value: 0, ..., 1
\item \strong{COVTEMPOR} \emph{numeric} temporal lobe coverage Range value: 0, ..., 1
\item \strong{COVOTHER} \emph{numeric} other coverage failure Range value: 0, ..., 1
\item \strong{MEANTSNR} \emph{numeric} mean temporal SNR in brain Range value: 12.2656, ..., 160.7456
\item \strong{MEDTSNR} \emph{numeric} median temporal SNR in brain Range value: 10.9972, ..., 126.6814
\item \strong{SDTSNR} \emph{numeric} std deviation of temporal SNR in brain Range value: 6.8661, ..., 378.546
\item \strong{PENCIL} \emph{numeric} pencil artifact observed Range value: 0, ..., 1
\item \strong{VENETIAN} \emph{numeric} venetian blind artifact observed Range value: 0, ..., 1
\item \strong{PHASEDIR} \emph{character} observed phase direction , with options: Indeterminate, PA, AP, NoEval
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
