% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAYOADIRL_MRI_FMRI_NFQ}
\alias{MAYOADIRL_MRI_FMRI_NFQ}
\title{Jack Lab - fMRI Network Failure Quotient (NFQ)}
\format{
A data frame with 2996 observations and 35 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MAYOADIRL_MRI_FMRI_NFQ)
}
\description{
Jack Lab - fMRI Network Failure Quotient (NFQ) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{factor} Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{RID} \emph{numeric} Participant roster ID The 4 digit roster ID (RID) should be used to merge data Range value: 21, ..., 7117
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code Months from baseline rounded to nearest 6 months  
\item \strong{SCANDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{SERIESDATE} \emph{ } SeriesDate From Input DICOM files Date: YYY-MM-DD
\item \strong{SERIESTIME} \emph{ } SeriesTime From Input DICOM files hms: HH-MM-SS
\item \strong{MANUFACTURER} \emph{character} Manufacturer From Input DICOM files , with options: Philips_Medical_Systems, Philips_Healthcare, GE_MEDICAL_SYSTEMS, SIEMENS, Philips, Siemens_Healthineers
\item \strong{MANUFACTURERSMODELNAME} \emph{character} ManufacturersModelName From Input DICOM files  
\item \strong{REPETITIONTIME} \emph{numeric} RepetitionTime From Input DICOM files Range value: 607, ..., 3150
\item \strong{ECHOTIME} \emph{numeric} EchoTime From Input DICOM files Range value: 30, ..., 33
\item \strong{SOFTWAREVERSIONS} \emph{character} SoftwareVersions From Input DICOM files  
\item \strong{SERIESNUMBER} \emph{numeric} SeriesNumber From Input DICOM files Range value: 2, ..., 7801
\item \strong{PATIENTSAGE} \emph{character} PatientsAge From Input DICOM files  
\item \strong{WADDMN_TO_WAVDMN} \emph{numeric} wadDMN_to_wavDMN inter-region connectivity measures Range value: -0.4802, ..., 0.97149
\item \strong{WADDMN_TO_WPDMN} \emph{numeric} wadDMN_to_wpDMN inter-region connectivity measures Range value: -0.56463, ..., 0.87183
\item \strong{WADDMN_TO_WVDMN} \emph{numeric} wadDMN_to_wvDMN inter-region connectivity measures Range value: -0.66088, ..., 0.91702
\item \strong{WAVDMN_TO_WADDMN} \emph{numeric} wavDMN_to_wadDMN inter-region connectivity measures Range value: -0.4802, ..., 0.97149
\item \strong{WAVDMN_TO_WPDMN} \emph{numeric} wavDMN_to_wpDMN inter-region connectivity measures Range value: -0.44221, ..., 0.90101
\item \strong{WAVDMN_TO_WVDMN} \emph{numeric} wavDMN_to_wvDMN inter-region connectivity measures Range value: -0.44634, ..., 0.9
\item \strong{WPDMN_TO_WADDMN} \emph{numeric} wpDMN_to_wadDMN inter-region connectivity measures Range value: -0.56463, ..., 0.87183
\item \strong{WPDMN_TO_WAVDMN} \emph{numeric} wpDMN_to_wavDMN inter-region connectivity measures Range value: -0.44221, ..., 0.90101
\item \strong{WPDMN_TO_WVDMN} \emph{numeric} wpDMN_to_wvDMN inter-region connectivity measures Range value: -0.60345, ..., 0.93997
\item \strong{WVDMN_TO_WADDMN} \emph{numeric} wvDMN_to_wadDMN inter-region connectivity measures Range value: -0.66088, ..., 0.91702
\item \strong{WVDMN_TO_WAVDMN} \emph{numeric} wvDMN_to_wavDMN inter-region connectivity measures Range value: -0.44634, ..., 0.9
\item \strong{WVDMN_TO_WPDMN} \emph{numeric} wvDMN_to_wpDMN inter-region connectivity measures Range value: -0.60345, ..., 0.93997
\item \strong{WPDMN_MEDIAN} \emph{numeric} wpDMN_median within region connectivity estimates Range value: 0.030933, ..., 1.7037
\item \strong{WADDMN_MEDIAN} \emph{numeric} wadDMN_median within region connectivity estimates Range value: -0.16728, ..., 1.7865
\item \strong{WAVDMN_MEDIAN} \emph{numeric} wavDMN_median within region connectivity estimates Range value: 0.028275, ..., 2.0806
\item \strong{WVDMN_MEDIAN} \emph{numeric} wvDMN_median within region connectivity estimates Range value: -0.15038, ..., 2.0523
\item \strong{NFQ} \emph{numeric} calculated network failure quotient Range value: -2.4834, ..., 6.95
\item \strong{OVERALLQC} \emph{character} manual QC assessment for the input data DO NOT USE NFQ VALUES WHEN QC=4 1=excellent ; 2=good ; 3=fair ; 4=fail
\emph{Decoded Value: }\strong{Yes}
\item \strong{SLICETIMING} \emph{character} sliceTiming an underscore-separated list of times (in seconds) for the acquired image slices in the first volume. In ADNI all fMRI the slices are axial and here they are always counted from inferior to superior.   This is invariant to the way the data is stored in DICOM.  For Philips scanners the DICOM data do not indicate the slice acquisition order and it is inferred from knowledge of the protocols in use on the scan acquisition date if necessary.    Slice ordering should be determined from the Siemens CSA header.   GE slice ordering should be determined by the time for each slice in the DICOM header.  The notion of when time equals zero is determined by the vendor data if possible.  
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{imaging}
