% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAYOADIRL_MRI_MCH}
\alias{MAYOADIRL_MRI_MCH}
\title{Jack Lab - ADNI MRI MCH}
\format{
A data frame with 19220 observations and 24 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MAYOADIRL_MRI_MCH)
}
\description{
Jack Lab - ADNI MRI MCH data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI3, ADNI2, ADNIGO
\item \strong{RID} \emph{numeric} The 4 digit roster ID (RID) should be used to merge data Range value: 21, ..., 7123
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Months from baseline rounded to nearest 6 months  
\item \strong{SCANDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{LONI_STUDY} \emph{numeric} LONI Study ID Range value: 27356, ..., 228454
\item \strong{LONI_SERIES} \emph{character} LONI Series ID  
\item \strong{LONI_IMG_ID} \emph{character} LONI Image ID  
\item \strong{SERIES_NUM} \emph{numeric} Series Number Range value: 2, ..., 401029
\item \strong{CHOSEN} \emph{numeric} Indicates if the series was chosen for analysis (0) Not chosen, (1) Chosen
\item \strong{STUDY_QUALITY} \emph{numeric} Study QC Assessment (1) Pass, (4) Fail, (-1) Not evaluated
\item \strong{SERIES_QUALITY} \emph{numeric} Series QC Assessment (1) Excellent, (2) Good, (3) Fair, (4) Unusable, (-1) Not evaluated
\item \strong{UNIQUEID} \emph{numeric} Unique ID of MCH Range value: 1, ..., 604
\item \strong{TYPE} \emph{character} Type of finding , with options: MCH, Other, Hemorrhagic_Infarction, SS, Veinous_Angioma, Hemosiderin_deposition, Cavernous_Angioma, White_Matter_Hemorrhage
\item \strong{STATUS} \emph{character} Status Definite/Possible/None/Rescinded/Enlarged/Not Observed Definite/Not Observed Possible
\item \strong{FINDCOMMENTS} \emph{character} Comments Applicable to Specific Finding  
\item \strong{ATLASREGIONS} \emph{character} Atlas Regions of Finding  
\item \strong{RASLOCATIONS} \emph{character} RAS Location of Finding  
\item \strong{NOFINDINGS} \emph{numeric} No Findings (0) findings exist, (1) no findings exist
\item \strong{SCAN_COMMENT} \emph{character} Comments Applicable to Series Findings Analysis  
\item \strong{EVALDATE} \emph{ } Date of last evalutation POSIXct: YYY-MM-DD
\item \strong{SERIES_UID} \emph{character} Individual Dicom Image String  
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{imaging}
