% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAYOADIRL_MRI_QUALITY_ADNI3}
\alias{MAYOADIRL_MRI_QUALITY_ADNI3}
\title{MAYOADIRL_MRI_QUALITY_ADNI3}
\format{
A data frame with 28065 observations and 29 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MAYOADIRL_MRI_QUALITY_ADNI3)
}
\description{
MAYOADIRL_MRI_QUALITY_ADNI3 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI2, ADNI1, ADNIGO
\item \strong{RID} \emph{numeric}   Range value: 21, ..., 7125
\item \strong{PTID} \emph{character}    
\item \strong{SERIES_DATE} \emph{ }   POSIXct: YYY-MM-DD
\item \strong{STUDYINSTANCEUID} \emph{character}    
\item \strong{STUDY_OVERALLPASS} \emph{numeric}   Range value: 1, ..., 1
\item \strong{STUDY_COMMENTS} \emph{character}    
\item \strong{STUDY_PROTOCOL_STATUS} \emph{numeric}   Range value: -1, ..., 4
\item \strong{STUDY_PROTOCOL_COMMENT} \emph{character}    
\item \strong{STUDY_RESCAN_REQUESTED} \emph{logical}   A Boolean value: TRUE/ FALSE
\item \strong{SERIESINSTANCEUID} \emph{character}    
\item \strong{SERIES_DESCRIPTION} \emph{character}    
\item \strong{SERIES_TYPE} \emph{character}    
\item \strong{SERIES_QUALITY} \emph{numeric}   Range value: -1, ..., 4
\item \strong{SERIES_COMMENTS} \emph{character}    
\item \strong{PROTOCOL_STATUS} \emph{numeric}   Range value: -1, ..., 4
\item \strong{PROTOCOL_COMMENTS} \emph{character}    
\item \strong{SERIES_SELECTED} \emph{logical}   A Boolean value: TRUE/ FALSE
\item \strong{LONI_STUDY} \emph{numeric}   Range value: 107097, ..., 356386
\item \strong{LONI_SERIES} \emph{numeric}   Range value: 527282, ..., 5956068
\item \strong{LONI_IMAGE} \emph{numeric}   Range value: 808811, ..., 11229764
\item \strong{STUDY_MEDICAL_ABNORMALITIES} \emph{numeric}   Range value: -1, ..., 1
\item \strong{SERIAL} \emph{numeric}   Range value: -1, ..., 3
\item \strong{MEDICAL_EXCLUSION} \emph{numeric}   Range value: 0, ..., 0
\item \strong{RELEASE_FROM_QUARANTINE} \emph{numeric}   Range value: 1, ..., 1
\item \strong{PAY_SITE} \emph{numeric}   Range value: 0, ..., 1
\item \strong{QUALIFICATION} \emph{numeric}   Range value: 0, ..., 0
\item \strong{FIELD_STRENGTH} \emph{numeric}   Range value: 3, ..., 3
\item \strong{T1_ACCELERATED} \emph{numeric}   Range value: -1, ..., 1
}
}
\keyword{adni3}
\keyword{imaging}
