% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAYOADIRL_MRI_TBMSYN}
\alias{MAYOADIRL_MRI_TBMSYN}
\title{Jack Lab - TBM-SyN Based Scores}
\format{
A data frame with 2667 observations and 45 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MAYOADIRL_MRI_TBMSYN)
}
\description{
Jack Lab - TBM-SyN Based Scores data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{character}   , with options: ADNI2, ADNI3, ADNIGO
\item \strong{RID} \emph{numeric} Participant roster ID The 4 digit roster ID (RID) should be used to merge data Range value: 21, ..., 7016
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Months from baseline rounded to nearest 6 months  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{numeric} TBMSyN Version Number Range value: 2, ..., 2
\item \strong{PHASEBL} \emph{character} ADNI grant phase baseline , with options: ADNIGO, ADNI3, ADNI2
\item \strong{VISCODEBL} \emph{character} Visit code baseline  
\item \strong{VISCODE2BL} \emph{character} Translated visit code baseline  
\item \strong{DXBL} \emph{character} Participant diagnosis baseline , with options: CN, Dementia, MCI
\item \strong{DX} \emph{character} Participant diagnosis , with options: CN, MCI, Dementia
\item \strong{EXAMDATEBL} \emph{ } Examination Date baseline Date: YYY-MM-DD
\item \strong{SERIESTIME_BL} \emph{ } Series Time bseline hms: HH-MM-SS
\item \strong{SERIESTIME} \emph{ } Series Time hms: HH-MM-SS
\item \strong{SERIESNUMBER_BL} \emph{numeric} Series Number baseline Range value: 2, ..., 7701
\item \strong{SERIESNUMBER} \emph{numeric} Series Number Range value: 2, ..., 201029
\item \strong{SERIESDESCRIPTION_BL} \emph{character} Series Description baseline  
\item \strong{SERIESDESCRIPTION} \emph{character} Series Description  
\item \strong{LONI_SERIESID_BL} \emph{character} LONI Series ID baseline  
\item \strong{LONI_SERIESID} \emph{character} LONI Series ID  
\item \strong{LONI_STUDYID_BL} \emph{numeric} LONI Study ID baseline Range value: 27356, ..., 174807
\item \strong{LONI_STUDYID} \emph{numeric} LONI Study ID Range value: 28860, ..., 218890
\item \strong{LONI_IMAGEID_BL} \emph{character} LONI Image ID baseline  
\item \strong{LONI_IMAGEID} \emph{character} LONI Image ID  
\item \strong{SERIESINSTANCEUID_BL} \emph{character} DICOM SeriesInstanceUID baseline  
\item \strong{SERIESINSTANCEUID} \emph{character} DICOM SeriesInstanceUID  
\item \strong{FIELDSTRENGTH_BL} \emph{numeric} Field Strength Range value: 3, ..., 3
\item \strong{FIELDSTRENGTH} \emph{numeric} Field Strength Range value: 3, ..., 3
\item \strong{MANUFACTURER_BL} \emph{character} MRI Scanner Manufacturer baseline , with options: SIEMENS, PHILIPS MEDICAL SYSTEMS, GE MEDICAL SYSTEMS
\item \strong{MANUFACTURER} \emph{character} MRI Scanner Manufacturer (USE CAUTION IF CHANGED FROM BASELINE) , with options: SIEMENS, PHILIPS MEDICAL SYSTEMS, GE MEDICAL SYSTEMS
\item \strong{MODELNAME_BL} \emph{character} MRI Scanner Model Name baseline  
\item \strong{MODELNAME} \emph{character} MRI Scanner Model Name (USE CAUTION IF CHANGED FROM BASELINE)  
\item \strong{COILNAME_BL} \emph{character} Receive Coil Name baseline  
\item \strong{COILNAME} \emph{character} Receive Coil Name  (USE CAUTION IF CHANGED FROM BASELINE)  
\item \strong{SOFTWAREVERSIONS_BL} \emph{character} MRI Scanner Software Version baseline  
\item \strong{SOFTWAREVERSIONS} \emph{character} MRI Scanner Software Version (USE CAUTION IF CHANGED FROM BASELINE)  
\item \strong{QUALITY_BL} \emph{numeric} Series Quality baseline <blank> Not evaluated, (1) Excellent, (2) Good, (3) Fair, (4) Unusable
\item \strong{QUALITY} \emph{numeric} Series Quality <blank> Not evaluated, (1) Excellent, (2) Good, (3) Fair, (4) Unusable
\item \strong{STUDY_OVERALLPASS_BL} \emph{numeric} Study Overall Quality Assessment baseline (-1) Not evaluated, (0) Fail, (1) Pass
\item \strong{STUDY_OVERALLPASS} \emph{numeric} Study Overall Quality Assessment (-1) Not evaluated, (0) Fail, (1) Pass
\item \strong{RUNDATE} \emph{ } Run Date Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Status of data run. , with options: Partial
\item \strong{TBMSYNSCOR} \emph{numeric} The TBM SyN score is the mean over 31 ROIs, of the annualized log Jacobian determinant of the SyN deformation computed by warping the follow-up and baseline images together. Range value: -0.072333, ..., 0.046604
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{imaging}
