% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEDHIST}
\alias{MEDHIST}
\title{Medical History}
\format{
A data frame with 3083 observations and 40 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MEDHIST)
}
\description{
Medical History data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, f, m48, m36, m60, v01, v06
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: sc, f, m48, m36, m60, m12, m72, m24, m84
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{MHSOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{MHPSYCH} \emph{character} 1. Psychiatric 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH2NEURL} \emph{character} 2. Neurologic (other than AD) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH3HEAD} \emph{character} 3. Head, Eyes, Ears, Nose and Throat 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH4CARD} \emph{character} 4. Cardiovascular 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH5RESP} \emph{character} 5. Respiratory 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH6HEPAT} \emph{character} 6. Hepatic 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH7DERM} \emph{character} 7. Dermatologic-Connective Tissue 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH8MUSCL} \emph{character} 8. Musculoskeletal 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH9ENDO} \emph{character} 9. Endocrine-Metabolic 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH10GAST} \emph{character} 10. Gastrointestinal 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH11HEMA} \emph{character} 11. Hematopoietic-Lymphatic 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH12RENA} \emph{character} 12. Renal-Genitourinary 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH13ALLE} \emph{character} 13. Allergies or Drug Sensitivities 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH14ALCH} \emph{character} 14. Alcohol Abuse 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH14AALCH} \emph{numeric} 14a. During period of alcohol abuse, estimate the average number of drinks per day 0..30
\item \strong{MH14BALCH} \emph{numeric} 14b. Duration of abuse (years) 1..99
\item \strong{MH14CALCH} \emph{numeric} 14c. Period of time since end of abuse (years) 0..99
\item \strong{MH15DRUG} \emph{character} 15. Drug Abuse 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH15ADRUG} \emph{numeric} 15a.  Duration of abuse (years): 1..99
\item \strong{MH15BDRUG} \emph{numeric} 15b.  Period of time since end of abuse (years): 0..99
\item \strong{MH16SMOK} \emph{character} 16. Smoking 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH16ASMOK} \emph{numeric} 16a. During periods of smoking, the average number of packs/day 0..10
\item \strong{MH16BSMOK} \emph{numeric} 16b. Duration (years) 1..99
\item \strong{MH16CSMOK} \emph{numeric} 16c. If no longer smoking, provide period of time since stopped smoking (years) 0..99
\item \strong{MH17MALI} \emph{character} 17. Malignancy 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH18SURG} \emph{character} 18. Major Surgical Procedures 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MH19OTHR} \emph{character} 19. Other (if none, select 'No') 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 2934
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{medical_history}
