% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METACORES}
\alias{METACORES}
\title{ADNI Metadata-Specs}
\format{
A R6-class wrapper object created using \code{\link[metacore]{metacore}} function.
}
\source{
For more details about the metadata-specs see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Analysis-Meta-Specs', package = 'ADNIMERGE2')}.
}
\usage{
data(METACORES)
}
\description{
Metadata specifications for the ADNI study. It is generated to create analysis ready dataset using \href{https://pharmaverse.org/}{PHARMAVERSE} workflow for illustration purpose.
}
\examples{
\dontrun{
ADNIMERGE2::METACORES
}
}
\keyword{derived}
\keyword{derived_dataset}
\keyword{metadata}
\keyword{specs}
