% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MMSE}
\alias{MMSE}
\title{Mini Mental State Exam}
\format{
A data frame with 14599 observations and 59 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MMSE)
}
\description{
Mini Mental State Exam data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment DATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{DONE} \emph{character} Was assessment/procedure done? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 1=Participant Unable for Cognitive Reasons ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{SOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call/Video Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMDATE} \emph{character} Score (date) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMYEAR} \emph{character} Score (year) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMMONTH} \emph{character} Score (month) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMDAY} \emph{character} Score (day of the week) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMSEASON} \emph{character} Score (season) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMHOSPIT} \emph{character} Score (building) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMFLOOR} \emph{character} Score (floor) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMCITY} \emph{character} Score (city/town) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMAREA} \emph{character} Score (county) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMSTATE} \emph{character} Score (state) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WORDLIST} \emph{character} Which list was used? 1=Ball, Flag, Tree ; 2=Apple, Penny, Table
\emph{Decoded Value: }\strong{Yes}
\item \strong{WORD1} \emph{character} Ball (alt: Apple) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WORD2} \emph{character} Flag (alt: Penny) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WORD3} \emph{character} Tree (alt: Table) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMTRIALS} \emph{numeric} 13a. Enter number of trials 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1..6
\item \emph{ADNI3} 
}
\item \strong{MMD} \emph{character} 14. D 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MML} \emph{character} 15. L 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMR} \emph{character} 16. R 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMO} \emph{character} 17. O 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMW} \emph{character} 18. W 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMLTR1} \emph{character} Indicate 1st letter  
\item \strong{MMLTR2} \emph{character} Indicate 2nd letter  
\item \strong{MMLTR3} \emph{character} Indicate 3rd letter  
\item \strong{MMLTR4} \emph{character} Indicate 4th letter  
\item \strong{MMLTR5} \emph{character} Indicate 5th letter  
\item \strong{MMLTR6} \emph{character} Indicate 6th letter (if given) , with options: -, w, D, W, O, S, d, x, M, Y
\item \strong{MMLTR7} \emph{character} Indicate 7th letter (if given) , with options: -, _, x, W
\item \strong{WORLDSCORE} \emph{numeric} Score:  World Backwards Range value: 0, ..., 5
\item \strong{WORD1DL} \emph{character} Ball (alt: Apple) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WORD2DL} \emph{character} Flag (alt: Penny) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WORD3DL} \emph{character} Tree (alt: Table) 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMWATCH} \emph{character} What is this? [Point to a watch] 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMPENCIL} \emph{character} What is this? [Point to a pencil or pen] 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMREPEAT} \emph{character} No ifs, ands, or buts 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMHAND} \emph{character} Takes in right hand 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMFOLD} \emph{character} Folds in half 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMONFLR} \emph{character} Puts on floor 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMREAD} \emph{character} Close your eyes 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMWRITE} \emph{character} Please write a sentence 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMDRAW} \emph{character} Please copy this design 0=0 - Incorrect ; 1=1 - Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMSCORE} \emph{numeric} MMSE TOTAL SCORE 
\itemize{
\item \emph{ADNI1, ADNI2, ADNI3:} 0..30
\item \emph{ADNIGO, ADNI4} 
}
\item \strong{ID} \emph{numeric} Record ID Range value: 10, ..., 187107
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
