% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOCA}
\alias{MOCA}
\title{MoCA}
\format{
A data frame with 8962 observations and 59 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MOCA)
}
\description{
MoCA data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment MCDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{TRAILS} \emph{character} Trails 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CUBE} \emph{character} Copy Cube 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKCON} \emph{character} Draw Clock - Contour 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKNO} \emph{character} Draw Clock - Number 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKHAN} \emph{character} Draw Clock - Hands 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{LION} \emph{character} Lion 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RHINO} \emph{character} Rhinoceros 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CAMEL} \emph{character} Camel 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT1W1} \emph{character} Immediate (#1): Face 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT1W2} \emph{character} Immediate (#1): Velvet 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT1W3} \emph{character} Immediate (#1): Church 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT1W4} \emph{character} Immediate (#1): Daisy 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT1W5} \emph{character} Immediate (#1): Red 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT2W1} \emph{character} Immediate (#2): Face 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT2W2} \emph{character} Immediate (#2): Velvet 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT2W3} \emph{character} Immediate (#2): Church 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT2W4} \emph{character} Immediate (#2): Daisy 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{IMMT2W5} \emph{character} Immediate (#2): Red 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DIGFOR} \emph{character} Digits Forward 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DIGBACK} \emph{character} Digits Backward 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{LETTERS} \emph{numeric} List of Letters/Tapping:  # Errors 
\itemize{
\item \emph{ADNIGO, ADNI2:} 0..29
\item \emph{ADNI3, ADNI4} 
}
\item \strong{SERIAL1} \emph{character} Serial 7: 1st Subtraction 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SERIAL2} \emph{character} Serial 7: 2nd Subtraction 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SERIAL3} \emph{character} Serial 7: 3rd Subtraction 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SERIAL4} \emph{character} Serial 7: 4th Subtraction 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SERIAL5} \emph{character} Serial 7: 5th Subtraction 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{REPEAT1} \emph{character} Repeat Sentence: I only know that John is the one to help today. 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{REPEAT2} \emph{character} Repeat Sentence: The cat always hid under the couch when dogs were in the room. 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FFLUENCY} \emph{numeric} Letter Fluency - F: Total number of correct words 
\itemize{
\item \emph{ADNIGO, ADNI2:} 0..99
\item \emph{ADNI3, ADNI4} 
}
\item \strong{ABSTRAN} \emph{character} Abstraction: train-bicycle 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{ABSMEAS} \emph{character} Abstraction: watch-ruler 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DELW1} \emph{character} Delayed: Face 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DELW2} \emph{character} Delayed: Velvet 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DELW3} \emph{character} Delayed: Church 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DELW4} \emph{character} Delayed: Daisy 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DELW5} \emph{character} Delayed: Red 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DATE} \emph{character} Date 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MONTH} \emph{character} Month 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{YEAR} \emph{character} Year 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{DAY} \emph{character} Day 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PLACE} \emph{character} Place 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CITY} \emph{character} City 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MOCA} \emph{numeric} MOCA Total Score 
\itemize{
\item \emph{ADNI3:} 0..30
\item \emph{ADNI4} 
}
\item \strong{SOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call/Video Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 4, ..., 187102
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English ; s=Spanish
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
