% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MODHACH}
\alias{MODHACH}
\title{Modified Hachinski}
\format{
A data frame with 3462 observations and 22 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MODHACH)
}
\description{
Modified Hachinski data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 7126
\item \strong{VISCODE} \emph{character}   , with options: sc, f, v01
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: sc, f
\item \strong{VISDATE} \emph{ } Assessment DATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{HMONSET} \emph{character} 1. Abrupt Onset of Dementia 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 2=Present - 2 points ; 0=Absent
\item \emph{ADNI3:} 0=Absent ; 2=Present - 2 points
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HMSTEPWS} \emph{character} 2. Stepwise Deterioration of Dementia 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Present - 1 point ; 0=Absent
\item \emph{ADNI3:} 0=Absent ; 1=Present - 1 point
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HMSOMATC} \emph{character} 3. Somatic Complaints 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Present - 1 point ; 0=Absent
\item \emph{ADNI3:} 0=Absent ; 1=Present - 1 point
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HMEMOTIO} \emph{character} 4. Emotional Incontinence 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Present - 1 point ; 0=Absent
\item \emph{ADNI3:} 0=Absent ; 1=Present - 1 point
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HMHYPERT} \emph{character} 5. History of Hypertension 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Present - 1 point ; 0=Absent
\item \emph{ADNI3:} 0=Absent ; 1=Present - 1 point
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HMSTROKE} \emph{character} 6. History of Stroke 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 2=Present - 2 points ; 0=Absent
\item \emph{ADNI3:} 0=Absent ; 2=Present - 2 points
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HMNEURSM} \emph{character} 7. Focal Neurologic Symptoms 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 2=Present - 2 points ; 0=Absent
\item \emph{ADNI3:} 0=Absent ; 2=Present - 2 points
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HMNEURSG} \emph{character} 8. Focal Neurologic Signs 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 2=Present - 2 points ; 0=Absent
\item \emph{ADNI3:} 0=Absent ; 2=Present - 2 points
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HMSCORE} \emph{numeric} TOTAL SCORE 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..12
\item \emph{ADNI3} 
}
\item \strong{SOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call/Video Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 6, ..., 186318
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{assessments}
