% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRI3META}
\alias{MRI3META}
\title{3T MRI Scan Information}
\format{
A data frame with 9278 observations and 43 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRI3META)
}
\description{
3T MRI Scan Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 15, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character}    
\item \strong{FIELD_STRENGTH} \emph{character} Field Strength 3T=3 Tesla
\emph{Decoded Value: }\strong{Yes}
\item \strong{RECNO} \emph{character}   1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMCONDCT} \emph{character} Was the scan conducted? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMREASON} \emph{character} Reason why the scan was not conducted: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{MMSCOUT} \emph{character} 1. Tri-Planar Scout (if available, otherwise use an axial scout) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMSMPRAGE} \emph{character} 2. Straight Sagittal MPRAGE Sequence 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMRMPRAGE} \emph{character} 3. Repeat Straight Sagittal MPRAGE Sequence 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMB1HEAD} \emph{character} 4. B1 Calibration Head Coil Scan (Only applicable for phased array head coil on GE and Siemens systems) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMB1BODY} \emph{character} 5. B1 Calibration Body Coil Scan (Only applicable for phased array head coil on GE and Siemens systems) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMECHO} \emph{character} 6. Straight Axial Fast or Turbo Spin Echo 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMPHAN} \emph{character} 7. In new exam; Perform ADNI QC Scan. Localizer and Straight Sagittal MPRAGE (with increased slice thickness to ensure phantom coverage) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMTRNSFR} \emph{character} Was data transferred to LONI within 24 hours of scan? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMTRNDATE} \emph{character} Transfer Date  
\item \strong{MMTRNDATEYR_DRVD} \emph{numeric} Transfer Year (Derived from MMTRNDATE) Range value: 2005, ..., 2025
\item \strong{MMARCHIVE} \emph{character} Data Archived Locally 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMLPDONE} \emph{character} 9. Was a Lumbar Puncture completed prior to the MRI scan? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMLPINTER} \emph{character} If Yes, What was the interval between LP and MRI? 1=less than 6 hours ; 2=6-12 hours ; 3=13-24 hours ; 4=25-48 hours ; 5=49-72 hours ; 6=more than 72 hours
\emph{Decoded Value: }\strong{Yes}
\item \strong{SEDATIVE} \emph{numeric}   Range value: 0, ..., 1
\item \strong{OTRDEV} \emph{numeric}   Range value: 0, ..., 1
\item \strong{MOTION} \emph{numeric}   Range value: 0, ..., 1
\item \strong{EYEKEEP} \emph{numeric}   Range value: 0, ..., 2
\item \strong{EYEASK} \emph{numeric}   Range value: 0, ..., 2
\item \strong{EYEKEEP4} \emph{numeric}   Range value: 0, ..., 2
\item \strong{EYEASK4} \emph{numeric}   Range value: 0, ..., 2
\item \strong{EYEKEEP6} \emph{numeric}   Range value: 0, ..., 2
\item \strong{EYEASK6} \emph{numeric}   Range value: 0, ..., 2
\item \strong{MALFUNC} \emph{numeric}   Range value: 0, ..., 1
\item \strong{FLDSTRNGTH} \emph{numeric}   Range value: 2, ..., 2
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 187111
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character}   , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character}   , with options: e
\item \strong{HAS_QC_ERROR} \emph{numeric}   Range value: 0, ..., 0
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{imaging}
