% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRI_INFARCTS}
\alias{MRI_INFARCTS}
\title{MRI Infarcts: 2021-01-29}
\format{
A data frame with 8593 observations and 16 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRI_INFARCTS)
}
\description{
MRI Infarcts: 2021-01-29 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI1, ADNI3, ADNI2, ADNIGO, nv
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 6795
\item \strong{VISCODE} \emph{character} Visit code ADNI code
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{MRI.DATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{MRI.DATE1} \emph{ }   Date: YYY-MM-DD
\item \strong{ANALYZE.DATE} \emph{ }   Date: YYY-MM-DD
\item \strong{SIDE} \emph{character} Side of Infarct, if detected R/L or -
\item \strong{SIZE} \emph{character} Size of Infarct, if detected Small/Large or -
\item \strong{CVA.NUMBER} \emph{numeric}   Range value: 0, ..., 22
\item \strong{CVA.LOCATION} \emph{character}    
\item \strong{CVA.TYPE} \emph{character}   , with options: -, THROMBOSIS, HEMORRHAGE, UNANALYZABLE
\item \strong{XYZ} \emph{character}    
\item \strong{IMGFILE} \emph{character}    
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{imaging}
