% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRIB1CALIB}
\alias{MRIB1CALIB}
\title{MRI B1 Calibration}
\format{
A data frame with 4480 observations and 18 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRIB1CALIB)
}
\description{
MRI B1 Calibration data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 8, ..., 8974
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1435
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 17
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{FLDSTRENG} \emph{character} Field Strength 1.5=1.5T MRI ; 3=3T MRI
\emph{Decoded Value: }\strong{Yes}
\item \strong{COILTYPE} \emph{character} Coil Type  
\item \strong{SCAN} \emph{character} Scan 1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{LONISID} \emph{numeric} LONI Study ID Range value: 1945, ..., 29427
\item \strong{PRESENT} \emph{character}   0=Present and adequate ; 1=Present NOT adequate ; 2=Not present/not expected ; 3=Missing
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMMENTS} \emph{numeric}   Range value: Inf, ..., -Inf
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{imaging}
