% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRIFind}
\alias{MRIFind}
\title{MAYO ADIR LAB MRI Findings}
\format{
A data frame with 3166 observations and 42 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRIFind)
}
\description{
MAYO ADIR LAB MRI Findings data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI3, ADNI1, ADNI2, ADNIGO
\item \strong{ParticipantID} \emph{character} ADNI Participant ID; SSS_S_PPPP or SSS_S_PPPPP where SSS is the SITE ID and PPPP or PPPPP is the ParticipantID  
\item \strong{RID} \emph{numeric}   Range value: 21, ..., 10849
\item \strong{VISCODE} \emph{character}   , with options: 4_sc, 4_init, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character}    
\item \strong{StudyDate} \emph{ } Study Date YYYYMMDD
\item \strong{SeriesTime} \emph{ } Time Series Started on the Scanner HHMMSS ; <blank> Unknown
\item \strong{SeriesNumber} \emph{numeric} DICOM Series Number Range value: 2, ..., 5001
\item \strong{SeriesType} \emph{character} Series Acquisition Type , with options: dMRI, MET2starw, FLAIR, T2starw
\item \strong{SiteID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{UniqueID} \emph{numeric} Unique ID of Finding Range value: 1, ..., 218
\item \strong{FindingsType} \emph{character} Type of Finding  
\item \strong{Status} \emph{character} Finding Status Definite ; Possible ; Not Observed Possible ; Not Observed Definite ; Rescinded ; Resolved ; Failed Image ; Merged ; Enlarged ; NoEval ; None ; <blank>
\item \strong{FindingComments} \emph{character} Comment(s) Applicable to Specific Finding  
\item \strong{AtlasRegions} \emph{character} Atlas Region of Finding  
\item \strong{RASLocations} \emph{character} RAS Location of Finding, Typically Relative to Isocenter  
\item \strong{ARIARegion} \emph{logical} ARIA Region(s) of Finding A Boolean value: TRUE/ FALSE
\item \strong{ARIALength} \emph{logical} ARIA Length of Finding A Boolean value: TRUE/ FALSE
\item \strong{GMProbability} \emph{character} List of Gray Matter Probabilities for each voxel in the finding  
\item \strong{WMProbability} \emph{character} List of White Matter Probabilities for each voxel in the finding  
\item \strong{CSFProbability} \emph{character} List of CSF Probabilities for each voxel in the finding  
\item \strong{MaxDiameter} \emph{numeric} Maximum Diameter of Relevant Finding Range value: 0, ..., 24
\item \strong{SliceCount} \emph{numeric} Extent of Slices on which finding is observed Range value: 0, ..., 17
\item \strong{NoFindings} \emph{numeric} No Findings Associated with Series (-1) Not assessed ; (0) Findings identified ; (1) No findings identified
\item \strong{EvalDate} \emph{ } Evaluation Date YYYYMMDD:HHMMSS
\item \strong{ScannerManufacturer} \emph{character} DICOM Scanner Manufacturer Unknown, GE, Philips, Siemens, Other
\item \strong{ScannerModel} \emph{character} DICOM Scanner Model  
\item \strong{SoftwareVersion} \emph{character} DICOM Software Version; may be extracted from vendor private tags  
\item \strong{MagneticFieldStrength} \emph{numeric} Scanner Field Strength <blank>, 1.5, 3, 7
\item \strong{Acceleration} \emph{character}   , with options: Accelerated, Ultrafast, Unaccelerated
\item \strong{AcquisitionType} \emph{numeric}   Range value: 2, ..., 3
\item \strong{AcquisitionPlane} \emph{character}   , with options: Axial, Sagittal
\item \strong{NumberVolumes} \emph{numeric}   Range value: 1, ..., 127
\item \strong{SlicesPerVolume} \emph{numeric}   Range value: 35, ..., 211
\item \strong{SliceThickness} \emph{numeric}   Range value: 1, ..., 4
\item \strong{ReceiveCoilName} \emph{character} Receive Coil Name  
\item \strong{StudyInstanceUID} \emph{character} DICOM StudyInstanceUID  
\item \strong{SeriesInstanceUID} \emph{character} DICOM SeriesInstanceUID  
\item \strong{LONIStudy} \emph{numeric} LONI Study ID Range value: 242527, ..., 363432
\item \strong{LONISeries} \emph{numeric} LONI Series ID Range value: 5162601, ..., 6020790
\item \strong{LONIImage} \emph{numeric} LONI Image ID Range value: 10248885, ..., 11328357
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
