% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRIMETA}
\alias{MRIMETA}
\title{1.5T MRI Scan Information}
\format{
A data frame with 4740 observations and 36 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRIMETA)
}
\description{
1.5T MRI Scan Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 6837
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{FIELD_STRENGTH} \emph{character} Field Strength 
\itemize{
\item \emph{ADNI1, ADNIGO:} 1.5T=1.5 Tesla
\item \emph{ADNI2, ADNI3:} 1.5T=1.5 Tesla ; 3T=3 Tesla
\item \emph{ADNI4:} 3T=3 Tesla
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RECNO} \emph{character}   1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMCONDCT} \emph{character} Was the scan conducted? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMREASON} \emph{character} If No, please provide reason why the scan was not conducted: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMSCOUT} \emph{character} 1. Tri-Planar Scout (if available, otherwise use an axial scout) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMSMPRAGE} \emph{character} 2. Straight Sagittal MPRAGE Sequence 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMRMPRAGE} \emph{character} 3. Repeat Straight Sagittal MPRAGE Sequence 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMB1HEAD} \emph{character} 4. B1 Calibration Head Coil Scan (Only applicable for phased array head coil on GE and Siemens systems) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMB1BODY} \emph{character} 5. B1 Calibration Body Coil Scan (Only applicable for phased array head coil on GE and Siemens systems) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMECHO} \emph{character} 6. Straight Axial Fast or Turbo Spin Echo 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMPHAN} \emph{character} 7. In new exam; Perform ADNI QC Scan. Localizer and Straight Sagittal MPRAGE (with increased slice thickness to ensure phantom coverage) 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMTRNSFR} \emph{character} Was data transferred to LONI within 24 hours of scan? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMTRNDATE} \emph{character} Transfer Date  
\item \strong{MMTRNDATEYR_DRVD} \emph{numeric} Transfer Year (Derived from MMTRNDATE) Range value: 2005, ..., 2023
\item \strong{MMARCHIVE} \emph{character} Data Archived Locally 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMLPDONE} \emph{character} Was a Lumbar Puncture completed prior to the MRI scan? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MMLPINTER} \emph{character} If Yes, What was the interval between LP and MRI? 1=less than 6 hours ; 2=6-12 hours ; 3=13-24 hours ; 4=25-48 hours ; 5=49-72 hours ; 6=more than 72 hours
\emph{Decoded Value: }\strong{Yes}
\item \strong{SEDATIVE} \emph{character} Was sedation medication consumed for use during MRI scan? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{OTRDEV} \emph{character} Other Protocol Variations 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MOTION} \emph{character} Patient Motion Problems: 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{EYEKEEP} \emph{character} Did the participant keep their eyes open (MRI Tech:  ask the subject right after the scan) 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No ; 2=N/A Scanner manufacturer is GE
\item \emph{ADNI3:} 0=No ; 1=Yes ; 2=N/A Scanner manufacturer is GE
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{EYEASK} \emph{character} Was the participant instructed to open their eyes? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No ; 2=N/A Scanner manufacturer is GE
\item \emph{ADNI3:} 0=No ; 1=Yes ; 2=N/A Scanner manufacturer is GE
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{MALFUNC} \emph{character} Scanner Malfunction: 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FLDSTRNGTH} \emph{character} Field Strength 1=1.5 Tesla ; 2=3 Tesla
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 6, ..., 176409
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{UPDATE_STAMP} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{imaging}
