% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRIMPPRO}
\alias{MRIMPPRO}
\title{MRI MPRAGE Process}
\format{
A data frame with 3269 observations and 17 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRIMPPRO)
}
\description{
MRI MPRAGE Process data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 6546
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1430
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 2
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{FLDSTRENG} \emph{numeric} Field Strength Range value: 1.5, ..., 3
\item \strong{COILTYPE} \emph{character} Coil Type  
\item \strong{SCAN} \emph{character} Scan 1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{LONISID} \emph{numeric} LONI Study ID Range value: 1945, ..., 29427
\item \strong{COMMENTS} \emph{character}    
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{imaging}
