% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRIMPRANK}
\alias{MRIMPRANK}
\title{MRI MPRAGE Ranking}
\format{
A data frame with 9412 observations and 18 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRIMPRANK)
}
\description{
MRI MPRAGE Ranking data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 8, ..., 19134
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1435
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 34
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{FLDSTRENG} \emph{numeric} Field Strength Range value: 1.5, ..., 3
\item \strong{COILTYPE} \emph{character} Coil Type  
\item \strong{SCAN} \emph{character} Scan 1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{LONISID} \emph{numeric} LONI Study ID Range value: 1945, ..., 29427
\item \strong{LONIUID} \emph{character} LONI Unique Series ID  
\item \strong{RANK} \emph{numeric} MPRAGE Rankings: The MPRAGE rankings are done in a qualitative manner.A scan with a 1 ranking is considered a high quality scan.  A 2 ranking is considered a medium quality scan.  3 ranking is a low quality scan.3 is the lowest ranking a scan can receive and still be considered a passing scan.  Anything below 3 would be a failure for quality.  Choosing MPRAGES: For scans ranked 2 and 3 the scan ranked 2 (whatever is the lower ranking of the two) should always be the "preferred/chosen"MPRAGE.  If the rankings are tied in ranking the QC-er would chose the"sharper of the two". If the two scans are identical the first MPRAGE was always chosen. The preferred scan is indicated by the LONIUID field in the MRIMPPRO table. Range value: 1, ..., 5
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{imaging}
