% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRINCLUSIO}
\alias{MRINCLUSIO}
\title{MRI Subject Inclusion}
\format{
A data frame with 9805 observations and 39 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRINCLUSIO)
}
\description{
MRI Subject Inclusion data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 386, ..., 70898
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5296
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 5
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{LONISID} \emph{numeric} LONI Study ID Range value: 1945, ..., 135496
\item \strong{FLDSTRENG} \emph{character} Field Strength 1.5=1.5T MRI ; 3=3T MRI
\emph{Decoded Value: }\strong{Yes}
\item \strong{SCAN} \emph{character} Scan 1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{QCTYPE} \emph{character} QC Type 1=Clipy ; 2=Grinder
\emph{Decoded Value: }\strong{Yes}
\item \strong{NOMEDICAL} \emph{character} Image Medical Issues (check all that apply): 1=None
\emph{Decoded Value: }\strong{Yes}
\item \strong{SURGERY} \emph{character}   1=Surgery
\emph{Decoded Value: }\strong{Yes}
\item \strong{HEMMORHAGE} \emph{character}   1=Hemmorhage
\emph{Decoded Value: }\strong{Yes}
\item \strong{DEVANOMALY} \emph{character}   1=Development Anomaly
\emph{Decoded Value: }\strong{Yes}
\item \strong{LESION} \emph{character}   1=Space Occupying Lesion
\emph{Decoded Value: }\strong{Yes}
\item \strong{ATROPHY} \emph{character}   1=Atypical Atrophy
\emph{Decoded Value: }\strong{Yes}
\item \strong{INFARCTION} \emph{character}   1=Hemispheric Infarction
\emph{Decoded Value: }\strong{Yes}
\item \strong{TRAUMA} \emph{character}   1=Prior Head Trauma
\emph{Decoded Value: }\strong{Yes}
\item \strong{METALLIC} \emph{character}   1=Metallic Artifact
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPH} \emph{character}   1=NPH
\emph{Decoded Value: }\strong{Yes}
\item \strong{EDEMA} \emph{character}   1=Cerebral Edema
\emph{Decoded Value: }\strong{Yes}
\item \strong{OTHER} \emph{character}   1=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{INCLUSION} \emph{character} Inclusion 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{STEREOMARK} \emph{character} Stereotactic Marker present -1 =Not Evaluated ; 0=Missing ; 1=Present Correct ; 2=Present Incorrect
\emph{Decoded Value: }\strong{Yes}
\item \strong{SERIAL} \emph{character} Serial scan -1 =Not Evaluated ; 0=Baseline ; 1=Matches Previous ; 2=Doesn't Match
\emph{Decoded Value: }\strong{Yes}
\item \strong{MEDEXCL} \emph{character} Medical Exclusion (see Comments for details) -1 =Not Evaluated ; 0=No ; 1=Yes (not exclusion) ; 2=Yes (potential exclusion)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PASS} \emph{character} Pass: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{QUARANTINE} \emph{character} Release from Quarantine: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PAYPATIENT} \emph{character} Pay Patient: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PAYSITE} \emph{character} Pay Site: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{RESCAN} \emph{character} Rescan 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMMENTS} \emph{character} Comments  
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{imaging}
