% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRINFQ}
\alias{MRINFQ}
\title{MAYO ADIR LAB MRI Network Failure Quotient}
\format{
A data frame with 781 observations and 47 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRINFQ)
}
\description{
MAYO ADIR LAB MRI Network Failure Quotient data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI2, ADNI3, ADNI1, ADNIGO
\item \strong{ParticipantID} \emph{character} ADNI Participant ID; SSS_S_PPPP or SSS_S_PPPPP where SSS is the SITE ID and PPPP or PPPPP is the ParticipantID  
\item \strong{RID} \emph{numeric}   Range value: 21, ..., 10849
\item \strong{VISCODE} \emph{character}   , with options: 4_sc, 4_init, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character}    
\item \strong{StudyDate} \emph{ } Study Date YYYYMMDD
\item \strong{SeriesTime} \emph{ } Time Series Started on the Scanner HHMMSS ; <blank> Unknown
\item \strong{SeriesNumber} \emph{numeric} DICOM Series Number Range value: 4, ..., 90
\item \strong{SeriesType} \emph{character} Series Acquisition Type , with options: BOLD
\item \strong{SiteID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{RepetitionTime} \emph{numeric} Repetition Time Range value: 607, ..., 3000
\item \strong{EchoTime} \emph{numeric} Echo Time Range value: 30, ..., 35
\item \strong{wadDMN_to_wavDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.29222, ..., 0.95566
\item \strong{wadDMN_to_wpDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.62843, ..., 0.92319
\item \strong{wadDMN_to_wvDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.18897, ..., 0.92799
\item \strong{wavDMN_to_wadDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.29222, ..., 0.95566
\item \strong{wavDMN_to_wpDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.54247, ..., 0.88718
\item \strong{wavDMN_to_wvDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.34795, ..., 0.89694
\item \strong{wpDMN_to_wadDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.62843, ..., 0.92319
\item \strong{wpDMN_to_wavDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.54247, ..., 0.88718
\item \strong{wpDMN_to_wvDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.47204, ..., 0.94144
\item \strong{wvDMN_to_wadDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.18897, ..., 0.92799
\item \strong{wvDMN_to_wavDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.34795, ..., 0.89694
\item \strong{wvDMN_to_wpDMN} \emph{numeric} Inter-region Connectivity Measure Range value: -0.47204, ..., 0.94144
\item \strong{wpDMN_median} \emph{numeric} Within Region Connectivity Estimate Range value: 0.016631, ..., 1.3674
\item \strong{wadDMN_median} \emph{numeric} Within Region Connectivity Estimate Range value: -0.13334, ..., 1.3771
\item \strong{wavDMN_median} \emph{numeric} Within Region Connectivity Estimate Range value: 0.025276, ..., 1.6033
\item \strong{wvDMN_median} \emph{numeric} Within Region Connectivity Estimate Range value: 0.048329, ..., 2.0323
\item \strong{NFQ} \emph{numeric} Network Failure Quotient, calculated Range value: -0.56048, ..., 2.0035
\item \strong{SliceTiming} \emph{character} Slice Timing; an underscore-separated list of time (in seconds) for the acquired image slices in the first volume.  
\item \strong{ScannerManufacturer} \emph{character} DICOM Scanner Manufacturer Unknown, GE, Philips, Siemens, Other
\item \strong{ScannerModel} \emph{character} DICOM Scanner Model  
\item \strong{SoftwareVersion} \emph{character} DICOM Software Version; may be extracted from vendor private tags  
\item \strong{MagneticFieldStrength} \emph{numeric} Scanner Field Strength <blank>, 1.5, 3, 7
\item \strong{Acceleration} \emph{character}   , with options: Ultrafast, Accelerated, Unaccelerated
\item \strong{AcquisitionType} \emph{numeric}   Range value: 2, ..., 2
\item \strong{AcquisitionPlane} \emph{character}   , with options: Axial
\item \strong{NumberVolumes} \emph{numeric}   Range value: 97, ..., 482
\item \strong{SlicesPerVolume} \emph{numeric}   Range value: 44, ..., 64
\item \strong{SliceThickness} \emph{numeric}   Range value: 2.5, ..., 3.4
\item \strong{ReceiveCoilName} \emph{character} Receive Coil Name  
\item \strong{StudyInstanceUID} \emph{character} DICOM StudyInstanceUID  
\item \strong{SeriesInstanceUID} \emph{character} DICOM SeriesInstanceUID  
\item \strong{LONIStudy} \emph{numeric} LONI Study ID Range value: 242527, ..., 363480
\item \strong{LONISeries} \emph{numeric} LONI Series ID Range value: 5162602, ..., 6021855
\item \strong{LONIImage} \emph{numeric} LONI Image ID Range value: 10248886, ..., 11329733
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
