% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRIQC}
\alias{MRIQC}
\title{MAYO ADIR LAB MRI Quality}
\format{
A data frame with 90250 observations and 26 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRIQC)
}
\description{
MAYO ADIR LAB MRI Quality data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{image_id} \emph{numeric} LONI Image ID Range value: 7024, ..., 11475374
\item \strong{ParticipantID} \emph{character} ADNI Participant ID; SSS_S_PPPP or SSS_S_PPPPP where SSS is the SITE ID and PPPP or PPPPP is the ParticipantID  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{StudyDate} \emph{ } Study Date YYYYMMDD
\item \strong{SeriesTime} \emph{ } Time Series Started on the Scanner HHMMSS ; <blank> Unknown
\item \strong{SeriesNumber} \emph{numeric} DICOM Series Number Range value: 1, ..., 1101015
\item \strong{SeriesType} \emph{character} Series Acquisition Type  
\item \strong{MRIProtocolPhase} \emph{character} MRI Protocol Phase ADNI1 ; ADNIGO/ADNI2 ; ADNI3 ; ADNI4 ; UNKNOWN
\item \strong{SeriesDescription} \emph{character} DICOM SeriesDescription  
\item \strong{ScannerManufacturer} \emph{character} DICOM Scanner Manufacturer Unknown ; GE ; Philips ; Siemens ; Other
\item \strong{ScannerModel} \emph{character} DICOM Scanner Model  
\item \strong{SoftwareVersion} \emph{character} DICOM Software Version; may be extracted from vendor private tags  
\item \strong{MagneticFieldStrength} \emph{numeric} Scanner Field Strength <blank> ; 1.5 ; 3 ; 7
\item \strong{Acceleration} \emph{character} Acceleration Unaccelerated ; Accelerated ; Ultrafast ; Undetermined
\item \strong{AcquisitionType} \emph{character} Acquisition Type 2D ; 3D ; Unknown
\item \strong{AcquisitionPlane} \emph{character} Acquisition Plane Axial ; Sagittal ; Coronal ; NA ; Unknown
\item \strong{NumberVolumes} \emph{numeric} Number of Volumes Range value: 0, ..., 1952
\item \strong{SlicesPerVolume} \emph{numeric} Slices per Volume Range value: 0, ..., 9599
\item \strong{SliceThickness} \emph{numeric} Slice Thickness Range value: 0.9, ..., 24
\item \strong{ReceiveCoilName} \emph{character} Receive Coil Name  
\item \strong{StudyInstanceUID} \emph{character} DICOM StudyInstanceUID  
\item \strong{SeriesInstanceUID} \emph{character} DICOM SeriesInstanceUID  
\item \strong{LONIStudy} \emph{numeric} LONI Study ID Range value: 1945, ..., 377773
\item \strong{LONISeries} \emph{numeric} LONI Series ID Range value: 8722, ..., 6118226
\item \strong{LONIImage} \emph{numeric} LONI Image ID Range value: 7024, ..., 11475374
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
