% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRIQSM}
\alias{MRIQSM}
\title{MAYO ADIR LAB MRI Quantitative Susceptibility Mapping}
\format{
A data frame with 23778 observations and 39 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRIQSM)
}
\description{
MAYO ADIR LAB MRI Quantitative Susceptibility Mapping data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI3, ADNI2, ADNI1, ADNIGO
\item \strong{ParticipantID} \emph{character} ADNI Participant ID; SSS_S_PPPP or SSS_S_PPPPP where SSS is the SITE ID and PPPP or PPPPP is the ParticipantID  
\item \strong{RID} \emph{numeric}   Range value: 21, ..., 10849
\item \strong{VISCODE} \emph{character}   , with options: 4_sc, 4_init, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character}    
\item \strong{StudyDate} \emph{ } Study Date YYYYMMDD
\item \strong{SeriesTime} \emph{ } Time Series Started on the Scanner HHMMSS ; <blank> Unknown
\item \strong{SeriesNumber} \emph{numeric} DICOM Series Number Range value: 4, ..., 48
\item \strong{SeriesType} \emph{character} Series Acquisition Type , with options: MET2starw
\item \strong{SiteID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{PostProcessQC} \emph{numeric} Post Processing Series Quality Assessment (-1) Not assessed ; (1) Pass ; (4) Fail
\item \strong{Atlas} \emph{character} ADIR Lab Name for the Atlas Used for Analysis Example: ADIR122 refers to the 122 ROI atlas
\item \strong{Sharpening} \emph{character} Sharpening; tissue types included WM, GM, GMWM
\item \strong{RegionNumber} \emph{numeric} Region Number Range value: 2, ..., 49
\item \strong{RegionName} \emph{character} Region Name  
\item \strong{RegionSide} \emph{character} Region Side (L) Left ; (R) Right
\item \strong{Median} \emph{numeric} Median of the voxel values in the given image in the given ROI Range value: -0.0375901, ..., 0.172673
\item \strong{TotalVoxels} \emph{numeric} Total number of Voxels in the given image  in the given ROI Range value: 60, ..., 111829
\item \strong{VoxelsIgnoredZero} \emph{numeric} Number of Voxels Ignored due to having a Value of Zero Range value: 0, ..., 12111
\item \strong{VoxelsIgnoredMask} \emph{numeric} Number of Voxels Ignored due to being Outside Chosen Sharpening Mask Range value: 0, ..., 88011
\item \strong{Sum} \emph{numeric} Sum of the voxel values in the given image in the given ROI Range value: -1066.75, ..., 772.445
\item \strong{VoxelVolume} \emph{numeric} Volume of a single voxel in the given image Range value: 1, ..., 1
\item \strong{ScannerManufacturer} \emph{character} DICOM Scanner Manufacturer Unknown, GE, Philips, Siemens, Other
\item \strong{ScannerModel} \emph{character} DICOM Scanner Model , with options: MAGNETOM Vida, Prisma_fit, Prisma, Skyra, MAGNETOM Skyra, MAGNETOM Prisma Fit, MAGNETOM Prisma, MAGNETOM Skyra fit, MAGNETOM Cima.X
\item \strong{SoftwareVersion} \emph{character} DICOM Software Version; may be extracted from vendor private tags , with options: syngo MR XA60, syngo MR E11, syngo MR XA30, syngo MR XA61
\item \strong{MagneticFieldStrength} \emph{numeric} Scanner Field Strength <blank>, 1.5, 3, 7
\item \strong{Acceleration} \emph{character}   , with options: Accelerated
\item \strong{AcquisitionType} \emph{numeric}   Range value: 3, ..., 3
\item \strong{AcquisitionPlane} \emph{character}   , with options: Axial
\item \strong{NumberVolumes} \emph{numeric}   Range value: 5, ..., 5
\item \strong{SlicesPerVolume} \emph{numeric}   Range value: 88, ..., 104
\item \strong{SliceThickness} \emph{numeric}   Range value: 1.799999, ..., 1.8
\item \strong{ReceiveCoilName} \emph{character} Receive Coil Name , with options: HeadNeck_20_TCS, NOT SET, Head_32, HeadNeck_20, HeadNeck_64, HeadNeck_16
\item \strong{StudyInstanceUID} \emph{character} DICOM StudyInstanceUID  
\item \strong{SeriesInstanceUID} \emph{character} DICOM SeriesInstanceUID  
\item \strong{LONIStudy} \emph{numeric} LONI Study ID Range value: 242527, ..., 363432
\item \strong{LONISeries} \emph{numeric} LONI Series ID Range value: 5162614, ..., 6019064
\item \strong{LONIImage} \emph{numeric} LONI Image ID Range value: 10248902, ..., 11326165
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
