% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRIREAD}
\alias{MRIREAD}
\title{MRI Clinical Read}
\format{
A data frame with 4106 observations and 16 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRIREAD)
}
\description{
MRI Clinical Read data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{ID} \emph{numeric}   Range value: 6, ..., 8254
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1435
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, f, bl, m06, m12, uns1, m18, m24, m36, m48
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{MRREAD} \emph{numeric} Was the 1.5T MRI done? 1=Yes ; 0=No
\item \strong{MRUPLD} \emph{character} Upload 1.5T MRI Clinical Read:  
\item \strong{MRCOMPAT} \emph{numeric} Is the MRI compatible with the Inclusion/Exclusion Criteria? 1=Yes ; 0=No
\item \strong{MRCOMSPE} \emph{character} If No, explain:  
\item \strong{MR3TREAD} \emph{numeric} Was a clinical read of the 3T MRI done? 1=Yes ; 0=No
\item \strong{MR3TUPLD} \emph{character} Upload 3T MRI Clinical Read :  
\item \strong{MRSUITABL} \emph{numeric} Is the participant clinically suitable to remain in the study? 1=Yes ; 0=No
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
