% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRISERIAL}
\alias{MRISERIAL}
\title{MRI Serial}
\format{
A data frame with 3204 observations and 23 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(MRISERIAL)
}
\description{
MRI Serial data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 6416
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1430
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code , with options: m06, m12, sc, nv, m18, uns1, m24, m36, m48
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 13
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{FLDSTRENG} \emph{numeric} Field Strength Range value: 1.5, ..., 3
\item \strong{COILTYPE} \emph{character} Coil Type  
\item \strong{SCAN} \emph{character} Scan 1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{LONISID} \emph{numeric} LONI Study ID Range value: 3151, ..., 29427
\item \strong{LONIUID} \emph{character} LONI Unique Series ID  
\item \strong{BLLONIUID} \emph{character} LONI Unique Series ID for the baseline scan.  
\item \strong{BLSDATE} \emph{numeric} Baseline scan date. Range value: 20050826, ..., 20080815
\item \strong{SAMESUBJ} \emph{character}   1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{REGISTRAT} \emph{character}   1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{GROSSINHOM} \emph{character}   1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{COMMENTS} \emph{character}    
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{imaging}
