% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEUROBAT}
\alias{NEUROBAT}
\title{Neuropsychological Battery}
\format{
A data frame with 17622 observations and 84 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(NEUROBAT)
}
\description{
Neuropsychological Battery data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment DATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{SOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call/Video Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKCIRC} \emph{character} 1. Approximately circular face 
\itemize{
\item \emph{ADNI1, ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKSYM} \emph{character} 2. Symmetry of number placement 
\itemize{
\item \emph{ADNI1, ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKNUM} \emph{character} 3. Correctness of numbers 
\itemize{
\item \emph{ADNI1, ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKHAND} \emph{character} 4. Presence of the two hands 
\itemize{
\item \emph{ADNI1, ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKTIME} \emph{character} 5. Presence of the two hands, set to ten after eleven 
\itemize{
\item \emph{ADNI1, ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLOCKSCOR} \emph{numeric} Total Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..5
\item \emph{ADNI3, ADNI4} 
}
\item \strong{COPYCIRC} \emph{character} 1. Approximately circular face 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{COPYSYM} \emph{character} 2. Symmetry of number placement 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{COPYNUM} \emph{character} 3. Correctness of numbers 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{COPYHAND} \emph{character} 4. Presence of the two hands 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{COPYTIME} \emph{character} 5. Presence of the two hands, set to ten after eleven 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Correct ; 0=Incorrect
\item \emph{ADNI3, ADNI4:} 0=Incorrect ; 1=Correct
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{COPYSCOR} \emph{numeric} Total Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..5
\item \emph{ADNI3, ADNI4} 
}
\item \strong{LMSTORY} \emph{character} Which Logical Memory story was used? 1=Anna Thompson ; 2=Greg Fortune
\emph{Decoded Value: }\strong{Yes}
\item \strong{LIMMTOTAL} \emph{numeric} Total Number of Story Units Recalled 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..25
\item \emph{ADNI3, ADNI4} 
}
\item \strong{LIMMEND} \emph{numeric} Time Ended (24-hour clock) Range value: 33, ..., 2041
\item \strong{AVTOT1} \emph{numeric} Trial 1 Total 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVERR1} \emph{numeric} Total Intrusions Range value: 0, ..., 13
\item \strong{AVTOT2} \emph{numeric} Trial 2 Total 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVERR2} \emph{numeric} Total Intrusions Range value: 0, ..., 12
\item \strong{AVTOT3} \emph{numeric} Trial 3 Total 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVERR3} \emph{numeric} Total Intrusions Range value: 0, ..., 14
\item \strong{AVTOT4} \emph{numeric} Trial 4 Total 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVERR4} \emph{numeric} Total Intrusions Range value: 0, ..., 14
\item \strong{AVTOT5} \emph{numeric} Trial 5 Total 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVERR5} \emph{numeric} Total Intrusions Range value: 0, ..., 15
\item \strong{AVTOT6} \emph{numeric} Trial 6 Total 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVERR6} \emph{numeric} Total Intrusions Range value: 0, ..., 13
\item \strong{AVTOTB} \emph{numeric} List B Total 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVERRB} \emph{numeric} Total Intrusions Range value: 0, ..., 12
\item \strong{AVENDED} \emph{numeric} AVLT Time Trial B Ended Range value: 100, ..., 1925
\item \strong{DSPANFOR} \emph{numeric} Forward: Total Correct 0..12
\item \strong{DSPANFLTH} \emph{numeric} Forward: Length 3..8
\item \strong{DSPANBAC} \emph{numeric} Backward: Total Correct 0..12
\item \strong{DSPANBLTH} \emph{numeric} Backward: Length 2..7
\item \strong{CATANIMSC} \emph{numeric} Category Fluency (Animals) - Total Correct 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..77
\item \emph{ADNI3, ADNI4} 
}
\item \strong{CATANPERS} \emph{numeric} Category Fluency (Animals) - Perseverations 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..77
\item \emph{ADNI3, ADNI4} 
}
\item \strong{CATANINTR} \emph{numeric} Category Fluency (Animals) - Intrusions 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..77
\item \emph{ADNI3, ADNI4} 
}
\item \strong{CATVEGESC} \emph{numeric} Category Fluency (Vegetables) - Total Correct 0..77
\item \strong{CATVGPERS} \emph{numeric} Category Fluency (Vegetables) - Perseverations 0..77
\item \strong{CATVGINTR} \emph{numeric} Category Fluency (Vegetables) - Intrusions 0..77
\item \strong{TRAASCOR} \emph{numeric} Part A - Time to Complete 
\itemize{
\item \emph{ADNI1, ADNI2, ADNI3, ADNI4} 
\item \emph{ADNIGO:} 1..150
}
\item \strong{TRAAERRCOM} \emph{numeric} Errors of Commission Range value: 0, ..., 24
\item \strong{TRAAERROM} \emph{numeric} Errors of Omission 
\itemize{
\item \emph{ADNI1, ADNI3, ADNI4} 
\item \emph{ADNIGO:} 0..24
\item \emph{ADNI2:} 0..77
}
\item \strong{TRABSCOR} \emph{numeric} Part B - Time to complete 
\itemize{
\item \emph{ADNI1, ADNI2, ADNI3, ADNI4} 
\item \emph{ADNIGO:} 1..300
}
\item \strong{TRABERRCOM} \emph{numeric} Errors of Commission Range value: 0, ..., 25
\item \strong{TRABERROM} \emph{numeric} Errors of Omission 
\itemize{
\item \emph{ADNI1, ADNI2, ADNI3, ADNI4} 
\item \emph{ADNIGO:} 0..24
}
\item \strong{DIGITSCOR} \emph{numeric} Total Correct Range value: 0, ..., 83
\item \strong{LDELBEGIN} \emph{numeric} Time Began (24-hour clock) Range value: 8, ..., 2112
\item \strong{LDELTOTAL} \emph{numeric} Total Number of Story Units Recalled 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..25
\item \emph{ADNI3, ADNI4} 
}
\item \strong{LDELCUE} \emph{character} Reminder given? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{BNTND} \emph{character} Check here if: 1=Boston Naming Not Done - Spanish speaking
\emph{Decoded Value: }\strong{Yes}
\item \strong{BNTSPONT} \emph{numeric} 1. Total correct without a cue 0..30
\item \strong{BNTSTIM} \emph{numeric} 2. Total semantic cues given 0..30
\item \strong{BNTCSTIM} \emph{numeric} 3. Total correct with a semantic cue 0..30
\item \strong{BNTPHON} \emph{numeric} 4. Total phonemic cues given 0..30
\item \strong{BNTCPHON} \emph{numeric} 5. Number of correct responses following a phonemic cue 0..30
\item \strong{BNTTOTAL} \emph{numeric} Total Number Correct (1+3) 0..30
\item \strong{AVDELBEGAN} \emph{numeric} Time Began (24-hour clock) Range value: 30, ..., 1910
\item \strong{AVDEL30MIN} \emph{numeric} 30 Minute Delay Total 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVDELERR1} \emph{numeric} Total Intrusions Range value: 0, ..., 30
\item \strong{AVDELTOT} \emph{numeric} Recognition Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..15
\item \emph{ADNI3, ADNI4} 
}
\item \strong{AVDELERR2} \emph{numeric} Total Intrusions Range value: 0, ..., 17
\item \strong{ANARTERR} \emph{numeric} ANART Total Score (Total # of errors) 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..50
\item \emph{ADNI3, ADNI4} 
}
\item \strong{ANARTND} \emph{character} Check here if: 1=ANART Not Done - Spanish speaking
\emph{Decoded Value: }\strong{Yes}
\item \strong{ANART} \emph{character} Was ANART conducted? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{MINTSEMCUE} \emph{numeric} MINT Total Correct - with Semantic Cue Range value: 0, ..., 32
\item \strong{MINTTOTAL} \emph{numeric} MINT Total Correct (Uncued + Correct with Semantic cue) Range value: 0, ..., 32
\item \strong{MINTUNCUED} \emph{numeric} MINT Total Uncued Correct Range value: 0, ..., 32
\item \strong{ID} \emph{numeric} Record ID Range value: 6, ..., 187101
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: sc, annual, bl
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
