% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEUROEXM}
\alias{NEUROEXM}
\title{Neurological Exam}
\format{
A data frame with 5752 observations and 29 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(NEUROEXM)
}
\description{
Neurological Exam data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{NXVISUAL} \emph{character} 1. Significant Visual Impairment 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXAUDITO} \emph{character} 2. Significant Auditory Impairment 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXTREMOR} \emph{character} 3. Tremor 1=Absent ; 2=Present
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXCONSCI} \emph{character} 4. Level of Consciousness 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXNERVE} \emph{character} 5. Cranial Nerves 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXMOTOR} \emph{character} 6. Motor Strength 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXFINGER} \emph{character} 7a. Cerebellar - Finger to Nose 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXHEEL} \emph{character} 7b. Cerebellar - Heel to Shin 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXSENSOR} \emph{character} 8. Sensory 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXTENDON} \emph{character} 9. Deep Tendon Reflexes 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXPLANTA} \emph{character} 10. Plantar Reflexes 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXGAIT} \emph{character} 11. Gait 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXOTHER} \emph{character} 12. Other 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{NXABNORM} \emph{character} 14. Based on Neurological Examination, clinician must check appropriate box below: 1=Findings consistent with eligibility for study ; 2=Participant is not eligible for study
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 6, ..., 186320
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{medical_history}
