% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEUROPATH}
\alias{NEUROPATH}
\title{NACC Neuropathology Data Form (Version 11)}
\format{
A data frame with 121 observations and 165 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(NEUROPATH)
}
\description{
NACC Neuropathology Data Form (Version 11) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI1, ADNI3, ADNIGO
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 8, ..., 6891
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: AUT
\item \strong{NPFORMMO} \emph{numeric}   Range value: 1, ..., 12
\item \strong{NPFORMDY} \emph{numeric} Date Form Completed - Day Range value: 1, ..., 31
\item \strong{NPFORMYR} \emph{numeric} Date Form Completed - Year Range value: 2008, ..., 2025
\item \strong{NPFORMDATE_DRVD} \emph{ } 2. Date Form Completed (Derived from NPFORMMO,NPFORMDY,NPFORMYR) Date: YYY-MM-DD
\item \strong{NPDAGE} \emph{numeric} Age at death (years) Range value: 59, ..., 97
\item \strong{NPDODYR} \emph{numeric} Date of Death - Year Range value: 2007, ..., 2023
\item \strong{NPPMIH} \emph{numeric} Postmortem interval (PMI) (hours) Range value: 2, ..., 9999
\item \strong{NPFIX} \emph{numeric} 8. Fixative 1=1 Formalin ; 2=2 Paraformaldehyde ; 7=7 Other (Specify)
\item \strong{NPFIXX} \emph{character} Specify: , with options: First Paraform, then Formalin
\item \strong{NPWBRWT} \emph{numeric} a. Whole brain weight (if half brain, multiply weight by two) (9999=unknown) Range value: 850, ..., 9999
\item \strong{NPWBRF} \emph{numeric} b. Does the value in Question 9a represent fresh or fixed weight? (Check One) 1=1 Fresh ; 2=2 Fixed ; 8=8 Not applicable
\item \strong{NPGRCCA} \emph{numeric} c1. Cerebral cortex atrophy 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPGRLA} \emph{numeric} c2. Lobar atrophy (significant frontal and/or temporal atrophy) 0=0 None ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPGRHA} \emph{numeric} c3. Hippocampus atrophy 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPGRSNH} \emph{numeric} c4. Substantia nigra hypopigmentation 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPGRLCH} \emph{numeric} c5. L. ceruleus hypopigmentation 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPAVAS} \emph{numeric} c6. Atherosclerosis (of the circle of Willis) 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPTAN} \emph{numeric} a. Tau antibody (Check One) 1=1 Non-phospho specific ; 2=2 PHF1 ; 3=3 CP13 ; 4=4 AT8 ; 7=7 Other (Specify) ; 8=8 Not assessed
\item \strong{NPTANX} \emph{logical} Specify: A Boolean value: TRUE/ FALSE
\item \strong{NPABAN} \emph{numeric} b. Amyloid beta antibody (Check One) 1=1 4G8 ; 2=2 10D5 ; 7=7 Other (Specify) ; 8=8 Not assessed
\item \strong{NPABANX} \emph{logical} Specify: A Boolean value: TRUE/ FALSE
\item \strong{NPASAN} \emph{numeric} c. Alpha synuclein antibody 1=1 Non-phospho specific (e.g. LB509) ; 2=2 Phospho-specific (e.g. pSYN#64) ; 7=7 Other (Specify) ; 8=8 Not assessed
\item \strong{NPASANX} \emph{logical} Specify: A Boolean value: TRUE/ FALSE
\item \strong{NPTDPAN} \emph{numeric} d. TDP-43 antibody (Check One) 1=1 Non-phospho specific ; 2=2 Phospho-specific ; 7=7 Other (Specify) ; 8=8 Not assessed
\item \strong{NPTDPANX} \emph{logical} Specify: A Boolean value: TRUE/ FALSE
\item \strong{NPHISMB} \emph{numeric} e1. Modified Bielschowsky 0=0 No ; 1=1 Yes
\item \strong{NPHISG} \emph{numeric} e2. Gallyas 0=0 No ; 1=1 Yes
\item \strong{NPHISSS} \emph{numeric} e3. Other Silver stain 0=0 No ; 1=1 Yes
\item \strong{NPHIST} \emph{numeric} e4. Thioflavin 0=0 No ; 1=1 Yes
\item \strong{NPHISO} \emph{numeric} e5. Other 0=0 No ; 1=1 Yes
\item \strong{NPHISOX} \emph{character} Specify: , with options: Double IHC, AT8 + H31L21, Neurofilament IHC, Dual IHC Aß (H31L21) + Tau (AT8), LFB/PAS, p62, Neurofilament IHC
\item \strong{NPTHAL} \emph{numeric} a. Thal phase for amyloid plaques by immunohisto-chemistry (IHC) (A score - Check One) 0=0 Phase 0 (A0) ; 1=1 Phase 1 (A1) ; 2=2 Phase 2 (A1) ; 3=3 Phase 3 (A2) ; 4=4 Phase 4 (A3) ; 5=5 Phase 5 (A3) ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPBRAAK} \emph{numeric} b. Braak stage for neurofibrillary degeneration (B score - Check One) 0=0 Sage 0: AD-type neurofibrillary degeneration not present (B0) ; 1=1 Stage I (B1) ; 2=2 Stage II (B1) ; 3=3 Stage III (B2) ; 4=4 Stage IV (B2) ; 5=5 Stage V (B3) ; 6=6 Stage VI (B3) ; 7=7 The presence of a tauopathy (other than aging/AD) precludes Braak staging ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPNEUR} \emph{numeric} c. CERAD score for density of neocortical neuritic plaque (plaques with argrophilic dystrophic neurities, with or without dense amyloid cores). Score without respect to age or diagnosis. (C score - Check One) 0=0 No neuritic plaques (C0) ; 1=1 Sparse neuritic plaques (C1) ; 2=2 Moderate neuritic plaques (C2) ; 3=3 Frequent neuritic plaques (C3) ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPADNC} \emph{numeric} d. NIA-AA Alzheimer's disease neuropathologic change (ADNC) (Check One) 0=0 Not AD ; 1=1 Low ADNC ; 2=2 Intermediate ADNC ; 3=3 High ADNC ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPDIFF} \emph{numeric} e1. CERAD semi-quantitative score for diffuse plaques (plaques with non-compact amyloid and no apparent dystrophic neurites). Score from the neocortical field with the highest plaque density and without respect to age or diagnosis. (Check One) 0=0 No diffuse plaques ; 1=1 Sparse diffuse plaques ; 2=2 Moderate diffuse plaques ; 3=3 Frequent diffuse plaques ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPAMY} \emph{numeric} e2. Cerebral amyloid angiopathy (Check One) 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPINF} \emph{numeric} a. Old infarcts observed grossly, including lacunes? (Check One) 0=0 No (SKIP TO QUESTION 12b) ; 1=1 Yes (COMPLETE QUESTIONS 12a1-12a4) ; 8=8 Not assessed (SKIP TO QUESTION 12b) ; 9=9 Missing/unknown (SKIP TO QUESTION 12b)
\item \strong{NPINF1A} \emph{character} Number of infarcts -- Cerebral cortex , with options: ., 0, 1, 5
\item \strong{NPINF1B} \emph{character} Size of largest (greatest dimension in cm) , with options: ., 4, 4.5, 1
\item \strong{NPINF1D} \emph{character} Size of next (greatest dimension in cm) , with options: ., 4
\item \strong{NPINF1F} \emph{character} Size of next (greatest dimension in cm) , with options: ., 3
\item \strong{NPINF2A} \emph{character} Number of infarcts -- Subcortical cerebral white matter and peri-ventricular white matter , with options: ., 1, 0
\item \strong{NPINF2B} \emph{character} Size of largest (greatest dimension in cm) , with options: ., 1.2, 0.3
\item \strong{NPINF2D} \emph{logical} Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE
\item \strong{NPINF2F} \emph{logical} Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE
\item \strong{NPINF3A} \emph{character} Number of infarcts -- Deep cerebral gray matter or internal capsule , with options: ., 0, 1, 2
\item \strong{NPINF3B} \emph{character} Size of largest (greatest dimension in cm) , with options: ., 2, 1.5, 1, 0.3
\item \strong{NPINF3D} \emph{character} Size of next (greatest dimension in cm) , with options: ., 0.5
\item \strong{NPINF3F} \emph{logical} Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE
\item \strong{NPINF4A} \emph{character} Number of infarcts -- Brainstem or cerebellum , with options: ., 0, 1
\item \strong{NPINF4B} \emph{character} Size of largest (greatest dimension in cm) , with options: ., 0.5, 0.6
\item \strong{NPINF4D} \emph{logical} Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE
\item \strong{NPINF4F} \emph{logical} Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE
\item \strong{NPHEMO} \emph{numeric} b. Were single or multiple old hemorrhages observed grossly? 0=0 No (SKIP TO QUESTION 12c) ; 1=1 Yes (COMPLETE QUESTIONS 12b1-12b3) ; 8=8 Not assessed (SKIP TO QUESTION 12c) ; 9=9 Missing/unknown (SKIP TO QUESTION 12c)
\item \strong{NPHEMO1} \emph{character} b1. Subdural or epidural hemorrhage 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPHEMO2} \emph{character} b2. Primary parenchymal hemorrhage Include those >5mm. If ?5mm, include as microbleed; see Question 12d. 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPHEMO3} \emph{character} b3. Secondary parenchymal hemorrhage (e.g., tumor, vascular malformation) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOLD} \emph{numeric} c. Old microinfarcts (not observed grossly)? (Check One) 0=0 No (SKIP TO QUESTION 12d) ; 1=1 Yes (COMPLETE QUESTIONS 12c1-12c4) ; 8=8 Not assessed (SKIP TO QUESTION 12d) ; 9=9 Missing/unknown (SKIP TO QUESTION 12d)
\item \strong{NPOLD1} \emph{character} c1. Number in screening sections of cerebral cortex (gray matter of cerebral cortex) 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOLD2} \emph{character} c2. Number in screening sections of subcortical white matter and periventricular white matter 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOLD3} \emph{character} c3. Number in screening sections of subcortical gray matter 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOLD4} \emph{character} c4. Number in brainstem and cerebellum 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOLDD} \emph{numeric} d. Old cerebral microbleeds? (Check One) Include old hemorrahages that are greater than or equal to 5mn. 0=0 No (SKIP TO QUESTION 12e) ; 1=1 Yes (COMPLETE QUESTIONS 12d1-12d4) ; 8=8 Not assessed (SKIP TO QUESTION 12e) ; 9=9 Missing/unknown (SKIP TO QUESTION 12e)
\item \strong{NPOLDD1} \emph{character} d1. Number in screening sections of cerebral cortex 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOLDD2} \emph{character} d2. Number in screening sections of subcortical white matter and periventricular white matter 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOLDD3} \emph{character} d3. Number in screening sections of subcortical gray matter 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOLDD4} \emph{character} d4. Number in brainstem and cerebellum 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPARTER} \emph{numeric} e. Arteriolosclerosis? (Check One) (Assessed in subcortical white or gray matter) 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPWMR} \emph{numeric} f. White matter rarefaction? (Check One) (H&E or myelin stain may be used) 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH} \emph{numeric} g. Other pathologic changes related to ischemic or vascular disease not previously specified? 0=0 No (SKIP TO QUESTION 13) ; 1=1 Yes (COMPLETE QUESTIONS 12g1-12g12) ; 8=8 Not assessed (SKIP TO QUESTION 13) ; 9=9 Missing/unknown (SKIP TO QUESTION 13)
\item \strong{NPNEC} \emph{character} g1. Laminar necrosis 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH2} \emph{character} g2. Acute neuronal necrosis 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH3} \emph{character} g3. Acute/subacute gross infarcts 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH4} \emph{character} g4. Acute/subacute microinfarcts 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH5} \emph{character} g5. Acute/subacute gross hemorrhage 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH6} \emph{character} g6. Acute/subacute microhemorrhage 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH7} \emph{character} g7. Vascular malformation of any type 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH8} \emph{character} g8. Aneurysm of any type 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH9} \emph{character} g9. Vasculitis of any type 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH10} \emph{character} g10. CADASIL 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATH11} \emph{character} g11. Mineralization of blood vessels 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPATHO} \emph{character} g12. Other 0=0 No ; 1=1 Yes
\item \strong{NPPATHOX} \emph{character} Specify: , with options: Mural/Luminal protein deposits, Coxiella Burnetii, Remote ischemic foci in CA1, subacute corticospinal tract degeneration, FOCAL SUBACUTE THROMBOSIS, R PRECENTRAL GYRUS
\item \strong{NPLBOD} \emph{numeric} Is there evidence of Lewy body pathology? (Check One) 0=0 No ; 1=1 Brainstem predominant ; 2=2 Limbic (transitional) ; 3=3 Neocortical (diffuse) ; 4=4 Amygdala predominant ; 5=5 Olfactory bulb ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPNLOSS} \emph{numeric} Neuron loss in the substantia nigra 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 severe ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPHIPSCL} \emph{numeric} Hippocampal sclerosis (CA1 and/or subiculum) 0=0 None ; 1=1 Unilateral ; 2=2 Bilateral ; 3=3 Present but laterality not assessed ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPTDPA} \emph{numeric} a. Spinal cord 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPTDPB} \emph{numeric} b. Amygdala 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPTDPC} \emph{numeric} c. Hippocampus 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPTDPD} \emph{numeric} d. Entorhinal/inferior temporal cortex 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPTDPE} \emph{numeric} e. Neocortex 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDTAU} \emph{numeric} a. FTLD with tau pathology (FTLD-tau) or other tauopathy (Check One) 0=0 No (SKIP TO QUESTION 17c) ; 1=1 Yes (COMPLETE QUESTIONS 17b1-17b10) ; 8=8 Not assessed (SKIP TO QUESTION 17c) ; 9=9 Missing/unknown (SKIP TO QUESTION 17c)
\item \strong{NPPICK} \emph{character} b1. FTLD-tau (PiD) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDT2} \emph{character} b2. Other 3R tauopathy (Includes MAPT mutation tauopathy) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPCORT} \emph{character} b3. FTLD-tau (CBD) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPROG} \emph{character} b4. FTLD-tau (PSP) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDT5} \emph{character} b5. Argyrophilic grains 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDT6} \emph{character} b6. Other 4R tauopathy (Includes sporadic multiple systems tauopathy, globular glial tauopathy, MAPT mutation tauopathy) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDT7} \emph{character} b7. Chronic traumatic encephalopathy 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDT8} \emph{character} b8. Amyotrophic lateral sclerosis (ALS)/Parkinsonism-dementia complex of Guam 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDT9} \emph{character} b9. Tangle dominant disease 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDT10} \emph{character} b10. Other 3R + 4F tauopathy (Includes unclassifiable, focal, glial only, MAPT mutation tauopathy, NOS) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPFTDTDP} \emph{numeric} c. FTLD with TDP-43 pathology (FTLD-TDP)? (Check One) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPALSMND} \emph{numeric} d. ALS/motor neuron disease (MND) present? (Check One) 0=0 No ; 1=1 Yes, with TDP-43 inclusions in motor neurons ; 2=2 Yes, with FUS inclusions in motor neurons ; 3=3 Yes, with SOD1 inclusions in motor neurons ; 4=4 Yes, with other inclusions ; 5=5 Yes, with no specific inclusions ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOFTD} \emph{numeric} e. Other FTLD? (Check One) 0=0 No (SKIP TO QUESTION 18a) ; 1=1 Yes (COMPLETE QUESTIONS 17f1 - 17f5) ; 8=8 Not assessed (SKIP TO QUESTION 18a) ; 9=9 Missing/unknown (SKIP TO QUESTION 18a)
\item \strong{NPOFTD1} \emph{logical} f1. Atypical FTLD-U (aFTLD-U) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOFTD2} \emph{logical} f2. NIFID (neuronal intermediate filament inclusions disease) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOFTD3} \emph{logical} f3. BIBD (basophilic inclusion body disease) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOFTD4} \emph{logical} f4. FTLD-UPS (ubiquitin-proteasome system [ubiquitin or p62 positive, tau/TDP-43/FUS negative inclusions]) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPOFTD5} \emph{logical} f5. FTLD-NOS (includes dementia lacking distinctive histology (DLDH) and FTLD with no inclusions (FTLD-NI) detected by tau, TDP-43, or ubiquitin/p62 IHC) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXA} \emph{numeric} a. Pigment-spheroid degeneration/NBIA 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXB} \emph{numeric} b. Multiple system atrophy 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXC} \emph{numeric} c. Prion disease 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXD} \emph{numeric} d. Trinucleotide disease (Huntington disease, SCA, other) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXE} \emph{numeric} e. Malformation of cortical development 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXF} \emph{numeric} f. Metabolic/storage disorder of any type 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXG} \emph{numeric} g. WM disease, leukodystrophy 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXH} \emph{numeric} h. WM disease, multiple sclerosis or other demyelinating disease 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXI} \emph{numeric} i. Contusion/traumatic brain injury of any type, acute 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXJ} \emph{numeric} j. Contusion/traumatic brain injury of any type, chronic 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXK} \emph{numeric} k. Neoplasm, primary 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXL} \emph{numeric} l. Neoplasm, metastatic 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXM} \emph{numeric} m. Infectious process of any type (encephalitis, abscess, etc.) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXN} \emph{numeric} n. Herniation, any site 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXO} \emph{numeric} o. Trisomy 21/Down syndrome 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXP} \emph{numeric} p. AD-related genes (dominantly inherited); do not include APOE or other polymorphisms or genetic risk factors. 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXQ} \emph{numeric} q. FTLD-related genes (dominantly inherited); do not include polymorphisms or genetic risk factors. 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown
\item \strong{NPPDXR} \emph{numeric} r. Other (Specify) 0=0 No ; 1=1 Yes
\item \strong{NPPDXRX} \emph{character} Specify:  
\item \strong{NPPDXS} \emph{numeric} s. Other (Specify) 0=0 No ; 1=1 Yes
\item \strong{NPPDXSX} \emph{character} Specify:  
\item \strong{NPPDXT} \emph{numeric} t. Other (Specify) 0=0 No ; 1=1 Yes
\item \strong{NPPDXTX} \emph{character} Specify: , with options: Vascular mineralization, mild, hippo, pallidum, LATE-NC, ARTAG, Moderate, Widespread, Leptomeningeal hemosiderin bilateral anterior temporal lobes, Meningioma, transitional, WHO grade 1, 0.5cm, TDP-43 Proteinopathy, Joseph's Stage 3, LATE-NC Stage 2, ARTAG, Mild, Amygdala, Basal Forebrain, Olfactory Cortex, Parietal White Matter, LBD, very mild, substantia nigra only
\item \strong{NPBNKA} \emph{numeric} a. Banked frozen brain or half brain 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPBNKB} \emph{numeric} b. Banked frozen wedge of cerebellum or other sample for future DNA prep 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPBNKC} \emph{numeric} c. Formalin- or paraformaldehyde-fixed brain 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPBNKD} \emph{numeric} d. Paraffin-embedded blocks of brain regions 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPBNKE} \emph{numeric} e. Banked postmortem CSF 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPBNKF} \emph{numeric} f. Banked postmortem blood or serum 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPBNKG} \emph{numeric} g. Banked DNA 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPFAUT} \emph{numeric} h. Full autopsy performed? 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPFAUT1} \emph{character} If full autopsy, first major finding , with options: Leukemia, Acute and Chronic Congestive Heart Failure
\item \strong{NPFAUT2} \emph{character} If full autopsy, second major finding , with options: AD, Severe Atherosclerosis - Coronary, Aorta
\item \strong{NPFAUT3} \emph{character} If full autopsy, third major finding , with options: 0, Erythroid Hyperplasia with megaloblastic changes
\item \strong{NPFAUT4} \emph{character} If full autopsy, fourth major finding , with options: 0, .
\item \strong{FORMVER} \emph{numeric}   Range value: 10, ..., 11
\item \strong{NPARTAG} \emph{character}   0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPATGSEV} \emph{character}   1=Mild ; 2=Moderate ; 3=Severe ; 8=Not assesed ; 9=Missing/unknown
\item \strong{NPATGAMY} \emph{character}   0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPATGAM1} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{NPATGAM2} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{NPATGAM3} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{NPATGAM4} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{NPATGAM5} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{NPATGFRN} \emph{character}   0=0 No ; 1=1 Yes ; 9=9 Missing/unknown
\item \strong{NPATGFR1} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{NPATGFR2} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{NPATGFR3} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{NPATGFR4} \emph{character}   0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{neuropath}
\keyword{undefined_phase}
