% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NPI}
\alias{NPI}
\title{Neuropsychiatric Inventory}
\format{
A data frame with 8225 observations and 169 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(NPI)
}
\description{
Neuropsychiatric Inventory data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI2, ADNIGO, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{NPISOURC} \emph{character} Information source 
\itemize{
\item \emph{ADNI2:} 1=Participant visit ; 2=Telephone call
\item \emph{ADNI3, ADNI4:} 1=Participant visit ; 2=Telephone call/Video Call
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SPID} \emph{numeric} Study Partner ID: Range value: 1, ..., 10
\item \strong{NPIA} \emph{character} A. Delusions 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA1} \emph{character} 1. Does  P believe that he/she is in danger - that others are planning to hurt him/her? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA2} \emph{character} 2. Does  P believe that others are stealing from him/her? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA3} \emph{character} 3. Does  P believe that his/her spouse is having an affair? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA4} \emph{character} 4. Does  P believe that unwelcome guests are living in his/her house? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA5} \emph{character} 5. Does  P believe that his/her spouse or others are not who they claim to be? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA6} \emph{character} 6. Does  P believe that his/her house is not his/her home? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA7} \emph{character} 7. Does  P believe that family members plan to abandon him/her? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA8} \emph{character} 8. Does  P believe that television or magazine figures are actually present in the home?  [Does he/she try to talk to or interact with them?] 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA9} \emph{character} 9. Does  P believe any other unusual things that I haven't asked about? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA10A} \emph{character} 10a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA10B} \emph{character} 10b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA10C} \emph{character} 10c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIATOT} \emph{numeric} A. Delusions: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIB} \emph{character} B.  Hallucinations 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB1} \emph{character} 1. Does  P describe hearing voices or act as if he/she hears voices? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB2} \emph{character} 2. Does  P talk to people who are not there? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB3} \emph{character} 3. Does  P describe seeing things not seen by others or behave as if he/she is seeing things not seen by others (people, animals, lights, etc.)? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB4} \emph{character} 4. Does  P report smelling odors not smelled by others? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB5} \emph{character} 5. Does  P describe feeling things on his/her skin or otherwise appear to be feeling things crawling or touching him/her? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB6} \emph{character} 6. Does  P describe tastes that are without any known cause? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB7} \emph{character} 7. Does  P describe any other unusual sensory experiences? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB8A} \emph{character} 8a.  Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB8B} \emph{character} 8b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB8C} \emph{character} 8c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIBTOT} \emph{numeric} B. Hallucinations: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIC} \emph{character} C: Agitation/Aggression 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC1} \emph{character} 1. Does  P get upset with those trying to care for him/her or resist activities such as bathing or changing clothes? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC2} \emph{character} 2. Is  P stubborn, having to have things his/her way? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC3} \emph{character} 3. Is  P uncooperative, resistive to help from others? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC4} \emph{character} 4. Does  P have any other behaviors that make him/her hard to handle? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC5} \emph{character} 5. Does  P shout or curse angrily? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC6} \emph{character} 6. Does  P slam doors, kick furniture, throw things? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC7} \emph{character} 7. Does  P attempt to hurt or hit others? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC8} \emph{character} 8. Does  P have any other aggressive or agitated behaviors? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC9A} \emph{character} 9a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC9B} \emph{character} 9b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC9C} \emph{character} 9c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPICTOT} \emph{numeric} C. Agitation/Aggression: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPID} \emph{character} D: Depression/Dysphoria 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID1} \emph{character} 1. Does  P have periods of tearfulness or sobbing that seem to indicate sadness? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID2} \emph{character} 2. Does  P say or act as if he/she is sad or low in spirits? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID3} \emph{character} 3. Does  P put him/herself down or say that he/she feels like a failure? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID4} \emph{character} 4. Does  P say that he/she is a bad person or deserves to be punished? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID5} \emph{character} 5. Does  P seem very discouraged or say that he/ she has not future? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID6} \emph{character} 6. Does  P say he/she is a burden to the family or that the family would be better off without him/her? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID7} \emph{character} 7. Does  P express a wish for death or talk about killing him/herself? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID8} \emph{character} 8. Does  P show any other signs of depression or sadness? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID9A} \emph{character} 9a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID9B} \emph{character} 9b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID9C} \emph{character} 9c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIDTOT} \emph{numeric} D. Depression/Dysphoria: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIE} \emph{character} E. Anxiety 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE1} \emph{character} 1. Does  P say that he/she is worried about planned events? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE2} \emph{character} 2. Does  P have periods of feeling shaky, unable to relax, or feeling excessively tense? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE3} \emph{character} 3. Does  P have periods of [or complain of] shortness of breath, gasping, or sighing for no apparent reason other than nervousness? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE4} \emph{character} 4. Does  P complain of butterflies in his/her stomach, or of racing or pounding of the heart in association with nervousness? [Symptoms not explained by ill health]? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE5} \emph{character} 5. Does  P avoid certain places or situations that make him.her more nervous such as riding in the car, meeting with friends, or being in crowds? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE6} \emph{character} 6. Does  P become nervous or upset when separated from you [or his/her caregiver]? [Does he/she cling to you to keep from being separated?] 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE7} \emph{character} 7. Does  P show any other signs of anxiety? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE8A} \emph{character} 8a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE8B} \emph{character} 8b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE8C} \emph{character} 8c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIETOT} \emph{numeric} E. Anxiety: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIF} \emph{character} F. Elation/Euphoria 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF1} \emph{character} 1. Does  P appear to feel too good or too happy, different from his/her usual self? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF2} \emph{character} 2. Does  P find humor and laugh at things that others do not find funny? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF3} \emph{character} 3. Does  P seem to have a childish sense of humor with a tendency to giggle or laugh inappropriately (such as when something unfortunate happens to others)? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF4} \emph{character} 4. Does  P tell jokes or make remarks that have little humor for others but seem funny to him/her? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF5} \emph{character} 5. Does he/she play childish pranks such as pinching or playing "keep away" for the fun of it? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF6} \emph{character} 6. Does  P "talk big" or claim to have more abilities or wealth than is true? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF7} \emph{character} 7. Does  P show any other signs of feeling too good or being too happy? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF8A} \emph{character} 8a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF8B} \emph{character} 8b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF8C} \emph{character} 8c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIFTOT} \emph{numeric} F. Elation/Euphoria: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIG} \emph{character} G. Apathy/Indifference 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG1} \emph{character} 1. Does  P seem less spontaneous and less active than usual? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG2} \emph{character} 2. Is  P less likely to initiate a conversation? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG3} \emph{character} 3. Is  P less affectionate or lacking in emotions when compared to his/her usual self? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG4} \emph{character} 4. Does  P contribute less to household chores? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG5} \emph{character} 5. Does  P seem less interested in the activities and plans of others? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG6} \emph{character} 6. Has  P lost interest in friends and family members? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG7} \emph{character} 7. Is  P less enthusiastic about his/her usual interests? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG8} \emph{character} 8. Does  P show any other signs that he/she doesn't care about doing new things? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG9A} \emph{character} 9a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG9B} \emph{character} 9b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG9C} \emph{character} 9c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIGTOT} \emph{numeric} G. Apathy/Indifference: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIH} \emph{character} H. Disinhibition 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH1} \emph{character} 1. Does  P act impulsively without appearing to consider the consequences? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH2} \emph{character} 2. Does  P talk to total strangers as if he/she knew them? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH3} \emph{character} 3. Does  P say things to people that are insensitive or hurt their feelings? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH4} \emph{character} 4. Does  P say crude things or make sexual remarks that they would not usually have said? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH5} \emph{character} 5. Does  P talk openly about very personal or private matters not usually discussed in public? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH6} \emph{character} 6. Does  P take liberties or touch or hug others in a way that is out of character for him/her? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH7} \emph{character} 7. Does  P show any other signs of loss of control of his/her impulses? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH8A} \emph{character} 8a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH8B} \emph{character} 8b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH8C} \emph{character} 8c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIHTOT} \emph{numeric} H. Disinhibition: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPII} \emph{character} I. Irritability/Lability 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII1} \emph{character} 1. Does  P have a bad temper, flying "off the handle" easily over little things? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII2} \emph{character} 2. Does  P rapidly change moods from one to another, being fine one minute and angry the next? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII3} \emph{character} 3. Does  P have sudden flashes of anger? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII4} \emph{character} 4. Is  P impatient, having trouble coping with delays or waiting for planned activities? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII5} \emph{character} 5. Is  P cranky and irritable? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII6} \emph{character} 6. Is  P argumentative and difficult to get along with? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII7} \emph{character} 7. Does  P show any other signs of irritability? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII8A} \emph{character} 8a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII8B} \emph{character} 8b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII8C} \emph{character} 8c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIITOT} \emph{numeric} I. Irritability/Lability: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIJ} \emph{character} J. Aberrant Motor Behavior 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ1} \emph{character} 1. Does  P pace around the house without apparent purpose? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ2} \emph{character} 2. Does  P rummage around opening and unpacking drawers or closets? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ3} \emph{character} 3. Does  P repeatedly put on and take off clothing? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ4} \emph{character} 4. Does  P have repetitive activities or "habits" that he/she performs over and over? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ5} \emph{character} 5. Does  P engage in repetitive activities such as handling buttons, picking, wrapping string, etc? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ6} \emph{character} 6. Has  P fidget excessively, seem unable to sit still, or bounce his/her feet or tap his/her fingers a lot? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ7} \emph{character} 7. Does  P do any other activities over and over? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ8A} \emph{character} 8a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ8B} \emph{character} 8b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ8C} \emph{character} 8c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJTOT} \emph{numeric} J. Aberrant Motor Behavior: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIK} \emph{character} K. Sleep 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK1} \emph{character} 1. Does  P have difficulty falling asleep? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK2} \emph{character} 2. Does  P get up during the night (do not count if the patient gets up once or twice per night only to go to the bathroom and falls back asleep immediately)? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK3} \emph{character} 3. Does  P wander, pace, or get involved in inappropriate activities at night? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK4} \emph{character} 4. Does  P awaken you during the night? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK5} \emph{character} 5. Does  P awaken at night, dress, and plan to go out thinking that it is morning and time to start the day? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK6} \emph{character} 6. Does  P awaken too early in the morning (earlier than was his/her habit)? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK7} \emph{character} 7. Does  P sleep excessively during the day? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK8} \emph{character} 8. Does  P have any other night-time behaviors that bother you that we haven't talked about? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK9A} \emph{character} 9a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK9B} \emph{character} 9b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK9C} \emph{character} 9c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIKTOT} \emph{numeric} K. Sleep: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPIL} \emph{character} L. Appetite and eating disorders: 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL1} \emph{character} 1. Has he/she had a loss of appetite? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL2} \emph{character} 2. Has he/she had an increase in appetite? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL3} \emph{character} 3. Has he/she had a loss of weight? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL4} \emph{character} 4. Has he/she gained weight? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL5} \emph{character} 5. Has he/she had a change in eating behavior such as putting too much food in his/her mouth at once? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL6} \emph{character} 6. Has he/she had a change in the kind of food he/she likes such as eating too many sweets or other specific types of food? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL7} \emph{character} 7. Has he/she developed eating behaviors such as eating exactly the same types of food each day or eating the food in exactly the same order? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL8} \emph{character} 8. Have there been any other changes in appetite or eating that I haven't asked about? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL9A} \emph{character} 9a. Frequency Ratings 
\itemize{
\item \emph{ADNI2:} 1=1: Occasionally ;  2=2: Often ;  3=3: Frequently ; 4=4: Very frequently
\item \emph{ADNI3, ADNI4:} 1=1: Occasionally ; 2=2: Often ; 3=3: Frequently ; 4=4: Very frequently
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL9B} \emph{character} 9b. Severity Ratings 1=1: Mild ; 2=2: Moderate ; 3=3: Marked
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL9C} \emph{character} 9c. Caregiver Distress 0=Not at all ; 1=Minimally ; 2=Mildly ; 3=Moderately ; 4=Severely ; 5=Very severely or extremely
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPILTOT} \emph{numeric} L. Appetite and eating disorders: Item score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..12
\item \emph{ADNI4} 
}
\item \strong{NPITOTAL} \emph{numeric} NPI Total Score 
\itemize{
\item \emph{ADNI2, ADNI3:} 0..144
\item \emph{ADNI4} 
}
\item \strong{ID} \emph{numeric} Record ID Range value: 2, ..., 187116
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{assessments}
