% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NPIQ}
\alias{NPIQ}
\title{Neuropsychiatric Inventory Q}
\format{
A data frame with 7355 observations and 42 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(NPIQ)
}
\description{
Neuropsychiatric Inventory Q data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10793
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment NPDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{SOURCE} \emph{character} Information Source 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Participant Visit ; 2=Telephone Call
\item \emph{ADNI3, ADNI4:} 1=Participant Visit ; 2=Telephone Call/Video Call
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIA} \emph{character} A. Delusions:  Does  P believe that others are stealing from him/her, or planning to harm him/her in some way? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIASEV} \emph{character} A. Delusions:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIB} \emph{character} B. Hallucinations:  Does  P act as if he/she hears voices? Does he/she talk to people who are not there? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIBSEV} \emph{character} B. Hallucinations:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIC} \emph{character} C. Agitation/Aggression:  Is  P stubborn and resistive to help from others? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPICSEV} \emph{character} C. Agitation/Aggression:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPID} \emph{character} D. Depression/Dysphoria:  Does  P act as if he/she is sad or in low spirits? Does he/she cry? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIDSEV} \emph{character} D. Depression/Dysphoria:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIE} \emph{character} E. Anxiety:  Does  P become upset when separated from you? Does he/she have any other signs of nervousness, such as shortness of breath, sighing, being unable to relax, or feeling excessively tense? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIESEV} \emph{character} E. Anxiety:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIF} \emph{character} F. Elation/Euphoria:  Does  P appear to feel too good or act excessively happy? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIFSEV} \emph{character} F. Elation/Euphoria:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIG} \emph{character} G. Apathy/Indifference:  Does  P seem less interested in his/her usual activities and in the activities and plans of others? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIGSEV} \emph{character} G. Apathy/Indifference:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIH} \emph{character} H. Disinhibition:  Does  P seem to act impulsively? For example, does  P talk to strangers as if he/she knows them, or does  P say things that may hurt people's feelings? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIHSEV} \emph{character} H. Disinhibition:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPII} \emph{character} I. Irritability/Lability:  Is  P impatient or cranky? Does he/she have difficulty coping with delays or waiting for planned activities? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIISEV} \emph{character} I. Irritability/Lability:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIJ} \emph{numeric} J. Aberrant Motor Behavior:  Does  P engage in repetitive activities, such as pacing around the house, handling buttons, wrapping strings, or doing other things repeatedly? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\item \strong{NPIJSEV} \emph{character} J. Aberrant Motor Behavior:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIK} \emph{numeric} K. Sleep:  Does  P awaken you during the night, rise too early in the morning, or take excessive naps during the day? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\item \strong{NPIKSEV} \emph{character} K. Sleep:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPIL} \emph{character} L. Appetite:  Has  P lost or gained weight, or had a change in the food he/she likes? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0=No ; 1=Yes ; 2=N/A
\item \emph{ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPILSEV} \emph{character} L. Appetite:  Severity Ratings 1=1 - Mild (noticeable, but not a significant change). ; 2=2 - Moderate (significant, but not a dramatic change). ; 3=3 - Severe (very marked or prominent. A dramatic change).
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPISCORE} \emph{numeric} NPI-Q Total Score 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 0..36
\item \emph{ADNI4} 
}
\item \strong{SPID} \emph{numeric} Study Partner ID: Range value: 1, ..., 3
\item \strong{ID} \emph{numeric} Record ID Range value: 4, ..., 186247
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{assessments}
