% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NPSTATUS}
\alias{NPSTATUS}
\title{Neuropathology Status}
\format{
A data frame with 11597 observations and 23 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(NPSTATUS)
}
\description{
Neuropathology Status data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI2, ADNIGO, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10900
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Assessment NPSDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD
\item \strong{NPDECIDE} \emph{character} Has a definitive decision about brain donation been made? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPDEC} \emph{character} If a decision has been made, has the participant agreed to register for brain donation? 
\itemize{
\item \emph{ADNI2, ADNI3:} 1=No - provide reason for nonparticipation ; 2=Yes - signed provisional consent form on file ; 3=Yes - consent form being reviewed ; 4=Yes - consent form to be completed at time of autopsy ; 5=Other
\item \emph{ADNI4:} 1=No – provide reason for nonregistration ; 2=Yes – signed registration form on file ; 3=Yes – registration form being reviewed ; 4=Yes – registration form to be completed at time of autopsy
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPDECNO} \emph{character} If No, select reason(s) for nonregistration. 
\itemize{
\item \emph{ADNI2} 
\item \emph{ADNI3, ADNI4:} 1=Not interested, personal choice, or not comfortable. ; 2=Family does not want autopsy, participant does not want to burden family after death ; 3=Religious or other beliefs ; 4=Wants body buried/cremated intact, does not want body altered after death ; 99=Other (specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPDECCF} \emph{character} If there is agreement to register, is the agreement for autopsy for a study other than ADNI? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPUNDEC} \emph{character} If participant is undecided, have they been given information (brochure, discussion with clinician) regarding autopsy/brain donation? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPCONS} \emph{character} Was registration for brain donation discussed at this visit? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPCONSNO} \emph{character} If no, why not: 
\itemize{
\item \emph{ADNI3:} 1=Site error ; 2=Previous definitive no response from participant ; 3=Site not participating in brain donation ; 99=Other reason(s)
\item \emph{ADNI4:} 1=Site error ; 2=Previous definitive "no" response from participant ; 3=Site not participating in brain donation (prior approval from ADNI leadership required) ; 99=Other reason(s)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NPSOURC} \emph{character} Information source 1=Participant visit ; 2=Telephone call
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 13, ..., 187108
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1, v2
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{enrollment}
