% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NV}
\alias{NV}
\title{[DERIVED]  Derived  Nervous System Finding}
\format{
A data frame with 41046 observations and 23 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(NV)
}
\description{
Nervous system finding derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier ; , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbrevation , with character options: NV
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{NVSEQ} \emph{integer} Sequence Number Range value: 1, ..., 85
\item \strong{NVGRPID} \emph{character} Group ID Study Protocol of Data Collection; , with character options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{NVLNKID} \emph{character} Link ID  
\item \strong{NVTESTCD} \emph{character} Short Name of Nervous System Test  
\item \strong{NVTEST} \emph{character} Name of Nervous System Test  
\item \strong{NVCAT} \emph{character} Category for Nervous System , with character options: FDG-8MM, PIBPETSV, AMYLOIDPET, AMYLOID-STATUS, TAUPET, TAUPETPVC
\item \strong{NVSCAT} \emph{character} Subcategory for Nervous System  
\item \strong{NVORRES} \emph{character} Result or Finding in Original Units  
\item \strong{NVSTRESC} \emph{character} Character Result/Finding in Std Format  
\item \strong{NVSTRESN} \emph{numeric} Numeric Result/Finding in Standard Units Range value: -39, ..., 346
\item \strong{NVSTAT} \emph{character} Completion Status , with character options: NOT DONE
\item \strong{NVREASND} \emph{character} Reason Not Done , with character options: NOT ASSESSED, CANNOT BE PROCESSED, FAIL
\item \strong{NVMETHOD} \emph{character} Method of Test or Examination , with character options: FDG PET, PET
\item \strong{NVBLFL} \emph{character} Baseline Flag , with character options: Y
\item \strong{NVDRVFL} \emph{character} Derived Flag , with character options: Yes
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: 1, ..., Inf
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{EPOCH} \emph{character} Epoch , with character options: Baseline, Follow-up
\item \strong{NVDTC} \emph{Date} Date/Time of Collection Date: YYY-MM-DD
\item \strong{NVDY} \emph{integer} Study Day of Collection Range value: -112, ..., 7092
}
}
\keyword{derived_dataset}
