% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{original_study_protocol}
\alias{original_study_protocol}
\title{ADNI Original Protocol Version}
\usage{
original_study_protocol(RID)
}
\arguments{
\item{RID}{Subject RID}
}
\value{
A character vector with the same size as RID with original study
 protocol (i.e. ADNI study phases).
}
\description{
This function is used to identify the original ADNI study phase of subject
 (i.e. when subject enrolled in ADNI study as new-enrollee) based on their RID.
}
\examples{
\dontrun{
RID <- c(0001, 1250, 7015, 10002)
origprot <- original_study_protocol(RID = RID)
origprot
}
}
\seealso{
Other ADNI study protocol/phase: 
\code{\link{adni_phase}()},
\code{\link{adni_phase_order_num}()},
\code{\link{adni_pttype}()},
\code{\link{adni_study_track}()},
\code{\link{convert_adni_phase_order_num}()},
\code{\link{create_col_protocol}()},
\code{\link{create_orig_protocol}()}
}
\concept{ADNI study protocol/phase}
\keyword{adni_procotol_fun}
