% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PACC}
\alias{PACC}
\title{[DERIVED]  Derived  Modified Preclinical Alzheimer’s Cognitive Composite (PACC) Score}
\format{
A data frame with 19571 observations and 26 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-PACC', package = 'ADNIMERGE2')}
}
\usage{
data(PACC)
}
\description{
Modified preclinical alzheimer’s cognitive composite (pacc) score derived dataset.
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character} Original Study Protocol First time participation in ADNI study phase; , with character options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{character} Study Data Collection Protocol , with character options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{RID} \emph{numeric} Subject Identifier Range value: 1, ..., 10898
\item \strong{VISCODE} \emph{character} Study visit code  
\item \strong{ADASQ4SCORE} \emph{numeric} Delayed Word Recall Score - portion of ADAS-Cog Range value: 0, ..., 10
\item \strong{ADASQ4SCORE_VISDATE} \emph{Date} ADAS-Cog Collection/Visit Date Date: YYY-MM-DD
\item \strong{MMSE} \emph{numeric} Mini-Mental Examination Total Score Range value: 0, ..., 30
\item \strong{MMSE_VISDATE} \emph{Date} MMSE Collection/Visit Date Date: YYY-MM-DD
\item \strong{LDELTOTL} \emph{numeric} Logical Memory - Delayed RecallScore Range value: 0, ..., 25
\item \strong{LDELTOTL_VISDATE} \emph{Date} LDELTOTL Collection/Visit Date Date: YYY-MM-DD
\item \strong{DIGITSCR} \emph{numeric} Digit Symbol Substitution Score Range value: 0, ..., 83
\item \strong{DIGITSCR_VISDATE} \emph{Date} DIGITSCR Collection/Visit Date Date: YYY-MM-DD
\item \strong{TRABSCOR} \emph{numeric} Trial B score Range value: 0, ..., 996
\item \strong{LOG.TRABSCOR} \emph{numeric} Trial B score - Log Scale Range value: 0, ..., 6.90475076996184
\item \strong{TRABSCOR_VISDATE} \emph{Date} TRABSCOR Collection/Visit Date Date: YYY-MM-DD
\item \strong{PACC_VISDATE} \emph{Date} PACC Collection/Visit Date Either common date across non-missing PACC component collection date, the maximum relative date or the corresponding study visit date when all component collection date is not available.; Date: YYY-MM-DD
\item \strong{ADASQ4SCOREZ} \emph{numeric} ADAS Q4 Standardized Score Standardized by corresponding baseline score among cognitive normal (CN) new enrollee in ADNI study. Also adjusted for directional effect (i.e. multiplied by -1); Range value: -3.7210201569984, ..., 1.47622571846727
\item \strong{MMSEZ} \emph{numeric} Mini-Mental State Examiniation Standardized Score Standardized by corresponding baseline score among cognitive normal (CN) new enrollee in ADNI study.; Range value: -24.6850367577629, ..., 0.835279966381236
\item \strong{LDELTOTLZ} \emph{numeric} Logical Memory - Delayed Recall Standardized Score Standardized by corresponding baseline score among cognitive normal (CN) new enrollee in ADNI study.; Range value: -3.64701404766878, ..., 3.50702383439261
\item \strong{DIGITSCRZ} \emph{numeric} Digit Symbol Substitution Standardized Score Standardized by corresponding baseline score among cognitive normal (CN) new enrollee in ADNI1 study phase.; Range value: -4.48438541990256, ..., 3.65180778861111
\item \strong{LOG.TRABSCORZ} \emph{numeric} Logical Memory - Delayed Recall Log Standardized Score Standardized by corresponding log transformed baseline score among cognitive normal (CN) new enrollee in ADNI study. Also adjusted for directional effect (i.e. multiplied by -1); Range value: -6.35360930903085, ..., 10.5810477057065
\item \strong{mPACCdigit} \emph{numeric} Modified PACC Digit Score Range value: -56.812114, ..., 7.615753
\item \strong{mPACCtrailsB} \emph{numeric} Modified PACC Trials B Score Range value: -56.812114, ..., 12.950277
\item \strong{ENRLDT} \emph{Date} Enrollment Date Date: YYY-MM-DD
\item \strong{ENRLFG} \emph{character} Enrollment Flag , with character options: Yes
\item \strong{DX} \emph{character} Baseline Diagnostics Status Based on DXSUM eCRF and require to use enrollment flag. Please see `get_adni_blscreen_dxsum` function for more information.; , with character options: CN, Dementia, MCI
}
}
\keyword{derived_dataset}
