% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PEDQCV}
\alias{PEDQCV}
\title{Brief Perceived Ethnic Discrimination Questionnaire-Community Version}
\format{
A data frame with 1253 observations and 34 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PEDQCV)
}
\description{
Brief Perceived Ethnic Discrimination Questionnaire-Community Version data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNI2, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{DONE} \emph{character} Was assessment/procedure done? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 1=Participant Unable for Cognitive Reasons ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA1} \emph{character} 1. Have you been treated unfairly by teachers, principals, or other staff at school? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA2} \emph{character} 2. Have others thought you couldn't do things or handle a job? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA3} \emph{character} 3. Have others threatened to hurt you (ex: said they would hit you)? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA4} \emph{character} 4. Have others actually hurt you or tried to hurt you (ex: kicked or hit you)? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA5} \emph{character} 5. Have policeman or security officers been unfair to you? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA6} \emph{character} 6. Have others threatened to damage your property? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA7} \emph{character} 7. Have others actually damaged your property? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA8} \emph{character} 8. Have others made you feel like an outsider who doesn't fit in because of your dress, speech, or other characteristics related to your ethnicity? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA9} \emph{character} 9. Have you been treated unfairly by co-workers or classmates? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA10} \emph{character} 10. Have others hinted that you are dishonest or can't be trusted? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA11} \emph{character} 11. Have people been nice to your face, but said bad things about you behind your back? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA12} \emph{character} 12. Have people who speak a different language made you feel like an outsider? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA13} \emph{character} 13. Have others ignored you or not paid attention to you? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA14} \emph{character} 14. Has your boss or supervisor been unfair to you? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA15} \emph{character} 15. Have others hinted that you must not be clean 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA16} \emph{character} 16. Have people not trusted you? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PEDQA17} \emph{character} 17. Has it been hinted that you must be lazy? 1=Never ; 2=Rarely ; 3=Sometimes ; 4=Often ; 5=Very often
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 74, ..., 128212
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
