% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PETC3}
\alias{PETC3}
\title{PETC3}
\format{
A data frame with 434 observations and 25 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PETC3)
}
\description{
PETC3 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNIGO, ADNI1, ADNI3
\item \strong{COLPROT} \emph{character}   , with options: ADNI3
\item \strong{ID} \emph{numeric}   Range value: 1840, ..., 174179
\item \strong{RID} \emph{numeric}   Range value: 69, ..., 7121
\item \strong{PTID} \emph{character}    
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 63
\item \strong{VISCODE} \emph{character}   , with options: init, bl
\item \strong{VISCODE2} \emph{character}    
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{SCANDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{LONIUPDT} \emph{ }   Date: YYY-MM-DD
\item \strong{LONIUID} \emph{character}    
\item \strong{REVDT} \emph{ }   Date: YYY-MM-DD
\item \strong{ALLFRAME} \emph{numeric}   Range value: 0, ..., 1
\item \strong{UNUSABL} \emph{character}   , with options: 3rd, 5,6, 3-6, 3, 6, 1-6
\item \strong{UNRSN} \emph{numeric}   Range value: 1, ..., 2
\item \strong{UNRSNSPEC} \emph{character}   , with options: PET acquisition started 16 minutes late, corrupt/streak artifacts, outside acquisition window, late scan start, late scan staT, Late scan start. Last 2 frns outside acq window, acquired outside protocol window, FDG PET ACQ started 75 min post inj
\item \strong{SCANQLTY} \emph{numeric}   Range value: 0, ..., 1
\item \strong{REPROCREQ} \emph{numeric}   Range value: 0, ..., 1
\item \strong{PROCERR} \emph{character}   , with options: 5|99, 2|3|4, 8, 2, 99, 1, 2|3|9, 1|2
\item \strong{PROCERRSPEC} \emph{character}   , with options: Scatter error, Reprocess with offsets, Incomplete reconstruction, Header strip/filtering, Header strip
\item \strong{RESCANREQ} \emph{numeric}   Range value: 0, ..., 1
\item \strong{COMM} \emph{character}    
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{imaging}
