% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PETMETA3}
\alias{PETMETA3}
\title{FDG PET Scan Information}
\format{
A data frame with 701 observations and 32 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PETMETA3)
}
\description{
FDG PET Scan Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI1, ADNIGO, ADNI3
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 7121
\item \strong{VISCODE} \emph{character}   , with options: init, bl
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{SCANDATE} \emph{ } Scan Date/Dose Date Date: YYY-MM-DD
\item \strong{DONE} \emph{character} Was the scan conducted? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Not called for by protocol ; 8=Software/scanner error ; 99=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDSPECIFY} \emph{character} If Other, specify:  
\item \strong{RADTRACER} \emph{character} If scan not done, was radiotracer administered? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{QCTIME} \emph{ } Time of today's Scanner QC hms: HH-MM-SS
\item \strong{PMBLTIME} \emph{ } Time of blood glucose measurement hms: HH-MM-SS
\item \strong{PMBLGLUC} \emph{numeric} Blood Glucose (pre-FDG) Range value: 5, ..., 178
\item \strong{ASSAYTIME} \emph{ } Time of FDG dose assay hms: HH-MM-SS
\item \strong{FDGDOS} \emph{numeric} Injected FDG dose assay Range value: 4.3, ..., 10.8
\item \strong{INJTIME} \emph{ } Time of FDG injection hms: HH-MM-SS
\item \strong{SCANTIME} \emph{ } Emission Scan Start Time hms: HH-MM-SS
\item \strong{VARIAT} \emph{character} Any variations from protocol during FDG uptake? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PROTID} \emph{character} Predefined Acquisition Protocol Name  
\item \strong{MOTION} \emph{character} Participant motion problems 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{OTHERVAR} \emph{character} Other protocol variations 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{RECON} \emph{character} Select reconstruction used 1=3D or LOR Ramla (Philips) ; 2=FORE/2D-OSEM (Siemens) ; 4=3D-OSEM (Siemens) ; 5=3D Iterative (GE) ; 6=FORE Iterative (GE)
\emph{Decoded Value: }\strong{Yes}
\item \strong{SHARP} \emph{character} For Phillips scanners:  was "smooth" parameter set to "sharp"? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{ARCHIVE} \emph{character} Data archived locally? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{LONI} \emph{character} Data transferred to LONI? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{TRANDATE} \emph{ } Transfer Date Date: YYY-MM-DD
\item \strong{ID} \emph{numeric}   Range value: 501, ..., 174814
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 63
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{imaging}
