% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PETMETA_ADNI1}
\alias{PETMETA_ADNI1}
\title{PET Scan Information}
\format{
A data frame with 1957 observations and 105 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PETMETA_ADNI1)
}
\description{
PET Scan Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 3958
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1435
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code , with options: bl, m06, m12, m18, m24, m36, m48
\item \strong{VISCODE2} \emph{character}   , with options: bl, m06, m12, m18, m24, m36, m48
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 2
\item \strong{PMCONDCT} \emph{character} Was the scan conducted? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMREASON} \emph{character} Reason why the scan was not conducted: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMCONSENT} \emph{character} Did the participant consent to quantitative scan? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMTYPE} \emph{character} Type of scan conducted 1=Qualitative ; 2=Quantitative
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{PMSCANNER} \emph{character} GE 1=check here
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMGEMODEL} \emph{character} If GE, Scanner Model: 1=Advance ; 2=Discovery LS ; 3=Discovery ST ; 4=Discovery RX ; 5=Discovery STE
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSIEMENS} \emph{character} Siemens/CTI 1=check here
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSIEMODEL} \emph{character} If Siemens/CTI, Scanner Model: 1=ACCEL ; 2=Biograph ; 3=BioGraph BGO ; 4=BioGraph HiRez ; 5=EXACT ; 6=HR+ ; 7=HRRT
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMPHILLIPS} \emph{character} Phillips 1=check here
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMPHMODEL} \emph{character} If Phillips, Scanner Model: 1=Allegro ; 2=Allegro-Neuro ; 3=Gemini ; 4=Gemini GLX ; 5=Gemini-TF
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMQCTIME} \emph{character} Time of today's Scanner QC  
\item \strong{PMBLTIME} \emph{character} Time of blood glucose measurement  
\item \strong{PMBLGLUC} \emph{numeric} Blood Glucose (pre-FDG) Range value: 50, ..., 250
\item \strong{PMFDGTIME} \emph{character} Time of FDG dose assay  
\item \strong{PMFDGDOS} \emph{numeric} FDG dose assay Range value: 3.8, ..., 54
\item \strong{PMFDGVOL} \emph{numeric} FDG Volume Range value: 0.1, ..., 10
\item \strong{PMINJTIME} \emph{character} Time of FDG injection  
\item \strong{PMSCTIME} \emph{character} Time scan started (emission)  
\item \strong{PMVARIAT} \emph{character} Any variations from protocol during FDG uptake? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMPROTID} \emph{character} Predefined acquisition protocol ID  
\item \strong{PMFRAME} \emph{character} Which framing rate was used? 1=6 frames, 5 min/frame (6x300s) ; 2=1 frame, 30 min (1x1800s) ; 3=Quantitative
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMMOTION} \emph{character} Subject motion problems: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMMALFUN} \emph{character} Scanner malfunction 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMOTHER} \emph{character} Other protocol variations: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMRECON} \emph{character} Check which of the following reconstructions was used: 2=FORE/2D-OSEM ; 4=OSEM3D-OP ; 1=3D-Ramla ; 3=3D Back-projection
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSUBSET} \emph{character} # subsets: 14=14 ; 16=16 ; 98=N/A ; 99=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSUBSPE} \emph{numeric} If Other, specify Range value: 0, ..., 92
\item \strong{PMITERAT} \emph{character} # iterations: 2=2 ; 4=4 ; 6=6 ; 99=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMITERSPE} \emph{numeric} If Other, specify: Range value: 0, ..., 16
\item \strong{PMRAMLA} \emph{character} If Ramla, Lambda=0.016? 1=Check here to confirm
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMFILTER} \emph{character} If Back Projection, Ramp filter? 1=Check here to confirm
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMMODEON} \emph{character} If FORE/2D-OSEM, Brain Mode "ON" for PET only scanners or TRIM "ON" for PET/CT scanners? 1=Check here to confirm
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSMOOTH} \emph{character} No post-process smoothing: 1=Check here to confirm
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMDECAY} \emph{character} Decay Correction 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMSCATTR} \emph{character} Scatter Correction: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMATTEN} \emph{character} Attenuation Correction: 1=CT ; 2=Ge-68+Segmentation ; 3=Cs-137+Segmentation
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMWITHDRAW} \emph{character} Clock for blood sample withdrawal time 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMWITHTIME} \emph{character} If No, provide the time difference , with options: offset seems to be 44 sec (16:01:11 vs 16:01:55), 39 seconds, ~50 sec off, Subject was a hard stick, Nurse had trouble drawing blood
\item \strong{PMCOUNT} \emph{character} Clock for blood sample count time 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMCOUNTIME} \emph{character} If No, provide the time difference , with options: 39 seconds, ~50 sec off
\item \strong{PMBG1TIME} \emph{character} Sample Count Time (24h) Background #1  
\item \strong{PMBG1VOL} \emph{numeric} Sample Plasma Volume Counted Background #1 Range value: 200, ..., 250
\item \strong{PMBG1DUR} \emph{numeric} Sample Count Duration Background #1 Range value: 60, ..., 300
\item \strong{PMBG1COUNT} \emph{numeric} Sample Count Rate Background #1 Range value: 1, ..., 279
\item \strong{PMP1DTIME} \emph{character} Sample Draw Time (24h) Plasma 1  
\item \strong{PMP1CTIME} \emph{character} Sample Count Time (24h) Plasma 1  
\item \strong{PMP1BGL} \emph{numeric} Sample BGL Plasma 1 Range value: 63, ..., 168
\item \strong{PMP1VOL} \emph{numeric} Sample Plasma Volume Counted Plasma 1 Range value: 200, ..., 250
\item \strong{PMP1DUR} \emph{numeric} Sample Count Duration Plasma 1 Range value: 25, ..., 60
\item \strong{PMP1COUNT} \emph{numeric} Sample Count Rate Plasma 1 Range value: 3, ..., 9999
\item \strong{PMP2DTIME} \emph{character} Sample Draw Time (24h) Plasma 2  
\item \strong{PMP2CTIME} \emph{character} Sample Count Time (24h) Plasma 2  
\item \strong{PMP2BGL} \emph{numeric} Sample BGL Plasma 2 Range value: 66, ..., 164
\item \strong{PMP2VOL} \emph{numeric} Sample Plasma Volume Counted Plasma 2 Range value: 111, ..., 250
\item \strong{PMP2DUR} \emph{numeric} Sample Count Duration Plasma 2 Range value: 11, ..., 60
\item \strong{PMP2COUNT} \emph{numeric} Sample Count Rate Plasma 2 Range value: 8, ..., 9999
\item \strong{PMP3DTIME} \emph{character} Sample Draw Time (24h) Plasma 3  
\item \strong{PMP3CTIME} \emph{character} Sample Count Time (24h) Plasma 3  
\item \strong{PMP3BGL} \emph{numeric} Sample BGL Plasma 3 Range value: 53, ..., 165
\item \strong{PMP3VOL} \emph{numeric} Sample Plasma Volume Counted Plasma 3 Range value: 60, ..., 250
\item \strong{PMP3DUR} \emph{numeric} Sample Count Duration Plasma 3 Range value: 29, ..., 200
\item \strong{PMP3COUNT} \emph{numeric} Sample Count Rate Plasma 3 Range value: 7, ..., 9999
\item \strong{PMP4DTIME} \emph{character} Sample Draw Time (24h) Plasma 4  
\item \strong{PMP4CTIME} \emph{character} Sample Count Time (24h) Plasma 4  
\item \strong{PMP4BGL} \emph{numeric} Sample BGL Plasma 4 Range value: 63, ..., 168
\item \strong{PMP4VOL} \emph{numeric} Sample Plasma Volume Counted Plasma 4 Range value: 99, ..., 250
\item \strong{PMP4DUR} \emph{numeric} Sample Count Duration Plasma 4 Range value: 30, ..., 99
\item \strong{PMP4COUNT} \emph{numeric} Sample Count Rate Plasma 4 Range value: 6, ..., 9999
\item \strong{PMP5DTIME} \emph{character} Sample Draw Time (24h) Plasma 5  
\item \strong{PMP5CTIME} \emph{character} Sample Count Time (24h) Plasma 5  
\item \strong{PMP5BGL} \emph{numeric} Sample BGL Plasma 5 Range value: 60, ..., 177
\item \strong{PMP5VOL} \emph{numeric} Sample Plasma Volume Counted Plasma 5 Range value: 99, ..., 250
\item \strong{PMP5DUR} \emph{numeric} Sample Count Duration Plasma 5 Range value: 39, ..., 99
\item \strong{PMP5COUNT} \emph{numeric} Sample Count Rate Plasma 5 Range value: 1, ..., 9999
\item \strong{PMBG2CTIME} \emph{character} Sample Count Time (24h) Background 2  
\item \strong{PMBG2VOL} \emph{numeric} Sample Plasma Volume Counted Background 2 Range value: 200, ..., 250
\item \strong{PMBG2DUR} \emph{numeric} Sample Count Duration Background 2 Range value: 1, ..., 300
\item \strong{PMBG2COUNT} \emph{numeric} Sample Count Rate Background 2 Range value: 1, ..., 283
\item \strong{PMPIPVOL} \emph{character} Was the pipetted plasma volume 200 uL? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMPIPVOLSP} \emph{character} If No, denote volume used: , with options: 250 microL -  have only fixed vol. pipettes, 250 uL fixed volume pipette, 250 micro liter (fixed volume pipette), 250 microliters, 250 microL fixed volume pipette, 250 microliter fixed volume pipette, 250 ul fixed volume pipette, 250;  fixed volume pipette, 250 uL fixed volume  pipette, 250 uL  fixed volume pipette
\item \strong{PMCTIME} \emph{character} Was the plasma sample count time 1 minute? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMCTIMESP} \emph{character} If No, denote count time used , with options: fixed counts 10000: 25-39 sec
\item \strong{PMPHACT} \emph{numeric} Phantom Activity at Time of Scan Range value: 0.25, ..., 3
\item \strong{PMPHVOL} \emph{numeric} Phantom Volume Range value: 5969, ..., 8538
\item \strong{PMPHCOUNT} \emph{numeric} Average Counts from Phantom Image ROI Range value: 5.7e-05, ..., 11264
\item \strong{PMALIQVOL} \emph{numeric} Aliquot Volume Range value: 50, ..., 1500
\item \strong{PMALIQACT} \emph{numeric} Aliquot Count Rate Range value: 137, ..., 6138
\item \strong{PMTRNSFR} \emph{character} Was data transferred to LONI within 24 hours of scan? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMTRNDATE} \emph{ } Transfer Date Date: YYY-MM-DD
\item \strong{PMARCHIVE} \emph{character} Data Archived Locally 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMLPDONE} \emph{character} Was a Lumbar Puncture completed prior to the PET scan? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PMLPINTER} \emph{character} If Yes, What was the interval between LP and PET? 1=less than 6 hours ; 2=6-12 hours ; 3=13-24 hours ; 4=25-48 hours ; 5=49-72 hours ; 6=more than 72 hours
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{imaging}
