% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PETQC}
\alias{PETQC}
\title{PET QC Tracking}
\format{
A data frame with 3951 observations and 39 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PETQC)
}
\description{
PET QC Tracking data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 3990
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 5296
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character}    
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 4
\item \strong{SCAN} \emph{character} Scan 1=Original Scan ; 2=Rescan
\emph{Decoded Value: }\strong{Yes}
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{PQFILENO} \emph{numeric} Upload Number 1..25
\item \strong{PQDATTRA} \emph{character} Date images uploaded to LONI  
\item \strong{LONIUID} \emph{character} LONI Unique Series ID  
\item \strong{PQDATE} \emph{character} Date reviewed  
\item \strong{PQACQFR} \emph{character} Acquisition 
\itemize{
\item \emph{ADNI1:} 1=6 frames, 5 min/frame (qualitative) ; 2=1 frame, 30 min (qualitative) ; 3=Quantitative
\item \emph{ADNIGO:} 1=6 frames, 5 min/frame ; 2=2 scans, 15 min each
\item \emph{ADNI2:} 1=6 frames, 5 min/frame ; 2=2 scans, 15 min each ; 3=1 scan, 30min
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQCORALG} \emph{character} Correct Reconstruction Algorithm? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQRECON} \emph{character} Check which of the following reconstruction algorithms was used 
\itemize{
\item \emph{ADNI1:} 2=FORE/2D-OSEM ; 4=OSEM3D-OP ; 1=3D-Ramla ; 3=3D Back-projection
\item \emph{ADNIGO:} 2=FORE/2D-OSEM (Siemens) ; 4=OSEM3D (Siemens) ; 5=3D Iterative (GE) ; 1=3D-Ramla (Philips) ; 3=3D Back-projection (GE)
\item \emph{ADNI2:} 2=FORE/2D-OSEM (Siemens) ; 4=OSEM3D (Siemens) ; 5=3D Iterative (GE) ; 6=FORE Iterative (GE) ; 1=3D or LOR Ramla (Philips) ; 3=3D Back-projection (GE)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQCORNUM} \emph{character} If FORE/2D-OSEM or OSEM3D or 3D Iterative or FORE Iterative, correct # subsets and iterations? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQCORFIL} \emph{character} Correct Filters/TRIM? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQCORTHI} \emph{character} Correct Slice Thickness (PET/CT systems)? 1=Yes ; 0=No ; 2=N/A (PET only system)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQVOXEL} \emph{character} Acceptable Voxel Size? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQALLFRAME} \emph{character} Were all frames acceptable? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQUNUSABLE} \emph{character} Indicate which frames were unacceptable:  
\item \strong{PQREASON} \emph{character} Indicate why frames were unacceptable: 1=Subject motion ; 2=Other (Please Specify)
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQREASSP} \emph{character} If Other, Specify:  
\item \strong{PASS} \emph{character} Pass QC? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQPROERR} \emph{character} A. Processing Error(s) 1=Incorrect reconstruction algorithm ; 2=Incorrect filtering ; 3=Incorrect slice thickness ; 4=Incorrect voxel size ; 5=Attenuation Error ; 6=Normalization Error ; 7=Inadequate counts ; 8=Inadequate field of view ; 9=Incorrect acquisition protocol ; 10=Unfixable image artifact (e.g. attenuation, normalization, etc.) ; 99=Other (specify)
\item \strong{PQERRSP} \emph{character} Specify  
\item \strong{PQACTION} \emph{character} B. QC outcome 
\itemize{
\item \emph{ADNI1:} 1=Request Reprocessing ; 2=Request Rescan ; 3=Unuseable - Don't request rescan (provide details) ; 4=Allow Rescan (rescan not requested)
\item \emph{ADNIGO, ADNI2:} 1=Request Reprocessing ; 2=Request Rescan ; 3=Unusable - Don't request rescan (provide details) ; 4=Allow Rescan (rescan not requested)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PQNONESP} \emph{character} If unusable, reason for not rescanning  
\item \strong{PQISSUES} \emph{character} Imaging issues 
\itemize{
\item \emph{ADNI1:} 1=Motion - none ; 2=Motion - mild ; 3=Motion - mild-to-moderate ; 4=Motion - moderate ; 5=Motion - moderate-to-severe ; 6=Top Cutoff - none ; 7=Top Cutoff - 0-3mm ; 8=Top Cutoff - 3-6mm ; 9=Top Cutoff - 6-9mm ; 10=Top Cutoff - >9mm ; 11=Bottom Cutoff - none ; 12=Bottom Cutoff - 0-3mm ; 13=Bottom Cutoff - 3-6mm ; 14=Bottom Cutoff - 6-9mm ; 15=Bottom Cutoff - >9mm ; 99=Other (specify)
\item \emph{ADNIGO, ADNI2:} 1=Motion - minimal ; 2=Motion - mild ; 3=Motion - mild-to-moderate ; 4=Motion - moderate ; 5=Motion - moderate-to-severe ; 6=Top Cutoff - none ; 7=Top Cutoff - 0-3mm ; 8=Top Cutoff - 3-6mm ; 9=Top Cutoff - 6-9mm ; 10=Top Cutoff - >9mm ; 11=Bottom Cutoff - none ; 12=Bottom Cutoff - 0-3mm ; 13=Bottom Cutoff - 3-6mm ; 14=Bottom Cutoff - 6-9mm ; 15=Bottom Cutoff - >9mm ; 99=Other (specify)
}
\item \strong{PQISSOTH} \emph{character} Other, specify , with options: Single frame - motion N/A, ., front of head cut off, high in FOV
\item \strong{COMMENTS} \emph{character} Additional QC Comments  
\item \strong{QUARANTINE} \emph{character} Release from quarantine? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{PAYSITE} \emph{character} Pay Site? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{imaging}
