% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHYSICAL}
\alias{PHYSICAL}
\title{Physical Exam}
\format{
A data frame with 5751 observations and 29 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PHYSICAL)
}
\description{
Physical Exam data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{PXGENAPP} \emph{character} 1. General Appearance 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXHEADEY} \emph{character} 2. Head, Eyes, Ears, Nose and Throat 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXNECK} \emph{character} 3. Neck 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXCHEST} \emph{character} 4. Chest 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXHEART} \emph{character} 5. Heart 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXABDOM} \emph{character} 6. Abdomen 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXEXTREM} \emph{character} 7. Extremities 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXEDEMA} \emph{character} 8. Edema 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXPERIPH} \emph{character} 9. Peripheral Vascular 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXSKIN} \emph{character} 10. Skin and Appendages 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXMUSCUL} \emph{character} 11. Musculoskeletal 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXBACK} \emph{character} 12. Back 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXOTHER} \emph{character} 13. Other 1=Normal ; 2=Abnormal
\emph{Decoded Value: }\strong{Yes}
\item \strong{PXABNORM} \emph{character} 15. Based on the Physical Examination, clinician must check appropriate box below: 1=Findings consistent with eligibility for study ; 2=Participant is not eligible for study
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 6, ..., 186319
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{medical_history}
