% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PIBMETA}
\alias{PIBMETA}
\title{PIB Scan Information}
\format{
A data frame with 257 observations and 76 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PIBMETA)
}
\description{
PIB Scan Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 520
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 57, ..., 1430
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 124
\item \strong{VISCODE} \emph{character} Visit code , with options: m12, bl, m24, m36, m48
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{PBCONDCT} \emph{numeric} Was the scan conducted? 1=Yes ; 0=No
\item \strong{PBREASON} \emph{numeric} Reason why the scan was not conducted: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify)
\item \strong{PBOTHSPE} \emph{character} If Other, specify:  
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{PBGE} \emph{numeric} GE 1=check here
\item \strong{PBGEMODEL} \emph{numeric} If GE, Scanner Model: 1=Advance ; 2=Discovery LS ; 3=Discovery ST ; 4=Discovery RX ; 5=Discovery STE
\item \strong{PBSIEMENS} \emph{numeric} Siemens/CTI 1=check here
\item \strong{PBSIEMODEL} \emph{numeric} If Siemens/CTI, Scanner Model: 1=ACCEL ; 2=Biograph ; 3=BioGraph BGO ; 4=BioGraph HiRez ; 5=EXACT ; 6=HR+ ; 7=HRRT
\item \strong{PBPHILLIPS} \emph{numeric} Phillips 1=check here
\item \strong{PBPHMODEL} \emph{numeric} If Phillips, Scanner Model: 1=Allegro ; 2=Allegro-Neuro ; 3=Gemini ; 4=Gemini GLX ; 5=Gemini-TF
\item \strong{PBQCTIME} \emph{character} Time of today's Scanner QC  
\item \strong{PBPIBTIME} \emph{character} Time of PIB dose assay  
\item \strong{PBPIBDOS} \emph{numeric} PIB dose assay Range value: 5.8, ..., 29.8
\item \strong{PBRESTIME} \emph{character} Time of residual PIB assay  
\item \strong{PBRESID} \emph{numeric} Residual left in syringe Range value: 0.01, ..., 72
\item \strong{PBINJDOS} \emph{numeric} Net injected dose of PIB Range value: 5, ..., 17.1
\item \strong{PBPIBVOL} \emph{numeric} PIB volume Range value: 3, ..., 17
\item \strong{PBINJTIME} \emph{character} Time of PIB injection  
\item \strong{PBSCTIME} \emph{character} Time scan started (emission)  
\item \strong{PBSCANCOM} \emph{character} Provide an explanation if start time is not within the allowable window  
\item \strong{PBVARIAT} \emph{numeric} Any variations from protocol during PIB uptake? 1=Yes ; 0=No
\item \strong{PBVARSP} \emph{character} If Yes, describe: , with options: 9.4 mCi injected, Subject had to go to bathroom, Patient ate at 11:30, LOW PIB DOSE SYNTHESIZED, Participant fell asleep, scan not started in window., dose<15+/-1.5 mci due to Qc time of isotope and increased time setting upsubject in scanner.
\item \strong{PBPROTID} \emph{character} Predefined acquisition protocol ID  
\item \strong{PBSTATDYN} \emph{numeric} Indicate whether scan was static or dynamic: 1=Static (1 x 20 min) ; 2=Standard Dynamic (4 x 5 min) ; 3=Dynamic (specify)
\item \strong{PBNMFRAM1} \emph{numeric} 1. No. of Frames Range value: 4, ..., 8
\item \strong{PBDURATN1} \emph{numeric} Duration: Range value: 4, ..., 300
\item \strong{PBNMFRAM2} \emph{numeric} 2. No. of Frames: Range value: 0, ..., 12
\item \strong{PBDURATN2} \emph{numeric} Duration: Range value: 0, ..., 300
\item \strong{PBNMFRAM3} \emph{numeric} 3. No. of Frames: Range value: 0, ..., 27
\item \strong{PBDURATN3} \emph{numeric} Duration: Range value: 0, ..., 120
\item \strong{PBNMFRAM4} \emph{numeric} 4. No. of Frames: Range value: 0, ..., 16
\item \strong{PBDURATN4} \emph{numeric} Duration: Range value: 0, ..., 300
\item \strong{PBNMFRAM5} \emph{numeric} 5. No. of Frames: Range value: 0, ..., 8
\item \strong{PBDURATN5} \emph{numeric} Duration: Range value: 0, ..., 300
\item \strong{PBNMFRAM6} \emph{numeric} 6. No. of Frames: Range value: 0, ..., 6
\item \strong{PBDURATN6} \emph{numeric} Duration: Range value: 0, ..., 300
\item \strong{PBNMFRAM7} \emph{numeric} 7. No. of Frames: Range value: 0, ..., 16
\item \strong{PBDURATN7} \emph{numeric} Duration: Range value: 0, ..., 300
\item \strong{PBNMFRAM8} \emph{numeric} 8. No. of Frames: Range value: 0, ..., 0
\item \strong{PBDURATN8} \emph{numeric} Duration: Range value: 0, ..., 0
\item \strong{PBNMFRAM9} \emph{numeric} 9. No. of Frames: Range value: 0, ..., 0
\item \strong{PBDURATN9} \emph{numeric} Duration: Range value: 0, ..., 0
\item \strong{PBNMFRAM10} \emph{numeric} 10. No. of Frames: Range value: 0, ..., 0
\item \strong{PBDURATN10} \emph{numeric} Duration: Range value: 0, ..., 0
\item \strong{PBMOTION} \emph{numeric} Subject motion problems: 1=Yes ; 0=No
\item \strong{PBMOTSP} \emph{character} If yes, describe:  
\item \strong{PBMALFUN} \emph{numeric} Scanner malfunction 1=Yes ; 0=No
\item \strong{PBMALSP} \emph{character} If yes, describe: , with options: error with ACS creating and reading file
\item \strong{PBOTHER} \emph{numeric} Other protocol variations: 1=Yes ; 0=No
\item \strong{PBOTHSP} \emph{character} If yes, describe: , with options: Only 2 five minute frames were performed, only 2 5 minute frames were done according to the ADNI PIB protocol, A transmission scan was not performed therefore the scan was attenuation corrected using drawn (eliptical) regions of interest over the entire head (outside the skull)., The injected dose of C-11 PIB was 8.2 mCi instead of recommended dose of >10 mCi., MCH 60 min multiframe acquisition used.  Tech advised to use PIB ADNI acquisition protocol, Wrong Imaging Parameters were used, subject injected with only 13.1 mCi because of delay due to scanner error, Placed a lot of towels under the neck, to make patient comfortablt, at 70min into scan - subject put hand into scanner, less than +/-1.5mci of desired dose injected
\item \strong{PBRECON} \emph{numeric} Check which of the following reconstructions was used: 2=FORE/2D-OSEM ; 4=OSEM3D-OP ; 1=3D-Ramla ; 3=3D Back-projection
\item \strong{PBSUBSET} \emph{numeric} # subsets: 14=14 ; 16=16 ; 98=N/A ; 99=Other
\item \strong{PBSUBSPE} \emph{numeric} If Other, specify Range value: Inf, ..., -Inf
\item \strong{PBITERAT} \emph{numeric} # iterations: 2=2 ; 4=4 ; 6=6 ; 99=Other
\item \strong{PBITERSPE} \emph{numeric} If Other, specify: Range value: Inf, ..., -Inf
\item \strong{PBRAMLA} \emph{numeric} If Ramla, Lambda=0.016? 1=Check here to confirm
\item \strong{PBFILTER} \emph{numeric} If Back Projection, Ramp filter? 1=Check here to confirm
\item \strong{PBMODEON} \emph{numeric} If FORE/2D-OSEM, Brain Mode "ON" for PET only scanners or TRIM "ON" for PET/CT scanners? 1=Check here to confirm
\item \strong{PBSMOOTH} \emph{numeric} No post-process smoothing: 1=Check here to confirm
\item \strong{PBDECAY} \emph{numeric} Decay Correction 1=Yes ; 0=No
\item \strong{PBSCATTR} \emph{numeric} Scatter Correction: 1=Yes ; 0=No
\item \strong{PBMATTEN} \emph{numeric} Attenuation Correction: 1=CT ; 2=Ge-68+Segmentation ; 3=Cs-137+Segmentation
\item \strong{PBTRSNFR} \emph{numeric} Was data transferred to LONI within 24 hours of scan? 1=Yes ; 0=No
\item \strong{PBTRNDATE} \emph{character} Transfer Date  
\item \strong{PBTRNCOM} \emph{character} Comments  
\item \strong{PBARCHIVE} \emph{numeric} Data Archived Locally 1=Yes ; 0=No
\item \strong{PBARCMED} \emph{character} Archive Medium  
\item \strong{PBARCCOM} \emph{character} Comments  
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
