% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PIBPETSUVR}
\alias{PIBPETSUVR}
\title{U Pitt PIB PET Analysis}
\format{
A data frame with 224 observations and 22 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PIBPETSUVR)
}
\description{
U Pitt PIB PET Analysis data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric}   Range value: 57, ..., 1430
\item \strong{VISCODE} \emph{character}   , with options: m12, m36, m24, m48, bl
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{IMAGEUID} \emph{numeric}   Range value: 8935, ..., 26041
\item \strong{LONISID} \emph{character}    
\item \strong{LONIUID} \emph{character}    
\item \strong{ACG} \emph{numeric} Anterior cingulate Range value: 0.9, ..., 3.04
\item \strong{FRC} \emph{numeric} Frontal cortex Range value: 1.09, ..., 3.05
\item \strong{LTC} \emph{numeric} Lateral temporal cortex Range value: 1.03, ..., 2.74
\item \strong{PAR} \emph{numeric} Parietal cortex Range value: 1.05, ..., 2.78
\item \strong{PRC} \emph{numeric} Precuneus cortex Range value: 0.98, ..., 2.89
\item \strong{MTC} \emph{numeric} Mesial temporal cortex Range value: 0.75, ..., 1.49
\item \strong{OCC} \emph{numeric} Occipital cortex Range value: 0.98, ..., 2.05
\item \strong{OCP} \emph{numeric} Occipital pole Range value: 1.1, ..., 2.23
\item \strong{PON} \emph{numeric} Pons Range value: 1.09, ..., 2.62
\item \strong{AVS} \emph{numeric} Anterior ventral striatum Range value: 1, ..., 2.84
\item \strong{CER} \emph{numeric} Cerebellum Range value: 1, ..., 1
\item \strong{SMC} \emph{numeric} Sensory motor cortex Range value: 0.96, ..., 2.04
\item \strong{SWM} \emph{numeric} Sub-cortical white matter Range value: 1.15, ..., 2.75
\item \strong{THL} \emph{numeric} Thalamus Range value: 0.95, ..., 1.98
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
