% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PIBQC}
\alias{PIBQC}
\title{PIB QC Tracking}
\format{
A data frame with 294 observations and 34 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PIBQC)
}
\description{
PIB QC Tracking data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 590
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 57, ..., 1430
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 124
\item \strong{VISCODE} \emph{character} Visit code , with options: m12, bl, m24, m36, m48
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 4
\item \strong{EXAMDATE} \emph{ } PIB PET Scan Date Date: YYY-MM-DD
\item \strong{PQFILENO} \emph{numeric} Upload Number Range value: 1, ..., 4
\item \strong{PQDATTRA} \emph{character} Date images uploaded to LONI  
\item \strong{LONIUID} \emph{character} LONI Unique Series ID  
\item \strong{PQDATE} \emph{character} Date reviewed  
\item \strong{PQACQFR} \emph{numeric} Acquisition 1=4 frames, 5 min/frame (qualitative) ; 2=1 frame, 20 min (qualitative) ; 3=Quantitative
\item \strong{PQCORALG} \emph{numeric} Correct Reconstruction Algorithm? 1=Yes ; 0=No
\item \strong{PQRECON} \emph{numeric} Check which of the following reconstruction algorithms was used 2=FORE/2D-OSEM ; 4=OSEM3D-OP ; 1=3D-Ramla ; 3=3D Back-projection
\item \strong{PQCORNUM} \emph{numeric} If OSEM or Ramla, correct # iterations and subsets? 1=Yes ; 0=No
\item \strong{PQCORFIL} \emph{numeric} Correct Filters/TRIM? 1=Yes ; 0=No
\item \strong{PQCORTHI} \emph{numeric} Correct Slice Thickness (PET/CT systems)? 1=Yes ; 0=No ; 2=N/A (PET only system)
\item \strong{PQVOXEL} \emph{numeric} Acceptable Voxel Size? 1=Yes ; 0=No
\item \strong{PQALLFRAME} \emph{numeric} Were all frames acceptable? 1=Yes ; 0=No
\item \strong{PQUNUSABLE} \emph{numeric} Indicate which frames were unacceptable: Range value: 2, ..., 1234
\item \strong{PQREASON} \emph{numeric} Indicate why frames were unacceptable: 1=Subject motion ; 2=Other (Please Specify)
\item \strong{PQREASSP} \emph{character} If Other, Specify: , with options: Incorrect framing or binning, Corrupt
\item \strong{PASS} \emph{numeric} Pass QC? 1=Yes ; 0=No
\item \strong{PQPROERR} \emph{character} A. Processing Error(s) 1=Incorrect reconstruction algorithm ; 2=Incorrect filtering ; 3=Incorrect slice thickness ; 4=Incorrect voxel size ; 5=Attenuation Error ; 6=Normalization Error ; 7=Inadequate counts ; 8=Inadequate field of view ; 9=Incorrect acquisition protocol ; 10=Unfixable image artifact (e.g. attenuation, normalization, etc.) ; 99=Other (specify)
\item \strong{PQERRSP} \emph{character} Specify  
\item \strong{PQACTION} \emph{numeric} B. QC outcome 1=Request Reprocessing ; 3=Unuseable
\item \strong{PQISSUES} \emph{character} Imaging issues 1=Motion - none ; 2=Motion - mild ; 3=Motion - mild-to-moderate ; 4=Motion - moderate ; 5=Motion - moderate-to-severe ; 6=Top Cutoff - none ; 7=Top Cutoff - 0-3mm ; 8=Top Cutoff - 3-6mm ; 9=Top Cutoff - 6-9mm ; 10=Top Cutoff - >9mm ; 11=Bottom Cutoff - none ; 12=Bottom Cutoff - 0-3mm ; 13=Bottom Cutoff - 3-6mm ; 14=Bottom Cutoff - 6-9mm ; 15=Bottom Cutoff - >9mm ; 99=Other (specify)
\item \strong{PQISSOTH} \emph{numeric} Other, specify Range value: Inf, ..., -Inf
\item \strong{COMMENTS} \emph{character} Additional QC Comments  
\item \strong{QUARANTINE} \emph{numeric} Release from quarantine? 1=Yes ; 0=No
\item \strong{PAYSITE} \emph{numeric} Pay Site? 1=Yes ; 0=No
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
