% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PICSLASHS}
\alias{PICSLASHS}
\title{ASHS volume data}
\format{
A data frame with 3916 observations and 68 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PICSLASHS)
}
\description{
ASHS volume data data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNIGO, ADNI1, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor} Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Site and Unique participant ID  
\item \strong{RID} \emph{numeric} roster participant ID Range value: 21, ..., 10796
\item \strong{VISCODE} \emph{character} visit code  
\item \strong{VISCODE2} \emph{character} visit code translated  
\item \strong{EXAMDATE} \emph{ } Date of MRI scan Date: YYY-MM-DD
\item \strong{PROCESSDATE} \emph{ } Date image was procesed for data in this row Date: YYY-MM-DD
\item \strong{IMAGEUID_T1} \emph{numeric} Unique ID of T1 dicom Range value: 239888, ..., 11309435
\item \strong{IMAGEUID_T2} \emph{numeric} Unique ID of T2 dicom Range value: 239883, ..., 11306907
\item \strong{LEFT_CA1_VOL} \emph{numeric} left hemisphere ROI CA1 volume Range value: 43.03, ..., 5584.101
\item \strong{LEFT_CA1_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI CA1 appears Range value: 2, ..., 46
\item \strong{LEFT_CA2_VOL} \emph{numeric} left hemisphere ROI CA2 volume Range value: 0.234, ..., 9.46
\item \strong{LEFT_CA2_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI CA2 appears Range value: 1, ..., 10
\item \strong{LEFT_CA3_VOL} \emph{numeric} left hemisphere ROI CA3 volume Range value: 0.467, ..., 195.923
\item \strong{LEFT_CA3_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI CA3 appears Range value: 1, ..., 11
\item \strong{LEFT_DG_VOL} \emph{numeric} left hemisphere ROI dentate gyrus volume Range value: 8.85, ..., 2731.324
\item \strong{LEFT_DG_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI dentate gyrus appears Range value: 1, ..., 38
\item \strong{LEFT_MISC_VOL} \emph{numeric} left hemisphere ROI MISC volume Range value: 7.629, ..., 793.426
\item \strong{LEFT_MISC_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI MISC appears Range value: 3, ..., 35
\item \strong{LEFT_SUB_VOL} \emph{numeric} left hemisphere ROI Subiculum volume Range value: 24.585, ..., 1701.356
\item \strong{LEFT_SUB_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI Subiculum appears Range value: 1, ..., 43
\item \strong{LEFT_ERC_VOL} \emph{numeric} left hemisphere ROI Entorhinal cortex volume Range value: 0.916, ..., 785.523
\item \strong{LEFT_ERC_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI Entorhinal cortex appears Range value: 1, ..., 19
\item \strong{LEFT_BA35_VOL} \emph{numeric} left hemisphere ROI BA35 volume Range value: 38.452, ..., 995.853
\item \strong{LEFT_BA35_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI BA35 appears Range value: 2, ..., 25
\item \strong{LEFT_BA36_VOL} \emph{numeric} left hemisphere ROI BA36 volume Range value: 538.94, ..., 4696.956
\item \strong{LEFT_BA36_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI BA36 appears Range value: 5, ..., 27
\item \strong{LEFT_PHC_VOL} \emph{numeric} left hemisphere ROI parahippocampal cortex volume Range value: 0.305, ..., 5310.664
\item \strong{LEFT_PHC_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI parahippocampal cortex appears Range value: 1, ..., 18
\item \strong{LEFT_SULCUS_VOL} \emph{numeric} left hemisphere ROI sulcus volume Range value: 13.41, ..., 1256.594
\item \strong{LEFT_SULCUS_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI sulcus appears Range value: 1, ..., 41
\item \strong{LEFT_CA_VOL} \emph{numeric} left hemisphere ROI CA (CA1+CA2+CA3) volume Range value: 43.03, ..., 5629.267
\item \strong{LEFT_CA_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI CA (CA1+CA2+CA3) appears Range value: 2, ..., 46
\item \strong{LEFT_HIPP_VOL} \emph{numeric} left hemisphere ROI CA1+CA2+CA3+DG volume Range value: 51.88, ..., 8125.907
\item \strong{LEFT_HIPP_NS} \emph{numeric} Number of coronal slices in which left hemisphere ROI CA1+CA2+CA3+DG appears Range value: 2, ..., 46
\item \strong{RIGHT_CA1_VOL} \emph{numeric} right hemisphere ROI CA1 volume Range value: 188.599, ..., 6692.515
\item \strong{RIGHT_CA1_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI CA1 appears Range value: 5, ..., 50
\item \strong{RIGHT_CA2_VOL} \emph{numeric} right hemisphere ROI CA2 volume Range value: 0.305, ..., 19.531
\item \strong{RIGHT_CA2_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI CA2 appears Range value: 1, ..., 11
\item \strong{RIGHT_CA3_VOL} \emph{numeric} right hemisphere ROI CA3 volume Range value: 0.234, ..., 171.509
\item \strong{RIGHT_CA3_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI CA3 appears Range value: 1, ..., 10
\item \strong{RIGHT_DG_VOL} \emph{numeric} right hemisphere ROI dentate gyrus volume Range value: 67.293, ..., 5831.613
\item \strong{RIGHT_DG_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI dentate gyrus appears Range value: 3, ..., 41
\item \strong{RIGHT_MISC_VOL} \emph{numeric} right hemisphere ROI MISC volume Range value: 2.136, ..., 552.064
\item \strong{RIGHT_MISC_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI MISC appears Range value: 2, ..., 44
\item \strong{RIGHT_SUB_VOL} \emph{numeric} right hemisphere ROI Subiculum volume Range value: 9.155, ..., 2147.22
\item \strong{RIGHT_SUB_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI Subiculum appears Range value: 2, ..., 46
\item \strong{RIGHT_ERC_VOL} \emph{numeric} right hemisphere ROI Entorhinal cortex volume Range value: 1.221, ..., 2073.673
\item \strong{RIGHT_ERC_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI Entorhinal cortex appears Range value: 1, ..., 24
\item \strong{RIGHT_BA35_VOL} \emph{numeric} right hemisphere ROI BA35 volume Range value: 2.747, ..., 4026.19
\item \strong{RIGHT_BA35_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI BA35 appears Range value: 2, ..., 28
\item \strong{RIGHT_BA36_VOL} \emph{numeric} right hemisphere ROI BA36 volume Range value: 4.883, ..., 8778.09
\item \strong{RIGHT_BA36_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI BA36 appears Range value: 1, ..., 29
\item \strong{RIGHT_PHC_VOL} \emph{numeric} right hemisphere ROI parahippocampal cortex volume Range value: 0.61, ..., 5197.762
\item \strong{RIGHT_PHC_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI parahippocampal cortex appears Range value: 1, ..., 18
\item \strong{RIGHT_SULCUS_VOL} \emph{numeric} right hemisphere ROI sulcus volume Range value: 3.505, ..., 2192.081
\item \strong{RIGHT_SULCUS_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI sulcus appears Range value: 2, ..., 44
\item \strong{RIGHT_CA_VOL} \emph{numeric} right hemisphere ROI CA (CA1+CA2+CA3) volume Range value: 188.599, ..., 6752.635
\item \strong{RIGHT_CA_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI CA (CA1+CA2+CA3) appears Range value: 5, ..., 50
\item \strong{RIGHT_HIPP_VOL} \emph{numeric} right hemisphere ROI CA1+CA2+CA3+DG volume Range value: 276.794, ..., 12584.248
\item \strong{RIGHT_HIPP_NS} \emph{numeric} Number of coronal slices in which right hemisphere ROI CA1+CA2+CA3+DG appears Range value: 5, ..., 50
\item \strong{ICV} \emph{numeric} Intercranial volume derived from T1 image Range value: 541618.623, ..., 1944066.216
\item \strong{SLICE_THICKNESS} \emph{numeric} Coronal slice thickness Range value: 0.983051, ..., 14
\item \strong{T2_QC_R} \emph{character} manual QC score for right hemisphere image quality -1=no image ; 0=Unusable (substructures and the SRLM are not identifiable) or image has motion and looks blurry/wiggly ; 1=Bad Quality (SRLM is hard to identify and substructures are blurry) or image has motion ; 2=Adequate Quality (some substructures are clear, SRLM is mostly visible) ; 3=intermediate quality between 2 & 4 ; 4=Good Quality (most substructures are clear, SRLM is continuously visible)
\emph{Decoded Value: }\strong{Yes}
\item \strong{T2_QC_L} \emph{character} manual QC score for left hemisphere image quality -1=no image ; 0=Unusable (substructures and the SRLM are not identifiable) or image has motion and looks blurry/wiggly ; 1=Bad Quality (SRLM is hard to identify and substructures are blurry) or image has motion ; 2=Adequate Quality (some substructures are clear, SRLM is mostly visible) ; 3=intermediate quality between 2 & 4 ; 4=Good Quality (most substructures are clear, SRLM is continuously visible)
\emph{Decoded Value: }\strong{Yes}
\item \strong{T2_QC_COMMENT} \emph{character} QC comment If value is "QC from MAYOADIRL_MRI_IMAGEQC", QC rating mapped from MAYOADIRL scale. MAYOADIRL scale is 1: Excellent, 2: Good, 3: Fair, 4: Unusable, -1: Not evaluated. MAYOADIRL scale mapped to PICSLASHS scale is -1:-1, 4:0, 3:2, 2:3, 1:4. Any comments from sheet MAYOADIRL_MRI_IMAGEQC included in comment.
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{imaging}
