% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POSAMYDISC}
\alias{POSAMYDISC}
\title{Post-Amyloid PET Disclosure}
\format{
A data frame with 578 observations and 28 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(POSAMYDISC)
}
\description{
Post-Amyloid PET Disclosure data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNIGO, ADNI2, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10878
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{RECALL} \emph{character} Which of the following best describes your result? 0=Not Elevated ; 1=Elevated ; 2=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{RISKAD} \emph{character} For CN and MCI only: Did your amyloid PET scan result indicate that you have an increased risk for developing Alzheimer's disease dementia? Would you say___ 0=No, my amyloid PET scan result did not indicate an increased risk for Alzheimer's disease dementia ; 1=Yes, my amyloid PET scan result indicated an increased risk for Alzheimer's disease dementia ; 2=Not applicable (For DEM participants only) ; 99=I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{CONSAD} \emph{character} For DEM only: Was your amyloid PET scan result consistent with Alzheimer's disease? 0=No ; 1=Yes ; 2=Not applicable (For CN/MCI participants only) ; 99=I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{DEVDEMAD} \emph{numeric} On a scale of 0-100\%, where 0\% is no chance and 100\% is definite, what do you think is your chance of developing dementia caused by Alzheimer's disease in your lifetime? Range value: 0, ..., 100
\item \strong{DEMADLVL} \emph{character} Would you say that your risk of developing dementia caused by Alzheimer's disease is very low (1), low (2), average (3), high (4), or very high (5)? 1=very low (1) ; 2=low (2) ; 3=average (3) ; 4=high (4) ; 5=very high (5) ; 6=Not applicable (For DEM participants only)
\emph{Decoded Value: }\strong{Yes}
\item \strong{DEMADAGE} \emph{character} What do you think your chances are of developing dementia caused by Alzheimer's disease compared to others your age? Much lower (1), A little lower (2), The same (3), A little higher (4), Much higher (5) 1=much lower (1) ; 2=a little lower (2) ; 3=the same (3) ; 4=a little higher (4) ; 5=much higher (5) ; 6=Not applicable (For DEM participants only)
\emph{Decoded Value: }\strong{Yes}
\item \strong{USEFUL} \emph{character} Learning my amyloid result was useful to me. 1=Strongly disagree ; 2=Disagree ; 3=Neutral ; 4=Agree ; 5=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{REGRET} \emph{character} I regret learning my amyloid result. 1=Strongly disagree ; 2=Disagree ; 3=Neutral ; 4=Agree ; 5=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISITSAT} \emph{character} Overall, how satisfied were you with the visit where you learned your amyloid brain PET scan results? 1=Very dissatisfied ; 2=Dissatisfied ; 3=Neither satisfied nor dissatisfied ; 4=Satisfied ; 5=Very satisfied
\emph{Decoded Value: }\strong{Yes}
\item \strong{CLININTER} \emph{character} The clinician helped me understand what the elevated and not elevated results meant. 1=Strongly disagree ; 2=Disagree ; 3=Neutral ; 4=Agree ; 5=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{ELAV} \emph{character} I understood the meaning of an elevated amyloid result. 1=Strongly disagree ; 2=Disagree ; 3=Neutral ; 4=Agree ; 5=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{NONELAV} \emph{character} I understood the meaning of a not elevated amyloid result. 1=Strongly disagree ; 2=Disagree ; 3=Neutral ; 4=Agree ; 5=Strongly agree
\emph{Decoded Value: }\strong{Yes}
\item \strong{SHARE} \emph{character} Did you share your amyloid result with any of the following people? 1=Your ADNI study partner ; 10=Your financial advisor/banker ; 11=Your boss, supervisor, manager, or any person you report to in an employment or volunteer position ; 12=Other ; 2=Your primary care physician ; 3=Other health care providers ; 4=Your children ; 5=Your sibling(s) ; 6=Your life partner/spouse ; 7=Other family members or loved one(s) you have close relationships with ; 8=Spiritual or faith-based advisor (e.g., priest, rabbi, etc.) ; 9=Your attorney
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 4453, ..., 128107
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{enrollment}
