% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PREAMYDISC}
\alias{PREAMYDISC}
\title{Pre-Amyloid PET Disclosure}
\format{
A data frame with 806 observations and 25 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PREAMYDISC)
}
\description{
Pre-Amyloid PET Disclosure data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI2, ADNIGO, ADNI3, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_bl, 4_init, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{PETLEARN} \emph{character} Have you ever learned the result of a PET (positron emission tomography) scan that measures brain amyloid? This may also have been referred to as an "amyloid PET scan" or a "brain amyloid scan." 0=No ; 1=Yes ; 99=I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{PETRESULT} \emph{character} If yes, what was your amyloid PET scan result? 0=Not Elevated ; 1=Elevated ; 99=I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{RISKAD} \emph{character} For CN/MCI: If yes, did your amyloid PET scan result indicate that you have an increased risk for developing Alzheimer's disease dementia? 0=No, my amyloid PET scan result did not indicate an increased risk for Alzheimer's disease dementia. ; 1=Yes, my amyloid PET scan result indicated an increased risk for Alzheimer's disease dementia. ; 2=Not applicable (For DEM) ; 99=I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{CONSAD} \emph{character} For dementia: If yes, was your amyloid PET scan result consistent with Alzheimer's disease dementia? 0=No, my amyloid PET scan result was not consistent with Alzheimer's disease ; 1=Yes, my amyloid PET scan result was consistent with Alzheimer's disease ; 2=Not applicable (For CN/MCI) ; 99=I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{DESCBHC} \emph{character} Which of the terms best describes what you've been told about your brain health and cognition? 1="I am cognitively unimpaired" ; 2="I have mild cognitive impairment" ; 3="I have dementia" ; 4=Not applicable (For CN/MCI) ; 99=Don't remember/ Don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{RESULT} \emph{character} What do you think the result of your amyloid PET scan will be? 0=Not Elevated ; 1=Elevated ; 99=I don't know
\emph{Decoded Value: }\strong{Yes}
\item \strong{CONFLVL} \emph{numeric} On a scale of 0-100\%, how confident are you that's what your result will be? Range value: 0, ..., 100
\item \strong{DEVDEMAD} \emph{numeric} On a scale of 0-100\%, where 0\% is no chance and 100\% is definite, what do you think is your chance of developing dementia caused by Alzheimer's disease in your lifetime? Range value: 0, ..., 100
\item \strong{DEMADLVL} \emph{character} Would you say that your risk of developing dementia caused by Alzheimer's disease is very low (1), low (2), average (3), high (4), or very high (5)? 1=very low (1) ; 2=low (2) ; 3=average (3) ; 4=high (4) ; 5=very high (5) ; 6=Not applicable (For DEM)
\emph{Decoded Value: }\strong{Yes}
\item \strong{DEMADAGE} \emph{character} What do you think your chances are of developing dementia caused by Alzheimer's disease compared to others your age? Much lower (1), A little lower (2), The same (3), A little higher (4), Much higher (5) 1=much lower (1) ; 2=a little lower (2) ; 3=the same (3) ; 4=a little higher (4) ; 5=much higher (5) ; 6=Not applicable (For DEM)
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 1285, ..., 128000
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{enrollment}
