% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PSS}
\alias{PSS}
\title{Perceived Stress Scale 10 Item}
\format{
A data frame with 1254 observations and 27 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PSS)
}
\description{
Perceived Stress Scale 10 Item data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNI2, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{DONE} \emph{character} Was assessment/procedure done? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 1=Participant Unable for Cognitive Reasons ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS1} \emph{character} 1. In the last month, how often have you been upset because of something that happened unexpectedly? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS2} \emph{character} 2. In the last month, how often have you felt that you were unable to control the important things in your life? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS3} \emph{character} 3. In the last month, how often have you felt nervous and stressed? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS4} \emph{character} 4. In the last month, how often have you felt confident about your ability to handle your personal problems? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS5} \emph{character} 5. In the last month, how often have you felt that things were going your way? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS6} \emph{character} 6. In the last month, how often have you found that you could not cope with all the things that you had to do? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS7} \emph{character} 7. In the last month, how often have you been able to control irritations in your life? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS8} \emph{character} 8. In the last month, how often have you felt that you were on top of things? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS9} \emph{character} 9. In the last month, how often have you been angered because of things that happened that were outside of your control? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{PSS10} \emph{character} 10. In the last month, how often have you felt difficulties were piling up so high that you could not overcome them? 0=Never ; 1=Almost Never ; 2=Sometimes ; 3=Fairly Often ; 4=Very Often
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 73, ..., 128211
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
