% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PTDEMOG}
\alias{PTDEMOG}
\title{Participant Demographic Information}
\format{
A data frame with 6219 observations and 85 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PTDEMOG)
}
\description{
Participant Demographic Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{PTSOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTGENDER} \emph{character} 1. What sex was the participant assigned at birth, on their original birth certificate 1=Male ; 2=Female
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTDOB} \emph{character} 4. Participant Date of Birth  
\item \strong{PTDOBYY} \emph{numeric} 2b. Participant Year of Birth 1900..1980
\item \strong{PTHAND} \emph{character} 5. Participant Handedness 1=Right ; 2=Left
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTMARRY} \emph{character} 6. Participant Marital Status 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 1=Married ; 2=Widowed ; 3=Divorced ; 4=Never married ; 5=Unknown
\item \emph{ADNI4:} 1=Married ; 2=Widowed ; 3=Divorced ; 4=Never married ; 5=Unknown ; 6=Domestic Partnership
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTEDUCAT} \emph{numeric} 7. Participant Education 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 0..20
\item \emph{ADNI3, ADNI4} 
}
\item \strong{PTWORKHS} \emph{character} 7a. Does the participant have a work history sufficient to exclude mental retardation? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTWORK} \emph{character} 8a. Primary occupation during most of adult life? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3} 
\item \emph{ADNI4:} 1=Professional and Higher Executive Occupations, Chief Executives ; 2=Middle Professional Occupations/ Small Business Owners ; 3=Managers ; 4=Support Personnel, Drafters, Technicians ; 5=Arts, Design, Entertainment, Sports Occupations – Non-Professional ; 6=Aides / Assistants / Clerks ; 7=Laborers ; 8=Other - Never Employed ; 9=Missing/Unknown
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTNOTRT} \emph{character} 9. Participant Retired? 
\itemize{
\item \emph{ADNI1, ADNIGO:} 1=Yes ; 0=No
\item \emph{ADNI2:} 1=Yes ; 0=No ; 2=Not Applicable
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes ; 2=Not Applicable
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTRTYR} \emph{numeric} Retirement Date Range value: 1926, ..., 9999
\item \strong{PTHOME} \emph{character} 10. Type of Participant residence 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 1=House (owned or rented) ; 2=Condo/Co-op (owned) ; 3=Apartment (rented) ; 4=Mobile Home ; 5=Retirement Community ; 6=Assisted Living ; 7=Skilled Nursing Facility ; 8=Other (specify)
\item \emph{ADNI4:} 10=House (owned) ; 2=Condo/Co-op (owned) ; 3=Apartment (rented) ; 4=Mobile Home ; 5=Retirement Community ; 6=Assisted Living ; 7=Skilled Nursing Facility ; 8=Other (specify) ; 9=House (rented)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTTLANG} \emph{character} 11. Language to be used for testing the Participant: 1=English ; 2=Spanish
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTPLANG} \emph{character} 12. What is the participant's primary language? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 1=English ; 2=Spanish ; 3=Other (specify)
\item \emph{ADNI4:} 1=English ; 2=Spanish ; 3=Other
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTADBEG} \emph{numeric} 11b. Year of onset of Alzheimer's disease symptoms (best estimate) 
\itemize{
\item \emph{ADNI1:} 1985..2007
\item \emph{ADNIGO:} 1985..2012
}
\item \strong{PTCOGBEG} \emph{numeric} 23a. Year of onset of cognitive symptoms (best estimate) 
\itemize{
\item \emph{ADNI2:} 9999=NA - Cognitively Normal ; 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012
\item \emph{ADNI3:} 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012 ; 2013=2013 ; 2014=2014 ; 2015=2015 ; 2016=2016 ; 2017=2017 ; 2018=2018 ; 2019=2019 ; 2020=2020 ; 2021=2021 ; 2022=2022 ; 9999=NA - Cognitively Normal
\item \emph{ADNI4:} 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012 ; 2013=2013 ; 2014=2014 ; 2015=2015 ; 2016=2016 ; 2017=2017 ; 2018=2018 ; 2019=2019 ; 2020=2020 ; 2021=2021 ; 2022=2022 ; 2023=2023 ; 2024=2024 ; 2025=2025 ; 9999=NA - Cognitively Normal
}
\item \strong{PTADDX} \emph{numeric} 23b. Year of Alzheimer's Disease diagnosis 
\itemize{
\item \emph{ADNI2:} 9999=NA - Not Diagnosed with AD ; 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012
\item \emph{ADNI3:} 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012 ; 2013=2013 ; 2014=2014 ; 2015=2015 ; 2016=2016 ; 2017=2017 ; 2018=2018 ; 2019=2019 ; 2020=2020 ; 2021=2021 ; 2022=2022 ; 9999=NA - Not Diagnosed with AD
\item \emph{ADNI4:} 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012 ; 2013=2013 ; 2014=2014 ; 2015=2015 ; 2016=2016 ; 2017=2017 ; 2018=2018 ; 2019=2019 ; 2020=2020 ; 2021=2021 ; 2022=2022 ; 2023=2023 ; 2024=2024 ; 2025=2025 ; 9999=NA - Not Diagnosed with AD
}
\item \strong{PTETHCAT} \emph{character} 24. Ethnic Category 1=Hispanic or Latino ; 2=Not Hispanic or Latino ; 3=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTRACCAT} \emph{character} 25. Racial Categories 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2, ADNI3:} 1=American Indian or Alaskan Native ; 2=Asian ; 3=Native Hawaiian or Other Pacific Islander ; 4=Black or African American ; 5=White ; 6=More than one race ; 7=Unknown
\item \emph{ADNI4:} 1=American Indian or Alaskan Native ; 2=Asian ; 3=Native Hawaiian or Other Pacific Islander ; 4=Black or African American ; 5=White ; 7=Unknown ; 8=Native Hawaiian ; 9=Other Pacific Islander
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTIDENT} \emph{character} 2. Which term(s) best describes the participant's current gender identity? 1=Man ; 2=Woman ; 3=Transgender man ; 4=Transgender woman ; 5=Non-binary/Gender queer ; 6=Two-Spirit (if respondent is AIAN) ; 7=I use a different term ; 8=Don't know ; 9=Prefer not to say
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTORIENT} \emph{numeric} 3. Which term(s) best describes the participants current sexual orientation 1=Gay, Lesbian ; 2=Straight ; that is, not gay or lesbian, etc. ; 3=Bisexual ; 4=None of these options describe me, and I’d like to see additional options
\item \strong{PTORIENTOT} \emph{character} 3a. Are any of the following options a closer description of how you think of yourself (sexual orientation)? 1=Queer ; 10=I mean something else, please specify: ; 11=Prefer not to answer ; 2=Omnisexual, sapiosexual, or pansexual ; 3=Asexual ; 4=Two-spirit ; 5=Have not figured out or are in the process of figuring out my sexuality ; 6=Mostly straight, but sometimes attracted to people of my own sex ; 7=Do not think of myself as having sexuality ; 8=Do not use labels to identify myself ; 9=Don’t know the answer
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTENGSPK} \emph{character} 13. Do you consider yourself a monolingual English speaker? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTNLANG} \emph{character} 14. What is your native language? 1=English ; 2=Spanish ; 3=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTENGSPKAGE} \emph{numeric} 14a. For participant's whose native language is not English, how old where you when you learned to speak English? Range value: 0, ..., 46
\item \strong{PTCLANG} \emph{character} 15. What was your primary-caretaker's preferred language? 1=English ; 2=Spanish ; 3=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGSP} \emph{character} 16. What is the language you prefer to speak? 1=English ; 2=Spanish ; 3=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGWR} \emph{character} 17. What is the language you prefer to write? 1=English ; 2=Spanish ; 3=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTSPTIM} \emph{numeric} 18. What percentage of the time during an average day do you speak English? Range value: 0, ..., 100
\item \strong{PTSPOTTIM} \emph{numeric} 19. What percentage of the time during an average day do you speak another language? Range value: 0, ..., 100
\item \strong{PTLANGPR1} \emph{character} Language 1 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGSP1} \emph{character} Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGRD1} \emph{character} Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGWR1} \emph{character} Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGUN1} \emph{character} Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGPR2} \emph{character} Language 2 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGSP2} \emph{character} Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGRD2} \emph{character} Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGWR2} \emph{character} Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGUN2} \emph{character} Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGPR3} \emph{character} Language 3 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGSP3} \emph{character} Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGRD3} \emph{character} Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGWR3} \emph{character} Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGUN3} \emph{character} Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGPR4} \emph{character} Language 4 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGSP4} \emph{character} Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGRD4} \emph{character} Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGWR4} \emph{character} Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGUN4} \emph{character} Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGPR5} \emph{character} Language 5 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGSP5} \emph{character} Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGRD5} \emph{character} Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGWR5} \emph{character} Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGUN5} \emph{character} Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGPR6} \emph{character} Language 6 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGSP6} \emph{character} Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGRD6} \emph{character} Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGWR6} \emph{character} Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGUN6} \emph{character} Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTLANGTTL} \emph{numeric} 21. Total number of languages spoken (based on table above) Range value: 1, ..., 7
\item \strong{PTETHCATH} \emph{character} 24a. If Hispanic or Latino is selected, please provide country/region of origin. : 1=Mexican, Mexican Am., Chicano ; 10=Other Latin culture or origin, please specify: ; 2=Puerto Rican ; 3=Cuban ; 4=Another Hispanic, Latino, or Spanish origin ; 5=Unknown Hispanic or Latino origin ; 6=Dominican ; 7=Mexican ; 8=South American ; 9=Central American
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTASIAN} \emph{character} 25a. If Asian is selected, please provide country/region of origin. 1=Chinese ; 2=Filipino ; 3=Asian Indian ; 4=Vietnamese ; 5=Korean ; 6=Japanese ; 7=Other Asian
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTOPI} \emph{character} 25b. If Other Pacific Islander is selected, please provide further detail 1=Samoan ; 2=Chamorro ; 3=Other Pacific islander – for example, Tongan, Fijian, Marshallese, etc
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTBORN} \emph{character} 26. Were you born in the United States/Canada? 0=No ; 1=Yes ; 2=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTBIRPL} \emph{character} 26a. If NO to 26: Where were you born? 1=Puerto Rico ; 2=Dominican Republic ; 3=Cuba ; 4=Mexico ; 5=South America ; 6=Central America ; 7=Unknown ; 8=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTIMMAGE} \emph{numeric} 26b. If NO to 26: How old were you when you moved to the United States/Canada? Range value: 0, ..., 57
\item \strong{PTIMMWHY} \emph{character} 26c. What motivated you to migrate to the United States/Canada? 1=To escape our past or future persecution based on race, religion, nationality, and/or membership in a particular social group or political opinion ; 10=For jobs and business opportunities ; 11=Marriage ; 12=To meet new people from different backgrounds ; 13=To make money to send back to my family back home in my home country ; 14=Other ; 2=To escape environmental conflict or violence (i.e., war or escaping gang violence) ; 3=To escape domestic/intimate partner violence ; 4=To find refuge after being displaced due to environmental factors (i.e., natural disasters, erosion, and other environmental factors caused by climate change) ; 5=To seek better healthcare services ; 6=To seek better living conditions ; 7=To offer more opportunities for my children ; 8=Family reunification ; 9=For better educational opportunities for my family and/or me
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTBIRPR} \emph{character} 27. Were your parents born in the United States/Canada? 0=No ; 1=Yes ; 2=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTBIRGR} \emph{character} 28. Were your grandparents born in the United States/Canada? 0=No ; 1=Yes ; 2=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 12, ..., 186877
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{subject_characteristics}
