% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PUBLICITY}
\alias{PUBLICITY}
\title{Recruitment Tracking}
\format{
A data frame with 1382 observations and 23 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(PUBLICITY)
}
\description{
Recruitment Tracking data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI3, ADNI2, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10900
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_sc, 4_init
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{REFERRAL} \emph{character} Do you have a referral code for this participant? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{HEARD} \emph{character} 1.  How did you hear about the study? 
\itemize{
\item \emph{ADNI4:} 1=I received something in the mail. (Use the referral code or the phone number on the mailer as the referral code above.) ; 10=My community group newsletter/bulletin (Provide the group name below.) ; 11=Saw a flyer (Let us know where you saw the flyer below.) ; 12=Saw a printed advertisement (Share the name of the publication where you saw the ad below.) ; 13=Read a printed news article, report (Include the name of the publication and let us know if you are a subscriber below.) ; 14=Picked up a brochure (Let us know where you got the brochure below.) ; 15=Saw a video (Let us know where you saw the video below.) ; 16=Reading materials in a wait room (Let us know where you were waiting below.) ; 17=I received a phone call ; 2=I have received an email. (Include referral code above.) ; 3=I attended a research event/talk presentation. (Specify the event/talk below.) ; 4=A speaker came to one of my meetings/events. (Name speaker and meeting) ; 5=Listening to the radio. (Specify radio show below and if you are a regular listener.) ; 6=Watching TV (Specify below what you were watching and if you are a regular viewer.) ; 7=Website (Specify the website below.) ; 8=Facebook (Specify whether it was a Facebook ad or something a friend shared with you.) ; 9=Other social media (Let us know more below.) ; 99=Other (Please provide any additional information you can regarding how you heard about the study if applicable.)
\item \emph{ADNI4:} 1=I received something in the mail. Specify below: referral code on mailer. ; 10=I saw it in my community group newsletter/bulletin. Specify below: name or type of community group and newsletter/bulletin (including if in person or virtual). ; 11=I saw a flyer. Specify below: referral code on flyer and/or name or type of place you saw it. ; 12=I saw another type of printed advertisement. Specify below: referral code on the printed advertisement and/or type of printed advertisement. ; 13=I saw a news article/report. Specify below: Online or in print. Specify name of publication and if you are a subscriber. ; 14=I picked up a brochure. Specify below: referral code on brochure and/or name or type of place you saw it. ; 15=I saw a video (non-social media). Specify below: describe video and where you saw it. ; 16=I saw it in reading materials in a waiting room. Specify below: referral code in the reading material and where you were waiting. ; 17=I received a phone call. Specify below: organization or type of organization who called. ; 18=I saw it on YouTube. Specify below: describe images/words used in the video. ; 19=I saw it on Instagram. Specify below: describe images/video/words used. ; 2=I received an email. Specify below: name of organization who emailed you and referral code in email. ; 20=I saw it on Twitter. Specify below: describe images/video/words used. ; 21=I saw it on another social media website. Specify below: name of social media site and describe images/video/words used. ; 22=I saw it on another non-social media website. Specify below: name of the website or type of website (e.g. website list of AD studies, etc.). ; 23=I attended a health fair. Specify below: name or type of event and where it was held (including if in person or virtual). ; 24=I attended a memory screening. Specify below: name or type of event and where it was held (including if in person or virtual). ; 3=I attended a research/educational event/talk presentation. Specify below: name or type of event/talk and where it was held (including if in person or virtual). ; 4=A speaker came to one of my meetings/events. Specify below: name of speaker, name or type of meeting/event, and where it was held (including if in person or virtual). ; 5=I heard about it on the radio. Specify below: name of radio show and if you are a regular listener. ; 6=I saw it on TV. Specify below: name of tv show and if you are a regular watcher. ; 8=I saw it on Facebook. Specify below: describe images/video/words used. ; 99=I heard about it some other way. Below, please provide any additional information you can regarding how you heard about the study if applicable.
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{WHOMHRD} \emph{character} 2.  From whom did you hear about the study? 
\itemize{
\item \emph{ADNI4:} 1=My general practitioner ; 10=Alzheimer’s Association or Trial Match ; 11=Alzheimer’s Prevention Initiative/Registry “API/R” ; 12=Clinical Trials.gov ; 13=NIH, NIA Alzheimer’s Disease Education and Referral Center (ADEAR) ; 14=The Center for Information and Study on Clinical Research Participation (CISCRP) ; 15=My alma mater ; 16=My trusted research institution ; 17=My community group ; 18=A trusted advisor ; 19=A news report ; 2=A family member’s general practitioner ; 20=A blogger ; 21=A trusted medical resource ; 3=My specialty doctor ; 4=A family member’s specialty doctor ; 5=An Alzheimer’s research doctor ; 6=An Alzheimer’s research team member ; 7=A friend ; 8=A family member ; 9=Brain Health Registry “BHR” ; 99=Other (please specify)
\item \emph{ADNI4:} 1=My general practitioner ; 10=Alzheimer’s Association or Trial Match ; 11=Alzheimer’s Prevention Initiative/Registry "API/R" ; 12=Clinical Trials.gov ; 13=NIH, NIA Alzheimer’s Disease Education and Referral Center (ADEAR) ; 14=The Center for Information and Study on Clinical Research Participation (CISCRP) ; 15=My alma mater ; 16=My trusted research institution ; 17=My community group ; 18=A trusted advisor ; 19=A news report ; 2=A family member’s general practitioner ; 20=A blogger ; 21=A trusted medical resource ; 22=Community health clinic ; 23=Religious/charitable organization ; 24=Senior center ; 25=Social services organization ; 3=My specialty doctor ; 4=A family member’s specialty doctor ; 5=An Alzheimer’s research doctor ; 6=An Alzheimer’s research team member ; 7=A friend ; 8=A family member ; 99=Other (please specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{HUBSITE} \emph{character} 2a. If yes, is your clinical site a Hub Site/Hub City? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTTYPE} \emph{character} 1. Is this participant a rollover? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{RECRUIT} \emph{character} 2. Was this participant a direct recruit from the site? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{REFCRL} \emph{character} 2b. If yes, was this participant referred to your site via the site Community Research Liaison (CRL) staff member? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{REFEBISU} \emph{character} 3. Was the participant referred to your clinical site via participation in Remote (online) ADNI cohorts (such as the Remote Digital cohort or the Remote Blood cohort)? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 318, ..., 128074
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1, v2
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{enrollment}
