% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QS}
\alias{QS}
\title{[DERIVED]  Derived  Questionnaires}
\format{
A data frame with 320021 observations and 20 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(QS)
}
\description{
Questionnaires derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier ; , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbrevation , with character options: QS
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{QSSEQ} \emph{integer} Sequence Number Range value: 1, ..., 414
\item \strong{QSGRPID} \emph{character} Group ID Study Protocol of Data Collection; , with character options: ADNI1, ADNI2, ADNIGO, ADNI3, ADNI4
\item \strong{QSTESTCD} \emph{character} Question Short Name  
\item \strong{QSTEST} \emph{character} Question Name  
\item \strong{QSCAT} \emph{character} Category of Question  
\item \strong{QSORRES} \emph{character} Finding in Original Units  
\item \strong{QSSTRESC} \emph{character} Character Result/Finding in Std Format  
\item \strong{QSSTRESN} \emph{numeric} Numeric Finding in Standard Units Range value: -500, ..., 996
\item \strong{QSSTAT} \emph{character} Completion Status , with character options: NOT DONE
\item \strong{QSREASND} \emph{character} Reason Not Performed , with character options: Unknown, Box not checked, Other Reason, Participant Unable for Cognitive Reasons, Participant is unable to complete the GDS based on the clinician's best judgement., Site Error, Participant Unable for Other Reasons (Physical, auditory, etc.), Box checked, Participant/Study Partner Refused, Equipment related issue
\item \strong{QSBLFL} \emph{character} Baseline Flag , with character options: Y
\item \strong{QSDRVFL} \emph{character} Derived Flag , with character options: Yes
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -2, ..., 49
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{EPOCH} \emph{character} Epoch , with character options: Follow-up, Baseline, Screening
\item \strong{QSDTC} \emph{Date} Date/Time of Finding Date: YYY-MM-DD
\item \strong{QSDY} \emph{integer} Study Day of Finding Range value: -365625, ..., 35075
}
}
\keyword{derived_dataset}
