% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RECADV}
\alias{RECADV}
\title{Adverse Events/Hospitalizations - Log}
\format{
A data frame with 19416 observations and 39 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RECADV)
}
\description{
Adverse Events/Hospitalizations - Log data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 42
\item \strong{AENUMBER} \emph{numeric} Adverse Event Number Range value: 1, ..., 384
\item \strong{AEHBLSYM} \emph{character} Check here if: 
\itemize{
\item \emph{ADNI1, ADNIGO:} 1=this symptom was reported on the Baseline Symptoms Checklist, but has worsened in chronicity or severity.
\item \emph{ADNI2:} 1=This symptom was reported on the Pre-Existing Symptoms Checklist, but has worsened in chronicity or severity.
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHONSDT} \emph{character} Onset Date  
\item \strong{AEHONSHR} \emph{numeric} Estimated Onset Time: Range value: 0, ..., 2359
\item \strong{AEHONGNG} \emph{character} Is the event ongoing? 
\itemize{
\item \emph{ADNI1, ADNIGO:} 1=Yes ; 0=No
\item \emph{ADNI2:} 1=Yes, event is ongoing ; 0=No, event resolved ; 2=Resolved with Sequelae
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHCSDT} \emph{character} Cease Date  
\item \strong{AEHCSHR} \emph{numeric} Estimated Cease Time: Range value: 0, ..., 2359
\item \strong{AEHCHRON} \emph{character} Chronicity 1= Single Occurrence ; 2= Intermittent ; 3= Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHSEVR} \emph{character} Severity 1= Mild ; 2= Moderate ; 3= Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHSERIO} \emph{character} Serious? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHSAE} \emph{character} Check here if: 1=SAE prior to Baseline Visit
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHREASN} \emph{character} Reason for Qualifying as Serious Adverse Event: 
\itemize{
\item \emph{ADNI1, ADNIGO} 
\item \emph{ADNI2:} 1=Life-Threatening ; 2=Prolongation of hospitalization ; 3=Hospitalization ; 4=Death ; 5=Congenital anomaly or birth defect ; 6=Persistent or significant disability/incapacity ; 7=Other medically important conditions
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHLIFE} \emph{character} Life-Threatening? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHIMG} \emph{character} Related to Imaging 1= Definitely ; 2= Possibly ; 3= Not Related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHLUMB} \emph{character} Related to Lumbar Puncture 1= Definitely ; 2= Possibly ; 3= Not Related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHAV45} \emph{character} Investigator Judgment of Relatedness to Florbetapir F18 1= Definitely ; 2= Possibly ; 3= Not Related
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHCMEDS} \emph{character} Was treatment required? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHINHOS} \emph{character} Did this event occur while the participant was being hospitalized for another event? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHPRO} \emph{character} If Yes, did this event prolong hospitalization? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHHOSP} \emph{character} If No, did this event require hospitalization? 3=No ; 2=Yes - Outpatient ; 1=Yes - Inpatient
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHOUTDT} \emph{character} If Outpatient, provide the date of visit  
\item \strong{AEHADMDT} \emph{character} If hospitalized, provide Admission Date  
\item \strong{AEHDISDT} \emph{character} Discharge Date  
\item \strong{AEHDEATH} \emph{character} Did this event result in death? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{AEHDTHDT} \emph{numeric} If death, date of death Range value: 2006, ..., 2018
\item \strong{AEHALZHI} \emph{character} Was diagnosis of Alzheimer's confirmed at autopsy? 0=No ; 1=Yes ; 2=No postmortem brain exam
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 22528
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{medical_history}
