% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RECBLLOG}
\alias{RECBLLOG}
\title{Documentation of Baseline Symptoms Log}
\format{
A data frame with 12848 observations and 21 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RECBLLOG)
}
\description{
Documentation of Baseline Symptoms Log data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 29
\item \strong{BSXSYMNO} \emph{character} Symptom Number 1=1. Nausea ; 2=2. Vomiting ; 3=3. Diarrhea ; 4=4. Constipation ; 5=5. Abdominal discomfort ; 6=6. Sweating ; 7=7. Dizziness ; 8=8. Low energy ; 9=9. Drowsiness ; 10=10. Blurred vision ; 11=11. Headache ; 12=12. Dry mouth ; 13=13. Shortness of breath ; 14=14. Coughing ; 15=15. Palpitations ; 16=16. Chest pain ; 17=17. Urinary discomfort ; 18=18. Urinary frequency ; 19=19. Ankle swelling ; 20=20. Musculoskeletal pain ; 21=21. Rash ; 22=22. Insomnia ; 23=23. Depressed mood ; 24=24. Crying ; 25=25. Elevated mood ; 26=26. Wandering ; 27=27. Fall ; 28=28. Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{BSXSEVER} \emph{character} Severity 1=Mild ; 2=Moderate ; 3=Severe
\emph{Decoded Value: }\strong{Yes}
\item \strong{BSXCHRON} \emph{character} Chronicity 1=Single occurrence ; 2=Intermittent ; 3=Persistent
\emph{Decoded Value: }\strong{Yes}
\item \strong{BSXONSET} \emph{character} Date of Onset  
\item \strong{BSXONSETYR_DRVD} \emph{numeric} Year of Onset (Derived from BSXONSET) Range value: 1900, ..., 2013
\item \strong{BSXCONTD} \emph{character} Is the symptom ongoing? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{BSXCEASE} \emph{character} Date Ceased  
\item \strong{BSXCEASEYR_DRVD} \emph{numeric} Year Ceased (Derived from BSXCEASE) Range value: 1970, ..., 2016
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 14658
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{medical_history}
