% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RECCMEDS}
\alias{RECCMEDS}
\title{Concurrent Medications Log}
\format{
A data frame with 77466 observations and 38 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RECCMEDS)
}
\description{
Concurrent Medications Log data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 104
\item \strong{CMMED} \emph{character} Concurrent Medication  
\item \strong{CMMEDO} \emph{character} If medication not found in search list, hand enter for approval  
\item \strong{CMMEDID} \emph{numeric} medlist ID Range value: 1, ..., 1544
\item \strong{CMDOSE} \emph{character} Dose  
\item \strong{CMUNITSID} \emph{character}    
\item \strong{CMUNITS} \emph{numeric} Units 
\itemize{
\item \emph{ADNI2:} 1=drop (gtt) ; 2=gram (gm) ; 3=international units (IU) ; 4=microgram (mcg) ; 5=milligram (mg) ;  6=milliliter (ml) ; 7=percent (\%) ; 8=puff ; 9=spray (spray) ; 10=tablespoon (15mL) (tbsp) ; 11=tablets (tabs) ; 12=teaspoon (5mL) (tsp) ; 13=other (OT)
\item \emph{ADNI3:} 1=drop (gtt) ; 2=gram (gm) ; 3=international units (IU) ; 4=microgram (mcg) ; 5=milligram (mg) ; 6=milliliter (ml) ; 7=percent (\%) ; 8=puff ; 9=spray (spray) ; 10=tablespoon (15mL) (tbsp) ; 11=tablets (tabs) ; 12=teaspoon (5mL) (tsp) ; 13=other (OT)
\item \emph{ADNI4:} 1=drop (gtt) ; 10=tablespoon (15mL) (tbsp) ; 11=tablets (tabs) ; 12=teaspoon (5mL) (tsp) ; 13=other (OT) ; 2=gram (gm) ; 3=international units (IU) ; 4=microgram (mcg) ; 5=milligram (mg) ; 6=milliliter (ml) ; 7=percent (\%) ; 8=puff ; 9=spray (spray)
}
\item \strong{CMUNITO} \emph{character} If other, specify:  
\item \strong{CMFREQID} \emph{character}    
\item \strong{CMFREQNC} \emph{numeric} Frequency 
\itemize{
\item \emph{ADNI1} 
\item \emph{ADNI2:} 1=2 times a day (bid) ; 2=3 times a day (tid) ; 3=as needed (prn) ; 4=continuous (cont) ; 5=every 12 hours (q12h) ; 6=every 4 hours (q4h) ; 7=every morning (qam) ; 8=every night at bedtime (qhs) ; 9=every other day (qod) ; 10= once a day (daily) (qd) ; 11=once (1x) ; 12=once a month (1xm) ; 13=other (OT)
\item \emph{ADNI3:} 1=2 times a day (bid) ; 2=3 times a day (tid) ; 3=as needed (prn) ; 4=continuous (cont) ; 5=every 12 hours (q12h) ; 6=every 4 hours (q4h) ; 7=every morning (qam) ; 8=every night at bedtime (qhs) ; 9=every other day (qod) ; 10=once a day (daily) (qd) ; 11=once (1x) ; 12=once a month (1xm) ; 13=other (OT)
\item \emph{ADNI4:} 1=2 times a day (bid) ; 10=once a day (daily) (qd) ; 11=once (1x) ; 12=once a month (1xm) ; 13=other (OT) ; 2=3 times a day (tid) ; 3=as needed (prn) ; 4=continuous (cont) ; 5=every 12 hours (q12h) ; 6=every 4 hours (q4h) ; 7=every morning (qam) ; 8=every night at bedtime (qhs) ; 9=every other day (qod)
}
\item \strong{CMFREQNCO} \emph{character} If other, specify:  
\item \strong{CMROUTEID} \emph{character}    
\item \strong{CMROUTE} \emph{numeric} Route 
\itemize{
\item \emph{ADNI1} 
\item \emph{ADNI2:} 1=both ears (au) ; 2=each eye (oe) ; 3=intramuscular (im) ; 4=intravenous (iv) ; 5=nasal (nas, ns) ; 6=oral (po) ; 7=subcutaneous (sc, sq) ; 8=sublingual (sl, subling) ; 11=topical (top, tp) ; 9=transdermal (td, tdrm) ; 10=other (OT)
\item \emph{ADNI3:} 1=both ears (au) ; 2=each eye (oe) ; 3=intramuscular (im) ; 4=intravenous (iv) ; 5=nasal (nas, ns) ; 6=oral (po) ; 7=subcutaneous (sc, sq) ; 8=sublingual (sl, subling) ; 9=transdermal (td, tdrm) ; 10=other (OT) ; 11=topical (top, tp)
\item \emph{ADNI4:} 1=both ears (au) ; 10=other (OT) ; 11=topical (top, tp) ; 2=each eye (oe) ; 3=intramuscular (im) ; 4=intravenous (iv) ; 5=nasal (nas, ns) ; 6=oral (po) ; 7=subcutaneous (sc, sq) ; 8=sublingual (sl, subling) ; 9=transdermal (td, tdrm)
}
\item \strong{CMROUTEO} \emph{character} If other, specify:  
\item \strong{CMREASON} \emph{numeric} Reason Prescribed 
\itemize{
\item \emph{ADNI1} 
\item \emph{ADNIGO, ADNI2, ADNI3, ADNI4:} 1=Adverse Event ; 2=Therapeutic Use ; 3=Prophylaxis/Non-therapeutic Use
}
\item \strong{CMREASONO} \emph{character} If other, specify:  
\item \strong{CMEVNUM} \emph{character} If Adverse Event, enter event number:  
\item \strong{CMBGN} \emph{character} Date Began  
\item \strong{CMBGNYR_DRVD} \emph{numeric} Year Began (Derived from CMBGN) Range value: 1930, ..., 2924
\item \strong{CMCONT} \emph{numeric} Is the medication continuing? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\item \strong{CMEND} \emph{character} Date Ended  
\item \strong{CMENDYR_DRVD} \emph{numeric} Year Ended (Derived from CMEND) Range value: 1985, ..., 2918
\item \strong{CMCOMM} \emph{character} Comments  
\item \strong{ID} \emph{numeric} Record ID Range value: 6, ..., 186563
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\item \strong{HAS_QC_ERROR} \emph{numeric} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{medical_history}
