% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RECFHQ}
\alias{RECFHQ}
\title{Family History Questionnaire Subtable}
\format{
A data frame with 7407 observations and 16 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RECFHQ)
}
\description{
Family History Questionnaire Subtable data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 4, ..., 7474
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5296
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 21
\item \strong{FHQGNDR} \emph{character} Gender 1=Male ; 2=Female
\emph{Decoded Value: }\strong{Yes}
\item \strong{FHQSIB} \emph{character} Dementia 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No ; 2=Don't Know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{FHQSIBAD} \emph{character} Alzheimer's Disease 
\itemize{
\item \emph{ADNI1:} 1=Yes ; 0=No
\item \emph{ADNIGO, ADNI2:} 1=Yes ; 0=No ; 2=Don't Know
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{subject_characteristics}
