% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RECMHIST}
\alias{RECMHIST}
\title{Medical History}
\format{
A data frame with 30810 observations and 18 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RECMHIST)
}
\description{
Medical History data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, f, m48, m36, m60, v01, v06
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: sc, f, m48, m36, m60, m12, m24, m72, m84
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 48
\item \strong{MHNUM} \emph{character} System Number 1=1. Psychiatric ; 2=2. Neurologic ; 3=3. Head, Eyes, Ears, Nose, Throat ; 4=4. Cardiovascular ; 5=5. Respiratory ; 6=6. Hepatic ; 7=7. Dermatologic-Connective Tissue ; 8=8. Musculoskeletal ; 9=9. Endocrine-Metabolic ; 10=10. Gastrointestinal ; 11=11. Hematopoietic-Lymphatic ; 12=12. Renal-Genitourinary ; 13=13. Allergies or Drug Sensitivities ; 14=14. Alcohol Abuse ; 15=15. Drug Abuse ; 16=16. Smoking ; 17=17. Malignancy ; 18=18. Major Surgical Procedures ; 19=19. Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{MHDESC} \emph{character} Description of condition/problem  
\item \strong{MHDTONSET} \emph{character} Date of onset:  
\item \strong{MHCUR} \emph{character} Is the condition/problem current? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{MHSTAB} \emph{character} If Yes, is the condition/problem stable? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric}   Range value: 8, ..., 32736
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{medical_history}
