% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REGISTRY}
\alias{REGISTRY}
\title{Registry}
\format{
A data frame with 28858 observations and 27 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(REGISTRY)
}
\description{
Registry data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 1, ..., 10900
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Date Date: YYY-MM-DD
\item \strong{RGCONDCT} \emph{character} Was this visit conducted? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{VISTYPE} \emph{character} Visit Type 
\itemize{
\item \emph{ADNIGO, ADNI2:} 1=Standard ; 2=Non-standard ; 3=Not done
\item \emph{ADNI3, ADNI4:} 1=In Clinic Visit ; 2=Telephone Visit ; 3=Telephone Neuropathology Follow-up ; 4=Not done
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RGSTATUS} \emph{character} Was a screening visit/ADNI2 initial visit performed? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{RGREASON} \emph{character} If visit is not done, or telephone visit is done in place of an In-clinic visit, provide reason. 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn Consent ; 6=Other (specify)
\item \emph{ADNI3:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn Consent ; 6=Other (specify) ; 10=COVID-19 Participant's Illness ; 11=COVID-19 Participant - pandemic disruptions ; 12=COVID-19 Site closure due to pandemic
\item \emph{ADNI4:} 1=Illness ; 10=COVID-19 Participant's Illness ; 11=COVID-19 pandemic disruptions ; 12=COVID-19 Site closure due to pandemic ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn Consent ; 6=Other (specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RGSOURCE} \emph{character} Who participated in this visit? 1=Participant ; 2=Study Partner ; 3=Both
\emph{Decoded Value: }\strong{Yes}
\item \strong{RGRESCRN} \emph{character} Is this a rescreen? 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RGPREVID} \emph{character} If Yes, what was the participant's initial ID number?  
\item \strong{CHANGTR} \emph{character} Would you like to change this visit's track or the participant's visit schedule? 0=No ; 1=Yes, change this visit's track ; 2=Yes, change ADNI1 continuing participant's visit schedule
\emph{Decoded Value: }\strong{Yes}
\item \strong{CGTRACK} \emph{character} Please select from the following visit's tracks: -4=No selection ; 1=standard eMCI visit ; 2=standard ADNI1 continuing visit(start at m36) ; 3=standard ADNI1 continuing visit(start at m48) ; 4=standard ADNI1 continuing visit(start at m60) ; 5=flexible ;
\emph{Decoded Value: }\strong{Yes}
\item \strong{CGTRACK2} \emph{character} Please select from the following visit schedule: -4=No selection ; 2=start ADNIGO at m36 ; 3=start ADNIGO at m48 ; 4=start ADNIGO at m60 ;
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTSTATUS} \emph{character} Participant Status: 1=Active ; 2=Discontinued from follow up, will not return
\emph{Decoded Value: }\strong{Yes}
\item \strong{PTTYPE} \emph{character} Type of transfer - Indicate if visit is related to a Permanent or Temporary Transfer 1=First visit conducted at new Home Site - Permanent transfer (not returning to site) ; 2=Visit conducted at alternate site - Temporary transfer (one or more visits done elsewhere, participant will return to original site) ; 3=Visit conducted  at Home Site
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 2, ..., 187099
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: annual, sc
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{enrollment}
