% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMOTE_DATADIC}
\alias{REMOTE_DATADIC}
\title{REMOTE_DATADIC}
\format{
A data frame with 321 observations and 12 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(REMOTE_DATADIC)
}
\description{
REMOTE_DATADIC . More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{PHASE} \emph{character}    
\item \strong{CRFNAME} \emph{character}    
\item \strong{TBLNAME} \emph{character}    
\item \strong{FLDNAME} \emph{character}    
\item \strong{TEXT} \emph{character}    
\item \strong{TYPE} \emph{character}    
\item \strong{LENGTH} \emph{character}    
\item \strong{CODE} \emph{character}    
\item \strong{UNITS} \emph{character}    
\item \strong{STATUS} \emph{character}    
\item \strong{CODE_CHANGES} \emph{character}    
\item \strong{MAPPING_NOTES} \emph{character}    
}
}
\keyword{adni4}
\keyword{data_dict}
