% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_multiple_values}
\alias{replace_multiple_values}
\title{Replace/Substitute Multiple Values}
\usage{
replace_multiple_values(input_string, code, decode)
}
\arguments{
\item{input_string}{Input string}

\item{code}{Coded values}

\item{decode}{Actual string values}
}
\value{
A character vector similar to \code{input_string} with replaced values.
}
\description{
This function is used to replace string values.
}
\examples{
\dontrun{
input_string <- c("-2", "1", "1", "-1")
code <- c("null", "1", "-1", "-2")
decode <- c(
  "pending enrollment", "randomized - assigned a scan category",
  "screen failed", "screen failed after randomization"
)
replaced_values <- replace_multiple_values(
  input_string = input_string,
  code = code,
  decode = decode
)
replaced_values
}
}
\concept{function to replace values}
\keyword{utils_fun}
