% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_values_dataset}
\alias{replace_values_dataset}
\title{Replace Values for Multiple Variable - Wrapper}
\usage{
replace_values_dataset(.data, phaseVar = "PHASE", input_values)
}
\arguments{
\item{.data}{Data.frame}

\item{phaseVar}{Variable name for the ADNI study protocol, Default: "PHASE"}

\item{input_values}{A nested list values of each columns associated with corresponding ADNI
 study phase and the format will be: [column_name][[phase_name]]$values
\itemize{
  \item \emph{code} Value that will be replaced
  \item \emph{decode} Values that will replace coded value (\emph{code})
}}
}
\value{
A data frame with replaced values
}
\description{
This function is used to replace the value of multiple columns in the dataset
 based on the provided corresponding input values.
}
\examples{
\dontrun{
data_dict_dd <- get_factor_levels_datadict(
  .datadic = ADNIMERGE2::DATADIC,
  tbl_name = NULL,
  nested_value = TRUE
)
all_fld_name <- get_factor_fldname(
  .datadic = data_dict_dd,
  tbl_name = "REGISTRY",
  dd_fldnames = colnames(ADNIMERGE2::REGISTRY)
)
input_values <- collect_value_mapping(
  .datadic = data_dict_dd,
  tbl_name = "REGISTRY",
  all_fld_name = all_fld_name
)
result_dataset <- replace_values_dataset(
  .data = ADNIMERGE2::REGISTRY,
  phaseVar = "COLPROT",
  input_values = input_values
)
}
}
\seealso{
Other functions to replace values: 
\code{\link{convert_to_missing_value}()},
\code{\link{replace_values_phase_specific}()},
\code{\link{replace_values_single_var}()}
}
\concept{functions to replace values}
\keyword{adni_replace_fun}
