% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_values_phase_specific}
\alias{replace_values_phase_specific}
\title{Replace Phase Specific Values for Single Variable}
\usage{
replace_values_phase_specific(
  .data,
  fld_name,
  phase,
  phaseVar = "PHASE",
  code,
  decode
)
}
\arguments{
\item{.data}{Data.frame}

\item{fld_name}{Variable name}

\item{phase}{ADNI study phase/protocol name}

\item{phaseVar}{Variable name for the ADNI study protocol, Default: "PHASE"}

\item{code}{Values that will be replaced}

\item{decode}{Values that will replace coded value (\code{code})}
}
\value{
A data.frame with replaced values for the provided variable.
}
\description{
This function is used to recode/replace phase specific values of a
 single variable in a dataset.
}
\examples{
\dontrun{
data_dict_dd <- get_factor_levels_datadict(
  data_dict = ADNIMERGE2::DATADIC,
  tbl_name = "REGISTRY", # NULL
  nested_value = TRUE
)
input_values <- collect_value_mapping(
  data_dict = data_dict_dd,
  tbl_name = "REGISTRY",
  all_fld_name = "VISTYPE"
)
decode <- input_values$VISTYPE$ADNI1$decode
code <- input_values$VISTYPE$ADNI1$code
result_dataset <- replace_values_phase_specific(
  .data = ADNIMERGE2::REGISTRY,
  phaseVar = "COLPROT",
  fld_name = "VISTYPE",
  phase = "ADNI1",
  code = code,
  decode = decode
)
}
}
\seealso{
Other functions to replace values: 
\code{\link{convert_to_missing_value}()},
\code{\link{replace_values_dataset}()},
\code{\link{replace_values_single_var}()}
}
\concept{functions to replace values}
\keyword{adni_replace_fun}
