% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_values_single_var}
\alias{replace_values_single_var}
\title{Replace Values for Single Variable Across Multiple Phase}
\usage{
replace_values_single_var(.data, fld_name, phaseVar = "PHASE", input_values)
}
\arguments{
\item{.data}{Data frame}

\item{fld_name}{Variable name}

\item{phaseVar}{Variable name for the ADNI study protocol, Default: "PHASE"}

\item{input_values}{A list value associated with each ADNI study phase and the format will be
  \emph{[phase_name]$values}.
\itemize{
  \item \emph{code}: Value that will be replaced, see more \code{\link{collect_value_mapping_single_var}()}
  \item \emph{decode}: Values that will replace the coded values, \code{code}
}}
}
\value{
A data frame with replaced values.
}
\description{
This function is used to replace values for a single column in the dataset
 across different ADNI study phases.
}
\examples{
\dontrun{
data_dict_dd <- get_factor_levels_datadict(
  data_dict = ADNIMERGE2::DATADIC,
  tbl_name = NULL,
  nested_value = TRUE
)
input_values <- collect_value_mapping(
  data_dict = data_dict_dd,
  tbl_name = "REGISTRY",
  all_fld_name = "VISTYPE"
)
input_values <- input_values$VISTYPE
result_dataset <- replace_values_single_var(
  .data = ADNIMERGE2::REGISTRY,
  phaseVar = "COLPROT",
  fld_name = "VISTYPE",
  input_values = input_values
)
}
}
\seealso{
Other functions to replace values: 
\code{\link{convert_to_missing_value}()},
\code{\link{replace_values_dataset}()},
\code{\link{replace_values_phase_specific}()}
}
\concept{functions to replace values}
\keyword{adni_replace_fun}
