% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RMT_APOERES}
\alias{RMT_APOERES}
\title{ApoE Genotyping - Results}
\format{
A data frame with 612 observations and 4 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RMT_APOERES)
}
\description{
ApoE Genotyping - Results data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{PHASE} \emph{character}   , with options: ADNI4
\item \strong{ADNIOnlineID} \emph{character} Participant Identifier for the Remote Cohort  
\item \strong{GENOTYPE} \emph{character} Apolipoprotein E (Apo-E) Genotype , with options: 3/3, 3/4, 2/4, 2/3, 4/4, 2/2
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{remotely_collected_data}
