% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RMT_ECOG12SP}
\alias{RMT_ECOG12SP}
\title{DIG Remotely administered Assessment of Everyday Cognition (Ecog) Questionnaire – Informant-Report Version}
\format{
A data frame with 4675 observations and 21 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RMT_ECOG12SP)
}
\description{
[DIG] Remotely administered Assessment of Everyday Cognition (Ecog) Questionnaire – Informant-Report Version data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{RMT_PHASE} \emph{character} Phase of the study in which this observation was collected , with options: ADNI4
\item \strong{ADNIOnlineID} \emph{character} Participant Identifier for the Remote Cohort  
\item \strong{PTID} \emph{character} Participant Identifier for the in-clinic ADNI study (if applicable)  
\item \strong{RMT_Timepoint} \emph{character} Timepoint of observation , with options: m00, m06, m12, m24, m18
\item \strong{RMT_StatusDate} \emph{ } Date of completion Date: YYY-MM-DD
\item \strong{RMT_SP_CONCERN} \emph{numeric} Are you concerned that they have a memory or other thinking problem? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_SP_MEMREF} \emph{numeric} Compared to one year ago, do you feel that their memory has declined substantially? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_ecogsp_Duration} \emph{numeric} Amount of time elapsed between beginning and completing the assessment (seconds) Range value: 21, ..., 86281
\item \strong{RMT_ecogsp1} \emph{numeric} Remembering where they have placed personal items or objects 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp2} \emph{numeric} Remembering the current date or day of the week 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp3} \emph{numeric} Communicating or expressing ideas in a conversation 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp4} \emph{numeric} Understanding spoken directions or instructions 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp5} \emph{numeric} Difficulty calculating distances when driving or driving too close to another vehicle 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp6} \emph{numeric} Finding their way around a familiar home 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp7} \emph{numeric} Anticipating weather changes and planning accordingly 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp8} \emph{numeric} Planning a trip, vacation, or outing 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp9} \emph{numeric} Keeping a living and work space (such as a home office, garage, or pantry) organized 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp10} \emph{numeric} Managing bill payments 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp11} \emph{numeric} Doing two things at once 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{RMT_ecogsp12} \emph{numeric} Working on a particular task while talking to someone at the same time 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{remotely_collected_data}
\keyword{undefined_phase}
