% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RMT_PTDEMOG}
\alias{RMT_PTDEMOG}
\title{DIG Remotely administered participant demographics questionnaire}
\format{
A data frame with 21766 observations and 23 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RMT_PTDEMOG)
}
\description{
[DIG] Remotely administered participant demographics questionnaire data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{RMT_PHASE} \emph{character} Phase of the study in which this observation was collected , with options: ADNI4
\item \strong{ADNIOnlineID} \emph{character} Participant Identifier for the Remote Cohort  
\item \strong{PTID} \emph{character} Participant Identifier for the in-clinic ADNI study (if applicable)  
\item \strong{RMT_Timepoint} \emph{character} Timepoint of observation , with options: m00, m06, m12, m18, m24
\item \strong{RMT_StatusDate} \emph{ } Date of completion Date: YYY-MM-DD
\item \strong{Age_Baseline} \emph{numeric} The participant's de-identified age at their baseline date in this dataset. Ages 90 and above are represented simply as 90. Range value: 54, ..., 90
\item \strong{Gender} \emph{numeric} Gender of the participant 0=Male ; 1=Female ; 2=Other ; 3=Prefer not to say
\item \strong{LatinoEthnicity} \emph{numeric} Do you consider yourself Hispanic/Latino or not Hispanic/Latino? Check all that apply. 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{Latino_MX} \emph{numeric} Did the participant select Mexican, Mexican American, Chicano/a 0=No ; 1=Yes
\item \strong{Latino_PR} \emph{numeric} Did the participant select Puerto Rican 0=No ; 1=Yes
\item \strong{Latino_CB} \emph{numeric} Did the participant select Cuban 0=No ; 1=Yes
\item \strong{Latino_Other} \emph{numeric} Did the participant select another Hispanic, Latino/a, or Spanish origin 0=No ; 1=Yes
\item \strong{Race_AmerIndian} \emph{numeric} Did the participant select American Indian or Alaska Native 0=No ; 1=Yes
\item \strong{Race_Asian} \emph{numeric} Did the participant select Asian 0=No ; 1=Yes
\item \strong{Race_AfricanAmerican} \emph{numeric} Did the participant select African American 0=No ; 1=Yes
\item \strong{Race_PacificIslander} \emph{numeric} Did the participant select Pacific Islander 0=No ; 1=Yes
\item \strong{Race_White} \emph{numeric} Did the participant select White 0=No ; 1=Yes
\item \strong{Race_Unknown} \emph{numeric} Did the participant select Unknown? 0=No ; 1=Yes
\item \strong{Race_PreferNotSay} \emph{numeric} Did the participant select Prefer Not to Say? 0=No ; 1=Yes
\item \strong{RMT_Education} \emph{numeric} Participant Years of Education 99=Prefer Not to Say
\item \strong{RMT_ADI_NATRANK_v2021} \emph{numeric} National percentile of block group ADI score Range value: 1, ..., 100
\item \strong{RMT_ADI_STATERANK_v2021} \emph{numeric} State-specific decile of block group ADI score Range value: 1, ..., 10
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{remotely_collected_data}
\keyword{undefined_phase}
