% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RMT_Screening}
\alias{RMT_Screening}
\title{DIG Remotely administered screening questionnaire}
\format{
A data frame with 12227 observations and 26 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RMT_Screening)
}
\description{
[DIG] Remotely administered screening questionnaire data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{RMT_PHASE} \emph{character} Phase of the study in which this observation was collected , with options: ADNI4
\item \strong{ADNIOnlineID} \emph{character} Participant Identifier for the Remote Cohort  
\item \strong{PTID} \emph{character} Participant Identifier for the in-clinic ADNI study (if applicable)  
\item \strong{RMT_Timepoint} \emph{character} Timepoint of observation , with options: m00, m06, m12, m18
\item \strong{RMT_StatusDate} \emph{ } Date of completion Date: YYY-MM-DD
\item \strong{RMT_MCI_Dx} \emph{numeric} Have you been diagnosed with Mild Cognitive Impairment (MCI) or a memory problem by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say
\item \strong{RMT_AD_Dx} \emph{numeric} Have you been diagnosed with Alzheimer's Disease (AD) by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say
\item \strong{RMT_DEM_Dx} \emph{numeric} Have you been diagnosed with dementia by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say
\item \strong{RMT_CImed} \emph{numeric} Have you ever been prescribed a medication for cognitive impairment or memory problems by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say
\item \strong{RMT_FamAD} \emph{numeric} Have any of your biological parents, full sibling(s), or biological children ever been diagnosed with Alzheimer's Disease or dementia by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say
\item \strong{RMT_NursHm} \emph{numeric} Do you reside in a skilled nursing facility? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_MntlHlth} \emph{numeric} In the past year, have you been diagnosed with any of the following conditions: schizophrenia, bipolar disorder, or major depression? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_BrainCndtn} \emph{numeric} Have you ever been diagnosed with any of the following conditions: Parkinson's disease, Huntington's disease, a brain tumor, a seizure disorder, or multiple sclerosis? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_AlchlDrug} \emph{numeric} In the past two years, have you been diagnosed by a healthcare provider as having alcohol use disorder or drug use disorder? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_Metal} \emph{numeric} Do you have a pacemaker, an implanted medical device (such as aneurysm clips, artificial heart valve, IUD, etc.), metal fragments (for example, due to military service or metal working), or other foreign object(s) in your body?Note: if the foreign object is an artificial knee or hip replacement, please answer no 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_Metal_rplc} \emph{numeric} Is the metal in your body an artificial knee or hip replacement? 0=No ; 1=Yes
\item \strong{RMT_Metal_other} \emph{numeric} Do you have any other metal in your body (such as pacemaker, aneurysm clips, metal fragments, etc.)? 0=No ; 1=Yes
\item \strong{RMT_MemStudy} \emph{numeric} Are you willing to participate in a 5-year study that includes memory and thinking tests? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_MRIPET} \emph{numeric} Are you willing to participate in a 5-year study that includes MRI and PET scans (pictures of your brain, at no cost to you)? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_Clausphba} \emph{numeric} Do you experience claustrophobia (fear of confined spaces) that would prevent you from completing a MRI or PET scan (pictures of your brain)? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say
\item \strong{RMT_BLDwill} \emph{numeric} Are you willing to have a blood draw at a nearby local clinic (Quest Diagnostics Patient Service Center) at no cost to you? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_Web} \emph{numeric} Do you regularly have access to the internet? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{RMT_Device} \emph{numeric} Do you have a computer, smartphone, or tablet device (such as an iPad, Chromebook, etc.)? 0=No ; 1=Yes ; 3=Prefer not to say
\item \strong{Referral_BLD} \emph{numeric} Has this participant been referred to the remote blood cohort? 0=No ; 1=Yes
\item \strong{Referral_CLIN} \emph{numeric} Has this participant been referred to the in-clinic ADNI study? 0=No ; 1=Yes
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{remotely_collected_data}
\keyword{undefined_phase}
