% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RMT_STORYTELLER}
\alias{RMT_STORYTELLER}
\title{Novoic Storyteller Data}
\format{
A data frame with 14823 observations and 219 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RMT_STORYTELLER)
}
\description{
Novoic Storyteller Data data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{RMT_PHASE} \emph{character} Phase of the study in which this observation was collected , with options: ADNI4
\item \strong{ADNIOnlineID} \emph{character} Participant Identifier for the Remote Cohort  
\item \strong{PTID} \emph{character} Participant Identifier for the in-clinic ADNI study (if applicable)  
\item \strong{RMT_Timepoint} \emph{character} Timepoint of observation , with options: m00, m06, m12, m18, m24
\item \strong{EXAMDATE} \emph{ } Date of examination Date: YYY-MM-DD
\item \strong{SCREENER_SCORE} \emph{numeric} Overall G-match score for the screener Range value: 0, ..., 100
\item \strong{SCREENER_RECOMMEND} \emph{numeric} Overall recommendation for the screener 0=No ; 1=Yes
\item \strong{QC_ATTEMPTS} \emph{numeric} Number of attempts at the quality control task Range value: 1, ..., 11
\item \strong{QC_PASSED} \emph{numeric} Whether the quality control task was passed 0=No ; 1=Yes
\item \strong{USER_AGENT} \emph{character} User agent string  
\item \strong{STORY1IMM_LANGUAGE} \emph{character} Language for story 1 immediate recall , with options: en-US
\item \strong{STORY1IMM_PLAYBACK_COUNT} \emph{numeric} Audio stimulus playback count for story 1 immediate recall Range value: 0, ..., 5
\item \strong{STORY1IMM_AUDIO_LENGTH} \emph{numeric} Participant audio length for story 1 immediate recall Range value: 0.76, ..., 305.18
\item \strong{STORY1IMM_NUM_PRONOUNS} \emph{numeric} Number of pronouns for story 1 immediate recall Range value: 0, ..., 77
\item \strong{STORY1IMM_NUM_NOUNS} \emph{numeric} Number of nouns for story 1 immediate recall Range value: 0, ..., 46
\item \strong{STORY1IMM_NUM_PROPER_NOUNS} \emph{numeric} Number of proper nouns for story 1 immediate recall Range value: 0, ..., 13
\item \strong{STORY1IMM_NUM_VERBS} \emph{numeric} Number of verbs for story 1 immediate recall Range value: 0, ..., 56
\item \strong{STORY1IMM_NUM_ADVERBS} \emph{numeric} Number of adverbs for story 1 immediate recall Range value: 0, ..., 28
\item \strong{STORY1IMM_NUM_ADJECTIVES} \emph{numeric} Number of adjectives for story 1 immediate recall Range value: 0, ..., 18
\item \strong{STORY1IMM_NUM_WORDS} \emph{numeric} Number of words for story 1 immediate recall Range value: 0, ..., 350
\item \strong{STORY1IMM_IDEA_DENSITY} \emph{numeric} Idea density for story 1 immediate recall Range value: 0, ..., 1
\item \strong{STORY1IMM_UNIQUE_WORDS_RATIO} \emph{numeric} Unique words ratio for story 1 immediate recall Range value: 0.357142857142857, ..., 1
\item \strong{STORY1IMM_MEAN_LETTER_PER_WORD} \emph{numeric} Mean number of letters per word for story 1 immediate recall Range value: 2, ..., 9
\item \strong{STORY1IMM_GMATCH} \emph{numeric} G-match score (source text vs transcript) for story 1 immediate recall Range value: 0, ..., 100
\item \strong{STORY1IMM_SPEECH_DURATION} \emph{numeric} Speech duration for story 1 immediate recall Range value: 0, ..., 112.6
\item \strong{STORY1IMM_NONSPEECH_PROPORTION} \emph{numeric} Non-speech proportion of time for story 1 immediate recall Range value: 0, ..., 0.912912912912913
\item \strong{STORY1IMM_SPEECH_RATE} \emph{numeric} Speech rate for story 1 immediate recall Range value: 0.0338983050847458, ..., 5.55555555555556
\item \strong{STORY1IMM_ARTICULATION_RATE} \emph{numeric} Articulation rate for story 1 immediate recall Range value: 0.0338983050847458, ..., 5.55555555555556
\item \strong{STORY1IMM_NUM_PAUSES} \emph{numeric} Number of pauses for story 1 immediate recall Range value: 0, ..., 35
\item \strong{STORY1IMM_MEAN_PAUSE_LENGTH} \emph{numeric} Mean length of pauses for story 1 immediate recall Range value: 0.3, ..., 39.8
\item \strong{STORY1IMM_AMPLITUDE_SD} \emph{numeric} Standard deviation of the amplitude for story 1 immediate recall Range value: 0, ..., 0.295156925916672
\item \strong{STORY1IMM_F0_MEAN} \emph{numeric} Formant F0 mean for story 1 immediate recall Range value: 61.038, ..., 240.867
\item \strong{STORY1IMM_F0_SD} \emph{numeric} Formant F0 standard deviation for story 1 immediate recall Range value: 0, ..., 77.268382971904
\item \strong{STORY1IMM_INTENSITY_MEAN} \emph{numeric} Intensity mean for story 1 immediate recall Range value: 0, ..., 0.415349274873734
\item \strong{STORY1IMM_INTENSITY_SD} \emph{numeric} Intensity standard deviation for story 1 immediate recall Range value: 0, ..., 0.249458640813828
\item \strong{STORY1IMM_MFCC_1_MEAN} \emph{numeric} MFCC 1 mean for story 1 immediate recall Range value: -1131.37109375, ..., -6.6077995300293
\item \strong{STORY1IMM_MFCC_2_MEAN} \emph{numeric} MFCC 2 mean for story 1 immediate recall Range value: -87.7940826416016, ..., 222.97785949707
\item \strong{STORY1IMM_MFCC_3_MEAN} \emph{numeric} MFCC 3 mean for story 1 immediate recall Range value: -101.867546081543, ..., 93.3259811401367
\item \strong{STORY1IMM_MFCC_4_MEAN} \emph{numeric} MFCC 4 mean for story 1 immediate recall Range value: -34.630313873291, ..., 96.9562377929688
\item \strong{STORY1IMM_MFCC_5_MEAN} \emph{numeric} MFCC 5 mean for story 1 immediate recall Range value: -48.5658760070801, ..., 52.7833595275879
\item \strong{STORY1IMM_MFCC_6_MEAN} \emph{numeric} MFCC 6 mean for story 1 immediate recall Range value: -44.1751670837402, ..., 52.3673057556152
\item \strong{STORY1IMM_MFCC_7_MEAN} \emph{numeric} MFCC 7 mean for story 1 immediate recall Range value: -40.6176223754883, ..., 29.2108592987061
\item \strong{STORY1IMM_MFCC_8_MEAN} \emph{numeric} MFCC 8 mean for story 1 immediate recall Range value: -34.4426116943359, ..., 38.9434051513672
\item \strong{STORY1IMM_MFCC_9_MEAN} \emph{numeric} MFCC 9 mean for story 1 immediate recall Range value: -36.2972297668457, ..., 30.849494934082
\item \strong{STORY1IMM_MFCC_10_MEAN} \emph{numeric} MFCC 10 mean for story 1 immediate recall Range value: -31.7630462646484, ..., 34.6205101013184
\item \strong{STORY1IMM_MFCC_11_MEAN} \emph{numeric} MFCC 11 mean for story 1 immediate recall Range value: -31.8291187286377, ..., 17.6796360015869
\item \strong{STORY1IMM_MFCC_12_MEAN} \emph{numeric} MFCC 12 mean for story 1 immediate recall Range value: -21.5977916717529, ..., 30.8843936920166
\item \strong{STORY1IMM_MFCC_13_MEAN} \emph{numeric} MFCC 13 mean for story 1 immediate recall Range value: -50.1643333435059, ..., 12.4952554702759
\item \strong{STORY1IMM_MFCC_1_SD} \emph{numeric} MFCC 1 standard deviation for story 1 immediate recall Range value: 0, ..., 194.822341918945
\item \strong{STORY1IMM_MFCC_2_SD} \emph{numeric} MFCC 2 standard deviation for story 1 immediate recall Range value: 0, ..., 102.484924316406
\item \strong{STORY1IMM_MFCC_3_SD} \emph{numeric} MFCC 3 standard deviation for story 1 immediate recall Range value: 0, ..., 67.6520538330078
\item \strong{STORY1IMM_MFCC_4_SD} \emph{numeric} MFCC 4 standard deviation for story 1 immediate recall Range value: 0, ..., 48.763599395752
\item \strong{STORY1IMM_MFCC_5_SD} \emph{numeric} MFCC 5 standard deviation for story 1 immediate recall Range value: 0, ..., 41.1321830749512
\item \strong{STORY1IMM_MFCC_6_SD} \emph{numeric} MFCC 6 standard deviation for story 1 immediate recall Range value: 0, ..., 33.4563980102539
\item \strong{STORY1IMM_MFCC_7_SD} \emph{numeric} MFCC 7 standard deviation for story 1 immediate recall Range value: 0, ..., 31.499885559082
\item \strong{STORY1IMM_MFCC_8_SD} \emph{numeric} MFCC 8 standard deviation for story 1 immediate recall Range value: 0, ..., 24.8378448486328
\item \strong{STORY1IMM_MFCC_9_SD} \emph{numeric} MFCC 9 standard deviation for story 1 immediate recall Range value: 0, ..., 24.9872531890869
\item \strong{STORY1IMM_MFCC_10_SD} \emph{numeric} MFCC 10 standard deviation for story 1 immediate recall Range value: 0, ..., 21.022367477417
\item \strong{STORY1IMM_MFCC_11_SD} \emph{numeric} MFCC 11 standard deviation for story 1 immediate recall Range value: 0, ..., 19.2079486846924
\item \strong{STORY1IMM_MFCC_12_SD} \emph{numeric} MFCC 12 standard deviation for story 1 immediate recall Range value: 0, ..., 18.1088771820068
\item \strong{STORY1IMM_MFCC_13_SD} \emph{numeric} MFCC 13 standard deviation for story 1 immediate recall Range value: 0, ..., 19.2777976989746
\item \strong{STORY1IMM_FORMANTS_1_MEAN} \emph{numeric} Formant F1 mean for story 1 immediate recall Range value: 0, ..., 2240.35
\item \strong{STORY1IMM_FORMANTS_2_MEAN} \emph{numeric} Formant F2 mean for story 1 immediate recall Range value: 0, ..., 3182.5
\item \strong{STORY1IMM_FORMANTS_3_MEAN} \emph{numeric} Formant F3 mean for story 1 immediate recall Range value: 0, ..., 4135.35
\item \strong{STORY1IMM_FORMANTS_4_MEAN} \emph{numeric} Formant F4 mean for story 1 immediate recall Range value: 0, ..., 5116.33
\item \strong{STORY1IMM_FORMANTS_1_SD} \emph{numeric} Formant F1 standard deviation for story 1 immediate recall Range value: 0, ..., 1034.12
\item \strong{STORY1IMM_FORMANTS_2_SD} \emph{numeric} Formant F2 standard deviation for story 1 immediate recall Range value: 0, ..., 1289.39
\item \strong{STORY1IMM_FORMANTS_3_SD} \emph{numeric} Formant F3 standard deviation for story 1 immediate recall Range value: 0, ..., 1685.79
\item \strong{STORY1IMM_FORMANTS_4_SD} \emph{numeric} Formant F4 standard deviation for story 1 immediate recall Range value: 0, ..., 2198.59
\item \strong{STORY1IMM_LOCAL_JITTER} \emph{numeric} Jitter (local) for story 1 immediate recall Range value: 0, ..., 0.0443882271257532
\item \strong{STORY1IMM_LOCAL_ABS_JITTER} \emph{numeric} Jitter (local, absolute) for story 1 immediate recall Range value: 0, ..., 0.00041727400262465
\item \strong{STORY1IMM_RAP_JITTER} \emph{numeric} Jitter (rap) for story 1 immediate recall Range value: 0, ..., 0.0265944777953175
\item \strong{STORY1IMM_PPQ5_JITTER} \emph{numeric} Jitter (PPQ5) for story 1 immediate recall Range value: 0, ..., 0.0174450080485077
\item \strong{STORY1IMM_DDP_JITTER} \emph{numeric} Jitter (DDP) for story 1 immediate recall Range value: 0, ..., 0.0797834333859526
\item \strong{STORY1IMM_LOCAL_SHIMMER} \emph{numeric} Shimmer (local) for story 1 immediate recall Range value: 0, ..., 8.11020828635767
\item \strong{STORY1IMM_LOCAL_DB_SHIMMER} \emph{numeric} Shimmer (local, dB) for story 1 immediate recall Range value: 0, ..., 2.49877482652664
\item \strong{STORY1IMM_APQ3_SHIMMER} \emph{numeric} Shimmer (APQ3) for story 1 immediate recall Range value: 0, ..., 3.26254028119208
\item \strong{STORY1IMM_APQ5_SHIMMER} \emph{numeric} Shimmer (APQ5) for story 1 immediate recall Range value: 0, ..., 4.28702720374124
\item \strong{STORY1IMM_APQ11_SHIMMER} \emph{numeric} Shimmer (APQ11) for story 1 immediate recall Range value: 0, ..., 7.7851804203585
\item \strong{STORY2IMM_LANGUAGE} \emph{character} Language for story 2 immediate recall , with options: en-US
\item \strong{STORY2IMM_PLAYBACK_COUNT} \emph{numeric} Audio stimulus playback count for story 2 immediate recall Range value: 0, ..., 5
\item \strong{STORY2IMM_AUDIO_LENGTH} \emph{numeric} Participant audio length for story 2 immediate recall Range value: 0.08, ..., 305.52
\item \strong{STORY2IMM_NUM_PRONOUNS} \emph{numeric} Number of pronouns for story 2 immediate recall Range value: 0, ..., 62
\item \strong{STORY2IMM_NUM_NOUNS} \emph{numeric} Number of nouns for story 2 immediate recall Range value: 0, ..., 70
\item \strong{STORY2IMM_NUM_PROPER_NOUNS} \emph{numeric} Number of proper nouns for story 2 immediate recall Range value: 0, ..., 17
\item \strong{STORY2IMM_NUM_VERBS} \emph{numeric} Number of verbs for story 2 immediate recall Range value: 0, ..., 69
\item \strong{STORY2IMM_NUM_ADVERBS} \emph{numeric} Number of adverbs for story 2 immediate recall Range value: 0, ..., 34
\item \strong{STORY2IMM_NUM_ADJECTIVES} \emph{numeric} Number of adjectives for story 2 immediate recall Range value: 0, ..., 22
\item \strong{STORY2IMM_NUM_WORDS} \emph{numeric} Number of words for story 2 immediate recall Range value: 0, ..., 363
\item \strong{STORY2IMM_IDEA_DENSITY} \emph{numeric} Idea density for story 2 immediate recall Range value: 0, ..., 1
\item \strong{STORY2IMM_UNIQUE_WORDS_RATIO} \emph{numeric} Unique words ratio for story 2 immediate recall Range value: 0.235955056179775, ..., 1
\item \strong{STORY2IMM_MEAN_LETTER_PER_WORD} \emph{numeric} Mean number of letters per word for story 2 immediate recall Range value: 2, ..., 12
\item \strong{STORY2IMM_GMATCH} \emph{numeric} G-match score (source text vs transcript) for story 2 immediate recall Range value: 0, ..., 100
\item \strong{STORY2IMM_SPEECH_DURATION} \emph{numeric} Speech duration for story 2 immediate recall Range value: 0, ..., 124
\item \strong{STORY2IMM_NONSPEECH_PROPORTION} \emph{numeric} Non-speech proportion of time for story 2 immediate recall Range value: 0, ..., 0.898190045248869
\item \strong{STORY2IMM_SPEECH_RATE} \emph{numeric} Speech rate for story 2 immediate recall Range value: 0.0925925925925926, ..., 6.36363636363636
\item \strong{STORY2IMM_ARTICULATION_RATE} \emph{numeric} Articulation rate for story 2 immediate recall Range value: 0.0925925925925926, ..., 6.36363636363636
\item \strong{STORY2IMM_NUM_PAUSES} \emph{numeric} Number of pauses for story 2 immediate recall Range value: 0, ..., 38
\item \strong{STORY2IMM_MEAN_PAUSE_LENGTH} \emph{numeric} Mean length of pauses for story 2 immediate recall Range value: 0.3, ..., 42.3
\item \strong{STORY2IMM_AMPLITUDE_SD} \emph{numeric} Standard deviation of the amplitude for story 2 immediate recall Range value: 0, ..., 0.291553527116776
\item \strong{STORY2IMM_F0_MEAN} \emph{numeric} Formant F0 mean for story 2 immediate recall Range value: 61.0378759895198, ..., 237.99677681199
\item \strong{STORY2IMM_F0_SD} \emph{numeric} Formant F0 standard deviation for story 2 immediate recall Range value: 0, ..., 79.0865961852582
\item \strong{STORY2IMM_INTENSITY_MEAN} \emph{numeric} Intensity mean for story 2 immediate recall Range value: 0, ..., 0.413347005844116
\item \strong{STORY2IMM_INTENSITY_SD} \emph{numeric} Intensity standard deviation for story 2 immediate recall Range value: 0, ..., 0.246163427829742
\item \strong{STORY2IMM_MFCC_1_MEAN} \emph{numeric} MFCC 1 mean for story 2 immediate recall Range value: -1131.37109375, ..., -3.91779088973999
\item \strong{STORY2IMM_MFCC_2_MEAN} \emph{numeric} MFCC 2 mean for story 2 immediate recall Range value: -87.0389022827148, ..., 229.237380981445
\item \strong{STORY2IMM_MFCC_3_MEAN} \emph{numeric} MFCC 3 mean for story 2 immediate recall Range value: -95.5741119384766, ..., 77.5792846679688
\item \strong{STORY2IMM_MFCC_4_MEAN} \emph{numeric} MFCC 4 mean for story 2 immediate recall Range value: -37.4888496398926, ..., 90.4929351806641
\item \strong{STORY2IMM_MFCC_5_MEAN} \emph{numeric} MFCC 5 mean for story 2 immediate recall Range value: -48.589672088623, ..., 49.888256072998
\item \strong{STORY2IMM_MFCC_6_MEAN} \emph{numeric} MFCC 6 mean for story 2 immediate recall Range value: -39.4962539672851, ..., 50.9638519287109
\item \strong{STORY2IMM_MFCC_7_MEAN} \emph{numeric} MFCC 7 mean for story 2 immediate recall Range value: -42.2606010437012, ..., 28.8483142852783
\item \strong{STORY2IMM_MFCC_8_MEAN} \emph{numeric} MFCC 8 mean for story 2 immediate recall Range value: -38.1706962585449, ..., 35.4149894714355
\item \strong{STORY2IMM_MFCC_9_MEAN} \emph{numeric} MFCC 9 mean for story 2 immediate recall Range value: -38.6608734130859, ..., 34.9120216369629
\item \strong{STORY2IMM_MFCC_10_MEAN} \emph{numeric} MFCC 10 mean for story 2 immediate recall Range value: -31.0014114379883, ..., 34.4861831665039
\item \strong{STORY2IMM_MFCC_11_MEAN} \emph{numeric} MFCC 11 mean for story 2 immediate recall Range value: -35.1237106323242, ..., 17.449031829834
\item \strong{STORY2IMM_MFCC_12_MEAN} \emph{numeric} MFCC 12 mean for story 2 immediate recall Range value: -24.2628021240234, ..., 26.7934436798096
\item \strong{STORY2IMM_MFCC_13_MEAN} \emph{numeric} MFCC 13 mean for story 2 immediate recall Range value: -50.6267318725586, ..., 12.5135879516602
\item \strong{STORY2IMM_MFCC_1_SD} \emph{numeric} MFCC 1 standard deviation for story 2 immediate recall Range value: 0, ..., 192.498779296875
\item \strong{STORY2IMM_MFCC_2_SD} \emph{numeric} MFCC 2 standard deviation for story 2 immediate recall Range value: 0, ..., 105.969032287598
\item \strong{STORY2IMM_MFCC_3_SD} \emph{numeric} MFCC 3 standard deviation for story 2 immediate recall Range value: 0, ..., 62.268970489502
\item \strong{STORY2IMM_MFCC_4_SD} \emph{numeric} MFCC 4 standard deviation for story 2 immediate recall Range value: 0, ..., 50.8089256286621
\item \strong{STORY2IMM_MFCC_5_SD} \emph{numeric} MFCC 5 standard deviation for story 2 immediate recall Range value: 0, ..., 40.167537689209
\item \strong{STORY2IMM_MFCC_6_SD} \emph{numeric} MFCC 6 standard deviation for story 2 immediate recall Range value: 0, ..., 30.9231510162354
\item \strong{STORY2IMM_MFCC_7_SD} \emph{numeric} MFCC 7 standard deviation for story 2 immediate recall Range value: 0, ..., 31.9191837310791
\item \strong{STORY2IMM_MFCC_8_SD} \emph{numeric} MFCC 8 standard deviation for story 2 immediate recall Range value: 0, ..., 24.2332668304443
\item \strong{STORY2IMM_MFCC_9_SD} \emph{numeric} MFCC 9 standard deviation for story 2 immediate recall Range value: 0, ..., 25.6944770812988
\item \strong{STORY2IMM_MFCC_10_SD} \emph{numeric} MFCC 10 standard deviation for story 2 immediate recall Range value: 0, ..., 21.327579498291
\item \strong{STORY2IMM_MFCC_11_SD} \emph{numeric} MFCC 11 standard deviation for story 2 immediate recall Range value: 0, ..., 18.3624801635742
\item \strong{STORY2IMM_MFCC_12_SD} \emph{numeric} MFCC 12 standard deviation for story 2 immediate recall Range value: 0, ..., 16.8758640289307
\item \strong{STORY2IMM_MFCC_13_SD} \emph{numeric} MFCC 13 standard deviation for story 2 immediate recall Range value: 0, ..., 19.2984504699707
\item \strong{STORY2IMM_FORMANTS_1_MEAN} \emph{numeric} Formant F1 mean for story 2 immediate recall Range value: 0, ..., 2249.13
\item \strong{STORY2IMM_FORMANTS_2_MEAN} \emph{numeric} Formant F2 mean for story 2 immediate recall Range value: 0, ..., 3144.03
\item \strong{STORY2IMM_FORMANTS_3_MEAN} \emph{numeric} Formant F3 mean for story 2 immediate recall Range value: 0, ..., 4117.63
\item \strong{STORY2IMM_FORMANTS_4_MEAN} \emph{numeric} Formant F4 mean for story 2 immediate recall Range value: 0, ..., 5147.57
\item \strong{STORY2IMM_FORMANTS_1_SD} \emph{numeric} Formant F1 standard deviation for story 2 immediate recall Range value: 0, ..., 1020.2
\item \strong{STORY2IMM_FORMANTS_2_SD} \emph{numeric} Formant F2 standard deviation for story 2 immediate recall Range value: 0, ..., 1245.77
\item \strong{STORY2IMM_FORMANTS_3_SD} \emph{numeric} Formant F3 standard deviation for story 2 immediate recall Range value: 0, ..., 1627.55
\item \strong{STORY2IMM_FORMANTS_4_SD} \emph{numeric} Formant F4 standard deviation for story 2 immediate recall Range value: 0, ..., 2131.08
\item \strong{STORY2IMM_LOCAL_JITTER} \emph{numeric} Jitter (local) for story 2 immediate recall Range value: 0, ..., 0.0260047111135237
\item \strong{STORY2IMM_LOCAL_ABS_JITTER} \emph{numeric} Jitter (local, absolute) for story 2 immediate recall Range value: 0, ..., 0.00022897620443122
\item \strong{STORY2IMM_RAP_JITTER} \emph{numeric} Jitter (rap) for story 2 immediate recall Range value: 0, ..., 0.0109727925492098
\item \strong{STORY2IMM_PPQ5_JITTER} \emph{numeric} Jitter (PPQ5) for story 2 immediate recall Range value: 0, ..., 0.0332189871466471
\item \strong{STORY2IMM_DDP_JITTER} \emph{numeric} Jitter (DDP) for story 2 immediate recall Range value: 0, ..., 0.0329183776476296
\item \strong{STORY2IMM_LOCAL_SHIMMER} \emph{numeric} Shimmer (local) for story 2 immediate recall Range value: 0, ..., 4.214980693118
\item \strong{STORY2IMM_LOCAL_DB_SHIMMER} \emph{numeric} Shimmer (local, dB) for story 2 immediate recall Range value: 0, ..., 3.52182507514954
\item \strong{STORY2IMM_APQ3_SHIMMER} \emph{numeric} Shimmer (APQ3) for story 2 immediate recall Range value: 0, ..., 1.86314556953646
\item \strong{STORY2IMM_APQ5_SHIMMER} \emph{numeric} Shimmer (APQ5) for story 2 immediate recall Range value: 0, ..., 3.88903059903243
\item \strong{STORY2IMM_APQ11_SHIMMER} \emph{numeric} Shimmer (APQ11) for story 2 immediate recall Range value: 0, ..., 9.35855869683073
\item \strong{STORY1DEL_LANGUAGE} \emph{character} Language for story 1 delayed recall , with options: en-US, en
\item \strong{STORY1DEL_PLAYBACK_COUNT} \emph{numeric} Audio stimulus playback count for story 1 delayed recall Range value: 0, ..., 3
\item \strong{STORY1DEL_AUDIO_LENGTH} \emph{numeric} Participant audio length for story 1 delayed recall Range value: 0.1, ..., 305.62
\item \strong{STORY1DEL_NUM_PRONOUNS} \emph{numeric} Number of pronouns for story 1 delayed recall Range value: 0, ..., 116
\item \strong{STORY1DEL_NUM_NOUNS} \emph{numeric} Number of nouns for story 1 delayed recall Range value: 0, ..., 109
\item \strong{STORY1DEL_NUM_PROPER_NOUNS} \emph{numeric} Number of proper nouns for story 1 delayed recall Range value: 0, ..., 21
\item \strong{STORY1DEL_NUM_VERBS} \emph{numeric} Number of verbs for story 1 delayed recall Range value: 0, ..., 75
\item \strong{STORY1DEL_NUM_ADVERBS} \emph{numeric} Number of adverbs for story 1 delayed recall Range value: 0, ..., 60
\item \strong{STORY1DEL_NUM_ADJECTIVES} \emph{numeric} Number of adjectives for story 1 delayed recall Range value: 0, ..., 28
\item \strong{STORY1DEL_NUM_WORDS} \emph{numeric} Number of words for story 1 delayed recall Range value: 0, ..., 668
\item \strong{STORY1DEL_IDEA_DENSITY} \emph{numeric} Idea density for story 1 delayed recall Range value: 0, ..., 1
\item \strong{STORY1DEL_UNIQUE_WORDS_RATIO} \emph{numeric} Unique words ratio for story 1 delayed recall Range value: 0.366766467065868, ..., 1
\item \strong{STORY1DEL_MEAN_LETTER_PER_WORD} \emph{numeric} Mean number of letters per word for story 1 delayed recall Range value: 1, ..., 12
\item \strong{STORY1DEL_GMATCH} \emph{numeric} G-match score (source text vs transcript) for story 1 delayed recall Range value: 0, ..., 100
\item \strong{STORY1DEL_GMATCH_IMM_DEL} \emph{numeric} G-match score (immediate transcript vs delayed transcript) for story 1 delayed recall Range value: 0, ..., 100
\item \strong{STORY1DEL_SPEECH_DURATION} \emph{numeric} Speech duration for story 1 delayed recall Range value: 0, ..., 231.5
\item \strong{STORY1DEL_NONSPEECH_PROPORTION} \emph{numeric} Non-speech proportion of time for story 1 delayed recall Range value: 0, ..., 0.934402332361516
\item \strong{STORY1DEL_SPEECH_RATE} \emph{numeric} Speech rate for story 1 delayed recall Range value: 0.0551724137931034, ..., 7.5
\item \strong{STORY1DEL_ARTICULATION_RATE} \emph{numeric} Articulation rate for story 1 delayed recall Range value: 0.0840336134453782, ..., 7.5
\item \strong{STORY1DEL_NUM_PAUSES} \emph{numeric} Number of pauses for story 1 delayed recall Range value: 0, ..., 48
\item \strong{STORY1DEL_MEAN_PAUSE_LENGTH} \emph{numeric} Mean length of pauses for story 1 delayed recall Range value: 0.3, ..., 64.2
\item \strong{STORY1DEL_AMPLITUDE_SD} \emph{numeric} Standard deviation of the amplitude for story 1 delayed recall Range value: 0, ..., 0.275285631418228
\item \strong{STORY1DEL_F0_MEAN} \emph{numeric} Formant F0 mean for story 1 delayed recall Range value: 61.0377, ..., 241.025
\item \strong{STORY1DEL_F0_SD} \emph{numeric} Formant F0 standard deviation for story 1 delayed recall Range value: 0, ..., 87.8908692043104
\item \strong{STORY1DEL_INTENSITY_MEAN} \emph{numeric} Intensity mean for story 1 delayed recall Range value: 0, ..., 0.163705304265022
\item \strong{STORY1DEL_INTENSITY_SD} \emph{numeric} Intensity standard deviation for story 1 delayed recall Range value: 0, ..., 0.224303364753723
\item \strong{STORY1DEL_MFCC_1_MEAN} \emph{numeric} MFCC 1 mean for story 1 delayed recall Range value: -1131.37109375, ..., -6.73255968093872
\item \strong{STORY1DEL_MFCC_2_MEAN} \emph{numeric} MFCC 2 mean for story 1 delayed recall Range value: -86.9030990600586, ..., 239.915786743164
\item \strong{STORY1DEL_MFCC_3_MEAN} \emph{numeric} MFCC 3 mean for story 1 delayed recall Range value: -94.8164138793945, ..., 87.6731414794922
\item \strong{STORY1DEL_MFCC_4_MEAN} \emph{numeric} MFCC 4 mean for story 1 delayed recall Range value: -36.1702308654785, ..., 104.859107971191
\item \strong{STORY1DEL_MFCC_5_MEAN} \emph{numeric} MFCC 5 mean for story 1 delayed recall Range value: -59.839168548584, ..., 50.9919586181641
\item \strong{STORY1DEL_MFCC_6_MEAN} \emph{numeric} MFCC 6 mean for story 1 delayed recall Range value: -41.1012496948242, ..., 50.9464225769043
\item \strong{STORY1DEL_MFCC_7_MEAN} \emph{numeric} MFCC 7 mean for story 1 delayed recall Range value: -40.8954391479492, ..., 31.0081806182861
\item \strong{STORY1DEL_MFCC_8_MEAN} \emph{numeric} MFCC 8 mean for story 1 delayed recall Range value: -31.3429355621338, ..., 36.5341949462891
\item \strong{STORY1DEL_MFCC_9_MEAN} \emph{numeric} MFCC 9 mean for story 1 delayed recall Range value: -36.9382019042969, ..., 24.6565589904785
\item \strong{STORY1DEL_MFCC_10_MEAN} \emph{numeric} MFCC 10 mean for story 1 delayed recall Range value: -28.2096557617188, ..., 33.9763984680176
\item \strong{STORY1DEL_MFCC_11_MEAN} \emph{numeric} MFCC 11 mean for story 1 delayed recall Range value: -38.313362121582, ..., 18.4537830352783
\item \strong{STORY1DEL_MFCC_12_MEAN} \emph{numeric} MFCC 12 mean for story 1 delayed recall Range value: -23.663948059082, ..., 28.4471168518066
\item \strong{STORY1DEL_MFCC_13_MEAN} \emph{numeric} MFCC 13 mean for story 1 delayed recall Range value: -50.3445472717285, ..., 13.6552019119263
\item \strong{STORY1DEL_MFCC_1_SD} \emph{numeric} MFCC 1 standard deviation for story 1 delayed recall Range value: 0, ..., 190.576751708984
\item \strong{STORY1DEL_MFCC_2_SD} \emph{numeric} MFCC 2 standard deviation for story 1 delayed recall Range value: 0, ..., 102.492218017578
\item \strong{STORY1DEL_MFCC_3_SD} \emph{numeric} MFCC 3 standard deviation for story 1 delayed recall Range value: 0, ..., 64.6785202026367
\item \strong{STORY1DEL_MFCC_4_SD} \emph{numeric} MFCC 4 standard deviation for story 1 delayed recall Range value: 0, ..., 50.8623428344727
\item \strong{STORY1DEL_MFCC_5_SD} \emph{numeric} MFCC 5 standard deviation for story 1 delayed recall Range value: 0, ..., 43.9329948425293
\item \strong{STORY1DEL_MFCC_6_SD} \emph{numeric} MFCC 6 standard deviation for story 1 delayed recall Range value: 0, ..., 29.9269313812256
\item \strong{STORY1DEL_MFCC_7_SD} \emph{numeric} MFCC 7 standard deviation for story 1 delayed recall Range value: 0, ..., 30.547571182251
\item \strong{STORY1DEL_MFCC_8_SD} \emph{numeric} MFCC 8 standard deviation for story 1 delayed recall Range value: 0, ..., 26.5493087768555
\item \strong{STORY1DEL_MFCC_9_SD} \emph{numeric} MFCC 9 standard deviation for story 1 delayed recall Range value: 0, ..., 24.248218536377
\item \strong{STORY1DEL_MFCC_10_SD} \emph{numeric} MFCC 10 standard deviation for story 1 delayed recall Range value: 0, ..., 21.4484958648682
\item \strong{STORY1DEL_MFCC_11_SD} \emph{numeric} MFCC 11 standard deviation for story 1 delayed recall Range value: 0, ..., 20.0861186981201
\item \strong{STORY1DEL_MFCC_12_SD} \emph{numeric} MFCC 12 standard deviation for story 1 delayed recall Range value: 0, ..., 18.094799041748
\item \strong{STORY1DEL_MFCC_13_SD} \emph{numeric} MFCC 13 standard deviation for story 1 delayed recall Range value: 0, ..., 19.0477294921875
\item \strong{STORY1DEL_FORMANTS_1_MEAN} \emph{numeric} Formant F1 mean for story 1 delayed recall Range value: 0, ..., 2243.04
\item \strong{STORY1DEL_FORMANTS_2_MEAN} \emph{numeric} Formant F2 mean for story 1 delayed recall Range value: 0, ..., 3160.56
\item \strong{STORY1DEL_FORMANTS_3_MEAN} \emph{numeric} Formant F3 mean for story 1 delayed recall Range value: 0, ..., 4132.79
\item \strong{STORY1DEL_FORMANTS_4_MEAN} \emph{numeric} Formant F4 mean for story 1 delayed recall Range value: 0, ..., 5150.73
\item \strong{STORY1DEL_FORMANTS_1_SD} \emph{numeric} Formant F1 standard deviation for story 1 delayed recall Range value: 0, ..., 1046.26
\item \strong{STORY1DEL_FORMANTS_2_SD} \emph{numeric} Formant F2 standard deviation for story 1 delayed recall Range value: 0, ..., 1383.15
\item \strong{STORY1DEL_FORMANTS_3_SD} \emph{numeric} Formant F3 standard deviation for story 1 delayed recall Range value: 0, ..., 1737.71
\item \strong{STORY1DEL_FORMANTS_4_SD} \emph{numeric} Formant F4 standard deviation for story 1 delayed recall Range value: 0, ..., 2192.01
\item \strong{STORY1DEL_LOCAL_JITTER} \emph{numeric} Jitter (local) for story 1 delayed recall Range value: 0, ..., 0.0499121719738789
\item \strong{STORY1DEL_LOCAL_ABS_JITTER} \emph{numeric} Jitter (local, absolute) for story 1 delayed recall Range value: 0, ..., 0.00034419061500214
\item \strong{STORY1DEL_RAP_JITTER} \emph{numeric} Jitter (rap) for story 1 delayed recall Range value: 0, ..., 0.023187141762905
\item \strong{STORY1DEL_PPQ5_JITTER} \emph{numeric} Jitter (PPQ5) for story 1 delayed recall Range value: 0, ..., 0.0544022479432006
\item \strong{STORY1DEL_DDP_JITTER} \emph{numeric} Jitter (DDP) for story 1 delayed recall Range value: 0, ..., 0.0695614252887149
\item \strong{STORY1DEL_LOCAL_SHIMMER} \emph{numeric} Shimmer (local) for story 1 delayed recall Range value: 0, ..., 3.65037231263716
\item \strong{STORY1DEL_LOCAL_DB_SHIMMER} \emph{numeric} Shimmer (local, dB) for story 1 delayed recall Range value: 0, ..., 3.27713578939438
\item \strong{STORY1DEL_APQ3_SHIMMER} \emph{numeric} Shimmer (APQ3) for story 1 delayed recall Range value: 0, ..., 1.15465222648567
\item \strong{STORY1DEL_APQ5_SHIMMER} \emph{numeric} Shimmer (APQ5) for story 1 delayed recall Range value: 0, ..., 2.3392710108845
\item \strong{STORY1DEL_APQ11_SHIMMER} \emph{numeric} Shimmer (APQ11) for story 1 delayed recall Range value: 0.00464633656313258, ..., 5.23005269083713
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{remotely_collected_data}
\keyword{undefined_phase}
