% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RS}
\alias{RS}
\title{[DERIVED]  Derived  Clinical Classification}
\format{
A data frame with 15881 observations and 17 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(RS)
}
\description{
Clinical classification derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier ; , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbrevation , with character options: RS
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{RSSEQ} \emph{integer} Sequence Number Range value: 1, ..., 20
\item \strong{RSGRPID} \emph{factor} Group ID Study Protocol of Data Collection; , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{RSTESTCD} \emph{character} Assessment Short Name , with character options: DX
\item \strong{RSTEST} \emph{character} Assessment Name , with character options: Clinical Diagnostics Status
\item \strong{RSCAT} \emph{character} Category for Assessment , with character options: DIAGNOSTICS
\item \strong{RSORRES} \emph{character} Result or Finding in Original Units , with character options: DEM, MCI, CN
\item \strong{RSSTRESC} \emph{character} Character Result/Finding in Std Format , with character options: DEM, MCI, CN
\item \strong{RSBLFL} \emph{character} Baseline Flag , with character options: Y
\item \strong{RSEVAL} \emph{numeric} Evaluator SITEID will be used instead of actual evaluator initial.; Range value: 1, ..., 941
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -2, ..., 999
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{EPOCH} \emph{character} Epoch , with character options: Baseline, Follow-up, Screening
\item \strong{RSDTC} \emph{Date} Date/Time of Assessment Date: YYY-MM-DD
\item \strong{RSDY} \emph{integer} Study Day of Assessment Range value: -1849, ..., 7208
}
}
\keyword{derived_dataset}
