% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RURALITY}
\alias{RURALITY}
\title{Rurality – RUCA & RUCC}
\format{
A data frame with 1514 observations and 20 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RURALITY)
}
\description{
Rurality – RUCA & RUCC data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI3, ADNI4, ADNI2, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{RURDATE} \emph{ } Date of collection: Date: YYY-MM-DD
\item \strong{RUCA} \emph{character} RUCA Designation 1=Rural ; 2=Urban ; 3=Not Coded ; 99=Zip code not known/refused
\emph{Decoded Value: }\strong{Yes}
\item \strong{RUCC} \emph{character} RUCC: 1=1 ; 10=10 ; 2=2 ; 3=3 ; 4=4 ; 5=5 ; 6=6 ; 7=7 ; 8=8 ; 9=9 ; 99=County name not known/refused
\emph{Decoded Value: }\strong{Yes}
\item \strong{RURREV} \emph{character} Did the participant move since their last visit? 0=No ; 1=Yes ; 2=Not applicable
\emph{Decoded Value: }\strong{Yes}
\item \strong{RUCA_2010} \emph{character} RUCA_2010 1=1 ; 10=10 ; 2=2 ; 3=3 ; 4=4 ; 5=5 ; 6=6 ; 7=7 ; 8=8 ; 9=9 ; 99=99
\emph{Decoded Value: }\strong{Yes}
\item \strong{RUCC_2023} \emph{character} RUCC_2023 1=1 ; 10=10 ; 2=2 ; 3=3 ; 4=4 ; 5=5 ; 6=6 ; 7=7 ; 8=8 ; 9=9 ; 99=County name not known/refused
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 1269, ..., 128245
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1, v2, v3, v4
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{subject_characteristics}
