% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RYFF}
\alias{RYFF}
\title{RYFF Psychological Well-Being Purpose in Life}
\format{
A data frame with 901 observations and 29 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(RYFF)
}
\description{
RYFF Psychological Well-Being Purpose in Life data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI2, ADNIGO, ADNI3, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_bl, 4_init, 4_m12, 4_m24
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{RYFF1} \emph{character} 1. I feel good when I think of what I've done in the past and what I hope to do in the future. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF2} \emph{character} 2. I live life one day at a time and don't really think about the future. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF3} \emph{character} 3. I tend to focus on the present, because the future nearly always brings me problems. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF4} \emph{character} 4. I have a sense of direction and purpose in life. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF5} \emph{character} 5. My daily activities often seem trivial and unimportant to me. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF6} \emph{character} 6. I don't have a good sense of what it is I'm trying to accomplish in life. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF7} \emph{character} 7. I used to set goals for myself, but that now seems like a waste of time. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF8} \emph{character} 8. I enjoy making plans for the future and working to make them a reality. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF9} \emph{character} 9. I am an active person in carrying out the plans I set for myself. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF10} \emph{character} 10. Some people wander aimlessly through life, but I am not one of them. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF11} \emph{character} 11. I sometimes feel as if I've done all there is to do in life. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF12} \emph{character} 12. My aims in life have been more a source of satisfaction than frustration to me. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF13} \emph{character} 13. I find it satisfying to think about what I have accomplished in life. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{RYFF14} \emph{character} 14. In the final analysis, I'm not so sure that my life adds up to much. 1=STRONGLY DISAGREE ; 2=DISAGREE SOMEWHAT ; 3=DISAGREE SLIGHTLY ; 4=AGREE SLIGHTLY ; 5=AGREE SOMEWHAT ; 6=STRONGLY AGREE
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 1286, ..., 128218
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
