% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SC}
\alias{SC}
\title{[DERIVED]  Derived  Subject Characteristics}
\format{
A data frame with 129462 observations and 16 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(SC)
}
\description{
Subject characteristics derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier ; , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbrevation , with character options: SC
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{SCSEQ} \emph{integer} Sequence Number Range value: 1, ..., 133
\item \strong{SCGRPID} \emph{factor} Group ID Study Protocol of Data Collection; , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{SCTESTCD} \emph{character} Subject Characteristic Short Name  
\item \strong{SCTEST} \emph{character} Subject Characteristic  
\item \strong{SCCAT} \emph{character} Category for Subject Characteristic , with character options: Demographic Records, Area Deprivation Index, Rurality
\item \strong{SCORRES} \emph{character} Result or Finding in Original Units  
\item \strong{SCSTRESN} \emph{numeric} Numeric Result/Finding in Standard Units Range value: 0, ..., 9999
\item \strong{SCBLFL} \emph{character} Baseline Flag , with character options: Y
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -2, ..., 49
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{EPOCH} \emph{character} Epoch , with character options: Screening, Follow-up, Baseline
\item \strong{SCDTC} \emph{Date} Date/Time of Collection Date: YYY-MM-DD
\item \strong{SCDY} \emph{integer} Study Day of Collection Range value: -363, ..., 7208
}
}
\keyword{derived_dataset}
