% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacc.R
\name{set_as_tibble}
\alias{set_as_tibble}
\title{Make Similar Format}
\usage{
set_as_tibble(.data)
}
\arguments{
\item{.data}{Data.frame}
}
\value{
A tibble/data.frame object with upper-case column name and character type.
}
\description{
A function to convert a data.frame object into tibble format with upper-case
column name and character type.
}
\keyword{utils_fun}
