% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_datadict_tbl}
\alias{set_datadict_tbl}
\title{Set \code{datadict_tbl} object class}
\usage{
set_datadict_tbl(.data)
}
\arguments{
\item{.data}{Input data}
}
\value{
The same as the input data with additional class type of \code{datadict_tbl}.
}
\description{
Set \code{datadict_tbl} object class
}
\keyword{adni_datadic_fun}
