% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPAP_AVID_FLORBETAPIR}
\alias{SPAP_AVID_FLORBETAPIR}
\title{sPAP Avid ADNI Florbetapir summaries}
\format{
A data frame with 604 observations and 21 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(SPAP_AVID_FLORBETAPIR)
}
\description{
sPAP Avid ADNI Florbetapir summaries data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 4530
\item \strong{VISCODE} \emph{character} Visit code , with options: m60, v06, m48, m36, bl, v03
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: m60, m72, m48, m36, bl
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{DIAGRP2} \emph{character} clinical diagnosis close to the AV45 scan available during the data download in August 2012 , with options: NL, MCI, EMCI, AD
\item \strong{SPAP_GLOBAL_SUVR} \emph{numeric} Average six region SUVr normalized by entire cerebellum for flroetbapir (AV-45) images calcualted using sPAP software. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.77, ..., 2.18
\item \strong{SPAP_FRONTAL_SUVR} \emph{numeric} medial orbital frontal normalized to entire cerebellum in sPAP. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.51, ..., 2.16
\item \strong{SPAP_TEMPORAL_SUVR} \emph{numeric} lateral temporal normalized to entire cerebellum in sPAP. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.91, ..., 2.15
\item \strong{SPAP_ANTERIOR_CINGULATE_SUVR} \emph{numeric} anterior cingulate normalized to entire cerebellum in sPAP. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.66, ..., 2.2
\item \strong{SPAP_POSTERIOR_CINGULATE_SUVR} \emph{numeric} posterior cingulate normalized to entire cerebellum is sPAP. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.71, ..., 2.28
\item \strong{SPAP_PARIETAL_SUVR} \emph{numeric} parietal region normalized to entire cerebellum is sPAP. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.7, ..., 2.03
\item \strong{SPAP_PRECUNEUS_SUVR} \emph{numeric} precuneus region normalized to entire cerebellum is sPAP. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.69, ..., 2.45
\item \strong{AVID_STAGE_4_GLOBAL_SUVR} \emph{numeric} Average six region SUVr normalized by entire cerebellmu for florbetapir (AV-45) images calcualted using AVID semi-automated quantification. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.78, ..., 2.25
\item \strong{AVID_STAGE_4_FRONTAL_MEDIAL_ORBITAL_SUVR} \emph{numeric} medial orbital frontal normalized to entire cerebellum using Avid semi-automated method. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.53, ..., 2.07
\item \strong{AVID_STAGE_4_TEMPORAL_SUVR} \emph{numeric} lateral temporal normalized to entire cerebellum using Avid semi-automated method. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.78, ..., 2.04
\item \strong{AVID_STAGE_4_PARIETAL_SUVR} \emph{numeric} parietal normalized to entire cerebellum using Avid semi-automated method. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.53, ..., 2.09
\item \strong{AVID_STAGE_4_PRECUNEUS_SUVR} \emph{numeric} precuneus normalized to entire cerebellum using Avid semi-automated method. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.78, ..., 2.56
\item \strong{AVID_STAGE_4_ANTERIOR_CINGULATE_SUVR} \emph{numeric} anterior cingulate normalized to entire cerebellum using Avid semi-automated method. Refer to sPAP_Avid_Florbetapir_Analysis_Methods PDF on LONI for details Range value: 0.66, ..., 2.39
\item \strong{AVID_STAGE_4_POSTERIOR_CINGULATE_SUVR} \emph{numeric} s Range value: 0.69, ..., 2.41
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
