% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_strings}
\alias{split_strings}
\title{Split Strings With Two Split Parameters}
\usage{
split_strings(input_string, spliter1 = "; |;", spliter2 = "=| = ")
}
\arguments{
\item{input_string}{Input string/character}

\item{spliter1}{First split parameter/pattern, Default: '; |;'}

\item{spliter2}{Second split parameter/pattern, Default: '=| = '}
}
\value{
A data.frame that includes prefix and suffix values.
}
\description{
This function is used to split a string using two split parameters and return a data.frame.
}
\examples{
\dontrun{
input_string <- "1=BUTTER; 2=ARM; 3=SHORE; 4=LETTER; 5=QUEEN; 6=CABIN"
split_strings(
  input_string = input_string,
  spliter1 = "; ",
  spliter2 = "="
)
}
}
\seealso{
\code{\link{get_factor_levels_datadict}()}
}
\concept{Split strings}
\keyword{utils_fun}
