% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STAIAD}
\alias{STAIAD}
\title{Mini-State Trait Anxiety Inventory (STAI) for Adults™}
\format{
A data frame with 360 observations and 22 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(STAIAD)
}
\description{
Mini-State Trait Anxiety Inventory (STAI) for Adults™ data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI4, ADNI2, ADNI3, ADNIGO, ADNI1
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 69, ..., 10875
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_bl, 4_init, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{METHOD} \emph{character} Method of collection 1=In-person/Self ; 2=By phone/Interview
\emph{Decoded Value: }\strong{Yes}
\item \strong{STAIAD1} \emph{character} 1. I felt calm. 1=Not at All ; 2=Somewhat ; 3=Moderately So ; 4=Very Much
\emph{Decoded Value: }\strong{Yes}
\item \strong{STAIAD2} \emph{character} 2. I was tense. 1=Not at All ; 2=Somewhat ; 3=Moderately So ; 4=Very Much
\emph{Decoded Value: }\strong{Yes}
\item \strong{STAIAD3} \emph{character} 3. I felt upset. 1=Not at All ; 2=Somewhat ; 3=Moderately So ; 4=Very Much
\emph{Decoded Value: }\strong{Yes}
\item \strong{STAIAD4} \emph{character} 4. I was relaxed. 1=Not at All ; 2=Somewhat ; 3=Moderately So ; 4=Very Much
\emph{Decoded Value: }\strong{Yes}
\item \strong{STAIAD5} \emph{character} 5. I felt content. 1=Not at All ; 2=Somewhat ; 3=Moderately So ; 4=Very Much
\emph{Decoded Value: }\strong{Yes}
\item \strong{STAIAD6} \emph{character} 6. I was worried. 1=Not at All ; 2=Somewhat ; 3=Moderately So ; 4=Very Much
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 1372, ..., 127834
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
