% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAUMETA}
\alias{TAUMETA}
\title{Tau AV-1451 PET Scan Information}
\format{
A data frame with 3067 observations and 40 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(TAUMETA)
}
\description{
Tau AV-1451 PET Scan Information data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI1, ADNIGO, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{SCANDATE} \emph{ } Scan Date/Dose Date Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 2
\item \strong{DONE} \emph{character} Was the scan conducted? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 
\itemize{
\item \emph{ADNI2:} 1=Participant Unable for Cognitive Reasons ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 4=Equipment related issue ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\item \emph{ADNI3:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Not called for by protocol ; 8=Software/scanner error ; 99=Other (specify)
\item \emph{ADNI4:} 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Not called for by protocol ; 8=Software/scanner error ; 9=Tracer issues ; 99=Other (specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{TRACERISS} \emph{numeric} Tracer issues 1=Dose failure ; 2=Late delivery or other issues with delivery ; 3=Incorrect tracer delivered by radiopharmacy ; 4=Incorrect tracer ordered by site ; 99=Other (specify)
\item \strong{RADTRACER} \emph{character} If scan not done, was radiotracer administered? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{TRACERTYPE} \emph{character} Tau Imaging radiotracer used 
\itemize{
\item \emph{ADNI2, ADNI3:} 1=Flortaucipir
\item \emph{ADNI4:} 1=Flortaucipir ; 2=MK-6240 ; 3=PI-2620
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{QCTIME} \emph{character} Time of today's Scanner QC  
\item \strong{ASSAYTIME} \emph{character} Time of dose assay  
\item \strong{DOSETAU} \emph{numeric} Flortaucipir :  Injected dose assay Range value: 4.9, ..., 13.5
\item \strong{DOSEMK} \emph{numeric} MK6240:  Injected dose assay Range value: 3.5, ..., 5.6
\item \strong{DOSEPI} \emph{numeric} PI-2620: Injected dose assay Range value: 4.8, ..., 5.7
\item \strong{INJTIME} \emph{character} Time of injection  
\item \strong{SCANTIME} \emph{character} Emission Scan Start Time  
\item \strong{VARIAT} \emph{character} Any variations from protocol during uptake? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PROTID} \emph{character} Predefined Acquisition Protocol Name  
\item \strong{TAUFRAME} \emph{character} For the acquisition, which framing rate was used? 1=6 frames, 5 min/frame (6x300s) ; 2=2 scans,15 min each (2x900s) (scans started as close together as possible, only for BioGraph scanners without list-mode) ; 3= 1 frame, 30 min (1 x 1200s) (only for Biograph 47-slice scanners) ; 8=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{MOTION} \emph{character} Participant motion problems 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{OTHERVAR} \emph{character} Other protocol variations 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{RECON} \emph{character} Select reconstruction used 
\itemize{
\item \emph{ADNI2:} 1=3D or LOR Ramla (Philips) ; 2=FORE/2D-OSEM (Siemens) ; 3=3D Back-projection (GE) ; 4=OSEM3D (Siemens) (If HHRT scanners using OP, select OSEM3D) ; 5=3D Iterative (GE) ; 6=FORE Iterative (GE)
\item \emph{ADNI3:} 1=3D or LOR Ramla (Philips) ; 2=FORE/2D-OSEM (Siemens) ; 4=3D-OSEM (Siemens) ; 5=3D Iterative (GE) ; 6=FORE Iterative (GE)
\item \emph{ADNI4:} 1=3D or LOR Ramla (for Philips Gemini scanners) ; 2=FORE/2D-OSEM (Siemens) ; 4=3D-OSEM (Siemens) ; 5=3D Iterative (GE) ; 6=FORE Iterative (GE) ; 7=3D-OSEM (Philips Vereos and Ingenuity Scanners)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{SHARP} \emph{character} For Philips Gemini scanners: was "smooth" parameter set to "sharp"? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{TAUSMOOTH} \emph{character} Was post-process smoothing performed: 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{TAURAMLA} \emph{numeric} Lambda Range value: -1, ..., 1
\item \strong{ARCHIVE} \emph{character} Data archived locally? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{LONI} \emph{character} Data transferred to LONI? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{TRANDATE} \emph{ } Transfer Date Date: YYY-MM-DD
\item \strong{ID} \emph{numeric} Record ID Range value: 2, ..., 186569
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{imaging}
