% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAUQC}
\alias{TAUQC}
\title{Tau AV-1451 PET QC}
\format{
A data frame with 2573 observations and 29 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(TAUQC)
}
\description{
Tau AV-1451 PET QC data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI1, ADNIGO, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10888
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{SCANDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 3
\item \strong{LONIUPDT} \emph{ } Date uploaded to LONI Date: YYY-MM-DD
\item \strong{LONIUID} \emph{character} LONI Unique Image ID  
\item \strong{REVDT} \emph{ } Date Reviewed Date: YYY-MM-DD
\item \strong{ALLFRAME} \emph{character} Were all frames acceptable? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{UNUSABL} \emph{character} If No, indicate which frames were unacceptable  
\item \strong{UNRSN} \emph{character} If No, indicate why frames were unacceptable 1=Subject motion ; 2=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{UNRSNSPEC} \emph{character} If Other, Specify  
\item \strong{SCANQLTY} \emph{character} Scan Pass QC? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{TECHERR} \emph{character} If No, was this due to technologist error? 1=Yes ; 0=No
\emph{Decoded Value: }\strong{Yes}
\item \strong{REPROCREQ} \emph{character} Reprocessing requested? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PROCERR} \emph{character} If Yes, select processing error(s) 
\itemize{
\item \emph{ADNI2, ADNI3:} 1=Incorrect reconstruction algorithm ; 2=Incorrect filtering ; 3=Incorrect slice thickness ; 4=Incorrect voxel size ; 5=Attenuation Error ; 6=Normalization Error ; 7=Inadequate counts ; 8=Inadequate field of view ; 9=Incorrect acquisition protocol ; 10=Unfixable image artifact (e.g. attenuation, normalization, etc.) ; 99=Other (specify)
\item \emph{ADNI4:} 1=Incorrect reconstruction algorithm ; 10=Unfixable image artifact (e.g. attenuation, normalization, etc.) ; 2=Incorrect filtering ; 3=Incorrect slice thickness ; 4=Incorrect voxel size ; 5=Attenuation Error ; 6=Normalization Error ; 7=Inadequate counts ; 8=Inadequate field of view ; 9=Incorrect acquisition protocol ; 99=Other (specify)
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{PROCERRSPEC} \emph{character} If Other, Specify  
\item \strong{RESCANREQ} \emph{character} Rescan requested? 
\itemize{
\item \emph{ADNI2:} 1=Yes ; 0=No
\item \emph{ADNI3, ADNI4:} 0=No ; 1=Yes
}
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 2, ..., 187095
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{imaging}
