% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TBM22}
\alias{TBM22}
\title{Paul Thompson's Lab, Imaging Genetics Center (IGC), University of Southern California - Cross-sectional and longitudinal tensor-based morphometry Versions 2.0 and 2.1: 2020-03-17}
\format{
A data frame with 9354 observations and 18 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(TBM22)
}
\description{
Paul Thompson's Lab, Imaging Genetics Center (IGC), University of Southern California - Cross-sectional and longitudinal tensor-based morphometry Versions 2.0 and 2.1: 2020-03-17 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 3, ..., 6891
\item \strong{VISCODE} \emph{character} Visit code , with options: m06, m12, m24, m18, m36, m48, m03, m72, m60
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{numeric}   Range value: 2, ..., 2.1
\item \strong{IMAGEUID_1} \emph{numeric} IMAGE ID of the screening scan Range value: 31084, ..., 1353817
\item \strong{IMAGEUID_2} \emph{numeric} IMAGE ID of the follow-up scan; see VISCODE Range value: 31008, ..., 1523844
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   , with options: complete, Complete
\item \strong{MEASURE_1} \emph{numeric} Numerical summary of cumulative temporal lobe atrophy  average within a statistically defined region-of-interest (p<0.00001) inside the temporal lobes; see VISCODE; summaries are scaled by 1000 (e.g. 1000: no change, 1200: 20\% increase, 800: 20\% loss) Range value: 885.309, ..., 1036.37
\item \strong{MEASURE_2} \emph{numeric} Numerical summary of cumulative temporal lobe atrophy  average within an anatomically defined region-of-interest including bilateral temporal lobes; see VISCODE; summaries are scaled by 1000 (e.g. 1000: no change, 1200: 20\% increase, 800: 20\% loss) Range value: 945.397, ..., 1023.9
\item \strong{SEQUENCE} \emph{character}    
\item \strong{ACCEL_NONACCEL} \emph{character}   , with options: nonaccel, accel
\item \strong{DESCRIPTION} \emph{character}   , with options: MPR; GradWarp; B1 Correction; N3; Scaled, MPR-R; GradWarp; B1 Correction; N3; Scaled, MPR; ; N3; Scaled, MPR-R; GradWarp; N3; Scaled, MPR; GradWarp; N3; Scaled, MPR-R; ; N3; Scaled, MT1; GradWarp; N3m, MT1; N3m
\item \strong{FIELD_STRENGTH} \emph{character}   , with options: 1.5T, 3.0T
\item \strong{EXTRA_NOTES_ON_PROCESSED_IMAGES} \emph{character}   , with options: score from TBM_11_17_13 upload csv is replaced with this score, uses sc mask, uses 6p image and mask, score from TBM_11_17_21 upload csv is replaced with this score, score from TBM_12_13_21 upload csv is replaced with this score; uses 6p image and mask, score from TBM_12_13_21 upload csv is replaced with this score, score from TBM_04_20_20 upload csv is replaced with this score; uses 6p image and mask
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{imaging}
