% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCV}
\alias{TCV}
\title{Total Cranial Vault Segmentation}
\format{
A data frame with 810 observations and 11 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(TCV)
}
\description{
Total Cranial Vault Segmentation data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant ID Range value: 2, ..., 1435
\item \strong{VISCODE} \emph{character} Visit Code sc=screen
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{LONIUID} \emph{character} Study ID  
\item \strong{RUNDATE} \emph{ } Run Date Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Status complete ; partial
\item \strong{VERSION} \emph{numeric} Version Number Range value: 1, ..., 1
\item \strong{T2TCV} \emph{numeric} T2 Total Intracranial Volume Range value: 1002.95, ..., 1922.29
\item \strong{GRADE} \emph{numeric} Grade 3 = excellent ; 2 = good ; 1 = unsatisfactory
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
