% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TREATDIS}
\alias{TREATDIS}
\title{Early Discontinuation and Withdrawal}
\format{
A data frame with 1889 observations and 18 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(TREATDIS)
}
\description{
Early Discontinuation and Withdrawal data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5295
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 6
\item \strong{EXAMDATE} \emph{character} Examination Date  
\item \strong{EXAMDATEYR_DRVD} \emph{numeric} Year of Discontinuation/Withdrawal/Completion oF ADNI2 (Derived from EXAMDATE) Range value: 2000, ..., 2019
\item \strong{WDRAWTYPE} \emph{numeric} 1. Is this a Full or Partial Withdrawal? 
\itemize{
\item \emph{ADNI1:} 1=Full ; 2=Partial ; 3=Completion of ADNI 1, not continuing to extension protocol
\item \emph{ADNIGO:} 1=Full ; 2=Partial ; 3=Completion of ADNIGO, not continuing to extension protocol
\item \emph{ADNI2:} 1=Full ; 2=Partial
}
\item \strong{WDPARTIAL} \emph{character} If Partial, what is the participant withdrawing from? 
\itemize{
\item \emph{ADNI1:} 1=1.5 T MRI ; 2=3.0 T MRI ; 3=PET ; 4=Lumbar Puncture ; 5=PIB
\item \emph{ADNIGO:} 1=1.5 T MRI ; 2=3.0 T MRI ; 3=FDG PET ; 6=AV-45 PET ; 4=Lumbar Puncture ; 7=consent to store/share samples/clinical data ; 8=DNA sample collection ; 9=consent to store and share DNA ; 10=RNA sample collection ; 11=consent to store and share RNA ; 12=provisional autopsy consent
\item \emph{ADNI2:} 1=1.5 T MRI ; 2=3.0 T MRI ; 3=FDG PET ; 6=AV-45 PET ; 20 =Tau AV-1451 F 18 ; 4=Lumbar Puncture ; 7=consent to store/share samples/clinical data ; 8=DNA sample collection ; 9=consent to store and share DNA ; 10=RNA sample collection ; 11=consent to store and share RNA ; 12=provisional autopsy consent ; 13=Full Clinic Visits (agrees to only telephone follow-up) ; 21 = Cogstate Brief Battery
}
\item \strong{WDREASON} \emph{character} Reason for Withdrawal 
\itemize{
\item \emph{ADNI1:} 1=Adverse Event ; 2=Death ; 3=Safety Risk ; 4=Protocol Violation ; 5=Non-Compliance ; 6=Investigator Judgment ; 7=Consent Withdrawn ; 8=Study Terminated ; 9=Loss of Study Partner ; 10=Lost to Follow-Up ; 11=Coordinating Center Request
\item \emph{ADNIGO, ADNI2:} 1=Death (SAE required, provide details below) ; 2=Withdrawn Consent ; 3=Perceived Safety Risk ; 4=Protocol Violations ; 5=Non-Compliance with Protocol ; 6=Adverse Event (provide detailed explanation below) ; 7=Elects to enroll in clinical trial ; 8=Contraindication to Procedure (e.g. pacemaker) ; 9=Elects to start an excluded FDA Approved treatment for AD (patient/caregiver) ; 10=Participant Burden (provide detailed explanation below) ; 11=Loss of Study Partner with no replacement available ; 12=Study Partner Burden ; 13=Study Partner unwilling or unable to participate (provide detailed explanation below) ; 14=Subject Moved: unable to transfer or to travel distance for visits ; 15=Subject Moved to Nursing Facility ; 16=Study Partner Moved: unable to transfer or to travel distance for visits ; 17=Study Terminated (Mandated by Project Director or IRB) ; 18=Non-responsive (Registered mail returned, not showing for appointments, not answering phone calls) ; 19=Did not understand consent ; 20=Compensation inadequate/ not competitive ;  21=Other (provide detailed explanation below) ;
}
\item \strong{WDFOLLOWUP} \emph{character} 2. Follow-up:  (check all that apply) 
\itemize{
\item \emph{ADNI1:} 1=Agrees to return for all follow-up visits. ; 2=Agrees to return for final visit. ; 3=Agrees to unscheduled visit. ; 4=Refuses/unable to return for any future visits.
\item \emph{ADNIGO:} 1=Agrees to return for all follow-up visits. ; 2=Agrees to return for final visit. ; 4=Refuses/unable to return for any future visits.
\item \emph{ADNI2:} 1=Agrees to return for all follow-up visits. ; 5=Agrees to follow-up phone visits only ; 2=Agrees to return for a final in-clinic visit. ; 4=Refuses/unable to return for any future visits.
}
\item \strong{ID} \emph{numeric}   Range value: 2, ..., 2346
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 127
\item \strong{USERDATE} \emph{ } Date record created Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } Date record last updated Date: YYY-MM-DD
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{enrollment}
