% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UASPMVBM}
\alias{UASPMVBM}
\title{UA (Gene Alexander) MRI SPM voxel based morphometry (VBM) analysis}
\format{
A data frame with 4091 observations and 128 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UASPMVBM)
}
\description{
UA (Gene Alexander) MRI SPM voxel based morphometry (VBM) analysis data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{ID} \emph{numeric}   Range value: 1, ..., 4091
\item \strong{RID} \emph{numeric}   Range value: 3, ..., 1435
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character}   , with options: sc, m06, m12, bl, f
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   1 = 1.5T cross-sectional screen ; 2 = 1.5T MO_06 longitudinal ; 3 = 1.5T MO_12 longitudinal ; 4 = 3T cross-sectional baseline ; 5 = 3T MO_06 longitudinal ; 6 = 3T MO_12 longitudinal
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{IMAGEUID} \emph{numeric} imageUID Range value: 31084, ..., 125949
\item \strong{FLDSTRENG} \emph{numeric} Field Strength Range value: 1.5, ..., 3
\item \strong{PRECENTL} \emph{numeric} Precentral_L. 1) Mean gray matter values extracted from 1.5T ADNI MPRAGE MRI segmented gray matter maps using longitudinal voxel based morphometry (VBM) processing steps (Chetelat et al., Neuroimage, 2005, 27:934-946) adapted for SPM5 with customized tissue-specific priors derived from a representative sample of 300 ADNI subjects (ADNI300). The regions of interest (ROI) were defined by Anatomical Automatic Labeling (AAL, Tzourio-Mazoyer et al., Neuroimage, 2002, 15:273-89, http://www.cyceron.fr/freeware/) after warping the standard AAL ROI template into the customized sample-specific ADNI300 template space. The ADNI300-customized AAL ROI template was applied to spatially normalized and segmented SPM5 gray matter maps for the longitudinal VBM generated baseline and 6 month (indicated by RECNO=2) or 12 month (indicated by RECNO=3) follow up scans without Gaussian smoothing to reduce averaging across regional borders. 2) The ADNI300 template was created from AD, MCI, and Normal subjects proportional to the distribution of these subject groups enrolled in the whole ADNI study and the groups were matched in age and gender.  3) Regional values will be subject to change due to study alterations in image pre- or post-processing, inclusion of additional analysis time points, and/or modifications or updates in image analysis procedures. 4) SPM analyses were performed by ADNI Analysis Team Members: Gene E. Alexander, Ph.D., Kewei Chen, Ph.D., and Eric M. Reiman, M.D. Range value: 0.109464, ..., 0.519062
\item \strong{PRECENTR} \emph{numeric} Precentral_R Range value: 0.087301, ..., 0.519366
\item \strong{FRONTSUPL} \emph{numeric} Frontal_Sup_L Range value: 0.072548, ..., 0.572452
\item \strong{FRONTSUPR} \emph{numeric} Frontal_Sup_R Range value: 0.06832, ..., 0.495008
\item \strong{FRONTSORBL} \emph{numeric} Frontal_Sup_Orb_L Range value: 0.079956, ..., 0.664292
\item \strong{FRONTSORBR} \emph{numeric} Frontal_Sup_Orb_R Range value: 0.070114, ..., 0.646994
\item \strong{FRONTMIDL} \emph{numeric} Frontal_Mid_L Range value: 0.093176, ..., 0.60747
\item \strong{FRONTMIDR} \emph{numeric} Frontal_Mid_R Range value: 0.0698, ..., 0.626904
\item \strong{FRTMIDORBL} \emph{numeric} Frontal_Mid_Orb_L Range value: 0.073239, ..., 0.75648
\item \strong{FRTMIDORBR} \emph{numeric} Frontal_Mid_Orb_R Range value: 0.061234, ..., 0.745696
\item \strong{FRONTINOPL} \emph{numeric} Frontal_Inf_Oper_L Range value: 0.131869, ..., 0.626053
\item \strong{FRONTINOPR} \emph{numeric} Frontal_Inf_Oper_R Range value: 0.115524, ..., 0.628719
\item \strong{FRONTINTRL} \emph{numeric} Frontal_Inf_Tri_L Range value: 0.139826, ..., 0.563734
\item \strong{FRONTINTRR} \emph{numeric} Frontal_Inf_Tri_R Range value: 0.110499, ..., 0.569364
\item \strong{FRONTINOBL} \emph{numeric} Frontal_Inf_Orb_L Range value: 0.113868, ..., 0.750045
\item \strong{FRONTINOBR} \emph{numeric} Frontal_Inf_Orb_R Range value: 0.092592, ..., 0.696247
\item \strong{ROLANDOPL} \emph{numeric} Rolandic_Oper_L Range value: 0.152697, ..., 0.668278
\item \strong{ROLANDOPR} \emph{numeric} Rolandic_Oper_R Range value: 0.163112, ..., 0.689594
\item \strong{SUPMOTORL} \emph{numeric} Supp_Motor_Area_L Range value: 0.083472, ..., 0.622019
\item \strong{SUPMOTORR} \emph{numeric} Supp_Motor_Area_R Range value: 0.113827, ..., 0.607832
\item \strong{OLFACTL} \emph{numeric} Olfactory_L Range value: 0.171352, ..., 0.813431
\item \strong{OLFACTR} \emph{numeric} Olfactory_R Range value: 0.172559, ..., 0.758829
\item \strong{FRONTSMEDL} \emph{numeric} Frontal_Sup_Medial_L Range value: 0.074001, ..., 0.581061
\item \strong{FRONTSMEDR} \emph{numeric} Frontal_Sup_Medial_R Range value: 0.084855, ..., 0.660669
\item \strong{FRTMEDORBL} \emph{numeric} Frontal_Med_Orb_L Range value: 0.129765, ..., 0.845215
\item \strong{FRTMEDORBR} \emph{numeric} Frontal_Med_Orb_R Range value: 0.150819, ..., 0.810541
\item \strong{RECTUSL} \emph{numeric} Rectus_L Range value: 0.136071, ..., 0.755095
\item \strong{RECTUSR} \emph{numeric} Rectus_R Range value: 0.15509, ..., 0.742195
\item \strong{INSULAL} \emph{numeric} Insula_L Range value: 0.159247, ..., 0.76472
\item \strong{INSULAR} \emph{numeric} Insula_R Range value: 0.149596, ..., 0.760286
\item \strong{CINGANTL} \emph{numeric} Cingulum_Ant_L Range value: 0.130653, ..., 0.660412
\item \strong{CINGANTR} \emph{numeric} Cingulum_Ant_R Range value: 0.200086, ..., 0.667653
\item \strong{CINGMIDL} \emph{numeric} Cingulum_Mid_L Range value: 0.203155, ..., 0.648047
\item \strong{CINGMIDR} \emph{numeric} Cingulum_Mid_R Range value: 0.227015, ..., 0.736913
\item \strong{CINGPOSTL} \emph{numeric} Cingulum_Post_L Range value: 0.141583, ..., 0.730277
\item \strong{CINGPOSTR} \emph{numeric} Cingulum_Post_R Range value: 0.120043, ..., 0.623823
\item \strong{HIPPL} \emph{numeric} Hippocampus_L Range value: 0.195745, ..., 0.862557
\item \strong{HIPPR} \emph{numeric} Hippocampus_R Range value: 0.079445, ..., 0.838027
\item \strong{PARAHIPPL} \emph{numeric} ParaHippocampal_L Range value: 0.194895, ..., 0.785074
\item \strong{PARAHIPPR} \emph{numeric}   Range value: 0.269733, ..., 0.811848
\item \strong{AMYGDL} \emph{numeric}   Range value: 0.125053, ..., 0.970368
\item \strong{AMYGDR} \emph{numeric} Amygdala_R Range value: 0.157966, ..., 0.860282
\item \strong{CALCARINEL} \emph{numeric} Calcarine_L Range value: 0.135687, ..., 0.776054
\item \strong{CALCARINER} \emph{numeric} Calcarine_R Range value: 0.142091, ..., 0.721373
\item \strong{CUNEUSL} \emph{numeric}   Range value: 0.106092, ..., 0.711471
\item \strong{CUNEUSR} \emph{numeric}   Range value: 0.120579, ..., 0.694838
\item \strong{LINGUALL} \emph{numeric}   Range value: 0.2063, ..., 0.734729
\item \strong{LINGUALR} \emph{numeric}   Range value: 0.180107, ..., 0.715062
\item \strong{OCCSUPL} \emph{numeric}   Range value: 0.100288, ..., 0.558346
\item \strong{OCCSUPR} \emph{numeric}   Range value: 0.110552, ..., 0.547554
\item \strong{OCCMIDL} \emph{numeric}   Range value: 0.149659, ..., 0.69116
\item \strong{OCCMIDR} \emph{numeric}   Range value: 0.128717, ..., 0.671495
\item \strong{OCCINFL} \emph{numeric}   Range value: 0.15837, ..., 0.783026
\item \strong{OCCINFR} \emph{numeric}   Range value: 0.105393, ..., 0.814613
\item \strong{FUSIFORML} \emph{numeric}   Range value: 0.243454, ..., 0.814792
\item \strong{FUSIFORMR} \emph{numeric}   Range value: 0.214296, ..., 0.806602
\item \strong{POSTCENTL} \emph{numeric}   Range value: 0.132477, ..., 0.539285
\item \strong{POSTCENTR} \emph{numeric}   Range value: 0.098583, ..., 0.525414
\item \strong{PARIETSUPL} \emph{numeric}   Range value: 0.106263, ..., 0.645473
\item \strong{PARIETSUPR} \emph{numeric}   Range value: 0.082337, ..., 0.599451
\item \strong{PARIETINFL} \emph{numeric}   Range value: 0.161056, ..., 0.660393
\item \strong{PARIETINFR} \emph{numeric}   Range value: 0.1459, ..., 0.691876
\item \strong{SUPRAMARGL} \emph{numeric}   Range value: 0.200084, ..., 0.687163
\item \strong{SUPRAMARGR} \emph{numeric}   Range value: 0.152732, ..., 0.625606
\item \strong{ANGULARL} \emph{numeric}   Range value: 0.178645, ..., 0.71849
\item \strong{ANGULARR} \emph{numeric}   Range value: 0.15023, ..., 0.641721
\item \strong{PRECUNEUSL} \emph{numeric}   Range value: 0.152047, ..., 0.633392
\item \strong{PRECUNEUSR} \emph{numeric}   Range value: 0.164068, ..., 0.710013
\item \strong{PARCENTLBL} \emph{numeric}   Range value: 0.081076, ..., 0.550229
\item \strong{PARCENTLBR} \emph{numeric}   Range value: 0.095236, ..., 0.600817
\item \strong{CAUDATEL} \emph{numeric}   Range value: 0.101717, ..., 0.757027
\item \strong{CAUDATER} \emph{numeric}   Range value: 0.128286, ..., 0.719425
\item \strong{PUTAMENL} \emph{numeric}   Range value: 0.23163, ..., 0.857368
\item \strong{PUTAMENR} \emph{numeric}   Range value: 0.241973, ..., 0.752296
\item \strong{PALLIDUML} \emph{numeric}   Range value: 0.025142, ..., 0.584371
\item \strong{PALLIDUMR} \emph{numeric}   Range value: 0.001927, ..., 0.636137
\item \strong{THALAMUSL} \emph{numeric}   Range value: 0.159741, ..., 0.583169
\item \strong{THALAMUSR} \emph{numeric}   Range value: 0.13538, ..., 0.611456
\item \strong{HESCHLL} \emph{numeric}   Range value: 0.123076, ..., 0.703712
\item \strong{HESCHLR} \emph{numeric}   Range value: 0.094019, ..., 0.605857
\item \strong{TEMPSUPL} \emph{numeric}   Range value: 0.213788, ..., 0.643675
\item \strong{TEMPSUPR} \emph{numeric}   Range value: 0.152122, ..., 0.618064
\item \strong{TEMPPLSUPL} \emph{numeric}   Range value: 0.130266, ..., 0.721552
\item \strong{TEMPPLSUPR} \emph{numeric}   Range value: 0.103663, ..., 0.641147
\item \strong{TEMPMIDL} \emph{numeric}   Range value: 0.216623, ..., 0.677134
\item \strong{TEMPMIDR} \emph{numeric}   Range value: 0.166822, ..., 0.679876
\item \strong{TEMPPLMIDL} \emph{numeric}   Range value: 0.088095, ..., 0.800392
\item \strong{TEMPPLMIDR} \emph{numeric}   Range value: 0.071037, ..., 0.723863
\item \strong{TEMPINFL} \emph{numeric}   Range value: 0.190501, ..., 0.710179
\item \strong{TEMPINFR} \emph{numeric}   Range value: 0.147539, ..., 0.749946
\item \strong{CEREBCR1L} \emph{numeric}   Range value: 0.227836, ..., 0.800763
\item \strong{CEREBCR1R} \emph{numeric}   Range value: 0.176948, ..., 0.825526
\item \strong{CEREBCR2L} \emph{numeric}   Range value: 0.181098, ..., 1.47167
\item \strong{CEREBCR2R} \emph{numeric}   Range value: 0.156126, ..., 1.1756
\item \strong{CEREB3L} \emph{numeric}   Range value: 0.175525, ..., 0.800511
\item \strong{CEREB3R} \emph{numeric}   Range value: 0.182603, ..., 0.682015
\item \strong{CEREB45L} \emph{numeric}   Range value: 0.297412, ..., 0.78897
\item \strong{CEREB45R} \emph{numeric}   Range value: 0.341734, ..., 0.858938
\item \strong{CEREB6L} \emph{numeric}   Range value: 0.360521, ..., 0.912022
\item \strong{CEREB6R} \emph{numeric}   Range value: 0.36296, ..., 0.859137
\item \strong{CEREB7BL} \emph{numeric}   Range value: 0.22627, ..., 1.84525
\item \strong{CEREB7BR} \emph{numeric}   Range value: 0.189325, ..., 1.63047
\item \strong{CEREB8L} \emph{numeric}   Range value: 0.243068, ..., 2.01657
\item \strong{CEREB8R} \emph{numeric}   Range value: 0.218044, ..., 1.68857
\item \strong{CEREB9L} \emph{numeric}   Range value: 0.260033, ..., 1.40136
\item \strong{CEREB9R} \emph{numeric}   Range value: 0.126837, ..., 1.1916
\item \strong{CEREB10L} \emph{numeric}   Range value: 0.075762, ..., 0.71962
\item \strong{CEREB10R} \emph{numeric}   Range value: 0.10381, ..., 0.752279
\item \strong{VERMIS12} \emph{numeric}   Range value: 0.08083, ..., 0.818248
\item \strong{VERMIS3} \emph{numeric}   Range value: 0.155702, ..., 0.796297
\item \strong{VERMIS45} \emph{numeric}   Range value: 0.207042, ..., 0.710177
\item \strong{VERMIS6} \emph{numeric}   Range value: 0.187682, ..., 0.841051
\item \strong{VERMIS7} \emph{numeric}   Range value: 0.210638, ..., 0.79367
\item \strong{VERMIS8} \emph{numeric}   Range value: 0.286865, ..., 0.994217
\item \strong{VERMIS9} \emph{numeric}   Range value: 0.365639, ..., 0.956039
\item \strong{VERMIS10} \emph{numeric}   Range value: 0.042025, ..., 0.684707
\item \strong{ETIV} \emph{numeric}   Range value: 1.33109, ..., 2.62237
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
