% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCBERKELEY_AMY_6MM}
\alias{UCBERKELEY_AMY_6MM}
\title{UC Berkeley - Amyloid PET 6mm Res analysis}
\format{
A data frame with 4582 observations and 345 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCBERKELEY_AMY_6MM)
}
\description{
UC Berkeley - Amyloid PET 6mm Res analysis data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI2, ADNI3, ADNIGO, ADNI4
\item \strong{LONIUID} \emph{numeric} LONI Image ID Range value: 193251, ..., 11432065
\item \strong{PTID} \emph{character} ADNI Participant ID consisting of 3-digit "SITEID" and 4-digit "RID"  
\item \strong{RID} \emph{numeric} Participant roster ID (unique participant identifier that should be used to merge data) Range value: 21, ..., 10875
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{SCANDATE} \emph{ } Scan acquisition date Date: YYY-MM-DD
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{PROCESSDATE} \emph{ } Date quantification pipeline was run Date: YYY-MM-DD
\item \strong{IMAGE_RESOLUTION} \emph{character} Spatial resolution in mm^3 FWHM of Step4 Preprocessed PET image. See UCBERKELEY PET 8mm-to-6mm Transformation Methods PDF on LONI for details. , with options: 6mm, 8mm
\item \strong{qc_flag} \emph{numeric} Quality control flag based on visual inspection: 2=Pass;1=Partial pass; 0=Fail; -1=Not assessed; -2=Cannot be processed; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: -2, ..., 2
\item \strong{TRACER} \emph{character} PET Radiotracer , with options: FBB, FBP, NAV
\item \strong{TRACER_SUVR_WARNING} \emph{character} Reminder of the incomparability of SUVRs from different tracers. See UCBERKELEY Amyloid Processing Methods PDF. , with options: USE CENTILOIDS (NOT SUVRs) TO COMPARE ACROSS TRACERS
\item \strong{AMYLOID_STATUS} \emph{numeric} Amyloid positivity determined by applying thresholds to cortical summary SUVR normalized by whole cerebellum; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0, ..., 1
\item \strong{AMYLOID_STATUS_COMPOSITE_REF} \emph{numeric} Amyloid positivity determined by applying thresholds to cortical summary SUVR normalized by composite ref region; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0, ..., 1
\item \strong{CENTILOIDS} \emph{numeric} Summary cortical SUVR normalized by whole cerebellum and transformed to Centiloids. All SUVRs in this file are normalized to the whole cerebellum unless otherwise noted. See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: -39, ..., 346
\item \strong{SUMMARY_SUVR} \emph{numeric} Summary cortical SUVR normalized by whole cerebellum. All SUVRs in this file are normalized to the whole cerebellum unless otherwise noted. See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0.797, ..., 2.842
\item \strong{SUMMARY_VOLUME} \emph{numeric} Summary cortical ROI size in mm^3 Range value: 179914, ..., 423018
\item \strong{WHOLECEREBELLUM_SUVR} \emph{numeric} Reference region - mean SUVR of whole cerebellum.  Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 1, ..., 1
\item \strong{WHOLECEREBELLUM_VOLUME} \emph{numeric} Reference region - volume of whole cerebellum.  Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 85374, ..., 180735
\item \strong{COMPOSITE_REF_SUVR} \emph{numeric} Reference region - mean SUVR of composite ref region (avg of whole cerebellum, brainstem, eroded WM).  Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 1.03, ..., 2.362
\item \strong{COMPOSITE_REF_VOLUME} \emph{numeric} Reference region - volume of composite ref region (avg of whole cerebellum, brainstem, eroded WM).  Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 173936, ..., 592522
\item \strong{CEREBELLUM_CORTEX_SUVR} \emph{numeric} Reference region - mean SUVR of bilateral-cerebellum-cortex defined by Freesurfer (previously CEREBELLUMGREYMATTER_SUVR); See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0.754, ..., 1.06
\item \strong{CEREBELLUM_CORTEX_VOLUME} \emph{numeric} Reference region - volume of bilateral-cerebellum-cortex defined by Freesurfer (previously CEREBELLUMGREYMATTER_VOLUME); See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 68832, ..., 142592
\item \strong{ERODED_SUBCORTICALWM_SUVR} \emph{numeric} Reference region - mean SUVR of eroded subcortical white matter.  Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 1.086, ..., 3.97
\item \strong{ERODED_SUBCORTICALWM_VOLUME} \emph{numeric} Reference region - volume of eroded subcortical white matter.  Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 55673, ..., 424418
\item \strong{BRAINSTEM_SUVR} \emph{numeric} Reference region - mean SUVR of brainstem.  Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0.974, ..., 2.116
\item \strong{BRAINSTEM_VOLUME} \emph{numeric} brain-stem ROI size in mm^3 Range value: 13764, ..., 31506
\item \strong{CC_ANTERIOR_SUVR} \emph{numeric} cc-anterior Range value: 0.564, ..., 2.676
\item \strong{CC_ANTERIOR_VOLUME} \emph{numeric} cc-anterior ROI size in mm^3 Range value: 67, ..., 1688
\item \strong{CC_CENTRAL_SUVR} \emph{numeric} cc-central Range value: 0.361, ..., 2.497
\item \strong{CC_CENTRAL_VOLUME} \emph{numeric} cc-central ROI size in mm^3 Range value: 217, ..., 1297
\item \strong{CC_MID_ANTERIOR_SUVR} \emph{numeric} cc-mid-anterior Range value: 0.391, ..., 2.526
\item \strong{CC_MID_ANTERIOR_VOLUME} \emph{numeric} cc-mid-anterior ROI size in mm^3 Range value: 59, ..., 1170
\item \strong{CC_MID_POSTERIOR_SUVR} \emph{numeric} cc-mid-posterior Range value: 0.432, ..., 2.48
\item \strong{CC_MID_POSTERIOR_VOLUME} \emph{numeric} cc-mid-posterior ROI size in mm^3 Range value: 179, ..., 1476
\item \strong{CC_POSTERIOR_SUVR} \emph{numeric} cc-posterior Range value: 0.825, ..., 3.454
\item \strong{CC_POSTERIOR_VOLUME} \emph{numeric} cc-posterior ROI size in mm^3 Range value: 218, ..., 2196
\item \strong{CSF_SUVR} \emph{numeric} csf Range value: 0.339, ..., 1.29
\item \strong{CSF_VOLUME} \emph{numeric} csf ROI size in mm^3 Range value: 573, ..., 6061
\item \strong{VENTRICLE_3RD_SUVR} \emph{numeric} 3rd-ventricle Range value: 0.356, ..., 1.213
\item \strong{VENTRICLE_3RD_VOLUME} \emph{numeric} 3rd-ventricle ROI size in mm^3 Range value: 474, ..., 5244
\item \strong{VENTRICLE_4TH_SUVR} \emph{numeric} 4th-ventricle Range value: 0.541, ..., 1.299
\item \strong{VENTRICLE_4TH_VOLUME} \emph{numeric} 4th-ventricle ROI size in mm^3 Range value: 561, ..., 5586
\item \strong{VENTRICLE_5TH_SUVR} \emph{numeric} 5th-ventricle Range value: 0.422, ..., 1.414
\item \strong{VENTRICLE_5TH_VOLUME} \emph{numeric} 5th-ventricle ROI size in mm^3 Range value: 1, ..., 11
\item \strong{WM_HYPOINTENSITIES_SUVR} \emph{numeric} wm-hypointensities Range value: 0.617, ..., 3.046
\item \strong{WM_HYPOINTENSITIES_VOLUME} \emph{numeric} wm-hypointensities ROI size in mm^3 Range value: 255, ..., 72336
\item \strong{NON_WM_HYPOINTENSITIES_SUVR} \emph{numeric} non-wm-hypointensities Range value: 0.427, ..., 2.76
\item \strong{NON_WM_HYPOINTENSITIES_VOLUME} \emph{numeric} non-wm-hypointensities ROI size in mm^3 Range value: 1, ..., 233
\item \strong{CTX_BANKSSTS_SUVR} \emph{numeric} ctx-bilateral-bankssts Range value: 0.915, ..., 3.049
\item \strong{CTX_BANKSSTS_VOLUME} \emph{numeric} ctx-bilateral-bankssts ROI size in mm^3 Range value: 2094, ..., 8963
\item \strong{CTX_CAUDALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-bilateral-caudalanteriorcingulate Range value: 0.602, ..., 2.95
\item \strong{CTX_CAUDALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-bilateral-caudalanteriorcingulate ROI size in mm^3 Range value: 963, ..., 8378
\item \strong{CTX_CAUDALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-bilateral-caudalmiddlefrontal Range value: 0.688, ..., 2.915
\item \strong{CTX_CAUDALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-caudalmiddlefrontal ROI size in mm^3 Range value: 5472, ..., 21954
\item \strong{CTX_CUNEUS_SUVR} \emph{numeric} ctx-bilateral-cuneus Range value: 0.783, ..., 2.557
\item \strong{CTX_CUNEUS_VOLUME} \emph{numeric} ctx-bilateral-cuneus ROI size in mm^3 Range value: 3366, ..., 10596
\item \strong{CTX_ENTORHINAL_SUVR} \emph{numeric} ctx-bilateral-entorhinal Range value: 0.505, ..., 1.639
\item \strong{CTX_ENTORHINAL_VOLUME} \emph{numeric} ctx-bilateral-entorhinal ROI size in mm^3 Range value: 746, ..., 6911
\item \strong{CTX_FRONTALPOLE_SUVR} \emph{numeric} ctx-bilateral-frontalpole Range value: 0.421, ..., 2.973
\item \strong{CTX_FRONTALPOLE_VOLUME} \emph{numeric} ctx-bilateral-frontalpole ROI size in mm^3 Range value: 982, ..., 3888
\item \strong{CTX_FUSIFORM_SUVR} \emph{numeric} ctx-bilateral-fusiform Range value: 0.803, ..., 2.24
\item \strong{CTX_FUSIFORM_VOLUME} \emph{numeric} ctx-bilateral-fusiform ROI size in mm^3 Range value: 8002, ..., 26997
\item \strong{CTX_INFERIORPARIETAL_SUVR} \emph{numeric} ctx-bilateral-inferiorparietal Range value: 0.774, ..., 2.818
\item \strong{CTX_INFERIORPARIETAL_VOLUME} \emph{numeric} ctx-bilateral-inferiorparietal ROI size in mm^3 Range value: 11643, ..., 37678
\item \strong{CTX_INFERIORTEMPORAL_SUVR} \emph{numeric} ctx-bilateral-inferiortemporal Range value: 0.804, ..., 2.499
\item \strong{CTX_INFERIORTEMPORAL_VOLUME} \emph{numeric} ctx-bilateral-inferiortemporal ROI size in mm^3 Range value: 8623, ..., 31454
\item \strong{CTX_INSULA_SUVR} \emph{numeric} ctx-bilateral-insula Range value: 0.721, ..., 2.614
\item \strong{CTX_INSULA_VOLUME} \emph{numeric} ctx-bilateral-insula ROI size in mm^3 Range value: 7644, ..., 33551
\item \strong{CTX_ISTHMUSCINGULATE_SUVR} \emph{numeric} ctx-bilateral-isthmuscingulate Range value: 0.806, ..., 3.141
\item \strong{CTX_ISTHMUSCINGULATE_VOLUME} \emph{numeric} ctx-bilateral-isthmuscingulate ROI size in mm^3 Range value: 1985, ..., 16944
\item \strong{CTX_LATERALOCCIPITAL_SUVR} \emph{numeric} ctx-bilateral-lateraloccipital Range value: 0.802, ..., 2.764
\item \strong{CTX_LATERALOCCIPITAL_VOLUME} \emph{numeric} ctx-bilateral-lateraloccipital ROI size in mm^3 Range value: 9370, ..., 35627
\item \strong{CTX_LATERALORBITOFRONTAL_SUVR} \emph{numeric} ctx-bilateral-lateralorbitofrontal Range value: 0.801, ..., 2.682
\item \strong{CTX_LATERALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-lateralorbitofrontal ROI size in mm^3 Range value: 7937, ..., 22861
\item \strong{CTX_LINGUAL_SUVR} \emph{numeric} ctx-bilateral-lingual Range value: 0.77, ..., 2.127
\item \strong{CTX_LINGUAL_VOLUME} \emph{numeric} ctx-bilateral-lingual ROI size in mm^3 Range value: 5946, ..., 20629
\item \strong{CTX_MEDIALORBITOFRONTAL_SUVR} \emph{numeric} ctx-bilateral-medialorbitofrontal Range value: 0.744, ..., 2.828
\item \strong{CTX_MEDIALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-medialorbitofrontal ROI size in mm^3 Range value: 6400, ..., 15313
\item \strong{CTX_MIDDLETEMPORAL_SUVR} \emph{numeric} ctx-bilateral-middletemporal Range value: 0.785, ..., 2.549
\item \strong{CTX_MIDDLETEMPORAL_VOLUME} \emph{numeric} ctx-bilateral-middletemporal ROI size in mm^3 Range value: 9286, ..., 32354
\item \strong{CTX_PARACENTRAL_SUVR} \emph{numeric} ctx-bilateral-paracentral Range value: 0.656, ..., 3.277
\item \strong{CTX_PARACENTRAL_VOLUME} \emph{numeric} ctx-bilateral-paracentral ROI size in mm^3 Range value: 4084, ..., 13660
\item \strong{CTX_PARAHIPPOCAMPAL_SUVR} \emph{numeric} ctx-bilateral-parahippocampal Range value: 0.547, ..., 1.819
\item \strong{CTX_PARAHIPPOCAMPAL_VOLUME} \emph{numeric} ctx-bilateral-parahippocampal ROI size in mm^3 Range value: 1751, ..., 6474
\item \strong{CTX_PARSOPERCULARIS_SUVR} \emph{numeric} ctx-bilateral-parsopercularis Range value: 0.784, ..., 2.951
\item \strong{CTX_PARSOPERCULARIS_VOLUME} \emph{numeric} ctx-bilateral-parsopercularis ROI size in mm^3 Range value: 4303, ..., 12836
\item \strong{CTX_PARSORBITALIS_SUVR} \emph{numeric} ctx-bilateral-parsorbitalis Range value: 0.744, ..., 2.803
\item \strong{CTX_PARSORBITALIS_VOLUME} \emph{numeric} ctx-bilateral-parsorbitalis ROI size in mm^3 Range value: 2438, ..., 7089
\item \strong{CTX_PARSTRIANGULARIS_SUVR} \emph{numeric} ctx-bilateral-parstriangularis Range value: 0.778, ..., 2.972
\item \strong{CTX_PARSTRIANGULARIS_VOLUME} \emph{numeric} ctx-bilateral-parstriangularis ROI size in mm^3 Range value: 4260, ..., 11719
\item \strong{CTX_PERICALCARINE_SUVR} \emph{numeric} ctx-bilateral-pericalcarine Range value: 0.854, ..., 2.808
\item \strong{CTX_PERICALCARINE_VOLUME} \emph{numeric} ctx-bilateral-pericalcarine ROI size in mm^3 Range value: 1847, ..., 13257
\item \strong{CTX_POSTCENTRAL_SUVR} \emph{numeric} ctx-bilateral-postcentral Range value: 0.712, ..., 2.707
\item \strong{CTX_POSTCENTRAL_VOLUME} \emph{numeric} ctx-bilateral-postcentral ROI size in mm^3 Range value: 7517, ..., 28111
\item \strong{CTX_POSTERIORCINGULATE_SUVR} \emph{numeric} ctx-bilateral-posteriorcingulate Range value: 0.601, ..., 3.356
\item \strong{CTX_POSTERIORCINGULATE_VOLUME} \emph{numeric} ctx-bilateral-posteriorcingulate ROI size in mm^3 Range value: 1978, ..., 18216
\item \strong{CTX_PRECENTRAL_SUVR} \emph{numeric} ctx-bilateral-precentral Range value: 0.704, ..., 2.82
\item \strong{CTX_PRECENTRAL_VOLUME} \emph{numeric} ctx-bilateral-precentral ROI size in mm^3 Range value: 13766, ..., 43751
\item \strong{CTX_PRECUNEUS_SUVR} \emph{numeric} ctx-bilateral-precuneus Range value: 0.788, ..., 3.386
\item \strong{CTX_PRECUNEUS_VOLUME} \emph{numeric} ctx-bilateral-precuneus ROI size in mm^3 Range value: 9272, ..., 27581
\item \strong{CTX_ROSTRALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-bilateral-rostralanteriorcingulate Range value: 0.77, ..., 2.933
\item \strong{CTX_ROSTRALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-bilateral-rostralanteriorcingulate ROI size in mm^3 Range value: 1558, ..., 8525
\item \strong{CTX_ROSTRALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-bilateral-rostralmiddlefrontal Range value: 0.786, ..., 3.131
\item \strong{CTX_ROSTRALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-rostralmiddlefrontal ROI size in mm^3 Range value: 12486, ..., 50046
\item \strong{CTX_SUPERIORFRONTAL_SUVR} \emph{numeric} ctx-bilateral-superiorfrontal Range value: 0.61, ..., 2.867
\item \strong{CTX_SUPERIORFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-superiorfrontal ROI size in mm^3 Range value: 23254, ..., 60833
\item \strong{CTX_SUPERIORPARIETAL_SUVR} \emph{numeric} ctx-bilateral-superiorparietal Range value: 0.634, ..., 2.824
\item \strong{CTX_SUPERIORPARIETAL_VOLUME} \emph{numeric} ctx-bilateral-superiorparietal ROI size in mm^3 Range value: 12027, ..., 36908
\item \strong{CTX_SUPERIORTEMPORAL_SUVR} \emph{numeric} ctx-bilateral-superiortemporal Range value: 0.755, ..., 2.602
\item \strong{CTX_SUPERIORTEMPORAL_VOLUME} \emph{numeric} ctx-bilateral-superiortemporal ROI size in mm^3 Range value: 11398, ..., 32667
\item \strong{CTX_SUPRAMARGINAL_SUVR} \emph{numeric} ctx-bilateral-supramarginal Range value: 0.762, ..., 3.014
\item \strong{CTX_SUPRAMARGINAL_VOLUME} \emph{numeric} ctx-bilateral-supramarginal ROI size in mm^3 Range value: 8902, ..., 31391
\item \strong{CTX_TEMPORALPOLE_SUVR} \emph{numeric} ctx-bilateral-temporalpole Range value: 0.535, ..., 1.864
\item \strong{CTX_TEMPORALPOLE_VOLUME} \emph{numeric} ctx-bilateral-temporalpole ROI size in mm^3 Range value: 1332, ..., 9644
\item \strong{CTX_TRANSVERSETEMPORAL_SUVR} \emph{numeric} ctx-bilateral-transversetemporal Range value: 0.661, ..., 2.907
\item \strong{CTX_TRANSVERSETEMPORAL_VOLUME} \emph{numeric} ctx-bilateral-transversetemporal ROI size in mm^3 Range value: 797, ..., 3541
\item \strong{ACCUMBENS_AREA_SUVR} \emph{numeric} bilateral-accumbens-area Range value: 0.555, ..., 2.832
\item \strong{ACCUMBENS_AREA_VOLUME} \emph{numeric} bilateral-accumbens-area ROI size in mm^3 Range value: 325, ..., 1636
\item \strong{AMYGDALA_SUVR} \emph{numeric} bilateral-amygdala Range value: 0.551, ..., 1.688
\item \strong{AMYGDALA_VOLUME} \emph{numeric} bilateral-amygdala ROI size in mm^3 Range value: 934, ..., 5040
\item \strong{CAUDATE_SUVR} \emph{numeric} bilateral-caudate Range value: 0.611, ..., 2.055
\item \strong{CAUDATE_VOLUME} \emph{numeric} bilateral-caudate ROI size in mm^3 Range value: 3588, ..., 13237
\item \strong{CEREBELLUM_WHITE_MATTER_SUVR} \emph{numeric} bilateral-cerebellum-white-matter Range value: 0.736, ..., 2.079
\item \strong{CEREBELLUM_WHITE_MATTER_VOLUME} \emph{numeric} bilateral-cerebellum-white-matter ROI size in mm^3 Range value: 13671, ..., 49420
\item \strong{CEREBRAL_WHITE_MATTER_SUVR} \emph{numeric} bilateral-cerebral-white-matter Range value: 1.13, ..., 3.527
\item \strong{CEREBRAL_WHITE_MATTER_VOLUME} \emph{numeric} bilateral-cerebral-white-matter ROI size in mm^3 Range value: 238555, ..., 718037
\item \strong{CHOROID_PLEXUS_SUVR} \emph{numeric} bilateral-choroid-plexus Range value: 0.378, ..., 1.56
\item \strong{CHOROID_PLEXUS_VOLUME} \emph{numeric} bilateral-choroid-plexus ROI size in mm^3 Range value: 459, ..., 3832
\item \strong{HIPPOCAMPUS_SUVR} \emph{numeric} bilateral-hippocampus Range value: 0.653, ..., 1.538
\item \strong{HIPPOCAMPUS_VOLUME} \emph{numeric} bilateral-hippocampus ROI size in mm^3 Range value: 3255, ..., 11921
\item \strong{OPTIC_CHIASM_SUVR} \emph{numeric} optic-chiasm Range value: 0.336, ..., 1.403
\item \strong{OPTIC_CHIASM_VOLUME} \emph{numeric} optic-chiasm ROI size in mm^3 Range value: 2, ..., 361
\item \strong{INF_LAT_VENT_SUVR} \emph{numeric} bilateral-inf-lat-vent Range value: 0.479, ..., 1.805
\item \strong{INF_LAT_VENT_VOLUME} \emph{numeric} bilateral-inf-lat-vent ROI size in mm^3 Range value: 164, ..., 16102
\item \strong{LATERAL_VENTRICLE_SUVR} \emph{numeric} bilateral-lateral-ventricle Range value: 0.284, ..., 1.441
\item \strong{LATERAL_VENTRICLE_VOLUME} \emph{numeric} bilateral-lateral-ventricle ROI size in mm^3 Range value: 5800, ..., 146015
\item \strong{PALLIDUM_SUVR} \emph{numeric} bilateral-pallidum Range value: 1.007, ..., 2.579
\item \strong{PALLIDUM_VOLUME} \emph{numeric} bilateral-pallidum ROI size in mm^3 Range value: 2306, ..., 7048
\item \strong{PUTAMEN_SUVR} \emph{numeric} bilateral-putamen Range value: 0.913, ..., 2.527
\item \strong{PUTAMEN_VOLUME} \emph{numeric} bilateral-putamen ROI size in mm^3 Range value: 4023, ..., 17943
\item \strong{THALAMUS_PROPER_SUVR} \emph{numeric} bilateral-thalamus-proper Range value: 0.784, ..., 1.828
\item \strong{THALAMUS_PROPER_VOLUME} \emph{numeric} bilateral-thalamus-proper ROI size in mm^3 Range value: 8557, ..., 24225
\item \strong{VENTRALDC_SUVR} \emph{numeric} bilateral-ventraldc Range value: 0.932, ..., 2.249
\item \strong{VENTRALDC_VOLUME} \emph{numeric} bilateral-ventraldc ROI size in mm^3 Range value: 4441, ..., 11382
\item \strong{VESSEL_SUVR} \emph{numeric} bilateral-vessel Range value: 0.537, ..., 2.058
\item \strong{VESSEL_VOLUME} \emph{numeric} bilateral-vessel ROI size in mm^3 Range value: 1, ..., 604
\item \strong{CTX_LH_BANKSSTS_SUVR} \emph{numeric} ctx-lh-bankssts Range value: 0.822, ..., 3.159
\item \strong{CTX_LH_BANKSSTS_VOLUME} \emph{numeric} ctx-lh-bankssts ROI size in mm^3 Range value: 960, ..., 5268
\item \strong{CTX_LH_CAUDALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-lh-caudalanteriorcingulate Range value: 0.642, ..., 3.029
\item \strong{CTX_LH_CAUDALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-lh-caudalanteriorcingulate ROI size in mm^3 Range value: 352, ..., 5165
\item \strong{CTX_LH_CAUDALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-lh-caudalmiddlefrontal Range value: 0.716, ..., 3.059
\item \strong{CTX_LH_CAUDALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-lh-caudalmiddlefrontal ROI size in mm^3 Range value: 2037, ..., 16456
\item \strong{CTX_LH_CUNEUS_SUVR} \emph{numeric} ctx-lh-cuneus Range value: 0.799, ..., 2.504
\item \strong{CTX_LH_CUNEUS_VOLUME} \emph{numeric} ctx-lh-cuneus ROI size in mm^3 Range value: 1411, ..., 6010
\item \strong{CTX_LH_ENTORHINAL_SUVR} \emph{numeric} ctx-lh-entorhinal Range value: 0.511, ..., 1.675
\item \strong{CTX_LH_ENTORHINAL_VOLUME} \emph{numeric} ctx-lh-entorhinal ROI size in mm^3 Range value: 139, ..., 3806
\item \strong{CTX_LH_FRONTALPOLE_SUVR} \emph{numeric} ctx-lh-frontalpole Range value: 0.43, ..., 2.957
\item \strong{CTX_LH_FRONTALPOLE_VOLUME} \emph{numeric} ctx-lh-frontalpole ROI size in mm^3 Range value: 417, ..., 1953
\item \strong{CTX_LH_FUSIFORM_SUVR} \emph{numeric} ctx-lh-fusiform Range value: 0.82, ..., 2.311
\item \strong{CTX_LH_FUSIFORM_VOLUME} \emph{numeric} ctx-lh-fusiform ROI size in mm^3 Range value: 3185, ..., 17168
\item \strong{CTX_LH_INFERIORPARIETAL_SUVR} \emph{numeric} ctx-lh-inferiorparietal Range value: 0.751, ..., 2.97
\item \strong{CTX_LH_INFERIORPARIETAL_VOLUME} \emph{numeric} ctx-lh-inferiorparietal ROI size in mm^3 Range value: 4994, ..., 17983
\item \strong{CTX_LH_INFERIORTEMPORAL_SUVR} \emph{numeric} ctx-lh-inferiortemporal Range value: 0.775, ..., 2.614
\item \strong{CTX_LH_INFERIORTEMPORAL_VOLUME} \emph{numeric} ctx-lh-inferiortemporal ROI size in mm^3 Range value: 4040, ..., 16913
\item \strong{CTX_LH_INSULA_SUVR} \emph{numeric} ctx-lh-insula Range value: 0.696, ..., 2.742
\item \strong{CTX_LH_INSULA_VOLUME} \emph{numeric} ctx-lh-insula ROI size in mm^3 Range value: 3855, ..., 27008
\item \strong{CTX_LH_ISTHMUSCINGULATE_SUVR} \emph{numeric} ctx-lh-isthmuscingulate Range value: 0.83, ..., 3.278
\item \strong{CTX_LH_ISTHMUSCINGULATE_VOLUME} \emph{numeric} ctx-lh-isthmuscingulate ROI size in mm^3 Range value: 615, ..., 14857
\item \strong{CTX_LH_LATERALOCCIPITAL_SUVR} \emph{numeric} ctx-lh-lateraloccipital Range value: 0.737, ..., 2.639
\item \strong{CTX_LH_LATERALOCCIPITAL_VOLUME} \emph{numeric} ctx-lh-lateraloccipital ROI size in mm^3 Range value: 4386, ..., 20415
\item \strong{CTX_LH_LATERALORBITOFRONTAL_SUVR} \emph{numeric} ctx-lh-lateralorbitofrontal Range value: 0.813, ..., 2.776
\item \strong{CTX_LH_LATERALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-lh-lateralorbitofrontal ROI size in mm^3 Range value: 3652, ..., 15394
\item \strong{CTX_LH_LINGUAL_SUVR} \emph{numeric} ctx-lh-lingual Range value: 0.727, ..., 2.184
\item \strong{CTX_LH_LINGUAL_VOLUME} \emph{numeric} ctx-lh-lingual ROI size in mm^3 Range value: 2540, ..., 10566
\item \strong{CTX_LH_MEDIALORBITOFRONTAL_SUVR} \emph{numeric} ctx-lh-medialorbitofrontal Range value: 0.739, ..., 2.823
\item \strong{CTX_LH_MEDIALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-lh-medialorbitofrontal ROI size in mm^3 Range value: 2937, ..., 9820
\item \strong{CTX_LH_MIDDLETEMPORAL_SUVR} \emph{numeric} ctx-lh-middletemporal Range value: 0.747, ..., 2.62
\item \strong{CTX_LH_MIDDLETEMPORAL_VOLUME} \emph{numeric} ctx-lh-middletemporal ROI size in mm^3 Range value: 3320, ..., 15289
\item \strong{CTX_LH_PARACENTRAL_SUVR} \emph{numeric} ctx-lh-paracentral Range value: 0.774, ..., 3.554
\item \strong{CTX_LH_PARACENTRAL_VOLUME} \emph{numeric} ctx-lh-paracentral ROI size in mm^3 Range value: 1709, ..., 9133
\item \strong{CTX_LH_PARAHIPPOCAMPAL_SUVR} \emph{numeric} ctx-lh-parahippocampal Range value: 0.585, ..., 2.127
\item \strong{CTX_LH_PARAHIPPOCAMPAL_VOLUME} \emph{numeric} ctx-lh-parahippocampal ROI size in mm^3 Range value: 885, ..., 4346
\item \strong{CTX_LH_PARSOPERCULARIS_SUVR} \emph{numeric} ctx-lh-parsopercularis Range value: 0.792, ..., 3.107
\item \strong{CTX_LH_PARSOPERCULARIS_VOLUME} \emph{numeric} ctx-lh-parsopercularis ROI size in mm^3 Range value: 2298, ..., 8062
\item \strong{CTX_LH_PARSORBITALIS_SUVR} \emph{numeric} ctx-lh-parsorbitalis Range value: 0.7, ..., 2.683
\item \strong{CTX_LH_PARSORBITALIS_VOLUME} \emph{numeric} ctx-lh-parsorbitalis ROI size in mm^3 Range value: 438, ..., 3562
\item \strong{CTX_LH_PARSTRIANGULARIS_SUVR} \emph{numeric} ctx-lh-parstriangularis Range value: 0.76, ..., 3.055
\item \strong{CTX_LH_PARSTRIANGULARIS_VOLUME} \emph{numeric} ctx-lh-parstriangularis ROI size in mm^3 Range value: 1547, ..., 6687
\item \strong{CTX_LH_PERICALCARINE_SUVR} \emph{numeric} ctx-lh-pericalcarine Range value: 0.796, ..., 2.652
\item \strong{CTX_LH_PERICALCARINE_VOLUME} \emph{numeric} ctx-lh-pericalcarine ROI size in mm^3 Range value: 755, ..., 11046
\item \strong{CTX_LH_POSTCENTRAL_SUVR} \emph{numeric} ctx-lh-postcentral Range value: 0.627, ..., 2.88
\item \strong{CTX_LH_POSTCENTRAL_VOLUME} \emph{numeric} ctx-lh-postcentral ROI size in mm^3 Range value: 4591, ..., 18003
\item \strong{CTX_LH_POSTERIORCINGULATE_SUVR} \emph{numeric} ctx-lh-posteriorcingulate Range value: 0.724, ..., 3.484
\item \strong{CTX_LH_POSTERIORCINGULATE_VOLUME} \emph{numeric} ctx-lh-posteriorcingulate ROI size in mm^3 Range value: 990, ..., 11595
\item \strong{CTX_LH_PRECENTRAL_SUVR} \emph{numeric} ctx-lh-precentral Range value: 0.671, ..., 2.977
\item \strong{CTX_LH_PRECENTRAL_VOLUME} \emph{numeric} ctx-lh-precentral ROI size in mm^3 Range value: 6606, ..., 31664
\item \strong{CTX_LH_PRECUNEUS_SUVR} \emph{numeric} ctx-lh-precuneus Range value: 0.805, ..., 3.599
\item \strong{CTX_LH_PRECUNEUS_VOLUME} \emph{numeric} ctx-lh-precuneus ROI size in mm^3 Range value: 4103, ..., 19194
\item \strong{CTX_LH_ROSTRALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-lh-rostralanteriorcingulate Range value: 0.771, ..., 3.013
\item \strong{CTX_LH_ROSTRALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-lh-rostralanteriorcingulate ROI size in mm^3 Range value: 571, ..., 6345
\item \strong{CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-lh-rostralmiddlefrontal Range value: 0.784, ..., 3.144
\item \strong{CTX_LH_ROSTRALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-lh-rostralmiddlefrontal ROI size in mm^3 Range value: 5531, ..., 24802
\item \strong{CTX_LH_SUPERIORFRONTAL_SUVR} \emph{numeric} ctx-lh-superiorfrontal Range value: 0.626, ..., 3.048
\item \strong{CTX_LH_SUPERIORFRONTAL_VOLUME} \emph{numeric} ctx-lh-superiorfrontal ROI size in mm^3 Range value: 11052, ..., 40962
\item \strong{CTX_LH_SUPERIORPARIETAL_SUVR} \emph{numeric} ctx-lh-superiorparietal Range value: 0.645, ..., 3.068
\item \strong{CTX_LH_SUPERIORPARIETAL_VOLUME} \emph{numeric} ctx-lh-superiorparietal ROI size in mm^3 Range value: 5411, ..., 24136
\item \strong{CTX_LH_SUPERIORTEMPORAL_SUVR} \emph{numeric} ctx-lh-superiortemporal Range value: 0.747, ..., 2.744
\item \strong{CTX_LH_SUPERIORTEMPORAL_VOLUME} \emph{numeric} ctx-lh-superiortemporal ROI size in mm^3 Range value: 4632, ..., 21023
\item \strong{CTX_LH_SUPRAMARGINAL_SUVR} \emph{numeric} ctx-lh-supramarginal Range value: 0.77, ..., 3.212
\item \strong{CTX_LH_SUPRAMARGINAL_VOLUME} \emph{numeric} ctx-lh-supramarginal ROI size in mm^3 Range value: 5209, ..., 21200
\item \strong{CTX_LH_TEMPORALPOLE_SUVR} \emph{numeric} ctx-lh-temporalpole Range value: 0.527, ..., 1.946
\item \strong{CTX_LH_TEMPORALPOLE_VOLUME} \emph{numeric} ctx-lh-temporalpole ROI size in mm^3 Range value: 48, ..., 5851
\item \strong{CTX_LH_TRANSVERSETEMPORAL_SUVR} \emph{numeric} ctx-lh-transversetemporal Range value: 0.664, ..., 3.141
\item \strong{CTX_LH_TRANSVERSETEMPORAL_VOLUME} \emph{numeric} ctx-lh-transversetemporal ROI size in mm^3 Range value: 441, ..., 2282
\item \strong{CTX_RH_BANKSSTS_SUVR} \emph{numeric} ctx-rh-bankssts Range value: 0.87, ..., 2.951
\item \strong{CTX_RH_BANKSSTS_VOLUME} \emph{numeric} ctx-rh-bankssts ROI size in mm^3 Range value: 873, ..., 4240
\item \strong{CTX_RH_CAUDALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-rh-caudalanteriorcingulate Range value: 0.515, ..., 2.88
\item \strong{CTX_RH_CAUDALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-rh-caudalanteriorcingulate ROI size in mm^3 Range value: 454, ..., 4843
\item \strong{CTX_RH_CAUDALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-rh-caudalmiddlefrontal Range value: 0.657, ..., 2.745
\item \strong{CTX_RH_CAUDALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-rh-caudalmiddlefrontal ROI size in mm^3 Range value: 2492, ..., 9630
\item \strong{CTX_RH_CUNEUS_SUVR} \emph{numeric} ctx-rh-cuneus Range value: 0.744, ..., 2.618
\item \strong{CTX_RH_CUNEUS_VOLUME} \emph{numeric} ctx-rh-cuneus ROI size in mm^3 Range value: 982, ..., 6284
\item \strong{CTX_RH_ENTORHINAL_SUVR} \emph{numeric} ctx-rh-entorhinal Range value: 0.495, ..., 1.598
\item \strong{CTX_RH_ENTORHINAL_VOLUME} \emph{numeric} ctx-rh-entorhinal ROI size in mm^3 Range value: 455, ..., 3722
\item \strong{CTX_RH_FRONTALPOLE_SUVR} \emph{numeric} ctx-rh-frontalpole Range value: 0.408, ..., 2.986
\item \strong{CTX_RH_FRONTALPOLE_VOLUME} \emph{numeric} ctx-rh-frontalpole ROI size in mm^3 Range value: 381, ..., 2370
\item \strong{CTX_RH_FUSIFORM_SUVR} \emph{numeric} ctx-rh-fusiform Range value: 0.777, ..., 2.242
\item \strong{CTX_RH_FUSIFORM_VOLUME} \emph{numeric} ctx-rh-fusiform ROI size in mm^3 Range value: 4093, ..., 14227
\item \strong{CTX_RH_INFERIORPARIETAL_SUVR} \emph{numeric} ctx-rh-inferiorparietal Range value: 0.759, ..., 2.739
\item \strong{CTX_RH_INFERIORPARIETAL_VOLUME} \emph{numeric} ctx-rh-inferiorparietal ROI size in mm^3 Range value: 6177, ..., 21817
\item \strong{CTX_RH_INFERIORTEMPORAL_SUVR} \emph{numeric} ctx-rh-inferiortemporal Range value: 0.782, ..., 2.501
\item \strong{CTX_RH_INFERIORTEMPORAL_VOLUME} \emph{numeric} ctx-rh-inferiortemporal ROI size in mm^3 Range value: 3238, ..., 16344
\item \strong{CTX_RH_INSULA_SUVR} \emph{numeric} ctx-rh-insula Range value: 0.748, ..., 2.48
\item \strong{CTX_RH_INSULA_VOLUME} \emph{numeric} ctx-rh-insula ROI size in mm^3 Range value: 3554, ..., 9794
\item \strong{CTX_RH_ISTHMUSCINGULATE_SUVR} \emph{numeric} ctx-rh-isthmuscingulate Range value: 0.777, ..., 3.009
\item \strong{CTX_RH_ISTHMUSCINGULATE_VOLUME} \emph{numeric} ctx-rh-isthmuscingulate ROI size in mm^3 Range value: 703, ..., 4091
\item \strong{CTX_RH_LATERALOCCIPITAL_SUVR} \emph{numeric} ctx-rh-lateraloccipital Range value: 0.748, ..., 2.873
\item \strong{CTX_RH_LATERALOCCIPITAL_VOLUME} \emph{numeric} ctx-rh-lateraloccipital ROI size in mm^3 Range value: 4984, ..., 19271
\item \strong{CTX_RH_LATERALORBITOFRONTAL_SUVR} \emph{numeric} ctx-rh-lateralorbitofrontal Range value: 0.79, ..., 2.706
\item \strong{CTX_RH_LATERALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-rh-lateralorbitofrontal ROI size in mm^3 Range value: 4176, ..., 9919
\item \strong{CTX_RH_LINGUAL_SUVR} \emph{numeric} ctx-rh-lingual Range value: 0.773, ..., 2.191
\item \strong{CTX_RH_LINGUAL_VOLUME} \emph{numeric} ctx-rh-lingual ROI size in mm^3 Range value: 3263, ..., 10478
\item \strong{CTX_RH_MEDIALORBITOFRONTAL_SUVR} \emph{numeric} ctx-rh-medialorbitofrontal Range value: 0.748, ..., 2.833
\item \strong{CTX_RH_MEDIALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-rh-medialorbitofrontal ROI size in mm^3 Range value: 3216, ..., 9115
\item \strong{CTX_RH_MIDDLETEMPORAL_SUVR} \emph{numeric} ctx-rh-middletemporal Range value: 0.781, ..., 2.537
\item \strong{CTX_RH_MIDDLETEMPORAL_VOLUME} \emph{numeric} ctx-rh-middletemporal ROI size in mm^3 Range value: 4756, ..., 17219
\item \strong{CTX_RH_PARACENTRAL_SUVR} \emph{numeric} ctx-rh-paracentral Range value: 0.553, ..., 3.046
\item \strong{CTX_RH_PARACENTRAL_VOLUME} \emph{numeric} ctx-rh-paracentral ROI size in mm^3 Range value: 1981, ..., 5470
\item \strong{CTX_RH_PARAHIPPOCAMPAL_SUVR} \emph{numeric} ctx-rh-parahippocampal Range value: 0.506, ..., 1.839
\item \strong{CTX_RH_PARAHIPPOCAMPAL_VOLUME} \emph{numeric} ctx-rh-parahippocampal ROI size in mm^3 Range value: 840, ..., 3420
\item \strong{CTX_RH_PARSOPERCULARIS_SUVR} \emph{numeric} ctx-rh-parsopercularis Range value: 0.769, ..., 2.801
\item \strong{CTX_RH_PARSOPERCULARIS_VOLUME} \emph{numeric} ctx-rh-parsopercularis ROI size in mm^3 Range value: 1716, ..., 7542
\item \strong{CTX_RH_PARSORBITALIS_SUVR} \emph{numeric} ctx-rh-parsorbitalis Range value: 0.749, ..., 2.913
\item \strong{CTX_RH_PARSORBITALIS_VOLUME} \emph{numeric} ctx-rh-parsorbitalis ROI size in mm^3 Range value: 810, ..., 3933
\item \strong{CTX_RH_PARSTRIANGULARIS_SUVR} \emph{numeric} ctx-rh-parstriangularis Range value: 0.774, ..., 2.905
\item \strong{CTX_RH_PARSTRIANGULARIS_VOLUME} \emph{numeric} ctx-rh-parstriangularis ROI size in mm^3 Range value: 2134, ..., 7512
\item \strong{CTX_RH_PERICALCARINE_SUVR} \emph{numeric} ctx-rh-pericalcarine Range value: 0.77, ..., 2.967
\item \strong{CTX_RH_PERICALCARINE_VOLUME} \emph{numeric} ctx-rh-pericalcarine ROI size in mm^3 Range value: 831, ..., 4617
\item \strong{CTX_RH_POSTCENTRAL_SUVR} \emph{numeric} ctx-rh-postcentral Range value: 0.73, ..., 2.52
\item \strong{CTX_RH_POSTCENTRAL_VOLUME} \emph{numeric} ctx-rh-postcentral ROI size in mm^3 Range value: 2926, ..., 13033
\item \strong{CTX_RH_POSTERIORCINGULATE_SUVR} \emph{numeric} ctx-rh-posteriorcingulate Range value: 0.456, ..., 3.228
\item \strong{CTX_RH_POSTERIORCINGULATE_VOLUME} \emph{numeric} ctx-rh-posteriorcingulate ROI size in mm^3 Range value: 552, ..., 6621
\item \strong{CTX_RH_PRECENTRAL_SUVR} \emph{numeric} ctx-rh-precentral Range value: 0.739, ..., 2.647
\item \strong{CTX_RH_PRECENTRAL_VOLUME} \emph{numeric} ctx-rh-precentral ROI size in mm^3 Range value: 6158, ..., 18058
\item \strong{CTX_RH_PRECUNEUS_SUVR} \emph{numeric} ctx-rh-precuneus Range value: 0.766, ..., 3.171
\item \strong{CTX_RH_PRECUNEUS_VOLUME} \emph{numeric} ctx-rh-precuneus ROI size in mm^3 Range value: 4969, ..., 14854
\item \strong{CTX_RH_ROSTRALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-rh-rostralanteriorcingulate Range value: 0.752, ..., 2.768
\item \strong{CTX_RH_ROSTRALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-rh-rostralanteriorcingulate ROI size in mm^3 Range value: 559, ..., 3508
\item \strong{CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-rh-rostralmiddlefrontal Range value: 0.77, ..., 3.239
\item \strong{CTX_RH_ROSTRALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-rh-rostralmiddlefrontal ROI size in mm^3 Range value: 6285, ..., 26835
\item \strong{CTX_RH_SUPERIORFRONTAL_SUVR} \emph{numeric} ctx-rh-superiorfrontal Range value: 0.596, ..., 2.677
\item \strong{CTX_RH_SUPERIORFRONTAL_VOLUME} \emph{numeric} ctx-rh-superiorfrontal ROI size in mm^3 Range value: 10729, ..., 32025
\item \strong{CTX_RH_SUPERIORPARIETAL_SUVR} \emph{numeric} ctx-rh-superiorparietal Range value: 0.623, ..., 2.574
\item \strong{CTX_RH_SUPERIORPARIETAL_VOLUME} \emph{numeric} ctx-rh-superiorparietal ROI size in mm^3 Range value: 6110, ..., 18993
\item \strong{CTX_RH_SUPERIORTEMPORAL_SUVR} \emph{numeric} ctx-rh-superiortemporal Range value: 0.748, ..., 2.447
\item \strong{CTX_RH_SUPERIORTEMPORAL_VOLUME} \emph{numeric} ctx-rh-superiortemporal ROI size in mm^3 Range value: 4975, ..., 15383
\item \strong{CTX_RH_SUPRAMARGINAL_SUVR} \emph{numeric} ctx-rh-supramarginal Range value: 0.751, ..., 2.787
\item \strong{CTX_RH_SUPRAMARGINAL_VOLUME} \emph{numeric} ctx-rh-supramarginal ROI size in mm^3 Range value: 3693, ..., 14663
\item \strong{CTX_RH_TEMPORALPOLE_SUVR} \emph{numeric} ctx-rh-temporalpole Range value: 0.503, ..., 1.825
\item \strong{CTX_RH_TEMPORALPOLE_VOLUME} \emph{numeric} ctx-rh-temporalpole ROI size in mm^3 Range value: 287, ..., 4666
\item \strong{CTX_RH_TRANSVERSETEMPORAL_SUVR} \emph{numeric} ctx-rh-transversetemporal Range value: 0.657, ..., 2.634
\item \strong{CTX_RH_TRANSVERSETEMPORAL_VOLUME} \emph{numeric} ctx-rh-transversetemporal ROI size in mm^3 Range value: 269, ..., 1566
\item \strong{LEFT_ACCUMBENS_AREA_SUVR} \emph{numeric} left-accumbens-area Range value: 0.559, ..., 2.971
\item \strong{LEFT_ACCUMBENS_AREA_VOLUME} \emph{numeric} left-accumbens-area ROI size in mm^3 Range value: 92, ..., 821
\item \strong{LEFT_AMYGDALA_SUVR} \emph{numeric} left-amygdala Range value: 0.465, ..., 1.736
\item \strong{LEFT_AMYGDALA_VOLUME} \emph{numeric} left-amygdala ROI size in mm^3 Range value: 339, ..., 2496
\item \strong{LEFT_CAUDATE_SUVR} \emph{numeric} left-caudate Range value: 0.615, ..., 2.23
\item \strong{LEFT_CAUDATE_VOLUME} \emph{numeric} left-caudate ROI size in mm^3 Range value: 1660, ..., 6903
\item \strong{LEFT_CEREBELLUM_CORTEX_SUVR} \emph{numeric} left-cerebellum-cortex Range value: 0.751, ..., 1.072
\item \strong{LEFT_CEREBELLUM_CORTEX_VOLUME} \emph{numeric} left-cerebellum-cortex ROI size in mm^3 Range value: 34655, ..., 69630
\item \strong{LEFT_CEREBELLUM_WHITE_MATTER_SUVR} \emph{numeric} left-cerebellum-white-matter Range value: 0.722, ..., 2.055
\item \strong{LEFT_CEREBELLUM_WHITE_MATTER_VOLUME} \emph{numeric} left-cerebellum-white-matter ROI size in mm^3 Range value: 6892, ..., 26253
\item \strong{LEFT_CEREBRAL_WHITE_MATTER_SUVR} \emph{numeric} left-cerebral-white-matter Range value: 1.138, ..., 3.634
\item \strong{LEFT_CEREBRAL_WHITE_MATTER_VOLUME} \emph{numeric} left-cerebral-white-matter ROI size in mm^3 Range value: 20485, ..., 354965
\item \strong{LEFT_CHOROID_PLEXUS_SUVR} \emph{numeric} left-choroid-plexus Range value: 0.328, ..., 1.578
\item \strong{LEFT_CHOROID_PLEXUS_VOLUME} \emph{numeric} left-choroid-plexus ROI size in mm^3 Range value: 130, ..., 2141
\item \strong{LEFT_HIPPOCAMPUS_SUVR} \emph{numeric} left-hippocampus Range value: 0.601, ..., 1.545
\item \strong{LEFT_HIPPOCAMPUS_VOLUME} \emph{numeric} left-hippocampus ROI size in mm^3 Range value: 1314, ..., 6080
\item \strong{LEFT_INF_LAT_VENT_SUVR} \emph{numeric} left-inf-lat-vent Range value: 0.525, ..., 1.871
\item \strong{LEFT_INF_LAT_VENT_VOLUME} \emph{numeric} left-inf-lat-vent ROI size in mm^3 Range value: 53, ..., 11328
\item \strong{LEFT_LATERAL_VENTRICLE_SUVR} \emph{numeric} left-lateral-ventricle Range value: 0.273, ..., 1.457
\item \strong{LEFT_LATERAL_VENTRICLE_VOLUME} \emph{numeric} left-lateral-ventricle ROI size in mm^3 Range value: 3016, ..., 82443
\item \strong{LEFT_PALLIDUM_SUVR} \emph{numeric} left-pallidum Range value: 0.99, ..., 2.626
\item \strong{LEFT_PALLIDUM_VOLUME} \emph{numeric} left-pallidum ROI size in mm^3 Range value: 1066, ..., 3513
\item \strong{LEFT_PUTAMEN_SUVR} \emph{numeric} left-putamen Range value: 0.891, ..., 2.533
\item \strong{LEFT_PUTAMEN_VOLUME} \emph{numeric} left-putamen ROI size in mm^3 Range value: 1788, ..., 9271
\item \strong{LEFT_THALAMUS_PROPER_SUVR} \emph{numeric} left-thalamus-proper Range value: 0.686, ..., 1.822
\item \strong{LEFT_THALAMUS_PROPER_VOLUME} \emph{numeric} left-thalamus-proper ROI size in mm^3 Range value: 4217, ..., 11791
\item \strong{LEFT_VENTRALDC_SUVR} \emph{numeric} left-ventraldc Range value: 0.944, ..., 2.346
\item \strong{LEFT_VENTRALDC_VOLUME} \emph{numeric} left-ventraldc ROI size in mm^3 Range value: 2259, ..., 5708
\item \strong{LEFT_VESSEL_SUVR} \emph{numeric} left-vessel Range value: 0.537, ..., 2.033
\item \strong{LEFT_VESSEL_VOLUME} \emph{numeric} left-vessel ROI size in mm^3 Range value: 1, ..., 537
\item \strong{RIGHT_ACCUMBENS_AREA_SUVR} \emph{numeric} right-accumbens-area Range value: 0.553, ..., 2.721
\item \strong{RIGHT_ACCUMBENS_AREA_VOLUME} \emph{numeric} right-accumbens-area ROI size in mm^3 Range value: 114, ..., 843
\item \strong{RIGHT_AMYGDALA_SUVR} \emph{numeric} right-amygdala Range value: 0.627, ..., 1.652
\item \strong{RIGHT_AMYGDALA_VOLUME} \emph{numeric} right-amygdala ROI size in mm^3 Range value: 483, ..., 2544
\item \strong{RIGHT_CAUDATE_SUVR} \emph{numeric} right-caudate Range value: 0.592, ..., 1.997
\item \strong{RIGHT_CAUDATE_VOLUME} \emph{numeric} right-caudate ROI size in mm^3 Range value: 1717, ..., 6880
\item \strong{RIGHT_CEREBELLUM_CORTEX_SUVR} \emph{numeric} right-cerebellum-cortex Range value: 0.756, ..., 1.048
\item \strong{RIGHT_CEREBELLUM_CORTEX_VOLUME} \emph{numeric} right-cerebellum-cortex ROI size in mm^3 Range value: 33950, ..., 73171
\item \strong{RIGHT_CEREBELLUM_WHITE_MATTER_SUVR} \emph{numeric} right-cerebellum-white-matter Range value: 0.75, ..., 2.105
\item \strong{RIGHT_CEREBELLUM_WHITE_MATTER_VOLUME} \emph{numeric} right-cerebellum-white-matter ROI size in mm^3 Range value: 6599, ..., 25269
\item \strong{RIGHT_CEREBRAL_WHITE_MATTER_SUVR} \emph{numeric} right-cerebral-white-matter Range value: 1.103, ..., 3.42
\item \strong{RIGHT_CEREBRAL_WHITE_MATTER_VOLUME} \emph{numeric} right-cerebral-white-matter ROI size in mm^3 Range value: 127204, ..., 363072
\item \strong{RIGHT_CHOROID_PLEXUS_SUVR} \emph{numeric} right-choroid-plexus Range value: 0.418, ..., 1.548
\item \strong{RIGHT_CHOROID_PLEXUS_VOLUME} \emph{numeric} right-choroid-plexus ROI size in mm^3 Range value: 126, ..., 2028
\item \strong{RIGHT_HIPPOCAMPUS_SUVR} \emph{numeric} right-hippocampus Range value: 0.614, ..., 1.532
\item \strong{RIGHT_HIPPOCAMPUS_VOLUME} \emph{numeric} right-hippocampus ROI size in mm^3 Range value: 852, ..., 5928
\item \strong{RIGHT_INF_LAT_VENT_SUVR} \emph{numeric} right-inf-lat-vent Range value: 0.384, ..., 1.748
\item \strong{RIGHT_INF_LAT_VENT_VOLUME} \emph{numeric} right-inf-lat-vent ROI size in mm^3 Range value: 58, ..., 7510
\item \strong{RIGHT_LATERAL_VENTRICLE_SUVR} \emph{numeric} right-lateral-ventricle Range value: 0.263, ..., 1.456
\item \strong{RIGHT_LATERAL_VENTRICLE_VOLUME} \emph{numeric} right-lateral-ventricle ROI size in mm^3 Range value: 2442, ..., 71992
\item \strong{RIGHT_PALLIDUM_SUVR} \emph{numeric} right-pallidum Range value: 0.991, ..., 2.534
\item \strong{RIGHT_PALLIDUM_VOLUME} \emph{numeric} right-pallidum ROI size in mm^3 Range value: 688, ..., 3535
\item \strong{RIGHT_PUTAMEN_SUVR} \emph{numeric} right-putamen Range value: 0.894, ..., 2.548
\item \strong{RIGHT_PUTAMEN_VOLUME} \emph{numeric} right-putamen ROI size in mm^3 Range value: 2106, ..., 8672
\item \strong{RIGHT_THALAMUS_PROPER_SUVR} \emph{numeric} right-thalamus-proper Range value: 0.791, ..., 1.871
\item \strong{RIGHT_THALAMUS_PROPER_VOLUME} \emph{numeric} right-thalamus-proper ROI size in mm^3 Range value: 3522, ..., 12434
\item \strong{RIGHT_VENTRALDC_SUVR} \emph{numeric} right-ventraldc Range value: 0.921, ..., 2.142
\item \strong{RIGHT_VENTRALDC_VOLUME} \emph{numeric} right-ventraldc ROI size in mm^3 Range value: 1892, ..., 5674
\item \strong{RIGHT_VESSEL_SUVR} \emph{numeric} right-vessel Range value: 0.712, ..., 2.149
\item \strong{RIGHT_VESSEL_VOLUME} \emph{numeric} right-vessel ROI size in mm^3 Range value: 1, ..., 300
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
