% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCBERKELEY_TAUPVC_6MM}
\alias{UCBERKELEY_TAUPVC_6MM}
\title{UC Berkeley - Tau PET PVC 6mm Res analysis}
\format{
A data frame with 2106 observations and 336 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCBERKELEY_TAUPVC_6MM)
}
\description{
UC Berkeley - Tau PET PVC 6mm Res analysis data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{LONIUID} \emph{numeric} LONI Image ID Range value: 1592895, ..., 11383999
\item \strong{PTID} \emph{character} ADNI Participant ID consisting of 3-digit "SITEID" and 4-digit "RID"  
\item \strong{RID} \emph{numeric} Participant roster ID (unique participant identifier that should be used to merge data) Range value: 21, ..., 10770
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{SCANDATE} \emph{ } SCAN acquisition date Date: YYY-MM-DD
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{PROCESSDATE} \emph{ } Date quantification pipeline was run Date: YYY-MM-DD
\item \strong{TRACER} \emph{character} PET Radiotracer , with options: FTP, MK6240, PI2620
\item \strong{TRACER_SUVR_WARNING} \emph{character} Reminder of the incomparability of SUVRs from different tracers. See UCBERKELEY Tau Processing Methods PDF. , with options: DO NOT COMPARE SUVRs ACROSS TRACERS
\item \strong{META_TEMPORAL_SUVR} \emph{numeric} Weighted mean SUVR of Jack et al temporal meta-ROI. All SUVRs in this file are normalized to the inferior cerebellar grey matter unless otherwise noted. See UCBERKELEY Tau PET Processing Methods PDF on LONI for details. Range value: 0.625, ..., 7.089
\item \strong{META_TEMPORAL_VOLUME} \emph{numeric} Jack et al temporal meta-ROI size in mm^3 Range value: 36639, ..., 97274
\item \strong{CTX_ENTORHINAL_SUVR} \emph{numeric} ctx-bilateral-entorhinal Range value: 0.602, ..., 5.771
\item \strong{CTX_ENTORHINAL_VOLUME} \emph{numeric} ctx-bilateral-entorhinal ROI size in mm^3 Range value: 746, ..., 5946
\item \strong{INFERIORCEREBELLUM_SUVR} \emph{numeric} Reference region - mean SUVR of inferior cerebellar grey matter Range value: 1, ..., 1
\item \strong{INFERIORCEREBELLUM_VOLUME} \emph{numeric} Inferior Cerebellar Grey Matter ROI size in mm^3 Range value: 40753, ..., 87660
\item \strong{CEREBRAL_WHITE_MATTER_SUVR} \emph{numeric} bilateral-cerebral-white-matter Range value: 0.46, ..., 1.859
\item \strong{CEREBRAL_WHITE_MATTER_VOLUME} \emph{numeric} bilateral-cerebral-white-matter ROI size in mm^3 Range value: 224785, ..., 718037
\item \strong{BRAINSTEM_SUVR} \emph{numeric} brain-stem Range value: 0.385, ..., 1.463
\item \strong{BRAINSTEM_VOLUME} \emph{numeric} brain-stem ROI size in mm^3 Range value: 13764, ..., 31388
\item \strong{CC_ANTERIOR_SUVR} \emph{numeric} cc-anterior Range value: 0.46, ..., 1.859
\item \strong{CC_ANTERIOR_VOLUME} \emph{numeric} cc-anterior ROI size in mm^3 Range value: 67, ..., 1763
\item \strong{CC_CENTRAL_SUVR} \emph{numeric} cc-central Range value: 0.46, ..., 1.859
\item \strong{CC_CENTRAL_VOLUME} \emph{numeric} cc-central ROI size in mm^3 Range value: 233, ..., 1380
\item \strong{CC_MID_ANTERIOR_SUVR} \emph{numeric} cc-mid-anterior Range value: 0.46, ..., 1.859
\item \strong{CC_MID_ANTERIOR_VOLUME} \emph{numeric} cc-mid-anterior ROI size in mm^3 Range value: 59, ..., 1241
\item \strong{CC_MID_POSTERIOR_SUVR} \emph{numeric} cc-mid-posterior Range value: 0.46, ..., 1.859
\item \strong{CC_MID_POSTERIOR_VOLUME} \emph{numeric} cc-mid-posterior ROI size in mm^3 Range value: 189, ..., 1659
\item \strong{CC_POSTERIOR_SUVR} \emph{numeric} cc-posterior Range value: 0.46, ..., 1.859
\item \strong{CC_POSTERIOR_VOLUME} \emph{numeric} cc-posterior ROI size in mm^3 Range value: 259, ..., 2196
\item \strong{CSF_SUVR} \emph{numeric} csf Range value: 0.007, ..., 0.844
\item \strong{CSF_VOLUME} \emph{numeric} csf ROI size in mm^3 Range value: 628, ..., 3763
\item \strong{VENTRICLE_3RD_SUVR} \emph{numeric} 3rd-ventricle Range value: 0.002, ..., 0.844
\item \strong{VENTRICLE_3RD_VOLUME} \emph{numeric} 3rd-ventricle ROI size in mm^3 Range value: 484, ..., 5244
\item \strong{VENTRICLE_4TH_SUVR} \emph{numeric} 4th-ventricle Range value: 0.001, ..., 0.855
\item \strong{VENTRICLE_4TH_VOLUME} \emph{numeric} 4th-ventricle ROI size in mm^3 Range value: 739, ..., 5081
\item \strong{VENTRICLE_5TH_SUVR} \emph{numeric} 5th-ventricle Range value: 0.185, ..., 0.655
\item \strong{VENTRICLE_5TH_VOLUME} \emph{numeric} 5th-ventricle ROI size in mm^3 Range value: 1, ..., 11
\item \strong{WM_HYPOINTENSITIES_SUVR} \emph{numeric} wm-hypointensities Range value: 0.001, ..., 3.215
\item \strong{WM_HYPOINTENSITIES_VOLUME} \emph{numeric} wm-hypointensities ROI size in mm^3 Range value: 268, ..., 72336
\item \strong{NON_WM_HYPOINTENSITIES_SUVR} \emph{numeric} non-wm-hypointensities Range value: 0.001, ..., 0.996
\item \strong{NON_WM_HYPOINTENSITIES_VOLUME} \emph{numeric} non-wm-hypointensities ROI size in mm^3 Range value: 1, ..., 233
\item \strong{CTX_BANKSSTS_SUVR} \emph{numeric} ctx-bilateral-bankssts Range value: 0.692, ..., 8.47
\item \strong{CTX_BANKSSTS_VOLUME} \emph{numeric} ctx-bilateral-bankssts ROI size in mm^3 Range value: 2121, ..., 8963
\item \strong{CTX_CAUDALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-bilateral-caudalanteriorcingulate Range value: 0.036, ..., 3.894
\item \strong{CTX_CAUDALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-bilateral-caudalanteriorcingulate ROI size in mm^3 Range value: 1309, ..., 8378
\item \strong{CTX_CAUDALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-bilateral-caudalmiddlefrontal Range value: 0.553, ..., 7.984
\item \strong{CTX_CAUDALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-caudalmiddlefrontal ROI size in mm^3 Range value: 5775, ..., 16670
\item \strong{CTX_CUNEUS_SUVR} \emph{numeric} ctx-bilateral-cuneus Range value: 0.538, ..., 9.75
\item \strong{CTX_CUNEUS_VOLUME} \emph{numeric} ctx-bilateral-cuneus ROI size in mm^3 Range value: 3366, ..., 10467
\item \strong{CTX_FRONTALPOLE_SUVR} \emph{numeric} ctx-bilateral-frontalpole Range value: 0.098, ..., 6.29
\item \strong{CTX_FRONTALPOLE_VOLUME} \emph{numeric} ctx-bilateral-frontalpole ROI size in mm^3 Range value: 982, ..., 3888
\item \strong{CTX_FUSIFORM_SUVR} \emph{numeric} ctx-bilateral-fusiform Range value: 0.459, ..., 8.054
\item \strong{CTX_FUSIFORM_VOLUME} \emph{numeric} ctx-bilateral-fusiform ROI size in mm^3 Range value: 9510, ..., 26688
\item \strong{CTX_INFERIORPARIETAL_SUVR} \emph{numeric} ctx-bilateral-inferiorparietal Range value: 0.811, ..., 8.66
\item \strong{CTX_INFERIORPARIETAL_VOLUME} \emph{numeric} ctx-bilateral-inferiorparietal ROI size in mm^3 Range value: 12378, ..., 37678
\item \strong{CTX_INFERIORTEMPORAL_SUVR} \emph{numeric} ctx-bilateral-inferiortemporal Range value: 0.687, ..., 8.074
\item \strong{CTX_INFERIORTEMPORAL_VOLUME} \emph{numeric} ctx-bilateral-inferiortemporal ROI size in mm^3 Range value: 9660, ..., 31454
\item \strong{CTX_INSULA_SUVR} \emph{numeric} ctx-bilateral-insula Range value: 0.635, ..., 4.639
\item \strong{CTX_INSULA_VOLUME} \emph{numeric} ctx-bilateral-insula ROI size in mm^3 Range value: 7460, ..., 27196
\item \strong{CTX_ISTHMUSCINGULATE_SUVR} \emph{numeric} ctx-bilateral-isthmuscingulate Range value: 0.564, ..., 7.722
\item \strong{CTX_ISTHMUSCINGULATE_VOLUME} \emph{numeric} ctx-bilateral-isthmuscingulate ROI size in mm^3 Range value: 1985, ..., 9811
\item \strong{CTX_LATERALOCCIPITAL_SUVR} \emph{numeric} ctx-bilateral-lateraloccipital Range value: 0.454, ..., 11.414
\item \strong{CTX_LATERALOCCIPITAL_VOLUME} \emph{numeric} ctx-bilateral-lateraloccipital ROI size in mm^3 Range value: 9370, ..., 35627
\item \strong{CTX_LATERALORBITOFRONTAL_SUVR} \emph{numeric} ctx-bilateral-lateralorbitofrontal Range value: 0.583, ..., 5.529
\item \strong{CTX_LATERALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-lateralorbitofrontal ROI size in mm^3 Range value: 8842, ..., 18822
\item \strong{CTX_LINGUAL_SUVR} \emph{numeric} ctx-bilateral-lingual Range value: 0.491, ..., 8.206
\item \strong{CTX_LINGUAL_VOLUME} \emph{numeric} ctx-bilateral-lingual ROI size in mm^3 Range value: 5946, ..., 20629
\item \strong{CTX_MEDIALORBITOFRONTAL_SUVR} \emph{numeric} ctx-bilateral-medialorbitofrontal Range value: 0.511, ..., 4.693
\item \strong{CTX_MEDIALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-medialorbitofrontal ROI size in mm^3 Range value: 6625, ..., 13876
\item \strong{CTX_MIDDLETEMPORAL_SUVR} \emph{numeric} ctx-bilateral-middletemporal Range value: 0.508, ..., 6.668
\item \strong{CTX_MIDDLETEMPORAL_VOLUME} \emph{numeric} ctx-bilateral-middletemporal ROI size in mm^3 Range value: 11213, ..., 31685
\item \strong{CTX_PARACENTRAL_SUVR} \emph{numeric} ctx-bilateral-paracentral Range value: 0.576, ..., 3.298
\item \strong{CTX_PARACENTRAL_VOLUME} \emph{numeric} ctx-bilateral-paracentral ROI size in mm^3 Range value: 4500, ..., 10986
\item \strong{CTX_PARAHIPPOCAMPAL_SUVR} \emph{numeric} ctx-bilateral-parahippocampal Range value: 0.279, ..., 5.406
\item \strong{CTX_PARAHIPPOCAMPAL_VOLUME} \emph{numeric} ctx-bilateral-parahippocampal ROI size in mm^3 Range value: 1751, ..., 5685
\item \strong{CTX_PARSOPERCULARIS_SUVR} \emph{numeric} ctx-bilateral-parsopercularis Range value: 0.559, ..., 6.747
\item \strong{CTX_PARSOPERCULARIS_VOLUME} \emph{numeric} ctx-bilateral-parsopercularis ROI size in mm^3 Range value: 4801, ..., 12836
\item \strong{CTX_PARSORBITALIS_SUVR} \emph{numeric} ctx-bilateral-parsorbitalis Range value: 0.661, ..., 6.378
\item \strong{CTX_PARSORBITALIS_VOLUME} \emph{numeric} ctx-bilateral-parsorbitalis ROI size in mm^3 Range value: 2452, ..., 6892
\item \strong{CTX_PARSTRIANGULARIS_SUVR} \emph{numeric} ctx-bilateral-parstriangularis Range value: 0.634, ..., 5.871
\item \strong{CTX_PARSTRIANGULARIS_VOLUME} \emph{numeric} ctx-bilateral-parstriangularis ROI size in mm^3 Range value: 4786, ..., 11719
\item \strong{CTX_PERICALCARINE_SUVR} \emph{numeric} ctx-bilateral-pericalcarine Range value: 0.313, ..., 5.844
\item \strong{CTX_PERICALCARINE_VOLUME} \emph{numeric} ctx-bilateral-pericalcarine ROI size in mm^3 Range value: 2165, ..., 8270
\item \strong{CTX_POSTCENTRAL_SUVR} \emph{numeric} ctx-bilateral-postcentral Range value: 0.673, ..., 3.712
\item \strong{CTX_POSTCENTRAL_VOLUME} \emph{numeric} ctx-bilateral-postcentral ROI size in mm^3 Range value: 10032, ..., 26525
\item \strong{CTX_POSTERIORCINGULATE_SUVR} \emph{numeric} ctx-bilateral-posteriorcingulate Range value: 0.624, ..., 6.589
\item \strong{CTX_POSTERIORCINGULATE_VOLUME} \emph{numeric} ctx-bilateral-posteriorcingulate ROI size in mm^3 Range value: 2054, ..., 18216
\item \strong{CTX_PRECENTRAL_SUVR} \emph{numeric} ctx-bilateral-precentral Range value: 0.569, ..., 3.127
\item \strong{CTX_PRECENTRAL_VOLUME} \emph{numeric} ctx-bilateral-precentral ROI size in mm^3 Range value: 13766, ..., 37386
\item \strong{CTX_PRECUNEUS_SUVR} \emph{numeric} ctx-bilateral-precuneus Range value: 0.717, ..., 8.131
\item \strong{CTX_PRECUNEUS_VOLUME} \emph{numeric} ctx-bilateral-precuneus ROI size in mm^3 Range value: 9272, ..., 27581
\item \strong{CTX_ROSTRALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-bilateral-rostralanteriorcingulate Range value: 0.497, ..., 4.24
\item \strong{CTX_ROSTRALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-bilateral-rostralanteriorcingulate ROI size in mm^3 Range value: 1558, ..., 7474
\item \strong{CTX_ROSTRALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-bilateral-rostralmiddlefrontal Range value: 0.661, ..., 9.23
\item \strong{CTX_ROSTRALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-rostralmiddlefrontal ROI size in mm^3 Range value: 16054, ..., 50046
\item \strong{CTX_SUPERIORFRONTAL_SUVR} \emph{numeric} ctx-bilateral-superiorfrontal Range value: 0.589, ..., 5.643
\item \strong{CTX_SUPERIORFRONTAL_VOLUME} \emph{numeric} ctx-bilateral-superiorfrontal ROI size in mm^3 Range value: 25550, ..., 57959
\item \strong{CTX_SUPERIORPARIETAL_SUVR} \emph{numeric} ctx-bilateral-superiorparietal Range value: 0.629, ..., 9.384
\item \strong{CTX_SUPERIORPARIETAL_VOLUME} \emph{numeric} ctx-bilateral-superiorparietal ROI size in mm^3 Range value: 12027, ..., 37095
\item \strong{CTX_SUPERIORTEMPORAL_SUVR} \emph{numeric} ctx-bilateral-superiortemporal Range value: 0.679, ..., 4.761
\item \strong{CTX_SUPERIORTEMPORAL_VOLUME} \emph{numeric} ctx-bilateral-superiortemporal ROI size in mm^3 Range value: 12461, ..., 32667
\item \strong{CTX_SUPRAMARGINAL_SUVR} \emph{numeric} ctx-bilateral-supramarginal Range value: 0.722, ..., 6.568
\item \strong{CTX_SUPRAMARGINAL_VOLUME} \emph{numeric} ctx-bilateral-supramarginal ROI size in mm^3 Range value: 10961, ..., 31391
\item \strong{CTX_TEMPORALPOLE_SUVR} \emph{numeric} ctx-bilateral-temporalpole Range value: 0.394, ..., 6.179
\item \strong{CTX_TEMPORALPOLE_VOLUME} \emph{numeric} ctx-bilateral-temporalpole ROI size in mm^3 Range value: 2178, ..., 9644
\item \strong{CTX_TRANSVERSETEMPORAL_SUVR} \emph{numeric} ctx-bilateral-transversetemporal Range value: 0.247, ..., 5.094
\item \strong{CTX_TRANSVERSETEMPORAL_VOLUME} \emph{numeric} ctx-bilateral-transversetemporal ROI size in mm^3 Range value: 1100, ..., 3541
\item \strong{ACCUMBENS_AREA_SUVR} \emph{numeric} bilateral-accumbens-area Range value: 0.186, ..., 4.112
\item \strong{ACCUMBENS_AREA_VOLUME} \emph{numeric} bilateral-accumbens-area ROI size in mm^3 Range value: 300, ..., 1636
\item \strong{AMYGDALA_SUVR} \emph{numeric} bilateral-amygdala Range value: 0.364, ..., 4.779
\item \strong{AMYGDALA_VOLUME} \emph{numeric} bilateral-amygdala ROI size in mm^3 Range value: 980, ..., 4740
\item \strong{CAUDATE_SUVR} \emph{numeric} bilateral-caudate Range value: 0.397, ..., 2.545
\item \strong{CAUDATE_VOLUME} \emph{numeric} bilateral-caudate ROI size in mm^3 Range value: 3588, ..., 12321
\item \strong{CEREBELLUM_CORTEX_SUVR} \emph{numeric} bilateral-cerebellum-cortex Range value: 0.489, ..., 1.357
\item \strong{CEREBELLUM_CORTEX_VOLUME} \emph{numeric} bilateral-cerebellum-cortex ROI size in mm^3 Range value: 68832, ..., 142592
\item \strong{CEREBELLUM_WHITE_MATTER_SUVR} \emph{numeric} bilateral-cerebellum-white-matter Range value: 0.378, ..., 1.868
\item \strong{CEREBELLUM_WHITE_MATTER_VOLUME} \emph{numeric} bilateral-cerebellum-white-matter ROI size in mm^3 Range value: 13671, ..., 48785
\item \strong{CHOROID_PLEXUS_SUVR} \emph{numeric} bilateral-choroid-plexus Range value: 0.504, ..., 16.369
\item \strong{CHOROID_PLEXUS_VOLUME} \emph{numeric} bilateral-choroid-plexus ROI size in mm^3 Range value: 459, ..., 3434
\item \strong{HIPPOCAMPUS_SUVR} \emph{numeric} bilateral-hippocampus Range value: 0.547, ..., 3.284
\item \strong{HIPPOCAMPUS_VOLUME} \emph{numeric} bilateral-hippocampus ROI size in mm^3 Range value: 3255, ..., 11461
\item \strong{OPTIC_CHIASM_SUVR} \emph{numeric} optic-chiasm Range value: 0.001, ..., 3.02
\item \strong{OPTIC_CHIASM_VOLUME} \emph{numeric} optic-chiasm ROI size in mm^3 Range value: 2, ..., 361
\item \strong{INF_LAT_VENT_SUVR} \emph{numeric} bilateral-inf-lat-vent Range value: 0.004, ..., 6.095
\item \strong{INF_LAT_VENT_VOLUME} \emph{numeric} bilateral-inf-lat-vent ROI size in mm^3 Range value: 164, ..., 16102
\item \strong{LATERAL_VENTRICLE_SUVR} \emph{numeric} bilateral-lateral-ventricle Range value: 0.005, ..., 1.263
\item \strong{LATERAL_VENTRICLE_VOLUME} \emph{numeric} bilateral-lateral-ventricle ROI size in mm^3 Range value: 5800, ..., 138561
\item \strong{PALLIDUM_SUVR} \emph{numeric} bilateral-pallidum Range value: 0.435, ..., 3.077
\item \strong{PALLIDUM_VOLUME} \emph{numeric} bilateral-pallidum ROI size in mm^3 Range value: 2454, ..., 6324
\item \strong{PUTAMEN_SUVR} \emph{numeric} bilateral-putamen Range value: 0.479, ..., 2.777
\item \strong{PUTAMEN_VOLUME} \emph{numeric} bilateral-putamen ROI size in mm^3 Range value: 4438, ..., 15492
\item \strong{THALAMUS_PROPER_SUVR} \emph{numeric} bilateral-thalamus-proper Range value: 0.482, ..., 2.052
\item \strong{THALAMUS_PROPER_VOLUME} \emph{numeric} bilateral-thalamus-proper ROI size in mm^3 Range value: 8557, ..., 20466
\item \strong{VENTRALDC_SUVR} \emph{numeric} bilateral-ventraldc Range value: 0.525, ..., 2.208
\item \strong{VENTRALDC_VOLUME} \emph{numeric} bilateral-ventraldc ROI size in mm^3 Range value: 4441, ..., 11382
\item \strong{VESSEL_SUVR} \emph{logical} bilateral-vessel A Boolean value: TRUE/ FALSE
\item \strong{VESSEL_VOLUME} \emph{numeric} bilateral-vessel ROI size in mm^3 Range value: 1, ..., 604
\item \strong{CTX_LH_BANKSSTS_SUVR} \emph{numeric} ctx-lh-bankssts Range value: 0.638, ..., 8.571
\item \strong{CTX_LH_BANKSSTS_VOLUME} \emph{numeric} ctx-lh-bankssts ROI size in mm^3 Range value: 779, ..., 4723
\item \strong{CTX_LH_CAUDALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-lh-caudalanteriorcingulate Range value: 0.083, ..., 3.875
\item \strong{CTX_LH_CAUDALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-lh-caudalanteriorcingulate ROI size in mm^3 Range value: 352, ..., 4960
\item \strong{CTX_LH_CAUDALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-lh-caudalmiddlefrontal Range value: 0.55, ..., 8.657
\item \strong{CTX_LH_CAUDALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-lh-caudalmiddlefrontal ROI size in mm^3 Range value: 2609, ..., 12065
\item \strong{CTX_LH_CUNEUS_SUVR} \emph{numeric} ctx-lh-cuneus Range value: 0.461, ..., 7.967
\item \strong{CTX_LH_CUNEUS_VOLUME} \emph{numeric} ctx-lh-cuneus ROI size in mm^3 Range value: 1411, ..., 5342
\item \strong{CTX_LH_ENTORHINAL_SUVR} \emph{numeric} ctx-lh-entorhinal Range value: 0.645, ..., 6.03
\item \strong{CTX_LH_ENTORHINAL_VOLUME} \emph{numeric} ctx-lh-entorhinal ROI size in mm^3 Range value: 291, ..., 3408
\item \strong{CTX_LH_FRONTALPOLE_SUVR} \emph{numeric} ctx-lh-frontalpole Range value: 0.156, ..., 6.617
\item \strong{CTX_LH_FRONTALPOLE_VOLUME} \emph{numeric} ctx-lh-frontalpole ROI size in mm^3 Range value: 486, ..., 1796
\item \strong{CTX_LH_FUSIFORM_SUVR} \emph{numeric} ctx-lh-fusiform Range value: 0.506, ..., 8.781
\item \strong{CTX_LH_FUSIFORM_VOLUME} \emph{numeric} ctx-lh-fusiform ROI size in mm^3 Range value: 3944, ..., 13678
\item \strong{CTX_LH_INFERIORPARIETAL_SUVR} \emph{numeric} ctx-lh-inferiorparietal Range value: 0.763, ..., 9.371
\item \strong{CTX_LH_INFERIORPARIETAL_VOLUME} \emph{numeric} ctx-lh-inferiorparietal ROI size in mm^3 Range value: 5047, ..., 17756
\item \strong{CTX_LH_INFERIORTEMPORAL_SUVR} \emph{numeric} ctx-lh-inferiortemporal Range value: 0.692, ..., 7.794
\item \strong{CTX_LH_INFERIORTEMPORAL_VOLUME} \emph{numeric} ctx-lh-inferiortemporal ROI size in mm^3 Range value: 4547, ..., 16512
\item \strong{CTX_LH_INSULA_SUVR} \emph{numeric} ctx-lh-insula Range value: 0.642, ..., 4.563
\item \strong{CTX_LH_INSULA_VOLUME} \emph{numeric} ctx-lh-insula ROI size in mm^3 Range value: 3825, ..., 21138
\item \strong{CTX_LH_ISTHMUSCINGULATE_SUVR} \emph{numeric} ctx-lh-isthmuscingulate Range value: 0.518, ..., 7.809
\item \strong{CTX_LH_ISTHMUSCINGULATE_VOLUME} \emph{numeric} ctx-lh-isthmuscingulate ROI size in mm^3 Range value: 615, ..., 6967
\item \strong{CTX_LH_LATERALOCCIPITAL_SUVR} \emph{numeric} ctx-lh-lateraloccipital Range value: 0.616, ..., 10.54
\item \strong{CTX_LH_LATERALOCCIPITAL_VOLUME} \emph{numeric} ctx-lh-lateraloccipital ROI size in mm^3 Range value: 4386, ..., 17628
\item \strong{CTX_LH_LATERALORBITOFRONTAL_SUVR} \emph{numeric} ctx-lh-lateralorbitofrontal Range value: 0.602, ..., 6.507
\item \strong{CTX_LH_LATERALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-lh-lateralorbitofrontal ROI size in mm^3 Range value: 3412, ..., 11834
\item \strong{CTX_LH_LINGUAL_SUVR} \emph{numeric} ctx-lh-lingual Range value: 0.404, ..., 7.338
\item \strong{CTX_LH_LINGUAL_VOLUME} \emph{numeric} ctx-lh-lingual ROI size in mm^3 Range value: 2540, ..., 11629
\item \strong{CTX_LH_MEDIALORBITOFRONTAL_SUVR} \emph{numeric} ctx-lh-medialorbitofrontal Range value: 0.49, ..., 4.624
\item \strong{CTX_LH_MEDIALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-lh-medialorbitofrontal ROI size in mm^3 Range value: 3183, ..., 7019
\item \strong{CTX_LH_MIDDLETEMPORAL_SUVR} \emph{numeric} ctx-lh-middletemporal Range value: 0.467, ..., 7.928
\item \strong{CTX_LH_MIDDLETEMPORAL_VOLUME} \emph{numeric} ctx-lh-middletemporal ROI size in mm^3 Range value: 4242, ..., 14767
\item \strong{CTX_LH_PARACENTRAL_SUVR} \emph{numeric} ctx-lh-paracentral Range value: 0.521, ..., 3.251
\item \strong{CTX_LH_PARACENTRAL_VOLUME} \emph{numeric} ctx-lh-paracentral ROI size in mm^3 Range value: 1868, ..., 7180
\item \strong{CTX_LH_PARAHIPPOCAMPAL_SUVR} \emph{numeric} ctx-lh-parahippocampal Range value: 0.527, ..., 5.488
\item \strong{CTX_LH_PARAHIPPOCAMPAL_VOLUME} \emph{numeric} ctx-lh-parahippocampal ROI size in mm^3 Range value: 938, ..., 3643
\item \strong{CTX_LH_PARSOPERCULARIS_SUVR} \emph{numeric} ctx-lh-parsopercularis Range value: 0.532, ..., 6.923
\item \strong{CTX_LH_PARSOPERCULARIS_VOLUME} \emph{numeric} ctx-lh-parsopercularis ROI size in mm^3 Range value: 2451, ..., 7630
\item \strong{CTX_LH_PARSORBITALIS_SUVR} \emph{numeric} ctx-lh-parsorbitalis Range value: 0.506, ..., 7.216
\item \strong{CTX_LH_PARSORBITALIS_VOLUME} \emph{numeric} ctx-lh-parsorbitalis ROI size in mm^3 Range value: 438, ..., 3294
\item \strong{CTX_LH_PARSTRIANGULARIS_SUVR} \emph{numeric} ctx-lh-parstriangularis Range value: 0.587, ..., 6.51
\item \strong{CTX_LH_PARSTRIANGULARIS_VOLUME} \emph{numeric} ctx-lh-parstriangularis ROI size in mm^3 Range value: 1834, ..., 5545
\item \strong{CTX_LH_PERICALCARINE_SUVR} \emph{numeric} ctx-lh-pericalcarine Range value: 0.292, ..., 4.847
\item \strong{CTX_LH_PERICALCARINE_VOLUME} \emph{numeric} ctx-lh-pericalcarine ROI size in mm^3 Range value: 941, ..., 4678
\item \strong{CTX_LH_POSTCENTRAL_SUVR} \emph{numeric} ctx-lh-postcentral Range value: 0.645, ..., 3.436
\item \strong{CTX_LH_POSTCENTRAL_VOLUME} \emph{numeric} ctx-lh-postcentral ROI size in mm^3 Range value: 4696, ..., 14589
\item \strong{CTX_LH_POSTERIORCINGULATE_SUVR} \emph{numeric} ctx-lh-posteriorcingulate Range value: 0.452, ..., 6.441
\item \strong{CTX_LH_POSTERIORCINGULATE_VOLUME} \emph{numeric} ctx-lh-posteriorcingulate ROI size in mm^3 Range value: 990, ..., 11595
\item \strong{CTX_LH_PRECENTRAL_SUVR} \emph{numeric} ctx-lh-precentral Range value: 0.563, ..., 3.627
\item \strong{CTX_LH_PRECENTRAL_VOLUME} \emph{numeric} ctx-lh-precentral ROI size in mm^3 Range value: 6606, ..., 25709
\item \strong{CTX_LH_PRECUNEUS_SUVR} \emph{numeric} ctx-lh-precuneus Range value: 0.699, ..., 8.728
\item \strong{CTX_LH_PRECUNEUS_VOLUME} \emph{numeric} ctx-lh-precuneus ROI size in mm^3 Range value: 4103, ..., 17038
\item \strong{CTX_LH_ROSTRALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-lh-rostralanteriorcingulate Range value: 0.593, ..., 4.188
\item \strong{CTX_LH_ROSTRALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-lh-rostralanteriorcingulate ROI size in mm^3 Range value: 571, ..., 5073
\item \strong{CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-lh-rostralmiddlefrontal Range value: 0.618, ..., 9.649
\item \strong{CTX_LH_ROSTRALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-lh-rostralmiddlefrontal ROI size in mm^3 Range value: 6861, ..., 23211
\item \strong{CTX_LH_SUPERIORFRONTAL_SUVR} \emph{numeric} ctx-lh-superiorfrontal Range value: 0.587, ..., 6.289
\item \strong{CTX_LH_SUPERIORFRONTAL_VOLUME} \emph{numeric} ctx-lh-superiorfrontal ROI size in mm^3 Range value: 13132, ..., 39158
\item \strong{CTX_LH_SUPERIORPARIETAL_SUVR} \emph{numeric} ctx-lh-superiorparietal Range value: 0.558, ..., 8.873
\item \strong{CTX_LH_SUPERIORPARIETAL_VOLUME} \emph{numeric} ctx-lh-superiorparietal ROI size in mm^3 Range value: 5411, ..., 23420
\item \strong{CTX_LH_SUPERIORTEMPORAL_SUVR} \emph{numeric} ctx-lh-superiortemporal Range value: 0.678, ..., 5.366
\item \strong{CTX_LH_SUPERIORTEMPORAL_VOLUME} \emph{numeric} ctx-lh-superiortemporal ROI size in mm^3 Range value: 5891, ..., 18312
\item \strong{CTX_LH_SUPRAMARGINAL_SUVR} \emph{numeric} ctx-lh-supramarginal Range value: 0.715, ..., 6.942
\item \strong{CTX_LH_SUPRAMARGINAL_VOLUME} \emph{numeric} ctx-lh-supramarginal ROI size in mm^3 Range value: 5543, ..., 17431
\item \strong{CTX_LH_TEMPORALPOLE_SUVR} \emph{numeric} ctx-lh-temporalpole Range value: 0.477, ..., 5.922
\item \strong{CTX_LH_TEMPORALPOLE_VOLUME} \emph{numeric} ctx-lh-temporalpole ROI size in mm^3 Range value: 48, ..., 5851
\item \strong{CTX_LH_TRANSVERSETEMPORAL_SUVR} \emph{numeric} ctx-lh-transversetemporal Range value: 0.228, ..., 4.831
\item \strong{CTX_LH_TRANSVERSETEMPORAL_VOLUME} \emph{numeric} ctx-lh-transversetemporal ROI size in mm^3 Range value: 601, ..., 2254
\item \strong{CTX_RH_BANKSSTS_SUVR} \emph{numeric} ctx-rh-bankssts Range value: 0.694, ..., 10.774
\item \strong{CTX_RH_BANKSSTS_VOLUME} \emph{numeric} ctx-rh-bankssts ROI size in mm^3 Range value: 952, ..., 4240
\item \strong{CTX_RH_CAUDALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-rh-caudalanteriorcingulate Range value: 0.093, ..., 3.907
\item \strong{CTX_RH_CAUDALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-rh-caudalanteriorcingulate ROI size in mm^3 Range value: 368, ..., 4843
\item \strong{CTX_RH_CAUDALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-rh-caudalmiddlefrontal Range value: 0.556, ..., 7.329
\item \strong{CTX_RH_CAUDALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-rh-caudalmiddlefrontal ROI size in mm^3 Range value: 2379, ..., 9461
\item \strong{CTX_RH_CUNEUS_SUVR} \emph{numeric} ctx-rh-cuneus Range value: 0.573, ..., 11.746
\item \strong{CTX_RH_CUNEUS_VOLUME} \emph{numeric} ctx-rh-cuneus ROI size in mm^3 Range value: 982, ..., 6165
\item \strong{CTX_RH_ENTORHINAL_SUVR} \emph{numeric} ctx-rh-entorhinal Range value: 0.552, ..., 5.84
\item \strong{CTX_RH_ENTORHINAL_VOLUME} \emph{numeric} ctx-rh-entorhinal ROI size in mm^3 Range value: 411, ..., 3452
\item \strong{CTX_RH_FRONTALPOLE_SUVR} \emph{numeric} ctx-rh-frontalpole Range value: 0.05, ..., 6.011
\item \strong{CTX_RH_FRONTALPOLE_VOLUME} \emph{numeric} ctx-rh-frontalpole ROI size in mm^3 Range value: 381, ..., 2370
\item \strong{CTX_RH_FUSIFORM_SUVR} \emph{numeric} ctx-rh-fusiform Range value: 0.407, ..., 7.601
\item \strong{CTX_RH_FUSIFORM_VOLUME} \emph{numeric} ctx-rh-fusiform ROI size in mm^3 Range value: 4673, ..., 14227
\item \strong{CTX_RH_INFERIORPARIETAL_SUVR} \emph{numeric} ctx-rh-inferiorparietal Range value: 0.791, ..., 9.632
\item \strong{CTX_RH_INFERIORPARIETAL_VOLUME} \emph{numeric} ctx-rh-inferiorparietal ROI size in mm^3 Range value: 6272, ..., 21817
\item \strong{CTX_RH_INFERIORTEMPORAL_SUVR} \emph{numeric} ctx-rh-inferiortemporal Range value: 0.682, ..., 8.895
\item \strong{CTX_RH_INFERIORTEMPORAL_VOLUME} \emph{numeric} ctx-rh-inferiortemporal ROI size in mm^3 Range value: 4566, ..., 16344
\item \strong{CTX_RH_INSULA_SUVR} \emph{numeric} ctx-rh-insula Range value: 0.629, ..., 4.712
\item \strong{CTX_RH_INSULA_VOLUME} \emph{numeric} ctx-rh-insula ROI size in mm^3 Range value: 3635, ..., 9212
\item \strong{CTX_RH_ISTHMUSCINGULATE_SUVR} \emph{numeric} ctx-rh-isthmuscingulate Range value: 0.615, ..., 7.641
\item \strong{CTX_RH_ISTHMUSCINGULATE_VOLUME} \emph{numeric} ctx-rh-isthmuscingulate ROI size in mm^3 Range value: 783, ..., 4091
\item \strong{CTX_RH_LATERALOCCIPITAL_SUVR} \emph{numeric} ctx-rh-lateraloccipital Range value: 0.292, ..., 12.431
\item \strong{CTX_RH_LATERALOCCIPITAL_VOLUME} \emph{numeric} ctx-rh-lateraloccipital ROI size in mm^3 Range value: 4984, ..., 18366
\item \strong{CTX_RH_LATERALORBITOFRONTAL_SUVR} \emph{numeric} ctx-rh-lateralorbitofrontal Range value: 0.564, ..., 5.269
\item \strong{CTX_RH_LATERALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-rh-lateralorbitofrontal ROI size in mm^3 Range value: 4176, ..., 9667
\item \strong{CTX_RH_LINGUAL_SUVR} \emph{numeric} ctx-rh-lingual Range value: 0.639, ..., 9.147
\item \strong{CTX_RH_LINGUAL_VOLUME} \emph{numeric} ctx-rh-lingual ROI size in mm^3 Range value: 3406, ..., 10478
\item \strong{CTX_RH_MEDIALORBITOFRONTAL_SUVR} \emph{numeric} ctx-rh-medialorbitofrontal Range value: 0.527, ..., 4.993
\item \strong{CTX_RH_MEDIALORBITOFRONTAL_VOLUME} \emph{numeric} ctx-rh-medialorbitofrontal ROI size in mm^3 Range value: 3216, ..., 7241
\item \strong{CTX_RH_MIDDLETEMPORAL_SUVR} \emph{numeric} ctx-rh-middletemporal Range value: 0.548, ..., 6.727
\item \strong{CTX_RH_MIDDLETEMPORAL_VOLUME} \emph{numeric} ctx-rh-middletemporal ROI size in mm^3 Range value: 5922, ..., 17219
\item \strong{CTX_RH_PARACENTRAL_SUVR} \emph{numeric} ctx-rh-paracentral Range value: 0.618, ..., 4.662
\item \strong{CTX_RH_PARACENTRAL_VOLUME} \emph{numeric} ctx-rh-paracentral ROI size in mm^3 Range value: 2261, ..., 5526
\item \strong{CTX_RH_PARAHIPPOCAMPAL_SUVR} \emph{numeric} ctx-rh-parahippocampal Range value: 0.586, ..., 5.711
\item \strong{CTX_RH_PARAHIPPOCAMPAL_VOLUME} \emph{numeric} ctx-rh-parahippocampal ROI size in mm^3 Range value: 1096, ..., 2800
\item \strong{CTX_RH_PARSOPERCULARIS_SUVR} \emph{numeric} ctx-rh-parsopercularis Range value: 0.593, ..., 6.516
\item \strong{CTX_RH_PARSOPERCULARIS_VOLUME} \emph{numeric} ctx-rh-parsopercularis ROI size in mm^3 Range value: 1783, ..., 7542
\item \strong{CTX_RH_PARSORBITALIS_SUVR} \emph{numeric} ctx-rh-parsorbitalis Range value: 0.528, ..., 5.569
\item \strong{CTX_RH_PARSORBITALIS_VOLUME} \emph{numeric} ctx-rh-parsorbitalis ROI size in mm^3 Range value: 1261, ..., 4301
\item \strong{CTX_RH_PARSTRIANGULARIS_SUVR} \emph{numeric} ctx-rh-parstriangularis Range value: 0.631, ..., 5.286
\item \strong{CTX_RH_PARSTRIANGULARIS_VOLUME} \emph{numeric} ctx-rh-parstriangularis ROI size in mm^3 Range value: 2268, ..., 7512
\item \strong{CTX_RH_PERICALCARINE_SUVR} \emph{numeric} ctx-rh-pericalcarine Range value: 0.156, ..., 6.99
\item \strong{CTX_RH_PERICALCARINE_VOLUME} \emph{numeric} ctx-rh-pericalcarine ROI size in mm^3 Range value: 831, ..., 4617
\item \strong{CTX_RH_POSTCENTRAL_SUVR} \emph{numeric} ctx-rh-postcentral Range value: 0.702, ..., 4.752
\item \strong{CTX_RH_POSTCENTRAL_VOLUME} \emph{numeric} ctx-rh-postcentral ROI size in mm^3 Range value: 4677, ..., 13254
\item \strong{CTX_RH_POSTERIORCINGULATE_SUVR} \emph{numeric} ctx-rh-posteriorcingulate Range value: 0.721, ..., 6.73
\item \strong{CTX_RH_POSTERIORCINGULATE_VOLUME} \emph{numeric} ctx-rh-posteriorcingulate ROI size in mm^3 Range value: 994, ..., 6621
\item \strong{CTX_RH_PRECENTRAL_SUVR} \emph{numeric} ctx-rh-precentral Range value: 0.575, ..., 3.159
\item \strong{CTX_RH_PRECENTRAL_VOLUME} \emph{numeric} ctx-rh-precentral ROI size in mm^3 Range value: 6369, ..., 18058
\item \strong{CTX_RH_PRECUNEUS_SUVR} \emph{numeric} ctx-rh-precuneus Range value: 0.71, ..., 7.6
\item \strong{CTX_RH_PRECUNEUS_VOLUME} \emph{numeric} ctx-rh-precuneus ROI size in mm^3 Range value: 5009, ..., 14854
\item \strong{CTX_RH_ROSTRALANTERIORCINGULATE_SUVR} \emph{numeric} ctx-rh-rostralanteriorcingulate Range value: 0.336, ..., 4.304
\item \strong{CTX_RH_ROSTRALANTERIORCINGULATE_VOLUME} \emph{numeric} ctx-rh-rostralanteriorcingulate ROI size in mm^3 Range value: 559, ..., 3508
\item \strong{CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR} \emph{numeric} ctx-rh-rostralmiddlefrontal Range value: 0.697, ..., 8.814
\item \strong{CTX_RH_ROSTRALMIDDLEFRONTAL_VOLUME} \emph{numeric} ctx-rh-rostralmiddlefrontal ROI size in mm^3 Range value: 7884, ..., 26835
\item \strong{CTX_RH_SUPERIORFRONTAL_SUVR} \emph{numeric} ctx-rh-superiorfrontal Range value: 0.589, ..., 5.001
\item \strong{CTX_RH_SUPERIORFRONTAL_VOLUME} \emph{numeric} ctx-rh-superiorfrontal ROI size in mm^3 Range value: 10729, ..., 28618
\item \strong{CTX_RH_SUPERIORPARIETAL_SUVR} \emph{numeric} ctx-rh-superiorparietal Range value: 0.587, ..., 9.934
\item \strong{CTX_RH_SUPERIORPARIETAL_VOLUME} \emph{numeric} ctx-rh-superiorparietal ROI size in mm^3 Range value: 6110, ..., 18993
\item \strong{CTX_RH_SUPERIORTEMPORAL_SUVR} \emph{numeric} ctx-rh-superiortemporal Range value: 0.68, ..., 5.169
\item \strong{CTX_RH_SUPERIORTEMPORAL_VOLUME} \emph{numeric} ctx-rh-superiortemporal ROI size in mm^3 Range value: 5963, ..., 15383
\item \strong{CTX_RH_SUPRAMARGINAL_SUVR} \emph{numeric} ctx-rh-supramarginal Range value: 0.731, ..., 7.764
\item \strong{CTX_RH_SUPRAMARGINAL_VOLUME} \emph{numeric} ctx-rh-supramarginal ROI size in mm^3 Range value: 4622, ..., 14663
\item \strong{CTX_RH_TEMPORALPOLE_SUVR} \emph{numeric} ctx-rh-temporalpole Range value: 0.319, ..., 6.325
\item \strong{CTX_RH_TEMPORALPOLE_VOLUME} \emph{numeric} ctx-rh-temporalpole ROI size in mm^3 Range value: 287, ..., 4666
\item \strong{CTX_RH_TRANSVERSETEMPORAL_SUVR} \emph{numeric} ctx-rh-transversetemporal Range value: 0.08, ..., 5.439
\item \strong{CTX_RH_TRANSVERSETEMPORAL_VOLUME} \emph{numeric} ctx-rh-transversetemporal ROI size in mm^3 Range value: 461, ..., 1566
\item \strong{LEFT_ACCUMBENS_AREA_SUVR} \emph{numeric} left-accumbens-area Range value: 0.051, ..., 4.736
\item \strong{LEFT_ACCUMBENS_AREA_VOLUME} \emph{numeric} left-accumbens-area ROI size in mm^3 Range value: 116, ..., 821
\item \strong{LEFT_AMYGDALA_SUVR} \emph{numeric} left-amygdala Range value: 0.367, ..., 5.24
\item \strong{LEFT_AMYGDALA_VOLUME} \emph{numeric} left-amygdala ROI size in mm^3 Range value: 375, ..., 2468
\item \strong{LEFT_CAUDATE_SUVR} \emph{numeric} left-caudate Range value: 0.399, ..., 2.459
\item \strong{LEFT_CAUDATE_VOLUME} \emph{numeric} left-caudate ROI size in mm^3 Range value: 1871, ..., 5745
\item \strong{LEFT_CEREBELLUM_CORTEX_SUVR} \emph{numeric} left-cerebellum-cortex Range value: 0.443, ..., 1.356
\item \strong{LEFT_CEREBELLUM_CORTEX_VOLUME} \emph{numeric} left-cerebellum-cortex ROI size in mm^3 Range value: 34655, ..., 69630
\item \strong{LEFT_CEREBELLUM_WHITE_MATTER_SUVR} \emph{numeric} left-cerebellum-white-matter Range value: 0.378, ..., 1.868
\item \strong{LEFT_CEREBELLUM_WHITE_MATTER_VOLUME} \emph{numeric} left-cerebellum-white-matter ROI size in mm^3 Range value: 6892, ..., 26253
\item \strong{LEFT_CEREBRAL_WHITE_MATTER_SUVR} \emph{numeric} left-cerebral-white-matter Range value: 0.46, ..., 1.859
\item \strong{LEFT_CEREBRAL_WHITE_MATTER_VOLUME} \emph{numeric} left-cerebral-white-matter ROI size in mm^3 Range value: 30465, ..., 354965
\item \strong{LEFT_CHOROID_PLEXUS_SUVR} \emph{numeric} left-choroid-plexus Range value: 0.275, ..., 18.556
\item \strong{LEFT_CHOROID_PLEXUS_VOLUME} \emph{numeric} left-choroid-plexus ROI size in mm^3 Range value: 130, ..., 1916
\item \strong{LEFT_HIPPOCAMPUS_SUVR} \emph{numeric} left-hippocampus Range value: 0.425, ..., 3.66
\item \strong{LEFT_HIPPOCAMPUS_VOLUME} \emph{numeric} left-hippocampus ROI size in mm^3 Range value: 1314, ..., 6080
\item \strong{LEFT_INF_LAT_VENT_SUVR} \emph{numeric} left-inf-lat-vent Range value: 0.004, ..., 6.095
\item \strong{LEFT_INF_LAT_VENT_VOLUME} \emph{numeric} left-inf-lat-vent ROI size in mm^3 Range value: 58, ..., 11328
\item \strong{LEFT_LATERAL_VENTRICLE_SUVR} \emph{numeric} left-lateral-ventricle Range value: 0.005, ..., 1.263
\item \strong{LEFT_LATERAL_VENTRICLE_VOLUME} \emph{numeric} left-lateral-ventricle ROI size in mm^3 Range value: 3254, ..., 79775
\item \strong{LEFT_PALLIDUM_SUVR} \emph{numeric} left-pallidum Range value: 0.438, ..., 3.087
\item \strong{LEFT_PALLIDUM_VOLUME} \emph{numeric} left-pallidum ROI size in mm^3 Range value: 1090, ..., 3436
\item \strong{LEFT_PUTAMEN_SUVR} \emph{numeric} left-putamen Range value: 0.482, ..., 2.86
\item \strong{LEFT_PUTAMEN_VOLUME} \emph{numeric} left-putamen ROI size in mm^3 Range value: 1945, ..., 7466
\item \strong{LEFT_THALAMUS_PROPER_SUVR} \emph{numeric} left-thalamus-proper Range value: 0.485, ..., 1.895
\item \strong{LEFT_THALAMUS_PROPER_VOLUME} \emph{numeric} left-thalamus-proper ROI size in mm^3 Range value: 3953, ..., 10060
\item \strong{LEFT_VENTRALDC_SUVR} \emph{numeric} left-ventraldc Range value: 0.525, ..., 2.208
\item \strong{LEFT_VENTRALDC_VOLUME} \emph{numeric} left-ventraldc ROI size in mm^3 Range value: 2259, ..., 5708
\item \strong{LEFT_VESSEL_SUVR} \emph{logical} left-vessel A Boolean value: TRUE/ FALSE
\item \strong{LEFT_VESSEL_VOLUME} \emph{numeric} left-vessel ROI size in mm^3 Range value: 1, ..., 537
\item \strong{RIGHT_ACCUMBENS_AREA_SUVR} \emph{numeric} right-accumbens-area Range value: 0.212, ..., 4.141
\item \strong{RIGHT_ACCUMBENS_AREA_VOLUME} \emph{numeric} right-accumbens-area ROI size in mm^3 Range value: 114, ..., 815
\item \strong{RIGHT_AMYGDALA_SUVR} \emph{numeric} right-amygdala Range value: 0.36, ..., 5.144
\item \strong{RIGHT_AMYGDALA_VOLUME} \emph{numeric} right-amygdala ROI size in mm^3 Range value: 595, ..., 2453
\item \strong{RIGHT_CAUDATE_SUVR} \emph{numeric} right-caudate Range value: 0.394, ..., 2.641
\item \strong{RIGHT_CAUDATE_VOLUME} \emph{numeric} right-caudate ROI size in mm^3 Range value: 1717, ..., 6880
\item \strong{RIGHT_CEREBELLUM_CORTEX_SUVR} \emph{numeric} right-cerebellum-cortex Range value: 0.534, ..., 1.357
\item \strong{RIGHT_CEREBELLUM_CORTEX_VOLUME} \emph{numeric} right-cerebellum-cortex ROI size in mm^3 Range value: 34177, ..., 73171
\item \strong{RIGHT_CEREBELLUM_WHITE_MATTER_SUVR} \emph{numeric} right-cerebellum-white-matter Range value: 0.378, ..., 1.868
\item \strong{RIGHT_CEREBELLUM_WHITE_MATTER_VOLUME} \emph{numeric} right-cerebellum-white-matter ROI size in mm^3 Range value: 6779, ..., 22532
\item \strong{RIGHT_CEREBRAL_WHITE_MATTER_SUVR} \emph{numeric} right-cerebral-white-matter Range value: 0.46, ..., 1.859
\item \strong{RIGHT_CEREBRAL_WHITE_MATTER_VOLUME} \emph{numeric} right-cerebral-white-matter ROI size in mm^3 Range value: 129854, ..., 363072
\item \strong{RIGHT_CHOROID_PLEXUS_SUVR} \emph{numeric} right-choroid-plexus Range value: 0.537, ..., 14.78
\item \strong{RIGHT_CHOROID_PLEXUS_VOLUME} \emph{numeric} right-choroid-plexus ROI size in mm^3 Range value: 126, ..., 1912
\item \strong{RIGHT_HIPPOCAMPUS_SUVR} \emph{numeric} right-hippocampus Range value: 0.548, ..., 3.575
\item \strong{RIGHT_HIPPOCAMPUS_VOLUME} \emph{numeric} right-hippocampus ROI size in mm^3 Range value: 852, ..., 5775
\item \strong{RIGHT_INF_LAT_VENT_SUVR} \emph{numeric} right-inf-lat-vent Range value: 0.004, ..., 6.095
\item \strong{RIGHT_INF_LAT_VENT_VOLUME} \emph{numeric} right-inf-lat-vent ROI size in mm^3 Range value: 58, ..., 7228
\item \strong{RIGHT_LATERAL_VENTRICLE_SUVR} \emph{numeric} right-lateral-ventricle Range value: 0.005, ..., 1.263
\item \strong{RIGHT_LATERAL_VENTRICLE_VOLUME} \emph{numeric} right-lateral-ventricle ROI size in mm^3 Range value: 2442, ..., 71992
\item \strong{RIGHT_PALLIDUM_SUVR} \emph{numeric} right-pallidum Range value: 0.41, ..., 3.187
\item \strong{RIGHT_PALLIDUM_VOLUME} \emph{numeric} right-pallidum ROI size in mm^3 Range value: 688, ..., 3207
\item \strong{RIGHT_PUTAMEN_SUVR} \emph{numeric} right-putamen Range value: 0.477, ..., 2.899
\item \strong{RIGHT_PUTAMEN_VOLUME} \emph{numeric} right-putamen ROI size in mm^3 Range value: 2106, ..., 8026
\item \strong{RIGHT_THALAMUS_PROPER_SUVR} \emph{numeric} right-thalamus-proper Range value: 0.475, ..., 2.224
\item \strong{RIGHT_THALAMUS_PROPER_VOLUME} \emph{numeric} right-thalamus-proper ROI size in mm^3 Range value: 3522, ..., 10406
\item \strong{RIGHT_VENTRALDC_SUVR} \emph{numeric} right-ventraldc Range value: 0.525, ..., 2.208
\item \strong{RIGHT_VENTRALDC_VOLUME} \emph{numeric} right-ventraldc ROI size in mm^3 Range value: 1892, ..., 5674
\item \strong{RIGHT_VESSEL_SUVR} \emph{logical} right-vessel A Boolean value: TRUE/ FALSE
\item \strong{RIGHT_VESSEL_VOLUME} \emph{numeric} right-vessel ROI size in mm^3 Range value: 1, ..., 236
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
