% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCBERKELEYFDG_8mm}
\alias{UCBERKELEYFDG_8mm}
\title{UC Berkeley - FDG analysis: 2023-02-17}
\format{
A data frame with 7524 observations and 11 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCBERKELEYFDG_8mm)
}
\description{
UC Berkeley - FDG analysis: 2023-02-17 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 7088
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{ROINAME} \emph{character} Region of interest (MetaROI) or reference region (Top50PonsVermis). For methods, see Landau et al., Neurology, 2010 or UCBERKELEY FDG Methods PDF on LONI. , with options: MetaROI, Top50PonsVermis
\item \strong{MEAN} \emph{numeric} Mean of FDG-PET within ROI Range value: 0.713, ..., 1.512
\item \strong{MAX} \emph{numeric} Maximum of FDG-PET within ROI Range value: 0.98, ..., 2.073
\item \strong{STDEV} \emph{numeric} Standard deviation of FDG-PET within ROI Range value: 0.056, ..., 0.445
\item \strong{TOTVOX} \emph{numeric} Total number of voxels within MNI-space ROI (1142 in target region and 1779 in reference region) Range value: 1142, ..., 1779
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
