% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCD_ADNI1_WMH}
\alias{UCD_ADNI1_WMH}
\title{Univerity of California Davis White Matter Hyperintensity Volumes (ADNI1)}
\format{
A data frame with 3469 observations and 11 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCD_ADNI1_WMH)
}
\description{
Univerity of California Davis White Matter Hyperintensity Volumes (ADNI1) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID ADNI code
\item \strong{EXAMDATE} \emph{ } Examination Date MM/DD/YYYY
\item \strong{VISCODE} \emph{character} Visit code , with options: sc, m06, m12, m24, m18, m36, m48, uns1, nv
\item \strong{WHITMATHYP} \emph{numeric} Adjusted WMH Range value: 0, ..., 75.9831
\item \strong{MANUFACTURER} \emph{character} MRI Machine Manufacturer Text field
\item \strong{MODEL} \emph{character} MRI Machine Model Text field
\item \strong{MAGSTRENGTH} \emph{numeric} MRI Machine Magnetic Field Strength Text field
\item \strong{SEGPROCIMG} \emph{character} Images used for segmentation Text field
\item \strong{SEGPROC} \emph{character} Analysis method used for segmentation Text field
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{imaging}
