% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCD_WMH}
\alias{UCD_WMH}
\title{UCD - White Matter Hyperintensity Volumes}
\format{
A data frame with 7437 observations and 28 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCD_WMH)
}
\description{
UCD - White Matter Hyperintensity Volumes data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI2, ADNI3, ADNIGO, ADNI4, ADNI1
\item \strong{PTID} \emph{character} Patient ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10805
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{IMAGEUID} \emph{numeric} LONI image ID Range value: 116847, ..., 11327737
\item \strong{RUNDATE} \emph{ } Date results generated Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} Analysis Status , with options: BAD FLAIR
\item \strong{MANUFACTURER} \emph{character} MRI Scanner Manufacturers , with options: Philips Medical Systems, SIEMENS, GE MEDICAL SYSTEMS, siemens, Philips, Siemens Healthineers, philips, ge, Philips Healthcare, Siemens
\item \strong{MANUFACTURERSMODELNAME} \emph{character} MRI Scanner Model Names  
\item \strong{MAGNETICFIELDSTRENGTH} \emph{numeric} MRI Scanner Magnetic Field Strength (1.5 or 3.0 Tesla) Range value: 3, ..., 3
\item \strong{MRACQUISITIONTYPEFLAIR} \emph{numeric} Flair Acquisition Type (2D or 3D) Range value: 2, ..., 3
\item \strong{CEREBRUM_TCV} \emph{numeric} Total Cerebrum Cranial Volume (supratentorial portion) Range value: 780.078, ..., 1688.14
\item \strong{CEREBRUM_TCB} \emph{numeric} Total Cerebrum Brain Volume Range value: 608.385, ..., 1284.6
\item \strong{CEREBRUM_TCC} \emph{numeric} Total Cerebrum Cerebrospinal Fluid Volume Range value: 139.01, ..., 526.299
\item \strong{CEREBRUM_GRAY} \emph{numeric} Total Cerebrum Gray Matter Volume Range value: 311.958, ..., 723.72
\item \strong{CEREBRUM_WHITE} \emph{numeric} Total Cerebrum White Matter Volume Range value: 203.61, ..., 639.017
\item \strong{LEFT_HIPPO} \emph{numeric} Segmented Left Hippocampus Volume Range value: 1.07321, ..., 4.9068
\item \strong{RIGHT_HIPPO} \emph{numeric} Segmented Right Hippocampus Volume Range value: 1.229, ..., 4.7508
\item \strong{TOTAL_HIPPO} \emph{numeric} Segmented Total Hippocampi Volume Range value: 2.44542, ..., 9.6504
\item \strong{TOTAL_CSF} \emph{numeric} Total Brain Cerebrospinal Fluid Volume Range value: 167.735, ..., 626.027
\item \strong{TOTAL_GRAY} \emph{numeric} Total Brain Gray Matter Volume Range value: 378.749, ..., 862.618
\item \strong{TOTAL_WHITE} \emph{numeric} Total Brain White Matter Volume Range value: 220.306, ..., 702.226
\item \strong{TOTAL_WMH} \emph{numeric} Total White Matter Hyperintensity Volume Range value: 0, ..., 129.837
\item \strong{TOTAL_BRAIN} \emph{numeric} Total Brain Volume (includes both cerebrum and infratentorial region) Range value: 913.145, ..., 1913.59
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{imaging}
