% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSDVOL}
\alias{UCSDVOL}
\title{Anders Dale Lab (UCSD) - Derived Volumes}
\format{
A data frame with 3128 observations and 25 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSDVOL)
}
\description{
Anders Dale Lab (UCSD) - Derived Volumes data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 1435
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character}   , with options: sc, m06, m12, m24, m18, m36
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{LONISID} \emph{numeric} LONISID Range value: 1945, ..., 28453
\item \strong{CDATE} \emph{character} Creation Date , with options: 12/09/2010
\item \strong{QCPASS} \emph{numeric}   Range value: 0, ..., 1
\item \strong{BRAIN} \emph{numeric} Whole Brain. A summary measure of total brain parenchyma, including the Cerebral-Cortex, Cerebellum-Cortex, Thalamus-Proper, CaudatePutamen, Pallidum, Hippocampus, Amygdala, Accumbens-area, VentralDC, Cerebral-White-Matter, Cerebellum-White-Matter, and WM-hypointensities. As such, it does not include ventricles or other CSF spaces, and is not intended to be used as a measure of intracranial volume. Range value: 629016, ..., 1361820
\item \strong{EICV} \emph{numeric} Estimated Intracranial Volume, aka eTIV. To generate the mask, the baseline image is automatically segmented; all thus-defined brain and ventricular voxels are given the value 1 with all other voxels 0. This binary mask is then repeatedly smoothed with a Gaussian kernel to produce a simply connected uniform mask, covering all sulci, whose boundary tapers smoothly from 1 to 0 over the length of a few voxels. Ideally, the mask would end on the skull and include the brain stem down to where it starts to bend with the neck. The smoothing can be controlled fairly nicely to begin tapering at the skull, so that voxels with a mask value less than 1 can  be considered outside the ICV and therefore ignored. Range value: 1059560, ..., 1951440
\item \strong{VENTRICLES} \emph{numeric} Ventricles Range value: 7552.46, ..., 168830
\item \strong{LHIPPOC} \emph{numeric} Left Hippocampus Range value: 1287.93, ..., 5514.94
\item \strong{RHIPPOC} \emph{numeric} Right Hippocampus Range value: 569.549, ..., 5750.4
\item \strong{LINFLATVEN} \emph{numeric} Left inferior lateral ventricle Range value: 232, ..., 8490.65
\item \strong{RINFLATVEN} \emph{numeric} Right inferior lateral ventricle Range value: 136.612, ..., 11628.5
\item \strong{LMIDTEMP} \emph{numeric} Left Middle Temporal Range value: 1.43278, ..., 2.951
\item \strong{RMIDTEMP} \emph{numeric} Right Middle Temporal Range value: 1.40978, ..., 3.1
\item \strong{LINFTEMP} \emph{numeric} Left Inferior Temporal Range value: 1.51919, ..., 3.1769
\item \strong{RINFTEMP} \emph{numeric} Right Inferior Temporal Range value: 1.55386, ..., 3.14129
\item \strong{LFUSIFORM} \emph{numeric} Left Fusiform Range value: 1.50766, ..., 2.928
\item \strong{RFUSIFORM} \emph{numeric} Right Fusiform Range value: 1.396, ..., 2.83648
\item \strong{LENTORHIN} \emph{numeric} Left Entorhinal Range value: 1.45156, ..., 4.08892
\item \strong{RENTORHIN} \emph{numeric} Right Entorhinal Range value: 1.4584, ..., 4.44725
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
