% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFASLFS_V2}
\alias{UCSFASLFS_V2}
\title{ASL Perfusion CBF by FreeSurfer ROI: 2015-11-02}
\format{
A data frame with 675 observations and 702 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFASLFS_V2)
}
\description{
ASL Perfusion CBF by FreeSurfer ROI: 2015-11-02 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI2, ADNIGO
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 5290
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{VERSION} \emph{ }   Date: YYY-MM-DD
\item \strong{LONIUID} \emph{numeric}   Range value: 95820, ..., 233842
\item \strong{IMAGEUID} \emph{numeric}   Range value: 207522, ..., 448375
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{RAWQC} \emph{character} A QC rating of the raw Perfusion Weighted Image , with options: Pass, Fail
\item \strong{ST1MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Brain-Stem Range value: 1, ..., 1
\item \strong{ST1MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Brain-Stem Range value: 247352, ..., 1456570
\item \strong{ST1MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Brain-Stem Range value: 7, ..., 184918
\item \strong{ST1AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Brain-Stem Range value: 11024, ..., 223723
\item \strong{ST1SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Brain-Stem Range value: 31930, ..., 268590
\item \strong{ST1CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Brain-Stem Range value: 11251, ..., 22740
\item \strong{ST2MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: CC_Anterior Range value: 1, ..., 2366
\item \strong{ST2MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: CC_Anterior Range value: 10, ..., 945604
\item \strong{ST2MD} \emph{numeric} Median CBF value for FreeSurfer ROI: CC_Anterior Range value: 5, ..., 33543
\item \strong{ST2AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: CC_Anterior Range value: 5, ..., 148871
\item \strong{ST2SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: CC_Anterior Range value: 2.71044, ..., 198908
\item \strong{ST2CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: CC_Anterior Range value: 66, ..., 630
\item \strong{ST3MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: CC_Central Range value: 1, ..., 7899
\item \strong{ST3MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: CC_Central Range value: 10, ..., 1061320
\item \strong{ST3MD} \emph{numeric} Median CBF value for FreeSurfer ROI: CC_Central Range value: 4, ..., 121940
\item \strong{ST3AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: CC_Central Range value: 4.71, ..., 161881
\item \strong{ST3SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: CC_Central Range value: 2.51386, ..., 228490
\item \strong{ST3CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: CC_Central Range value: 34, ..., 376
\item \strong{ST4MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: CC_Mid_Anterior Range value: 1, ..., 3647
\item \strong{ST4MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: CC_Mid_Anterior Range value: 10, ..., 976652
\item \strong{ST4MD} \emph{numeric} Median CBF value for FreeSurfer ROI: CC_Mid_Anterior Range value: 4, ..., 80297
\item \strong{ST4AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: CC_Mid_Anterior Range value: 4.89796, ..., 181492
\item \strong{ST4SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: CC_Mid_Anterior Range value: 2.58961, ..., 260398
\item \strong{ST4CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: CC_Mid_Anterior Range value: 35, ..., 339
\item \strong{ST5MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: CC_Mid_Posterior Range value: 1, ..., 5052
\item \strong{ST5MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: CC_Mid_Posterior Range value: 10, ..., 1202170
\item \strong{ST5MD} \emph{numeric} Median CBF value for FreeSurfer ROI: CC_Mid_Posterior Range value: 4, ..., 120060
\item \strong{ST5AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: CC_Mid_Posterior Range value: 4.97222, ..., 279778
\item \strong{ST5SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: CC_Mid_Posterior Range value: 2.60849, ..., 304477
\item \strong{ST5CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: CC_Mid_Posterior Range value: 52, ..., 372
\item \strong{ST6MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: CC_Posterior Range value: 1, ..., 5463
\item \strong{ST6MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: CC_Posterior Range value: 10, ..., 1417750
\item \strong{ST6MD} \emph{numeric} Median CBF value for FreeSurfer ROI: CC_Posterior Range value: 5, ..., 118703
\item \strong{ST6AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: CC_Posterior Range value: 5.52143, ..., 221727
\item \strong{ST6SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: CC_Posterior Range value: 2.74967, ..., 347537
\item \strong{ST6CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: CC_Posterior Range value: 74, ..., 765
\item \strong{ST7MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: CSF Range value: 1, ..., 308854
\item \strong{ST7MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: CSF Range value: 105208, ..., 1547790
\item \strong{ST7MD} \emph{numeric} Median CBF value for FreeSurfer ROI: CSF Range value: 7, ..., 737174
\item \strong{ST7AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: CSF Range value: 8648.36, ..., 701037
\item \strong{ST7SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: CSF Range value: 20577.2, ..., 445422
\item \strong{ST7CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: CSF Range value: 185, ..., 1706
\item \strong{ST8MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: 5th-Ventricle Range value: 1, ..., 223470
\item \strong{ST8MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: 5th-Ventricle Range value: 2, ..., 352617
\item \strong{ST8MD} \emph{numeric} Median CBF value for FreeSurfer ROI: 5th-Ventricle Range value: 1, ..., 226392
\item \strong{ST8AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: 5th-Ventricle Range value: 1.5, ..., 223470
\item \strong{ST8SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: 5th-Ventricle Range value: 0, ..., 89052.9
\item \strong{ST8CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: 5th-Ventricle Range value: 1, ..., 35
\item \strong{ST9MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: 4th-Ventricle Range value: 1, ..., 207769
\item \strong{ST9MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: 4th-Ventricle Range value: 185409, ..., 1283530
\item \strong{ST9MD} \emph{numeric} Median CBF value for FreeSurfer ROI: 4th-Ventricle Range value: 6, ..., 618950
\item \strong{ST9AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: 4th-Ventricle Range value: 25800, ..., 636000
\item \strong{ST9SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: 4th-Ventricle Range value: 33207.5, ..., 307298
\item \strong{ST9CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: 4th-Ventricle Range value: 61, ..., 1197
\item \strong{ST11MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: 1, ..., 181276
\item \strong{ST11MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: 109975, ..., 1132500
\item \strong{ST11MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: 8, ..., 530477
\item \strong{ST11AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: 12147.6, ..., 563077
\item \strong{ST11SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: 19790.7, ..., 250568
\item \strong{ST11CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Accumbens-area Range value: 170, ..., 751
\item \strong{ST12MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Amygdala Range value: 1, ..., 94720
\item \strong{ST12MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Amygdala Range value: 143597, ..., 1381520
\item \strong{ST12MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Amygdala Range value: 6112, ..., 522772
\item \strong{ST12AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Amygdala Range value: 26038.8, ..., 513624
\item \strong{ST12SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Amygdala Range value: 25764.6, ..., 318782
\item \strong{ST12CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Amygdala Range value: 395, ..., 1929
\item \strong{ST13MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: 1, ..., 154757
\item \strong{ST13MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: 223449, ..., 1552950
\item \strong{ST13MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: 38879, ..., 920726
\item \strong{ST13AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: 61600.2, ..., 891118
\item \strong{ST13SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: 32060, ..., 345197
\item \strong{ST13CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-bankssts Range value: 872, ..., 3199
\item \strong{ST14MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 1, ..., 193278
\item \strong{ST14MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 132001, ..., 1301040
\item \strong{ST14MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 6, ..., 562226
\item \strong{ST14AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 9607.39, ..., 545305
\item \strong{ST14SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 26258.3, ..., 237913
\item \strong{ST14CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 670, ..., 2772
\item \strong{ST15MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 1, ..., 25261
\item \strong{ST15MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 164339, ..., 1689200
\item \strong{ST15MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 6, ..., 626448
\item \strong{ST15AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 16035.4, ..., 617341
\item \strong{ST15SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 33394.6, ..., 434720
\item \strong{ST15CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 2252, ..., 7869
\item \strong{ST16MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Caudate Range value: 1, ..., 1774
\item \strong{ST16MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Caudate Range value: 101795, ..., 943672
\item \strong{ST16MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Caudate Range value: 6, ..., 414275
\item \strong{ST16AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Caudate Range value: 4446.15, ..., 400331
\item \strong{ST16SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Caudate Range value: 13922.6, ..., 178102
\item \strong{ST16CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Caudate Range value: 1940, ..., 4856
\item \strong{ST17MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Cerebellum-Cortex Range value: 1, ..., 1
\item \strong{ST17MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Cerebellum-Cortex Range value: 319968, ..., 1688290
\item \strong{ST17MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Cerebellum-Cortex Range value: 5, ..., 548159
\item \strong{ST17AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Cerebellum-Cortex Range value: 8179.87, ..., 493052
\item \strong{ST17SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Cerebellum-Cortex Range value: 62350.1, ..., 343190
\item \strong{ST17CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Cerebellum-Cortex Range value: 24620, ..., 48967
\item \strong{ST18MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Cerebellum-White-Matter Range value: 1, ..., 1953
\item \strong{ST18MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Cerebellum-White-Matter Range value: 253436, ..., 1346540
\item \strong{ST18MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Cerebellum-White-Matter Range value: 5, ..., 289088
\item \strong{ST18AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Cerebellum-White-Matter Range value: 1565.92, ..., 245499
\item \strong{ST18SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Cerebellum-White-Matter Range value: 22901.9, ..., 235575
\item \strong{ST18CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Cerebellum-White-Matter Range value: 5627, ..., 18865
\item \strong{ST20MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Cerebral-White-Matter Range value: 1, ..., 1
\item \strong{ST20MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Cerebral-White-Matter Range value: 305548, ..., 1688490
\item \strong{ST20MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Cerebral-White-Matter Range value: 784, ..., 97492.5
\item \strong{ST20AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Cerebral-White-Matter Range value: 18514.4, ..., 189588
\item \strong{ST20SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Cerebral-White-Matter Range value: 33543.3, ..., 220275
\item \strong{ST20CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Cerebral-White-Matter Range value: 72714, ..., 188444
\item \strong{ST21MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-choroid-plexus Range value: 1, ..., 4372
\item \strong{ST21MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-choroid-plexus Range value: 181858, ..., 1203830
\item \strong{ST21MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-choroid-plexus Range value: 7, ..., 584468
\item \strong{ST21AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-choroid-plexus Range value: 18451.6, ..., 501642
\item \strong{ST21SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-choroid-plexus Range value: 40428.5, ..., 296852
\item \strong{ST21CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-choroid-plexus Range value: 405, ..., 1774
\item \strong{ST22MIN} \emph{logical} Min CBF value for FreeSurfer ROI: ctx-lh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22MAX} \emph{logical} Max CBF value for FreeSurfer ROI: ctx-lh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22MD} \emph{logical} Median CBF value for FreeSurfer ROI: ctx-lh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22AVG} \emph{logical} Mean CBF value for FreeSurfer ROI: ctx-lh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22SD} \emph{logical} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22CT} \emph{logical} Count of CBF voxels in FreeSurfer ROI: ctx-lh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST23MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: 1, ..., 238959
\item \strong{ST23MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: 210121, ..., 1424620
\item \strong{ST23MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: 29268, ..., 724374
\item \strong{ST23AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: 39246.7, ..., 710901
\item \strong{ST23SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: 37988.8, ..., 365095
\item \strong{ST23CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-cuneus Range value: 1045, ..., 3498
\item \strong{ST24MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: 1, ..., 104369
\item \strong{ST24MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: 178164, ..., 1684840
\item \strong{ST24MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: 6, ..., 534753
\item \strong{ST24AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: 23356.1, ..., 523066
\item \strong{ST24SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: 44039.3, ..., 444789
\item \strong{ST24CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-entorhinal Range value: 489, ..., 2459
\item \strong{ST25MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: 1, ..., 202972
\item \strong{ST25MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: 63739, ..., 1301210
\item \strong{ST25MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: 6, ..., 695453
\item \strong{ST25AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: 1708.5, ..., 696147
\item \strong{ST25SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: 8717.02, ..., 228050
\item \strong{ST25CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-frontalpole Range value: 325, ..., 941
\item \strong{ST26MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: 1, ..., 68763
\item \strong{ST26MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: 248101, ..., 1688750
\item \strong{ST26MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: 6, ..., 567076
\item \strong{ST26AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: 76430.6, ..., 567038
\item \strong{ST26SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: 43087.5, ..., 345195
\item \strong{ST26CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-fusiform Range value: 3420, ..., 10424
\item \strong{ST29MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Hippocampus Range value: 1, ..., 10374
\item \strong{ST29MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Hippocampus Range value: 188709, ..., 1682730
\item \strong{ST29MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Hippocampus Range value: 13304, ..., 530690
\item \strong{ST29AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Hippocampus Range value: 39353.3, ..., 515812
\item \strong{ST29SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Hippocampus Range value: 40350, ..., 276272
\item \strong{ST29CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Hippocampus Range value: 1311, ..., 4529
\item \strong{ST30MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Inf-Lat-Vent Range value: 1, ..., 181765
\item \strong{ST30MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Inf-Lat-Vent Range value: 136687, ..., 1215610
\item \strong{ST30MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Inf-Lat-Vent Range value: 7, ..., 602544
\item \strong{ST30AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Inf-Lat-Vent Range value: 36955.8, ..., 573791
\item \strong{ST30SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Inf-Lat-Vent Range value: 33785.7, ..., 269000
\item \strong{ST30CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Inf-Lat-Vent Range value: 21, ..., 1826
\item \strong{ST31MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 1, ..., 33985
\item \strong{ST31MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 268449, ..., 1556630
\item \strong{ST31MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 8, ..., 644589
\item \strong{ST31AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 41756.6, ..., 649920
\item \strong{ST31SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 41953.1, ..., 329068
\item \strong{ST31CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 4989, ..., 12687
\item \strong{ST32MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 1, ..., 5764
\item \strong{ST32MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 317991, ..., 1689140
\item \strong{ST32MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 7, ..., 531378
\item \strong{ST32AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 36595.4, ..., 533460
\item \strong{ST32SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 62855.8, ..., 366098
\item \strong{ST32CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 4001, ..., 11480
\item \strong{ST34MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 1, ..., 55662
\item \strong{ST34MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 154476, ..., 1430930
\item \strong{ST34MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 5, ..., 879898
\item \strong{ST34AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 2793.15, ..., 858320
\item \strong{ST34SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 19920.1, ..., 302884
\item \strong{ST34CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 915, ..., 2789
\item \strong{ST35MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 1, ..., 15665
\item \strong{ST35MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 270485, ..., 1683840
\item \strong{ST35MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 6, ..., 622300
\item \strong{ST35AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 15721.3, ..., 624607
\item \strong{ST35SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 54490.9, ..., 297464
\item \strong{ST35CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 4665, ..., 11827
\item \strong{ST36MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 1, ..., 14978
\item \strong{ST36MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 234895, ..., 1512550
\item \strong{ST36MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 41183, ..., 525190
\item \strong{ST36AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 54493, ..., 534321
\item \strong{ST36SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 39497.9, ..., 232561
\item \strong{ST36CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 3299, ..., 7304
\item \strong{ST37MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Lateral-Ventricle Range value: 1, ..., 12242
\item \strong{ST37MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Lateral-Ventricle Range value: 145290, ..., 947548
\item \strong{ST37MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Lateral-Ventricle Range value: 7, ..., 285760
\item \strong{ST37AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Lateral-Ventricle Range value: 11525.1, ..., 270168
\item \strong{ST37SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Lateral-Ventricle Range value: 26872.1, ..., 189074
\item \strong{ST37CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Lateral-Ventricle Range value: 271, ..., 18366
\item \strong{ST38MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: 1, ..., 167416
\item \strong{ST38MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: 199555, ..., 1653050
\item \strong{ST38MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: 7, ..., 764647
\item \strong{ST38AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: 72947.5, ..., 757371
\item \strong{ST38SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: 30048.4, ..., 363116
\item \strong{ST38CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-lingual Range value: 2782, ..., 7086
\item \strong{ST39MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 1, ..., 47245
\item \strong{ST39MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 226408, ..., 1543600
\item \strong{ST39MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 9, ..., 532686
\item \strong{ST39AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 53309.1, ..., 518811
\item \strong{ST39SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 42946.7, ..., 242617
\item \strong{ST39CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 2301, ..., 5597
\item \strong{ST40MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: 1, ..., 13775
\item \strong{ST40MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: 306350, ..., 1689120
\item \strong{ST40MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: 35222, ..., 555170
\item \strong{ST40AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: 85995.3, ..., 559806
\item \strong{ST40SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: 50785.2, ..., 352951
\item \strong{ST40CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-middletemporal Range value: 3722, ..., 11532
\item \strong{ST42MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Pallidum Range value: 1, ..., 2716
\item \strong{ST42MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Pallidum Range value: 66599, ..., 1276220
\item \strong{ST42MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Pallidum Range value: 5, ..., 383427
\item \strong{ST42AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Pallidum Range value: 405.608, ..., 345000
\item \strong{ST42SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Pallidum Range value: 3343.93, ..., 201642
\item \strong{ST42CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Pallidum Range value: 176, ..., 1251
\item \strong{ST43MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: 1, ..., 83506
\item \strong{ST43MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: 118307, ..., 1476380
\item \strong{ST43MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: 5, ..., 641499
\item \strong{ST43AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: 2328.39, ..., 621221
\item \strong{ST43SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: 11875.6, ..., 324099
\item \strong{ST43CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-paracentral Range value: 1431, ..., 4086
\item \strong{ST44MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: 1, ..., 157977
\item \strong{ST44MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: 105681, ..., 1685050
\item \strong{ST44MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: 10, ..., 662992
\item \strong{ST44AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: 11940.6, ..., 671518
\item \strong{ST44SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: 18082.9, ..., 340792
\item \strong{ST44CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-parahippocampal Range value: 812, ..., 2481
\item \strong{ST45MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: 1, ..., 47280
\item \strong{ST45MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: 172379, ..., 1686110
\item \strong{ST45MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: 7, ..., 598989
\item \strong{ST45AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: 11502.5, ..., 572512
\item \strong{ST45SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: 27284.1, ..., 271510
\item \strong{ST45CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-parsopercularis Range value: 1945, ..., 5238
\item \strong{ST46MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 1, ..., 150465
\item \strong{ST46MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 183970, ..., 1438280
\item \strong{ST46MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 7, ..., 647431
\item \strong{ST46AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 50914.7, ..., 655232
\item \strong{ST46SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 27160.5, ..., 280244
\item \strong{ST46CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 837, ..., 2089
\item \strong{ST47MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: 1, ..., 114513
\item \strong{ST47MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: 182932, ..., 1667700
\item \strong{ST47MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: 8, ..., 655342
\item \strong{ST47AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: 22746.4, ..., 655639
\item \strong{ST47SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: 33547.6, ..., 306099
\item \strong{ST47CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-parstriangularis Range value: 1534, ..., 3422
\item \strong{ST48MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: 1, ..., 228387
\item \strong{ST48MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: 223225, ..., 1328980
\item \strong{ST48MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: 7, ..., 813155
\item \strong{ST48AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: 115421, ..., 781379
\item \strong{ST48SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: 41664.9, ..., 322098
\item \strong{ST48CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-pericalcarine Range value: 636, ..., 2379
\item \strong{ST49MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: 1, ..., 15884
\item \strong{ST49MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: 249821, ..., 1689980
\item \strong{ST49MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: 9, ..., 612160
\item \strong{ST49AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: 28108.3, ..., 635013
\item \strong{ST49SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: 45990.5, ..., 378925
\item \strong{ST49CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-postcentral Range value: 4332, ..., 11907
\item \strong{ST50MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 1, ..., 77928
\item \strong{ST50MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 171272, ..., 1478120
\item \strong{ST50MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 7, ..., 695538
\item \strong{ST50AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 17302.7, ..., 686365
\item \strong{ST50SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 34650.8, ..., 276795
\item \strong{ST50CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 1302, ..., 3402
\item \strong{ST51MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: 1, ..., 6946
\item \strong{ST51MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: 264793, ..., 1689650
\item \strong{ST51MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: 9, ..., 574469
\item \strong{ST51AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: 66560.3, ..., 557113
\item \strong{ST51SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: 44952.4, ..., 378292
\item \strong{ST51CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-precentral Range value: 5663, ..., 15250
\item \strong{ST52MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: 1, ..., 45555
\item \strong{ST52MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: 132615, ..., 1426380
\item \strong{ST52MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: 9, ..., 716917
\item \strong{ST52AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: 33538.7, ..., 700420
\item \strong{ST52SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: 27034, ..., 234409
\item \strong{ST52CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-precuneus Range value: 4309, ..., 9703
\item \strong{ST53MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Putamen Range value: 1, ..., 2798
\item \strong{ST53MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Putamen Range value: 187877, ..., 1231980
\item \strong{ST53MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Putamen Range value: 4529, ..., 460688
\item \strong{ST53AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Putamen Range value: 24367.9, ..., 427020
\item \strong{ST53SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Putamen Range value: 33907.8, ..., 247977
\item \strong{ST53CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Putamen Range value: 2470, ..., 5832
\item \strong{ST54MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 1, ..., 88024
\item \strong{ST54MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 135383, ..., 1420350
\item \strong{ST54MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 6, ..., 563483
\item \strong{ST54AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 6721, ..., 557226
\item \strong{ST54SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 25579.2, ..., 302815
\item \strong{ST54CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 1024, ..., 3254
\item \strong{ST55MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 1, ..., 14848
\item \strong{ST55MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 252573, ..., 1688600
\item \strong{ST55MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 9, ..., 651651
\item \strong{ST55AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 18201.5, ..., 631020
\item \strong{ST55SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 29079.4, ..., 454438
\item \strong{ST55CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 6103, ..., 17898
\item \strong{ST56MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 1, ..., 8702
\item \strong{ST56MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 227751, ..., 1689660
\item \strong{ST56MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 5, ..., 568256
\item \strong{ST56AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 7128.58, ..., 551849
\item \strong{ST56SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 37908.4, ..., 290222
\item \strong{ST56CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 9430, ..., 21063
\item \strong{ST57MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: 1, ..., 15190
\item \strong{ST57MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: 274978, ..., 1689590
\item \strong{ST57MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: 6, ..., 530957
\item \strong{ST57AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: 11799.5, ..., 540226
\item \strong{ST57SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: 36917.9, ..., 327608
\item \strong{ST57CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiorparietal Range value: 5153, ..., 12423
\item \strong{ST58MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: 1, ..., 35921
\item \strong{ST58MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: 227231, ..., 1534150
\item \strong{ST58MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: 72426, ..., 605077
\item \strong{ST58AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: 78580.7, ..., 596460
\item \strong{ST58SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: 42041.5, ..., 312506
\item \strong{ST58CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiortemporal Range value: 4731, ..., 11186
\item \strong{ST59MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: 1, ..., 58481
\item \strong{ST59MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: 199059, ..., 1660050
\item \strong{ST59MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: 7, ..., 675222
\item \strong{ST59AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: 20327.4, ..., 691003
\item \strong{ST59SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: 39941.5, ..., 314955
\item \strong{ST59CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-supramarginal Range value: 4806, ..., 11445
\item \strong{ST60MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: 1, ..., 112117
\item \strong{ST60MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: 177763, ..., 1676190
\item \strong{ST60MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: 6, ..., 740758
\item \strong{ST60AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: 7547.29, ..., 730599
\item \strong{ST60SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: 21751.5, ..., 363725
\item \strong{ST60CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-temporalpole Range value: 937, ..., 2847
\item \strong{ST61MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: 1, ..., 2593
\item \strong{ST61MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: 148363, ..., 1458220
\item \strong{ST61MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: 6, ..., 502338
\item \strong{ST61AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: 5781.67, ..., 452562
\item \strong{ST61SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: 14017.8, ..., 323965
\item \strong{ST61CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Thalamus-Proper Range value: 3702, ..., 6585
\item \strong{ST62MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: 1, ..., 226045
\item \strong{ST62MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: 215687, ..., 1442440
\item \strong{ST62MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: 86426, ..., 808712
\item \strong{ST62AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: 90869.6, ..., 801106
\item \strong{ST62SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: 30771.1, ..., 273535
\item \strong{ST62CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-transversetemporal Range value: 436, ..., 1430
\item \strong{ST64MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-unknown Range value: 1, ..., 3569
\item \strong{ST64MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-unknown Range value: 182352, ..., 1677970
\item \strong{ST64MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-unknown Range value: 30926, ..., 437797
\item \strong{ST64AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-unknown Range value: 46268.3, ..., 447409
\item \strong{ST64SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-unknown Range value: 41342.6, ..., 331175
\item \strong{ST64CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-unknown Range value: 399, ..., 1597
\item \strong{ST65MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-VentralDC Range value: 1, ..., 3196
\item \strong{ST65MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-VentralDC Range value: 152699, ..., 1395680
\item \strong{ST65MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-VentralDC Range value: 7, ..., 70907
\item \strong{ST65AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-VentralDC Range value: 7634.46, ..., 173578
\item \strong{ST65SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-VentralDC Range value: 17737.7, ..., 219561
\item \strong{ST65CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-VentralDC Range value: 1331, ..., 3608
\item \strong{ST66MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-vessel Range value: 1, ..., 477217
\item \strong{ST66MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-vessel Range value: 51656, ..., 1125940
\item \strong{ST66MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Left-vessel Range value: 8, ..., 705908
\item \strong{ST66AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-vessel Range value: 9207.21, ..., 697100
\item \strong{ST66SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-vessel Range value: 10425.5, ..., 207812
\item \strong{ST66CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-vessel Range value: 12, ..., 307
\item \strong{ST68MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: non-WM-hypointensities Range value: 1, ..., 93746
\item \strong{ST68MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: non-WM-hypointensities Range value: 10, ..., 869971
\item \strong{ST68MD} \emph{numeric} Median CBF value for FreeSurfer ROI: non-WM-hypointensities Range value: 4, ..., 437988
\item \strong{ST68AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: non-WM-hypointensities Range value: 5.86667, ..., 410239
\item \strong{ST68SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: non-WM-hypointensities Range value: 0, ..., 249031
\item \strong{ST68CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: non-WM-hypointensities Range value: 1, ..., 314
\item \strong{ST69MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Optic-Chiasm Range value: 1, ..., 36109
\item \strong{ST69MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Optic-Chiasm Range value: 10, ..., 1008810
\item \strong{ST69MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Optic-Chiasm Range value: 5, ..., 514860
\item \strong{ST69AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Optic-Chiasm Range value: 5.41463, ..., 457224
\item \strong{ST69SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Optic-Chiasm Range value: 2.67326, ..., 327610
\item \strong{ST69CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Optic-Chiasm Range value: 72, ..., 353
\item \strong{ST70MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: 1, ..., 71775
\item \strong{ST70MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: 87677, ..., 981525
\item \strong{ST70MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: 7, ..., 516540
\item \strong{ST70AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: 4695.98, ..., 494562
\item \strong{ST70SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: 12949.3, ..., 260252
\item \strong{ST70CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Accumbens-area Range value: 132, ..., 758
\item \strong{ST71MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Amygdala Range value: 1, ..., 58957
\item \strong{ST71MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Amygdala Range value: 170092, ..., 1105690
\item \strong{ST71MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Amygdala Range value: 2788, ..., 551073
\item \strong{ST71AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Amygdala Range value: 32757.2, ..., 554888
\item \strong{ST71SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Amygdala Range value: 26966.5, ..., 234898
\item \strong{ST71CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Amygdala Range value: 453, ..., 1997
\item \strong{ST72MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: 1, ..., 142098
\item \strong{ST72MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: 214902, ..., 1559030
\item \strong{ST72MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: 90007, ..., 788205
\item \strong{ST72AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: 92240.7, ..., 776868
\item \strong{ST72SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: 29813.4, ..., 353014
\item \strong{ST72CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-bankssts Range value: 714, ..., 2759
\item \strong{ST73MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 1, ..., 83558
\item \strong{ST73MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 90951, ..., 1471850
\item \strong{ST73MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 6, ..., 508464
\item \strong{ST73AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 1289.26, ..., 508727
\item \strong{ST73SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 7378.52, ..., 319548
\item \strong{ST73CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 798, ..., 3007
\item \strong{ST74MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 1, ..., 19195
\item \strong{ST74MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 201750, ..., 1680890
\item \strong{ST74MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 5, ..., 619768
\item \strong{ST74AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 11895.6, ..., 614950
\item \strong{ST74SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 30679.4, ..., 341222
\item \strong{ST74CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 1814, ..., 7315
\item \strong{ST75MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Caudate Range value: 1, ..., 2491
\item \strong{ST75MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Caudate Range value: 139788, ..., 943152
\item \strong{ST75MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Caudate Range value: 9, ..., 486418
\item \strong{ST75AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Caudate Range value: 15637.6, ..., 471730
\item \strong{ST75SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Caudate Range value: 25052.2, ..., 163195
\item \strong{ST75CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Caudate Range value: 1964, ..., 4857
\item \strong{ST76MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Cerebellum-Cortex Range value: 1, ..., 1
\item \strong{ST76MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Cerebellum-Cortex Range value: 319364, ..., 1457810
\item \strong{ST76MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Cerebellum-Cortex Range value: 5, ..., 508727
\item \strong{ST76AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Cerebellum-Cortex Range value: 6534.82, ..., 498593
\item \strong{ST76SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Cerebellum-Cortex Range value: 54091.4, ..., 317658
\item \strong{ST76CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Cerebellum-Cortex Range value: 25352, ..., 49712
\item \strong{ST77MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Cerebellum-White-Matter Range value: 1, ..., 1897
\item \strong{ST77MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Cerebellum-White-Matter Range value: 236181, ..., 1228340
\item \strong{ST77MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Cerebellum-White-Matter Range value: 5, ..., 135098
\item \strong{ST77AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Cerebellum-White-Matter Range value: 331.522, ..., 208081
\item \strong{ST77SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Cerebellum-White-Matter Range value: 10675.1, ..., 227987
\item \strong{ST77CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Cerebellum-White-Matter Range value: 5370, ..., 19595
\item \strong{ST79MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Cerebral-White-Matter Range value: 1, ..., 1
\item \strong{ST79MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Cerebral-White-Matter Range value: 310993, ..., 1689240
\item \strong{ST79MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Cerebral-White-Matter Range value: 2388, ..., 86607.5
\item \strong{ST79AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Cerebral-White-Matter Range value: 22024.7, ..., 164158
\item \strong{ST79SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Cerebral-White-Matter Range value: 37139.8, ..., 205784
\item \strong{ST79CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Cerebral-White-Matter Range value: 72164, ..., 185529
\item \strong{ST80MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-choroid-plexus Range value: 1, ..., 9017
\item \strong{ST80MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-choroid-plexus Range value: 133829, ..., 1467570
\item \strong{ST80MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-choroid-plexus Range value: 8, ..., 513412
\item \strong{ST80AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-choroid-plexus Range value: 19030.6, ..., 441618
\item \strong{ST80SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-choroid-plexus Range value: 37048.5, ..., 285538
\item \strong{ST80CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-choroid-plexus Range value: 529, ..., 1903
\item \strong{ST81MIN} \emph{logical} Min CBF value for FreeSurfer ROI: ctx-rh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81MAX} \emph{logical} Max CBF value for FreeSurfer ROI: ctx-rh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81MD} \emph{logical} Median CBF value for FreeSurfer ROI: ctx-rh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81AVG} \emph{logical} Mean CBF value for FreeSurfer ROI: ctx-rh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81SD} \emph{logical} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81CT} \emph{logical} Count of CBF voxels in FreeSurfer ROI: ctx-rh-corpuscallosum A Boolean value: TRUE/ FALSE
\item \strong{ST82MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: 1, ..., 166819
\item \strong{ST82MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: 198449, ..., 1443440
\item \strong{ST82MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: 52087, ..., 781318
\item \strong{ST82AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: 59282.5, ..., 774636
\item \strong{ST82SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: 35199.3, ..., 284089
\item \strong{ST82CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-cuneus Range value: 894, ..., 3403
\item \strong{ST83MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: 1, ..., 150519
\item \strong{ST83MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: 214925, ..., 1680370
\item \strong{ST83MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: 6, ..., 738087
\item \strong{ST83AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: 9600.44, ..., 728826
\item \strong{ST83SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: 36810.3, ..., 449440
\item \strong{ST83CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-entorhinal Range value: 309, ..., 2823
\item \strong{ST84MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: 1, ..., 143877
\item \strong{ST84MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: 69220, ..., 1665280
\item \strong{ST84MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: 6, ..., 721657
\item \strong{ST84AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: 3960.46, ..., 698404
\item \strong{ST84SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: 14494.5, ..., 302977
\item \strong{ST84CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-frontalpole Range value: 434, ..., 1153
\item \strong{ST85MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: 1, ..., 15841
\item \strong{ST85MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: 319248, ..., 1688970
\item \strong{ST85MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: 5, ..., 572729
\item \strong{ST85AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: 8920.71, ..., 559966
\item \strong{ST85SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: 50816.1, ..., 340743
\item \strong{ST85CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-fusiform Range value: 3625, ..., 13573
\item \strong{ST88MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Hippocampus Range value: 1, ..., 8167
\item \strong{ST88MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Hippocampus Range value: 193468, ..., 1257120
\item \strong{ST88MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Hippocampus Range value: 34991, ..., 509002
\item \strong{ST88AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Hippocampus Range value: 56170.4, ..., 488099
\item \strong{ST88SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Hippocampus Range value: 40249.3, ..., 219600
\item \strong{ST88CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Hippocampus Range value: 1219, ..., 4557
\item \strong{ST89MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Inf-Lat-Vent Range value: 1, ..., 225673
\item \strong{ST89MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Inf-Lat-Vent Range value: 108165, ..., 1056920
\item \strong{ST89MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Inf-Lat-Vent Range value: 5, ..., 516428
\item \strong{ST89AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Inf-Lat-Vent Range value: 8833.63, ..., 485177
\item \strong{ST89SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Inf-Lat-Vent Range value: 23369.1, ..., 230303
\item \strong{ST89CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Inf-Lat-Vent Range value: 13, ..., 1577
\item \strong{ST90MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 1, ..., 16344
\item \strong{ST90MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 245653, ..., 1552640
\item \strong{ST90MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 21471, ..., 740860
\item \strong{ST90AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 72999.7, ..., 730407
\item \strong{ST90SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 38745.8, ..., 311965
\item \strong{ST90CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 5081, ..., 16900
\item \strong{ST91MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 1, ..., 2970
\item \strong{ST91MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 319823, ..., 1689970
\item \strong{ST91MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 6, ..., 530109
\item \strong{ST91AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 10645.1, ..., 536092
\item \strong{ST91SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 42295.1, ..., 467363
\item \strong{ST91CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 4058, ..., 11523
\item \strong{ST93MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 1, ..., 44662
\item \strong{ST93MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 75074, ..., 1474550
\item \strong{ST93MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 6, ..., 940047
\item \strong{ST93AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 3456.83, ..., 901525
\item \strong{ST93SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 11801.6, ..., 305566
\item \strong{ST93CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 713, ..., 2835
\item \strong{ST94MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 1, ..., 17833
\item \strong{ST94MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 317374, ..., 1633390
\item \strong{ST94MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 6, ..., 585329
\item \strong{ST94AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 20505.1, ..., 568502
\item \strong{ST94SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 51248.2, ..., 302442
\item \strong{ST94CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 4885, ..., 13253
\item \strong{ST95MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 1, ..., 15173
\item \strong{ST95MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 267798, ..., 1654760
\item \strong{ST95MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 27621, ..., 517806
\item \strong{ST95AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 47824, ..., 512439
\item \strong{ST95SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 43453.9, ..., 223762
\item \strong{ST95CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 3748, ..., 6979
\item \strong{ST96MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Lateral-Ventricle Range value: 1, ..., 3003
\item \strong{ST96MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Lateral-Ventricle Range value: 132580, ..., 969750
\item \strong{ST96MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Lateral-Ventricle Range value: 8, ..., 270296
\item \strong{ST96AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Lateral-Ventricle Range value: 13189, ..., 306018
\item \strong{ST96SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Lateral-Ventricle Range value: 24806.2, ..., 222905
\item \strong{ST96CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Lateral-Ventricle Range value: 324, ..., 11836
\item \strong{ST97MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: 1, ..., 113532
\item \strong{ST97MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: 209584, ..., 1313300
\item \strong{ST97MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: 7, ..., 694917
\item \strong{ST97AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: 48177.6, ..., 698688
\item \strong{ST97SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: 34303.6, ..., 343385
\item \strong{ST97CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-lingual Range value: 3209, ..., 7349
\item \strong{ST98MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 1, ..., 54209
\item \strong{ST98MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 234630, ..., 1683640
\item \strong{ST98MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 8, ..., 535703
\item \strong{ST98AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 35906.7, ..., 515235
\item \strong{ST98SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 47123.1, ..., 380539
\item \strong{ST98CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 2631, ..., 5696
\item \strong{ST99MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: 1, ..., 30610
\item \strong{ST99MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: 305939, ..., 1679800
\item \strong{ST99MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: 97270, ..., 667536
\item \strong{ST99AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: 107269, ..., 686553
\item \strong{ST99SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: 50284.9, ..., 339565
\item \strong{ST99CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-middletemporal Range value: 4150, ..., 12923
\item \strong{ST101MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Pallidum Range value: 1, ..., 4482
\item \strong{ST101MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Pallidum Range value: 84890, ..., 754596
\item \strong{ST101MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Pallidum Range value: 7, ..., 431556
\item \strong{ST101AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Pallidum Range value: 1804.43, ..., 361694
\item \strong{ST101SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Pallidum Range value: 9834.2, ..., 226361
\item \strong{ST101CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Pallidum Range value: 155, ..., 1208
\item \strong{ST102MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: 1, ..., 25271
\item \strong{ST102MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: 165570, ..., 1678130
\item \strong{ST102MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: 6, ..., 655556
\item \strong{ST102AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: 8207.92, ..., 641301
\item \strong{ST102SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: 25962.7, ..., 228796
\item \strong{ST102CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-paracentral Range value: 1719, ..., 4452
\item \strong{ST103MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: 1, ..., 74169
\item \strong{ST103MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: 67740, ..., 1556290
\item \strong{ST103MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: 7, ..., 618295
\item \strong{ST103AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: 4720.23, ..., 596140
\item \strong{ST103SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: 11178, ..., 307429
\item \strong{ST103CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-parahippocampal Range value: 867, ..., 2501
\item \strong{ST104MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: 1, ..., 51069
\item \strong{ST104MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: 154426, ..., 1476500
\item \strong{ST104MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: 9, ..., 689385
\item \strong{ST104AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: 43914.2, ..., 678451
\item \strong{ST104SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: 35582.1, ..., 227349
\item \strong{ST104CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-parsopercularis Range value: 1578, ..., 4538
\item \strong{ST105MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 1, ..., 124963
\item \strong{ST105MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 195796, ..., 1259920
\item \strong{ST105MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 6, ..., 585090
\item \strong{ST105AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 30566.5, ..., 588300
\item \strong{ST105SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 38757.9, ..., 257426
\item \strong{ST105CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 1182, ..., 2692
\item \strong{ST106MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: 1, ..., 49833
\item \strong{ST106MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: 214322, ..., 1308570
\item \strong{ST106MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: 9, ..., 586091
\item \strong{ST106AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: 47321.4, ..., 572147
\item \strong{ST106SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: 35522.4, ..., 263359
\item \strong{ST106CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-parstriangularis Range value: 1570, ..., 4566
\item \strong{ST107MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: 1, ..., 205929
\item \strong{ST107MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: 226671, ..., 1257660
\item \strong{ST107MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: 8, ..., 792511
\item \strong{ST107AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: 89475.3, ..., 764038
\item \strong{ST107SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: 39134.2, ..., 351933
\item \strong{ST107CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-pericalcarine Range value: 643, ..., 2802
\item \strong{ST108MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: 1, ..., 10068
\item \strong{ST108MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: 232209, ..., 1689460
\item \strong{ST108MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: 10, ..., 595872
\item \strong{ST108AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: 50827.4, ..., 568722
\item \strong{ST108SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: 38725.4, ..., 321616
\item \strong{ST108CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-postcentral Range value: 3732, ..., 9749
\item \strong{ST109MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 1, ..., 73730
\item \strong{ST109MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 209318, ..., 1457300
\item \strong{ST109MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 7, ..., 699092
\item \strong{ST109AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 17282.6, ..., 668770
\item \strong{ST109SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 36642.7, ..., 264478
\item \strong{ST109CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 1418, ..., 3684
\item \strong{ST110MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: 1, ..., 6703
\item \strong{ST110MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: 218303, ..., 1686370
\item \strong{ST110MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: 8, ..., 595589
\item \strong{ST110AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: 54690.1, ..., 555361
\item \strong{ST110SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: 40095.6, ..., 280355
\item \strong{ST110CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-precentral Range value: 5833, ..., 15212
\item \strong{ST111MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: 1, ..., 47321
\item \strong{ST111MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: 227647, ..., 1387280
\item \strong{ST111MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: 1831, ..., 669212
\item \strong{ST111AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: 41411.8, ..., 650149
\item \strong{ST111SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: 37337.6, ..., 259902
\item \strong{ST111CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-precuneus Range value: 4203, ..., 10238
\item \strong{ST112MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Putamen Range value: 1, ..., 3393
\item \strong{ST112MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Putamen Range value: 205845, ..., 1171160
\item \strong{ST112MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Putamen Range value: 23577.5, ..., 477002
\item \strong{ST112AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Putamen Range value: 49871.4, ..., 429708
\item \strong{ST112SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Putamen Range value: 44735.7, ..., 244427
\item \strong{ST112CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Putamen Range value: 2415, ..., 5624
\item \strong{ST113MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 1, ..., 63162
\item \strong{ST113MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 91336, ..., 1410580
\item \strong{ST113MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 5, ..., 564188
\item \strong{ST113AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 1056.53, ..., 541153
\item \strong{ST113SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 7242.15, ..., 283960
\item \strong{ST113CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 880, ..., 2903
\item \strong{ST114MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 1, ..., 21142
\item \strong{ST114MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 206465, ..., 1686220
\item \strong{ST114MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 8, ..., 671024
\item \strong{ST114AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 22590.9, ..., 638720
\item \strong{ST114SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 31773.5, ..., 300923
\item \strong{ST114CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 6453, ..., 17319
\item \strong{ST115MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 1, ..., 7719
\item \strong{ST115MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 221771, ..., 1655640
\item \strong{ST115MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 6, ..., 501637
\item \strong{ST115AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 9846.97, ..., 536426
\item \strong{ST115SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 39797.8, ..., 310783
\item \strong{ST115CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 8860, ..., 20837
\item \strong{ST116MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: 1, ..., 7949
\item \strong{ST116MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: 283421, ..., 1682120
\item \strong{ST116MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: 7, ..., 522668
\item \strong{ST116AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: 18732.9, ..., 537430
\item \strong{ST116SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: 47370.3, ..., 264438
\item \strong{ST116CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiorparietal Range value: 5160, ..., 13255
\item \strong{ST117MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: 1, ..., 36272
\item \strong{ST117MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: 232226, ..., 1557910
\item \strong{ST117MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: 75354.5, ..., 626457
\item \strong{ST117AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: 93291.1, ..., 635921
\item \strong{ST117SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: 42386, ..., 293786
\item \strong{ST117CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiortemporal Range value: 5117, ..., 11119
\item \strong{ST118MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: 1, ..., 33865
\item \strong{ST118MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: 211597, ..., 1556820
\item \strong{ST118MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: 55625.5, ..., 735666
\item \strong{ST118AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: 62261.8, ..., 732730
\item \strong{ST118SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: 37314.1, ..., 331403
\item \strong{ST118CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-supramarginal Range value: 3952, ..., 10690
\item \strong{ST119MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: 1, ..., 69457
\item \strong{ST119MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: 138378, ..., 1548620
\item \strong{ST119MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: 6, ..., 606830
\item \strong{ST119AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: 1423.9, ..., 651187
\item \strong{ST119SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: 10090.7, ..., 471102
\item \strong{ST119CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-temporalpole Range value: 368, ..., 2611
\item \strong{ST120MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: 1, ..., 2612
\item \strong{ST120MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: 112338, ..., 1559830
\item \strong{ST120MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: 7, ..., 517984
\item \strong{ST120AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: 5131.28, ..., 464852
\item \strong{ST120SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: 15553.8, ..., 304893
\item \strong{ST120CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Thalamus-Proper Range value: 3694, ..., 6708
\item \strong{ST121MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: 1, ..., 169311
\item \strong{ST121MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: 170839, ..., 1476150
\item \strong{ST121MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: 79027, ..., 893340
\item \strong{ST121AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: 84398.8, ..., 863160
\item \strong{ST121SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: 30276.7, ..., 324019
\item \strong{ST121CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-transversetemporal Range value: 362, ..., 1082
\item \strong{ST123MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-unknown Range value: 1, ..., 2853
\item \strong{ST123MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-unknown Range value: 187363, ..., 1514540
\item \strong{ST123MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-unknown Range value: 2777, ..., 528198
\item \strong{ST123AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-unknown Range value: 44508.4, ..., 536318
\item \strong{ST123SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-unknown Range value: 45843.6, ..., 313308
\item \strong{ST123CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-unknown Range value: 348, ..., 1641
\item \strong{ST124MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-VentralDC Range value: 1, ..., 2843
\item \strong{ST124MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-VentralDC Range value: 148069, ..., 1118120
\item \strong{ST124MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-VentralDC Range value: 8, ..., 91524
\item \strong{ST124AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-VentralDC Range value: 8753.9, ..., 177095
\item \strong{ST124SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-VentralDC Range value: 24300.7, ..., 210573
\item \strong{ST124CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-VentralDC Range value: 1520, ..., 3502
\item \strong{ST125MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-vessel Range value: 1, ..., 480376
\item \strong{ST125MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-vessel Range value: 10, ..., 796531
\item \strong{ST125MD} \emph{numeric} Median CBF value for FreeSurfer ROI: Right-vessel Range value: 5, ..., 613650
\item \strong{ST125AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-vessel Range value: 4.58182, ..., 578953
\item \strong{ST125SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-vessel Range value: 2.71273, ..., 190778
\item \strong{ST125CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-vessel Range value: 5, ..., 243
\item \strong{ST127MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: 3rd-Ventricle Range value: 1, ..., 230379
\item \strong{ST127MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: 3rd-Ventricle Range value: 71593, ..., 1120470
\item \strong{ST127MD} \emph{numeric} Median CBF value for FreeSurfer ROI: 3rd-Ventricle Range value: 7, ..., 515358
\item \strong{ST127AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: 3rd-Ventricle Range value: 8848.25, ..., 522540
\item \strong{ST127SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: 3rd-Ventricle Range value: 17887.9, ..., 272306
\item \strong{ST127CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: 3rd-Ventricle Range value: 78, ..., 1461
\item \strong{ST128MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: WM-hypointensities Range value: 1, ..., 419
\item \strong{ST128MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: WM-hypointensities Range value: 77532, ..., 987269
\item \strong{ST128MD} \emph{numeric} Median CBF value for FreeSurfer ROI: WM-hypointensities Range value: 6, ..., 135864
\item \strong{ST128AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: WM-hypointensities Range value: 1251.08, ..., 143033
\item \strong{ST128SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: WM-hypointensities Range value: 6637.88, ..., 127475
\item \strong{ST128CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: WM-hypointensities Range value: 593, ..., 41339
\item \strong{ST129MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-insula Range value: 1, ..., 95046
\item \strong{ST129MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-insula Range value: 198109, ..., 1458820
\item \strong{ST129MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-lh-insula Range value: 61993.5, ..., 563400
\item \strong{ST129AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-insula Range value: 62547.7, ..., 559115
\item \strong{ST129SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-insula Range value: 35109.7, ..., 228005
\item \strong{ST129CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-insula Range value: 3405, ..., 8220
\item \strong{ST130MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-insula Range value: 1, ..., 35364
\item \strong{ST130MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-insula Range value: 168781, ..., 1266580
\item \strong{ST130MD} \emph{numeric} Median CBF value for FreeSurfer ROI: ctx-rh-insula Range value: 67047, ..., 524834
\item \strong{ST130AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-insula Range value: 68903.3, ..., 532781
\item \strong{ST130SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-insula Range value: 30888.6, ..., 227281
\item \strong{ST130CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-insula Range value: 3150, ..., 7945
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adnigo}
\keyword{imaging}
