% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFASLFSCBF}
\alias{UCSFASLFSCBF}
\title{ASL Perfusion CBF by FreeSurfer ROI}
\format{
A data frame with 515 observations and 438 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFASLFSCBF)
}
\description{
ASL Perfusion CBF by FreeSurfer ROI data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNI2, ADNI3, ADNIGO
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI3
\item \strong{PTID} \emph{character} Participant Identifier for the in-clinic ADNI study  
\item \strong{RID} \emph{numeric} Participant Identifier for the in-clinic ADNI study Range value: 21, ..., 7092
\item \strong{VISCODE} \emph{character} VISCODE of the Image , with options: y1, y2, init, sc, y4, y3, y5
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{IMAGEUID} \emph{numeric} Image UID of image that was processed. Range value: 812264, ..., 1691094
\item \strong{EXAMDATE} \emph{ } When the image was collected. Date: YYY-MM-DD
\item \strong{RUNDATE} \emph{ } When the image was processed. Date: YYY-MM-DD
\item \strong{ASLFLAVOR} \emph{character} The dimensionality and pulse sequence of the ASL acquisition , with options: SIEMENS3DPASL, 3DPCASL, SIEMENS2DPASL, PHILIPS2DPASL
\item \strong{CBFQC} \emph{character} Checks the qCBF overlaid on the PD image , with options: Pass, Fail
\item \strong{SEGQC} \emph{character} Checks the PD (M0) to T1 registration , with options: Pass, Fail
\item \strong{REGQC} \emph{character} Checks the FreeSurfer aparc+aseg in T1 space , with options: Pass, Fail
\item \strong{FSOVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. , with options: Pass, Partial
\item \strong{FSTEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) , with options: Pass, Fail
\item \strong{FSFRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) , with options: Pass, Fail
\item \strong{FSPARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) , with options: Pass, Fail
\item \strong{FSINSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) , with options: Pass, Fail
\item \strong{FSOCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) , with options: Pass, Fail
\item \strong{FSBGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) , with options: Pass, Fail
\item \strong{FSCWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87) , with options: Pass, Fail
\item \strong{FSVENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) , with options: Pass, Fail
\item \strong{ST1MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Brain-Stem Range value: -10386, ..., 689.853
\item \strong{ST1MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Brain-Stem Range value: 37.1248, ..., 20369.3
\item \strong{ST1AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Brain-Stem Range value: -47.3686, ..., 2284.07
\item \strong{ST1SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Brain-Stem Range value: 3.7962, ..., 1752.37
\item \strong{ST1CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Brain-Stem Range value: 12779, ..., 26796
\item \strong{ST11MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: -1101.28, ..., 1664.18
\item \strong{ST11MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: -5.22206e-08, ..., 12129.3
\item \strong{ST11AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: -0.114073, ..., 6170.38
\item \strong{ST11SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Accumbens-area Range value: 0.00660534, ..., 2997.92
\item \strong{ST11CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Accumbens-area Range value: 109, ..., 885
\item \strong{ST12MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Amygdala Range value: -420.299, ..., 1596.58
\item \strong{ST12MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Amygdala Range value: 5.79293, ..., 7632.48
\item \strong{ST12AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Amygdala Range value: 0.0323392, ..., 3739.37
\item \strong{ST12SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Amygdala Range value: 0.344373, ..., 1400.07
\item \strong{ST12CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Amygdala Range value: 640, ..., 2494
\item \strong{ST13MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: -134.222, ..., 1377.05
\item \strong{ST13MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: 24.8028, ..., 13009
\item \strong{ST13AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: 0.455634, ..., 3781.37
\item \strong{ST13SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-bankssts Range value: 2.00993, ..., 2440.89
\item \strong{ST13CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-bankssts Range value: 837, ..., 3507
\item \strong{ST14MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: -1258.53, ..., 1590.42
\item \strong{ST14MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 5.50916, ..., 8054.17
\item \strong{ST14AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 0.0549773, ..., 3461.48
\item \strong{ST14SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 0.368075, ..., 1594.26
\item \strong{ST14CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-caudalanteriorcingulate Range value: 381, ..., 3036
\item \strong{ST15MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: -2773.18, ..., 725.752
\item \strong{ST15MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 23.6794, ..., 7234
\item \strong{ST15AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: -40.739, ..., 1437.95
\item \strong{ST15SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 3.49624, ..., 899.627
\item \strong{ST15CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-caudalmiddlefrontal Range value: 2631, ..., 8356
\item \strong{ST16MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Caudate Range value: -1800.19, ..., 492.951
\item \strong{ST16MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Caudate Range value: 0.000105793, ..., 8595.62
\item \strong{ST16AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Caudate Range value: -0.126091, ..., 1573.57
\item \strong{ST16SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Caudate Range value: 4.99648e-05, ..., 1168.09
\item \strong{ST16CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Caudate Range value: 31, ..., 5551
\item \strong{ST23MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: -262.2, ..., 1187.89
\item \strong{ST23MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: 19.8083, ..., 16098.2
\item \strong{ST23AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: 3.17391, ..., 2459.44
\item \strong{ST23SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-cuneus Range value: 2.44041, ..., 1382.04
\item \strong{ST23CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-cuneus Range value: 1124, ..., 5140
\item \strong{ST24MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: -14217, ..., 1010.56
\item \strong{ST24MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: 30.8344, ..., 16793.2
\item \strong{ST24AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: -0.266852, ..., 4420.85
\item \strong{ST24SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-entorhinal Range value: 2.76962, ..., 2913.02
\item \strong{ST24CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-entorhinal Range value: 560, ..., 3448
\item \strong{ST25MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: -432.447, ..., 1420.46
\item \strong{ST25MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: 16.4994, ..., 4824.88
\item \strong{ST25AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: -27.9854, ..., 2246.92
\item \strong{ST25SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-frontalpole Range value: 2.05131, ..., 989.867
\item \strong{ST25CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-frontalpole Range value: 432, ..., 1515
\item \strong{ST26MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: -911.943, ..., 604.549
\item \strong{ST26MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: 34.1372, ..., 14496.6
\item \strong{ST26AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: 13.6224, ..., 2023.22
\item \strong{ST26SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-fusiform Range value: 3.964, ..., 1834.1
\item \strong{ST26CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-fusiform Range value: 5214, ..., 12961
\item \strong{ST29MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Hippocampus Range value: -5086, ..., 1258.39
\item \strong{ST29MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Hippocampus Range value: 25.9055, ..., 10936.2
\item \strong{ST29AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Hippocampus Range value: 4.36806, ..., 3425.35
\item \strong{ST29SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Hippocampus Range value: 2.6025, ..., 1755.47
\item \strong{ST29CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Hippocampus Range value: 1783, ..., 5000
\item \strong{ST31MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: -581.128, ..., 552.477
\item \strong{ST31MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 22.6054, ..., 16031.8
\item \strong{ST31AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 10.5849, ..., 2757.97
\item \strong{ST31SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 2.67486, ..., 2099.71
\item \strong{ST31CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-inferiorparietal Range value: 5320, ..., 16988
\item \strong{ST32MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: -1652.41, ..., 533.192
\item \strong{ST32MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 31.9236, ..., 18555.1
\item \strong{ST32AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 15.7119, ..., 1972.66
\item \strong{ST32SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 3.6132, ..., 1945.27
\item \strong{ST32CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-inferiortemporal Range value: 5373, ..., 15909
\item \strong{ST34MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: -47.5824, ..., 1351.11
\item \strong{ST34MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 18.2932, ..., 6398.51
\item \strong{ST34AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: -2.54398, ..., 2387.44
\item \strong{ST34SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 3.26648, ..., 950.083
\item \strong{ST34CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-isthmuscingulate Range value: 926, ..., 3586
\item \strong{ST35MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: -817.912, ..., 581.824
\item \strong{ST35MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 24.1187, ..., 18763.2
\item \strong{ST35AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: -37.3492, ..., 2192.79
\item \strong{ST35SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 2.65266, ..., 1567.97
\item \strong{ST35CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-lateraloccipital Range value: 4781, ..., 15277
\item \strong{ST36MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: -981.61, ..., 673.204
\item \strong{ST36MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 36.5844, ..., 15934
\item \strong{ST36AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: -5.63457, ..., 2889.01
\item \strong{ST36SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 4.1784, ..., 2379.69
\item \strong{ST36CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-lateralorbitofrontal Range value: 3978, ..., 9208
\item \strong{ST38MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: -318.722, ..., 950.697
\item \strong{ST38MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: 26.4187, ..., 8998.86
\item \strong{ST38AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: 0.265909, ..., 2412.48
\item \strong{ST38SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lingual Range value: 2.51561, ..., 1217.26
\item \strong{ST38CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-lingual Range value: 2653, ..., 10326
\item \strong{ST39MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: -4053.31, ..., 883.087
\item \strong{ST39MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 37.7604, ..., 17166.9
\item \strong{ST39AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: -10.1664, ..., 2816.58
\item \strong{ST39SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 4.50344, ..., 2098.33
\item \strong{ST39CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-medialorbitofrontal Range value: 2720, ..., 7065
\item \strong{ST40MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: -580.534, ..., 828.282
\item \strong{ST40MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: 35.4303, ..., 32569.4
\item \strong{ST40AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: 13.5305, ..., 2908.3
\item \strong{ST40SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-middletemporal Range value: 3.66558, ..., 2614.12
\item \strong{ST40CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-middletemporal Range value: 5683, ..., 14790
\item \strong{ST42MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Pallidum Range value: -160.748, ..., 1162.01
\item \strong{ST42MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Pallidum Range value: 0.000882827, ..., 6071.73
\item \strong{ST42AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Pallidum Range value: -0.248449, ..., 2056.36
\item \strong{ST42SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Pallidum Range value: 0.000266829, ..., 1003.17
\item \strong{ST42CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Pallidum Range value: 16, ..., 3077
\item \strong{ST43MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: -1960.91, ..., 872.599
\item \strong{ST43MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: 11.7829, ..., 7268.87
\item \strong{ST43AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: -10.3757, ..., 1911.02
\item \strong{ST43SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-paracentral Range value: 3.7534, ..., 998.78
\item \strong{ST43CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-paracentral Range value: 1214, ..., 5181
\item \strong{ST44MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: -366.917, ..., 1469.31
\item \strong{ST44MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: 31.161, ..., 12036.8
\item \strong{ST44AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: 14.4367, ..., 3941.76
\item \strong{ST44SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parahippocampal Range value: 2.4459, ..., 2296.77
\item \strong{ST44CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-parahippocampal Range value: 748, ..., 2889
\item \strong{ST45MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: -185.578, ..., 1146.01
\item \strong{ST45MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: 23.4366, ..., 6454.4
\item \strong{ST45AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: 4.23882, ..., 2304.63
\item \strong{ST45SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parsopercularis Range value: 2.2482, ..., 1046.03
\item \strong{ST45CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-parsopercularis Range value: 1689, ..., 6530
\item \strong{ST46MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: -360.533, ..., 1068.65
\item \strong{ST46MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 19.2553, ..., 5961.37
\item \strong{ST46AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 5.66897, ..., 1973.64
\item \strong{ST46SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 1.99132, ..., 926.282
\item \strong{ST46CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-parsorbitalis Range value: 1048, ..., 3151
\item \strong{ST47MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: -68.9963, ..., 925.296
\item \strong{ST47MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: 20.7563, ..., 8401.12
\item \strong{ST47AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: 10.0803, ..., 2725.04
\item \strong{ST47SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parstriangularis Range value: 3.46746, ..., 1436.69
\item \strong{ST47CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-parstriangularis Range value: 1479, ..., 5249
\item \strong{ST48MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: -229.263, ..., 1124.04
\item \strong{ST48MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: 16.2649, ..., 10002.6
\item \strong{ST48AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: 1.61882, ..., 2450.99
\item \strong{ST48SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-pericalcarine Range value: 1.99046, ..., 1095.95
\item \strong{ST48CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-pericalcarine Range value: 740, ..., 3222
\item \strong{ST49MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: -1960.91, ..., 749.681
\item \strong{ST49MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: 36.1975, ..., 20008.9
\item \strong{ST49AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: 6.95792, ..., 1835.4
\item \strong{ST49SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-postcentral Range value: 3.60346, ..., 1988.61
\item \strong{ST49CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-postcentral Range value: 3942, ..., 12342
\item \strong{ST50MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: -787.932, ..., 1487.9
\item \strong{ST50MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 0.328431, ..., 7337.72
\item \strong{ST50AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: -0.0136273, ..., 2480.54
\item \strong{ST50SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 0.0314972, ..., 924.733
\item \strong{ST50CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-posteriorcingulate Range value: 510, ..., 4241
\item \strong{ST51MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: -6731.8, ..., 576.196
\item \strong{ST51MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: 38.7107, ..., 11515.1
\item \strong{ST51AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: -2.54097, ..., 1667.27
\item \strong{ST51SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-precentral Range value: 3.91496, ..., 1245.49
\item \strong{ST51CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-precentral Range value: 5996, ..., 16161
\item \strong{ST52MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: -397.204, ..., 803.273
\item \strong{ST52MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: 19.4117, ..., 6224.24
\item \strong{ST52AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: 6.46745, ..., 2115.85
\item \strong{ST52SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-precuneus Range value: 2.20749, ..., 883.763
\item \strong{ST52CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-precuneus Range value: 2702, ..., 12012
\item \strong{ST53MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Putamen Range value: -43.5165, ..., 1083.26
\item \strong{ST53MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Putamen Range value: 0.000428547, ..., 11324.4
\item \strong{ST53AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Putamen Range value: -0.000310907, ..., 2727.57
\item \strong{ST53SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Putamen Range value: 0.00022524, ..., 1425.66
\item \strong{ST53CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Putamen Range value: 13, ..., 6099
\item \strong{ST54MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: -3470.64, ..., 1260.02
\item \strong{ST54MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 3.5454, ..., 8950.78
\item \strong{ST54AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 0.10126, ..., 3522.23
\item \strong{ST54SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 0.30721, ..., 1666.53
\item \strong{ST54CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-rostralanteriorcingulate Range value: 996, ..., 3811
\item \strong{ST55MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: -858.937, ..., 597.677
\item \strong{ST55MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 31.5792, ..., 9806.44
\item \strong{ST55AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 4.37067, ..., 1800.54
\item \strong{ST55SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 3.90851, ..., 982.984
\item \strong{ST55CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-rostralmiddlefrontal Range value: 7828, ..., 19547
\item \strong{ST56MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: -2447.68, ..., 446.359
\item \strong{ST56MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 21.0229, ..., 14259.9
\item \strong{ST56AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: -13.6599, ..., 1941.35
\item \strong{ST56SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 5.25059, ..., 1442.75
\item \strong{ST56CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiorfrontal Range value: 12219, ..., 27170
\item \strong{ST57MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: -874.278, ..., 522.183
\item \strong{ST57MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: 19.5387, ..., 15005.7
\item \strong{ST57AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: 5.51794, ..., 1465.57
\item \strong{ST57SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiorparietal Range value: 1.99672, ..., 1128.83
\item \strong{ST57CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiorparietal Range value: 5684, ..., 16382
\item \strong{ST58MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: -259.142, ..., 842.547
\item \strong{ST58MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: 46.0675, ..., 23018.5
\item \strong{ST58AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: 1.80499, ..., 3627.06
\item \strong{ST58SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiortemporal Range value: 4.88059, ..., 2900.32
\item \strong{ST58CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiortemporal Range value: 6259, ..., 17100
\item \strong{ST59MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: -1058.06, ..., 957.443
\item \strong{ST59MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: 23.3221, ..., 20891
\item \strong{ST59AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: 8.26686, ..., 2326.81
\item \strong{ST59SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-supramarginal Range value: 2.07704, ..., 1807.66
\item \strong{ST59CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-supramarginal Range value: 4450, ..., 17069
\item \strong{ST60MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: -684.113, ..., 1000.5
\item \strong{ST60MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: 19.8087, ..., 12656.7
\item \strong{ST60AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: -19.1862, ..., 3922.07
\item \strong{ST60SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-temporalpole Range value: 2.53074, ..., 2424.05
\item \strong{ST60CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-temporalpole Range value: 540, ..., 3698
\item \strong{ST61MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: -3548.21, ..., 705.001
\item \strong{ST61MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: 29.0451, ..., 12409.4
\item \strong{ST61AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: 7.81628, ..., 3447.89
\item \strong{ST61SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Thalamus-Proper Range value: 3.78034, ..., 2426.4
\item \strong{ST61CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Left-Thalamus-Proper Range value: 3525, ..., 8760
\item \strong{ST62MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: -0.0300484, ..., 2108.75
\item \strong{ST62MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: 0.000192979, ..., 10107.5
\item \strong{ST62AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: -0.00259514, ..., 3659.3
\item \strong{ST62SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-transversetemporal Range value: 5.67909e-05, ..., 1758.3
\item \strong{ST62CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-transversetemporal Range value: 11, ..., 1812
\item \strong{ST70MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: -1531.63, ..., 1495.42
\item \strong{ST70MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: -4.6536e-06, ..., 10784.9
\item \strong{ST70AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: -0.211484, ..., 4724.21
\item \strong{ST70SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Accumbens-area Range value: 0.00232823, ..., 2592.37
\item \strong{ST70CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Accumbens-area Range value: 146, ..., 1057
\item \strong{ST71MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Amygdala Range value: -1593.97, ..., 1468.85
\item \strong{ST71MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Amygdala Range value: 0.957156, ..., 6749.71
\item \strong{ST71AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Amygdala Range value: -0.00592455, ..., 3295.49
\item \strong{ST71SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Amygdala Range value: 0.0915562, ..., 1121.07
\item \strong{ST71CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Amygdala Range value: 507, ..., 2399
\item \strong{ST72MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: -101.519, ..., 1179.66
\item \strong{ST72MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: 8.60025, ..., 8255.24
\item \strong{ST72AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: 0.239414, ..., 3067
\item \strong{ST72SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-bankssts Range value: 1.16927, ..., 1770.94
\item \strong{ST72CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-bankssts Range value: 552, ..., 3026
\item \strong{ST73MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: -2092.54, ..., 1385.57
\item \strong{ST73MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 1.26261, ..., 6416.33
\item \strong{ST73AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 0.0367889, ..., 2832.67
\item \strong{ST73SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 0.12251, ..., 1196.49
\item \strong{ST73CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-caudalanteriorcingulate Range value: 401, ..., 3541
\item \strong{ST74MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: -1331.34, ..., 880.208
\item \strong{ST74MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 18.9668, ..., 19419
\item \strong{ST74AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 8.41395, ..., 1849.5
\item \strong{ST74SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 4.54585, ..., 1932
\item \strong{ST74CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-caudalmiddlefrontal Range value: 2424, ..., 7646
\item \strong{ST75MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Caudate Range value: -3116.32, ..., 546.055
\item \strong{ST75MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Caudate Range value: 0.000110223, ..., 7348.06
\item \strong{ST75AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Caudate Range value: -614.523, ..., 1770.6
\item \strong{ST75SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Caudate Range value: 0.00231673, ..., 1057.78
\item \strong{ST75CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Caudate Range value: 49, ..., 5957
\item \strong{ST82MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: -448.444, ..., 1607.78
\item \strong{ST82MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: 20.4848, ..., 10113.9
\item \strong{ST82AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: -1.94679, ..., 3383.2
\item \strong{ST82SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-cuneus Range value: 2.34043, ..., 1124.54
\item \strong{ST82CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-cuneus Range value: 1418, ..., 5745
\item \strong{ST83MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: -27297, ..., 1170.11
\item \strong{ST83MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: 22.7939, ..., 29169
\item \strong{ST83AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: -8.66998, ..., 2599.04
\item \strong{ST83SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-entorhinal Range value: 2.97063, ..., 2532.49
\item \strong{ST83CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-entorhinal Range value: 502, ..., 3359
\item \strong{ST84MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: -184.458, ..., 1189.75
\item \strong{ST84MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: 21.1316, ..., 13474.5
\item \strong{ST84AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: -19.4797, ..., 3594.19
\item \strong{ST84SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-frontalpole Range value: 1.97981, ..., 2402.77
\item \strong{ST84CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-frontalpole Range value: 451, ..., 1820
\item \strong{ST85MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: -856.025, ..., 466.559
\item \strong{ST85MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: 36, ..., 6796.3
\item \strong{ST85AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: -11.0208, ..., 1716.65
\item \strong{ST85SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-fusiform Range value: 4.09071, ..., 806.811
\item \strong{ST85CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-fusiform Range value: 3958, ..., 13483
\item \strong{ST88MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Hippocampus Range value: -8081, ..., 846.721
\item \strong{ST88MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Hippocampus Range value: 30.7417, ..., 7299.7
\item \strong{ST88AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Hippocampus Range value: 8.55313, ..., 2413.33
\item \strong{ST88SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Hippocampus Range value: 3.31855, ..., 1312.65
\item \strong{ST88CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Hippocampus Range value: 1935, ..., 5232
\item \strong{ST90MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: -5661, ..., 565.312
\item \strong{ST90MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 31.9671, ..., 14021
\item \strong{ST90AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 8.80525, ..., 2941.8
\item \strong{ST90SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 4.34012, ..., 2439.72
\item \strong{ST90CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-inferiorparietal Range value: 6353, ..., 18756
\item \strong{ST91MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: -1489.29, ..., 749.763
\item \strong{ST91MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 33.1381, ..., 13105.8
\item \strong{ST91AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 5.32134, ..., 2371.08
\item \strong{ST91SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 3.66874, ..., 1881.72
\item \strong{ST91CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-inferiortemporal Range value: 4533, ..., 15379
\item \strong{ST93MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: -95.4517, ..., 1181.82
\item \strong{ST93MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 18.8343, ..., 9597.38
\item \strong{ST93AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 1.26154, ..., 2230.05
\item \strong{ST93SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 2.13332, ..., 1131.02
\item \strong{ST93CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-isthmuscingulate Range value: 1035, ..., 3193
\item \strong{ST94MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: -891.208, ..., 605.193
\item \strong{ST94MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 19.4936, ..., 10243.7
\item \strong{ST94AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: -39.3239, ..., 1992.83
\item \strong{ST94SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 2.56347, ..., 1392.38
\item \strong{ST94CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-lateraloccipital Range value: 5805, ..., 17889
\item \strong{ST95MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: -567.112, ..., 656.322
\item \strong{ST95MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 37.2618, ..., 11213
\item \strong{ST95AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: -2.45757, ..., 2230.11
\item \strong{ST95SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 4.35594, ..., 1893.45
\item \strong{ST95CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-lateralorbitofrontal Range value: 3984, ..., 9612
\item \strong{ST97MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: -679.948, ..., 1145.75
\item \strong{ST97MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: 30.6829, ..., 14203.9
\item \strong{ST97AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: -0.666828, ..., 2645.56
\item \strong{ST97SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lingual Range value: 3.99141, ..., 1034.35
\item \strong{ST97CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-lingual Range value: 3358, ..., 9324
\item \strong{ST98MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: -1507.12, ..., 852.717
\item \strong{ST98MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 33.0811, ..., 13133.2
\item \strong{ST98AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 2.5509, ..., 2850.1
\item \strong{ST98SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 4.68474, ..., 2309.07
\item \strong{ST98CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-medialorbitofrontal Range value: 2977, ..., 6873
\item \strong{ST99MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: -526.589, ..., 852.886
\item \strong{ST99MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: 31.3705, ..., 25528.9
\item \strong{ST99AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: 14.9986, ..., 3381.72
\item \strong{ST99SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-middletemporal Range value: 3.8289, ..., 2623.49
\item \strong{ST99CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-middletemporal Range value: 5998, ..., 17367
\item \strong{ST101MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Pallidum Range value: -1722.89, ..., 1128.2
\item \strong{ST101MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Pallidum Range value: 1.7482e-05, ..., 7321.6
\item \strong{ST101AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Pallidum Range value: -242.21, ..., 2685.41
\item \strong{ST101SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Pallidum Range value: 0.000181311, ..., 1710.97
\item \strong{ST101CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Pallidum Range value: 30, ..., 2947
\item \strong{ST102MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: -1960.91, ..., 889.309
\item \strong{ST102MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: 12.0444, ..., 8620.74
\item \strong{ST102AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: -9.70174, ..., 1985.94
\item \strong{ST102SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-paracentral Range value: 2.3976, ..., 1129.13
\item \strong{ST102CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-paracentral Range value: 1684, ..., 5316
\item \strong{ST103MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: -431.758, ..., 1349.19
\item \strong{ST103MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: 34.3939, ..., 8462.38
\item \strong{ST103AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: 17.9239, ..., 3323.01
\item \strong{ST103SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parahippocampal Range value: 2.83994, ..., 2039.43
\item \strong{ST103CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-parahippocampal Range value: 870, ..., 2699
\item \strong{ST104MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: -283.268, ..., 1021.44
\item \strong{ST104MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: 15.7735, ..., 12680.9
\item \strong{ST104AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: 0.640445, ..., 4098.32
\item \strong{ST104SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parsopercularis Range value: 1.97579, ..., 2674.61
\item \strong{ST104CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-parsopercularis Range value: 1079, ..., 5059
\item \strong{ST105MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: -421.957, ..., 1103.62
\item \strong{ST105MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 18.7307, ..., 14163.1
\item \strong{ST105AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 11.0068, ..., 3025.15
\item \strong{ST105SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 2.13051, ..., 2429.63
\item \strong{ST105CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-parsorbitalis Range value: 1330, ..., 3897
\item \strong{ST106MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: -118.964, ..., 1359.91
\item \strong{ST106MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: 29.2297, ..., 15723.6
\item \strong{ST106AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: 12.8057, ..., 2940.52
\item \strong{ST106SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parstriangularis Range value: 4.28091, ..., 2243.93
\item \strong{ST106CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-parstriangularis Range value: 1837, ..., 5980
\item \strong{ST107MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: -322.857, ..., 1373.84
\item \strong{ST107MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: 19.0724, ..., 9570.57
\item \strong{ST107AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: -2.88251, ..., 3100.51
\item \strong{ST107SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-pericalcarine Range value: 1.75565, ..., 826.11
\item \strong{ST107CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-pericalcarine Range value: 952, ..., 4175
\item \strong{ST108MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: -6731.8, ..., 787.884
\item \strong{ST108MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: 27.1875, ..., 19497.3
\item \strong{ST108AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: 4.17135, ..., 2371.66
\item \strong{ST108SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-postcentral Range value: 4.69948, ..., 2176.17
\item \strong{ST108CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-postcentral Range value: 3680, ..., 10825
\item \strong{ST109MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: -524.789, ..., 1199.29
\item \strong{ST109MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 3.41654, ..., 6149.49
\item \strong{ST109AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: -1.1807, ..., 2363.93
\item \strong{ST109SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 0.281129, ..., 1104.62
\item \strong{ST109CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-posteriorcingulate Range value: 434, ..., 4340
\item \strong{ST110MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: -6731.8, ..., 711.55
\item \strong{ST110MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: 34.3098, ..., 17516
\item \strong{ST110AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: -6.79888, ..., 2069.72
\item \strong{ST110SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-precentral Range value: 5.2841, ..., 1935.02
\item \strong{ST110CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-precentral Range value: 4560, ..., 16577
\item \strong{ST111MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: -1960.91, ..., 1225.11
\item \strong{ST111MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: 25.9128, ..., 7828.61
\item \strong{ST111AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: -5.27515, ..., 2346.59
\item \strong{ST111SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-precuneus Range value: 3.11296, ..., 787.513
\item \strong{ST111CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-precuneus Range value: 3129, ..., 12564
\item \strong{ST112MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Putamen Range value: -1661.38, ..., 1054.15
\item \strong{ST112MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Putamen Range value: 0.000813951, ..., 7899.2
\item \strong{ST112AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Putamen Range value: -0.0098058, ..., 4013.6
\item \strong{ST112SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Putamen Range value: 0.000366438, ..., 1699.17
\item \strong{ST112CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Putamen Range value: 36, ..., 6611
\item \strong{ST113MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: -1368.43, ..., 1364
\item \strong{ST113MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 9.0397, ..., 11158.6
\item \strong{ST113AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 0.166082, ..., 4000.62
\item \strong{ST113SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 0.568884, ..., 2335.18
\item \strong{ST113CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-rostralanteriorcingulate Range value: 652, ..., 3522
\item \strong{ST114MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: -1248.49, ..., 585.624
\item \strong{ST114MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 36.794, ..., 15579.8
\item \strong{ST114AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: -2.40488, ..., 2142.98
\item \strong{ST114SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 4.02282, ..., 1599.62
\item \strong{ST114CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-rostralmiddlefrontal Range value: 8197, ..., 19821
\item \strong{ST115MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: -4033.2, ..., 478.605
\item \strong{ST115MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 22.7068, ..., 13753.5
\item \strong{ST115AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: -9.0192, ..., 1940.53
\item \strong{ST115SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 4.98537, ..., 1458.74
\item \strong{ST115CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiorfrontal Range value: 11219, ..., 27825
\item \strong{ST116MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: -1960.91, ..., 445.435
\item \strong{ST116MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: 21.2389, ..., 13563.3
\item \strong{ST116AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: -29.4729, ..., 1656.78
\item \strong{ST116SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiorparietal Range value: 2.5012, ..., 898.772
\item \strong{ST116CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiorparietal Range value: 4329, ..., 15945
\item \strong{ST117MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: -216.665, ..., 911.719
\item \strong{ST117MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: 36.97, ..., 17132.9
\item \strong{ST117AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: 14.5037, ..., 3322.99
\item \strong{ST117SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiortemporal Range value: 4.69229, ..., 1919.5
\item \strong{ST117CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiortemporal Range value: 4579, ..., 15324
\item \strong{ST118MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: -23115, ..., 673.776
\item \strong{ST118MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: 35.1016, ..., 16619
\item \strong{ST118AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: 0.559278, ..., 3125.45
\item \strong{ST118SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-supramarginal Range value: 4.83052, ..., 1716.98
\item \strong{ST118CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-supramarginal Range value: 3958, ..., 13781
\item \strong{ST119MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: -1792, ..., 1100
\item \strong{ST119MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: 21.8429, ..., 10677
\item \strong{ST119AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: -36.2806, ..., 3816.36
\item \strong{ST119SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-temporalpole Range value: 2.4972, ..., 1770.99
\item \strong{ST119CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-temporalpole Range value: 696, ..., 3454
\item \strong{ST120MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: -2697.22, ..., 613.051
\item \strong{ST120MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: 31.7976, ..., 7194.02
\item \strong{ST120AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: 16.2942, ..., 1945.56
\item \strong{ST120SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Thalamus-Proper Range value: 3.84997, ..., 1264.9
\item \strong{ST120CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: Right-Thalamus-Proper Range value: 3950, ..., 8734
\item \strong{ST121MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: -7.15097, ..., 1660.36
\item \strong{ST121MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: 0.000376913, ..., 9129.96
\item \strong{ST121AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: -0.0760365, ..., 3531.79
\item \strong{ST121SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-transversetemporal Range value: 0.0198411, ..., 1397.22
\item \strong{ST121CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-transversetemporal Range value: 84, ..., 1454
\item \strong{ST129MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-lh-insula Range value: -60.6214, ..., 937.785
\item \strong{ST129MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-lh-insula Range value: 0.66215, ..., 15740
\item \strong{ST129AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-lh-insula Range value: -0.0147535, ..., 4018.43
\item \strong{ST129SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-insula Range value: 0.106534, ..., 1767.19
\item \strong{ST129CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-lh-insula Range value: 185, ..., 9529
\item \strong{ST130MIN} \emph{numeric} Min CBF value for FreeSurfer ROI: ctx-rh-insula Range value: -467.584, ..., 985.536
\item \strong{ST130MAX} \emph{numeric} Max CBF value for FreeSurfer ROI: ctx-rh-insula Range value: 11.0806, ..., 12001.2
\item \strong{ST130AVG} \emph{numeric} Mean CBF value for FreeSurfer ROI: ctx-rh-insula Range value: -0.357995, ..., 3643.65
\item \strong{ST130SD} \emph{numeric} Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-insula Range value: 0.418108, ..., 2033.01
\item \strong{ST130CT} \emph{numeric} Count of CBF voxels in FreeSurfer ROI: ctx-rh-insula Range value: 415, ..., 10551
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{imaging}
