% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFASLQC}
\alias{UCSFASLQC}
\title{ASL Perfusion Raw QC}
\format{
A data frame with 1011 observations and 7 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFASLQC)
}
\description{
ASL Perfusion Raw QC data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{PTID} \emph{character} Subject Identifier  
\item \strong{LONIUID} \emph{numeric} The LONI Series UID for the image reviewed. Range value: 93192, ..., 256044
\item \strong{IMAGEUID} \emph{numeric} The LONI Image UID for the image reviewed. Range value: 196079, ..., 487455
\item \strong{QCRating} \emph{character}   Pass/Fail
\item \strong{QCDate} \emph{ }   Date: YYY-MM-DD
\item \strong{QCComments} \emph{character}    
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
