% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFATRPHY}
\alias{UCSFATRPHY}
\title{UCSF (Colin Studholme) Regional Atrophy Rates}
\format{
A data frame with 1721 observations and 13 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFATRPHY)
}
\description{
UCSF (Colin Studholme) Regional Atrophy Rates data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 1
\item \strong{RID} \emph{numeric}   Range value: 3, ..., 1430
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 127
\item \strong{VISCODE1} \emph{character}   , with options: sc, bl
\item \strong{EXAMDATE1} \emph{ }   Date: YYY-MM-DD
\item \strong{IMAGEUID1} \emph{numeric}   Range value: 8723, ..., 50770
\item \strong{VISCODE2} \emph{character}   , with options: m06, m12
\item \strong{EXAMDATE2} \emph{ }   Date: YYY-MM-DD
\item \strong{IMAGEUID2} \emph{numeric}   Range value: 12037, ..., 60591
\item \strong{FLDSTRENG} \emph{numeric} Field Strength Range value: 1.5, ..., 3
\item \strong{PCTCHANGE} \emph{numeric} Summaries of regional atrophy rates between first and last scan (i.e. 12 month interval). There is one percent change for each subject summarizing changes in both temporal lobes together (minus superior temporal gyrus). Range value: -3.58032, ..., 2.4082
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
