% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFFSL}
\alias{UCSFFSL}
\title{Longitudinal FreeSurfer (FreeSurfer Version 4.4): 2016-02-01}
\format{
A data frame with 3570 observations and 376 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFFSL)
}
\description{
Longitudinal FreeSurfer (FreeSurfer Version 4.4): 2016-02-01 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 3, ..., 1430
\item \strong{VISCODE} \emph{character} Visit code , with options: m06, m12, m24, sc, m18, m36, m48, f, bl, m60
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: m06, m12, m24, sc, m18, m36, m48, f, bl, m60
\item \strong{EXAMDATE} \emph{ } The scan date for the image processed Date: YYY-MM-DD
\item \strong{VERSION} \emph{ }   Date: YYY-MM-DD
\item \strong{FLDSTRENG} \emph{numeric} The field strength. 1.5 or 3 Range value: 1.5, ..., 1.5
\item \strong{LONISID} \emph{numeric} The LONI study UID for the image processed Range value: 1948, ..., 44565
\item \strong{LONIUID} \emph{numeric} The LONI series UID for the image processed. Range value: 8723, ..., 146809
\item \strong{IMAGEUID} \emph{numeric} The ImageUID for the image processed. Range value: 31863, ..., 322396
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   , with options: complete, partial
\item \strong{BASETP1} \emph{numeric} Whether or not the base image includes a timepoint 1 image. 0 = No, 1 = Yes
\item \strong{BASETP2} \emph{numeric} Whether or not the base image includes a timepoint 2 image. 0 = No, 1 = Yes
\item \strong{BASETP3} \emph{numeric} Whether or not the base image includes a timepoint 3 image. 0 = No, 1 = Yes
\item \strong{BASETP4} \emph{numeric} Whether or not the base image includes a timepoint 4 image. 0 = No, 1 = Yes
\item \strong{BASETP5} \emph{numeric} Whether or not the base image includes a timepoint 5 image. 0 = No, 1 = Yes
\item \strong{BASETP6} \emph{numeric} Whether or not the base image includes a timepoint 6 image. 0 = No, 1 = Yes
\item \strong{BASETP7} \emph{numeric} Whether or not the base image includes a timepoint 7 image. 0 = No, 1 = Yes
\item \strong{BASETP8} \emph{numeric} Whether or not the base image includes a timepoint 8 image. 0 = No, 1 = Yes
\item \strong{OVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. Pass/Fail/Partial
\item \strong{TEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) Pass/Fail
\item \strong{FRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) Pass/Fail
\item \strong{PARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) Pass/Fail
\item \strong{INSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) Pass/Fail
\item \strong{OCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) Pass/Fail
\item \strong{BGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) Pass/Fail
\item \strong{CWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79) Pass/Fail
\item \strong{VENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) , with options: Pass, Fail
\item \strong{ST100SV} \emph{logical} Volume (WM Parcellation) of RightNonWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST101SV} \emph{numeric} Volume (WM Parcellation) of RightPallidum Range value: 309, ..., 2317
\item \strong{ST102CV} \emph{numeric} Volume (Cortical Parcellation) of RightParacentral Range value: 1899, ..., 5944
\item \strong{ST102SA} \emph{numeric} Surface Area of RightParacentral Range value: 981, ..., 2711
\item \strong{ST102TA} \emph{numeric} Cortical Thickness Average of RightParacentral Range value: 1.385, ..., 3.281
\item \strong{ST102TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParacentral Range value: 0.363, ..., 1.458
\item \strong{ST103CV} \emph{numeric} Volume (Cortical Parcellation) of RightParahippocampal Range value: 592, ..., 3042
\item \strong{ST103SA} \emph{numeric} Surface Area of RightParahippocampal Range value: 360, ..., 1055
\item \strong{ST103TA} \emph{numeric} Cortical Thickness Average of RightParahippocampal Range value: 0.701, ..., 3.365
\item \strong{ST103TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParahippocampal Range value: 0.456, ..., 1.922
\item \strong{ST104CV} \emph{numeric} Volume (Cortical Parcellation) of RightParsOpercularis Range value: 1523, ..., 7323
\item \strong{ST104SA} \emph{numeric} Surface Area of RightParsOpercularis Range value: 803, ..., 2847
\item \strong{ST104TA} \emph{numeric} Cortical Thickness Average of RightParsOpercularis Range value: 1.368, ..., 2.917
\item \strong{ST104TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsOpercularis Range value: 0.323, ..., 1.509
\item \strong{ST105CV} \emph{numeric} Volume (Cortical Parcellation) of RightParsOrbitalis Range value: 920, ..., 3744
\item \strong{ST105SA} \emph{numeric} Surface Area of RightParsOrbitalis Range value: 308, ..., 1613
\item \strong{ST105TA} \emph{numeric} Cortical Thickness Average of RightParsOrbitalis Range value: 1.508, ..., 3.416
\item \strong{ST105TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsOrbitalis Range value: 0.406, ..., 1.294
\item \strong{ST106CV} \emph{numeric} Volume (Cortical Parcellation) of RightParsTriangularis Range value: 1437, ..., 5845
\item \strong{ST106SA} \emph{numeric} Surface Area of RightParsTriangularis Range value: 763, ..., 2555
\item \strong{ST106TA} \emph{numeric} Cortical Thickness Average of RightParsTriangularis Range value: 1.307, ..., 3.247
\item \strong{ST106TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsTriangularis Range value: 0.32, ..., 1.344
\item \strong{ST107CV} \emph{numeric} Volume (Cortical Parcellation) of RightPericalcarine Range value: 695, ..., 4131
\item \strong{ST107SA} \emph{numeric} Surface Area of RightPericalcarine Range value: 523, ..., 2430
\item \strong{ST107TA} \emph{numeric} Cortical Thickness Average of RightPericalcarine Range value: 1.096, ..., 2.495
\item \strong{ST107TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPericalcarine Range value: 0.186, ..., 1.361
\item \strong{ST108CV} \emph{numeric} Volume (Cortical Parcellation) of RightPostcentral Range value: 4756, ..., 12487
\item \strong{ST108SA} \emph{numeric} Surface Area of RightPostcentral Range value: 2215, ..., 6691
\item \strong{ST108TA} \emph{numeric} Cortical Thickness Average of RightPostcentral Range value: 1.31, ..., 2.753
\item \strong{ST108TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPostcentral Range value: 0.398, ..., 1.224
\item \strong{ST109CV} \emph{numeric} Volume (Cortical Parcellation) of RightPosteriorCingulate Range value: 1117, ..., 4607
\item \strong{ST109SA} \emph{numeric} Surface Area of RightPosteriorCingulate Range value: 730, ..., 2130
\item \strong{ST109TA} \emph{numeric} Cortical Thickness Average of RightPosteriorCingulate Range value: 0.989, ..., 3.3
\item \strong{ST109TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPosteriorCingulate Range value: 0.423, ..., 1.441
\item \strong{ST10CV} \emph{numeric} Volume (Cortical Parcellation) of Icv Range value: 1153430, ..., 2110290
\item \strong{ST110CV} \emph{numeric} Volume (Cortical Parcellation) of RightPrecentral Range value: 6097, ..., 19531
\item \strong{ST110SA} \emph{numeric} Surface Area of RightPrecentral Range value: 3247, ..., 8900
\item \strong{ST110TA} \emph{numeric} Cortical Thickness Average of RightPrecentral Range value: 1.352, ..., 3.717
\item \strong{ST110TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPrecentral Range value: 0.391, ..., 1.565
\item \strong{ST111CV} \emph{numeric} Volume (Cortical Parcellation) of RightPrecuneus Range value: 4104, ..., 12036
\item \strong{ST111SA} \emph{numeric} Surface Area of RightPrecuneus Range value: 2517, ..., 5668
\item \strong{ST111TA} \emph{numeric} Cortical Thickness Average of RightPrecuneus Range value: 1.352, ..., 2.731
\item \strong{ST111TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPrecuneus Range value: 0.373, ..., 1.308
\item \strong{ST112SV} \emph{numeric} Volume (WM Parcellation) of RightPutamen Range value: 1890, ..., 7198
\item \strong{ST113CV} \emph{numeric} Volume (Cortical Parcellation) of RightRostralAnteriorCingulate Range value: 0, ..., 2879
\item \strong{ST113SA} \emph{numeric} Surface Area of RightRostralAnteriorCingulate Range value: 243, ..., 1265
\item \strong{ST113TA} \emph{numeric} Cortical Thickness Average of RightRostralAnteriorCingulate Range value: 0, ..., 3.912
\item \strong{ST113TS} \emph{numeric} Cortical Thickness Standard Deviation of RightRostralAnteriorCingulate Range value: 0, ..., 1.324
\item \strong{ST114CV} \emph{numeric} Volume (Cortical Parcellation) of RightRostralMiddleFrontal Range value: 6116, ..., 20637
\item \strong{ST114SA} \emph{numeric} Surface Area of RightRostralMiddleFrontal Range value: 3287, ..., 9365
\item \strong{ST114TA} \emph{numeric} Cortical Thickness Average of RightRostralMiddleFrontal Range value: 1.376, ..., 2.997
\item \strong{ST114TS} \emph{numeric} Cortical Thickness Standard Deviation of RightRostralMiddleFrontal Range value: 0.37, ..., 1.738
\item \strong{ST115CV} \emph{numeric} Volume (Cortical Parcellation) of RightSuperiorFrontal Range value: 7447, ..., 28356
\item \strong{ST115SA} \emph{numeric} Surface Area of RightSuperiorFrontal Range value: 3528, ..., 10319
\item \strong{ST115TA} \emph{numeric} Cortical Thickness Average of RightSuperiorFrontal Range value: 1.527, ..., 3.483
\item \strong{ST115TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorFrontal Range value: 0.425, ..., 1.763
\item \strong{ST116CV} \emph{numeric} Volume (Cortical Parcellation) of RightSuperiorParietal Range value: 6015, ..., 16887
\item \strong{ST116SA} \emph{numeric} Surface Area of RightSuperiorParietal Range value: 2878, ..., 7581
\item \strong{ST116TA} \emph{numeric} Cortical Thickness Average of RightSuperiorParietal Range value: 1.287, ..., 3.247
\item \strong{ST116TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorParietal Range value: 0.307, ..., 1.326
\item \strong{ST117CV} \emph{numeric} Volume (Cortical Parcellation) of RightSuperiorTemporal Range value: 4362, ..., 13160
\item \strong{ST117SA} \emph{numeric} Surface Area of RightSuperiorTemporal Range value: 2449, ..., 6647
\item \strong{ST117TA} \emph{numeric} Cortical Thickness Average of RightSuperiorTemporal Range value: 1.472, ..., 3.045
\item \strong{ST117TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorTemporal Range value: 0.473, ..., 1.416
\item \strong{ST118CV} \emph{numeric} Volume (Cortical Parcellation) of RightSupramarginal Range value: 4188, ..., 14020
\item \strong{ST118SA} \emph{numeric} Surface Area of RightSupramarginal Range value: 2287, ..., 5908
\item \strong{ST118TA} \emph{numeric} Cortical Thickness Average of RightSupramarginal Range value: 1.474, ..., 2.849
\item \strong{ST118TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSupramarginal Range value: 0.419, ..., 1.296
\item \strong{ST119CV} \emph{numeric} Volume (Cortical Parcellation) of RightTemporalPole Range value: 660, ..., 3150
\item \strong{ST119SA} \emph{numeric} Surface Area of RightTemporalPole Range value: 148, ..., 965
\item \strong{ST119TA} \emph{numeric} Cortical Thickness Average of RightTemporalPole Range value: 1.288, ..., 4.646
\item \strong{ST119TS} \emph{numeric} Cortical Thickness Standard Deviation of RightTemporalPole Range value: 0.289, ..., 1.423
\item \strong{ST11SV} \emph{numeric} Volume (WM Parcellation) of LeftAccumbensArea Range value: 12, ..., 765
\item \strong{ST120SV} \emph{numeric} Volume (WM Parcellation) of RightThalamus Range value: 3725, ..., 10172
\item \strong{ST121CV} \emph{numeric} Volume (Cortical Parcellation) of RightTransverseTemporal Range value: 329, ..., 1617
\item \strong{ST121SA} \emph{numeric} Surface Area of RightTransverseTemporal Range value: 173, ..., 639
\item \strong{ST121TA} \emph{numeric} Cortical Thickness Average of RightTransverseTemporal Range value: 1.064, ..., 4.192
\item \strong{ST121TS} \emph{numeric} Cortical Thickness Standard Deviation of RightTransverseTemporal Range value: 0.163, ..., 1.292
\item \strong{ST122SV} \emph{logical} Volume (WM Parcellation) of RightUndetermined A Boolean value: TRUE/ FALSE
\item \strong{ST123CV} \emph{numeric} Volume (Cortical Parcellation) of RightUnknown Range value: 2302, ..., 9554
\item \strong{ST123SA} \emph{numeric} Surface Area of RightUnknown Range value: 3802, ..., 8432
\item \strong{ST123TA} \emph{numeric} Cortical Thickness Average of RightUnknown Range value: 0.395, ..., 1.513
\item \strong{ST123TS} \emph{numeric} Cortical Thickness Standard Deviation of RightUnknown Range value: 0.709, ..., 1.772
\item \strong{ST124SV} \emph{numeric} Volume (WM Parcellation) of RightVentralDC Range value: 2183, ..., 5733
\item \strong{ST125SV} \emph{numeric} Volume (WM Parcellation) of RightVessel Range value: 2, ..., 521
\item \strong{ST126SV} \emph{logical} Volume (WM Parcellation) of RightWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST127SV} \emph{numeric} Volume (WM Parcellation) of ThirdVentricle Range value: 625, ..., 5173
\item \strong{ST128SV} \emph{numeric} Volume (WM Parcellation) of WMHypoIntensities Range value: 1069, ..., 62310
\item \strong{ST129CV} \emph{numeric} Volume (Cortical Parcellation) of LeftInsula Range value: 2719, ..., 7888
\item \strong{ST129SA} \emph{numeric} Surface Area of LeftInsula Range value: 1499, ..., 3014
\item \strong{ST129TA} \emph{numeric} Cortical Thickness Average of LeftInsula Range value: 1.502, ..., 3.333
\item \strong{ST129TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInsula Range value: 0.455, ..., 1.466
\item \strong{ST12SV} \emph{numeric} Volume (WM Parcellation) of LeftAmygdala Range value: 343, ..., 1613
\item \strong{ST130CV} \emph{numeric} Volume (Cortical Parcellation) of RightInsula Range value: 2858, ..., 7632
\item \strong{ST130SA} \emph{numeric} Surface Area of RightInsula Range value: 1383, ..., 3069
\item \strong{ST130TA} \emph{numeric} Cortical Thickness Average of RightInsula Range value: 1.456, ..., 3.575
\item \strong{ST130TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInsula Range value: 0.406, ..., 1.357
\item \strong{ST13CV} \emph{numeric} Volume (Cortical Parcellation) of LeftBankssts Range value: 874, ..., 6014
\item \strong{ST13SA} \emph{numeric} Surface Area of LeftBankssts Range value: 509, ..., 2650
\item \strong{ST13TA} \emph{numeric} Cortical Thickness Average of LeftBankssts Range value: 1.29, ..., 3.155
\item \strong{ST13TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftBankssts Range value: 0.293, ..., 1.341
\item \strong{ST14CV} \emph{numeric} Volume (Cortical Parcellation) of LeftCaudalAnteriorCingulate Range value: 478, ..., 3148
\item \strong{ST14SA} \emph{numeric} Surface Area of LeftCaudalAnteriorCingulate Range value: 349, ..., 1402
\item \strong{ST14TA} \emph{numeric} Cortical Thickness Average of LeftCaudalAnteriorCingulate Range value: 1.448, ..., 3.746
\item \strong{ST14TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCaudalAnteriorCingulate Range value: 0.385, ..., 1.489
\item \strong{ST15CV} \emph{numeric} Volume (Cortical Parcellation) of LeftCaudalMiddleFrontal Range value: 2296, ..., 9379
\item \strong{ST15SA} \emph{numeric} Surface Area of LeftCaudalMiddleFrontal Range value: 1202, ..., 4094
\item \strong{ST15TA} \emph{numeric} Cortical Thickness Average of LeftCaudalMiddleFrontal Range value: 1.267, ..., 3.345
\item \strong{ST15TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCaudalMiddleFrontal Range value: 0.259, ..., 1.444
\item \strong{ST16SV} \emph{numeric} Volume (WM Parcellation) of LeftCaudate Range value: 1359, ..., 7198
\item \strong{ST17SV} \emph{numeric} Volume (WM Parcellation) of LeftCerebellumCortex Range value: 20849, ..., 62978
\item \strong{ST18SV} \emph{numeric} Volume (WM Parcellation) of LeftCerebellumWM Range value: 10791, ..., 51951
\item \strong{ST19SV} \emph{numeric} Volume (WM Parcellation) of LeftCerebralCortex Range value: 96144, ..., 236449
\item \strong{ST1SV} \emph{numeric} Volume (WM Parcellation) of Brainstem Range value: 13778, ..., 33268
\item \strong{ST20SV} \emph{numeric} Volume (WM Parcellation) of LeftCerebralWM Range value: 161817, ..., 382583
\item \strong{ST21SV} \emph{numeric} Volume (WM Parcellation) of LeftChoroidPlexus Range value: 1024, ..., 5720
\item \strong{ST22CV} \emph{logical} Volume (Cortical Parcellation) of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22SA} \emph{logical} Surface Area of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22TA} \emph{logical} Cortical Thickness Average of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22TS} \emph{logical} Cortical Thickness Standard Deviation of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST23CV} \emph{numeric} Volume (Cortical Parcellation) of LeftCuneus Range value: 818, ..., 4761
\item \strong{ST23SA} \emph{numeric} Surface Area of LeftCuneus Range value: 495, ..., 2342
\item \strong{ST23TA} \emph{numeric} Cortical Thickness Average of LeftCuneus Range value: 1.049, ..., 2.669
\item \strong{ST23TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCuneus Range value: 0.284, ..., 1.321
\item \strong{ST24CV} \emph{numeric} Volume (Cortical Parcellation) of LeftEntorhinal Range value: 488, ..., 3753
\item \strong{ST24SA} \emph{numeric} Surface Area of LeftEntorhinal Range value: 138, ..., 843
\item \strong{ST24TA} \emph{numeric} Cortical Thickness Average of LeftEntorhinal Range value: 1.348, ..., 4.287
\item \strong{ST24TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftEntorhinal Range value: 0.461, ..., 1.437
\item \strong{ST25CV} \emph{numeric} Volume (Cortical Parcellation) of LeftFrontalPole Range value: 156, ..., 1172
\item \strong{ST25SA} \emph{numeric} Surface Area of LeftFrontalPole Range value: 53, ..., 337
\item \strong{ST25TA} \emph{numeric} Cortical Thickness Average of LeftFrontalPole Range value: 1.254, ..., 3.785
\item \strong{ST25TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftFrontalPole Range value: 0.227, ..., 1.081
\item \strong{ST26CV} \emph{numeric} Volume (Cortical Parcellation) of LeftFusiform Range value: 3402, ..., 12965
\item \strong{ST26SA} \emph{numeric} Surface Area of LeftFusiform Range value: 1632, ..., 5521
\item \strong{ST26TA} \emph{numeric} Cortical Thickness Average of LeftFusiform Range value: 1.484, ..., 3.081
\item \strong{ST26TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftFusiform Range value: 0.434, ..., 1.188
\item \strong{ST27SA} \emph{numeric} Surface Area of LeftHemisphere Range value: 53766, ..., 122415
\item \strong{ST28CV} \emph{logical} Volume (Cortical Parcellation) of LeftHemisphereWM A Boolean value: TRUE/ FALSE
\item \strong{ST29SV} \emph{numeric} Volume (WM Parcellation) of LeftHippocampus Range value: 1345, ..., 5330
\item \strong{ST2SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumAnterior Range value: 29, ..., 2343
\item \strong{ST30SV} \emph{numeric} Volume (WM Parcellation) of LeftInferiorLateralVentricle Range value: 23, ..., 7994
\item \strong{ST31CV} \emph{numeric} Volume (Cortical Parcellation) of LeftInferiorParietal Range value: 4309, ..., 17500
\item \strong{ST31SA} \emph{numeric} Surface Area of LeftInferiorParietal Range value: 1936, ..., 7806
\item \strong{ST31TA} \emph{numeric} Cortical Thickness Average of LeftInferiorParietal Range value: 1.347, ..., 3.117
\item \strong{ST31TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInferiorParietal Range value: 0.296, ..., 1.343
\item \strong{ST32CV} \emph{numeric} Volume (Cortical Parcellation) of LeftInferiorTemporal Range value: 3918, ..., 14817
\item \strong{ST32SA} \emph{numeric} Surface Area of LeftInferiorTemporal Range value: 1479, ..., 5635
\item \strong{ST32TA} \emph{numeric} Cortical Thickness Average of LeftInferiorTemporal Range value: 1.556, ..., 3.255
\item \strong{ST32TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInferiorTemporal Range value: 0.453, ..., 1.377
\item \strong{ST33SV} \emph{logical} Volume (WM Parcellation) of LeftInterior A Boolean value: TRUE/ FALSE
\item \strong{ST34CV} \emph{numeric} Volume (Cortical Parcellation) of LeftIsthmusCingulate Range value: 1111, ..., 3432
\item \strong{ST34SA} \emph{numeric} Surface Area of LeftIsthmusCingulate Range value: 437, ..., 1625
\item \strong{ST34TA} \emph{numeric} Cortical Thickness Average of LeftIsthmusCingulate Range value: 1.426, ..., 3.291
\item \strong{ST34TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftIsthmusCingulate Range value: 0.417, ..., 1.914
\item \strong{ST35CV} \emph{numeric} Volume (Cortical Parcellation) of LeftLateralOccipital Range value: 5102, ..., 16149
\item \strong{ST35SA} \emph{numeric} Surface Area of LeftLateralOccipital Range value: 2539, ..., 7768
\item \strong{ST35TA} \emph{numeric} Cortical Thickness Average of LeftLateralOccipital Range value: 1.291, ..., 2.782
\item \strong{ST35TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLateralOccipital Range value: 0.281, ..., 1.334
\item \strong{ST36CV} \emph{numeric} Volume (Cortical Parcellation) of LeftLateralOrbitofrontal Range value: 2866, ..., 8842
\item \strong{ST36SA} \emph{numeric} Surface Area of LeftLateralOrbitofrontal Range value: 1271, ..., 3795
\item \strong{ST36TA} \emph{numeric} Cortical Thickness Average of LeftLateralOrbitofrontal Range value: 1.637, ..., 3.467
\item \strong{ST36TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLateralOrbitofrontal Range value: 0.453, ..., 1.259
\item \strong{ST37SV} \emph{numeric} Volume (WM Parcellation) of LeftLateralVentricle Range value: 3052, ..., 88479
\item \strong{ST38CV} \emph{numeric} Volume (Cortical Parcellation) of LeftLingual Range value: 2357, ..., 9036
\item \strong{ST38SA} \emph{numeric} Surface Area of LeftLingual Range value: 1567, ..., 4962
\item \strong{ST38TA} \emph{numeric} Cortical Thickness Average of LeftLingual Range value: 1.363, ..., 3.074
\item \strong{ST38TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLingual Range value: 0.373, ..., 1.134
\item \strong{ST39CV} \emph{numeric} Volume (Cortical Parcellation) of LeftMedialOrbitofrontal Range value: 1982, ..., 6219
\item \strong{ST39SA} \emph{numeric} Surface Area of LeftMedialOrbitofrontal Range value: 684, ..., 2750
\item \strong{ST39TA} \emph{numeric} Cortical Thickness Average of LeftMedialOrbitofrontal Range value: 1.408, ..., 3.12
\item \strong{ST39TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftMedialOrbitofrontal Range value: 0.382, ..., 1.299
\item \strong{ST3SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumCentral Range value: 66, ..., 3700
\item \strong{ST40CV} \emph{numeric} Volume (Cortical Parcellation) of LeftMiddleTemporal Range value: 3444, ..., 14083
\item \strong{ST40SA} \emph{numeric} Surface Area of LeftMiddleTemporal Range value: 1571, ..., 5461
\item \strong{ST40TA} \emph{numeric} Cortical Thickness Average of LeftMiddleTemporal Range value: 1.451, ..., 3.131
\item \strong{ST40TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftMiddleTemporal Range value: 0.357, ..., 1.317
\item \strong{ST41SV} \emph{logical} Volume (WM Parcellation) of LeftNonWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST42SV} \emph{numeric} Volume (WM Parcellation) of LeftPallidum Range value: 434, ..., 2868
\item \strong{ST43CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParacentral Range value: 1262, ..., 5636
\item \strong{ST43SA} \emph{numeric} Surface Area of LeftParacentral Range value: 839, ..., 2255
\item \strong{ST43TA} \emph{numeric} Cortical Thickness Average of LeftParacentral Range value: 1.303, ..., 3.429
\item \strong{ST43TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParacentral Range value: 0.335, ..., 1.374
\item \strong{ST44CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParahippocampal Range value: 806, ..., 3043
\item \strong{ST44SA} \emph{numeric} Surface Area of LeftParahippocampal Range value: 355, ..., 1259
\item \strong{ST44TA} \emph{numeric} Cortical Thickness Average of LeftParahippocampal Range value: 1.295, ..., 3.74
\item \strong{ST44TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParahippocampal Range value: 0.408, ..., 1.492
\item \strong{ST45CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParsOpercularis Range value: 1739, ..., 7951
\item \strong{ST45SA} \emph{numeric} Surface Area of LeftParsOpercularis Range value: 775, ..., 3693
\item \strong{ST45TA} \emph{numeric} Cortical Thickness Average of LeftParsOpercularis Range value: 1.351, ..., 2.982
\item \strong{ST45TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsOpercularis Range value: 0.254, ..., 1.295
\item \strong{ST46CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParsOrbitalis Range value: 754, ..., 3013
\item \strong{ST46SA} \emph{numeric} Surface Area of LeftParsOrbitalis Range value: 215, ..., 1154
\item \strong{ST46TA} \emph{numeric} Cortical Thickness Average of LeftParsOrbitalis Range value: 1.571, ..., 3.826
\item \strong{ST46TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsOrbitalis Range value: 0.371, ..., 1.3
\item \strong{ST47CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParsTriangularis Range value: 1071, ..., 5452
\item \strong{ST47SA} \emph{numeric} Surface Area of LeftParsTriangularis Range value: 540, ..., 2234
\item \strong{ST47TA} \emph{numeric} Cortical Thickness Average of LeftParsTriangularis Range value: 1.346, ..., 3.477
\item \strong{ST47TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsTriangularis Range value: 0.267, ..., 1.213
\item \strong{ST48CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPericalcarine Range value: 712, ..., 3906
\item \strong{ST48SA} \emph{numeric} Surface Area of LeftPericalcarine Range value: 631, ..., 2015
\item \strong{ST48TA} \emph{numeric} Cortical Thickness Average of LeftPericalcarine Range value: 1.04, ..., 2.869
\item \strong{ST48TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPericalcarine Range value: 0.192, ..., 1.198
\item \strong{ST49CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPostcentral Range value: 5033, ..., 14648
\item \strong{ST49SA} \emph{numeric} Surface Area of LeftPostcentral Range value: 2527, ..., 8112
\item \strong{ST49TA} \emph{numeric} Cortical Thickness Average of LeftPostcentral Range value: 1.268, ..., 2.814
\item \strong{ST49TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPostcentral Range value: 0.318, ..., 1.28
\item \strong{ST4SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumMidAnterior Range value: 101, ..., 3804
\item \strong{ST50CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPosteriorCingulate Range value: 1566, ..., 4828
\item \strong{ST50SA} \emph{numeric} Surface Area of LeftPosteriorCingulate Range value: 745, ..., 2095
\item \strong{ST50TA} \emph{numeric} Cortical Thickness Average of LeftPosteriorCingulate Range value: 1.414, ..., 3.233
\item \strong{ST50TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPosteriorCingulate Range value: 0.407, ..., 1.652
\item \strong{ST51CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPrecentral Range value: 5693, ..., 19234
\item \strong{ST51SA} \emph{numeric} Surface Area of LeftPrecentral Range value: 2806, ..., 7554
\item \strong{ST51TA} \emph{numeric} Cortical Thickness Average of LeftPrecentral Range value: 1.373, ..., 3.171
\item \strong{ST51TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPrecentral Range value: 0.314, ..., 1.391
\item \strong{ST52CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPrecuneus Range value: 4166, ..., 12683
\item \strong{ST52SA} \emph{numeric} Surface Area of LeftPrecuneus Range value: 2350, ..., 5436
\item \strong{ST52TA} \emph{numeric} Cortical Thickness Average of LeftPrecuneus Range value: 1.344, ..., 2.658
\item \strong{ST52TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPrecuneus Range value: 0.325, ..., 1.299
\item \strong{ST53SV} \emph{numeric} Volume (WM Parcellation) of LeftPutamen Range value: 1200, ..., 7116
\item \strong{ST54CV} \emph{numeric} Volume (Cortical Parcellation) of LeftRostralAnteriorCingulate Range value: 786, ..., 3771
\item \strong{ST54SA} \emph{numeric} Surface Area of LeftRostralAnteriorCingulate Range value: 302, ..., 1361
\item \strong{ST54TA} \emph{numeric} Cortical Thickness Average of LeftRostralAnteriorCingulate Range value: 1.459, ..., 3.901
\item \strong{ST54TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftRostralAnteriorCingulate Range value: 0.374, ..., 1.391
\item \strong{ST55CV} \emph{numeric} Volume (Cortical Parcellation) of LeftRostralMiddleFrontal Range value: 5192, ..., 17414
\item \strong{ST55SA} \emph{numeric} Surface Area of LeftRostralMiddleFrontal Range value: 2623, ..., 8667
\item \strong{ST55TA} \emph{numeric} Cortical Thickness Average of LeftRostralMiddleFrontal Range value: 1.242, ..., 3.207
\item \strong{ST55TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftRostralMiddleFrontal Range value: 0.282, ..., 1.2
\item \strong{ST56CV} \emph{numeric} Volume (Cortical Parcellation) of LeftSuperiorFrontal Range value: 7074, ..., 28034
\item \strong{ST56SA} \emph{numeric} Surface Area of LeftSuperiorFrontal Range value: 4189, ..., 10329
\item \strong{ST56TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorFrontal Range value: 1.434, ..., 3.346
\item \strong{ST56TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorFrontal Range value: 0.351, ..., 1.327
\item \strong{ST57CV} \emph{numeric} Volume (Cortical Parcellation) of LeftSuperiorParietal Range value: 6009, ..., 17166
\item \strong{ST57SA} \emph{numeric} Surface Area of LeftSuperiorParietal Range value: 2802, ..., 7685
\item \strong{ST57TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorParietal Range value: 1.189, ..., 3.096
\item \strong{ST57TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorParietal Range value: 0.256, ..., 1.355
\item \strong{ST58CV} \emph{numeric} Volume (Cortical Parcellation) of LeftSuperiorTemporal Range value: 4132, ..., 18177
\item \strong{ST58SA} \emph{numeric} Surface Area of LeftSuperiorTemporal Range value: 2406, ..., 6198
\item \strong{ST58TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorTemporal Range value: 1.391, ..., 2.954
\item \strong{ST58TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorTemporal Range value: 0.377, ..., 1.355
\item \strong{ST59CV} \emph{numeric} Volume (Cortical Parcellation) of LeftSupramarginal Range value: 4481, ..., 13506
\item \strong{ST59SA} \emph{numeric} Surface Area of LeftSupramarginal Range value: 2020, ..., 6240
\item \strong{ST59TA} \emph{numeric} Cortical Thickness Average of LeftSupramarginal Range value: 1.38, ..., 3.176
\item \strong{ST59TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSupramarginal Range value: 0.359, ..., 1.344
\item \strong{ST5SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumMidPosterior Range value: 43, ..., 2648
\item \strong{ST60CV} \emph{numeric} Volume (Cortical Parcellation) of LeftTemporalPole Range value: 491, ..., 3722
\item \strong{ST60SA} \emph{numeric} Surface Area of LeftTemporalPole Range value: 146, ..., 999
\item \strong{ST60TA} \emph{numeric} Cortical Thickness Average of LeftTemporalPole Range value: 1.29, ..., 4.396
\item \strong{ST60TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftTemporalPole Range value: 0.269, ..., 1.278
\item \strong{ST61SV} \emph{numeric} Volume (WM Parcellation) of LeftThalamus Range value: 4053, ..., 10823
\item \strong{ST62CV} \emph{numeric} Volume (Cortical Parcellation) of LeftTransverseTemporal Range value: 352, ..., 1810
\item \strong{ST62SA} \emph{numeric} Surface Area of LeftTransverseTemporal Range value: 223, ..., 1042
\item \strong{ST62TA} \emph{numeric} Cortical Thickness Average of LeftTransverseTemporal Range value: 1.133, ..., 3.536
\item \strong{ST62TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftTransverseTemporal Range value: 0.2, ..., 1.608
\item \strong{ST63SV} \emph{logical} Volume (WM Parcellation) of LeftUndetermined A Boolean value: TRUE/ FALSE
\item \strong{ST64CV} \emph{numeric} Volume (Cortical Parcellation) of LeftUnknown Range value: 3389, ..., 9489
\item \strong{ST64SA} \emph{numeric} Surface Area of LeftUnknown Range value: 4207, ..., 8682
\item \strong{ST64TA} \emph{numeric} Cortical Thickness Average of LeftUnknown Range value: 0.59, ..., 1.49
\item \strong{ST64TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftUnknown Range value: 0.857, ..., 1.72
\item \strong{ST65SV} \emph{numeric} Volume (WM Parcellation) of LeftVentralDC Range value: 2086, ..., 7611
\item \strong{ST66SV} \emph{numeric} Volume (WM Parcellation) of LeftVessel Range value: 2, ..., 490
\item \strong{ST67SV} \emph{logical} Volume (WM Parcellation) of LeftWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST68SV} \emph{numeric} Volume (WM Parcellation) of NonWMHypoIntensities Range value: 1, ..., 1231
\item \strong{ST69SV} \emph{numeric} Volume (WM Parcellation) of OpticChiasm Range value: 78, ..., 687
\item \strong{ST6SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumPosterior Range value: 54, ..., 1484
\item \strong{ST70SV} \emph{numeric} Volume (WM Parcellation) of RightAccumbensArea Range value: 41, ..., 776
\item \strong{ST71SV} \emph{numeric} Volume (WM Parcellation) of RightAmygdala Range value: 161, ..., 1971
\item \strong{ST72CV} \emph{numeric} Volume (Cortical Parcellation) of RightBankssts Range value: 759, ..., 4078
\item \strong{ST72SA} \emph{numeric} Surface Area of RightBankssts Range value: 512, ..., 1966
\item \strong{ST72TA} \emph{numeric} Cortical Thickness Average of RightBankssts Range value: 1.369, ..., 2.939
\item \strong{ST72TS} \emph{numeric} Cortical Thickness Standard Deviation of RightBankssts Range value: 0.265, ..., 1.338
\item \strong{ST73CV} \emph{numeric} Volume (Cortical Parcellation) of RightCaudalAnteriorCingulate Range value: 3, ..., 3553
\item \strong{ST73SA} \emph{numeric} Surface Area of RightCaudalAnteriorCingulate Range value: 402, ..., 1590
\item \strong{ST73TA} \emph{numeric} Cortical Thickness Average of RightCaudalAnteriorCingulate Range value: 0.002, ..., 3.762
\item \strong{ST73TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCaudalAnteriorCingulate Range value: 0.067, ..., 1.401
\item \strong{ST74CV} \emph{numeric} Volume (Cortical Parcellation) of RightCaudalMiddleFrontal Range value: 1966, ..., 10695
\item \strong{ST74SA} \emph{numeric} Surface Area of RightCaudalMiddleFrontal Range value: 1086, ..., 4684
\item \strong{ST74TA} \emph{numeric} Cortical Thickness Average of RightCaudalMiddleFrontal Range value: 1.321, ..., 3.5
\item \strong{ST74TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCaudalMiddleFrontal Range value: 0.32, ..., 1.587
\item \strong{ST75SV} \emph{numeric} Volume (WM Parcellation) of RightCaudate Range value: 1131, ..., 6378
\item \strong{ST76SV} \emph{numeric} Volume (WM Parcellation) of RightCerebellumCortex Range value: 22794, ..., 51265
\item \strong{ST77SV} \emph{numeric} Volume (WM Parcellation) of RightCerebellumWM Range value: 10755, ..., 53115
\item \strong{ST78SV} \emph{numeric} Volume (WM Parcellation) of RightCerebralCortex Range value: 103732, ..., 254274
\item \strong{ST79SV} \emph{numeric} Volume (WM Parcellation) of RightCerebralWM Range value: 156938, ..., 388359
\item \strong{ST7SV} \emph{numeric} Volume (WM Parcellation) of Csf Range value: 662, ..., 4346
\item \strong{ST80SV} \emph{numeric} Volume (WM Parcellation) of RightChoroidPlexus Range value: 1225, ..., 5050
\item \strong{ST81CV} \emph{logical} Volume (Cortical Parcellation) of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81SA} \emph{logical} Surface Area of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81TA} \emph{logical} Cortical Thickness Average of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81TS} \emph{logical} Cortical Thickness Standard Deviation of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST82CV} \emph{numeric} Volume (Cortical Parcellation) of RightCuneus Range value: 734, ..., 5266
\item \strong{ST82SA} \emph{numeric} Surface Area of RightCuneus Range value: 491, ..., 2416
\item \strong{ST82TA} \emph{numeric} Cortical Thickness Average of RightCuneus Range value: 1.277, ..., 2.817
\item \strong{ST82TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCuneus Range value: 0.264, ..., 1.402
\item \strong{ST83CV} \emph{numeric} Volume (Cortical Parcellation) of RightEntorhinal Range value: 238, ..., 3382
\item \strong{ST83SA} \emph{numeric} Surface Area of RightEntorhinal Range value: 135, ..., 850
\item \strong{ST83TA} \emph{numeric} Cortical Thickness Average of RightEntorhinal Range value: 1.047, ..., 4.542
\item \strong{ST83TS} \emph{numeric} Cortical Thickness Standard Deviation of RightEntorhinal Range value: 0.431, ..., 1.881
\item \strong{ST84CV} \emph{numeric} Volume (Cortical Parcellation) of RightFrontalPole Range value: 198, ..., 1443
\item \strong{ST84SA} \emph{numeric} Surface Area of RightFrontalPole Range value: 64, ..., 499
\item \strong{ST84TA} \emph{numeric} Cortical Thickness Average of RightFrontalPole Range value: 1.407, ..., 3.944
\item \strong{ST84TS} \emph{numeric} Cortical Thickness Standard Deviation of RightFrontalPole Range value: 0.228, ..., 1.244
\item \strong{ST85CV} \emph{numeric} Volume (Cortical Parcellation) of RightFusiform Range value: 3281, ..., 13506
\item \strong{ST85SA} \emph{numeric} Surface Area of RightFusiform Range value: 1717, ..., 5368
\item \strong{ST85TA} \emph{numeric} Cortical Thickness Average of RightFusiform Range value: 1.494, ..., 3.349
\item \strong{ST85TS} \emph{numeric} Cortical Thickness Standard Deviation of RightFusiform Range value: 0.434, ..., 1.335
\item \strong{ST86SA} \emph{numeric} Surface Area of RightHemisphere Range value: 61188.2, ..., 127391
\item \strong{ST87CV} \emph{logical} Volume (Cortical Parcellation) of RightHemisphereWM A Boolean value: TRUE/ FALSE
\item \strong{ST88SV} \emph{numeric} Volume (WM Parcellation) of RightHippocampus Range value: 636, ..., 5338
\item \strong{ST89SV} \emph{numeric} Volume (WM Parcellation) of RightInferiorLateralVentricle Range value: 2, ..., 9831
\item \strong{ST8SV} \emph{numeric} Volume (WM Parcellation) of FifthVentricle Range value: 1, ..., 206
\item \strong{ST90CV} \emph{numeric} Volume (Cortical Parcellation) of RightInferiorParietal Range value: 5859, ..., 18632
\item \strong{ST90SA} \emph{numeric} Surface Area of RightInferiorParietal Range value: 2665, ..., 8571
\item \strong{ST90TA} \emph{numeric} Cortical Thickness Average of RightInferiorParietal Range value: 1.367, ..., 2.749
\item \strong{ST90TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInferiorParietal Range value: 0.329, ..., 1.335
\item \strong{ST91CV} \emph{numeric} Volume (Cortical Parcellation) of RightInferiorTemporal Range value: 3651, ..., 15180
\item \strong{ST91SA} \emph{numeric} Surface Area of RightInferiorTemporal Range value: 1393, ..., 5767
\item \strong{ST91TA} \emph{numeric} Cortical Thickness Average of RightInferiorTemporal Range value: 1.611, ..., 3.37
\item \strong{ST91TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInferiorTemporal Range value: 0.488, ..., 1.302
\item \strong{ST92SV} \emph{logical} Volume (WM Parcellation) of RightInterior A Boolean value: TRUE/ FALSE
\item \strong{ST93CV} \emph{numeric} Volume (Cortical Parcellation) of RightIsthmusCingulate Range value: 968, ..., 3445
\item \strong{ST93SA} \emph{numeric} Surface Area of RightIsthmusCingulate Range value: 453, ..., 1425
\item \strong{ST93TA} \emph{numeric} Cortical Thickness Average of RightIsthmusCingulate Range value: 1.178, ..., 3.187
\item \strong{ST93TS} \emph{numeric} Cortical Thickness Standard Deviation of RightIsthmusCingulate Range value: 0.405, ..., 2.018
\item \strong{ST94CV} \emph{numeric} Volume (Cortical Parcellation) of RightLateralOccipital Range value: 4733, ..., 16735
\item \strong{ST94SA} \emph{numeric} Surface Area of RightLateralOccipital Range value: 2594, ..., 7463
\item \strong{ST94TA} \emph{numeric} Cortical Thickness Average of RightLateralOccipital Range value: 1.356, ..., 2.938
\item \strong{ST94TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLateralOccipital Range value: 0.346, ..., 1.268
\item \strong{ST95CV} \emph{numeric} Volume (Cortical Parcellation) of RightLateralOrbitofrontal Range value: 2656, ..., 9057
\item \strong{ST95SA} \emph{numeric} Surface Area of RightLateralOrbitofrontal Range value: 1554, ..., 3834
\item \strong{ST95TA} \emph{numeric} Cortical Thickness Average of RightLateralOrbitofrontal Range value: 1.652, ..., 3.339
\item \strong{ST95TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLateralOrbitofrontal Range value: 0.494, ..., 1.981
\item \strong{ST96SV} \emph{numeric} Volume (WM Parcellation) of RightLateralVentricle Range value: 2216, ..., 72583
\item \strong{ST97CV} \emph{numeric} Volume (Cortical Parcellation) of RightLingual Range value: 2937, ..., 9109
\item \strong{ST97SA} \emph{numeric} Surface Area of RightLingual Range value: 1644, ..., 4516
\item \strong{ST97TA} \emph{numeric} Cortical Thickness Average of RightLingual Range value: 1.244, ..., 2.824
\item \strong{ST97TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLingual Range value: 0.387, ..., 1.448
\item \strong{ST98CV} \emph{numeric} Volume (Cortical Parcellation) of RightMedialOrbitofrontal Range value: 240, ..., 7096
\item \strong{ST98SA} \emph{numeric} Surface Area of RightMedialOrbitofrontal Range value: 940, ..., 2899
\item \strong{ST98TA} \emph{numeric} Cortical Thickness Average of RightMedialOrbitofrontal Range value: 0.234, ..., 2.992
\item \strong{ST98TS} \emph{numeric} Cortical Thickness Standard Deviation of RightMedialOrbitofrontal Range value: 0.395, ..., 1.209
\item \strong{ST99CV} \emph{numeric} Volume (Cortical Parcellation) of RightMiddleTemporal Range value: 4496, ..., 15374
\item \strong{ST99SA} \emph{numeric} Surface Area of RightMiddleTemporal Range value: 1635, ..., 5419
\item \strong{ST99TA} \emph{numeric} Cortical Thickness Average of RightMiddleTemporal Range value: 1.557, ..., 3.271
\item \strong{ST99TS} \emph{numeric} Cortical Thickness Standard Deviation of RightMiddleTemporal Range value: 0.439, ..., 1.305
\item \strong{ST9SV} \emph{numeric} Volume (WM Parcellation) of FourthVentricle Range value: 505, ..., 6098
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
