% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFFSL51}
\alias{UCSFFSL51}
\title{Longitudinal FreeSurfer (5.1) - Final Run w/ All A: 2022-03-01}
\format{
A data frame with 3311 observations and 365 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFFSL51)
}
\description{
Longitudinal FreeSurfer (5.1) - Final Run w/ All A: 2022-03-01 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI2, ADNIGO
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 5289
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{ }   Date: YYY-MM-DD
\item \strong{LONISID} \emph{numeric} The LONI study UID for the image processed Range value: 27508, ..., 107491
\item \strong{LONIUID} \emph{numeric} The LONI series UID for the image processed. Range value: 88111, ..., 531719
\item \strong{IMAGEUID} \emph{numeric} The ImageUID for the image processed. Range value: 221959, ..., 1051225
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   complete/partial
\item \strong{OVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. Pass/Fail/Partial
\item \strong{TEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) Pass/Fail
\item \strong{FRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) Pass/Fail
\item \strong{PARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) Pass/Fail
\item \strong{INSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) Pass/Fail
\item \strong{OCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) Pass/Fail
\item \strong{BGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) Pass/Fail
\item \strong{CWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87) Pass/Fail
\item \strong{VENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) Pass/Fail
\item \strong{LHIPQC} \emph{logical} QC rating of the left hippocampal subfields. Fail affects the following regions: Left CA1 (ST131HS); Left CA2-3 (ST132HS); Left CA4-DG (ST133HS); Left Fimbria (ST134HS); Left Hippocampal Fissure (ST135HS); Left Presubiculum (ST136HS); Left Subiculum (ST137HS); Left Tail (ST138HS) Pass/Fail
\item \strong{RHIPQC} \emph{logical} QC rating of the right hippocampal subfields. Fail affects the following regions: Right CA1 (ST139HS); Right CA2-3 (ST140HS); Right CA4-DG (ST141HS); Right Fimbria (ST142HS); Right Hippocampal Fissure (ST143HS); Right Presubiculum (ST144HS); Right Subiculum (ST145HS); Right Tail (ST146HS) Pass/Fail
\item \strong{ST149SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalGM Range value: 245258, ..., 564040
\item \strong{ST28SA} \emph{numeric} Surface Area (aparc.stats) of LeftHemisphereWM Range value: 61404.3, ..., 106255
\item \strong{ST87SA} \emph{numeric} Surface Area (aparc.stats) of RightHemisphereWM Range value: 66707.5, ..., 113275
\item \strong{ST1SV} \emph{numeric} Subcortical Volume (aseg.stats) of Brainstem Range value: 13925, ..., 28323
\item \strong{ST2SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumAnterior Range value: 121, ..., 1289
\item \strong{ST3SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumCentral Range value: 111, ..., 627
\item \strong{ST4SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidAnterior Range value: 93, ..., 618
\item \strong{ST5SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidPosterior Range value: 137, ..., 588
\item \strong{ST6SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumPosterior Range value: 213, ..., 1461
\item \strong{ST7SV} \emph{numeric} Subcortical Volume (aseg.stats) of Csf Range value: 697, ..., 3359
\item \strong{ST8SV} \emph{numeric} Subcortical Volume (aseg.stats) of FifthVentricle Range value: 1, ..., 143
\item \strong{ST9SV} \emph{numeric} Subcortical Volume (aseg.stats) of FourthVentricle Range value: 666, ..., 6233
\item \strong{ST10CV} \emph{numeric} Cortical Volume (aparc.stats) of Icv Range value: 1142800, ..., 2080860
\item \strong{ST11SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAccumbensArea Range value: 147, ..., 874
\item \strong{ST12SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAmygdala Range value: 443, ..., 2175
\item \strong{ST13SA} \emph{numeric} Surface Area (aparc.stats) of LeftBankssts Range value: 554, ..., 1714
\item \strong{ST13TA} \emph{numeric} Thickness Average (aparc.stats) of LeftBankssts Range value: 1.45, ..., 3.093
\item \strong{ST13TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftBankssts Range value: 0.301, ..., 0.896
\item \strong{ST13CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftBankssts Range value: 954, ..., 4238
\item \strong{ST14SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 344, ..., 1235
\item \strong{ST14TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 1.676, ..., 3.831
\item \strong{ST14TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 0.287, ..., 1.272
\item \strong{ST14CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 649, ..., 3455
\item \strong{ST15SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1278, ..., 3488
\item \strong{ST15TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1.516, ..., 2.844
\item \strong{ST15TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalMiddleFrontal Range value: 0.381, ..., 0.908
\item \strong{ST15CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalMiddleFrontal Range value: 2620, ..., 9583
\item \strong{ST16SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCaudate Range value: 2141, ..., 7096
\item \strong{ST17SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumCortex Range value: 31923, ..., 63755
\item \strong{ST18SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumWM Range value: 5651, ..., 20283
\item \strong{ST21SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftChoroidPlexus Range value: 1099, ..., 3977
\item \strong{ST23SA} \emph{numeric} Surface Area (aparc.stats) of LeftCuneus Range value: 691, ..., 2159
\item \strong{ST23TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCuneus Range value: 1.337, ..., 2.498
\item \strong{ST23TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCuneus Range value: 0.317, ..., 1.059
\item \strong{ST23CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCuneus Range value: 1549, ..., 4343
\item \strong{ST24SA} \emph{numeric} Surface Area (aparc.stats) of LeftEntorhinal Range value: 209, ..., 877
\item \strong{ST24TA} \emph{numeric} Thickness Average (aparc.stats) of LeftEntorhinal Range value: 1.413, ..., 4.465
\item \strong{ST24TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftEntorhinal Range value: 0.409, ..., 1.385
\item \strong{ST24CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftEntorhinal Range value: 607, ..., 3344
\item \strong{ST25SA} \emph{numeric} Surface Area (aparc.stats) of LeftFrontalPole Range value: 75, ..., 355
\item \strong{ST25TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFrontalPole Range value: 1.514, ..., 3.663
\item \strong{ST25TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFrontalPole Range value: 0.248, ..., 1.006
\item \strong{ST25CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFrontalPole Range value: 324, ..., 1380
\item \strong{ST26SA} \emph{numeric} Surface Area (aparc.stats) of LeftFusiform Range value: 1962, ..., 4677
\item \strong{ST26TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFusiform Range value: 1.499, ..., 3.294
\item \strong{ST26TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFusiform Range value: 0.454, ..., 1.009
\item \strong{ST26CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFusiform Range value: 4171, ..., 15079
\item \strong{ST29SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftHippocampus Range value: 1091, ..., 5069
\item \strong{ST30SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftInferiorLateralVentricle Range value: 7, ..., 7991
\item \strong{ST31SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorParietal Range value: 2603, ..., 6179
\item \strong{ST31TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorParietal Range value: 1.418, ..., 2.885
\item \strong{ST31TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorParietal Range value: 0.348, ..., 0.925
\item \strong{ST31CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorParietal Range value: 4962, ..., 16531
\item \strong{ST32SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorTemporal Range value: 1747, ..., 4881
\item \strong{ST32TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorTemporal Range value: 1.613, ..., 3.424
\item \strong{ST32TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorTemporal Range value: 0.464, ..., 1.115
\item \strong{ST32CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorTemporal Range value: 4001, ..., 15915
\item \strong{ST34SA} \emph{numeric} Surface Area (aparc.stats) of LeftIsthmusCingulate Range value: 567, ..., 1637
\item \strong{ST34TA} \emph{numeric} Thickness Average (aparc.stats) of LeftIsthmusCingulate Range value: 1.746, ..., 3.333
\item \strong{ST34TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftIsthmusCingulate Range value: 0.465, ..., 1.276
\item \strong{ST34CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftIsthmusCingulate Range value: 1125, ..., 3960
\item \strong{ST35SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOccipital Range value: 3183, ..., 6422
\item \strong{ST35TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOccipital Range value: 1.49, ..., 2.751
\item \strong{ST35TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOccipital Range value: 0.402, ..., 0.907
\item \strong{ST35CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOccipital Range value: 5849, ..., 17839
\item \strong{ST36SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOrbitofrontal Range value: 1560, ..., 3684
\item \strong{ST36TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOrbitofrontal Range value: 1.769, ..., 3.282
\item \strong{ST36TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOrbitofrontal Range value: 0.485, ..., 1.033
\item \strong{ST36CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOrbitofrontal Range value: 4250, ..., 10516
\item \strong{ST37SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftLateralVentricle Range value: 2878, ..., 70095
\item \strong{ST38SA} \emph{numeric} Surface Area (aparc.stats) of LeftLingual Range value: 1763, ..., 4416
\item \strong{ST38TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLingual Range value: 1.313, ..., 2.459
\item \strong{ST38TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLingual Range value: 0.367, ..., 0.824
\item \strong{ST38CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLingual Range value: 2853, ..., 10293
\item \strong{ST39SA} \emph{numeric} Surface Area (aparc.stats) of LeftMedialOrbitofrontal Range value: 1148, ..., 2766
\item \strong{ST39TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMedialOrbitofrontal Range value: 1.599, ..., 3.206
\item \strong{ST39TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMedialOrbitofrontal Range value: 0.488, ..., 1.166
\item \strong{ST39CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMedialOrbitofrontal Range value: 3010, ..., 7315
\item \strong{ST40SA} \emph{numeric} Surface Area (aparc.stats) of LeftMiddleTemporal Range value: 1662, ..., 4661
\item \strong{ST40TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMiddleTemporal Range value: 1.579, ..., 3.294
\item \strong{ST40TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMiddleTemporal Range value: 0.449, ..., 0.995
\item \strong{ST40CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMiddleTemporal Range value: 3845, ..., 14136
\item \strong{ST42SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPallidum Range value: 582, ..., 2231
\item \strong{ST43SA} \emph{numeric} Surface Area (aparc.stats) of LeftParacentral Range value: 900, ..., 1969
\item \strong{ST43TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParacentral Range value: 1.571, ..., 2.87
\item \strong{ST43TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParacentral Range value: 0.399, ..., 0.917
\item \strong{ST43CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParacentral Range value: 1586, ..., 5365
\item \strong{ST44SA} \emph{numeric} Surface Area (aparc.stats) of LeftParahippocampal Range value: 413, ..., 1250
\item \strong{ST44TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParahippocampal Range value: 1.429, ..., 3.743
\item \strong{ST44TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParahippocampal Range value: 0.427, ..., 1.273
\item \strong{ST44CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParahippocampal Range value: 649, ..., 3549
\item \strong{ST45SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOpercularis Range value: 996, ..., 2750
\item \strong{ST45TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOpercularis Range value: 1.694, ..., 2.991
\item \strong{ST45TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOpercularis Range value: 0.349, ..., 0.829
\item \strong{ST45CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOpercularis Range value: 1958, ..., 7210
\item \strong{ST46SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOrbitalis Range value: 348, ..., 886
\item \strong{ST46TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOrbitalis Range value: 1.675, ..., 3.585
\item \strong{ST46TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOrbitalis Range value: 0.388, ..., 1.068
\item \strong{ST46CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOrbitalis Range value: 876, ..., 3222
\item \strong{ST47SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsTriangularis Range value: 710, ..., 1893
\item \strong{ST47TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsTriangularis Range value: 1.627, ..., 2.902
\item \strong{ST47TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsTriangularis Range value: 0.345, ..., 0.915
\item \strong{ST47CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsTriangularis Range value: 1436, ..., 5728
\item \strong{ST48SA} \emph{numeric} Surface Area (aparc.stats) of LeftPericalcarine Range value: 733, ..., 1950
\item \strong{ST48TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPericalcarine Range value: 1.159, ..., 2.185
\item \strong{ST48TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPericalcarine Range value: 0.228, ..., 1.21
\item \strong{ST48CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPericalcarine Range value: 1018, ..., 3214
\item \strong{ST49SA} \emph{numeric} Surface Area (aparc.stats) of LeftPostcentral Range value: 2988, ..., 5982
\item \strong{ST49TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPostcentral Range value: 1.458, ..., 2.435
\item \strong{ST49TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPostcentral Range value: 0.428, ..., 0.987
\item \strong{ST49CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPostcentral Range value: 5462, ..., 14173
\item \strong{ST50SA} \emph{numeric} Surface Area (aparc.stats) of LeftPosteriorCingulate Range value: 695, ..., 1825
\item \strong{ST50TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPosteriorCingulate Range value: 1.854, ..., 3.226
\item \strong{ST50TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPosteriorCingulate Range value: 0.417, ..., 1.073
\item \strong{ST50CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPosteriorCingulate Range value: 1742, ..., 4828
\item \strong{ST51SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecentral Range value: 3206, ..., 7210
\item \strong{ST51TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecentral Range value: 1.686, ..., 2.869
\item \strong{ST51TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecentral Range value: 0.477, ..., 0.918
\item \strong{ST51CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecentral Range value: 6978, ..., 19166
\item \strong{ST52SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecuneus Range value: 2495, ..., 5158
\item \strong{ST52TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecuneus Range value: 1.46, ..., 2.764
\item \strong{ST52TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecuneus Range value: 0.416, ..., 0.88
\item \strong{ST52CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecuneus Range value: 4198, ..., 12358
\item \strong{ST53SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPutamen Range value: 2746, ..., 11340
\item \strong{ST54SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralAnteriorCingulate Range value: 320, ..., 1296
\item \strong{ST54TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralAnteriorCingulate Range value: 1.982, ..., 3.944
\item \strong{ST54TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralAnteriorCingulate Range value: 0.418, ..., 1.372
\item \strong{ST54CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralAnteriorCingulate Range value: 1209, ..., 3955
\item \strong{ST55SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralMiddleFrontal Range value: 2865, ..., 8458
\item \strong{ST55TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralMiddleFrontal Range value: 1.584, ..., 2.701
\item \strong{ST55TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralMiddleFrontal Range value: 0.444, ..., 0.919
\item \strong{ST55CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralMiddleFrontal Range value: 7224, ..., 20623
\item \strong{ST56SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorFrontal Range value: 4773, ..., 10003
\item \strong{ST56TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorFrontal Range value: 1.742, ..., 3.17
\item \strong{ST56TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorFrontal Range value: 0.456, ..., 0.932
\item \strong{ST56CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorFrontal Range value: 10671, ..., 26370
\item \strong{ST57SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorParietal Range value: 3615, ..., 7051
\item \strong{ST57TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorParietal Range value: 1.378, ..., 2.667
\item \strong{ST57TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorParietal Range value: 0.359, ..., 0.971
\item \strong{ST57CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorParietal Range value: 5708, ..., 17852
\item \strong{ST58SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorTemporal Range value: 2592, ..., 4895
\item \strong{ST58TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorTemporal Range value: 1.575, ..., 3.151
\item \strong{ST58TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorTemporal Range value: 0.5, ..., 0.948
\item \strong{ST58CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorTemporal Range value: 5426, ..., 15168
\item \strong{ST59SA} \emph{numeric} Surface Area (aparc.stats) of LeftSupramarginal Range value: 2430, ..., 5390
\item \strong{ST59TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSupramarginal Range value: 1.495, ..., 2.877
\item \strong{ST59TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSupramarginal Range value: 0.411, ..., 0.92
\item \strong{ST59CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSupramarginal Range value: 4843, ..., 14726
\item \strong{ST60SA} \emph{numeric} Surface Area (aparc.stats) of LeftTemporalPole Range value: 259, ..., 880
\item \strong{ST60TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTemporalPole Range value: 1.341, ..., 4.684
\item \strong{ST60TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTemporalPole Range value: 0.298, ..., 1.254
\item \strong{ST60CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTemporalPole Range value: 685, ..., 3678
\item \strong{ST61SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftThalamus Range value: 4279, ..., 8913
\item \strong{ST62SA} \emph{numeric} Surface Area (aparc.stats) of LeftTransverseTemporal Range value: 259, ..., 747
\item \strong{ST62TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTransverseTemporal Range value: 1.363, ..., 2.96
\item \strong{ST62TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTransverseTemporal Range value: 0.243, ..., 0.743
\item \strong{ST62CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTransverseTemporal Range value: 494, ..., 1879
\item \strong{ST65SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVentralDC Range value: 2398, ..., 4926
\item \strong{ST66SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVessel Range value: 3, ..., 729
\item \strong{ST68SV} \emph{numeric} Subcortical Volume (aseg.stats) of NonWMHypoIntensities Range value: 1, ..., 1715
\item \strong{ST69SV} \emph{numeric} Subcortical Volume (aseg.stats) of OpticChiasm Range value: 64, ..., 666
\item \strong{ST70SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAccumbensArea Range value: 133, ..., 887
\item \strong{ST71SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAmygdala Range value: 458, ..., 2345
\item \strong{ST72SA} \emph{numeric} Surface Area (aparc.stats) of RightBankssts Range value: 528, ..., 1458
\item \strong{ST72TA} \emph{numeric} Thickness Average (aparc.stats) of RightBankssts Range value: 1.577, ..., 3.012
\item \strong{ST72TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightBankssts Range value: 0.308, ..., 0.839
\item \strong{ST72CV} \emph{numeric} Cortical Volume (aparc.stats) of RightBankssts Range value: 959, ..., 3542
\item \strong{ST73SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalAnteriorCingulate Range value: 367, ..., 1419
\item \strong{ST73TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalAnteriorCingulate Range value: 1.848, ..., 3.649
\item \strong{ST73TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalAnteriorCingulate Range value: 0.342, ..., 1.235
\item \strong{ST73CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalAnteriorCingulate Range value: 722, ..., 3613
\item \strong{ST74SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalMiddleFrontal Range value: 990, ..., 3284
\item \strong{ST74TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalMiddleFrontal Range value: 1.641, ..., 2.888
\item \strong{ST74TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalMiddleFrontal Range value: 0.409, ..., 0.875
\item \strong{ST74CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalMiddleFrontal Range value: 2206, ..., 9628
\item \strong{ST75SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCaudate Range value: 2147, ..., 7553
\item \strong{ST76SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumCortex Range value: 32213, ..., 67160
\item \strong{ST77SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumWM Range value: 6038, ..., 20448
\item \strong{ST80SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightChoroidPlexus Range value: 1210, ..., 4357
\item \strong{ST82SA} \emph{numeric} Surface Area (aparc.stats) of RightCuneus Range value: 680, ..., 2174
\item \strong{ST82TA} \emph{numeric} Thickness Average (aparc.stats) of RightCuneus Range value: 1.335, ..., 2.529
\item \strong{ST82TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCuneus Range value: 0.278, ..., 0.907
\item \strong{ST82CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCuneus Range value: 1132, ..., 4604
\item \strong{ST83SA} \emph{numeric} Surface Area (aparc.stats) of RightEntorhinal Range value: 171, ..., 882
\item \strong{ST83TA} \emph{numeric} Thickness Average (aparc.stats) of RightEntorhinal Range value: 1.439, ..., 4.599
\item \strong{ST83TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightEntorhinal Range value: 0.354, ..., 1.392
\item \strong{ST83CV} \emph{numeric} Cortical Volume (aparc.stats) of RightEntorhinal Range value: 309, ..., 3725
\item \strong{ST84SA} \emph{numeric} Surface Area (aparc.stats) of RightFrontalPole Range value: 138, ..., 581
\item \strong{ST84TA} \emph{numeric} Thickness Average (aparc.stats) of RightFrontalPole Range value: 1.619, ..., 3.422
\item \strong{ST84TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFrontalPole Range value: 0.313, ..., 1.106
\item \strong{ST84CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFrontalPole Range value: 479, ..., 1756
\item \strong{ST85SA} \emph{numeric} Surface Area (aparc.stats) of RightFusiform Range value: 1976, ..., 4656
\item \strong{ST85TA} \emph{numeric} Thickness Average (aparc.stats) of RightFusiform Range value: 1.425, ..., 3.228
\item \strong{ST85TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFusiform Range value: 0.462, ..., 0.973
\item \strong{ST85CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFusiform Range value: 3475, ..., 14353
\item \strong{ST88SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightHippocampus Range value: 1259, ..., 5195
\item \strong{ST89SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightInferiorLateralVentricle Range value: 10, ..., 6571
\item \strong{ST90SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorParietal Range value: 2946, ..., 7333
\item \strong{ST90TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorParietal Range value: 1.445, ..., 2.889
\item \strong{ST90TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorParietal Range value: 0.344, ..., 0.969
\item \strong{ST90CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorParietal Range value: 5885, ..., 21822
\item \strong{ST91SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorTemporal Range value: 1857, ..., 4686
\item \strong{ST91TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorTemporal Range value: 1.575, ..., 3.381
\item \strong{ST91TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorTemporal Range value: 0.455, ..., 1.048
\item \strong{ST91CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorTemporal Range value: 3435, ..., 15525
\item \strong{ST93SA} \emph{numeric} Surface Area (aparc.stats) of RightIsthmusCingulate Range value: 462, ..., 1503
\item \strong{ST93TA} \emph{numeric} Thickness Average (aparc.stats) of RightIsthmusCingulate Range value: 1.549, ..., 3.218
\item \strong{ST93TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightIsthmusCingulate Range value: 0.486, ..., 1.17
\item \strong{ST93CV} \emph{numeric} Cortical Volume (aparc.stats) of RightIsthmusCingulate Range value: 1109, ..., 3898
\item \strong{ST94SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOccipital Range value: 2810, ..., 6419
\item \strong{ST94TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOccipital Range value: 1.363, ..., 2.772
\item \strong{ST94TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOccipital Range value: 0.365, ..., 0.85
\item \strong{ST94CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOccipital Range value: 5798, ..., 16591
\item \strong{ST95SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOrbitofrontal Range value: 1620, ..., 3767
\item \strong{ST95TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOrbitofrontal Range value: 1.823, ..., 3.216
\item \strong{ST95TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOrbitofrontal Range value: 0.552, ..., 1.098
\item \strong{ST95CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOrbitofrontal Range value: 4461, ..., 9627
\item \strong{ST96SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightLateralVentricle Range value: 2372, ..., 67939
\item \strong{ST97SA} \emph{numeric} Surface Area (aparc.stats) of RightLingual Range value: 2003, ..., 4251
\item \strong{ST97TA} \emph{numeric} Thickness Average (aparc.stats) of RightLingual Range value: 1.471, ..., 2.49
\item \strong{ST97TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLingual Range value: 0.43, ..., 0.856
\item \strong{ST97CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLingual Range value: 3666, ..., 9606
\item \strong{ST98SA} \emph{numeric} Surface Area (aparc.stats) of RightMedialOrbitofrontal Range value: 1160, ..., 2515
\item \strong{ST98TA} \emph{numeric} Thickness Average (aparc.stats) of RightMedialOrbitofrontal Range value: 1.595, ..., 3.392
\item \strong{ST98TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMedialOrbitofrontal Range value: 0.49, ..., 1.246
\item \strong{ST98CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMedialOrbitofrontal Range value: 2609, ..., 7081
\item \strong{ST99SA} \emph{numeric} Surface Area (aparc.stats) of RightMiddleTemporal Range value: 2058, ..., 4845
\item \strong{ST99TA} \emph{numeric} Thickness Average (aparc.stats) of RightMiddleTemporal Range value: 1.626, ..., 3.409
\item \strong{ST99TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMiddleTemporal Range value: 0.494, ..., 1.019
\item \strong{ST99CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMiddleTemporal Range value: 4836, ..., 17127
\item \strong{ST101SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPallidum Range value: 831, ..., 2201
\item \strong{ST102SA} \emph{numeric} Surface Area (aparc.stats) of RightParacentral Range value: 1013, ..., 2394
\item \strong{ST102TA} \emph{numeric} Thickness Average (aparc.stats) of RightParacentral Range value: 1.592, ..., 3.037
\item \strong{ST102TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParacentral Range value: 0.424, ..., 1.033
\item \strong{ST102CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParacentral Range value: 1924, ..., 5681
\item \strong{ST103SA} \emph{numeric} Surface Area (aparc.stats) of RightParahippocampal Range value: 405, ..., 1249
\item \strong{ST103TA} \emph{numeric} Thickness Average (aparc.stats) of RightParahippocampal Range value: 1.359, ..., 3.577
\item \strong{ST103TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParahippocampal Range value: 0.345, ..., 1.207
\item \strong{ST103CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParahippocampal Range value: 668, ..., 3479
\item \strong{ST104SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOpercularis Range value: 820, ..., 2139
\item \strong{ST104TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOpercularis Range value: 1.732, ..., 2.972
\item \strong{ST104TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOpercularis Range value: 0.35, ..., 0.78
\item \strong{ST104CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOpercularis Range value: 2132, ..., 5814
\item \strong{ST105SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOrbitalis Range value: 440, ..., 1210
\item \strong{ST105TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOrbitalis Range value: 1.694, ..., 3.487
\item \strong{ST105TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOrbitalis Range value: 0.414, ..., 1.05
\item \strong{ST105CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOrbitalis Range value: 1289, ..., 3478
\item \strong{ST106SA} \emph{numeric} Surface Area (aparc.stats) of RightParsTriangularis Range value: 781, ..., 2253
\item \strong{ST106TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsTriangularis Range value: 1.664, ..., 2.778
\item \strong{ST106TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsTriangularis Range value: 0.361, ..., 0.929
\item \strong{ST106CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsTriangularis Range value: 1994, ..., 5980
\item \strong{ST107SA} \emph{numeric} Surface Area (aparc.stats) of RightPericalcarine Range value: 775, ..., 2299
\item \strong{ST107TA} \emph{numeric} Thickness Average (aparc.stats) of RightPericalcarine Range value: 1.142, ..., 2.033
\item \strong{ST107TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPericalcarine Range value: 0.222, ..., 0.8
\item \strong{ST107CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPericalcarine Range value: 1024, ..., 3576
\item \strong{ST108SA} \emph{numeric} Surface Area (aparc.stats) of RightPostcentral Range value: 2639, ..., 5932
\item \strong{ST108TA} \emph{numeric} Thickness Average (aparc.stats) of RightPostcentral Range value: 1.405, ..., 2.378
\item \strong{ST108TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPostcentral Range value: 0.448, ..., 0.901
\item \strong{ST108CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPostcentral Range value: 4623, ..., 13923
\item \strong{ST109SA} \emph{numeric} Surface Area (aparc.stats) of RightPosteriorCingulate Range value: 693, ..., 1746
\item \strong{ST109TA} \emph{numeric} Thickness Average (aparc.stats) of RightPosteriorCingulate Range value: 1.743, ..., 3.162
\item \strong{ST109TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPosteriorCingulate Range value: 0.419, ..., 1.046
\item \strong{ST109CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPosteriorCingulate Range value: 1587, ..., 4761
\item \strong{ST110SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecentral Range value: 3364, ..., 6928
\item \strong{ST110TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecentral Range value: 1.621, ..., 2.836
\item \strong{ST110TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecentral Range value: 0.461, ..., 0.863
\item \strong{ST110CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecentral Range value: 7085, ..., 18994
\item \strong{ST111SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecuneus Range value: 2302, ..., 5438
\item \strong{ST111TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecuneus Range value: 1.446, ..., 2.774
\item \strong{ST111TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecuneus Range value: 0.423, ..., 0.859
\item \strong{ST111CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecuneus Range value: 4592, ..., 12757
\item \strong{ST112SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPutamen Range value: 2829, ..., 10600
\item \strong{ST113SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralAnteriorCingulate Range value: 291, ..., 1131
\item \strong{ST113TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralAnteriorCingulate Range value: 1.952, ..., 3.812
\item \strong{ST113TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralAnteriorCingulate Range value: 0.451, ..., 1.356
\item \strong{ST113CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralAnteriorCingulate Range value: 762, ..., 3638
\item \strong{ST114SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralMiddleFrontal Range value: 3702, ..., 8547
\item \strong{ST114TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralMiddleFrontal Range value: 1.572, ..., 2.621
\item \strong{ST114TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralMiddleFrontal Range value: 0.426, ..., 0.871
\item \strong{ST114CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralMiddleFrontal Range value: 8116, ..., 21638
\item \strong{ST115SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorFrontal Range value: 4501, ..., 9436
\item \strong{ST115TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorFrontal Range value: 1.699, ..., 2.951
\item \strong{ST115TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorFrontal Range value: 0.478, ..., 0.91
\item \strong{ST115CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorFrontal Range value: 9823, ..., 26615
\item \strong{ST116SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorParietal Range value: 3382, ..., 7433
\item \strong{ST116TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorParietal Range value: 1.314, ..., 2.655
\item \strong{ST116TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorParietal Range value: 0.274, ..., 1.293
\item \strong{ST116CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorParietal Range value: 6870, ..., 18860
\item \strong{ST117SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorTemporal Range value: 2187, ..., 4619
\item \strong{ST117TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorTemporal Range value: 1.528, ..., 3.27
\item \strong{ST117TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorTemporal Range value: 0.51, ..., 1.072
\item \strong{ST117CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorTemporal Range value: 5022, ..., 15729
\item \strong{ST118SA} \emph{numeric} Surface Area (aparc.stats) of RightSupramarginal Range value: 2155, ..., 5383
\item \strong{ST118TA} \emph{numeric} Thickness Average (aparc.stats) of RightSupramarginal Range value: 1.567, ..., 2.842
\item \strong{ST118TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSupramarginal Range value: 0.437, ..., 0.885
\item \strong{ST118CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSupramarginal Range value: 4939, ..., 15352
\item \strong{ST119SA} \emph{numeric} Surface Area (aparc.stats) of RightTemporalPole Range value: 202, ..., 794
\item \strong{ST119TA} \emph{numeric} Thickness Average (aparc.stats) of RightTemporalPole Range value: 1.238, ..., 4.548
\item \strong{ST119TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTemporalPole Range value: 0.25, ..., 1.379
\item \strong{ST119CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTemporalPole Range value: 537, ..., 3378
\item \strong{ST120SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightThalamus Range value: 3939, ..., 9132
\item \strong{ST121SA} \emph{numeric} Surface Area (aparc.stats) of RightTransverseTemporal Range value: 166, ..., 520
\item \strong{ST121TA} \emph{numeric} Thickness Average (aparc.stats) of RightTransverseTemporal Range value: 1.308, ..., 3.006
\item \strong{ST121TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTransverseTemporal Range value: 0.205, ..., 0.747
\item \strong{ST121CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTransverseTemporal Range value: 345, ..., 1534
\item \strong{ST124SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVentralDC Range value: 2273, ..., 4832
\item \strong{ST125SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVessel Range value: 1, ..., 538
\item \strong{ST127SV} \emph{numeric} Subcortical Volume (aseg.stats) of ThirdVentricle Range value: 554, ..., 4000
\item \strong{ST128SV} \emph{numeric} Subcortical Volume (aseg.stats) of WMHypoIntensities Range value: 592, ..., 65957
\item \strong{ST129SA} \emph{numeric} Surface Area (aparc.stats) of LeftInsula Range value: 1523, ..., 3499
\item \strong{ST129TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInsula Range value: 2.025, ..., 3.491
\item \strong{ST129TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInsula Range value: 0.647, ..., 1.223
\item \strong{ST129CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInsula Range value: 3363, ..., 9817
\item \strong{ST130SA} \emph{numeric} Surface Area (aparc.stats) of RightInsula Range value: 1339, ..., 3225
\item \strong{ST130TA} \emph{numeric} Thickness Average (aparc.stats) of RightInsula Range value: 1.811, ..., 3.527
\item \strong{ST130TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInsula Range value: 0.602, ..., 1.218
\item \strong{ST130CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInsula Range value: 3324, ..., 9548
\item \strong{ST131HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA1 Range value: 132.161, ..., 514.916
\item \strong{ST132HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA2_3 Range value: 420.858, ..., 1301.25
\item \strong{ST133HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA4_DG Range value: 231.625, ..., 750.445
\item \strong{ST134HS} \emph{numeric} Hippocampal Subfields Volume of LeftFimbria Range value: 0.211275, ..., 141.111
\item \strong{ST135HS} \emph{numeric} Hippocampal Subfields Volume of LeftHippocampalFissure Range value: 4.55686, ..., 140.262
\item \strong{ST136HS} \emph{numeric} Hippocampal Subfields Volume of LeftPresubiculum Range value: 145.654, ..., 645.895
\item \strong{ST137HS} \emph{numeric} Hippocampal Subfields Volume of LeftSubiculum Range value: 230.404, ..., 844.06
\item \strong{ST138HS} \emph{numeric} Hippocampal Subfields Volume of LeftTail Range value: 119.786, ..., 505.331
\item \strong{ST139HS} \emph{numeric} Hippocampal Subfields Volume of RightCA1 Range value: 153.32, ..., 507.932
\item \strong{ST140HS} \emph{numeric} Hippocampal Subfields Volume of RightCA2_3 Range value: 474.562, ..., 1456
\item \strong{ST141HS} \emph{numeric} Hippocampal Subfields Volume of RightCA4_DG Range value: 245.044, ..., 789.64
\item \strong{ST142HS} \emph{numeric} Hippocampal Subfields Volume of RightFimbria Range value: 0.0745125, ..., 116.526
\item \strong{ST143HS} \emph{numeric} Hippocampal Subfields Volume of RightHippocampalFissure Range value: 5.849, ..., 184.616
\item \strong{ST144HS} \emph{numeric} Hippocampal Subfields Volume of RightPresubiculum Range value: 167.327, ..., 642.021
\item \strong{ST145HS} \emph{numeric} Hippocampal Subfields Volume of RightSubiculum Range value: 214.024, ..., 858.434
\item \strong{ST146HS} \emph{numeric} Hippocampal Subfields Volume of RightTail Range value: 141.966, ..., 503.163
\item \strong{ST147SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalGM Range value: 122991, ..., 279301
\item \strong{ST148SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalGM Range value: 121579, ..., 284739
\item \strong{ST150SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalWM Range value: 28532.3, ..., 339190
\item \strong{ST151SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalWM Range value: 143892, ..., 343271
\item \strong{ST152SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalWM Range value: 258858, ..., 677568
\item \strong{ST153SV} \emph{numeric} Subcortical Volume (aseg.stats) of SubcorticalGM Range value: 117181, ..., 215070
\item \strong{ST154SV} \emph{numeric} Subcortical Volume (aseg.stats) of TotalGM Range value: 362439, ..., 776500
\item \strong{ST155SV} \emph{numeric} Subcortical Volume (aseg.stats) of SupraTentorial Range value: 749726, ..., 1396280
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adnigo}
\keyword{imaging}
