% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFFSL51Y1}
\alias{UCSFFSL51Y1}
\title{Longitudinal FreeSurfer (5.1) - Year 1 Base Image: 2016-08-01}
\format{
A data frame with 1356 observations and 365 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFFSL51Y1)
}
\description{
Longitudinal FreeSurfer (5.1) - Year 1 Base Image: 2016-08-01 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNIGO, ADNI2, nv
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 4795
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code , with options: m60, m72, m48, m06, m12, scmri, m03, nv
\item \strong{EXAMDATE} \emph{ } The scan date for the image processed Date: YYY-MM-DD
\item \strong{VERSION} \emph{ }   Date: YYY-MM-DD
\item \strong{LONISID} \emph{numeric} The LONI study UID for the image processed Range value: 27744, ..., 58989
\item \strong{LONIUID} \emph{numeric} The LONI series UID for the image processed. Range value: 88859, ..., 191072
\item \strong{IMAGEUID} \emph{numeric} The ImageUID for the image processed. Range value: 222683, ..., 375918
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   complete/partial
\item \strong{OVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. Pass/Fail/Partial
\item \strong{TEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) Pass/Fail
\item \strong{FRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) Pass/Fail
\item \strong{PARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) Pass/Fail
\item \strong{INSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) Pass/Fail
\item \strong{OCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) Pass/Fail
\item \strong{BGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) Pass/Fail
\item \strong{CWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87) Pass/Fail
\item \strong{VENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) Pass/Fail
\item \strong{LHIPQC} \emph{logical} QC rating of the left hippocampal subfields. Fail affects the following regions: Left CA1 (ST131HS); Left CA2-3 (ST132HS); Left CA4-DG (ST133HS); Left Fimbria (ST134HS); Left Hippocampal Fissure (ST135HS); Left Presubiculum (ST136HS); Left Subiculum (ST137HS); Left Tail (ST138HS) Pass/Fail
\item \strong{RHIPQC} \emph{logical} QC rating of the right hippocampal subfields. Fail affects the following regions: Right CA1 (ST139HS); Right CA2-3 (ST140HS); Right CA4-DG (ST141HS); Right Fimbria (ST142HS); Right Hippocampal Fissure (ST143HS); Right Presubiculum (ST144HS); Right Subiculum (ST145HS); Right Tail (ST146HS) Pass/Fail
\item \strong{ST149SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalGM Range value: 262521, ..., 559862
\item \strong{ST28SA} \emph{numeric} Surface Area (aparc.stats) of LeftHemisphereWM Range value: 62253.7, ..., 106486
\item \strong{ST87SA} \emph{numeric} Surface Area (aparc.stats) of RightHemisphereWM Range value: 67294.2, ..., 112585
\item \strong{ST1SV} \emph{numeric} Subcortical Volume (aseg.stats) of Brainstem Range value: 14245, ..., 29356
\item \strong{ST2SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumAnterior Range value: 367, ..., 1224
\item \strong{ST3SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumCentral Range value: 142, ..., 585
\item \strong{ST4SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidAnterior Range value: 133, ..., 620
\item \strong{ST5SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidPosterior Range value: 143, ..., 602
\item \strong{ST6SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumPosterior Range value: 285, ..., 1265
\item \strong{ST7SV} \emph{numeric} Subcortical Volume (aseg.stats) of Csf Range value: 729, ..., 3332
\item \strong{ST8SV} \emph{numeric} Subcortical Volume (aseg.stats) of FifthVentricle Range value: 1, ..., 146
\item \strong{ST9SV} \emph{numeric} Subcortical Volume (aseg.stats) of FourthVentricle Range value: 592, ..., 5544
\item \strong{ST10CV} \emph{numeric} Cortical Volume (aparc.stats) of Icv Range value: 1168740, ..., 2073760
\item \strong{ST11SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAccumbensArea Range value: 137, ..., 859
\item \strong{ST12SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAmygdala Range value: 494, ..., 2196
\item \strong{ST13SA} \emph{numeric} Surface Area (aparc.stats) of LeftBankssts Range value: 587, ..., 1513
\item \strong{ST13TA} \emph{numeric} Thickness Average (aparc.stats) of LeftBankssts Range value: 1.611, ..., 3.006
\item \strong{ST13TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftBankssts Range value: 0.275, ..., 0.856
\item \strong{ST13CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftBankssts Range value: 1105, ..., 4238
\item \strong{ST14SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 344, ..., 1117
\item \strong{ST14TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 1.657, ..., 3.831
\item \strong{ST14TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 0.272, ..., 1.122
\item \strong{ST14CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 776, ..., 3206
\item \strong{ST15SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1378, ..., 3372
\item \strong{ST15TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1.626, ..., 2.981
\item \strong{ST15TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalMiddleFrontal Range value: 0.4, ..., 0.865
\item \strong{ST15CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalMiddleFrontal Range value: 3292, ..., 9560
\item \strong{ST16SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCaudate Range value: 2261, ..., 7059
\item \strong{ST17SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumCortex Range value: 32234, ..., 62151
\item \strong{ST18SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumWM Range value: 7184, ..., 21285
\item \strong{ST21SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftChoroidPlexus Range value: 1202, ..., 4084
\item \strong{ST23SA} \emph{numeric} Surface Area (aparc.stats) of LeftCuneus Range value: 691, ..., 1995
\item \strong{ST23TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCuneus Range value: 1.342, ..., 2.498
\item \strong{ST23TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCuneus Range value: 0.324, ..., 1.124
\item \strong{ST23CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCuneus Range value: 1595, ..., 4351
\item \strong{ST24SA} \emph{numeric} Surface Area (aparc.stats) of LeftEntorhinal Range value: 238, ..., 713
\item \strong{ST24TA} \emph{numeric} Thickness Average (aparc.stats) of LeftEntorhinal Range value: 1.6, ..., 4.6
\item \strong{ST24TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftEntorhinal Range value: 0.423, ..., 1.244
\item \strong{ST24CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftEntorhinal Range value: 578, ..., 3302
\item \strong{ST25SA} \emph{numeric} Surface Area (aparc.stats) of LeftFrontalPole Range value: 75, ..., 383
\item \strong{ST25TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFrontalPole Range value: 1.748, ..., 3.361
\item \strong{ST25TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFrontalPole Range value: 0.254, ..., 1.023
\item \strong{ST25CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFrontalPole Range value: 366, ..., 1127
\item \strong{ST26SA} \emph{numeric} Surface Area (aparc.stats) of LeftFusiform Range value: 2088, ..., 4203
\item \strong{ST26TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFusiform Range value: 1.623, ..., 3.212
\item \strong{ST26TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFusiform Range value: 0.501, ..., 0.957
\item \strong{ST26CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFusiform Range value: 4732, ..., 13730
\item \strong{ST29SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftHippocampus Range value: 1259, ..., 5051
\item \strong{ST30SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftInferiorLateralVentricle Range value: 6, ..., 5302
\item \strong{ST31SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorParietal Range value: 2761, ..., 6682
\item \strong{ST31TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorParietal Range value: 1.535, ..., 2.842
\item \strong{ST31TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorParietal Range value: 0.361, ..., 1.207
\item \strong{ST31CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorParietal Range value: 5623, ..., 16714
\item \strong{ST32SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorTemporal Range value: 1885, ..., 4760
\item \strong{ST32TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorTemporal Range value: 1.817, ..., 3.267
\item \strong{ST32TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorTemporal Range value: 0.506, ..., 1.026
\item \strong{ST32CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorTemporal Range value: 5853, ..., 15977
\item \strong{ST34SA} \emph{numeric} Surface Area (aparc.stats) of LeftIsthmusCingulate Range value: 577, ..., 1576
\item \strong{ST34TA} \emph{numeric} Thickness Average (aparc.stats) of LeftIsthmusCingulate Range value: 1.775, ..., 3.286
\item \strong{ST34TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftIsthmusCingulate Range value: 0.548, ..., 1.282
\item \strong{ST34CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftIsthmusCingulate Range value: 1420, ..., 3960
\item \strong{ST35SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOccipital Range value: 3232, ..., 6904
\item \strong{ST35TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOccipital Range value: 1.513, ..., 2.631
\item \strong{ST35TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOccipital Range value: 0.407, ..., 1.036
\item \strong{ST35CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOccipital Range value: 5942, ..., 17548
\item \strong{ST36SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOrbitofrontal Range value: 1802, ..., 3507
\item \strong{ST36TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOrbitofrontal Range value: 1.74, ..., 3.043
\item \strong{ST36TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOrbitofrontal Range value: 0.485, ..., 1.006
\item \strong{ST36CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOrbitofrontal Range value: 4495, ..., 9441
\item \strong{ST37SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftLateralVentricle Range value: 2958, ..., 64376
\item \strong{ST38SA} \emph{numeric} Surface Area (aparc.stats) of LeftLingual Range value: 1978, ..., 4390
\item \strong{ST38TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLingual Range value: 1.526, ..., 2.31
\item \strong{ST38TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLingual Range value: 0.43, ..., 0.843
\item \strong{ST38CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLingual Range value: 3933, ..., 10389
\item \strong{ST39SA} \emph{numeric} Surface Area (aparc.stats) of LeftMedialOrbitofrontal Range value: 1214, ..., 2731
\item \strong{ST39TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMedialOrbitofrontal Range value: 1.696, ..., 2.863
\item \strong{ST39TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMedialOrbitofrontal Range value: 0.556, ..., 1.126
\item \strong{ST39CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMedialOrbitofrontal Range value: 2894, ..., 7225
\item \strong{ST40SA} \emph{numeric} Surface Area (aparc.stats) of LeftMiddleTemporal Range value: 1919, ..., 4427
\item \strong{ST40TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMiddleTemporal Range value: 1.699, ..., 3.217
\item \strong{ST40TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMiddleTemporal Range value: 0.471, ..., 0.936
\item \strong{ST40CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMiddleTemporal Range value: 4979, ..., 14121
\item \strong{ST42SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPallidum Range value: 629, ..., 2305
\item \strong{ST43SA} \emph{numeric} Surface Area (aparc.stats) of LeftParacentral Range value: 893, ..., 1880
\item \strong{ST43TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParacentral Range value: 1.709, ..., 2.728
\item \strong{ST43TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParacentral Range value: 0.411, ..., 1.005
\item \strong{ST43CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParacentral Range value: 2091, ..., 5203
\item \strong{ST44SA} \emph{numeric} Surface Area (aparc.stats) of LeftParahippocampal Range value: 424, ..., 1032
\item \strong{ST44TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParahippocampal Range value: 1.664, ..., 3.754
\item \strong{ST44TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParahippocampal Range value: 0.496, ..., 1.198
\item \strong{ST44CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParahippocampal Range value: 1035, ..., 3360
\item \strong{ST45SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOpercularis Range value: 1021, ..., 2498
\item \strong{ST45TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOpercularis Range value: 1.541, ..., 2.948
\item \strong{ST45TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOpercularis Range value: 0.365, ..., 0.831
\item \strong{ST45CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOpercularis Range value: 2434, ..., 6624
\item \strong{ST46SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOrbitalis Range value: 377, ..., 866
\item \strong{ST46TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOrbitalis Range value: 1.85, ..., 3.331
\item \strong{ST46TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOrbitalis Range value: 0.41, ..., 1.06
\item \strong{ST46CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOrbitalis Range value: 1203, ..., 2805
\item \strong{ST47SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsTriangularis Range value: 739, ..., 1871
\item \strong{ST47TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsTriangularis Range value: 1.784, ..., 2.778
\item \strong{ST47TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsTriangularis Range value: 0.368, ..., 0.839
\item \strong{ST47CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsTriangularis Range value: 1711, ..., 4797
\item \strong{ST48SA} \emph{numeric} Surface Area (aparc.stats) of LeftPericalcarine Range value: 813, ..., 1923
\item \strong{ST48TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPericalcarine Range value: 1.143, ..., 2.024
\item \strong{ST48TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPericalcarine Range value: 0.228, ..., 1.21
\item \strong{ST48CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPericalcarine Range value: 1030, ..., 2980
\item \strong{ST49SA} \emph{numeric} Surface Area (aparc.stats) of LeftPostcentral Range value: 2814, ..., 5812
\item \strong{ST49TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPostcentral Range value: 1.491, ..., 2.404
\item \strong{ST49TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPostcentral Range value: 0.462, ..., 1.053
\item \strong{ST49CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPostcentral Range value: 5718, ..., 14763
\item \strong{ST50SA} \emph{numeric} Surface Area (aparc.stats) of LeftPosteriorCingulate Range value: 709, ..., 1638
\item \strong{ST50TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPosteriorCingulate Range value: 1.734, ..., 3.199
\item \strong{ST50TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPosteriorCingulate Range value: 0.415, ..., 1.007
\item \strong{ST50CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPosteriorCingulate Range value: 1662, ..., 4442
\item \strong{ST51SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecentral Range value: 3218, ..., 7110
\item \strong{ST51TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecentral Range value: 1.741, ..., 2.801
\item \strong{ST51TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecentral Range value: 0.479, ..., 1.242
\item \strong{ST51CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecentral Range value: 7265, ..., 18972
\item \strong{ST52SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecuneus Range value: 2404, ..., 4988
\item \strong{ST52TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecuneus Range value: 1.509, ..., 2.781
\item \strong{ST52TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecuneus Range value: 0.46, ..., 0.857
\item \strong{ST52CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecuneus Range value: 4571, ..., 12023
\item \strong{ST53SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPutamen Range value: 3254, ..., 11243
\item \strong{ST54SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralAnteriorCingulate Range value: 393, ..., 1338
\item \strong{ST54TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralAnteriorCingulate Range value: 1.777, ..., 3.7
\item \strong{ST54TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralAnteriorCingulate Range value: 0.473, ..., 1.348
\item \strong{ST54CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralAnteriorCingulate Range value: 1318, ..., 3993
\item \strong{ST55SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralMiddleFrontal Range value: 3627, ..., 8481
\item \strong{ST55TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralMiddleFrontal Range value: 1.668, ..., 2.621
\item \strong{ST55TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralMiddleFrontal Range value: 0.47, ..., 0.848
\item \strong{ST55CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralMiddleFrontal Range value: 8528, ..., 21414
\item \strong{ST56SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorFrontal Range value: 4872, ..., 8785
\item \strong{ST56TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorFrontal Range value: 1.729, ..., 3.072
\item \strong{ST56TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorFrontal Range value: 0.507, ..., 0.846
\item \strong{ST56CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorFrontal Range value: 11652, ..., 26416
\item \strong{ST57SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorParietal Range value: 3723, ..., 7136
\item \strong{ST57TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorParietal Range value: 1.401, ..., 2.645
\item \strong{ST57TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorParietal Range value: 0.38, ..., 0.996
\item \strong{ST57CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorParietal Range value: 7284, ..., 17703
\item \strong{ST58SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorTemporal Range value: 2598, ..., 4926
\item \strong{ST58TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorTemporal Range value: 1.726, ..., 3.141
\item \strong{ST58TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorTemporal Range value: 0.493, ..., 0.936
\item \strong{ST58CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorTemporal Range value: 6488, ..., 14758
\item \strong{ST59SA} \emph{numeric} Surface Area (aparc.stats) of LeftSupramarginal Range value: 2284, ..., 5208
\item \strong{ST59TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSupramarginal Range value: 1.847, ..., 2.817
\item \strong{ST59TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSupramarginal Range value: 0.441, ..., 0.899
\item \strong{ST59CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSupramarginal Range value: 4933, ..., 14311
\item \strong{ST60SA} \emph{numeric} Surface Area (aparc.stats) of LeftTemporalPole Range value: 240, ..., 824
\item \strong{ST60TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTemporalPole Range value: 1.467, ..., 4.533
\item \strong{ST60TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTemporalPole Range value: 0.343, ..., 1.104
\item \strong{ST60CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTemporalPole Range value: 711, ..., 3752
\item \strong{ST61SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftThalamus Range value: 4546, ..., 8796
\item \strong{ST62SA} \emph{numeric} Surface Area (aparc.stats) of LeftTransverseTemporal Range value: 282, ..., 747
\item \strong{ST62TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTransverseTemporal Range value: 1.396, ..., 2.986
\item \strong{ST62TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTransverseTemporal Range value: 0.269, ..., 0.753
\item \strong{ST62CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTransverseTemporal Range value: 583, ..., 1827
\item \strong{ST65SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVentralDC Range value: 2593, ..., 4963
\item \strong{ST66SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVessel Range value: 16, ..., 652
\item \strong{ST68SV} \emph{numeric} Subcortical Volume (aseg.stats) of NonWMHypoIntensities Range value: 3, ..., 654
\item \strong{ST69SV} \emph{numeric} Subcortical Volume (aseg.stats) of OpticChiasm Range value: 107, ..., 547
\item \strong{ST70SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAccumbensArea Range value: 201, ..., 850
\item \strong{ST71SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAmygdala Range value: 531, ..., 2306
\item \strong{ST72SA} \emph{numeric} Surface Area (aparc.stats) of RightBankssts Range value: 565, ..., 1234
\item \strong{ST72TA} \emph{numeric} Thickness Average (aparc.stats) of RightBankssts Range value: 1.695, ..., 2.979
\item \strong{ST72TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightBankssts Range value: 0.31, ..., 0.759
\item \strong{ST72CV} \emph{numeric} Cortical Volume (aparc.stats) of RightBankssts Range value: 1216, ..., 3369
\item \strong{ST73SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalAnteriorCingulate Range value: 361, ..., 1245
\item \strong{ST73TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalAnteriorCingulate Range value: 1.784, ..., 3.561
\item \strong{ST73TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalAnteriorCingulate Range value: 0.402, ..., 1.118
\item \strong{ST73CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalAnteriorCingulate Range value: 765, ..., 3376
\item \strong{ST74SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalMiddleFrontal Range value: 1070, ..., 3284
\item \strong{ST74TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalMiddleFrontal Range value: 1.71, ..., 3.012
\item \strong{ST74TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalMiddleFrontal Range value: 0.393, ..., 0.882
\item \strong{ST74CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalMiddleFrontal Range value: 2570, ..., 9332
\item \strong{ST75SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCaudate Range value: 2465, ..., 7463
\item \strong{ST76SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumCortex Range value: 32157, ..., 66746
\item \strong{ST77SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumWM Range value: 7441, ..., 20403
\item \strong{ST80SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightChoroidPlexus Range value: 1203, ..., 4166
\item \strong{ST82SA} \emph{numeric} Surface Area (aparc.stats) of RightCuneus Range value: 660, ..., 2172
\item \strong{ST82TA} \emph{numeric} Thickness Average (aparc.stats) of RightCuneus Range value: 1.416, ..., 2.548
\item \strong{ST82TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCuneus Range value: 0.306, ..., 0.878
\item \strong{ST82CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCuneus Range value: 1211, ..., 4644
\item \strong{ST83SA} \emph{numeric} Surface Area (aparc.stats) of RightEntorhinal Range value: 181, ..., 725
\item \strong{ST83TA} \emph{numeric} Thickness Average (aparc.stats) of RightEntorhinal Range value: 1.308, ..., 4.575
\item \strong{ST83TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightEntorhinal Range value: 0.403, ..., 1.359
\item \strong{ST83CV} \emph{numeric} Cortical Volume (aparc.stats) of RightEntorhinal Range value: 517, ..., 3629
\item \strong{ST84SA} \emph{numeric} Surface Area (aparc.stats) of RightFrontalPole Range value: 129, ..., 452
\item \strong{ST84TA} \emph{numeric} Thickness Average (aparc.stats) of RightFrontalPole Range value: 1.564, ..., 3.554
\item \strong{ST84TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFrontalPole Range value: 0.31, ..., 1.044
\item \strong{ST84CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFrontalPole Range value: 501, ..., 1454
\item \strong{ST85SA} \emph{numeric} Surface Area (aparc.stats) of RightFusiform Range value: 1932, ..., 4405
\item \strong{ST85TA} \emph{numeric} Thickness Average (aparc.stats) of RightFusiform Range value: 1.704, ..., 3.362
\item \strong{ST85TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFusiform Range value: 0.544, ..., 0.992
\item \strong{ST85CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFusiform Range value: 4411, ..., 14597
\item \strong{ST88SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightHippocampus Range value: 1575, ..., 5403
\item \strong{ST89SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightInferiorLateralVentricle Range value: 25, ..., 5732
\item \strong{ST90SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorParietal Range value: 3170, ..., 7064
\item \strong{ST90TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorParietal Range value: 1.471, ..., 2.818
\item \strong{ST90TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorParietal Range value: 0.374, ..., 0.875
\item \strong{ST90CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorParietal Range value: 6400, ..., 19256
\item \strong{ST91SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorTemporal Range value: 1912, ..., 4468
\item \strong{ST91TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorTemporal Range value: 1.634, ..., 3.277
\item \strong{ST91TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorTemporal Range value: 0.49, ..., 0.991
\item \strong{ST91CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorTemporal Range value: 3609, ..., 15472
\item \strong{ST93SA} \emph{numeric} Surface Area (aparc.stats) of RightIsthmusCingulate Range value: 515, ..., 1414
\item \strong{ST93TA} \emph{numeric} Thickness Average (aparc.stats) of RightIsthmusCingulate Range value: 1.807, ..., 3.267
\item \strong{ST93TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightIsthmusCingulate Range value: 0.545, ..., 1.159
\item \strong{ST93CV} \emph{numeric} Cortical Volume (aparc.stats) of RightIsthmusCingulate Range value: 1226, ..., 3595
\item \strong{ST94SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOccipital Range value: 2810, ..., 6255
\item \strong{ST94TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOccipital Range value: 1.564, ..., 2.777
\item \strong{ST94TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOccipital Range value: 0.425, ..., 0.848
\item \strong{ST94CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOccipital Range value: 5798, ..., 17340
\item \strong{ST95SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOrbitofrontal Range value: 1727, ..., 3344
\item \strong{ST95TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOrbitofrontal Range value: 1.993, ..., 2.916
\item \strong{ST95TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOrbitofrontal Range value: 0.569, ..., 1.038
\item \strong{ST95CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOrbitofrontal Range value: 4606, ..., 8898
\item \strong{ST96SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightLateralVentricle Range value: 3131, ..., 53151
\item \strong{ST97SA} \emph{numeric} Surface Area (aparc.stats) of RightLingual Range value: 2020, ..., 4479
\item \strong{ST97TA} \emph{numeric} Thickness Average (aparc.stats) of RightLingual Range value: 1.51, ..., 2.427
\item \strong{ST97TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLingual Range value: 0.462, ..., 0.856
\item \strong{ST97CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLingual Range value: 3666, ..., 9421
\item \strong{ST98SA} \emph{numeric} Surface Area (aparc.stats) of RightMedialOrbitofrontal Range value: 1247, ..., 2434
\item \strong{ST98TA} \emph{numeric} Thickness Average (aparc.stats) of RightMedialOrbitofrontal Range value: 1.618, ..., 2.895
\item \strong{ST98TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMedialOrbitofrontal Range value: 0.48, ..., 1.129
\item \strong{ST98CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMedialOrbitofrontal Range value: 2756, ..., 6868
\item \strong{ST99SA} \emph{numeric} Surface Area (aparc.stats) of RightMiddleTemporal Range value: 2097, ..., 4657
\item \strong{ST99TA} \emph{numeric} Thickness Average (aparc.stats) of RightMiddleTemporal Range value: 1.638, ..., 3.415
\item \strong{ST99TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMiddleTemporal Range value: 0.487, ..., 0.974
\item \strong{ST99CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMiddleTemporal Range value: 5044, ..., 15752
\item \strong{ST101SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPallidum Range value: 785, ..., 2092
\item \strong{ST102SA} \emph{numeric} Surface Area (aparc.stats) of RightParacentral Range value: 1044, ..., 2119
\item \strong{ST102TA} \emph{numeric} Thickness Average (aparc.stats) of RightParacentral Range value: 1.7, ..., 2.952
\item \strong{ST102TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParacentral Range value: 0.424, ..., 0.821
\item \strong{ST102CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParacentral Range value: 2176, ..., 5711
\item \strong{ST103SA} \emph{numeric} Surface Area (aparc.stats) of RightParahippocampal Range value: 401, ..., 1105
\item \strong{ST103TA} \emph{numeric} Thickness Average (aparc.stats) of RightParahippocampal Range value: 1.571, ..., 3.561
\item \strong{ST103TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParahippocampal Range value: 0.44, ..., 1.107
\item \strong{ST103CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParahippocampal Range value: 1014, ..., 3164
\item \strong{ST104SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOpercularis Range value: 898, ..., 2139
\item \strong{ST104TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOpercularis Range value: 1.89, ..., 2.882
\item \strong{ST104TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOpercularis Range value: 0.357, ..., 0.835
\item \strong{ST104CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOpercularis Range value: 2096, ..., 5616
\item \strong{ST105SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOrbitalis Range value: 444, ..., 1057
\item \strong{ST105TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOrbitalis Range value: 1.654, ..., 3.374
\item \strong{ST105TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOrbitalis Range value: 0.447, ..., 1.046
\item \strong{ST105CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOrbitalis Range value: 1427, ..., 3476
\item \strong{ST106SA} \emph{numeric} Surface Area (aparc.stats) of RightParsTriangularis Range value: 837, ..., 2229
\item \strong{ST106TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsTriangularis Range value: 1.762, ..., 2.81
\item \strong{ST106TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsTriangularis Range value: 0.369, ..., 0.854
\item \strong{ST106CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsTriangularis Range value: 2153, ..., 6110
\item \strong{ST107SA} \emph{numeric} Surface Area (aparc.stats) of RightPericalcarine Range value: 831, ..., 2148
\item \strong{ST107TA} \emph{numeric} Thickness Average (aparc.stats) of RightPericalcarine Range value: 1.128, ..., 2.009
\item \strong{ST107TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPericalcarine Range value: 0.227, ..., 0.801
\item \strong{ST107CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPericalcarine Range value: 1112, ..., 3685
\item \strong{ST108SA} \emph{numeric} Surface Area (aparc.stats) of RightPostcentral Range value: 2883, ..., 5566
\item \strong{ST108TA} \emph{numeric} Thickness Average (aparc.stats) of RightPostcentral Range value: 1.42, ..., 2.346
\item \strong{ST108TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPostcentral Range value: 0.466, ..., 0.901
\item \strong{ST108CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPostcentral Range value: 4672, ..., 12727
\item \strong{ST109SA} \emph{numeric} Surface Area (aparc.stats) of RightPosteriorCingulate Range value: 715, ..., 1746
\item \strong{ST109TA} \emph{numeric} Thickness Average (aparc.stats) of RightPosteriorCingulate Range value: 1.953, ..., 3.237
\item \strong{ST109TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPosteriorCingulate Range value: 0.419, ..., 0.963
\item \strong{ST109CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPosteriorCingulate Range value: 1699, ..., 4531
\item \strong{ST110SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecentral Range value: 3408, ..., 6866
\item \strong{ST110TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecentral Range value: 1.682, ..., 2.783
\item \strong{ST110TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecentral Range value: 0.476, ..., 0.795
\item \strong{ST110CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecentral Range value: 7293, ..., 18458
\item \strong{ST111SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecuneus Range value: 2549, ..., 5382
\item \strong{ST111TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecuneus Range value: 1.437, ..., 2.714
\item \strong{ST111TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecuneus Range value: 0.424, ..., 0.821
\item \strong{ST111CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecuneus Range value: 5005, ..., 12571
\item \strong{ST112SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPutamen Range value: 3328, ..., 10086
\item \strong{ST113SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralAnteriorCingulate Range value: 337, ..., 1131
\item \strong{ST113TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralAnteriorCingulate Range value: 1.952, ..., 3.816
\item \strong{ST113TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralAnteriorCingulate Range value: 0.46, ..., 1.202
\item \strong{ST113CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralAnteriorCingulate Range value: 762, ..., 3793
\item \strong{ST114SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralMiddleFrontal Range value: 3898, ..., 8521
\item \strong{ST114TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralMiddleFrontal Range value: 1.673, ..., 2.562
\item \strong{ST114TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralMiddleFrontal Range value: 0.458, ..., 0.823
\item \strong{ST114CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralMiddleFrontal Range value: 8719, ..., 21442
\item \strong{ST115SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorFrontal Range value: 4609, ..., 8894
\item \strong{ST115TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorFrontal Range value: 1.855, ..., 3.109
\item \strong{ST115TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorFrontal Range value: 0.506, ..., 0.86
\item \strong{ST115CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorFrontal Range value: 10484, ..., 26367
\item \strong{ST116SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorParietal Range value: 3697, ..., 7257
\item \strong{ST116TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorParietal Range value: 1.364, ..., 2.736
\item \strong{ST116TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorParietal Range value: 0.305, ..., 1.194
\item \strong{ST116CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorParietal Range value: 7736, ..., 17940
\item \strong{ST117SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorTemporal Range value: 2567, ..., 4638
\item \strong{ST117TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorTemporal Range value: 1.55, ..., 3.131
\item \strong{ST117TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorTemporal Range value: 0.521, ..., 0.946
\item \strong{ST117CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorTemporal Range value: 5118, ..., 15725
\item \strong{ST118SA} \emph{numeric} Surface Area (aparc.stats) of RightSupramarginal Range value: 2258, ..., 4880
\item \strong{ST118TA} \emph{numeric} Thickness Average (aparc.stats) of RightSupramarginal Range value: 1.738, ..., 2.799
\item \strong{ST118TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSupramarginal Range value: 0.457, ..., 0.836
\item \strong{ST118CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSupramarginal Range value: 5146, ..., 13666
\item \strong{ST119SA} \emph{numeric} Surface Area (aparc.stats) of RightTemporalPole Range value: 224, ..., 768
\item \strong{ST119TA} \emph{numeric} Thickness Average (aparc.stats) of RightTemporalPole Range value: 1.325, ..., 4.509
\item \strong{ST119TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTemporalPole Range value: 0.297, ..., 1.278
\item \strong{ST119CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTemporalPole Range value: 500, ..., 3322
\item \strong{ST120SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightThalamus Range value: 3988, ..., 8516
\item \strong{ST121SA} \emph{numeric} Surface Area (aparc.stats) of RightTransverseTemporal Range value: 172, ..., 515
\item \strong{ST121TA} \emph{numeric} Thickness Average (aparc.stats) of RightTransverseTemporal Range value: 1.475, ..., 2.988
\item \strong{ST121TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTransverseTemporal Range value: 0.255, ..., 0.67
\item \strong{ST121CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTransverseTemporal Range value: 422, ..., 1339
\item \strong{ST124SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVentralDC Range value: 2382, ..., 4868
\item \strong{ST125SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVessel Range value: 7, ..., 567
\item \strong{ST127SV} \emph{numeric} Subcortical Volume (aseg.stats) of ThirdVentricle Range value: 632, ..., 3829
\item \strong{ST128SV} \emph{numeric} Subcortical Volume (aseg.stats) of WMHypoIntensities Range value: 670, ..., 63420
\item \strong{ST129SA} \emph{numeric} Surface Area (aparc.stats) of LeftInsula Range value: 1603, ..., 3397
\item \strong{ST129TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInsula Range value: 1.639, ..., 3.468
\item \strong{ST129TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInsula Range value: 0.678, ..., 1.205
\item \strong{ST129CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInsula Range value: 4140, ..., 9719
\item \strong{ST130SA} \emph{numeric} Surface Area (aparc.stats) of RightInsula Range value: 1637, ..., 3403
\item \strong{ST130TA} \emph{numeric} Thickness Average (aparc.stats) of RightInsula Range value: 1.87, ..., 3.449
\item \strong{ST130TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInsula Range value: 0.683, ..., 1.226
\item \strong{ST130CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInsula Range value: 3572, ..., 9743
\item \strong{ST131HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA1 Range value: 132.161, ..., 464.3
\item \strong{ST132HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA2_3 Range value: 420.858, ..., 1284.03
\item \strong{ST133HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA4_DG Range value: 234.143, ..., 733.838
\item \strong{ST134HS} \emph{numeric} Hippocampal Subfields Volume of LeftFimbria Range value: 0.243137, ..., 119.274
\item \strong{ST135HS} \emph{numeric} Hippocampal Subfields Volume of LeftHippocampalFissure Range value: 7.15048, ..., 178.526
\item \strong{ST136HS} \emph{numeric} Hippocampal Subfields Volume of LeftPresubiculum Range value: 153.531, ..., 660.662
\item \strong{ST137HS} \emph{numeric} Hippocampal Subfields Volume of LeftSubiculum Range value: 239.68, ..., 819.706
\item \strong{ST138HS} \emph{numeric} Hippocampal Subfields Volume of LeftTail Range value: 134.05, ..., 485.949
\item \strong{ST139HS} \emph{numeric} Hippocampal Subfields Volume of RightCA1 Range value: 182.521, ..., 525.054
\item \strong{ST140HS} \emph{numeric} Hippocampal Subfields Volume of RightCA2_3 Range value: 493.534, ..., 1366.95
\item \strong{ST141HS} \emph{numeric} Hippocampal Subfields Volume of RightCA4_DG Range value: 261.379, ..., 777.229
\item \strong{ST142HS} \emph{numeric} Hippocampal Subfields Volume of RightFimbria Range value: 0.228925, ..., 106.918
\item \strong{ST143HS} \emph{numeric} Hippocampal Subfields Volume of RightHippocampalFissure Range value: 11.2494, ..., 143.77
\item \strong{ST144HS} \emph{numeric} Hippocampal Subfields Volume of RightPresubiculum Range value: 187.956, ..., 690.345
\item \strong{ST145HS} \emph{numeric} Hippocampal Subfields Volume of RightSubiculum Range value: 265.105, ..., 871.319
\item \strong{ST146HS} \emph{numeric} Hippocampal Subfields Volume of RightTail Range value: 143.886, ..., 516.354
\item \strong{ST147SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalGM Range value: 132087, ..., 278897
\item \strong{ST148SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalGM Range value: 130434, ..., 280966
\item \strong{ST150SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalWM Range value: 163792, ..., 368544
\item \strong{ST151SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalWM Range value: 153725, ..., 340260
\item \strong{ST152SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalWM Range value: 318505, ..., 676778
\item \strong{ST153SV} \emph{numeric} Subcortical Volume (aseg.stats) of SubcorticalGM Range value: 118374, ..., 213253
\item \strong{ST154SV} \emph{numeric} Subcortical Volume (aseg.stats) of TotalGM Range value: 380895, ..., 770398
\item \strong{ST155SV} \emph{numeric} Subcortical Volume (aseg.stats) of SupraTentorial Range value: 751447, ..., 1396100
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni2}
\keyword{adnigo}
\keyword{imaging}
