% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFFSX}
\alias{UCSFFSX}
\title{Cross-Sectional FreeSurfer (FreeSurfer Version 4.3): 2015-11-02}
\format{
A data frame with 4486 observations and 367 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFFSX)
}
\description{
Cross-Sectional FreeSurfer (FreeSurfer Version 4.3): 2015-11-02 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1435
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{EXAMDATE} \emph{ } The scan date for the image processed Date: YYY-MM-DD
\item \strong{VERSION} \emph{ }   Date: YYY-MM-DD
\item \strong{LONISID} \emph{numeric} The LONI study UID for the image processed Range value: 1945, ..., 74960
\item \strong{FLDSTRENG} \emph{numeric} The field strength. 1.5 or 3 Range value: 1.5, ..., 1.5
\item \strong{LONIUID} \emph{numeric} The LONI series UID for the image processed. Range value: 8723, ..., 252869
\item \strong{IMAGEUID} \emph{numeric} The ImageUID for the image processed. Range value: 31863, ..., 488736
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   , with options: complete, partial
\item \strong{OVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. Pass/Fail/Partial
\item \strong{TEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) Pass/Fail
\item \strong{FRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) Pass/Fail
\item \strong{PARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) Pass/Fail
\item \strong{INSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) Pass/Fail
\item \strong{OCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) Pass/Fail
\item \strong{BGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) Pass/Fail
\item \strong{CWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79) Pass/Fail
\item \strong{VENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) , with options: Pass, Fail
\item \strong{ST100SV} \emph{logical} Volume (WM Parcellation) of RightNonWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST101SV} \emph{numeric} Volume (WM Parcellation) of RightPallidum Range value: 242, ..., 3040
\item \strong{ST102CV} \emph{numeric} Volume (Cortical Parcellation) of RightParacentral Range value: 1500, ..., 6562
\item \strong{ST102SA} \emph{numeric} Surface Area of RightParacentral Range value: 948, ..., 2863
\item \strong{ST102TA} \emph{numeric} Cortical Thickness Average of RightParacentral Range value: 1.204, ..., 2.712
\item \strong{ST102TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParacentral Range value: 0.279, ..., 0.968
\item \strong{ST103CV} \emph{numeric} Volume (Cortical Parcellation) of RightParahippocampal Range value: 803, ..., 3120
\item \strong{ST103SA} \emph{numeric} Surface Area of RightParahippocampal Range value: 397, ..., 1509
\item \strong{ST103TA} \emph{numeric} Cortical Thickness Average of RightParahippocampal Range value: 0.877, ..., 3.311
\item \strong{ST103TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParahippocampal Range value: 0.462, ..., 1.581
\item \strong{ST104CV} \emph{numeric} Volume (Cortical Parcellation) of RightParsOpercularis Range value: 976, ..., 7712
\item \strong{ST104SA} \emph{numeric} Surface Area of RightParsOpercularis Range value: 679, ..., 2948
\item \strong{ST104TA} \emph{numeric} Cortical Thickness Average of RightParsOpercularis Range value: 1.387, ..., 2.895
\item \strong{ST104TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsOpercularis Range value: 0.379, ..., 0.908
\item \strong{ST105CV} \emph{numeric} Volume (Cortical Parcellation) of RightParsOrbitalis Range value: 860, ..., 3645
\item \strong{ST105SA} \emph{numeric} Surface Area of RightParsOrbitalis Range value: 360, ..., 1253
\item \strong{ST105TA} \emph{numeric} Cortical Thickness Average of RightParsOrbitalis Range value: 1.296, ..., 3.508
\item \strong{ST105TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsOrbitalis Range value: 0.374, ..., 1.063
\item \strong{ST106CV} \emph{numeric} Volume (Cortical Parcellation) of RightParsTriangularis Range value: 894, ..., 6021
\item \strong{ST106SA} \emph{numeric} Surface Area of RightParsTriangularis Range value: 717, ..., 2642
\item \strong{ST106TA} \emph{numeric} Cortical Thickness Average of RightParsTriangularis Range value: 1.208, ..., 2.831
\item \strong{ST106TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsTriangularis Range value: 0.34, ..., 0.9
\item \strong{ST107CV} \emph{numeric} Volume (Cortical Parcellation) of RightPericalcarine Range value: 420, ..., 3644
\item \strong{ST107SA} \emph{numeric} Surface Area of RightPericalcarine Range value: 279, ..., 2513
\item \strong{ST107TA} \emph{numeric} Cortical Thickness Average of RightPericalcarine Range value: 0.951, ..., 2.738
\item \strong{ST107TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPericalcarine Range value: 0.176, ..., 1.227
\item \strong{ST108CV} \emph{numeric} Volume (Cortical Parcellation) of RightPostcentral Range value: 4245, ..., 21441
\item \strong{ST108SA} \emph{numeric} Surface Area of RightPostcentral Range value: 2297, ..., 13873
\item \strong{ST108TA} \emph{numeric} Cortical Thickness Average of RightPostcentral Range value: 1.198, ..., 2.253
\item \strong{ST108TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPostcentral Range value: 0.334, ..., 0.961
\item \strong{ST109CV} \emph{numeric} Volume (Cortical Parcellation) of RightPosteriorCingulate Range value: 856, ..., 4735
\item \strong{ST109SA} \emph{numeric} Surface Area of RightPosteriorCingulate Range value: 740, ..., 2600
\item \strong{ST109TA} \emph{numeric} Cortical Thickness Average of RightPosteriorCingulate Range value: 1.118, ..., 3.3
\item \strong{ST109TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPosteriorCingulate Range value: 0.368, ..., 1.133
\item \strong{ST10CV} \emph{numeric} Volume (Cortical Parcellation) of Icv Range value: 843290, ..., 2110290
\item \strong{ST110CV} \emph{numeric} Volume (Cortical Parcellation) of RightPrecentral Range value: 3773, ..., 16749
\item \strong{ST110SA} \emph{numeric} Surface Area of RightPrecentral Range value: 2254, ..., 7844
\item \strong{ST110TA} \emph{numeric} Cortical Thickness Average of RightPrecentral Range value: 1.272, ..., 2.717
\item \strong{ST110TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPrecentral Range value: 0.411, ..., 0.914
\item \strong{ST111CV} \emph{numeric} Volume (Cortical Parcellation) of RightPrecuneus Range value: 3972, ..., 12400
\item \strong{ST111SA} \emph{numeric} Surface Area of RightPrecuneus Range value: 2321, ..., 6040
\item \strong{ST111TA} \emph{numeric} Cortical Thickness Average of RightPrecuneus Range value: 1.345, ..., 2.569
\item \strong{ST111TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPrecuneus Range value: 0.375, ..., 1.073
\item \strong{ST112SV} \emph{numeric} Volume (WM Parcellation) of RightPutamen Range value: 1311, ..., 7924
\item \strong{ST113CV} \emph{numeric} Volume (Cortical Parcellation) of RightRostralAnteriorCingulate Range value: 152, ..., 3001
\item \strong{ST113SA} \emph{numeric} Surface Area of RightRostralAnteriorCingulate Range value: 203, ..., 1093
\item \strong{ST113TA} \emph{numeric} Cortical Thickness Average of RightRostralAnteriorCingulate Range value: 0.552, ..., 4.333
\item \strong{ST113TS} \emph{numeric} Cortical Thickness Standard Deviation of RightRostralAnteriorCingulate Range value: 0.414, ..., 1.344
\item \strong{ST114CV} \emph{numeric} Volume (Cortical Parcellation) of RightRostralMiddleFrontal Range value: 3558, ..., 21197
\item \strong{ST114SA} \emph{numeric} Surface Area of RightRostralMiddleFrontal Range value: 2530, ..., 8007
\item \strong{ST114TA} \emph{numeric} Cortical Thickness Average of RightRostralMiddleFrontal Range value: 1.419, ..., 2.67
\item \strong{ST114TS} \emph{numeric} Cortical Thickness Standard Deviation of RightRostralMiddleFrontal Range value: 0.409, ..., 0.921
\item \strong{ST115CV} \emph{numeric} Volume (Cortical Parcellation) of RightSuperiorFrontal Range value: 7578, ..., 27082
\item \strong{ST115SA} \emph{numeric} Surface Area of RightSuperiorFrontal Range value: 3259, ..., 10021
\item \strong{ST115TA} \emph{numeric} Cortical Thickness Average of RightSuperiorFrontal Range value: 1.468, ..., 2.994
\item \strong{ST115TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorFrontal Range value: 0.503, ..., 0.979
\item \strong{ST116CV} \emph{numeric} Volume (Cortical Parcellation) of RightSuperiorParietal Range value: 5687, ..., 18204
\item \strong{ST116SA} \emph{numeric} Surface Area of RightSuperiorParietal Range value: 3138, ..., 9185
\item \strong{ST116TA} \emph{numeric} Cortical Thickness Average of RightSuperiorParietal Range value: 1.24, ..., 2.617
\item \strong{ST116TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorParietal Range value: 0.268, ..., 1.076
\item \strong{ST117CV} \emph{numeric} Volume (Cortical Parcellation) of RightSuperiorTemporal Range value: 2154, ..., 13031
\item \strong{ST117SA} \emph{numeric} Surface Area of RightSuperiorTemporal Range value: 1643, ..., 5473
\item \strong{ST117TA} \emph{numeric} Cortical Thickness Average of RightSuperiorTemporal Range value: 1.274, ..., 2.989
\item \strong{ST117TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorTemporal Range value: 0.334, ..., 0.966
\item \strong{ST118CV} \emph{numeric} Volume (Cortical Parcellation) of RightSupramarginal Range value: 2633, ..., 13924
\item \strong{ST118SA} \emph{numeric} Surface Area of RightSupramarginal Range value: 1694, ..., 5819
\item \strong{ST118TA} \emph{numeric} Cortical Thickness Average of RightSupramarginal Range value: 1.426, ..., 2.818
\item \strong{ST118TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSupramarginal Range value: 0.377, ..., 1.212
\item \strong{ST119CV} \emph{numeric} Volume (Cortical Parcellation) of RightTemporalPole Range value: 314, ..., 3756
\item \strong{ST119SA} \emph{numeric} Surface Area of RightTemporalPole Range value: 125, ..., 796
\item \strong{ST119TA} \emph{numeric} Cortical Thickness Average of RightTemporalPole Range value: 1.33, ..., 4.613
\item \strong{ST119TS} \emph{numeric} Cortical Thickness Standard Deviation of RightTemporalPole Range value: 0.288, ..., 1.296
\item \strong{ST11SV} \emph{numeric} Volume (WM Parcellation) of LeftAccumbensArea Range value: 52, ..., 946
\item \strong{ST120SV} \emph{numeric} Volume (WM Parcellation) of RightThalamus Range value: 4138, ..., 11116
\item \strong{ST121CV} \emph{numeric} Volume (Cortical Parcellation) of RightTransverseTemporal Range value: 113, ..., 1380
\item \strong{ST121SA} \emph{numeric} Surface Area of RightTransverseTemporal Range value: 73, ..., 592
\item \strong{ST121TA} \emph{numeric} Cortical Thickness Average of RightTransverseTemporal Range value: 1.167, ..., 3.085
\item \strong{ST121TS} \emph{numeric} Cortical Thickness Standard Deviation of RightTransverseTemporal Range value: 0.194, ..., 0.875
\item \strong{ST122SV} \emph{logical} Volume (WM Parcellation) of RightUndetermined A Boolean value: TRUE/ FALSE
\item \strong{ST123CV} \emph{numeric} Volume (Cortical Parcellation) of RightUnknown Range value: 3088, ..., 10227
\item \strong{ST123SA} \emph{numeric} Surface Area of RightUnknown Range value: 3968, ..., 9908
\item \strong{ST123TA} \emph{numeric} Cortical Thickness Average of RightUnknown Range value: 0.486, ..., 1.592
\item \strong{ST123TS} \emph{numeric} Cortical Thickness Standard Deviation of RightUnknown Range value: 0.803, ..., 1.796
\item \strong{ST124SV} \emph{numeric} Volume (WM Parcellation) of RightVentralDC Range value: 2038, ..., 5491
\item \strong{ST125SV} \emph{numeric} Volume (WM Parcellation) of RightVessel Range value: 3, ..., 516
\item \strong{ST126SV} \emph{logical} Volume (WM Parcellation) of RightWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST127SV} \emph{numeric} Volume (WM Parcellation) of ThirdVentricle Range value: 599, ..., 5668
\item \strong{ST128SV} \emph{numeric} Volume (WM Parcellation) of WMHypoIntensities Range value: 1148, ..., 65054
\item \strong{ST129CV} \emph{numeric} Volume (Cortical Parcellation) of LeftInsula Range value: 2430, ..., 8162
\item \strong{ST129SA} \emph{numeric} Surface Area of LeftInsula Range value: 1399, ..., 2904
\item \strong{ST129TA} \emph{numeric} Cortical Thickness Average of LeftInsula Range value: 1.446, ..., 3.411
\item \strong{ST129TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInsula Range value: 0.534, ..., 1.343
\item \strong{ST12SV} \emph{numeric} Volume (WM Parcellation) of LeftAmygdala Range value: 380, ..., 1904
\item \strong{ST130CV} \emph{numeric} Volume (Cortical Parcellation) of RightInsula Range value: 2021, ..., 8339
\item \strong{ST130SA} \emph{numeric} Surface Area of RightInsula Range value: 1195, ..., 2696
\item \strong{ST130TA} \emph{numeric} Cortical Thickness Average of RightInsula Range value: 1.629, ..., 3.569
\item \strong{ST130TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInsula Range value: 0.59, ..., 1.358
\item \strong{ST13CV} \emph{numeric} Volume (Cortical Parcellation) of LeftBankssts Range value: 707, ..., 5891
\item \strong{ST13SA} \emph{numeric} Surface Area of LeftBankssts Range value: 515, ..., 2635
\item \strong{ST13TA} \emph{numeric} Cortical Thickness Average of LeftBankssts Range value: 1.245, ..., 2.851
\item \strong{ST13TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftBankssts Range value: 0.287, ..., 0.847
\item \strong{ST14CV} \emph{numeric} Volume (Cortical Parcellation) of LeftCaudalAnteriorCingulate Range value: 229, ..., 3285
\item \strong{ST14SA} \emph{numeric} Surface Area of LeftCaudalAnteriorCingulate Range value: 235, ..., 1556
\item \strong{ST14TA} \emph{numeric} Cortical Thickness Average of LeftCaudalAnteriorCingulate Range value: 1.008, ..., 3.688
\item \strong{ST14TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCaudalAnteriorCingulate Range value: 0.439, ..., 1.57
\item \strong{ST15CV} \emph{numeric} Volume (Cortical Parcellation) of LeftCaudalMiddleFrontal Range value: 1570, ..., 10847
\item \strong{ST15SA} \emph{numeric} Surface Area of LeftCaudalMiddleFrontal Range value: 1110, ..., 4425
\item \strong{ST15TA} \emph{numeric} Cortical Thickness Average of LeftCaudalMiddleFrontal Range value: 1.23, ..., 2.804
\item \strong{ST15TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCaudalMiddleFrontal Range value: 0.269, ..., 0.928
\item \strong{ST16SV} \emph{numeric} Volume (WM Parcellation) of LeftCaudate Range value: 1822, ..., 8426
\item \strong{ST17SV} \emph{numeric} Volume (WM Parcellation) of LeftCerebellumCortex Range value: 31323, ..., 65461
\item \strong{ST18SV} \emph{numeric} Volume (WM Parcellation) of LeftCerebellumWM Range value: 7569, ..., 26949
\item \strong{ST19SV} \emph{numeric} Volume (WM Parcellation) of LeftCerebralCortex Range value: 117038, ..., 266350
\item \strong{ST1SV} \emph{numeric} Volume (WM Parcellation) of Brainstem Range value: 14030, ..., 33819
\item \strong{ST20SV} \emph{numeric} Volume (WM Parcellation) of LeftCerebralWM Range value: 107685, ..., 322806
\item \strong{ST21SV} \emph{numeric} Volume (WM Parcellation) of LeftChoroidPlexus Range value: 975, ..., 4846
\item \strong{ST22CV} \emph{logical} Volume (Cortical Parcellation) of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22SA} \emph{logical} Surface Area of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22TA} \emph{logical} Cortical Thickness Average of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22TS} \emph{logical} Cortical Thickness Standard Deviation of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST23CV} \emph{numeric} Volume (Cortical Parcellation) of LeftCuneus Range value: 860, ..., 4880
\item \strong{ST23SA} \emph{numeric} Surface Area of LeftCuneus Range value: 412, ..., 2256
\item \strong{ST23TA} \emph{numeric} Cortical Thickness Average of LeftCuneus Range value: 1.13, ..., 2.712
\item \strong{ST23TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCuneus Range value: 0.145, ..., 1.077
\item \strong{ST24CV} \emph{numeric} Volume (Cortical Parcellation) of LeftEntorhinal Range value: 266, ..., 3705
\item \strong{ST24SA} \emph{numeric} Surface Area of LeftEntorhinal Range value: 136, ..., 825
\item \strong{ST24TA} \emph{numeric} Cortical Thickness Average of LeftEntorhinal Range value: 1.179, ..., 4.295
\item \strong{ST24TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftEntorhinal Range value: 0.448, ..., 1.492
\item \strong{ST25CV} \emph{numeric} Volume (Cortical Parcellation) of LeftFrontalPole Range value: 162, ..., 1411
\item \strong{ST25SA} \emph{numeric} Surface Area of LeftFrontalPole Range value: 47, ..., 341
\item \strong{ST25TA} \emph{numeric} Cortical Thickness Average of LeftFrontalPole Range value: 1.23, ..., 3.8
\item \strong{ST25TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftFrontalPole Range value: 0.197, ..., 1.104
\item \strong{ST26CV} \emph{numeric} Volume (Cortical Parcellation) of LeftFusiform Range value: 2600, ..., 12710
\item \strong{ST26SA} \emph{numeric} Surface Area of LeftFusiform Range value: 1643, ..., 5278
\item \strong{ST26TA} \emph{numeric} Cortical Thickness Average of LeftFusiform Range value: 1.388, ..., 3.172
\item \strong{ST26TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftFusiform Range value: 0.46, ..., 1.231
\item \strong{ST27SA} \emph{numeric} Surface Area of LeftHemisphere Range value: 57804.7, ..., 111492
\item \strong{ST28CV} \emph{logical} Volume (Cortical Parcellation) of LeftHemisphereWM A Boolean value: TRUE/ FALSE
\item \strong{ST29SV} \emph{numeric} Volume (WM Parcellation) of LeftHippocampus Range value: 1217, ..., 5468
\item \strong{ST2SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumAnterior Range value: 55, ..., 1351
\item \strong{ST30SV} \emph{numeric} Volume (WM Parcellation) of LeftInferiorLateralVentricle Range value: 40, ..., 7819
\item \strong{ST31CV} \emph{numeric} Volume (Cortical Parcellation) of LeftInferiorParietal Range value: 4036, ..., 16386
\item \strong{ST31SA} \emph{numeric} Surface Area of LeftInferiorParietal Range value: 2331, ..., 6541
\item \strong{ST31TA} \emph{numeric} Cortical Thickness Average of LeftInferiorParietal Range value: 1.311, ..., 2.897
\item \strong{ST31TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInferiorParietal Range value: 0.305, ..., 1.048
\item \strong{ST32CV} \emph{numeric} Volume (Cortical Parcellation) of LeftInferiorTemporal Range value: 3409, ..., 14711
\item \strong{ST32SA} \emph{numeric} Surface Area of LeftInferiorTemporal Range value: 1408, ..., 5164
\item \strong{ST32TA} \emph{numeric} Cortical Thickness Average of LeftInferiorTemporal Range value: 1.547, ..., 3.305
\item \strong{ST32TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInferiorTemporal Range value: 0.506, ..., 1.161
\item \strong{ST33SV} \emph{logical} Volume (WM Parcellation) of LeftInterior A Boolean value: TRUE/ FALSE
\item \strong{ST34CV} \emph{numeric} Volume (Cortical Parcellation) of LeftIsthmusCingulate Range value: 752, ..., 3588
\item \strong{ST34SA} \emph{numeric} Surface Area of LeftIsthmusCingulate Range value: 486, ..., 1475
\item \strong{ST34TA} \emph{numeric} Cortical Thickness Average of LeftIsthmusCingulate Range value: 1.137, ..., 3.162
\item \strong{ST34TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftIsthmusCingulate Range value: 0.411, ..., 1.316
\item \strong{ST35CV} \emph{numeric} Volume (Cortical Parcellation) of LeftLateralOccipital Range value: 4743, ..., 16335
\item \strong{ST35SA} \emph{numeric} Surface Area of LeftLateralOccipital Range value: 2481, ..., 7540
\item \strong{ST35TA} \emph{numeric} Cortical Thickness Average of LeftLateralOccipital Range value: 1.088, ..., 2.47
\item \strong{ST35TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLateralOccipital Range value: 0.301, ..., 1.044
\item \strong{ST36CV} \emph{numeric} Volume (Cortical Parcellation) of LeftLateralOrbitofrontal Range value: 2493, ..., 8846
\item \strong{ST36SA} \emph{numeric} Surface Area of LeftLateralOrbitofrontal Range value: 1021, ..., 3544
\item \strong{ST36TA} \emph{numeric} Cortical Thickness Average of LeftLateralOrbitofrontal Range value: 1.433, ..., 3.325
\item \strong{ST36TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLateralOrbitofrontal Range value: 0.472, ..., 1.082
\item \strong{ST37SV} \emph{numeric} Volume (WM Parcellation) of LeftLateralVentricle Range value: 3111, ..., 87716
\item \strong{ST38CV} \emph{numeric} Volume (Cortical Parcellation) of LeftLingual Range value: 1641, ..., 9324
\item \strong{ST38SA} \emph{numeric} Surface Area of LeftLingual Range value: 1280, ..., 5005
\item \strong{ST38TA} \emph{numeric} Cortical Thickness Average of LeftLingual Range value: 1.221, ..., 2.246
\item \strong{ST38TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLingual Range value: 0.371, ..., 1.213
\item \strong{ST39CV} \emph{numeric} Volume (Cortical Parcellation) of LeftMedialOrbitofrontal Range value: 1338, ..., 5858
\item \strong{ST39SA} \emph{numeric} Surface Area of LeftMedialOrbitofrontal Range value: 868, ..., 2304
\item \strong{ST39TA} \emph{numeric} Cortical Thickness Average of LeftMedialOrbitofrontal Range value: 1.167, ..., 3.066
\item \strong{ST39TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftMedialOrbitofrontal Range value: 0.39, ..., 1.154
\item \strong{ST3SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumCentral Range value: 22, ..., 680
\item \strong{ST40CV} \emph{numeric} Volume (Cortical Parcellation) of LeftMiddleTemporal Range value: 1958, ..., 15462
\item \strong{ST40SA} \emph{numeric} Surface Area of LeftMiddleTemporal Range value: 1394, ..., 4407
\item \strong{ST40TA} \emph{numeric} Cortical Thickness Average of LeftMiddleTemporal Range value: 1.406, ..., 3.193
\item \strong{ST40TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftMiddleTemporal Range value: 0.393, ..., 1.066
\item \strong{ST41SV} \emph{logical} Volume (WM Parcellation) of LeftNonWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST42SV} \emph{numeric} Volume (WM Parcellation) of LeftPallidum Range value: 253, ..., 2730
\item \strong{ST43CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParacentral Range value: 1380, ..., 5620
\item \strong{ST43SA} \emph{numeric} Surface Area of LeftParacentral Range value: 792, ..., 2394
\item \strong{ST43TA} \emph{numeric} Cortical Thickness Average of LeftParacentral Range value: 1.24, ..., 3.061
\item \strong{ST43TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParacentral Range value: 0.323, ..., 1.039
\item \strong{ST44CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParahippocampal Range value: 398, ..., 3585
\item \strong{ST44SA} \emph{numeric} Surface Area of LeftParahippocampal Range value: 431, ..., 1411
\item \strong{ST44TA} \emph{numeric} Cortical Thickness Average of LeftParahippocampal Range value: 0.543, ..., 3.512
\item \strong{ST44TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParahippocampal Range value: 0.404, ..., 1.307
\item \strong{ST45CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParsOpercularis Range value: 1440, ..., 8985
\item \strong{ST45SA} \emph{numeric} Surface Area of LeftParsOpercularis Range value: 974, ..., 3661
\item \strong{ST45TA} \emph{numeric} Cortical Thickness Average of LeftParsOpercularis Range value: 1.482, ..., 2.851
\item \strong{ST45TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsOpercularis Range value: 0.403, ..., 0.918
\item \strong{ST46CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParsOrbitalis Range value: 471, ..., 3054
\item \strong{ST46SA} \emph{numeric} Surface Area of LeftParsOrbitalis Range value: 299, ..., 1016
\item \strong{ST46TA} \emph{numeric} Cortical Thickness Average of LeftParsOrbitalis Range value: 1.157, ..., 3.371
\item \strong{ST46TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsOrbitalis Range value: 0.224, ..., 1.218
\item \strong{ST47CV} \emph{numeric} Volume (Cortical Parcellation) of LeftParsTriangularis Range value: 820, ..., 5758
\item \strong{ST47SA} \emph{numeric} Surface Area of LeftParsTriangularis Range value: 595, ..., 2131
\item \strong{ST47TA} \emph{numeric} Cortical Thickness Average of LeftParsTriangularis Range value: 1.371, ..., 2.992
\item \strong{ST47TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsTriangularis Range value: 0.381, ..., 1.055
\item \strong{ST48CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPericalcarine Range value: 454, ..., 4431
\item \strong{ST48SA} \emph{numeric} Surface Area of LeftPericalcarine Range value: 515, ..., 2179
\item \strong{ST48TA} \emph{numeric} Cortical Thickness Average of LeftPericalcarine Range value: 0.882, ..., 2.701
\item \strong{ST48TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPericalcarine Range value: 0.197, ..., 1.532
\item \strong{ST49CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPostcentral Range value: 4208, ..., 14798
\item \strong{ST49SA} \emph{numeric} Surface Area of LeftPostcentral Range value: 2353, ..., 8588
\item \strong{ST49TA} \emph{numeric} Cortical Thickness Average of LeftPostcentral Range value: 1.279, ..., 2.225
\item \strong{ST49TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPostcentral Range value: 0.371, ..., 1.032
\item \strong{ST4SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumMidAnterior Range value: 32, ..., 1099
\item \strong{ST50CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPosteriorCingulate Range value: 1064, ..., 4756
\item \strong{ST50SA} \emph{numeric} Surface Area of LeftPosteriorCingulate Range value: 701, ..., 2274
\item \strong{ST50TA} \emph{numeric} Cortical Thickness Average of LeftPosteriorCingulate Range value: 1.198, ..., 2.969
\item \strong{ST50TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPosteriorCingulate Range value: 0.415, ..., 1.094
\item \strong{ST51CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPrecentral Range value: 4968, ..., 19421
\item \strong{ST51SA} \emph{numeric} Surface Area of LeftPrecentral Range value: 3091, ..., 8215
\item \strong{ST51TA} \emph{numeric} Cortical Thickness Average of LeftPrecentral Range value: 1.295, ..., 2.71
\item \strong{ST51TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPrecentral Range value: 0.42, ..., 0.926
\item \strong{ST52CV} \emph{numeric} Volume (Cortical Parcellation) of LeftPrecuneus Range value: 3465, ..., 12275
\item \strong{ST52SA} \emph{numeric} Surface Area of LeftPrecuneus Range value: 2230, ..., 5257
\item \strong{ST52TA} \emph{numeric} Cortical Thickness Average of LeftPrecuneus Range value: 1.346, ..., 2.738
\item \strong{ST52TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPrecuneus Range value: 0.336, ..., 0.929
\item \strong{ST53SV} \emph{numeric} Volume (WM Parcellation) of LeftPutamen Range value: 1126, ..., 7672
\item \strong{ST54CV} \emph{numeric} Volume (Cortical Parcellation) of LeftRostralAnteriorCingulate Range value: 309, ..., 3869
\item \strong{ST54SA} \emph{numeric} Surface Area of LeftRostralAnteriorCingulate Range value: 313, ..., 1258
\item \strong{ST54TA} \emph{numeric} Cortical Thickness Average of LeftRostralAnteriorCingulate Range value: 0.967, ..., 4.019
\item \strong{ST54TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftRostralAnteriorCingulate Range value: 0.415, ..., 1.455
\item \strong{ST55CV} \emph{numeric} Volume (Cortical Parcellation) of LeftRostralMiddleFrontal Range value: 3668, ..., 18349
\item \strong{ST55SA} \emph{numeric} Surface Area of LeftRostralMiddleFrontal Range value: 2668, ..., 8043
\item \strong{ST55TA} \emph{numeric} Cortical Thickness Average of LeftRostralMiddleFrontal Range value: 1.25, ..., 2.706
\item \strong{ST55TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftRostralMiddleFrontal Range value: 0.352, ..., 0.902
\item \strong{ST56CV} \emph{numeric} Volume (Cortical Parcellation) of LeftSuperiorFrontal Range value: 7584, ..., 27081
\item \strong{ST56SA} \emph{numeric} Surface Area of LeftSuperiorFrontal Range value: 3656, ..., 9932
\item \strong{ST56TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorFrontal Range value: 1.414, ..., 2.973
\item \strong{ST56TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorFrontal Range value: 0.456, ..., 0.976
\item \strong{ST57CV} \emph{numeric} Volume (Cortical Parcellation) of LeftSuperiorParietal Range value: 4941, ..., 17480
\item \strong{ST57SA} \emph{numeric} Surface Area of LeftSuperiorParietal Range value: 3272, ..., 8542
\item \strong{ST57TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorParietal Range value: 1.153, ..., 2.708
\item \strong{ST57TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorParietal Range value: 0.223, ..., 0.962
\item \strong{ST58CV} \emph{numeric} Volume (Cortical Parcellation) of LeftSuperiorTemporal Range value: 2803, ..., 17668
\item \strong{ST58SA} \emph{numeric} Surface Area of LeftSuperiorTemporal Range value: 2113, ..., 6019
\item \strong{ST58TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorTemporal Range value: 1.45, ..., 2.975
\item \strong{ST58TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorTemporal Range value: 0.437, ..., 0.972
\item \strong{ST59CV} \emph{numeric} Volume (Cortical Parcellation) of LeftSupramarginal Range value: 3184, ..., 13537
\item \strong{ST59SA} \emph{numeric} Surface Area of LeftSupramarginal Range value: 2353, ..., 5690
\item \strong{ST59TA} \emph{numeric} Cortical Thickness Average of LeftSupramarginal Range value: 1.411, ..., 2.83
\item \strong{ST59TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSupramarginal Range value: 0.365, ..., 0.866
\item \strong{ST5SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumMidPosterior Range value: 27, ..., 620
\item \strong{ST60CV} \emph{numeric} Volume (Cortical Parcellation) of LeftTemporalPole Range value: 474, ..., 3397
\item \strong{ST60SA} \emph{numeric} Surface Area of LeftTemporalPole Range value: 147, ..., 971
\item \strong{ST60TA} \emph{numeric} Cortical Thickness Average of LeftTemporalPole Range value: 1.43, ..., 4.5
\item \strong{ST60TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftTemporalPole Range value: 0.322, ..., 1.223
\item \strong{ST61SV} \emph{numeric} Volume (WM Parcellation) of LeftThalamus Range value: 4267, ..., 11315
\item \strong{ST62CV} \emph{numeric} Volume (Cortical Parcellation) of LeftTransverseTemporal Range value: 159, ..., 2242
\item \strong{ST62SA} \emph{numeric} Surface Area of LeftTransverseTemporal Range value: 112, ..., 1484
\item \strong{ST62TA} \emph{numeric} Cortical Thickness Average of LeftTransverseTemporal Range value: 1.181, ..., 2.936
\item \strong{ST62TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftTransverseTemporal Range value: 0.198, ..., 0.784
\item \strong{ST63SV} \emph{logical} Volume (WM Parcellation) of LeftUndetermined A Boolean value: TRUE/ FALSE
\item \strong{ST64CV} \emph{numeric} Volume (Cortical Parcellation) of LeftUnknown Range value: 2937, ..., 9804
\item \strong{ST64SA} \emph{numeric} Surface Area of LeftUnknown Range value: 4230, ..., 9112
\item \strong{ST64TA} \emph{numeric} Cortical Thickness Average of LeftUnknown Range value: 0.434, ..., 1.511
\item \strong{ST64TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftUnknown Range value: 0.899, ..., 1.746
\item \strong{ST65SV} \emph{numeric} Volume (WM Parcellation) of LeftVentralDC Range value: 2166, ..., 5915
\item \strong{ST66SV} \emph{numeric} Volume (WM Parcellation) of LeftVessel Range value: 3, ..., 516
\item \strong{ST67SV} \emph{logical} Volume (WM Parcellation) of LeftWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST68SV} \emph{numeric} Volume (WM Parcellation) of NonWMHypoIntensities Range value: 3, ..., 1100
\item \strong{ST69SV} \emph{numeric} Volume (WM Parcellation) of OpticChiasm Range value: 42, ..., 728
\item \strong{ST6SV} \emph{numeric} Volume (WM Parcellation) of CorpusCallosumPosterior Range value: 48, ..., 1654
\item \strong{ST70SV} \emph{numeric} Volume (WM Parcellation) of RightAccumbensArea Range value: 160, ..., 817
\item \strong{ST71SV} \emph{numeric} Volume (WM Parcellation) of RightAmygdala Range value: 349, ..., 2095
\item \strong{ST72CV} \emph{numeric} Volume (Cortical Parcellation) of RightBankssts Range value: 618, ..., 3958
\item \strong{ST72SA} \emph{numeric} Surface Area of RightBankssts Range value: 463, ..., 1772
\item \strong{ST72TA} \emph{numeric} Cortical Thickness Average of RightBankssts Range value: 1.268, ..., 2.932
\item \strong{ST72TS} \emph{numeric} Cortical Thickness Standard Deviation of RightBankssts Range value: 0.256, ..., 0.944
\item \strong{ST73CV} \emph{numeric} Volume (Cortical Parcellation) of RightCaudalAnteriorCingulate Range value: 7, ..., 3704
\item \strong{ST73SA} \emph{numeric} Surface Area of RightCaudalAnteriorCingulate Range value: 405, ..., 1366
\item \strong{ST73TA} \emph{numeric} Cortical Thickness Average of RightCaudalAnteriorCingulate Range value: 0.006, ..., 3.514
\item \strong{ST73TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCaudalAnteriorCingulate Range value: 0.098, ..., 1.474
\item \strong{ST74CV} \emph{numeric} Volume (Cortical Parcellation) of RightCaudalMiddleFrontal Range value: 1230, ..., 10269
\item \strong{ST74SA} \emph{numeric} Surface Area of RightCaudalMiddleFrontal Range value: 834, ..., 4769
\item \strong{ST74TA} \emph{numeric} Cortical Thickness Average of RightCaudalMiddleFrontal Range value: 1.341, ..., 2.887
\item \strong{ST74TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCaudalMiddleFrontal Range value: 0.32, ..., 0.955
\item \strong{ST75SV} \emph{numeric} Volume (WM Parcellation) of RightCaudate Range value: 1203, ..., 7387
\item \strong{ST76SV} \emph{numeric} Volume (WM Parcellation) of RightCerebellumCortex Range value: 31502, ..., 68021
\item \strong{ST77SV} \emph{numeric} Volume (WM Parcellation) of RightCerebellumWM Range value: 7431, ..., 23025
\item \strong{ST78SV} \emph{numeric} Volume (WM Parcellation) of RightCerebralCortex Range value: 117890, ..., 271706
\item \strong{ST79SV} \emph{numeric} Volume (WM Parcellation) of RightCerebralWM Range value: 126353, ..., 328034
\item \strong{ST7SV} \emph{numeric} Volume (WM Parcellation) of Csf Range value: 655, ..., 6291
\item \strong{ST80SV} \emph{numeric} Volume (WM Parcellation) of RightChoroidPlexus Range value: 866, ..., 5684
\item \strong{ST81CV} \emph{logical} Volume (Cortical Parcellation) of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81SA} \emph{logical} Surface Area of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81TA} \emph{logical} Cortical Thickness Average of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81TS} \emph{logical} Cortical Thickness Standard Deviation of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST82CV} \emph{numeric} Volume (Cortical Parcellation) of RightCuneus Range value: 563, ..., 4693
\item \strong{ST82SA} \emph{numeric} Surface Area of RightCuneus Range value: 441, ..., 2505
\item \strong{ST82TA} \emph{numeric} Cortical Thickness Average of RightCuneus Range value: 1.214, ..., 2.253
\item \strong{ST82TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCuneus Range value: 0.239, ..., 0.998
\item \strong{ST83CV} \emph{numeric} Volume (Cortical Parcellation) of RightEntorhinal Range value: 402, ..., 3955
\item \strong{ST83SA} \emph{numeric} Surface Area of RightEntorhinal Range value: 158, ..., 882
\item \strong{ST83TA} \emph{numeric} Cortical Thickness Average of RightEntorhinal Range value: 1.314, ..., 4.408
\item \strong{ST83TS} \emph{numeric} Cortical Thickness Standard Deviation of RightEntorhinal Range value: 0.374, ..., 1.58
\item \strong{ST84CV} \emph{numeric} Volume (Cortical Parcellation) of RightFrontalPole Range value: 264, ..., 1800
\item \strong{ST84SA} \emph{numeric} Surface Area of RightFrontalPole Range value: 72, ..., 515
\item \strong{ST84TA} \emph{numeric} Cortical Thickness Average of RightFrontalPole Range value: 1.272, ..., 3.89
\item \strong{ST84TS} \emph{numeric} Cortical Thickness Standard Deviation of RightFrontalPole Range value: 0.254, ..., 1.121
\item \strong{ST85CV} \emph{numeric} Volume (Cortical Parcellation) of RightFusiform Range value: 2856, ..., 13203
\item \strong{ST85SA} \emph{numeric} Surface Area of RightFusiform Range value: 1639, ..., 4749
\item \strong{ST85TA} \emph{numeric} Cortical Thickness Average of RightFusiform Range value: 1.474, ..., 3.18
\item \strong{ST85TS} \emph{numeric} Cortical Thickness Standard Deviation of RightFusiform Range value: 0.448, ..., 1.089
\item \strong{ST86SA} \emph{numeric} Surface Area of RightHemisphere Range value: 58959.5, ..., 113037
\item \strong{ST87CV} \emph{logical} Volume (Cortical Parcellation) of RightHemisphereWM A Boolean value: TRUE/ FALSE
\item \strong{ST88SV} \emph{numeric} Volume (WM Parcellation) of RightHippocampus Range value: 734, ..., 5818
\item \strong{ST89SV} \emph{numeric} Volume (WM Parcellation) of RightInferiorLateralVentricle Range value: 41, ..., 8760
\item \strong{ST8SV} \emph{numeric} Volume (WM Parcellation) of FifthVentricle Range value: 1, ..., 206
\item \strong{ST90CV} \emph{numeric} Volume (Cortical Parcellation) of RightInferiorParietal Range value: 4880, ..., 20098
\item \strong{ST90SA} \emph{numeric} Surface Area of RightInferiorParietal Range value: 3156, ..., 8275
\item \strong{ST90TA} \emph{numeric} Cortical Thickness Average of RightInferiorParietal Range value: 1.321, ..., 2.706
\item \strong{ST90TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInferiorParietal Range value: 0.311, ..., 0.932
\item \strong{ST91CV} \emph{numeric} Volume (Cortical Parcellation) of RightInferiorTemporal Range value: 3511, ..., 14767
\item \strong{ST91SA} \emph{numeric} Surface Area of RightInferiorTemporal Range value: 1517, ..., 4940
\item \strong{ST91TA} \emph{numeric} Cortical Thickness Average of RightInferiorTemporal Range value: 1.503, ..., 3.404
\item \strong{ST91TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInferiorTemporal Range value: 0.486, ..., 1.132
\item \strong{ST92SV} \emph{logical} Volume (WM Parcellation) of RightInterior A Boolean value: TRUE/ FALSE
\item \strong{ST93CV} \emph{numeric} Volume (Cortical Parcellation) of RightIsthmusCingulate Range value: 599, ..., 3358
\item \strong{ST93SA} \emph{numeric} Surface Area of RightIsthmusCingulate Range value: 460, ..., 1459
\item \strong{ST93TA} \emph{numeric} Cortical Thickness Average of RightIsthmusCingulate Range value: 1.221, ..., 3.132
\item \strong{ST93TS} \emph{numeric} Cortical Thickness Standard Deviation of RightIsthmusCingulate Range value: 0.299, ..., 1.328
\item \strong{ST94CV} \emph{numeric} Volume (Cortical Parcellation) of RightLateralOccipital Range value: 4444, ..., 16325
\item \strong{ST94SA} \emph{numeric} Surface Area of RightLateralOccipital Range value: 2618, ..., 7252
\item \strong{ST94TA} \emph{numeric} Cortical Thickness Average of RightLateralOccipital Range value: 1.312, ..., 2.544
\item \strong{ST94TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLateralOccipital Range value: 0.29, ..., 0.847
\item \strong{ST95CV} \emph{numeric} Volume (Cortical Parcellation) of RightLateralOrbitofrontal Range value: 1889, ..., 8663
\item \strong{ST95SA} \emph{numeric} Surface Area of RightLateralOrbitofrontal Range value: 1344, ..., 3603
\item \strong{ST95TA} \emph{numeric} Cortical Thickness Average of RightLateralOrbitofrontal Range value: 1.359, ..., 3.254
\item \strong{ST95TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLateralOrbitofrontal Range value: 0.464, ..., 1.108
\item \strong{ST96SV} \emph{numeric} Volume (WM Parcellation) of RightLateralVentricle Range value: 2320, ..., 72924
\item \strong{ST97CV} \emph{numeric} Volume (Cortical Parcellation) of RightLingual Range value: 2425, ..., 8808
\item \strong{ST97SA} \emph{numeric} Surface Area of RightLingual Range value: 1523, ..., 6647
\item \strong{ST97TA} \emph{numeric} Cortical Thickness Average of RightLingual Range value: 1.052, ..., 2.299
\item \strong{ST97TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLingual Range value: 0.374, ..., 0.934
\item \strong{ST98CV} \emph{numeric} Volume (Cortical Parcellation) of RightMedialOrbitofrontal Range value: 1448, ..., 6063
\item \strong{ST98SA} \emph{numeric} Surface Area of RightMedialOrbitofrontal Range value: 888, ..., 2521
\item \strong{ST98TA} \emph{numeric} Cortical Thickness Average of RightMedialOrbitofrontal Range value: 1.174, ..., 3.307
\item \strong{ST98TS} \emph{numeric} Cortical Thickness Standard Deviation of RightMedialOrbitofrontal Range value: 0.413, ..., 1.261
\item \strong{ST99CV} \emph{numeric} Volume (Cortical Parcellation) of RightMiddleTemporal Range value: 2378, ..., 15683
\item \strong{ST99SA} \emph{numeric} Surface Area of RightMiddleTemporal Range value: 1657, ..., 4837
\item \strong{ST99TA} \emph{numeric} Cortical Thickness Average of RightMiddleTemporal Range value: 1.366, ..., 3.286
\item \strong{ST99TS} \emph{numeric} Cortical Thickness Standard Deviation of RightMiddleTemporal Range value: 0.375, ..., 1.042
\item \strong{ST9SV} \emph{numeric} Volume (WM Parcellation) of FourthVentricle Range value: 674, ..., 6300
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
