% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFFSX51}
\alias{UCSFFSX51}
\title{Cross-Sectional FreeSurfer (5.1): 2019-11-08}
\format{
A data frame with 4896 observations and 366 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFFSX51)
}
\description{
Cross-Sectional FreeSurfer (5.1): 2019-11-08 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI2, ADNIGO, ADNI1, nv
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code , with options: v11, nv, v06, v21, v41, v31, v51, v02, v04, v05
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } The scan date for the image processed Date: YYY-MM-DD
\item \strong{VERSION} \emph{ }   Date: YYY-MM-DD
\item \strong{LONISID} \emph{numeric} The LONI study UID for the image processed Range value: 24831, ..., 107491
\item \strong{LONIUID} \emph{numeric} The LONI series UID for the image processed. Range value: 79478, ..., 531719
\item \strong{IMAGEUID} \emph{numeric} The ImageUID for the image processed. Range value: 221959, ..., 1051225
\item \strong{IMAGETYPE} \emph{character} Whether the T1 image was an accelerated T1 or not. Accelerated T1/Non-Accelerated T1
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   complete/partial
\item \strong{OVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. Pass/Fail/Partial
\item \strong{TEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) Pass/Fail
\item \strong{FRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) Pass/Fail
\item \strong{PARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) Pass/Fail
\item \strong{INSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) Pass/Fail
\item \strong{OCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) Pass/Fail
\item \strong{BGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) Pass/Fail
\item \strong{CWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87) Pass/Fail
\item \strong{VENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) Pass/Fail
\item \strong{LHIPQC} \emph{character} QC rating of the left hippocampal subfields. Fail affects the following regions: Left CA1 (ST131HS); Left CA2-3 (ST132HS); Left CA4-DG (ST133HS); Left Fimbria (ST134HS); Left Hippocampal Fissure (ST135HS); Left Presubiculum (ST136HS); Left Subiculum (ST137HS); Left Tail (ST138HS) Pass/Fail
\item \strong{RHIPQC} \emph{character} QC rating of the right hippocampal subfields. Fail affects the following regions: Right CA1 (ST139HS); Right CA2-3 (ST140HS); Right CA4-DG (ST141HS); Right Fimbria (ST142HS); Right Hippocampal Fissure (ST143HS); Right Presubiculum (ST144HS); Right Subiculum (ST145HS); Right Tail (ST146HS) Pass/Fail
\item \strong{ST101SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPallidum Range value: 506, ..., 2538
\item \strong{ST102CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParacentral Range value: 1975, ..., 7965
\item \strong{ST102SA} \emph{numeric} Surface Area (aparc.stats) of RightParacentral Range value: 978, ..., 4143
\item \strong{ST102TA} \emph{numeric} Thickness Average (aparc.stats) of RightParacentral Range value: 1.592, ..., 3.213
\item \strong{ST102TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParacentral Range value: 0.422, ..., 0.926
\item \strong{ST103CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParahippocampal Range value: 699, ..., 3892
\item \strong{ST103SA} \emph{numeric} Surface Area (aparc.stats) of RightParahippocampal Range value: 218, ..., 1266
\item \strong{ST103TA} \emph{numeric} Thickness Average (aparc.stats) of RightParahippocampal Range value: 1.489, ..., 3.569
\item \strong{ST103TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParahippocampal Range value: 0.438, ..., 1.219
\item \strong{ST104CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOpercularis Range value: 1839, ..., 5785
\item \strong{ST104SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOpercularis Range value: 709, ..., 2260
\item \strong{ST104TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOpercularis Range value: 1.7, ..., 2.95
\item \strong{ST104TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOpercularis Range value: 0.349, ..., 1.006
\item \strong{ST105CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOrbitalis Range value: 492, ..., 3601
\item \strong{ST105SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOrbitalis Range value: 136, ..., 1310
\item \strong{ST105TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOrbitalis Range value: 1.694, ..., 3.464
\item \strong{ST105TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOrbitalis Range value: 0.409, ..., 1.15
\item \strong{ST106CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsTriangularis Range value: 1797, ..., 6078
\item \strong{ST106SA} \emph{numeric} Surface Area (aparc.stats) of RightParsTriangularis Range value: 562, ..., 2378
\item \strong{ST106TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsTriangularis Range value: 1.672, ..., 2.995
\item \strong{ST106TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsTriangularis Range value: 0.369, ..., 1.135
\item \strong{ST107CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPericalcarine Range value: 823, ..., 4241
\item \strong{ST107SA} \emph{numeric} Surface Area (aparc.stats) of RightPericalcarine Range value: 594, ..., 2343
\item \strong{ST107TA} \emph{numeric} Thickness Average (aparc.stats) of RightPericalcarine Range value: 1.145, ..., 2.593
\item \strong{ST107TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPericalcarine Range value: 0.243, ..., 1.005
\item \strong{ST108CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPostcentral Range value: 4645, ..., 13977
\item \strong{ST108SA} \emph{numeric} Surface Area (aparc.stats) of RightPostcentral Range value: 2086, ..., 7464
\item \strong{ST108TA} \emph{numeric} Thickness Average (aparc.stats) of RightPostcentral Range value: 1.389, ..., 2.44
\item \strong{ST108TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPostcentral Range value: 0.43, ..., 1.075
\item \strong{ST109CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPosteriorCingulate Range value: 898, ..., 8216
\item \strong{ST109SA} \emph{numeric} Surface Area (aparc.stats) of RightPosteriorCingulate Range value: 603, ..., 4498
\item \strong{ST109TA} \emph{numeric} Thickness Average (aparc.stats) of RightPosteriorCingulate Range value: 1.02, ..., 3.337
\item \strong{ST109TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPosteriorCingulate Range value: 0.438, ..., 1.139
\item \strong{ST10CV} \emph{numeric} Cortical Volume (aparc.stats) of Icv Range value: 708913, ..., 2108790
\item \strong{ST110CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecentral Range value: 6801, ..., 18591
\item \strong{ST110SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecentral Range value: 2696, ..., 7877
\item \strong{ST110TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecentral Range value: 1.43, ..., 2.787
\item \strong{ST110TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecentral Range value: 0.469, ..., 0.869
\item \strong{ST111CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecuneus Range value: 4266, ..., 13971
\item \strong{ST111SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecuneus Range value: 2077, ..., 6149
\item \strong{ST111TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecuneus Range value: 1.403, ..., 2.79
\item \strong{ST111TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecuneus Range value: 0.441, ..., 1.016
\item \strong{ST112SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPutamen Range value: 1747, ..., 10019
\item \strong{ST113CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralAnteriorCingulate Range value: 547, ..., 3862
\item \strong{ST113SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralAnteriorCingulate Range value: 231, ..., 1197
\item \strong{ST113TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralAnteriorCingulate Range value: 1.632, ..., 4.049
\item \strong{ST113TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralAnteriorCingulate Range value: 0.4, ..., 1.565
\item \strong{ST114CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralMiddleFrontal Range value: 6383, ..., 22748
\item \strong{ST114SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralMiddleFrontal Range value: 2110, ..., 9005
\item \strong{ST114TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralMiddleFrontal Range value: 1.545, ..., 2.973
\item \strong{ST114TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralMiddleFrontal Range value: 0.44, ..., 1.118
\item \strong{ST115CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorFrontal Range value: 9969, ..., 28879
\item \strong{ST115SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorFrontal Range value: 4386, ..., 11371
\item \strong{ST115TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorFrontal Range value: 1.673, ..., 3.117
\item \strong{ST115TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorFrontal Range value: 0.496, ..., 1.073
\item \strong{ST116CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorParietal Range value: 6121, ..., 24067
\item \strong{ST116SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorParietal Range value: 3316, ..., 11855
\item \strong{ST116TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorParietal Range value: 1.3, ..., 2.804
\item \strong{ST116TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorParietal Range value: 0.3, ..., 1.263
\item \strong{ST117CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorTemporal Range value: 4488, ..., 15255
\item \strong{ST117SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorTemporal Range value: 2014, ..., 4788
\item \strong{ST117TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorTemporal Range value: 1.543, ..., 3.2
\item \strong{ST117TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorTemporal Range value: 0.487, ..., 0.989
\item \strong{ST118CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSupramarginal Range value: 4313, ..., 15014
\item \strong{ST118SA} \emph{numeric} Surface Area (aparc.stats) of RightSupramarginal Range value: 1864, ..., 6037
\item \strong{ST118TA} \emph{numeric} Thickness Average (aparc.stats) of RightSupramarginal Range value: 1.676, ..., 2.832
\item \strong{ST118TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSupramarginal Range value: 0.441, ..., 0.976
\item \strong{ST119CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTemporalPole Range value: 444, ..., 3685
\item \strong{ST119SA} \emph{numeric} Surface Area (aparc.stats) of RightTemporalPole Range value: 202, ..., 734
\item \strong{ST119TA} \emph{numeric} Thickness Average (aparc.stats) of RightTemporalPole Range value: 1.294, ..., 4.549
\item \strong{ST119TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTemporalPole Range value: 0.333, ..., 1.317
\item \strong{ST11SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAccumbensArea Range value: 20, ..., 1097
\item \strong{ST120SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightThalamus Range value: 2149, ..., 10290
\item \strong{ST121CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTransverseTemporal Range value: 381, ..., 1551
\item \strong{ST121SA} \emph{numeric} Surface Area (aparc.stats) of RightTransverseTemporal Range value: 166, ..., 596
\item \strong{ST121TA} \emph{numeric} Thickness Average (aparc.stats) of RightTransverseTemporal Range value: 1.242, ..., 3.045
\item \strong{ST121TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTransverseTemporal Range value: 0.203, ..., 0.894
\item \strong{ST124SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVentralDC Range value: 2286, ..., 5105
\item \strong{ST125SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVessel Range value: 1, ..., 443
\item \strong{ST127SV} \emph{numeric} Subcortical Volume (aseg.stats) of ThirdVentricle Range value: 517, ..., 5337
\item \strong{ST128SV} \emph{numeric} Subcortical Volume (aseg.stats) of WMHypoIntensities Range value: 584, ..., 56775
\item \strong{ST129CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInsula Range value: 3956, ..., 10141
\item \strong{ST129SA} \emph{numeric} Surface Area (aparc.stats) of LeftInsula Range value: 1289, ..., 3483
\item \strong{ST129TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInsula Range value: 2.104, ..., 3.495
\item \strong{ST129TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInsula Range value: 0.647, ..., 1.282
\item \strong{ST12SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAmygdala Range value: 296, ..., 2591
\item \strong{ST130CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInsula Range value: 3479, ..., 9962
\item \strong{ST130SA} \emph{numeric} Surface Area (aparc.stats) of RightInsula Range value: 1224, ..., 3561
\item \strong{ST130TA} \emph{numeric} Thickness Average (aparc.stats) of RightInsula Range value: 1.917, ..., 3.535
\item \strong{ST130TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInsula Range value: 0.602, ..., 1.273
\item \strong{ST13CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftBankssts Range value: 1049, ..., 4067
\item \strong{ST13SA} \emph{numeric} Surface Area (aparc.stats) of LeftBankssts Range value: 465, ..., 1662
\item \strong{ST13TA} \emph{numeric} Thickness Average (aparc.stats) of LeftBankssts Range value: 1.526, ..., 3.192
\item \strong{ST13TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftBankssts Range value: 0.304, ..., 0.902
\item \strong{ST14CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 192, ..., 5408
\item \strong{ST14SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 125, ..., 2623
\item \strong{ST14TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 1.097, ..., 3.935
\item \strong{ST14TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 0.304, ..., 1.604
\item \strong{ST15CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalMiddleFrontal Range value: 2733, ..., 9771
\item \strong{ST15SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalMiddleFrontal Range value: 969, ..., 3600
\item \strong{ST15TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1.532, ..., 3.011
\item \strong{ST15TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalMiddleFrontal Range value: 0.403, ..., 1.079
\item \strong{ST16SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCaudate Range value: 681, ..., 7224
\item \strong{ST17SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumCortex Range value: 31232, ..., 77103
\item \strong{ST18SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumWM Range value: 6028, ..., 25744
\item \strong{ST1SV} \emph{numeric} Subcortical Volume (aseg.stats) of Brainstem Range value: 13800, ..., 39756
\item \strong{ST21SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftChoroidPlexus Range value: 723, ..., 3725
\item \strong{ST23CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCuneus Range value: 760, ..., 5281
\item \strong{ST23SA} \emph{numeric} Surface Area (aparc.stats) of LeftCuneus Range value: 447, ..., 2143
\item \strong{ST23TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCuneus Range value: 1.363, ..., 3.185
\item \strong{ST23TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCuneus Range value: 0.28, ..., 1.082
\item \strong{ST24CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftEntorhinal Range value: 34, ..., 4239
\item \strong{ST24SA} \emph{numeric} Surface Area (aparc.stats) of LeftEntorhinal Range value: 11, ..., 1262
\item \strong{ST24TA} \emph{numeric} Thickness Average (aparc.stats) of LeftEntorhinal Range value: 1.442, ..., 4.475
\item \strong{ST24TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftEntorhinal Range value: 0.436, ..., 1.356
\item \strong{ST25CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFrontalPole Range value: 153, ..., 1524
\item \strong{ST25SA} \emph{numeric} Surface Area (aparc.stats) of LeftFrontalPole Range value: 59, ..., 384
\item \strong{ST25TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFrontalPole Range value: 1.484, ..., 3.727
\item \strong{ST25TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFrontalPole Range value: 0.263, ..., 1.08
\item \strong{ST26CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFusiform Range value: 2154, ..., 15382
\item \strong{ST26SA} \emph{numeric} Surface Area (aparc.stats) of LeftFusiform Range value: 1014, ..., 4622
\item \strong{ST26TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFusiform Range value: 1.484, ..., 3.267
\item \strong{ST26TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFusiform Range value: 0.473, ..., 1.061
\item \strong{ST28SA} \emph{numeric} Surface Area (aparc.stats) of LeftHemisphereWM Range value: 48423.8, ..., 108243
\item \strong{ST29SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftHippocampus Range value: 783, ..., 5991
\item \strong{ST2SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumAnterior Range value: 23, ..., 1300
\item \strong{ST30SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftInferiorLateralVentricle Range value: 11, ..., 7804
\item \strong{ST31CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorParietal Range value: 4481, ..., 16669
\item \strong{ST31SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorParietal Range value: 2247, ..., 6589
\item \strong{ST31TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorParietal Range value: 1.435, ..., 2.901
\item \strong{ST31TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorParietal Range value: 0.39, ..., 1.131
\item \strong{ST32CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorTemporal Range value: 3646, ..., 16183
\item \strong{ST32SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorTemporal Range value: 1283, ..., 4789
\item \strong{ST32TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorTemporal Range value: 1.6, ..., 3.373
\item \strong{ST32TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorTemporal Range value: 0.463, ..., 1.125
\item \strong{ST34CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftIsthmusCingulate Range value: 126, ..., 5563
\item \strong{ST34SA} \emph{numeric} Surface Area (aparc.stats) of LeftIsthmusCingulate Range value: 52, ..., 3286
\item \strong{ST34TA} \emph{numeric} Thickness Average (aparc.stats) of LeftIsthmusCingulate Range value: 1.527, ..., 3.191
\item \strong{ST34TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftIsthmusCingulate Range value: 0.531, ..., 1.231
\item \strong{ST35CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOccipital Range value: 5553, ..., 18263
\item \strong{ST35SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOccipital Range value: 3044, ..., 7580
\item \strong{ST35TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOccipital Range value: 1.437, ..., 2.692
\item \strong{ST35TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOccipital Range value: 0.398, ..., 1.154
\item \strong{ST36CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOrbitofrontal Range value: 3294, ..., 10214
\item \strong{ST36SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOrbitofrontal Range value: 864, ..., 3616
\item \strong{ST36TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOrbitofrontal Range value: 1.846, ..., 3.325
\item \strong{ST36TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOrbitofrontal Range value: 0.496, ..., 1.13
\item \strong{ST37SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftLateralVentricle Range value: 2726, ..., 75680
\item \strong{ST38CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLingual Range value: 1845, ..., 10008
\item \strong{ST38SA} \emph{numeric} Surface Area (aparc.stats) of LeftLingual Range value: 793, ..., 4507
\item \strong{ST38TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLingual Range value: 1.252, ..., 2.446
\item \strong{ST38TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLingual Range value: 0.301, ..., 0.926
\item \strong{ST39CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMedialOrbitofrontal Range value: 1073, ..., 8498
\item \strong{ST39SA} \emph{numeric} Surface Area (aparc.stats) of LeftMedialOrbitofrontal Range value: 439, ..., 2948
\item \strong{ST39TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMedialOrbitofrontal Range value: 1.593, ..., 3.565
\item \strong{ST39TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMedialOrbitofrontal Range value: 0.47, ..., 1.264
\item \strong{ST3SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumCentral Range value: 20, ..., 652
\item \strong{ST40CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMiddleTemporal Range value: 3747, ..., 15187
\item \strong{ST40SA} \emph{numeric} Surface Area (aparc.stats) of LeftMiddleTemporal Range value: 1483, ..., 4408
\item \strong{ST40TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMiddleTemporal Range value: 1.656, ..., 3.265
\item \strong{ST40TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMiddleTemporal Range value: 0.446, ..., 1.138
\item \strong{ST42SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPallidum Range value: 662, ..., 2490
\item \strong{ST43CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParacentral Range value: 1491, ..., 8234
\item \strong{ST43SA} \emph{numeric} Surface Area (aparc.stats) of LeftParacentral Range value: 797, ..., 4534
\item \strong{ST43TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParacentral Range value: 1.477, ..., 2.842
\item \strong{ST43TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParacentral Range value: 0.406, ..., 1.008
\item \strong{ST44CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParahippocampal Range value: 268, ..., 3899
\item \strong{ST44SA} \emph{numeric} Surface Area (aparc.stats) of LeftParahippocampal Range value: 94, ..., 1432
\item \strong{ST44TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParahippocampal Range value: 1.478, ..., 3.811
\item \strong{ST44TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParahippocampal Range value: 0.452, ..., 1.24
\item \strong{ST45CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOpercularis Range value: 2138, ..., 7442
\item \strong{ST45SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOpercularis Range value: 876, ..., 2747
\item \strong{ST45TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOpercularis Range value: 1.732, ..., 2.989
\item \strong{ST45TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOpercularis Range value: 0.334, ..., 1.111
\item \strong{ST46CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOrbitalis Range value: 868, ..., 3036
\item \strong{ST46SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOrbitalis Range value: 282, ..., 900
\item \strong{ST46TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOrbitalis Range value: 1.585, ..., 3.825
\item \strong{ST46TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOrbitalis Range value: 0.354, ..., 1.254
\item \strong{ST47CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsTriangularis Range value: 1450, ..., 5498
\item \strong{ST47SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsTriangularis Range value: 695, ..., 1932
\item \strong{ST47TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsTriangularis Range value: 1.617, ..., 2.905
\item \strong{ST47TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsTriangularis Range value: 0.345, ..., 0.987
\item \strong{ST48CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPericalcarine Range value: 835, ..., 3266
\item \strong{ST48SA} \emph{numeric} Surface Area (aparc.stats) of LeftPericalcarine Range value: 461, ..., 2018
\item \strong{ST48TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPericalcarine Range value: 1.15, ..., 2.407
\item \strong{ST48TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPericalcarine Range value: 0.231, ..., 1.221
\item \strong{ST49CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPostcentral Range value: 5108, ..., 16198
\item \strong{ST49SA} \emph{numeric} Surface Area (aparc.stats) of LeftPostcentral Range value: 2103, ..., 8405
\item \strong{ST49TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPostcentral Range value: 1.406, ..., 2.358
\item \strong{ST49TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPostcentral Range value: 0.451, ..., 0.982
\item \strong{ST4SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidAnterior Range value: 34, ..., 641
\item \strong{ST50CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPosteriorCingulate Range value: 1564, ..., 8797
\item \strong{ST50SA} \emph{numeric} Surface Area (aparc.stats) of LeftPosteriorCingulate Range value: 654, ..., 4744
\item \strong{ST50TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPosteriorCingulate Range value: 1.099, ..., 3.069
\item \strong{ST50TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPosteriorCingulate Range value: 0.354, ..., 1.096
\item \strong{ST51CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecentral Range value: 6421, ..., 18718
\item \strong{ST51SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecentral Range value: 3155, ..., 9185
\item \strong{ST51TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecentral Range value: 1.493, ..., 2.853
\item \strong{ST51TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecentral Range value: 0.464, ..., 0.914
\item \strong{ST52CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecuneus Range value: 4200, ..., 13145
\item \strong{ST52SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecuneus Range value: 1841, ..., 6238
\item \strong{ST52TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecuneus Range value: 1.488, ..., 2.806
\item \strong{ST52TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecuneus Range value: 0.454, ..., 0.99
\item \strong{ST53SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPutamen Range value: 2198, ..., 10872
\item \strong{ST54CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralAnteriorCingulate Range value: 131, ..., 4391
\item \strong{ST54SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralAnteriorCingulate Range value: 65, ..., 1674
\item \strong{ST54TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralAnteriorCingulate Range value: 1.051, ..., 4.24
\item \strong{ST54TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralAnteriorCingulate Range value: 0.409, ..., 1.398
\item \strong{ST55CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralMiddleFrontal Range value: 7088, ..., 22143
\item \strong{ST55SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralMiddleFrontal Range value: 2528, ..., 8679
\item \strong{ST55TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralMiddleFrontal Range value: 1.568, ..., 2.803
\item \strong{ST55TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralMiddleFrontal Range value: 0.439, ..., 1.218
\item \strong{ST56CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorFrontal Range value: 10433, ..., 28758
\item \strong{ST56SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorFrontal Range value: 4412, ..., 12474
\item \strong{ST56TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorFrontal Range value: 1.705, ..., 3.169
\item \strong{ST56TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorFrontal Range value: 0.481, ..., 1.15
\item \strong{ST57CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorParietal Range value: 5483, ..., 19630
\item \strong{ST57SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorParietal Range value: 2885, ..., 7882
\item \strong{ST57TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorParietal Range value: 1.408, ..., 2.651
\item \strong{ST57TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorParietal Range value: 0.381, ..., 1.079
\item \strong{ST58CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorTemporal Range value: 4377, ..., 15316
\item \strong{ST58SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorTemporal Range value: 1871, ..., 5157
\item \strong{ST58TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorTemporal Range value: 1.559, ..., 3.108
\item \strong{ST58TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorTemporal Range value: 0.495, ..., 0.959
\item \strong{ST59CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSupramarginal Range value: 4689, ..., 15118
\item \strong{ST59SA} \emph{numeric} Surface Area (aparc.stats) of LeftSupramarginal Range value: 1994, ..., 5721
\item \strong{ST59TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSupramarginal Range value: 1.523, ..., 2.891
\item \strong{ST59TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSupramarginal Range value: 0.428, ..., 1.027
\item \strong{ST5SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidPosterior Range value: 64, ..., 615
\item \strong{ST60CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTemporalPole Range value: 595, ..., 3750
\item \strong{ST60SA} \emph{numeric} Surface Area (aparc.stats) of LeftTemporalPole Range value: 241, ..., 737
\item \strong{ST60TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTemporalPole Range value: 1.297, ..., 4.7
\item \strong{ST60TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTemporalPole Range value: 0.291, ..., 1.314
\item \strong{ST61SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftThalamus Range value: 2273, ..., 11881
\item \strong{ST62CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTransverseTemporal Range value: 481, ..., 2014
\item \strong{ST62SA} \emph{numeric} Surface Area (aparc.stats) of LeftTransverseTemporal Range value: 207, ..., 785
\item \strong{ST62TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTransverseTemporal Range value: 1.383, ..., 2.905
\item \strong{ST62TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTransverseTemporal Range value: 0.229, ..., 0.845
\item \strong{ST65SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVentralDC Range value: 2198, ..., 5514
\item \strong{ST66SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVessel Range value: 2, ..., 639
\item \strong{ST68SV} \emph{numeric} Subcortical Volume (aseg.stats) of NonWMHypoIntensities Range value: 1, ..., 1331
\item \strong{ST69SV} \emph{numeric} Subcortical Volume (aseg.stats) of OpticChiasm Range value: 5, ..., 595
\item \strong{ST6SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumPosterior Range value: 27, ..., 1530
\item \strong{ST70SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAccumbensArea Range value: 133, ..., 911
\item \strong{ST71SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAmygdala Range value: 458, ..., 2586
\item \strong{ST72CV} \emph{numeric} Cortical Volume (aparc.stats) of RightBankssts Range value: 919, ..., 3628
\item \strong{ST72SA} \emph{numeric} Surface Area (aparc.stats) of RightBankssts Range value: 367, ..., 1501
\item \strong{ST72TA} \emph{numeric} Thickness Average (aparc.stats) of RightBankssts Range value: 1.534, ..., 3.013
\item \strong{ST72TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightBankssts Range value: 0.305, ..., 0.927
\item \strong{ST73CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalAnteriorCingulate Range value: 2, ..., 8350
\item \strong{ST73SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalAnteriorCingulate Range value: 3, ..., 3815
\item \strong{ST73TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalAnteriorCingulate Range value: 1.034, ..., 4.111
\item \strong{ST73TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalAnteriorCingulate Range value: 0.114, ..., 1.503
\item \strong{ST74CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalMiddleFrontal Range value: 2130, ..., 9246
\item \strong{ST74SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalMiddleFrontal Range value: 786, ..., 3338
\item \strong{ST74TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalMiddleFrontal Range value: 1.543, ..., 3.019
\item \strong{ST74TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalMiddleFrontal Range value: 0.398, ..., 1.019
\item \strong{ST75SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCaudate Range value: 519, ..., 7366
\item \strong{ST76SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumCortex Range value: 31387, ..., 79226
\item \strong{ST77SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumWM Range value: 6805, ..., 28552
\item \strong{ST7SV} \emph{numeric} Subcortical Volume (aseg.stats) of Csf Range value: 315, ..., 5418
\item \strong{ST80SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightChoroidPlexus Range value: 797, ..., 5295
\item \strong{ST82CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCuneus Range value: 861, ..., 6593
\item \strong{ST82SA} \emph{numeric} Surface Area (aparc.stats) of RightCuneus Range value: 575, ..., 3727
\item \strong{ST82TA} \emph{numeric} Thickness Average (aparc.stats) of RightCuneus Range value: 1.329, ..., 2.573
\item \strong{ST82TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCuneus Range value: 0.296, ..., 1.039
\item \strong{ST83CV} \emph{numeric} Cortical Volume (aparc.stats) of RightEntorhinal Range value: 275, ..., 3602
\item \strong{ST83SA} \emph{numeric} Surface Area (aparc.stats) of RightEntorhinal Range value: 121, ..., 891
\item \strong{ST83TA} \emph{numeric} Thickness Average (aparc.stats) of RightEntorhinal Range value: 1.355, ..., 4.611
\item \strong{ST83TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightEntorhinal Range value: 0.362, ..., 1.411
\item \strong{ST84CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFrontalPole Range value: 355, ..., 2073
\item \strong{ST84SA} \emph{numeric} Surface Area (aparc.stats) of RightFrontalPole Range value: 149, ..., 463
\item \strong{ST84TA} \emph{numeric} Thickness Average (aparc.stats) of RightFrontalPole Range value: 1.588, ..., 3.791
\item \strong{ST84TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFrontalPole Range value: 0.282, ..., 1.188
\item \strong{ST85CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFusiform Range value: 3953, ..., 17226
\item \strong{ST85SA} \emph{numeric} Surface Area (aparc.stats) of RightFusiform Range value: 1530, ..., 6592
\item \strong{ST85TA} \emph{numeric} Thickness Average (aparc.stats) of RightFusiform Range value: 1.554, ..., 3.28
\item \strong{ST85TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFusiform Range value: 0.446, ..., 1
\item \strong{ST87SA} \emph{numeric} Surface Area (aparc.stats) of RightHemisphereWM Range value: 55905.6, ..., 113666
\item \strong{ST88SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightHippocampus Range value: 1262, ..., 5806
\item \strong{ST89SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightInferiorLateralVentricle Range value: 21, ..., 8454
\item \strong{ST8SV} \emph{numeric} Subcortical Volume (aseg.stats) of FifthVentricle Range value: 1, ..., 207
\item \strong{ST90CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorParietal Range value: 5873, ..., 22049
\item \strong{ST90SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorParietal Range value: 2632, ..., 7916
\item \strong{ST90TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorParietal Range value: 1.452, ..., 2.902
\item \strong{ST90TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorParietal Range value: 0.374, ..., 0.988
\item \strong{ST91CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorTemporal Range value: 3354, ..., 15511
\item \strong{ST91SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorTemporal Range value: 1562, ..., 4729
\item \strong{ST91TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorTemporal Range value: 1.523, ..., 3.415
\item \strong{ST91TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorTemporal Range value: 0.482, ..., 1.097
\item \strong{ST93CV} \emph{numeric} Cortical Volume (aparc.stats) of RightIsthmusCingulate Range value: 261, ..., 4132
\item \strong{ST93SA} \emph{numeric} Surface Area (aparc.stats) of RightIsthmusCingulate Range value: 97, ..., 1973
\item \strong{ST93TA} \emph{numeric} Thickness Average (aparc.stats) of RightIsthmusCingulate Range value: 1.327, ..., 3.218
\item \strong{ST93TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightIsthmusCingulate Range value: 0.471, ..., 1.181
\item \strong{ST94CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOccipital Range value: 5650, ..., 17485
\item \strong{ST94SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOccipital Range value: 2646, ..., 6772
\item \strong{ST94TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOccipital Range value: 1.277, ..., 3.2
\item \strong{ST94TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOccipital Range value: 0.349, ..., 1.014
\item \strong{ST95CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOrbitofrontal Range value: 4041, ..., 10724
\item \strong{ST95SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOrbitofrontal Range value: 1345, ..., 3706
\item \strong{ST95TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOrbitofrontal Range value: 1.868, ..., 3.181
\item \strong{ST95TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOrbitofrontal Range value: 0.55, ..., 1.25
\item \strong{ST96SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightLateralVentricle Range value: 2003, ..., 66551
\item \strong{ST97CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLingual Range value: 3298, ..., 9620
\item \strong{ST97SA} \emph{numeric} Surface Area (aparc.stats) of RightLingual Range value: 1429, ..., 4469
\item \strong{ST97TA} \emph{numeric} Thickness Average (aparc.stats) of RightLingual Range value: 1.479, ..., 2.497
\item \strong{ST97TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLingual Range value: 0.376, ..., 0.957
\item \strong{ST98CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMedialOrbitofrontal Range value: 2192, ..., 7963
\item \strong{ST98SA} \emph{numeric} Surface Area (aparc.stats) of RightMedialOrbitofrontal Range value: 795, ..., 2473
\item \strong{ST98TA} \emph{numeric} Thickness Average (aparc.stats) of RightMedialOrbitofrontal Range value: 1.612, ..., 3.412
\item \strong{ST98TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMedialOrbitofrontal Range value: 0.482, ..., 1.222
\item \strong{ST99CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMiddleTemporal Range value: 4231, ..., 17046
\item \strong{ST99SA} \emph{numeric} Surface Area (aparc.stats) of RightMiddleTemporal Range value: 1774, ..., 5023
\item \strong{ST99TA} \emph{numeric} Thickness Average (aparc.stats) of RightMiddleTemporal Range value: 1.663, ..., 3.386
\item \strong{ST99TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMiddleTemporal Range value: 0.459, ..., 1.081
\item \strong{ST9SV} \emph{numeric} Subcortical Volume (aseg.stats) of FourthVentricle Range value: 543, ..., 6335
\item \strong{ST131HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA1 Range value: 106.503, ..., 601.008
\item \strong{ST132HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA2_3 Range value: 163.538, ..., 1327.07
\item \strong{ST133HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA4_DG Range value: 98.274, ..., 809.597
\item \strong{ST134HS} \emph{numeric} Hippocampal Subfields Volume of LeftFimbria Range value: 0.0902, ..., 131.295
\item \strong{ST135HS} \emph{numeric} Hippocampal Subfields Volume of LeftHippocampalFissure Range value: 2.04999, ..., 205.409
\item \strong{ST136HS} \emph{numeric} Hippocampal Subfields Volume of LeftPresubiculum Range value: 101.846, ..., 712.184
\item \strong{ST137HS} \emph{numeric} Hippocampal Subfields Volume of LeftSubiculum Range value: 152.292, ..., 935.025
\item \strong{ST138HS} \emph{numeric} Hippocampal Subfields Volume of LeftTail Range value: 68.378, ..., 539.759
\item \strong{ST139HS} \emph{numeric} Hippocampal Subfields Volume of RightCA1 Range value: 112.973, ..., 563.518
\item \strong{ST140HS} \emph{numeric} Hippocampal Subfields Volume of RightCA2_3 Range value: 334.062, ..., 1448.71
\item \strong{ST141HS} \emph{numeric} Hippocampal Subfields Volume of RightCA4_DG Range value: 189.601, ..., 941.514
\item \strong{ST142HS} \emph{numeric} Hippocampal Subfields Volume of RightFimbria Range value: 0.0367625, ..., 132.21
\item \strong{ST143HS} \emph{numeric} Hippocampal Subfields Volume of RightHippocampalFissure Range value: 1.44755, ..., 228.395
\item \strong{ST144HS} \emph{numeric} Hippocampal Subfields Volume of RightPresubiculum Range value: 132.63, ..., 689.449
\item \strong{ST145HS} \emph{numeric} Hippocampal Subfields Volume of RightSubiculum Range value: 194.494, ..., 917.576
\item \strong{ST146HS} \emph{numeric} Hippocampal Subfields Volume of RightTail Range value: 3.44019, ..., 537.036
\item \strong{ST147SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalGM Range value: 118569, ..., 279616
\item \strong{ST148SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalGM Range value: 123321, ..., 284310
\item \strong{ST149SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalGM Range value: 247294, ..., 559892
\item \strong{ST150SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalWM Range value: 124111, ..., 350784
\item \strong{ST151SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalWM Range value: 25546, ..., 350209
\item \strong{ST152SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalWM Range value: 227798, ..., 700367
\item \strong{ST153SV} \emph{numeric} Subcortical Volume (aseg.stats) of SubcorticalGM Range value: 116668, ..., 223265
\item \strong{ST154SV} \emph{numeric} Subcortical Volume (aseg.stats) of TotalGM Range value: 363962, ..., 773898
\item \strong{ST155SV} \emph{numeric} Subcortical Volume (aseg.stats) of SupraTentorial Range value: 635098, ..., 1394760
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adnigo}
\keyword{imaging}
