% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFFSX51_ADNI1_3T}
\alias{UCSFFSX51_ADNI1_3T}
\title{ADNI-1 3T Cross-Sectional FreeSurfer (5.1): 2016-02-01}
\format{
A data frame with 484 observations and 395 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFFSX51_ADNI1_3T)
}
\description{
ADNI-1 3T Cross-Sectional FreeSurfer (5.1): 2016-02-01 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 15, ..., 1387
\item \strong{VISCODE} \emph{character} Visit code , with options: m36, m06, m24, m12, bl, m18
\item \strong{EXAMDATE} \emph{ } The scan date for the image processed Date: YYY-MM-DD
\item \strong{VERSION} \emph{ }   Date: YYY-MM-DD
\item \strong{LONISID} \emph{numeric} The LONI study UID for the image processed Range value: 2613, ..., 24916
\item \strong{LONIUID} \emph{numeric} The LONI series UID for the image processed. Range value: 8833, ..., 79768
\item \strong{IMAGEUID} \emph{numeric} The ImageUID for the image processed. Range value: 32755, ..., 205567
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   complete/partial
\item \strong{OVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. Pass/Fail/Partial
\item \strong{TEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) Pass/Fail
\item \strong{FRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) Pass/Fail
\item \strong{PARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) Pass/Fail
\item \strong{INSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) Pass/Fail
\item \strong{OCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) Pass/Fail
\item \strong{BGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) Pass/Fail
\item \strong{CWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87) Pass/Fail
\item \strong{VENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) Pass/Fail
\item \strong{LHIPQC} \emph{logical} QC rating of the left hippocampal subfields. Fail affects the following regions: Left CA1 (ST131HS); Left CA2-3 (ST132HS); Left CA4-DG (ST133HS); Left Fimbria (ST134HS); Left Hippocampal Fissure (ST135HS); Left Presubiculum (ST136HS); Left Subiculum (ST137HS); Left Tail (ST138HS) Pass/Fail
\item \strong{RHIPQC} \emph{logical} QC rating of the right hippocampal subfields. Fail affects the following regions: Right CA1 (ST139HS); Right CA2-3 (ST140HS); Right CA4-DG (ST141HS); Right Fimbria (ST142HS); Right Hippocampal Fissure (ST143HS); Right Presubiculum (ST144HS); Right Subiculum (ST145HS); Right Tail (ST146HS) Pass/Fail
\item \strong{ST100SV} \emph{logical} Volume (aseg.stat) of RightNonWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST101SV} \emph{numeric} Volume (aseg.stat) of RightPallidum Range value: 672, ..., 2136
\item \strong{ST102CV} \emph{numeric} Volume (aparc.stat) of RightParacentral Range value: 2278, ..., 8039
\item \strong{ST102SA} \emph{numeric} Surface Area of RightParacentral Range value: 1017, ..., 3319
\item \strong{ST102TA} \emph{numeric} Cortical Thickness Average of RightParacentral Range value: 1.752, ..., 2.606
\item \strong{ST102TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParacentral Range value: 0.445, ..., 0.92
\item \strong{ST103CV} \emph{numeric} Volume (aparc.stat) of RightParahippocampal Range value: 939, ..., 2877
\item \strong{ST103SA} \emph{numeric} Surface Area of RightParahippocampal Range value: 380, ..., 922
\item \strong{ST103TA} \emph{numeric} Cortical Thickness Average of RightParahippocampal Range value: 1.544, ..., 3.469
\item \strong{ST103TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParahippocampal Range value: 0.461, ..., 1.161
\item \strong{ST104CV} \emph{numeric} Volume (aparc.stat) of RightParsOpercularis Range value: 2182, ..., 6837
\item \strong{ST104SA} \emph{numeric} Surface Area of RightParsOpercularis Range value: 884, ..., 2389
\item \strong{ST104TA} \emph{numeric} Cortical Thickness Average of RightParsOpercularis Range value: 1.956, ..., 2.853
\item \strong{ST104TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsOpercularis Range value: 0.357, ..., 0.923
\item \strong{ST105CV} \emph{numeric} Volume (aparc.stat) of RightParsOrbitalis Range value: 1249, ..., 3294
\item \strong{ST105SA} \emph{numeric} Surface Area of RightParsOrbitalis Range value: 406, ..., 1033
\item \strong{ST105TA} \emph{numeric} Cortical Thickness Average of RightParsOrbitalis Range value: 1.79, ..., 3.167
\item \strong{ST105TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsOrbitalis Range value: 0.409, ..., 1.071
\item \strong{ST106CV} \emph{numeric} Volume (aparc.stat) of RightParsTriangularis Range value: 2070, ..., 5669
\item \strong{ST106SA} \emph{numeric} Surface Area of RightParsTriangularis Range value: 790, ..., 2083
\item \strong{ST106TA} \emph{numeric} Cortical Thickness Average of RightParsTriangularis Range value: 1.839, ..., 2.658
\item \strong{ST106TS} \emph{numeric} Cortical Thickness Standard Deviation of RightParsTriangularis Range value: 0.401, ..., 0.856
\item \strong{ST107CV} \emph{numeric} Volume (aparc.stat) of RightPericalcarine Range value: 1112, ..., 3703
\item \strong{ST107SA} \emph{numeric} Surface Area of RightPericalcarine Range value: 727, ..., 2235
\item \strong{ST107TA} \emph{numeric} Cortical Thickness Average of RightPericalcarine Range value: 1.283, ..., 2.082
\item \strong{ST107TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPericalcarine Range value: 0.272, ..., 0.824
\item \strong{ST108CV} \emph{numeric} Volume (aparc.stat) of RightPostcentral Range value: 5115, ..., 13460
\item \strong{ST108SA} \emph{numeric} Surface Area of RightPostcentral Range value: 2856, ..., 7009
\item \strong{ST108TA} \emph{numeric} Cortical Thickness Average of RightPostcentral Range value: 1.463, ..., 2.294
\item \strong{ST108TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPostcentral Range value: 0.427, ..., 0.965
\item \strong{ST109CV} \emph{numeric} Volume (aparc.stat) of RightPosteriorCingulate Range value: 1463, ..., 6174
\item \strong{ST109SA} \emph{numeric} Surface Area of RightPosteriorCingulate Range value: 717, ..., 2935
\item \strong{ST109TA} \emph{numeric} Cortical Thickness Average of RightPosteriorCingulate Range value: 1.323, ..., 2.859
\item \strong{ST109TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPosteriorCingulate Range value: 0.418, ..., 1.075
\item \strong{ST10CV} \emph{numeric} Volume (aparc.stat) of Icv Range value: 1192530, ..., 2062120
\item \strong{ST110CV} \emph{numeric} Volume (aparc.stat) of RightPrecentral Range value: 8213, ..., 16905
\item \strong{ST110SA} \emph{numeric} Surface Area of RightPrecentral Range value: 3497, ..., 8090
\item \strong{ST110TA} \emph{numeric} Cortical Thickness Average of RightPrecentral Range value: 1.821, ..., 2.725
\item \strong{ST110TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPrecentral Range value: 0.442, ..., 0.819
\item \strong{ST111CV} \emph{numeric} Volume (aparc.stat) of RightPrecuneus Range value: 5938, ..., 13963
\item \strong{ST111SA} \emph{numeric} Surface Area of RightPrecuneus Range value: 2630, ..., 6734
\item \strong{ST111TA} \emph{numeric} Cortical Thickness Average of RightPrecuneus Range value: 1.631, ..., 2.573
\item \strong{ST111TS} \emph{numeric} Cortical Thickness Standard Deviation of RightPrecuneus Range value: 0.452, ..., 0.776
\item \strong{ST112SV} \emph{numeric} Volume (aseg.stat) of RightPutamen Range value: 3011, ..., 7365
\item \strong{ST113CV} \emph{numeric} Volume (aparc.stat) of RightRostralAnteriorCingulate Range value: 411, ..., 3471
\item \strong{ST113SA} \emph{numeric} Surface Area of RightRostralAnteriorCingulate Range value: 166, ..., 1374
\item \strong{ST113TA} \emph{numeric} Cortical Thickness Average of RightRostralAnteriorCingulate Range value: 1.937, ..., 4.083
\item \strong{ST113TS} \emph{numeric} Cortical Thickness Standard Deviation of RightRostralAnteriorCingulate Range value: 0.393, ..., 1.445
\item \strong{ST114CV} \emph{numeric} Volume (aparc.stat) of RightRostralMiddleFrontal Range value: 7958, ..., 20115
\item \strong{ST114SA} \emph{numeric} Surface Area of RightRostralMiddleFrontal Range value: 2593, ..., 7871
\item \strong{ST114TA} \emph{numeric} Cortical Thickness Average of RightRostralMiddleFrontal Range value: 1.662, ..., 2.674
\item \strong{ST114TS} \emph{numeric} Cortical Thickness Standard Deviation of RightRostralMiddleFrontal Range value: 0.473, ..., 0.945
\item \strong{ST115CV} \emph{numeric} Volume (aparc.stat) of RightSuperiorFrontal Range value: 12216, ..., 34908
\item \strong{ST115SA} \emph{numeric} Surface Area of RightSuperiorFrontal Range value: 4446, ..., 13409
\item \strong{ST115TA} \emph{numeric} Cortical Thickness Average of RightSuperiorFrontal Range value: 1.978, ..., 2.84
\item \strong{ST115TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorFrontal Range value: 0.513, ..., 1.134
\item \strong{ST116CV} \emph{numeric} Volume (aparc.stat) of RightSuperiorParietal Range value: 7369, ..., 16457
\item \strong{ST116SA} \emph{numeric} Surface Area of RightSuperiorParietal Range value: 3565, ..., 7303
\item \strong{ST116TA} \emph{numeric} Cortical Thickness Average of RightSuperiorParietal Range value: 1.587, ..., 2.443
\item \strong{ST116TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorParietal Range value: 0.427, ..., 0.848
\item \strong{ST117CV} \emph{numeric} Volume (aparc.stat) of RightSuperiorTemporal Range value: 5341, ..., 12972
\item \strong{ST117SA} \emph{numeric} Surface Area of RightSuperiorTemporal Range value: 2424, ..., 4344
\item \strong{ST117TA} \emph{numeric} Cortical Thickness Average of RightSuperiorTemporal Range value: 1.829, ..., 2.944
\item \strong{ST117TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSuperiorTemporal Range value: 0.498, ..., 0.901
\item \strong{ST118CV} \emph{numeric} Volume (aparc.stat) of RightSupramarginal Range value: 5085, ..., 13256
\item \strong{ST118SA} \emph{numeric} Surface Area of RightSupramarginal Range value: 2315, ..., 5578
\item \strong{ST118TA} \emph{numeric} Cortical Thickness Average of RightSupramarginal Range value: 1.743, ..., 2.766
\item \strong{ST118TS} \emph{numeric} Cortical Thickness Standard Deviation of RightSupramarginal Range value: 0.452, ..., 0.898
\item \strong{ST119CV} \emph{numeric} Volume (aparc.stat) of RightTemporalPole Range value: 784, ..., 3103
\item \strong{ST119SA} \emph{numeric} Surface Area of RightTemporalPole Range value: 215, ..., 564
\item \strong{ST119TA} \emph{numeric} Cortical Thickness Average of RightTemporalPole Range value: 1.6, ..., 4.435
\item \strong{ST119TS} \emph{numeric} Cortical Thickness Standard Deviation of RightTemporalPole Range value: 0.33, ..., 1.173
\item \strong{ST11SV} \emph{numeric} Volume (aseg.stat) of LeftAccumbensArea Range value: 44, ..., 874
\item \strong{ST120SV} \emph{numeric} Volume (aseg.stat) of RightThalamus Range value: 2973, ..., 8815
\item \strong{ST121CV} \emph{numeric} Volume (aparc.stat) of RightTransverseTemporal Range value: 439, ..., 1278
\item \strong{ST121SA} \emph{numeric} Surface Area of RightTransverseTemporal Range value: 174, ..., 543
\item \strong{ST121TA} \emph{numeric} Cortical Thickness Average of RightTransverseTemporal Range value: 1.502, ..., 2.865
\item \strong{ST121TS} \emph{numeric} Cortical Thickness Standard Deviation of RightTransverseTemporal Range value: 0.25, ..., 0.628
\item \strong{ST122SV} \emph{logical} Volume (aseg.stat) of RightUndetermined A Boolean value: TRUE/ FALSE
\item \strong{ST123CV} \emph{logical} Volume (aparc.stat) of RightUnknown A Boolean value: TRUE/ FALSE
\item \strong{ST123SA} \emph{logical} Surface Area of RightUnknown A Boolean value: TRUE/ FALSE
\item \strong{ST123TA} \emph{logical} Cortical Thickness Average of RightUnknown A Boolean value: TRUE/ FALSE
\item \strong{ST123TS} \emph{logical} Cortical Thickness Standard Deviation of RightUnknown A Boolean value: TRUE/ FALSE
\item \strong{ST124SV} \emph{numeric} Volume (aseg.stat) of RightVentralDC Range value: 2248, ..., 4641
\item \strong{ST125SV} \emph{numeric} Volume (aseg.stat) of RightVessel Range value: 2, ..., 554
\item \strong{ST126SV} \emph{logical} Volume (aseg.stat) of RightWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST127SV} \emph{numeric} Volume (aseg.stat) of ThirdVentricle Range value: 596, ..., 4664
\item \strong{ST128SV} \emph{numeric} Volume (aseg.stat) of WMHypoIntensities Range value: 1043, ..., 71649
\item \strong{ST129CV} \emph{numeric} Volume (aparc.stat) of LeftInsula Range value: 4492, ..., 10817
\item \strong{ST129SA} \emph{numeric} Surface Area of LeftInsula Range value: 1526, ..., 3847
\item \strong{ST129TA} \emph{numeric} Cortical Thickness Average of LeftInsula Range value: 2.341, ..., 3.419
\item \strong{ST129TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInsula Range value: 0.698, ..., 1.251
\item \strong{ST12SV} \emph{numeric} Volume (aseg.stat) of LeftAmygdala Range value: 485, ..., 2024
\item \strong{ST130CV} \emph{numeric} Volume (aparc.stat) of RightInsula Range value: 4489, ..., 9244
\item \strong{ST130SA} \emph{numeric} Surface Area of RightInsula Range value: 1521, ..., 3105
\item \strong{ST130TA} \emph{numeric} Cortical Thickness Average of RightInsula Range value: 2.289, ..., 3.477
\item \strong{ST130TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInsula Range value: 0.685, ..., 1.208
\item \strong{ST13CV} \emph{numeric} Volume (aparc.stat) of LeftBankssts Range value: 1180, ..., 5193
\item \strong{ST13SA} \emph{numeric} Surface Area of LeftBankssts Range value: 551, ..., 2146
\item \strong{ST13TA} \emph{numeric} Cortical Thickness Average of LeftBankssts Range value: 1.495, ..., 2.753
\item \strong{ST13TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftBankssts Range value: 0.323, ..., 0.771
\item \strong{ST14CV} \emph{numeric} Volume (aparc.stat) of LeftCaudalAnteriorCingulate Range value: 709, ..., 5935
\item \strong{ST14SA} \emph{numeric} Surface Area of LeftCaudalAnteriorCingulate Range value: 371, ..., 3841
\item \strong{ST14TA} \emph{numeric} Cortical Thickness Average of LeftCaudalAnteriorCingulate Range value: 1.168, ..., 3.66
\item \strong{ST14TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCaudalAnteriorCingulate Range value: 0.394, ..., 1.539
\item \strong{ST15CV} \emph{numeric} Volume (aparc.stat) of LeftCaudalMiddleFrontal Range value: 3103, ..., 9330
\item \strong{ST15SA} \emph{numeric} Surface Area of LeftCaudalMiddleFrontal Range value: 1266, ..., 3204
\item \strong{ST15TA} \emph{numeric} Cortical Thickness Average of LeftCaudalMiddleFrontal Range value: 1.736, ..., 2.819
\item \strong{ST15TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCaudalMiddleFrontal Range value: 0.427, ..., 1.006
\item \strong{ST16SV} \emph{numeric} Volume (aseg.stat) of LeftCaudate Range value: 942, ..., 6750
\item \strong{ST17SV} \emph{numeric} Volume (aseg.stat) of LeftCerebellumCortex Range value: 32850, ..., 61960
\item \strong{ST18SV} \emph{numeric} Volume (aseg.stat) of LeftCerebellumWM Range value: 7163, ..., 22432
\item \strong{ST19SV} \emph{logical} Volume (aseg.stat) of LeftCerebralCortex A Boolean value: TRUE/ FALSE
\item \strong{ST1SV} \emph{numeric} Volume (aseg.stat) of Brainstem Range value: 15212, ..., 27790
\item \strong{ST20SV} \emph{logical} Volume (aseg.stat) of LeftCerebralWM A Boolean value: TRUE/ FALSE
\item \strong{ST21SV} \emph{numeric} Volume (aseg.stat) of LeftChoroidPlexus Range value: 844, ..., 4126
\item \strong{ST22CV} \emph{logical} Volume (aparc.stat) of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22SA} \emph{logical} Surface Area of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22TA} \emph{logical} Cortical Thickness Average of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST22TS} \emph{logical} Cortical Thickness Standard Deviation of LeftCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST23CV} \emph{numeric} Volume (aparc.stat) of LeftCuneus Range value: 1730, ..., 4778
\item \strong{ST23SA} \emph{numeric} Surface Area of LeftCuneus Range value: 905, ..., 2366
\item \strong{ST23TA} \emph{numeric} Cortical Thickness Average of LeftCuneus Range value: 1.409, ..., 2.244
\item \strong{ST23TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftCuneus Range value: 0.305, ..., 1.101
\item \strong{ST24CV} \emph{numeric} Volume (aparc.stat) of LeftEntorhinal Range value: 304, ..., 3077
\item \strong{ST24SA} \emph{numeric} Surface Area of LeftEntorhinal Range value: 76, ..., 1035
\item \strong{ST24TA} \emph{numeric} Cortical Thickness Average of LeftEntorhinal Range value: 1.518, ..., 4.178
\item \strong{ST24TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftEntorhinal Range value: 0.447, ..., 1.242
\item \strong{ST25CV} \emph{numeric} Volume (aparc.stat) of LeftFrontalPole Range value: 405, ..., 1432
\item \strong{ST25SA} \emph{numeric} Surface Area of LeftFrontalPole Range value: 114, ..., 339
\item \strong{ST25TA} \emph{numeric} Cortical Thickness Average of LeftFrontalPole Range value: 1.776, ..., 3.954
\item \strong{ST25TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftFrontalPole Range value: 0.275, ..., 1.178
\item \strong{ST26CV} \emph{numeric} Volume (aparc.stat) of LeftFusiform Range value: 4667, ..., 13598
\item \strong{ST26SA} \emph{numeric} Surface Area of LeftFusiform Range value: 1799, ..., 4222
\item \strong{ST26TA} \emph{numeric} Cortical Thickness Average of LeftFusiform Range value: 1.896, ..., 3.045
\item \strong{ST26TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftFusiform Range value: 0.496, ..., 0.959
\item \strong{ST27SA} \emph{logical} Surface Area of LeftHemisphere A Boolean value: TRUE/ FALSE
\item \strong{ST28CV} \emph{logical} Volume (aparc.stat) of LeftHemisphereWM A Boolean value: TRUE/ FALSE
\item \strong{ST28SA} \emph{numeric} Surface Area of LeftHemisphereWM Range value: 57591.2, ..., 105344
\item \strong{ST29SV} \emph{numeric} Volume (aseg.stat) of LeftHippocampus Range value: 1517, ..., 4641
\item \strong{ST2SV} \emph{numeric} Volume (aseg.stat) of CorpusCallosumAnterior Range value: 6, ..., 1158
\item \strong{ST30SV} \emph{numeric} Volume (aseg.stat) of LeftInferiorLateralVentricle Range value: 37, ..., 6839
\item \strong{ST31CV} \emph{numeric} Volume (aparc.stat) of LeftInferiorParietal Range value: 5660, ..., 19299
\item \strong{ST31SA} \emph{numeric} Surface Area of LeftInferiorParietal Range value: 2611, ..., 7138
\item \strong{ST31TA} \emph{numeric} Cortical Thickness Average of LeftInferiorParietal Range value: 1.615, ..., 2.748
\item \strong{ST31TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInferiorParietal Range value: 0.43, ..., 0.96
\item \strong{ST32CV} \emph{numeric} Volume (aparc.stat) of LeftInferiorTemporal Range value: 5331, ..., 14311
\item \strong{ST32SA} \emph{numeric} Surface Area of LeftInferiorTemporal Range value: 1924, ..., 4351
\item \strong{ST32TA} \emph{numeric} Cortical Thickness Average of LeftInferiorTemporal Range value: 1.967, ..., 3.094
\item \strong{ST32TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftInferiorTemporal Range value: 0.554, ..., 1.051
\item \strong{ST33SV} \emph{logical} Volume (aseg.stat) of LeftInterior A Boolean value: TRUE/ FALSE
\item \strong{ST34CV} \emph{numeric} Volume (aparc.stat) of LeftIsthmusCingulate Range value: 671, ..., 5441
\item \strong{ST34SA} \emph{numeric} Surface Area of LeftIsthmusCingulate Range value: 407, ..., 2863
\item \strong{ST34TA} \emph{numeric} Cortical Thickness Average of LeftIsthmusCingulate Range value: 1.674, ..., 3.129
\item \strong{ST34TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftIsthmusCingulate Range value: 0.565, ..., 1.057
\item \strong{ST35CV} \emph{numeric} Volume (aparc.stat) of LeftLateralOccipital Range value: 6276, ..., 16199
\item \strong{ST35SA} \emph{numeric} Surface Area of LeftLateralOccipital Range value: 3157, ..., 7020
\item \strong{ST35TA} \emph{numeric} Cortical Thickness Average of LeftLateralOccipital Range value: 1.583, ..., 2.39
\item \strong{ST35TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLateralOccipital Range value: 0.424, ..., 0.832
\item \strong{ST36CV} \emph{numeric} Volume (aparc.stat) of LeftLateralOrbitofrontal Range value: 4570, ..., 15797
\item \strong{ST36SA} \emph{numeric} Surface Area of LeftLateralOrbitofrontal Range value: 1435, ..., 5881
\item \strong{ST36TA} \emph{numeric} Cortical Thickness Average of LeftLateralOrbitofrontal Range value: 2.011, ..., 3.033
\item \strong{ST36TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLateralOrbitofrontal Range value: 0.513, ..., 1.025
\item \strong{ST37SV} \emph{numeric} Volume (aseg.stat) of LeftLateralVentricle Range value: 3549, ..., 83251
\item \strong{ST38CV} \emph{numeric} Volume (aparc.stat) of LeftLingual Range value: 3335, ..., 9793
\item \strong{ST38SA} \emph{numeric} Surface Area of LeftLingual Range value: 1769, ..., 4854
\item \strong{ST38TA} \emph{numeric} Cortical Thickness Average of LeftLingual Range value: 1.502, ..., 2.298
\item \strong{ST38TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftLingual Range value: 0.449, ..., 0.82
\item \strong{ST39CV} \emph{numeric} Volume (aparc.stat) of LeftMedialOrbitofrontal Range value: 2819, ..., 7447
\item \strong{ST39SA} \emph{numeric} Surface Area of LeftMedialOrbitofrontal Range value: 994, ..., 3278
\item \strong{ST39TA} \emph{numeric} Cortical Thickness Average of LeftMedialOrbitofrontal Range value: 1.665, ..., 3.151
\item \strong{ST39TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftMedialOrbitofrontal Range value: 0.475, ..., 1.109
\item \strong{ST3SV} \emph{numeric} Volume (aseg.stat) of CorpusCallosumCentral Range value: 109, ..., 551
\item \strong{ST40CV} \emph{numeric} Volume (aparc.stat) of LeftMiddleTemporal Range value: 4857, ..., 13107
\item \strong{ST40SA} \emph{numeric} Surface Area of LeftMiddleTemporal Range value: 1877, ..., 3866
\item \strong{ST40TA} \emph{numeric} Cortical Thickness Average of LeftMiddleTemporal Range value: 1.82, ..., 3.061
\item \strong{ST40TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftMiddleTemporal Range value: 0.505, ..., 0.919
\item \strong{ST41SV} \emph{logical} Volume (aseg.stat) of LeftNonWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST42SV} \emph{numeric} Volume (aseg.stat) of LeftPallidum Range value: 798, ..., 2281
\item \strong{ST43CV} \emph{numeric} Volume (aparc.stat) of LeftParacentral Range value: 2021, ..., 7248
\item \strong{ST43SA} \emph{numeric} Surface Area of LeftParacentral Range value: 897, ..., 2963
\item \strong{ST43TA} \emph{numeric} Cortical Thickness Average of LeftParacentral Range value: 1.825, ..., 2.986
\item \strong{ST43TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParacentral Range value: 0.43, ..., 1.004
\item \strong{ST44CV} \emph{numeric} Volume (aparc.stat) of LeftParahippocampal Range value: 190, ..., 3745
\item \strong{ST44SA} \emph{numeric} Surface Area of LeftParahippocampal Range value: 59, ..., 1225
\item \strong{ST44TA} \emph{numeric} Cortical Thickness Average of LeftParahippocampal Range value: 1.373, ..., 3.555
\item \strong{ST44TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParahippocampal Range value: 0.407, ..., 1.138
\item \strong{ST45CV} \emph{numeric} Volume (aparc.stat) of LeftParsOpercularis Range value: 2469, ..., 7172
\item \strong{ST45SA} \emph{numeric} Surface Area of LeftParsOpercularis Range value: 951, ..., 2590
\item \strong{ST45TA} \emph{numeric} Cortical Thickness Average of LeftParsOpercularis Range value: 1.835, ..., 2.736
\item \strong{ST45TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsOpercularis Range value: 0.352, ..., 0.736
\item \strong{ST46CV} \emph{numeric} Volume (aparc.stat) of LeftParsOrbitalis Range value: 1205, ..., 4127
\item \strong{ST46SA} \emph{numeric} Surface Area of LeftParsOrbitalis Range value: 412, ..., 1510
\item \strong{ST46TA} \emph{numeric} Cortical Thickness Average of LeftParsOrbitalis Range value: 1.882, ..., 3.092
\item \strong{ST46TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsOrbitalis Range value: 0.426, ..., 1.008
\item \strong{ST47CV} \emph{numeric} Volume (aparc.stat) of LeftParsTriangularis Range value: 1804, ..., 4644
\item \strong{ST47SA} \emph{numeric} Surface Area of LeftParsTriangularis Range value: 784, ..., 1737
\item \strong{ST47TA} \emph{numeric} Cortical Thickness Average of LeftParsTriangularis Range value: 1.591, ..., 2.66
\item \strong{ST47TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftParsTriangularis Range value: 0.402, ..., 0.834
\item \strong{ST48CV} \emph{numeric} Volume (aparc.stat) of LeftPericalcarine Range value: 1044, ..., 6520
\item \strong{ST48SA} \emph{numeric} Surface Area of LeftPericalcarine Range value: 763, ..., 3417
\item \strong{ST48TA} \emph{numeric} Cortical Thickness Average of LeftPericalcarine Range value: 1.279, ..., 2.205
\item \strong{ST48TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPericalcarine Range value: 0.276, ..., 1.105
\item \strong{ST49CV} \emph{numeric} Volume (aparc.stat) of LeftPostcentral Range value: 5509, ..., 14650
\item \strong{ST49SA} \emph{numeric} Surface Area of LeftPostcentral Range value: 3017, ..., 7626
\item \strong{ST49TA} \emph{numeric} Cortical Thickness Average of LeftPostcentral Range value: 1.484, ..., 2.313
\item \strong{ST49TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPostcentral Range value: 0.45, ..., 0.945
\item \strong{ST4SV} \emph{numeric} Volume (aseg.stat) of CorpusCallosumMidAnterior Range value: 58, ..., 609
\item \strong{ST50CV} \emph{numeric} Volume (aparc.stat) of LeftPosteriorCingulate Range value: 1161, ..., 11108
\item \strong{ST50SA} \emph{numeric} Surface Area of LeftPosteriorCingulate Range value: 700, ..., 6621
\item \strong{ST50TA} \emph{numeric} Cortical Thickness Average of LeftPosteriorCingulate Range value: 1.415, ..., 2.821
\item \strong{ST50TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPosteriorCingulate Range value: 0.442, ..., 1.072
\item \strong{ST51CV} \emph{numeric} Volume (aparc.stat) of LeftPrecentral Range value: 8255, ..., 17971
\item \strong{ST51SA} \emph{numeric} Surface Area of LeftPrecentral Range value: 3298, ..., 7614
\item \strong{ST51TA} \emph{numeric} Cortical Thickness Average of LeftPrecentral Range value: 1.773, ..., 2.772
\item \strong{ST51TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPrecentral Range value: 0.474, ..., 0.822
\item \strong{ST52CV} \emph{numeric} Volume (aparc.stat) of LeftPrecuneus Range value: 4980, ..., 15097
\item \strong{ST52SA} \emph{numeric} Surface Area of LeftPrecuneus Range value: 2363, ..., 7434
\item \strong{ST52TA} \emph{numeric} Cortical Thickness Average of LeftPrecuneus Range value: 1.471, ..., 2.761
\item \strong{ST52TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftPrecuneus Range value: 0.422, ..., 0.819
\item \strong{ST53SV} \emph{numeric} Volume (aseg.stat) of LeftPutamen Range value: 3110, ..., 7305
\item \strong{ST54CV} \emph{numeric} Volume (aparc.stat) of LeftRostralAnteriorCingulate Range value: 996, ..., 4020
\item \strong{ST54SA} \emph{numeric} Surface Area of LeftRostralAnteriorCingulate Range value: 377, ..., 1324
\item \strong{ST54TA} \emph{numeric} Cortical Thickness Average of LeftRostralAnteriorCingulate Range value: 1.342, ..., 4.111
\item \strong{ST54TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftRostralAnteriorCingulate Range value: 0.43, ..., 1.723
\item \strong{ST55CV} \emph{numeric} Volume (aparc.stat) of LeftRostralMiddleFrontal Range value: 8027, ..., 17858
\item \strong{ST55SA} \emph{numeric} Surface Area of LeftRostralMiddleFrontal Range value: 3238, ..., 7611
\item \strong{ST55TA} \emph{numeric} Cortical Thickness Average of LeftRostralMiddleFrontal Range value: 1.573, ..., 2.688
\item \strong{ST55TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftRostralMiddleFrontal Range value: 0.442, ..., 0.887
\item \strong{ST56CV} \emph{numeric} Volume (aparc.stat) of LeftSuperiorFrontal Range value: 13493, ..., 28428
\item \strong{ST56SA} \emph{numeric} Surface Area of LeftSuperiorFrontal Range value: 4831, ..., 10607
\item \strong{ST56TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorFrontal Range value: 1.943, ..., 3.047
\item \strong{ST56TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorFrontal Range value: 0.51, ..., 1.111
\item \strong{ST57CV} \emph{numeric} Volume (aparc.stat) of LeftSuperiorParietal Range value: 7013, ..., 17283
\item \strong{ST57SA} \emph{numeric} Surface Area of LeftSuperiorParietal Range value: 3518, ..., 8449
\item \strong{ST57TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorParietal Range value: 1.538, ..., 2.639
\item \strong{ST57TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorParietal Range value: 0.421, ..., 0.968
\item \strong{ST58CV} \emph{numeric} Volume (aparc.stat) of LeftSuperiorTemporal Range value: 6187, ..., 16677
\item \strong{ST58SA} \emph{numeric} Surface Area of LeftSuperiorTemporal Range value: 2387, ..., 5889
\item \strong{ST58TA} \emph{numeric} Cortical Thickness Average of LeftSuperiorTemporal Range value: 1.718, ..., 2.955
\item \strong{ST58TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSuperiorTemporal Range value: 0.481, ..., 0.933
\item \strong{ST59CV} \emph{numeric} Volume (aparc.stat) of LeftSupramarginal Range value: 5819, ..., 13173
\item \strong{ST59SA} \emph{numeric} Surface Area of LeftSupramarginal Range value: 2476, ..., 5044
\item \strong{ST59TA} \emph{numeric} Cortical Thickness Average of LeftSupramarginal Range value: 1.69, ..., 2.727
\item \strong{ST59TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftSupramarginal Range value: 0.432, ..., 0.912
\item \strong{ST5SV} \emph{numeric} Volume (aseg.stat) of CorpusCallosumMidPosterior Range value: 61, ..., 516
\item \strong{ST60CV} \emph{numeric} Volume (aparc.stat) of LeftTemporalPole Range value: 939, ..., 3414
\item \strong{ST60SA} \emph{numeric} Surface Area of LeftTemporalPole Range value: 268, ..., 730
\item \strong{ST60TA} \emph{numeric} Cortical Thickness Average of LeftTemporalPole Range value: 1.995, ..., 4.396
\item \strong{ST60TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftTemporalPole Range value: 0.355, ..., 1.174
\item \strong{ST61SV} \emph{numeric} Volume (aseg.stat) of LeftThalamus Range value: 3386, ..., 9068
\item \strong{ST62CV} \emph{numeric} Volume (aparc.stat) of LeftTransverseTemporal Range value: 558, ..., 1906
\item \strong{ST62SA} \emph{numeric} Surface Area of LeftTransverseTemporal Range value: 254, ..., 863
\item \strong{ST62TA} \emph{numeric} Cortical Thickness Average of LeftTransverseTemporal Range value: 1.501, ..., 2.678
\item \strong{ST62TS} \emph{numeric} Cortical Thickness Standard Deviation of LeftTransverseTemporal Range value: 0.248, ..., 0.623
\item \strong{ST63SV} \emph{logical} Volume (aseg.stat) of LeftUndetermined A Boolean value: TRUE/ FALSE
\item \strong{ST64CV} \emph{logical} Volume (aparc.stat) of LeftUnknown A Boolean value: TRUE/ FALSE
\item \strong{ST64SA} \emph{logical} Surface Area of LeftUnknown A Boolean value: TRUE/ FALSE
\item \strong{ST64TA} \emph{logical} Cortical Thickness Average of LeftUnknown A Boolean value: TRUE/ FALSE
\item \strong{ST64TS} \emph{logical} Cortical Thickness Standard Deviation of LeftUnknown A Boolean value: TRUE/ FALSE
\item \strong{ST65SV} \emph{numeric} Volume (aseg.stat) of LeftVentralDC Range value: 2272, ..., 4879
\item \strong{ST66SV} \emph{numeric} Volume (aseg.stat) of LeftVessel Range value: 21, ..., 381
\item \strong{ST67SV} \emph{logical} Volume (aseg.stat) of LeftWMHypoIntensities A Boolean value: TRUE/ FALSE
\item \strong{ST68SV} \emph{numeric} Volume (aseg.stat) of NonWMHypoIntensities Range value: 1, ..., 779
\item \strong{ST69SV} \emph{numeric} Volume (aseg.stat) of OpticChiasm Range value: 78, ..., 488
\item \strong{ST6SV} \emph{numeric} Volume (aseg.stat) of CorpusCallosumPosterior Range value: 41, ..., 1439
\item \strong{ST70SV} \emph{numeric} Volume (aseg.stat) of RightAccumbensArea Range value: 109, ..., 987
\item \strong{ST71SV} \emph{numeric} Volume (aseg.stat) of RightAmygdala Range value: 573, ..., 2617
\item \strong{ST72CV} \emph{numeric} Volume (aparc.stat) of RightBankssts Range value: 991, ..., 3787
\item \strong{ST72SA} \emph{numeric} Surface Area of RightBankssts Range value: 553, ..., 1614
\item \strong{ST72TA} \emph{numeric} Cortical Thickness Average of RightBankssts Range value: 1.674, ..., 2.767
\item \strong{ST72TS} \emph{numeric} Cortical Thickness Standard Deviation of RightBankssts Range value: 0.308, ..., 0.769
\item \strong{ST73CV} \emph{numeric} Volume (aparc.stat) of RightCaudalAnteriorCingulate Range value: 1011, ..., 7533
\item \strong{ST73SA} \emph{numeric} Surface Area of RightCaudalAnteriorCingulate Range value: 421, ..., 3791
\item \strong{ST73TA} \emph{numeric} Cortical Thickness Average of RightCaudalAnteriorCingulate Range value: 1.303, ..., 3.371
\item \strong{ST73TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCaudalAnteriorCingulate Range value: 0.439, ..., 1.306
\item \strong{ST74CV} \emph{numeric} Volume (aparc.stat) of RightCaudalMiddleFrontal Range value: 2436, ..., 8891
\item \strong{ST74SA} \emph{numeric} Surface Area of RightCaudalMiddleFrontal Range value: 1024, ..., 3457
\item \strong{ST74TA} \emph{numeric} Cortical Thickness Average of RightCaudalMiddleFrontal Range value: 1.811, ..., 2.846
\item \strong{ST74TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCaudalMiddleFrontal Range value: 0.426, ..., 0.882
\item \strong{ST75SV} \emph{numeric} Volume (aseg.stat) of RightCaudate Range value: 535, ..., 6641
\item \strong{ST76SV} \emph{numeric} Volume (aseg.stat) of RightCerebellumCortex Range value: 34109, ..., 63633
\item \strong{ST77SV} \emph{numeric} Volume (aseg.stat) of RightCerebellumWM Range value: 7049, ..., 22200
\item \strong{ST78SV} \emph{logical} Volume (aseg.stat) of RightCerebralCortex A Boolean value: TRUE/ FALSE
\item \strong{ST79SV} \emph{logical} Volume (aseg.stat) of RightCerebralWM A Boolean value: TRUE/ FALSE
\item \strong{ST7SV} \emph{numeric} Volume (aseg.stat) of Csf Range value: 720, ..., 3327
\item \strong{ST80SV} \emph{numeric} Volume (aseg.stat) of RightChoroidPlexus Range value: 1215, ..., 4900
\item \strong{ST81CV} \emph{logical} Volume (aparc.stat) of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81SA} \emph{logical} Surface Area of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81TA} \emph{logical} Cortical Thickness Average of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST81TS} \emph{logical} Cortical Thickness Standard Deviation of RightCorpusCallosum A Boolean value: TRUE/ FALSE
\item \strong{ST82CV} \emph{numeric} Volume (aparc.stat) of RightCuneus Range value: 1774, ..., 4285
\item \strong{ST82SA} \emph{numeric} Surface Area of RightCuneus Range value: 887, ..., 2171
\item \strong{ST82TA} \emph{numeric} Cortical Thickness Average of RightCuneus Range value: 1.402, ..., 2.215
\item \strong{ST82TS} \emph{numeric} Cortical Thickness Standard Deviation of RightCuneus Range value: 0.341, ..., 0.766
\item \strong{ST83CV} \emph{numeric} Volume (aparc.stat) of RightEntorhinal Range value: 457, ..., 2887
\item \strong{ST83SA} \emph{numeric} Surface Area of RightEntorhinal Range value: 154, ..., 573
\item \strong{ST83TA} \emph{numeric} Cortical Thickness Average of RightEntorhinal Range value: 1.702, ..., 4.33
\item \strong{ST83TS} \emph{numeric} Cortical Thickness Standard Deviation of RightEntorhinal Range value: 0.485, ..., 1.255
\item \strong{ST84CV} \emph{numeric} Volume (aparc.stat) of RightFrontalPole Range value: 529, ..., 1847
\item \strong{ST84SA} \emph{numeric} Surface Area of RightFrontalPole Range value: 161, ..., 395
\item \strong{ST84TA} \emph{numeric} Cortical Thickness Average of RightFrontalPole Range value: 1.672, ..., 3.563
\item \strong{ST84TS} \emph{numeric} Cortical Thickness Standard Deviation of RightFrontalPole Range value: 0.292, ..., 0.986
\item \strong{ST85CV} \emph{numeric} Volume (aparc.stat) of RightFusiform Range value: 5297, ..., 12136
\item \strong{ST85SA} \emph{numeric} Surface Area of RightFusiform Range value: 1894, ..., 4124
\item \strong{ST85TA} \emph{numeric} Cortical Thickness Average of RightFusiform Range value: 1.938, ..., 3.131
\item \strong{ST85TS} \emph{numeric} Cortical Thickness Standard Deviation of RightFusiform Range value: 0.535, ..., 0.983
\item \strong{ST86SA} \emph{logical} Surface Area of RightHemisphere A Boolean value: TRUE/ FALSE
\item \strong{ST87CV} \emph{logical} Volume (aparc.stat) of RightHemisphereWM A Boolean value: TRUE/ FALSE
\item \strong{ST87SA} \emph{numeric} Surface Area of RightHemisphereWM Range value: 63064.9, ..., 111366
\item \strong{ST88SV} \emph{numeric} Volume (aseg.stat) of RightHippocampus Range value: 1690, ..., 4603
\item \strong{ST89SV} \emph{numeric} Volume (aseg.stat) of RightInferiorLateralVentricle Range value: 110, ..., 6791
\item \strong{ST8SV} \emph{numeric} Volume (aseg.stat) of FifthVentricle Range value: 1, ..., 174
\item \strong{ST90CV} \emph{numeric} Volume (aparc.stat) of RightInferiorParietal Range value: 6978, ..., 19130
\item \strong{ST90SA} \emph{numeric} Surface Area of RightInferiorParietal Range value: 3241, ..., 7098
\item \strong{ST90TA} \emph{numeric} Cortical Thickness Average of RightInferiorParietal Range value: 1.678, ..., 2.622
\item \strong{ST90TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInferiorParietal Range value: 0.449, ..., 0.935
\item \strong{ST91CV} \emph{numeric} Volume (aparc.stat) of RightInferiorTemporal Range value: 4079, ..., 13313
\item \strong{ST91SA} \emph{numeric} Surface Area of RightInferiorTemporal Range value: 1591, ..., 4517
\item \strong{ST91TA} \emph{numeric} Cortical Thickness Average of RightInferiorTemporal Range value: 1.884, ..., 3.158
\item \strong{ST91TS} \emph{numeric} Cortical Thickness Standard Deviation of RightInferiorTemporal Range value: 0.564, ..., 1.015
\item \strong{ST92SV} \emph{logical} Volume (aseg.stat) of RightInterior A Boolean value: TRUE/ FALSE
\item \strong{ST93CV} \emph{numeric} Volume (aparc.stat) of RightIsthmusCingulate Range value: 1264, ..., 3525
\item \strong{ST93SA} \emph{numeric} Surface Area of RightIsthmusCingulate Range value: 549, ..., 1736
\item \strong{ST93TA} \emph{numeric} Cortical Thickness Average of RightIsthmusCingulate Range value: 1.634, ..., 2.912
\item \strong{ST93TS} \emph{numeric} Cortical Thickness Standard Deviation of RightIsthmusCingulate Range value: 0.518, ..., 1.081
\item \strong{ST94CV} \emph{numeric} Volume (aparc.stat) of RightLateralOccipital Range value: 6177, ..., 16293
\item \strong{ST94SA} \emph{numeric} Surface Area of RightLateralOccipital Range value: 2917, ..., 6632
\item \strong{ST94TA} \emph{numeric} Cortical Thickness Average of RightLateralOccipital Range value: 1.634, ..., 2.873
\item \strong{ST94TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLateralOccipital Range value: 0.447, ..., 1.029
\item \strong{ST95CV} \emph{numeric} Volume (aparc.stat) of RightLateralOrbitofrontal Range value: 4579, ..., 9118
\item \strong{ST95SA} \emph{numeric} Surface Area of RightLateralOrbitofrontal Range value: 1351, ..., 3468
\item \strong{ST95TA} \emph{numeric} Cortical Thickness Average of RightLateralOrbitofrontal Range value: 1.879, ..., 3.215
\item \strong{ST95TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLateralOrbitofrontal Range value: 0.532, ..., 1.068
\item \strong{ST96SV} \emph{numeric} Volume (aseg.stat) of RightLateralVentricle Range value: 3695, ..., 61548
\item \strong{ST97CV} \emph{numeric} Volume (aparc.stat) of RightLingual Range value: 4119, ..., 9326
\item \strong{ST97SA} \emph{numeric} Surface Area of RightLingual Range value: 1848, ..., 4520
\item \strong{ST97TA} \emph{numeric} Cortical Thickness Average of RightLingual Range value: 1.584, ..., 2.314
\item \strong{ST97TS} \emph{numeric} Cortical Thickness Standard Deviation of RightLingual Range value: 0.451, ..., 0.809
\item \strong{ST98CV} \emph{numeric} Volume (aparc.stat) of RightMedialOrbitofrontal Range value: 2817, ..., 6895
\item \strong{ST98SA} \emph{numeric} Surface Area of RightMedialOrbitofrontal Range value: 973, ..., 2681
\item \strong{ST98TA} \emph{numeric} Cortical Thickness Average of RightMedialOrbitofrontal Range value: 1.689, ..., 3.128
\item \strong{ST98TS} \emph{numeric} Cortical Thickness Standard Deviation of RightMedialOrbitofrontal Range value: 0.523, ..., 1.058
\item \strong{ST99CV} \emph{numeric} Volume (aparc.stat) of RightMiddleTemporal Range value: 5596, ..., 14974
\item \strong{ST99SA} \emph{numeric} Surface Area of RightMiddleTemporal Range value: 2120, ..., 4460
\item \strong{ST99TA} \emph{numeric} Cortical Thickness Average of RightMiddleTemporal Range value: 2.136, ..., 3.118
\item \strong{ST99TS} \emph{numeric} Cortical Thickness Standard Deviation of RightMiddleTemporal Range value: 0.465, ..., 0.956
\item \strong{ST9SV} \emph{numeric} Volume (aseg.stat) of FourthVentricle Range value: 817, ..., 5180
\item \strong{ST131HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA1 Range value: 111.736, ..., 456.932
\item \strong{ST132HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA2_3 Range value: 366.395, ..., 1375.12
\item \strong{ST133HS} \emph{numeric} Hippocampal Subfields Volume of LeftCA4_DG Range value: 191.989, ..., 751.57
\item \strong{ST134HS} \emph{numeric} Hippocampal Subfields Volume of LeftFimbria Range value: 0.283825, ..., 141.091
\item \strong{ST135HS} \emph{numeric} Hippocampal Subfields Volume of LeftHippocampalFissure Range value: 2.12646, ..., 164.438
\item \strong{ST136HS} \emph{numeric} Hippocampal Subfields Volume of LeftPresubiculum Range value: 165.947, ..., 586.997
\item \strong{ST137HS} \emph{numeric} Hippocampal Subfields Volume of LeftSubiculum Range value: 231.42, ..., 768.6
\item \strong{ST138HS} \emph{numeric} Hippocampal Subfields Volume of LeftTail Range value: 105.745, ..., 449.677
\item \strong{ST139HS} \emph{numeric} Hippocampal Subfields Volume of RightCA1 Range value: 164.964, ..., 595.12
\item \strong{ST140HS} \emph{numeric} Hippocampal Subfields Volume of RightCA2_3 Range value: 472.269, ..., 1321.25
\item \strong{ST141HS} \emph{numeric} Hippocampal Subfields Volume of RightCA4_DG Range value: 241.851, ..., 709.459
\item \strong{ST142HS} \emph{numeric} Hippocampal Subfields Volume of RightFimbria Range value: 0.481375, ..., 124.775
\item \strong{ST143HS} \emph{numeric} Hippocampal Subfields Volume of RightHippocampalFissure Range value: 0.22255, ..., 130.961
\item \strong{ST144HS} \emph{numeric} Hippocampal Subfields Volume of RightPresubiculum Range value: 191.094, ..., 574.492
\item \strong{ST145HS} \emph{numeric} Hippocampal Subfields Volume of RightSubiculum Range value: 256.677, ..., 768.523
\item \strong{ST146HS} \emph{numeric} Hippocampal Subfields Volume of RightTail Range value: 160.268, ..., 502.573
\item \strong{ST147SV} \emph{numeric} Volume (aseg.stat) of LeftCorticalGM Range value: 140225, ..., 270535
\item \strong{ST148SV} \emph{numeric} Volume (aseg.stat) of RightCorticalGM Range value: 143489, ..., 265083
\item \strong{ST149SV} \emph{numeric} Volume (aseg.stat) of CorticalGM Range value: 283714, ..., 535618
\item \strong{ST150SV} \emph{numeric} Volume (aseg.stat) of LeftCorticalWM Range value: 138820, ..., 313999
\item \strong{ST151SV} \emph{numeric} Volume (aseg.stat) of RightCorticalWM Range value: 27810.3, ..., 330330
\item \strong{ST152SV} \emph{numeric} Volume (aseg.stat) of CorticalWM Range value: 224900, ..., 633676
\item \strong{ST153SV} \emph{numeric} Volume (aseg.stat) of SubcorticalGM Range value: 125351, ..., 205261
\item \strong{ST154SV} \emph{numeric} Volume (aseg.stat) of TotalGM Range value: 413811, ..., 727638
\item \strong{ST155SV} \emph{numeric} Volume (aseg.stat) of SupraTentorial Range value: 732041, ..., 1351590
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{imaging}
