% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFFSX6}
\alias{UCSFFSX6}
\title{Cross-Sectional FreeSurfer (6.0)}
\format{
A data frame with 2270 observations and 347 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFFSX6)
}
\description{
Cross-Sectional FreeSurfer (6.0) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI3, ADNI1, ADNIGO
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI3
\item \strong{PTID} \emph{character} Participant Identifier for the in-clinic ADNI study  
\item \strong{RID} \emph{numeric} Participant Identifier for the in-clinic ADNI study Range value: 21, ..., 7123
\item \strong{VISCODE} \emph{character} VISCODE of the Image , with options: init, y1, y2, sc, y4, y3, y5
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{IMAGEUID} \emph{numeric} Image UID of image that was processed. Range value: 808819, ..., 10298187
\item \strong{EXAMDATE} \emph{ } When the image was collected. Date: YYY-MM-DD
\item \strong{RUNDATE} \emph{ } When the image was processed. Date: YYY-MM-DD
\item \strong{VERSION} \emph{ } Version of the record Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} partial = No QC, not final. Complete = QC and finalized. , with options: complete, partial
\item \strong{OVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. , with options: Pass, Partial, Fail, Hippocampus Only
\item \strong{TEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) , with options: Pass, Fail
\item \strong{FRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) , with options: Pass, Fail
\item \strong{PARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) , with options: Pass, Fail
\item \strong{INSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) , with options: Pass, Fail
\item \strong{OCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) , with options: Pass, Fail
\item \strong{BGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) , with options: Pass, Fail
\item \strong{CWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87) , with options: Pass, Fail
\item \strong{VENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) , with options: Pass, Fail
\item \strong{HIPPOQC} \emph{character} QC rating of the hippocampus. Fail affects the following regions: LeftFusiform (ST26); RightFusiform (ST85); LeftEntorhinal (ST24); RightEntorhinal (ST83); LeftParahippocampal (ST44); RightParahippocampal (ST103); LeftAmygdala (ST12); RightAmygdala (ST71) , with options: Pass, Fail
\item \strong{ST101SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPallidum Range value: 986.5, ..., 2925.2
\item \strong{ST102CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParacentral Range value: 1849, ..., 5496
\item \strong{ST102SA} \emph{numeric} Surface Area (aparc.stats) of RightParacentral Range value: 959, ..., 2181
\item \strong{ST102TA} \emph{numeric} Thickness Average (aparc.stats) of RightParacentral Range value: 1.48, ..., 2.878
\item \strong{ST102TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParacentral Range value: 0.362, ..., 0.895
\item \strong{ST103CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParahippocampal Range value: 965, ..., 3204
\item \strong{ST103SA} \emph{numeric} Surface Area (aparc.stats) of RightParahippocampal Range value: 335, ..., 995
\item \strong{ST103TA} \emph{numeric} Thickness Average (aparc.stats) of RightParahippocampal Range value: 1.58, ..., 3.644
\item \strong{ST103TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParahippocampal Range value: 0.377, ..., 1.124
\item \strong{ST104CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOpercularis Range value: 1316, ..., 5637
\item \strong{ST104SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOpercularis Range value: 626, ..., 2018
\item \strong{ST104TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOpercularis Range value: 1.798, ..., 2.951
\item \strong{ST104TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOpercularis Range value: 0.277, ..., 0.882
\item \strong{ST105CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOrbitalis Range value: 629, ..., 3917
\item \strong{ST105SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOrbitalis Range value: 236, ..., 1173
\item \strong{ST105TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOrbitalis Range value: 1.641, ..., 3.512
\item \strong{ST105TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOrbitalis Range value: 0.34, ..., 1.045
\item \strong{ST106CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsTriangularis Range value: 1150, ..., 7010
\item \strong{ST106SA} \emph{numeric} Surface Area (aparc.stats) of RightParsTriangularis Range value: 482, ..., 2336
\item \strong{ST106TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsTriangularis Range value: 1.773, ..., 3.062
\item \strong{ST106TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsTriangularis Range value: 0.321, ..., 1.1
\item \strong{ST107CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPericalcarine Range value: 853, ..., 4217
\item \strong{ST107SA} \emph{numeric} Surface Area (aparc.stats) of RightPericalcarine Range value: 681, ..., 2462
\item \strong{ST107TA} \emph{numeric} Thickness Average (aparc.stats) of RightPericalcarine Range value: 1.131, ..., 2.241
\item \strong{ST107TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPericalcarine Range value: 0.211, ..., 0.833
\item \strong{ST108CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPostcentral Range value: 2792, ..., 12514
\item \strong{ST108SA} \emph{numeric} Surface Area (aparc.stats) of RightPostcentral Range value: 1590, ..., 5401
\item \strong{ST108TA} \emph{numeric} Thickness Average (aparc.stats) of RightPostcentral Range value: 1.328, ..., 2.65
\item \strong{ST108TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPostcentral Range value: 0.319, ..., 0.768
\item \strong{ST109CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPosteriorCingulate Range value: 644, ..., 5120
\item \strong{ST109SA} \emph{numeric} Surface Area (aparc.stats) of RightPosteriorCingulate Range value: 238, ..., 3301
\item \strong{ST109TA} \emph{numeric} Thickness Average (aparc.stats) of RightPosteriorCingulate Range value: 1.437, ..., 3.125
\item \strong{ST109TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPosteriorCingulate Range value: 0.38, ..., 1.504
\item \strong{ST10CV} \emph{numeric} Cortical Volume (aparc.stats) of Icv Range value: 655448.760866, ..., 3500154.120931
\item \strong{ST110CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecentral Range value: 4429, ..., 17399
\item \strong{ST110SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecentral Range value: 2324, ..., 7614
\item \strong{ST110TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecentral Range value: 1.557, ..., 2.804
\item \strong{ST110TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecentral Range value: 0.407, ..., 0.839
\item \strong{ST111CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecuneus Range value: 4484, ..., 14376
\item \strong{ST111SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecuneus Range value: 2128, ..., 6158
\item \strong{ST111TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecuneus Range value: 1.557, ..., 2.846
\item \strong{ST111TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecuneus Range value: 0.348, ..., 0.936
\item \strong{ST112SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPutamen Range value: 1405.4, ..., 8311.7
\item \strong{ST113CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralAnteriorCingulate Range value: 252, ..., 3554
\item \strong{ST113SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralAnteriorCingulate Range value: 124, ..., 1088
\item \strong{ST113TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralAnteriorCingulate Range value: 1.832, ..., 4.02
\item \strong{ST113TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralAnteriorCingulate Range value: 0.367, ..., 1.303
\item \strong{ST114CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralMiddleFrontal Range value: 5295, ..., 26180
\item \strong{ST114SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralMiddleFrontal Range value: 2170, ..., 9138
\item \strong{ST114TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralMiddleFrontal Range value: 1.695, ..., 3.017
\item \strong{ST114TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralMiddleFrontal Range value: 0.38, ..., 0.911
\item \strong{ST115CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorFrontal Range value: 8625, ..., 28204
\item \strong{ST115SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorFrontal Range value: 3881, ..., 9299
\item \strong{ST115TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorFrontal Range value: 1.832, ..., 3.018
\item \strong{ST115TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorFrontal Range value: 0.383, ..., 0.864
\item \strong{ST116CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorParietal Range value: 5128, ..., 17512
\item \strong{ST116SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorParietal Range value: 2318, ..., 7159
\item \strong{ST116TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorParietal Range value: 1.405, ..., 2.553
\item \strong{ST116TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorParietal Range value: 0.325, ..., 0.791
\item \strong{ST117CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorTemporal Range value: 3744, ..., 15319
\item \strong{ST117SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorTemporal Range value: 1699, ..., 5150
\item \strong{ST117TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorTemporal Range value: 1.456, ..., 3.287
\item \strong{ST117TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorTemporal Range value: 0.365, ..., 0.901
\item \strong{ST118CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSupramarginal Range value: 3252, ..., 16747
\item \strong{ST118SA} \emph{numeric} Surface Area (aparc.stats) of RightSupramarginal Range value: 1465, ..., 5914
\item \strong{ST118TA} \emph{numeric} Thickness Average (aparc.stats) of RightSupramarginal Range value: 1.654, ..., 2.942
\item \strong{ST118TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSupramarginal Range value: 0.362, ..., 0.849
\item \strong{ST119CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTemporalPole Range value: 600, ..., 4014
\item \strong{ST119SA} \emph{numeric} Surface Area (aparc.stats) of RightTemporalPole Range value: 188, ..., 698
\item \strong{ST119TA} \emph{numeric} Thickness Average (aparc.stats) of RightTemporalPole Range value: 1.476, ..., 4.717
\item \strong{ST119TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTemporalPole Range value: 0.308, ..., 1.25
\item \strong{ST11SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAccumbensArea Range value: 51.9, ..., 870.2
\item \strong{ST120SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightThalamus Range value: 2465.1, ..., 10073.9
\item \strong{ST121CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTransverseTemporal Range value: 223, ..., 1494
\item \strong{ST121SA} \emph{numeric} Surface Area (aparc.stats) of RightTransverseTemporal Range value: 153, ..., 495
\item \strong{ST121TA} \emph{numeric} Thickness Average (aparc.stats) of RightTransverseTemporal Range value: 1.252, ..., 3.159
\item \strong{ST121TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTransverseTemporal Range value: 0.161, ..., 0.696
\item \strong{ST124SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVentralDC Range value: 1801.1, ..., 5852
\item \strong{ST125SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVessel Range value: 1, ..., 322.2
\item \strong{ST127SV} \emph{numeric} Subcortical Volume (aseg.stats) of ThirdVentricle Range value: 606.1, ..., 4737.2
\item \strong{ST128SV} \emph{numeric} Subcortical Volume (aseg.stats) of WMHypoIntensities Range value: 163.9, ..., 90184.8
\item \strong{ST129CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInsula Range value: 3715, ..., 9743
\item \strong{ST129SA} \emph{numeric} Surface Area (aparc.stats) of LeftInsula Range value: 1611, ..., 3358
\item \strong{ST129TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInsula Range value: 1.808, ..., 3.692
\item \strong{ST129TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInsula Range value: 0.566, ..., 1.418
\item \strong{ST12SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAmygdala Range value: 364.1, ..., 2402.9
\item \strong{ST130CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInsula Range value: 3127, ..., 10787
\item \strong{ST130SA} \emph{numeric} Surface Area (aparc.stats) of RightInsula Range value: 1386, ..., 3365
\item \strong{ST130TA} \emph{numeric} Thickness Average (aparc.stats) of RightInsula Range value: 1.967, ..., 3.509
\item \strong{ST130TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInsula Range value: 0.553, ..., 1.268
\item \strong{ST13CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftBankssts Range value: 1078, ..., 3772
\item \strong{ST13SA} \emph{numeric} Surface Area (aparc.stats) of LeftBankssts Range value: 480, ..., 1511
\item \strong{ST13TA} \emph{numeric} Thickness Average (aparc.stats) of LeftBankssts Range value: 1.55, ..., 3.008
\item \strong{ST13TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftBankssts Range value: 0.232, ..., 0.907
\item \strong{ST14CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 376, ..., 3057
\item \strong{ST14SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 238, ..., 1043
\item \strong{ST14TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 1.146, ..., 3.791
\item \strong{ST14TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 0.213, ..., 1.42
\item \strong{ST15CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalMiddleFrontal Range value: 2102, ..., 8352
\item \strong{ST15SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalMiddleFrontal Range value: 986, ..., 3474
\item \strong{ST15TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1.639, ..., 2.869
\item \strong{ST15TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalMiddleFrontal Range value: 0.314, ..., 0.901
\item \strong{ST16SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCaudate Range value: 1046.9, ..., 6813.4
\item \strong{ST17SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumCortex Range value: 26542.1, ..., 68771.4
\item \strong{ST18SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumWM Range value: 6217.9, ..., 29917.2
\item \strong{ST1SV} \emph{numeric} Subcortical Volume (aseg.stats) of Brainstem Range value: 11729.8, ..., 33846.3
\item \strong{ST21SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftChoroidPlexus Range value: 189, ..., 1855.7
\item \strong{ST23CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCuneus Range value: 1151, ..., 4611
\item \strong{ST23SA} \emph{numeric} Surface Area (aparc.stats) of LeftCuneus Range value: 782, ..., 2140
\item \strong{ST23TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCuneus Range value: 1.388, ..., 2.398
\item \strong{ST23TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCuneus Range value: 0.229, ..., 0.921
\item \strong{ST24CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftEntorhinal Range value: 351, ..., 4052
\item \strong{ST24SA} \emph{numeric} Surface Area (aparc.stats) of LeftEntorhinal Range value: 194, ..., 884
\item \strong{ST24TA} \emph{numeric} Thickness Average (aparc.stats) of LeftEntorhinal Range value: 1.122, ..., 4.494
\item \strong{ST24TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftEntorhinal Range value: 0.406, ..., 1.264
\item \strong{ST25CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFrontalPole Range value: 432, ..., 1753
\item \strong{ST25SA} \emph{numeric} Surface Area (aparc.stats) of LeftFrontalPole Range value: 158, ..., 375
\item \strong{ST25TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFrontalPole Range value: 1.651, ..., 4.239
\item \strong{ST25TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFrontalPole Range value: 0.229, ..., 0.983
\item \strong{ST26CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFusiform Range value: 4101, ..., 14409
\item \strong{ST26SA} \emph{numeric} Surface Area (aparc.stats) of LeftFusiform Range value: 1725, ..., 4197
\item \strong{ST26TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFusiform Range value: 1.458, ..., 3.269
\item \strong{ST26TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFusiform Range value: 0.468, ..., 1.157
\item \strong{ST28SA} \emph{numeric} Surface Area (aparc.stats) of LeftHemisphereWM Range value: 46627.1, ..., 105355
\item \strong{ST29SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftHippocampus Range value: 970.7, ..., 5293
\item \strong{ST2SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumAnterior Range value: 61.5, ..., 1650.3
\item \strong{ST30SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftInferiorLateralVentricle Range value: 50.7, ..., 10185.2
\item \strong{ST31CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorParietal Range value: 5343, ..., 17055
\item \strong{ST31SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorParietal Range value: 2508, ..., 6299
\item \strong{ST31TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorParietal Range value: 1.648, ..., 2.783
\item \strong{ST31TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorParietal Range value: 0.343, ..., 0.872
\item \strong{ST32CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorTemporal Range value: 4561, ..., 17165
\item \strong{ST32SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorTemporal Range value: 1633, ..., 4800
\item \strong{ST32TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorTemporal Range value: 1.824, ..., 3.449
\item \strong{ST32TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorTemporal Range value: 0.463, ..., 0.986
\item \strong{ST34CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftIsthmusCingulate Range value: 756, ..., 4633
\item \strong{ST34SA} \emph{numeric} Surface Area (aparc.stats) of LeftIsthmusCingulate Range value: 342, ..., 2041
\item \strong{ST34TA} \emph{numeric} Thickness Average (aparc.stats) of LeftIsthmusCingulate Range value: 1.408, ..., 3.276
\item \strong{ST34TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftIsthmusCingulate Range value: 0.469, ..., 1.32
\item \strong{ST35CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOccipital Range value: 4816, ..., 17317
\item \strong{ST35SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOccipital Range value: 2783, ..., 6733
\item \strong{ST35TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOccipital Range value: 1.534, ..., 2.626
\item \strong{ST35TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOccipital Range value: 0.357, ..., 0.859
\item \strong{ST36CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOrbitofrontal Range value: 3493, ..., 9696
\item \strong{ST36SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOrbitofrontal Range value: 1474, ..., 3474
\item \strong{ST36TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOrbitofrontal Range value: 1.818, ..., 3.255
\item \strong{ST36TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOrbitofrontal Range value: 0.446, ..., 1.043
\item \strong{ST37SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftLateralVentricle Range value: 3756.3, ..., 75124.4
\item \strong{ST38CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLingual Range value: 2518, ..., 10018
\item \strong{ST38SA} \emph{numeric} Surface Area (aparc.stats) of LeftLingual Range value: 1556, ..., 4313
\item \strong{ST38TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLingual Range value: 1.566, ..., 2.49
\item \strong{ST38TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLingual Range value: 0.389, ..., 0.878
\item \strong{ST39CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMedialOrbitofrontal Range value: 2983, ..., 7172
\item \strong{ST39SA} \emph{numeric} Surface Area (aparc.stats) of LeftMedialOrbitofrontal Range value: 1175, ..., 2814
\item \strong{ST39TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMedialOrbitofrontal Range value: 1.677, ..., 3.204
\item \strong{ST39TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMedialOrbitofrontal Range value: 0.487, ..., 1.198
\item \strong{ST3SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumCentral Range value: 151.3, ..., 1236.1
\item \strong{ST40CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMiddleTemporal Range value: 3228, ..., 14853
\item \strong{ST40SA} \emph{numeric} Surface Area (aparc.stats) of LeftMiddleTemporal Range value: 1320, ..., 4444
\item \strong{ST40TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMiddleTemporal Range value: 1.765, ..., 3.265
\item \strong{ST40TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMiddleTemporal Range value: 0.441, ..., 0.965
\item \strong{ST42SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPallidum Range value: 740.5, ..., 2874.3
\item \strong{ST43CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParacentral Range value: 1617, ..., 5560
\item \strong{ST43SA} \emph{numeric} Surface Area (aparc.stats) of LeftParacentral Range value: 872, ..., 2094
\item \strong{ST43TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParacentral Range value: 1.534, ..., 2.956
\item \strong{ST43TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParacentral Range value: 0.377, ..., 0.802
\item \strong{ST44CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParahippocampal Range value: 987, ..., 3305
\item \strong{ST44SA} \emph{numeric} Surface Area (aparc.stats) of LeftParahippocampal Range value: 423, ..., 998
\item \strong{ST44TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParahippocampal Range value: 1.662, ..., 3.752
\item \strong{ST44TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParahippocampal Range value: 0.403, ..., 1.195
\item \strong{ST45CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOpercularis Range value: 2293, ..., 7005
\item \strong{ST45SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOpercularis Range value: 936, ..., 2551
\item \strong{ST45TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOpercularis Range value: 1.776, ..., 2.929
\item \strong{ST45TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOpercularis Range value: 0.295, ..., 0.854
\item \strong{ST46CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOrbitalis Range value: 887, ..., 3396
\item \strong{ST46SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOrbitalis Range value: 295, ..., 1029
\item \strong{ST46TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOrbitalis Range value: 2.009, ..., 3.545
\item \strong{ST46TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOrbitalis Range value: 0.339, ..., 1.073
\item \strong{ST47CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsTriangularis Range value: 1760, ..., 5477
\item \strong{ST47SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsTriangularis Range value: 737, ..., 2069
\item \strong{ST47TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsTriangularis Range value: 1.718, ..., 2.993
\item \strong{ST47TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsTriangularis Range value: 0.305, ..., 0.958
\item \strong{ST48CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPericalcarine Range value: 842, ..., 3282
\item \strong{ST48SA} \emph{numeric} Surface Area (aparc.stats) of LeftPericalcarine Range value: 717, ..., 2167
\item \strong{ST48TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPericalcarine Range value: 1.216, ..., 2.275
\item \strong{ST48TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPericalcarine Range value: 0.161, ..., 0.748
\item \strong{ST49CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPostcentral Range value: 4693, ..., 12459
\item \strong{ST49SA} \emph{numeric} Surface Area (aparc.stats) of LeftPostcentral Range value: 2168, ..., 5799
\item \strong{ST49TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPostcentral Range value: 1.406, ..., 2.413
\item \strong{ST49TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPostcentral Range value: 0.313, ..., 0.853
\item \strong{ST4SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidAnterior Range value: 183.4, ..., 1104.6
\item \strong{ST50CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPosteriorCingulate Range value: 948, ..., 4401
\item \strong{ST50SA} \emph{numeric} Surface Area (aparc.stats) of LeftPosteriorCingulate Range value: 432, ..., 1739
\item \strong{ST50TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPosteriorCingulate Range value: 1.387, ..., 3.047
\item \strong{ST50TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPosteriorCingulate Range value: 0.393, ..., 1.434
\item \strong{ST51CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecentral Range value: 5184, ..., 17392
\item \strong{ST51SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecentral Range value: 2451, ..., 6634
\item \strong{ST51TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecentral Range value: 1.612, ..., 2.871
\item \strong{ST51TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecentral Range value: 0.404, ..., 0.817
\item \strong{ST52CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecuneus Range value: 4018, ..., 12702
\item \strong{ST52SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecuneus Range value: 1937, ..., 5424
\item \strong{ST52TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecuneus Range value: 1.536, ..., 2.871
\item \strong{ST52TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecuneus Range value: 0.364, ..., 0.897
\item \strong{ST53SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPutamen Range value: 1442.4, ..., 9201.4
\item \strong{ST54CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralAnteriorCingulate Range value: 972, ..., 4887
\item \strong{ST54SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralAnteriorCingulate Range value: 336, ..., 1516
\item \strong{ST54TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralAnteriorCingulate Range value: 1.618, ..., 3.764
\item \strong{ST54TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralAnteriorCingulate Range value: 0.407, ..., 1.698
\item \strong{ST55CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralMiddleFrontal Range value: 6016, ..., 20179
\item \strong{ST55SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralMiddleFrontal Range value: 2866, ..., 7687
\item \strong{ST55TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralMiddleFrontal Range value: 1.636, ..., 2.987
\item \strong{ST55TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralMiddleFrontal Range value: 0.366, ..., 0.845
\item \strong{ST56CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorFrontal Range value: 10909, ..., 28373
\item \strong{ST56SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorFrontal Range value: 4480, ..., 9860
\item \strong{ST56TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorFrontal Range value: 1.671, ..., 3.016
\item \strong{ST56TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorFrontal Range value: 0.377, ..., 0.888
\item \strong{ST57CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorParietal Range value: 6210, ..., 18004
\item \strong{ST57SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorParietal Range value: 2822, ..., 7190
\item \strong{ST57TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorParietal Range value: 1.467, ..., 2.549
\item \strong{ST57TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorParietal Range value: 0.327, ..., 0.799
\item \strong{ST58CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorTemporal Range value: 5125, ..., 17367
\item \strong{ST58SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorTemporal Range value: 2103, ..., 5465
\item \strong{ST58TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorTemporal Range value: 1.623, ..., 3.256
\item \strong{ST58TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorTemporal Range value: 0.454, ..., 0.917
\item \strong{ST59CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSupramarginal Range value: 4606, ..., 18580
\item \strong{ST59SA} \emph{numeric} Surface Area (aparc.stats) of LeftSupramarginal Range value: 2061, ..., 6520
\item \strong{ST59TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSupramarginal Range value: 1.741, ..., 2.835
\item \strong{ST59TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSupramarginal Range value: 0.38, ..., 0.822
\item \strong{ST5SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidPosterior Range value: 164.6, ..., 1412.3
\item \strong{ST60CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTemporalPole Range value: 711, ..., 4008
\item \strong{ST60SA} \emph{numeric} Surface Area (aparc.stats) of LeftTemporalPole Range value: 212, ..., 873
\item \strong{ST60TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTemporalPole Range value: 1.716, ..., 4.609
\item \strong{ST60TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTemporalPole Range value: 0.361, ..., 1.215
\item \strong{ST61SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftThalamus Range value: 3465.5, ..., 9886.3
\item \strong{ST62CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTransverseTemporal Range value: 394, ..., 2042
\item \strong{ST62SA} \emph{numeric} Surface Area (aparc.stats) of LeftTransverseTemporal Range value: 265, ..., 752
\item \strong{ST62TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTransverseTemporal Range value: 1.296, ..., 2.894
\item \strong{ST62TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTransverseTemporal Range value: 0.181, ..., 0.767
\item \strong{ST65SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVentralDC Range value: 1930.2, ..., 6560
\item \strong{ST66SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVessel Range value: 1, ..., 390.5
\item \strong{ST68SV} \emph{numeric} Subcortical Volume (aseg.stats) of NonWMHypoIntensities Range value: 1, ..., 340
\item \strong{ST69SV} \emph{numeric} Subcortical Volume (aseg.stats) of OpticChiasm Range value: 1.1, ..., 405.7
\item \strong{ST6SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumPosterior Range value: 101.2, ..., 1881.8
\item \strong{ST70SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAccumbensArea Range value: 145.3, ..., 1025.1
\item \strong{ST71SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAmygdala Range value: 537.1, ..., 2501.6
\item \strong{ST72CV} \emph{numeric} Cortical Volume (aparc.stats) of RightBankssts Range value: 927, ..., 3323
\item \strong{ST72SA} \emph{numeric} Surface Area (aparc.stats) of RightBankssts Range value: 372, ..., 1322
\item \strong{ST72TA} \emph{numeric} Thickness Average (aparc.stats) of RightBankssts Range value: 1.503, ..., 3.129
\item \strong{ST72TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightBankssts Range value: 0.214, ..., 0.877
\item \strong{ST73CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalAnteriorCingulate Range value: 127, ..., 5797
\item \strong{ST73SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalAnteriorCingulate Range value: 82, ..., 3581
\item \strong{ST73TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalAnteriorCingulate Range value: 1.393, ..., 3.524
\item \strong{ST73TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalAnteriorCingulate Range value: 0.301, ..., 1.54
\item \strong{ST74CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalMiddleFrontal Range value: 2107, ..., 8227
\item \strong{ST74SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalMiddleFrontal Range value: 938, ..., 3155
\item \strong{ST74TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalMiddleFrontal Range value: 1.628, ..., 2.94
\item \strong{ST74TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalMiddleFrontal Range value: 0.309, ..., 0.811
\item \strong{ST75SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCaudate Range value: 783.3, ..., 7023.1
\item \strong{ST76SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumCortex Range value: 25533.3, ..., 71683.9
\item \strong{ST77SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumWM Range value: 7438.5, ..., 31801.8
\item \strong{ST7SV} \emph{numeric} Subcortical Volume (aseg.stats) of Csf Range value: 334.1, ..., 4001.5
\item \strong{ST80SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightChoroidPlexus Range value: 250.1, ..., 1812.7
\item \strong{ST82CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCuneus Range value: 1362, ..., 5770
\item \strong{ST82SA} \emph{numeric} Surface Area (aparc.stats) of RightCuneus Range value: 794, ..., 2493
\item \strong{ST82TA} \emph{numeric} Thickness Average (aparc.stats) of RightCuneus Range value: 1.327, ..., 2.295
\item \strong{ST82TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCuneus Range value: 0.284, ..., 0.922
\item \strong{ST83CV} \emph{numeric} Cortical Volume (aparc.stats) of RightEntorhinal Range value: 403, ..., 4119
\item \strong{ST83SA} \emph{numeric} Surface Area (aparc.stats) of RightEntorhinal Range value: 196, ..., 822
\item \strong{ST83TA} \emph{numeric} Thickness Average (aparc.stats) of RightEntorhinal Range value: 1.505, ..., 4.709
\item \strong{ST83TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightEntorhinal Range value: 0.346, ..., 1.452
\item \strong{ST84CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFrontalPole Range value: 527, ..., 1986
\item \strong{ST84SA} \emph{numeric} Surface Area (aparc.stats) of RightFrontalPole Range value: 171, ..., 409
\item \strong{ST84TA} \emph{numeric} Thickness Average (aparc.stats) of RightFrontalPole Range value: 1.703, ..., 3.83
\item \strong{ST84TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFrontalPole Range value: 0.285, ..., 1.057
\item \strong{ST85CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFusiform Range value: 3594, ..., 13646
\item \strong{ST85SA} \emph{numeric} Surface Area (aparc.stats) of RightFusiform Range value: 1179, ..., 4107
\item \strong{ST85TA} \emph{numeric} Thickness Average (aparc.stats) of RightFusiform Range value: 1.632, ..., 3.367
\item \strong{ST85TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFusiform Range value: 0.48, ..., 1.009
\item \strong{ST87sa} \emph{numeric} Surface Area (aparc.stats) of RightHemisphereWM Range value: 38437.7, ..., 106267
\item \strong{ST88SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightHippocampus Range value: 1618, ..., 5461.8
\item \strong{ST89SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightInferiorLateralVentricle Range value: 62.6, ..., 5879.5
\item \strong{ST8SV} \emph{numeric} Subcortical Volume (aseg.stats) of FifthVentricle Range value: 1, ..., 16.7
\item \strong{ST90CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorParietal Range value: 4602, ..., 20869
\item \strong{ST90SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorParietal Range value: 2023, ..., 7939
\item \strong{ST90TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorParietal Range value: 1.603, ..., 2.822
\item \strong{ST90TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorParietal Range value: 0.342, ..., 0.806
\item \strong{ST91CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorTemporal Range value: 4481, ..., 16579
\item \strong{ST91SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorTemporal Range value: 1423, ..., 4829
\item \strong{ST91TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorTemporal Range value: 1.806, ..., 3.42
\item \strong{ST91TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorTemporal Range value: 0.473, ..., 1.073
\item \strong{ST93CV} \emph{numeric} Cortical Volume (aparc.stats) of RightIsthmusCingulate Range value: 677, ..., 4172
\item \strong{ST93SA} \emph{numeric} Surface Area (aparc.stats) of RightIsthmusCingulate Range value: 361, ..., 2372
\item \strong{ST93TA} \emph{numeric} Thickness Average (aparc.stats) of RightIsthmusCingulate Range value: 1.283, ..., 3.175
\item \strong{ST93TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightIsthmusCingulate Range value: 0.476, ..., 1.363
\item \strong{ST94CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOccipital Range value: 4990, ..., 17967
\item \strong{ST94SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOccipital Range value: 2065, ..., 7263
\item \strong{ST94TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOccipital Range value: 1.644, ..., 2.813
\item \strong{ST94TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOccipital Range value: 0.414, ..., 0.884
\item \strong{ST95CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOrbitofrontal Range value: 3203, ..., 9642
\item \strong{ST95SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOrbitofrontal Range value: 1236, ..., 3562
\item \strong{ST95TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOrbitofrontal Range value: 1.812, ..., 3.339
\item \strong{ST95TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOrbitofrontal Range value: 0.501, ..., 1.103
\item \strong{ST96SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightLateralVentricle Range value: 3423, ..., 75143.8
\item \strong{ST97CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLingual Range value: 3232, ..., 10141
\item \strong{ST97SA} \emph{numeric} Surface Area (aparc.stats) of RightLingual Range value: 1761, ..., 4941
\item \strong{ST97TA} \emph{numeric} Thickness Average (aparc.stats) of RightLingual Range value: 1.569, ..., 2.432
\item \strong{ST97TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLingual Range value: 0.417, ..., 0.864
\item \strong{ST98CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMedialOrbitofrontal Range value: 2178, ..., 6980
\item \strong{ST98SA} \emph{numeric} Surface Area (aparc.stats) of RightMedialOrbitofrontal Range value: 909, ..., 2533
\item \strong{ST98TA} \emph{numeric} Thickness Average (aparc.stats) of RightMedialOrbitofrontal Range value: 1.722, ..., 3.281
\item \strong{ST98TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMedialOrbitofrontal Range value: 0.414, ..., 1.309
\item \strong{ST99CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMiddleTemporal Range value: 3999, ..., 17471
\item \strong{ST99SA} \emph{numeric} Surface Area (aparc.stats) of RightMiddleTemporal Range value: 1442, ..., 4877
\item \strong{ST99TA} \emph{numeric} Thickness Average (aparc.stats) of RightMiddleTemporal Range value: 1.686, ..., 3.311
\item \strong{ST99TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMiddleTemporal Range value: 0.431, ..., 0.99
\item \strong{ST9SV} \emph{numeric} Subcortical Volume (aseg.stats) of FourthVentricle Range value: 808.2, ..., 5287.4
\item \strong{ST147SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalGM Range value: 109831.777006, ..., 289315.99601
\item \strong{ST148SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalGM Range value: 89603.654726, ..., 290674.859217
\item \strong{ST149SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalGM Range value: 199435.431732, ..., 579990.855227
\item \strong{ST150SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalWM Range value: 106254.861555, ..., 335063.439838
\item \strong{ST151SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalWM Range value: 88956.080096, ..., 331634.613082
\item \strong{ST152SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalWM Range value: 195210.941651, ..., 666111.511832
\item \strong{ST153SV} \emph{numeric} Subcortical Volume (aseg.stats) of SubcorticalGM Range value: 23596, ..., 69781
\item \strong{ST154SV} \emph{numeric} Subcortical Volume (aseg.stats) of TotalGM Range value: 276362.431732, ..., 766520.855227
\item \strong{ST155SV} \emph{numeric} Subcortical Volume (aseg.stats) of SupraTentorial Range value: 446750.373382, ..., 1298772.642055
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{imaging}
