% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFFSX7}
\alias{UCSFFSX7}
\title{Cross-Sectional FreeSurfer (7.x)}
\format{
A data frame with 12151 observations and 348 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFFSX7)
}
\description{
Cross-Sectional FreeSurfer (7.x) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNIGO, ADNI3, ADNI4, ADNI1
\item \strong{COLPROT} \emph{character}   , with options: ADNI3, ADNI2, ADNI4, ADNIGO, ADNI1
\item \strong{PTID} \emph{character} Participant Identifier for the in-clinic ADNI study  
\item \strong{RID} \emph{numeric} Participant Identifier for the in-clinic ADNI study Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} VISCODE of the Image  
\item \strong{VISCODE2} \emph{character}    
\item \strong{IMAGEUID} \emph{numeric} Image UID of image that was processed. Range value: 31863, ..., 11452777
\item \strong{FIELD_STRENGTH} \emph{character} the field strength (tesla) of the image , with options: 3T, 1.5T
\item \strong{EXAMDATE} \emph{ } When the image was collected. Date: YYY-MM-DD
\item \strong{RUNDATE} \emph{ } When the image was processed. Date: YYY-MM-DD
\item \strong{STATUS} \emph{character} partial = No QC, not final. Complete = QC and finalized. , with options: partial, complete
\item \strong{FSVER} \emph{character} Specific version of FreeSurfer used. All FS 7.x results are compatible. , with options: 7.2.0, 7.4.1
\item \strong{OVERALLQC} \emph{character} An overall quality rating. Refer to additional QC variables for Partial rating. , with options: Partial, Pass, Hippocampus Only, Fail
\item \strong{TEMPQC} \emph{character} QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) , with options: Fail, Pass
\item \strong{FRONTQC} \emph{character} QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) , with options: Pass, Fail
\item \strong{PARQC} \emph{character} QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) , with options: Pass, Fail
\item \strong{INSULAQC} \emph{character} QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) , with options: Pass, Fail
\item \strong{OCCQC} \emph{character} QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) , with options: Pass, Fail
\item \strong{BGQC} \emph{character} QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) , with options: Pass, Fail
\item \strong{CWMQC} \emph{character} QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87) , with options: Pass, Fail
\item \strong{VENTQC} \emph{character} QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) , with options: Pass, Fail
\item \strong{HIPPOQC} \emph{character} QC rating of the hippocampus. Fail affects the following regions: LeftFusiform (ST26); RightFusiform (ST85); LeftEntorhinal (ST24); RightEntorhinal (ST83); LeftParahippocampal (ST44); RightParahippocampal (ST103); LeftAmygdala (ST12); RightAmygdala (ST71) , with options: Pass, Fail
\item \strong{ST101SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPallidum Range value: 881.8, ..., 3097.5
\item \strong{ST102CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParacentral Range value: 1589, ..., 9285
\item \strong{ST102SA} \emph{numeric} Surface Area (aparc.stats) of RightParacentral Range value: 903, ..., 4241
\item \strong{ST102TA} \emph{numeric} Thickness Average (aparc.stats) of RightParacentral Range value: 1.263, ..., 3.162
\item \strong{ST102TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParacentral Range value: 0.361, ..., 1.006
\item \strong{ST103CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParahippocampal Range value: 491, ..., 4818
\item \strong{ST103SA} \emph{numeric} Surface Area (aparc.stats) of RightParahippocampal Range value: 262, ..., 3128
\item \strong{ST103TA} \emph{numeric} Thickness Average (aparc.stats) of RightParahippocampal Range value: 1.334, ..., 3.36
\item \strong{ST103TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParahippocampal Range value: 0.339, ..., 1.194
\item \strong{ST104CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOpercularis Range value: 944, ..., 7417
\item \strong{ST104SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOpercularis Range value: 755, ..., 2512
\item \strong{ST104TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOpercularis Range value: 1.265, ..., 2.938
\item \strong{ST104TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOpercularis Range value: 0.281, ..., 1.01
\item \strong{ST105CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsOrbitalis Range value: 878, ..., 4667
\item \strong{ST105SA} \emph{numeric} Surface Area (aparc.stats) of RightParsOrbitalis Range value: 320, ..., 1501
\item \strong{ST105TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsOrbitalis Range value: 1.698, ..., 3.419
\item \strong{ST105TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsOrbitalis Range value: 0.359, ..., 1.072
\item \strong{ST106CV} \emph{numeric} Cortical Volume (aparc.stats) of RightParsTriangularis Range value: 1100, ..., 8028
\item \strong{ST106SA} \emph{numeric} Surface Area (aparc.stats) of RightParsTriangularis Range value: 781, ..., 2615
\item \strong{ST106TA} \emph{numeric} Thickness Average (aparc.stats) of RightParsTriangularis Range value: 1.35, ..., 2.852
\item \strong{ST106TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightParsTriangularis Range value: 0.297, ..., 0.96
\item \strong{ST107CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPericalcarine Range value: 753, ..., 4579
\item \strong{ST107SA} \emph{numeric} Surface Area (aparc.stats) of RightPericalcarine Range value: 508, ..., 3017
\item \strong{ST107TA} \emph{numeric} Thickness Average (aparc.stats) of RightPericalcarine Range value: 1.116, ..., 2.195
\item \strong{ST107TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPericalcarine Range value: 0.169, ..., 0.945
\item \strong{ST108CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPostcentral Range value: 3990, ..., 16246
\item \strong{ST108SA} \emph{numeric} Surface Area (aparc.stats) of RightPostcentral Range value: 2054, ..., 8042
\item \strong{ST108TA} \emph{numeric} Thickness Average (aparc.stats) of RightPostcentral Range value: 1.356, ..., 2.595
\item \strong{ST108TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPostcentral Range value: 0.332, ..., 1
\item \strong{ST109CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPosteriorCingulate Range value: 241, ..., 5261
\item \strong{ST109SA} \emph{numeric} Surface Area (aparc.stats) of RightPosteriorCingulate Range value: 207, ..., 3192
\item \strong{ST109TA} \emph{numeric} Thickness Average (aparc.stats) of RightPosteriorCingulate Range value: 1.043, ..., 3.086
\item \strong{ST109TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPosteriorCingulate Range value: 0.297, ..., 1.111
\item \strong{ST10CV} \emph{numeric} Cortical Volume (aparc.stats) of Icv Range value: 427249.652055, ..., 5158451.282703
\item \strong{ST110CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecentral Range value: 5810, ..., 19409
\item \strong{ST110SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecentral Range value: 2884, ..., 7536
\item \strong{ST110TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecentral Range value: 1.411, ..., 2.918
\item \strong{ST110TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecentral Range value: 0.377, ..., 1.054
\item \strong{ST111CV} \emph{numeric} Cortical Volume (aparc.stats) of RightPrecuneus Range value: 4555, ..., 15103
\item \strong{ST111SA} \emph{numeric} Surface Area (aparc.stats) of RightPrecuneus Range value: 2257, ..., 6311
\item \strong{ST111TA} \emph{numeric} Thickness Average (aparc.stats) of RightPrecuneus Range value: 1.357, ..., 2.682
\item \strong{ST111TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightPrecuneus Range value: 0.365, ..., 0.919
\item \strong{ST112SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightPutamen Range value: 1759.3, ..., 8524.3
\item \strong{ST113CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralAnteriorCingulate Range value: 135, ..., 3591
\item \strong{ST113SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralAnteriorCingulate Range value: 60, ..., 1116
\item \strong{ST113TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralAnteriorCingulate Range value: 1.31, ..., 3.725
\item \strong{ST113TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralAnteriorCingulate Range value: 0.349, ..., 1.438
\item \strong{ST114CV} \emph{numeric} Cortical Volume (aparc.stats) of RightRostralMiddleFrontal Range value: 4051, ..., 27288
\item \strong{ST114SA} \emph{numeric} Surface Area (aparc.stats) of RightRostralMiddleFrontal Range value: 2189, ..., 9572
\item \strong{ST114TA} \emph{numeric} Thickness Average (aparc.stats) of RightRostralMiddleFrontal Range value: 1.584, ..., 2.975
\item \strong{ST114TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightRostralMiddleFrontal Range value: 0.362, ..., 0.966
\item \strong{ST115CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorFrontal Range value: 7714, ..., 35844
\item \strong{ST115SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorFrontal Range value: 3520, ..., 13580
\item \strong{ST115TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorFrontal Range value: 1.492, ..., 3.019
\item \strong{ST115TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorFrontal Range value: 0.388, ..., 1.03
\item \strong{ST116CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorParietal Range value: 5730, ..., 20409
\item \strong{ST116SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorParietal Range value: 2752, ..., 7597
\item \strong{ST116TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorParietal Range value: 1.339, ..., 2.84
\item \strong{ST116TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorParietal Range value: 0.255, ..., 1.192
\item \strong{ST117CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSuperiorTemporal Range value: 4315, ..., 15796
\item \strong{ST117SA} \emph{numeric} Surface Area (aparc.stats) of RightSuperiorTemporal Range value: 2131, ..., 5518
\item \strong{ST117TA} \emph{numeric} Thickness Average (aparc.stats) of RightSuperiorTemporal Range value: 1.449, ..., 3.173
\item \strong{ST117TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSuperiorTemporal Range value: 0.309, ..., 1.012
\item \strong{ST118CV} \emph{numeric} Cortical Volume (aparc.stats) of RightSupramarginal Range value: 4192, ..., 14536
\item \strong{ST118SA} \emph{numeric} Surface Area (aparc.stats) of RightSupramarginal Range value: 1967, ..., 5867
\item \strong{ST118TA} \emph{numeric} Thickness Average (aparc.stats) of RightSupramarginal Range value: 1.501, ..., 2.83
\item \strong{ST118TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightSupramarginal Range value: 0.341, ..., 1.149
\item \strong{ST119CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTemporalPole Range value: 192, ..., 4791
\item \strong{ST119SA} \emph{numeric} Surface Area (aparc.stats) of RightTemporalPole Range value: 180, ..., 862
\item \strong{ST119TA} \emph{numeric} Thickness Average (aparc.stats) of RightTemporalPole Range value: 1.107, ..., 4.499
\item \strong{ST119TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTemporalPole Range value: 0.224, ..., 1.317
\item \strong{ST11SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAccumbensArea Range value: 47.6, ..., 843.2
\item \strong{ST120SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightThalamus Range value: 2546, ..., 11218
\item \strong{ST121CV} \emph{numeric} Cortical Volume (aparc.stats) of RightTransverseTemporal Range value: 197, ..., 1578
\item \strong{ST121SA} \emph{numeric} Surface Area (aparc.stats) of RightTransverseTemporal Range value: 138, ..., 613
\item \strong{ST121TA} \emph{numeric} Thickness Average (aparc.stats) of RightTransverseTemporal Range value: 1.179, ..., 3.046
\item \strong{ST121TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightTransverseTemporal Range value: 0.194, ..., 0.98
\item \strong{ST124SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVentralDC Range value: 1845.3, ..., 5456.7
\item \strong{ST125SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightVessel Range value: 1, ..., 434.9
\item \strong{ST127SV} \emph{numeric} Subcortical Volume (aseg.stats) of ThirdVentricle Range value: 515.1, ..., 4938.1
\item \strong{ST128SV} \emph{numeric} Subcortical Volume (aseg.stats) of WMHypoIntensities Range value: 222.9, ..., 74594.5
\item \strong{ST129CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInsula Range value: 2302, ..., 9955
\item \strong{ST129SA} \emph{numeric} Surface Area (aparc.stats) of LeftInsula Range value: 1296, ..., 4289
\item \strong{ST129TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInsula Range value: 1.452, ..., 3.491
\item \strong{ST129TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInsula Range value: 0.484, ..., 1.222
\item \strong{ST12SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftAmygdala Range value: 289.1, ..., 2455.5
\item \strong{ST130CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInsula Range value: 2567, ..., 9732
\item \strong{ST130SA} \emph{numeric} Surface Area (aparc.stats) of RightInsula Range value: 1134, ..., 4409
\item \strong{ST130TA} \emph{numeric} Thickness Average (aparc.stats) of RightInsula Range value: 1.506, ..., 3.433
\item \strong{ST130TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInsula Range value: 0.532, ..., 1.184
\item \strong{ST13CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftBankssts Range value: 726, ..., 5078
\item \strong{ST13SA} \emph{numeric} Surface Area (aparc.stats) of LeftBankssts Range value: 399, ..., 1896
\item \strong{ST13TA} \emph{numeric} Thickness Average (aparc.stats) of LeftBankssts Range value: 1.132, ..., 2.913
\item \strong{ST13TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftBankssts Range value: 0.248, ..., 0.847
\item \strong{ST14CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 160, ..., 4296
\item \strong{ST14SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 120, ..., 2626
\item \strong{ST14TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 1.214, ..., 3.517
\item \strong{ST14TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalAnteriorCingulate Range value: 0.259, ..., 1.403
\item \strong{ST15CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1830, ..., 10257
\item \strong{ST15SA} \emph{numeric} Surface Area (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1066, ..., 3929
\item \strong{ST15TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCaudalMiddleFrontal Range value: 1.408, ..., 2.984
\item \strong{ST15TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCaudalMiddleFrontal Range value: 0.23, ..., 0.822
\item \strong{ST16SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCaudate Range value: 863.6, ..., 7711.2
\item \strong{ST17SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumCortex Range value: 27119.6, ..., 71107.7
\item \strong{ST18SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCerebellumWM Range value: 7715.7, ..., 30930.1
\item \strong{ST1SV} \emph{numeric} Subcortical Volume (aseg.stats) of Brainstem Range value: 12977.4, ..., 35152
\item \strong{ST21SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftChoroidPlexus Range value: 107.6, ..., 2091.4
\item \strong{ST23CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftCuneus Range value: 1130, ..., 5318
\item \strong{ST23SA} \emph{numeric} Surface Area (aparc.stats) of LeftCuneus Range value: 405, ..., 2513
\item \strong{ST23TA} \emph{numeric} Thickness Average (aparc.stats) of LeftCuneus Range value: 1.284, ..., 2.484
\item \strong{ST23TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftCuneus Range value: 0.183, ..., 0.97
\item \strong{ST24CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftEntorhinal Range value: 2, ..., 4404
\item \strong{ST24SA} \emph{numeric} Surface Area (aparc.stats) of LeftEntorhinal Range value: 3, ..., 1212
\item \strong{ST24TA} \emph{numeric} Thickness Average (aparc.stats) of LeftEntorhinal Range value: 1.068, ..., 4.194
\item \strong{ST24TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftEntorhinal Range value: 0.168, ..., 1.168
\item \strong{ST25CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFrontalPole Range value: 156, ..., 2727
\item \strong{ST25SA} \emph{numeric} Surface Area (aparc.stats) of LeftFrontalPole Range value: 72, ..., 553
\item \strong{ST25TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFrontalPole Range value: 1.324, ..., 3.905
\item \strong{ST25TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFrontalPole Range value: 0.212, ..., 1.287
\item \strong{ST26CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftFusiform Range value: 2532, ..., 14289
\item \strong{ST26SA} \emph{numeric} Surface Area (aparc.stats) of LeftFusiform Range value: 1073, ..., 4512
\item \strong{ST26TA} \emph{numeric} Thickness Average (aparc.stats) of LeftFusiform Range value: 1.421, ..., 3.139
\item \strong{ST26TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftFusiform Range value: 0.423, ..., 0.988
\item \strong{ST28SA} \emph{numeric} Surface Area (aparc.stats) of LeftHemisphereWM Range value: 48668.9, ..., 110776
\item \strong{ST29SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftHippocampus Range value: 1188.7, ..., 5684.6
\item \strong{ST2SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumAnterior Range value: 32.9, ..., 1872.9
\item \strong{ST30SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftInferiorLateralVentricle Range value: 54.5, ..., 9047.8
\item \strong{ST31CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorParietal Range value: 4278, ..., 17194
\item \strong{ST31SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorParietal Range value: 2088, ..., 6500
\item \strong{ST31TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorParietal Range value: 1.345, ..., 2.732
\item \strong{ST31TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorParietal Range value: 0.314, ..., 0.979
\item \strong{ST32CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftInferiorTemporal Range value: 3405, ..., 17181
\item \strong{ST32SA} \emph{numeric} Surface Area (aparc.stats) of LeftInferiorTemporal Range value: 1175, ..., 5398
\item \strong{ST32TA} \emph{numeric} Thickness Average (aparc.stats) of LeftInferiorTemporal Range value: 1.542, ..., 3.197
\item \strong{ST32TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftInferiorTemporal Range value: 0.445, ..., 1.117
\item \strong{ST34CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftIsthmusCingulate Range value: 61, ..., 6312
\item \strong{ST34SA} \emph{numeric} Surface Area (aparc.stats) of LeftIsthmusCingulate Range value: 39, ..., 4276
\item \strong{ST34TA} \emph{numeric} Thickness Average (aparc.stats) of LeftIsthmusCingulate Range value: 1.045, ..., 2.88
\item \strong{ST34TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftIsthmusCingulate Range value: 0.426, ..., 1.211
\item \strong{ST35CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOccipital Range value: 3323, ..., 18127
\item \strong{ST35SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOccipital Range value: 2920, ..., 7415
\item \strong{ST35TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOccipital Range value: 1.243, ..., 2.62
\item \strong{ST35TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOccipital Range value: 0.296, ..., 1.065
\item \strong{ST36CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLateralOrbitofrontal Range value: 2600, ..., 10531
\item \strong{ST36SA} \emph{numeric} Surface Area (aparc.stats) of LeftLateralOrbitofrontal Range value: 1129, ..., 3835
\item \strong{ST36TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLateralOrbitofrontal Range value: 1.545, ..., 3.188
\item \strong{ST36TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLateralOrbitofrontal Range value: 0.421, ..., 0.977
\item \strong{ST37SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftLateralVentricle Range value: 3369.2, ..., 94302.7
\item \strong{ST38CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftLingual Range value: 1107, ..., 10465
\item \strong{ST38SA} \emph{numeric} Surface Area (aparc.stats) of LeftLingual Range value: 613, ..., 5735
\item \strong{ST38TA} \emph{numeric} Thickness Average (aparc.stats) of LeftLingual Range value: 1.088, ..., 2.494
\item \strong{ST38TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftLingual Range value: 0.315, ..., 0.89
\item \strong{ST39CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMedialOrbitofrontal Range value: 2147, ..., 8093
\item \strong{ST39SA} \emph{numeric} Surface Area (aparc.stats) of LeftMedialOrbitofrontal Range value: 1080, ..., 2918
\item \strong{ST39TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMedialOrbitofrontal Range value: 1.337, ..., 3.017
\item \strong{ST39TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMedialOrbitofrontal Range value: 0.396, ..., 1.181
\item \strong{ST3SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumCentral Range value: 123.4, ..., 1836
\item \strong{ST40CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftMiddleTemporal Range value: 3093, ..., 15598
\item \strong{ST40SA} \emph{numeric} Surface Area (aparc.stats) of LeftMiddleTemporal Range value: 1244, ..., 4722
\item \strong{ST40TA} \emph{numeric} Thickness Average (aparc.stats) of LeftMiddleTemporal Range value: 1.526, ..., 3.147
\item \strong{ST40TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftMiddleTemporal Range value: 0.406, ..., 0.991
\item \strong{ST42SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPallidum Range value: 775.3, ..., 3525.5
\item \strong{ST43CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParacentral Range value: 1633, ..., 7498
\item \strong{ST43SA} \emph{numeric} Surface Area (aparc.stats) of LeftParacentral Range value: 823, ..., 4298
\item \strong{ST43TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParacentral Range value: 1.359, ..., 2.877
\item \strong{ST43TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParacentral Range value: 0.367, ..., 0.982
\item \strong{ST44CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParahippocampal Range value: 21, ..., 7224
\item \strong{ST44SA} \emph{numeric} Surface Area (aparc.stats) of LeftParahippocampal Range value: 17, ..., 4381
\item \strong{ST44TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParahippocampal Range value: 0.96, ..., 3.564
\item \strong{ST44TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParahippocampal Range value: 0.175, ..., 1.192
\item \strong{ST45CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOpercularis Range value: 1377, ..., 8539
\item \strong{ST45SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOpercularis Range value: 856, ..., 3410
\item \strong{ST45TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOpercularis Range value: 1.461, ..., 2.904
\item \strong{ST45TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOpercularis Range value: 0.291, ..., 0.985
\item \strong{ST46CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsOrbitalis Range value: 414, ..., 3810
\item \strong{ST46SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsOrbitalis Range value: 161, ..., 1122
\item \strong{ST46TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsOrbitalis Range value: 1.613, ..., 3.555
\item \strong{ST46TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsOrbitalis Range value: 0.335, ..., 1.14
\item \strong{ST47CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftParsTriangularis Range value: 800, ..., 5781
\item \strong{ST47SA} \emph{numeric} Surface Area (aparc.stats) of LeftParsTriangularis Range value: 402, ..., 2105
\item \strong{ST47TA} \emph{numeric} Thickness Average (aparc.stats) of LeftParsTriangularis Range value: 1.505, ..., 2.923
\item \strong{ST47TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftParsTriangularis Range value: 0.282, ..., 1.051
\item \strong{ST48CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPericalcarine Range value: 595, ..., 4778
\item \strong{ST48SA} \emph{numeric} Surface Area (aparc.stats) of LeftPericalcarine Range value: 355, ..., 4098
\item \strong{ST48TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPericalcarine Range value: 1.123, ..., 2.422
\item \strong{ST48TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPericalcarine Range value: 0.164, ..., 1.061
\item \strong{ST49CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPostcentral Range value: 4205, ..., 16949
\item \strong{ST49SA} \emph{numeric} Surface Area (aparc.stats) of LeftPostcentral Range value: 2145, ..., 10232
\item \strong{ST49TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPostcentral Range value: 1.429, ..., 2.819
\item \strong{ST49TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPostcentral Range value: 0.362, ..., 1.07
\item \strong{ST4SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidAnterior Range value: 77.5, ..., 1251.2
\item \strong{ST50CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPosteriorCingulate Range value: 725, ..., 11302
\item \strong{ST50SA} \emph{numeric} Surface Area (aparc.stats) of LeftPosteriorCingulate Range value: 333, ..., 6909
\item \strong{ST50TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPosteriorCingulate Range value: 0.594, ..., 2.8
\item \strong{ST50TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPosteriorCingulate Range value: 0.352, ..., 1.033
\item \strong{ST51CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecentral Range value: 4510, ..., 19063
\item \strong{ST51SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecentral Range value: 2823, ..., 8718
\item \strong{ST51TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecentral Range value: 1.497, ..., 2.855
\item \strong{ST51TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecentral Range value: 0.401, ..., 0.866
\item \strong{ST52CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftPrecuneus Range value: 3677, ..., 13394
\item \strong{ST52SA} \emph{numeric} Surface Area (aparc.stats) of LeftPrecuneus Range value: 1662, ..., 7490
\item \strong{ST52TA} \emph{numeric} Thickness Average (aparc.stats) of LeftPrecuneus Range value: 1.442, ..., 2.71
\item \strong{ST52TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftPrecuneus Range value: 0.349, ..., 1.168
\item \strong{ST53SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftPutamen Range value: 1532, ..., 8618.8
\item \strong{ST54CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralAnteriorCingulate Range value: 177, ..., 4563
\item \strong{ST54SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralAnteriorCingulate Range value: 154, ..., 1597
\item \strong{ST54TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralAnteriorCingulate Range value: 1.347, ..., 3.803
\item \strong{ST54TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralAnteriorCingulate Range value: 0.379, ..., 1.622
\item \strong{ST55CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftRostralMiddleFrontal Range value: 4340, ..., 23950
\item \strong{ST55SA} \emph{numeric} Surface Area (aparc.stats) of LeftRostralMiddleFrontal Range value: 2019, ..., 8778
\item \strong{ST55TA} \emph{numeric} Thickness Average (aparc.stats) of LeftRostralMiddleFrontal Range value: 1.459, ..., 2.756
\item \strong{ST55TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftRostralMiddleFrontal Range value: 0.342, ..., 0.941
\item \strong{ST56CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorFrontal Range value: 8943, ..., 30991
\item \strong{ST56SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorFrontal Range value: 3982, ..., 13626
\item \strong{ST56TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorFrontal Range value: 1.442, ..., 2.981
\item \strong{ST56TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorFrontal Range value: 0.38, ..., 1.004
\item \strong{ST57CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorParietal Range value: 5459, ..., 19053
\item \strong{ST57SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorParietal Range value: 2931, ..., 7613
\item \strong{ST57TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorParietal Range value: 1.285, ..., 2.673
\item \strong{ST57TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorParietal Range value: 0.227, ..., 0.941
\item \strong{ST58CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSuperiorTemporal Range value: 4465, ..., 18228
\item \strong{ST58SA} \emph{numeric} Surface Area (aparc.stats) of LeftSuperiorTemporal Range value: 1896, ..., 5691
\item \strong{ST58TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSuperiorTemporal Range value: 1.544, ..., 3.107
\item \strong{ST58TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSuperiorTemporal Range value: 0.429, ..., 0.901
\item \strong{ST59CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftSupramarginal Range value: 3983, ..., 19102
\item \strong{ST59SA} \emph{numeric} Surface Area (aparc.stats) of LeftSupramarginal Range value: 2172, ..., 6822
\item \strong{ST59TA} \emph{numeric} Thickness Average (aparc.stats) of LeftSupramarginal Range value: 1.34, ..., 2.828
\item \strong{ST59TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftSupramarginal Range value: 0.331, ..., 1.046
\item \strong{ST5SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumMidPosterior Range value: 50.9, ..., 1863.5
\item \strong{ST60CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTemporalPole Range value: 225, ..., 4879
\item \strong{ST60SA} \emph{numeric} Surface Area (aparc.stats) of LeftTemporalPole Range value: 80, ..., 827
\item \strong{ST60TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTemporalPole Range value: 1.374, ..., 4.366
\item \strong{ST60TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTemporalPole Range value: 0.289, ..., 1.323
\item \strong{ST61SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftThalamus Range value: 3028, ..., 11089.4
\item \strong{ST62CV} \emph{numeric} Cortical Volume (aparc.stats) of LeftTransverseTemporal Range value: 432, ..., 2231
\item \strong{ST62SA} \emph{numeric} Surface Area (aparc.stats) of LeftTransverseTemporal Range value: 244, ..., 781
\item \strong{ST62TA} \emph{numeric} Thickness Average (aparc.stats) of LeftTransverseTemporal Range value: 1.26, ..., 3.003
\item \strong{ST62TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of LeftTransverseTemporal Range value: 0.167, ..., 0.789
\item \strong{ST65SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVentralDC Range value: 1530.2, ..., 5666.2
\item \strong{ST66SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftVessel Range value: 0.8, ..., 578
\item \strong{ST68SV} \emph{numeric} Subcortical Volume (aseg.stats) of NonWMHypoIntensities Range value: 0.9, ..., 269.6
\item \strong{ST69SV} \emph{numeric} Subcortical Volume (aseg.stats) of OpticChiasm Range value: 15.3, ..., 437.2
\item \strong{ST6SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorpusCallosumPosterior Range value: 54.3, ..., 2127.3
\item \strong{ST70SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAccumbensArea Range value: 21.6, ..., 865.5
\item \strong{ST71SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightAmygdala Range value: 328.7, ..., 2593.6
\item \strong{ST72CV} \emph{numeric} Cortical Volume (aparc.stats) of RightBankssts Range value: 584, ..., 3535
\item \strong{ST72SA} \emph{numeric} Surface Area (aparc.stats) of RightBankssts Range value: 409, ..., 1496
\item \strong{ST72TA} \emph{numeric} Thickness Average (aparc.stats) of RightBankssts Range value: 1.025, ..., 3.019
\item \strong{ST72TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightBankssts Range value: 0.234, ..., 0.937
\item \strong{ST73CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalAnteriorCingulate Range value: 117, ..., 4779
\item \strong{ST73SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalAnteriorCingulate Range value: 119, ..., 2730
\item \strong{ST73TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalAnteriorCingulate Range value: 1.161, ..., 3.639
\item \strong{ST73TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalAnteriorCingulate Range value: 0.269, ..., 1.362
\item \strong{ST74CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCaudalMiddleFrontal Range value: 1463, ..., 9527
\item \strong{ST74SA} \emph{numeric} Surface Area (aparc.stats) of RightCaudalMiddleFrontal Range value: 1022, ..., 3684
\item \strong{ST74TA} \emph{numeric} Thickness Average (aparc.stats) of RightCaudalMiddleFrontal Range value: 1.469, ..., 2.939
\item \strong{ST74TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCaudalMiddleFrontal Range value: 0.287, ..., 0.867
\item \strong{ST75SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCaudate Range value: 732, ..., 8166.3
\item \strong{ST76SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumCortex Range value: 31755.3, ..., 73352.3
\item \strong{ST77SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCerebellumWM Range value: 7088.2, ..., 27391.7
\item \strong{ST7SV} \emph{numeric} Subcortical Volume (aseg.stats) of Csf Range value: 267.4, ..., 6037.4
\item \strong{ST80SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightChoroidPlexus Range value: 99.3, ..., 2491.6
\item \strong{ST82CV} \emph{numeric} Cortical Volume (aparc.stats) of RightCuneus Range value: 1327, ..., 6143
\item \strong{ST82SA} \emph{numeric} Surface Area (aparc.stats) of RightCuneus Range value: 626, ..., 2588
\item \strong{ST82TA} \emph{numeric} Thickness Average (aparc.stats) of RightCuneus Range value: 1.3, ..., 2.54
\item \strong{ST82TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightCuneus Range value: 0.244, ..., 0.892
\item \strong{ST83CV} \emph{numeric} Cortical Volume (aparc.stats) of RightEntorhinal Range value: 308, ..., 4322
\item \strong{ST83SA} \emph{numeric} Surface Area (aparc.stats) of RightEntorhinal Range value: 177, ..., 1031
\item \strong{ST83TA} \emph{numeric} Thickness Average (aparc.stats) of RightEntorhinal Range value: 1.316, ..., 4.312
\item \strong{ST83TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightEntorhinal Range value: 0.409, ..., 1.211
\item \strong{ST84CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFrontalPole Range value: 320, ..., 2355
\item \strong{ST84SA} \emph{numeric} Surface Area (aparc.stats) of RightFrontalPole Range value: 122, ..., 545
\item \strong{ST84TA} \emph{numeric} Thickness Average (aparc.stats) of RightFrontalPole Range value: 1.433, ..., 3.996
\item \strong{ST84TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFrontalPole Range value: 0.287, ..., 1.219
\item \strong{ST85CV} \emph{numeric} Cortical Volume (aparc.stats) of RightFusiform Range value: 3498, ..., 14350
\item \strong{ST85SA} \emph{numeric} Surface Area (aparc.stats) of RightFusiform Range value: 1539, ..., 4285
\item \strong{ST85TA} \emph{numeric} Thickness Average (aparc.stats) of RightFusiform Range value: 1.409, ..., 3.086
\item \strong{ST85TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightFusiform Range value: 0.425, ..., 1.021
\item \strong{ST87sa} \emph{numeric} Surface Area (aparc.stats) of RightHemisphereWM Range value: 55054.4, ..., 109909
\item \strong{ST88SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightHippocampus Range value: 1047.6, ..., 5567.9
\item \strong{ST89SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightInferiorLateralVentricle Range value: 34.4, ..., 12740.1
\item \strong{ST8SV} \emph{numeric} Subcortical Volume (aseg.stats) of FifthVentricle Range value: 1, ..., 25.1
\item \strong{ST90CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorParietal Range value: 4606, ..., 21947
\item \strong{ST90SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorParietal Range value: 2609, ..., 8456
\item \strong{ST90TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorParietal Range value: 1.451, ..., 2.702
\item \strong{ST90TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorParietal Range value: 0.31, ..., 1.058
\item \strong{ST91CV} \emph{numeric} Cortical Volume (aparc.stats) of RightInferiorTemporal Range value: 3045, ..., 15795
\item \strong{ST91SA} \emph{numeric} Surface Area (aparc.stats) of RightInferiorTemporal Range value: 1447, ..., 4958
\item \strong{ST91TA} \emph{numeric} Thickness Average (aparc.stats) of RightInferiorTemporal Range value: 1.669, ..., 3.203
\item \strong{ST91TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightInferiorTemporal Range value: 0.459, ..., 1.009
\item \strong{ST93CV} \emph{numeric} Cortical Volume (aparc.stats) of RightIsthmusCingulate Range value: 358, ..., 4857
\item \strong{ST93SA} \emph{numeric} Surface Area (aparc.stats) of RightIsthmusCingulate Range value: 380, ..., 3816
\item \strong{ST93TA} \emph{numeric} Thickness Average (aparc.stats) of RightIsthmusCingulate Range value: 1.049, ..., 2.847
\item \strong{ST93TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightIsthmusCingulate Range value: 0.365, ..., 1.138
\item \strong{ST94CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOccipital Range value: 5232, ..., 19294
\item \strong{ST94SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOccipital Range value: 2791, ..., 7788
\item \strong{ST94TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOccipital Range value: 1.381, ..., 2.624
\item \strong{ST94TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOccipital Range value: 0.333, ..., 0.863
\item \strong{ST95CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLateralOrbitofrontal Range value: 2332, ..., 9996
\item \strong{ST95SA} \emph{numeric} Surface Area (aparc.stats) of RightLateralOrbitofrontal Range value: 1160, ..., 4069
\item \strong{ST95TA} \emph{numeric} Thickness Average (aparc.stats) of RightLateralOrbitofrontal Range value: 1.572, ..., 3.15
\item \strong{ST95TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLateralOrbitofrontal Range value: 0.455, ..., 0.991
\item \strong{ST96SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightLateralVentricle Range value: 1834.5, ..., 77644.3
\item \strong{ST97CV} \emph{numeric} Cortical Volume (aparc.stats) of RightLingual Range value: 2820, ..., 10445
\item \strong{ST97SA} \emph{numeric} Surface Area (aparc.stats) of RightLingual Range value: 1766, ..., 4918
\item \strong{ST97TA} \emph{numeric} Thickness Average (aparc.stats) of RightLingual Range value: 1.211, ..., 2.455
\item \strong{ST97TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightLingual Range value: 0.366, ..., 0.789
\item \strong{ST98CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMedialOrbitofrontal Range value: 560, ..., 8237
\item \strong{ST98SA} \emph{numeric} Surface Area (aparc.stats) of RightMedialOrbitofrontal Range value: 233, ..., 3072
\item \strong{ST98TA} \emph{numeric} Thickness Average (aparc.stats) of RightMedialOrbitofrontal Range value: 1.399, ..., 3.259
\item \strong{ST98TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMedialOrbitofrontal Range value: 0.412, ..., 1.225
\item \strong{ST99CV} \emph{numeric} Cortical Volume (aparc.stats) of RightMiddleTemporal Range value: 3468, ..., 17074
\item \strong{ST99SA} \emph{numeric} Surface Area (aparc.stats) of RightMiddleTemporal Range value: 1674, ..., 5222
\item \strong{ST99TA} \emph{numeric} Thickness Average (aparc.stats) of RightMiddleTemporal Range value: 1.479, ..., 3.168
\item \strong{ST99TS} \emph{numeric} Thickness Stardard Deviation (aparc.stats) of RightMiddleTemporal Range value: 0.35, ..., 0.958
\item \strong{ST9SV} \emph{numeric} Subcortical Volume (aseg.stats) of FourthVentricle Range value: 638.3, ..., 6263.4
\item \strong{ST147SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalGM Range value: 89121.5507, ..., 295642.067545
\item \strong{ST148SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalGM Range value: 91190.472591, ..., 298304.664373
\item \strong{ST149SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalGM Range value: 180312.023292, ..., 593946.731918
\item \strong{ST150SV} \emph{numeric} Subcortical Volume (aseg.stats) of LeftCorticalWM Range value: 27674.5, ..., 367026.5
\item \strong{ST151SV} \emph{numeric} Subcortical Volume (aseg.stats) of RightCorticalWM Range value: 127182.5, ..., 377982.5
\item \strong{ST152SV} \emph{numeric} Subcortical Volume (aseg.stats) of CorticalWM Range value: 184555, ..., 745009
\item \strong{ST153SV} \emph{numeric} Subcortical Volume (aseg.stats) of SubcorticalGM Range value: 28737, ..., 72446
\item \strong{ST154SV} \emph{numeric} Subcortical Volume (aseg.stats) of TotalGM Range value: 307855.023292, ..., 782888.731918
\item \strong{ST155SV} \emph{numeric} Subcortical Volume (aseg.stats) of SupraTentorial Range value: 537204, ..., 1337155
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{imaging}
