% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSFSNTVOL}
\alias{UCSFSNTVOL}
\title{UCSF SNT Hippocampal Volumes}
\format{
A data frame with 1874 observations and 16 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UCSFSNTVOL)
}
\description{
UCSF SNT Hippocampal Volumes data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric}   Range value: 3, ..., 1402
\item \strong{SITEID} \emph{numeric}   Range value: 1, ..., 127
\item \strong{VISCODE} \emph{character}   , with options: bl, m06, sc, m12
\item \strong{USERDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{RECNO} \emph{numeric}   Range value: 1, ..., 3
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{LONISID} \emph{numeric} LONI Study ID Range value: 1948, ..., 15737
\item \strong{LONIUID} \emph{numeric} LONI Unique Series ID Range value: 8723, ..., 49622
\item \strong{IMAGEUID} \emph{numeric} imageUID Range value: 30968, ..., 106232
\item \strong{FLDSTRENG} \emph{numeric} Field Strength Range value: 1.5, ..., 3
\item \strong{SNTDATE} \emph{ } Date Volume Generated Date: YYY-MM-DD
\item \strong{ATLAS} \emph{character} Name of Atlas Used , with options: HippoUCSF
\item \strong{LEFTHIPPO} \emph{numeric} Left Hippocampus Volume Range value: 816.761, ..., 2875.28
\item \strong{RIGHTHIPPO} \emph{numeric} Right Hippocampus Volume Range value: 592.838, ..., 3077.68
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
