% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UGOTPTAU181}
\alias{UGOTPTAU181}
\title{University of Gothenburg Longitudinal plasma P-tau181: 2020-06-18}
\format{
A data frame with 3758 observations and 9 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UGOTPTAU181)
}
\description{
University of Gothenburg Longitudinal plasma P-tau181: 2020-06-18 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VID} \emph{numeric} UPENN Visit Code Range value: 2, ..., 14
\item \strong{PLASMAPTAU181} \emph{numeric} Plasma Tau 181 Measurements Range value: 0.362, ..., 451.398
\item \strong{COMMENT} \emph{character} None , with options: <LLOQ
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
