% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENN_PLASMA_FUJIREBIO_QUANTERIX}
\alias{UPENN_PLASMA_FUJIREBIO_QUANTERIX}
\title{UPENN - Plasma Biomarkers (Aβ42, Aβ40, ptau217, NfL, GFAP) measured by Fujirebio & Quanterix ADNI1,GO,2,3,4}
\format{
A data frame with 1728 observations and 17 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENN_PLASMA_FUJIREBIO_QUANTERIX)
}
\description{
UPENN - Plasma Biomarkers (Aβ42, Aβ40, ptau217, NfL, GFAP) measured by Fujirebio & Quanterix [ADNI1,GO,2,3,4] data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{character}   , with options: ADNI1, ADNI2, ADNI4, ADNI3, ADNIGO
\item \strong{PTID} \emph{character} Patient ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10470
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{Primary} \emph{character} Sample Primary Type BLD=Blood
\item \strong{Additive} \emph{character} Sample Additive Type EDT=Ethylenediaminetetraacetic acid
\item \strong{pT217_F} \emph{numeric} pTau217 measured by Fujirebio -4=Insufficient sample
\item \strong{AB42_F} \emph{numeric} ABeta42 measured by Fujirebio -4=Insufficient sample
\item \strong{AB40_F} \emph{numeric} ABeta40 measured by Fujirebio -4=Insufficient sample
\item \strong{AB42_AB40_F} \emph{numeric} ABeta42/40 ratio measured by Fujirebio -4=Insufficient sample
\item \strong{pT217_AB42_F} \emph{numeric} pTau217/ABeta42 ratio measured by fujirebio -4=Insufficient sample
\item \strong{NfL_Q} \emph{numeric} Neurofilament light chain measured by Quanterix -4=Insufficient sample ; -5=Sample quantity not sufficient
\item \strong{GFAP_Q} \emph{numeric} Glial fibrillary acidic protein measured by Quanterix -4=Insufficient sample ; -5=Sample quantity not sufficient
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{biospecimen}
