% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENNBIOMK_MASTER}
\alias{UPENNBIOMK_MASTER}
\title{UPENN CSF Biomarker Master Alzbio3 (BATCH1-8)}
\format{
A data frame with 5876 observations and 15 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENNBIOMK_MASTER)
}
\description{
UPENN CSF Biomarker Master Alzbio3 (BATCH1-8) data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 3, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code , with options: bl, m12, m24, m48, m60, m84, m36, m72
\item \strong{BATCH} \emph{character} Name of LONI table, corresponding to analytical batch of results , with options: UPENNBIOMK, UPENNBIOMK2, MEDIAN, UPENNBIOMK3, UPENNBIOMK4, UPENNBIOMK6, UPENNBIOMK5, UPENNBIOMK7, UPENNBIOMK8
\item \strong{KIT} \emph{character} Reagents lot number  
\item \strong{STDS} \emph{character} Calibrators and Quality Controls lot number  
\item \strong{DRAWDTE} \emph{ } Date of CSF draw Date: YYY-MM-DD
\item \strong{RUNDATE} \emph{ } Date of analytical run Date: YYY-MM-DD
\item \strong{ABETA} \emph{numeric} Result rescaled to UPENNBIOMK Range value: 40.5, ..., 374
\item \strong{TAU} \emph{numeric} Result rescaled to UPENNBIOMK Range value: 13, ..., 495
\item \strong{PTAU} \emph{numeric} Result rescaled to UPENNBIOMK Range value: 6.9, ..., 213
\item \strong{ABETA_RAW} \emph{numeric} Original (raw) result, before rescaling Range value: 26.4, ..., 548
\item \strong{TAU_RAW} \emph{numeric} Original (raw) result, before rescaling Range value: 12.3, ..., 495
\item \strong{PTAU_RAW} \emph{numeric} Original (raw) result, before rescaling Range value: 5.22, ..., 115
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
