% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENNBIOMK_ROCHE_ELECSYS}
\alias{UPENNBIOMK_ROCHE_ELECSYS}
\title{UPENN CSF Biomarkers Elecsys ADNI1,GO,2,3}
\format{
A data frame with 3174 observations and 14 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENNBIOMK_ROCHE_ELECSYS)
}
\description{
UPENN CSF Biomarkers Elecsys [ADNI1,GO,2,3] data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 3, ..., 7073
\item \strong{VISCODE2} \emph{character} Months from baseline rounded to nearest 6 months  
\item \strong{EXAMDATE} \emph{ } Date sample collected Date: YYY-MM-DD
\item \strong{BATCH} \emph{character} Sample batch UPENNBIOMK9 = ADNI1/GO/2 batch ; UPENNBIOMK11 = ADNI3 1st batch ; UPENNBIOMK12 = ADNI3 2nd batch ; UPENNBIOMK13 = ADNI3 3rd & 4th batches
\emph{Decoded Value: }\strong{Yes}
\item \strong{RUNDATE} \emph{ } Date of experiment Date: YYY-MM-DD
\item \strong{ABETA40} \emph{numeric} ABETA40 result Range value: 841, ..., 37480
\item \strong{ABETA42} \emph{numeric} ABETA42 result Range value: 203, ..., 4779
\item \strong{TAU} \emph{numeric} TAU result Range value: 80.08, ..., 1018
\item \strong{PTAU} \emph{numeric} PTAU result Range value: 8, ..., 108.5
\item \strong{COMMENT} \emph{character} Supplementary notes for missing values and outlier , with options: Abeta42>1700, Tau>1300, PTau>120, Tau<80, PTau<8, Abeta42<200, PTau<8, Ptau<8, PTau>120, Sample Hemolyzed, Abeta42>1700, recalculation failed
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{biospecimen}
