% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENNBIOMKADNIDIAN2017}
\alias{UPENNBIOMKADNIDIAN2017}
\title{UPENNBIOMKADNIDIAN2017}
\format{
A data frame with 422 observations and 14 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENNBIOMKADNIDIAN2017)
}
\description{
UPENNBIOMKADNIDIAN2017 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric}   Range value: 42, ..., 5285
\item \strong{VISCODE} \emph{character}    
\item \strong{VISCODE2} \emph{character}    
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STUDY} \emph{character}   , with options: ADNI
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{ABETA} \emph{numeric}   Range value: 238, ..., 2982
\item \strong{AB40} \emph{numeric}   Range value: 5550, ..., 37590
\item \strong{TAU} \emph{numeric}   Range value: 110, ..., 946
\item \strong{PTAU} \emph{numeric}   Range value: 8, ..., 118
\item \strong{A4240} \emph{numeric}   Range value: 0.0199, ..., 0.1036
\item \strong{NOTE} \emph{character}   , with options: Abeta>1700, PTau<8
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
