% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENNMSMSABETA2CRM}
\alias{UPENNMSMSABETA2CRM}
\title{UPENNMSMSABETA2CRM}
\format{
A data frame with 1445 observations and 12 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENNMSMSABETA2CRM)
}
\description{
UPENNMSMSABETA2CRM data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 31, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VID} \emph{numeric} Visit code Range value: 2, ..., 14
\item \strong{RUNDATE} \emph{ } Run date Date: YYY-MM-DD
\item \strong{ABETA42} \emph{numeric} Amyloid Beta 1-42 Range value: 187, ..., 3738
\item \strong{ABETA40} \emph{numeric} Amyloid Beta 1-40 Range value: 2709, ..., 19252
\item \strong{ABETA38} \emph{numeric} Amyloid Beta 1-38 Range value: 561, ..., 4398
\item \strong{ABETA42CRM} \emph{numeric} Amyloid Beta 1-42 Range value: 199, ..., 3359
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
