% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENNPLASMA}
\alias{UPENNPLASMA}
\title{UPENN Plasma Biomarker Data}
\format{
A data frame with 2453 observations and 6 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENNPLASMA)
}
\description{
UPENN Plasma Biomarker Data data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Patient number Range value: 2, ..., 1427
\item \strong{VISCODE} \emph{character} Visit code , with options: bl, m12, m24, m36, m48, m06, m18
\item \strong{AB40} \emph{numeric} Abeta1-40 result in plasma pg/mL
\item \strong{AB42} \emph{numeric} Abeta1-42 result in plasma pg/mL
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{biospecimen}
\keyword{undefined_phase}
