% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENNROI_MARS}
\alias{UPENNROI_MARS}
\title{UPENNROI_MARS: 2016-06-01}
\format{
A data frame with 840 observations and 270 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENNROI_MARS)
}
\description{
UPENNROI_MARS: 2016-06-01 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric}   Range value: 2, ..., 1435
\item \strong{VISCODE} \emph{character}   , with options: sc
\item \strong{EXAMDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{VERSION} \emph{ }   Date: YYY-MM-DD
\item \strong{IMAGE_UID} \emph{character}    
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   , with options: COMPLETE
\item \strong{CBICA_ID} \emph{character}    
\item \strong{DATE} \emph{ }   Date: YYY-MM-DD
\item \strong{R702} \emph{numeric}   Range value: 1154010, ..., 2013350
\item \strong{R701} \emph{numeric}   Range value: 843218, ..., 1589360
\item \strong{R601} \emph{numeric}   Range value: 399295, ..., 832033
\item \strong{R604} \emph{numeric}   Range value: 356888, ..., 680505
\item \strong{R606} \emph{numeric}   Range value: 189508, ..., 404636
\item \strong{R607} \emph{numeric}   Range value: 172977, ..., 331536
\item \strong{R613} \emph{numeric}   Range value: 201650, ..., 415826
\item \strong{R614} \emph{numeric}   Range value: 175867, ..., 340199
\item \strong{R501} \emph{numeric}   Range value: 6608.58, ..., 18292
\item \strong{R502} \emph{numeric}   Range value: 95445.6, ..., 192006
\item \strong{R503} \emph{numeric}   Range value: 35555.4, ..., 66649.6
\item \strong{R504} \emph{numeric}   Range value: 217619, ..., 489606
\item \strong{R505} \emph{numeric}   Range value: 21644, ..., 45463.8
\item \strong{R506} \emph{numeric}   Range value: 79998.1, ..., 167032
\item \strong{R507} \emph{numeric}   Range value: 129805, ..., 247640
\item \strong{R508} \emph{numeric}   Range value: 140815, ..., 285599
\item \strong{R509} \emph{numeric}   Range value: 9070.28, ..., 159838
\item \strong{R510} \emph{numeric}   Range value: 44786.3, ..., 90020.1
\item \strong{R511} \emph{numeric}   Range value: 17432.4, ..., 33208.5
\item \strong{R512} \emph{numeric}   Range value: 105205, ..., 242629
\item \strong{R513} \emph{numeric}   Range value: 10900.6, ..., 23973.4
\item \strong{R514} \emph{numeric}   Range value: 38221, ..., 83394
\item \strong{R515} \emph{numeric}   Range value: 65324.8, ..., 126214
\item \strong{R516} \emph{numeric}   Range value: 66588.3, ..., 141867
\item \strong{R517} \emph{numeric}   Range value: 4038.95, ..., 88955.7
\item \strong{R518} \emph{numeric}   Range value: 43357.5, ..., 90236.2
\item \strong{R519} \emph{numeric}   Range value: 18123, ..., 33524.1
\item \strong{R520} \emph{numeric}   Range value: 112414, ..., 255152
\item \strong{R521} \emph{numeric}   Range value: 10184.2, ..., 22378
\item \strong{R522} \emph{numeric}   Range value: 41777, ..., 83762.2
\item \strong{R523} \emph{numeric}   Range value: 61396.4, ..., 123997
\item \strong{R524} \emph{numeric}   Range value: 73341.5, ..., 143732
\item \strong{R525} \emph{numeric}   Range value: 3174.13, ..., 73404.5
\item \strong{R401} \emph{numeric}   Range value: 12474, ..., 25558.4
\item \strong{R402} \emph{numeric}   Range value: 8811.82, ..., 18987.9
\item \strong{R403} \emph{numeric}   Range value: 7834.88, ..., 15389.4
\item \strong{R404} \emph{numeric}   Range value: 92281.9, ..., 246605
\item \strong{R405} \emph{numeric}   Range value: 125337, ..., 262626
\item \strong{R406} \emph{numeric}   Range value: 21644, ..., 45463.8
\item \strong{R407} \emph{numeric}   Range value: 44985.6, ..., 106931
\item \strong{R408} \emph{numeric}   Range value: 32732.4, ..., 66945
\item \strong{R409} \emph{numeric}   Range value: 56886.7, ..., 122774
\item \strong{R410} \emph{numeric}   Range value: 66264, ..., 137419
\item \strong{R411} \emph{numeric}   Range value: 61280.6, ..., 141982
\item \strong{R412} \emph{numeric}   Range value: 78438.9, ..., 160638
\item \strong{R413} \emph{numeric}   Range value: 6101.56, ..., 12628.2
\item \strong{R414} \emph{numeric}   Range value: 4432.47, ..., 9659.43
\item \strong{R415} \emph{numeric}   Range value: 3674.19, ..., 7388.54
\item \strong{R416} \emph{numeric}   Range value: 43985.8, ..., 122905
\item \strong{R417} \emph{numeric}   Range value: 61219.5, ..., 129567
\item \strong{R418} \emph{numeric}   Range value: 10900.6, ..., 23973.4
\item \strong{R419} \emph{numeric}   Range value: 20973.5, ..., 53817.3
\item \strong{R420} \emph{numeric}   Range value: 15503.5, ..., 33650.6
\item \strong{R421} \emph{numeric}   Range value: 26788.4, ..., 64225.2
\item \strong{R422} \emph{numeric}   Range value: 33719.2, ..., 74432.1
\item \strong{R423} \emph{numeric}   Range value: 25916.1, ..., 70491.8
\item \strong{R424} \emph{numeric}   Range value: 38817.6, ..., 81978.4
\item \strong{R425} \emph{numeric}   Range value: 6372.48, ..., 12930.3
\item \strong{R426} \emph{numeric}   Range value: 4379.35, ..., 9328.5
\item \strong{R427} \emph{numeric}   Range value: 4147.93, ..., 8000.89
\item \strong{R428} \emph{numeric}   Range value: 48296.1, ..., 127178
\item \strong{R429} \emph{numeric}   Range value: 64117.9, ..., 138653
\item \strong{R430} \emph{numeric}   Range value: 10184.2, ..., 22378
\item \strong{R431} \emph{numeric}   Range value: 20795.9, ..., 53114.2
\item \strong{R432} \emph{numeric}   Range value: 15734.9, ..., 35659.7
\item \strong{R433} \emph{numeric}   Range value: 26648, ..., 62840.3
\item \strong{R434} \emph{numeric}   Range value: 32544.8, ..., 62986.9
\item \strong{R435} \emph{numeric}   Range value: 27103, ..., 71489.9
\item \strong{R436} \emph{numeric}   Range value: 39621.4, ..., 79428.4
\item \strong{R301} \emph{numeric}   Range value: 10354.5, ..., 31021.8
\item \strong{R302} \emph{numeric}   Range value: 5863.04, ..., 16419.1
\item \strong{R303} \emph{numeric}   Range value: 51798.9, ..., 136479
\item \strong{R304} \emph{numeric}   Range value: 17366.5, ..., 50669.1
\item \strong{R305} \emph{numeric}   Range value: 6495.5, ..., 20000.7
\item \strong{R306} \emph{numeric}   Range value: 14172, ..., 33236.7
\item \strong{R307} \emph{numeric}   Range value: 5834.35, ..., 14599.3
\item \strong{R308} \emph{numeric}   Range value: 3376.49, ..., 11274.2
\item \strong{R309} \emph{numeric}   Range value: 19120, ..., 51363.3
\item \strong{R310} \emph{numeric}   Range value: 18565.3, ..., 45715.5
\item \strong{R311} \emph{numeric}   Range value: 45456.2, ..., 94398.8
\item \strong{R312} \emph{numeric}   Range value: 11150.2, ..., 33166.6
\item \strong{R313} \emph{numeric}   Range value: 9085.07, ..., 20986.1
\item \strong{R314} \emph{numeric}   Range value: 45979.7, ..., 110577
\item \strong{R315} \emph{numeric}   Range value: 4504.74, ..., 14023.2
\item \strong{R316} \emph{numeric}   Range value: 4953.06, ..., 15005.1
\item \strong{R317} \emph{numeric}   Range value: 2681.19, ..., 8198.38
\item \strong{R318} \emph{numeric}   Range value: 24421.1, ..., 68815
\item \strong{R319} \emph{numeric}   Range value: 8118.05, ..., 23434.1
\item \strong{R320} \emph{numeric}   Range value: 3484.9, ..., 10352.7
\item \strong{R321} \emph{numeric}   Range value: 7313.79, ..., 17851.2
\item \strong{R322} \emph{numeric}   Range value: 2600.73, ..., 7857.5
\item \strong{R323} \emph{numeric}   Range value: 2212.72, ..., 6338.87
\item \strong{R324} \emph{numeric}   Range value: 9204.61, ..., 26516.1
\item \strong{R325} \emph{numeric}   Range value: 8824.57, ..., 23461.8
\item \strong{R326} \emph{numeric}   Range value: 20654.1, ..., 50413.9
\item \strong{R327} \emph{numeric}   Range value: 5050.81, ..., 16670.6
\item \strong{R328} \emph{numeric}   Range value: 4235.04, ..., 10508.7
\item \strong{R329} \emph{numeric}   Range value: 19205, ..., 54942.2
\item \strong{R330} \emph{numeric}   Range value: 2228.44, ..., 7750.55
\item \strong{R331} \emph{numeric}   Range value: 5401.41, ..., 16518.6
\item \strong{R332} \emph{numeric}   Range value: 2903.38, ..., 8220.68
\item \strong{R333} \emph{numeric}   Range value: 27377.8, ..., 69981.1
\item \strong{R334} \emph{numeric}   Range value: 9248.48, ..., 28461.1
\item \strong{R335} \emph{numeric}   Range value: 3010.6, ..., 9968.48
\item \strong{R336} \emph{numeric}   Range value: 6052.67, ..., 15952.5
\item \strong{R337} \emph{numeric}   Range value: 2876.44, ..., 7264.37
\item \strong{R338} \emph{numeric}   Range value: 2.13431, ..., 6300.81
\item \strong{R339} \emph{numeric}   Range value: 9135.52, ..., 25883.6
\item \strong{R340} \emph{numeric}   Range value: 9536.59, ..., 23383.5
\item \strong{R341} \emph{numeric}   Range value: 21289.3, ..., 48015.6
\item \strong{R342} \emph{numeric}   Range value: 5358.72, ..., 16496.1
\item \strong{R343} \emph{numeric}   Range value: 4306.12, ..., 10880.2
\item \strong{R344} \emph{numeric}   Range value: 19968.1, ..., 55634.8
\item \strong{R345} \emph{numeric}   Range value: 2193.88, ..., 6272.63
\item \strong{R4} \emph{numeric}   Range value: 399.518, ..., 5171.28
\item \strong{R11} \emph{numeric}   Range value: 714.537, ..., 4851.28
\item \strong{R23} \emph{numeric}   Range value: 194.033, ..., 586.237
\item \strong{R30} \emph{numeric}   Range value: 204.043, ..., 692.609
\item \strong{R31} \emph{numeric}   Range value: 443.972, ..., 1563.61
\item \strong{R32} \emph{numeric}   Range value: 460.116, ..., 1549.01
\item \strong{R35} \emph{numeric}   Range value: 14668.1, ..., 31452
\item \strong{R36} \emph{numeric}   Range value: 2029.86, ..., 5485.72
\item \strong{R37} \emph{numeric}   Range value: 1923, ..., 4836.7
\item \strong{R38} \emph{numeric}   Range value: 33452.5, ..., 67770.2
\item \strong{R39} \emph{numeric}   Range value: 34220.1, ..., 66816
\item \strong{R40} \emph{numeric}   Range value: 9194.32, ..., 22599.1
\item \strong{R41} \emph{numeric}   Range value: 9498.2, ..., 23204.1
\item \strong{R47} \emph{numeric}   Range value: 1892.33, ..., 5501.21
\item \strong{R48} \emph{numeric}   Range value: 1833.76, ..., 5231.3
\item \strong{R49} \emph{numeric}   Range value: 73.8237, ..., 4910.25
\item \strong{R50} \emph{numeric}   Range value: 18.017, ..., 4744.54
\item \strong{R51} \emph{numeric}   Range value: 3076.58, ..., 70400.5
\item \strong{R52} \emph{numeric}   Range value: 3986.42, ..., 85181.5
\item \strong{R55} \emph{numeric}   Range value: 906.757, ..., 1962.36
\item \strong{R56} \emph{numeric}   Range value: 895.457, ..., 1908.72
\item \strong{R57} \emph{numeric}   Range value: 2688.63, ..., 6041.55
\item \strong{R58} \emph{numeric}   Range value: 2651.55, ..., 5851.9
\item \strong{R59} \emph{numeric}   Range value: 4379.35, ..., 9328.5
\item \strong{R60} \emph{numeric}   Range value: 4432.47, ..., 9659.43
\item \strong{R61} \emph{numeric}   Range value: 3215.32, ..., 6134.26
\item \strong{R62} \emph{numeric}   Range value: 3225.94, ..., 6401.05
\item \strong{R71} \emph{numeric}   Range value: 2562.5, ..., 7084.07
\item \strong{R72} \emph{numeric}   Range value: 1253.9, ..., 3140.25
\item \strong{R73} \emph{numeric}   Range value: 1047.69, ..., 4387.05
\item \strong{R75} \emph{numeric}   Range value: 2.12393, ..., 801.988
\item \strong{R76} \emph{numeric}   Range value: 0, ..., 809.807
\item \strong{R81} \emph{numeric}   Range value: 64117.9, ..., 138653
\item \strong{R82} \emph{numeric}   Range value: 61219.5, ..., 129567
\item \strong{R83} \emph{numeric}   Range value: 15734.9, ..., 35659.7
\item \strong{R84} \emph{numeric}   Range value: 15503.5, ..., 33650.6
\item \strong{R85} \emph{numeric}   Range value: 32544.8, ..., 62986.9
\item \strong{R86} \emph{numeric}   Range value: 33719.2, ..., 74432.1
\item \strong{R87} \emph{numeric}   Range value: 39621.4, ..., 79428.4
\item \strong{R88} \emph{numeric}   Range value: 38817.6, ..., 81978.4
\item \strong{R89} \emph{numeric}   Range value: 191.467, ..., 773.727
\item \strong{R90} \emph{numeric}   Range value: 253.946, ..., 778.709
\item \strong{R91} \emph{numeric}   Range value: 1916.75, ..., 4562.43
\item \strong{R92} \emph{numeric}   Range value: 1791.95, ..., 3860.64
\item \strong{R93} \emph{numeric}   Range value: 1546.83, ..., 3583.97
\item \strong{R94} \emph{numeric}   Range value: 1485.28, ..., 3322.5
\item \strong{R95} \emph{numeric}   Range value: 6608.58, ..., 18292
\item \strong{R100} \emph{numeric}   Range value: 1276.93, ..., 5534.38
\item \strong{R101} \emph{numeric}   Range value: 1804.73, ..., 7481.63
\item \strong{R102} \emph{numeric}   Range value: 1671.13, ..., 5692.73
\item \strong{R103} \emph{numeric}   Range value: 1603.84, ..., 5703.43
\item \strong{R104} \emph{numeric}   Range value: 505.71, ..., 2666.46
\item \strong{R105} \emph{numeric}   Range value: 589.562, ..., 2200.86
\item \strong{R106} \emph{numeric}   Range value: 3936.05, ..., 13597.3
\item \strong{R107} \emph{numeric}   Range value: 3528.28, ..., 13559
\item \strong{R108} \emph{numeric}   Range value: 1313.8, ..., 5268.33
\item \strong{R109} \emph{numeric}   Range value: 1309.52, ..., 5531.76
\item \strong{R112} \emph{numeric}   Range value: 1278.32, ..., 5179.46
\item \strong{R113} \emph{numeric}   Range value: 1628.22, ..., 4841.34
\item \strong{R114} \emph{numeric}   Range value: 2400.07, ..., 7960.08
\item \strong{R115} \emph{numeric}   Range value: 2110.19, ..., 7526.87
\item \strong{R116} \emph{numeric}   Range value: 852.971, ..., 3273.67
\item \strong{R117} \emph{numeric}   Range value: 681.533, ..., 3432.84
\item \strong{R118} \emph{numeric}   Range value: 577.449, ..., 2371.58
\item \strong{R119} \emph{numeric}   Range value: 560.442, ..., 2342.24
\item \strong{R120} \emph{numeric}   Range value: 1005.06, ..., 4546.04
\item \strong{R121} \emph{numeric}   Range value: 1077.31, ..., 4257.19
\item \strong{R122} \emph{numeric}   Range value: 4306.12, ..., 10880.2
\item \strong{R123} \emph{numeric}   Range value: 4235.04, ..., 10508.7
\item \strong{R124} \emph{numeric}   Range value: 974.252, ..., 2745.04
\item \strong{R125} \emph{numeric}   Range value: 966.814, ..., 2888.2
\item \strong{R128} \emph{numeric}   Range value: 3603.79, ..., 10574.3
\item \strong{R129} \emph{numeric}   Range value: 2911.19, ..., 9880.04
\item \strong{R132} \emph{numeric}   Range value: 6890.31, ..., 17273
\item \strong{R133} \emph{numeric}   Range value: 6035.33, ..., 16713.7
\item \strong{R134} \emph{numeric}   Range value: 4897.02, ..., 13656.8
\item \strong{R135} \emph{numeric}   Range value: 4605.85, ..., 12052.5
\item \strong{R136} \emph{numeric}   Range value: 920.437, ..., 3514.05
\item \strong{R137} \emph{numeric}   Range value: 640.647, ..., 3951.02
\item \strong{R138} \emph{numeric}   Range value: 1915.45, ..., 6353.33
\item \strong{R139} \emph{numeric}   Range value: 2537.19, ..., 6695.25
\item \strong{R140} \emph{numeric}   Range value: 1.06371, ..., 2592.92
\item \strong{R141} \emph{numeric}   Range value: 546.091, ..., 2810.95
\item \strong{R142} \emph{numeric}   Range value: 8645.66, ..., 24790
\item \strong{R143} \emph{numeric}   Range value: 8222.04, ..., 23604.1
\item \strong{R144} \emph{numeric}   Range value: 2162.3, ..., 7458.99
\item \strong{R145} \emph{numeric}   Range value: 2357.5, ..., 8047.69
\item \strong{R146} \emph{numeric}   Range value: 1780.64, ..., 6019.39
\item \strong{R147} \emph{numeric}   Range value: 1805.07, ..., 6618.79
\item \strong{R148} \emph{numeric}   Range value: 0, ..., 2004.97
\item \strong{R149} \emph{numeric}   Range value: 5.3101, ..., 1635.94
\item \strong{R150} \emph{numeric}   Range value: 1375.68, ..., 5252.24
\item \strong{R151} \emph{numeric}   Range value: 1177.18, ..., 4536.62
\item \strong{R152} \emph{numeric}   Range value: 2912.13, ..., 10840.7
\item \strong{R153} \emph{numeric}   Range value: 2073.87, ..., 8389.47
\item \strong{R154} \emph{numeric}   Range value: 6633.9, ..., 19166.6
\item \strong{R155} \emph{numeric}   Range value: 5103.32, ..., 21066.6
\item \strong{R156} \emph{numeric}   Range value: 1395.12, ..., 5597.45
\item \strong{R157} \emph{numeric}   Range value: 1540.93, ..., 6135.3
\item \strong{R160} \emph{numeric}   Range value: 2.13431, ..., 6300.81
\item \strong{R161} \emph{numeric}   Range value: 2212.72, ..., 6338.87
\item \strong{R162} \emph{numeric}   Range value: 984.21, ..., 4468.88
\item \strong{R163} \emph{numeric}   Range value: 1106.57, ..., 4721.81
\item \strong{R164} \emph{numeric}   Range value: 0, ..., 2282.97
\item \strong{R165} \emph{numeric}   Range value: 0, ..., 2887.65
\item \strong{R166} \emph{numeric}   Range value: 1981.63, ..., 5738.32
\item \strong{R167} \emph{numeric}   Range value: 2048.62, ..., 5890.62
\item \strong{R168} \emph{numeric}   Range value: 4901.13, ..., 15084.9
\item \strong{R169} \emph{numeric}   Range value: 4715.08, ..., 16000.7
\item \strong{R170} \emph{numeric}   Range value: 1705.31, ..., 4419.69
\item \strong{R171} \emph{numeric}   Range value: 1795.41, ..., 4793.7
\item \strong{R172} \emph{numeric}   Range value: 892.262, ..., 3044.76
\item \strong{R173} \emph{numeric}   Range value: 908.024, ..., 2934.76
\item \strong{R174} \emph{numeric}   Range value: 483.866, ..., 3876.34
\item \strong{R175} \emph{numeric}   Range value: 788.147, ..., 3946.11
\item \strong{R176} \emph{numeric}   Range value: 5563.97, ..., 13654.9
\item \strong{R177} \emph{numeric}   Range value: 6174.78, ..., 14358.9
\item \strong{R178} \emph{numeric}   Range value: 1612.77, ..., 4765.33
\item \strong{R179} \emph{numeric}   Range value: 1704.14, ..., 4425.55
\item \strong{R180} \emph{numeric}   Range value: 793.185, ..., 2557.69
\item \strong{R181} \emph{numeric}   Range value: 721.179, ..., 2349.44
\item \strong{R182} \emph{numeric}   Range value: 7454.03, ..., 17223.7
\item \strong{R183} \emph{numeric}   Range value: 5969.81, ..., 16771.9
\item \strong{R184} \emph{numeric}   Range value: 636.688, ..., 2845.41
\item \strong{R185} \emph{numeric}   Range value: 574.28, ..., 2818.86
\item \strong{R186} \emph{numeric}   Range value: 607.326, ..., 2836.46
\item \strong{R187} \emph{numeric}   Range value: 469.737, ..., 2674.97
\item \strong{R190} \emph{numeric}   Range value: 5762.64, ..., 17883.1
\item \strong{R191} \emph{numeric}   Range value: 5515.09, ..., 18163.2
\item \strong{R192} \emph{numeric}   Range value: 2136.97, ..., 7330.43
\item \strong{R193} \emph{numeric}   Range value: 2352.7, ..., 7786.26
\item \strong{R194} \emph{numeric}   Range value: 3811.12, ..., 11709.6
\item \strong{R195} \emph{numeric}   Range value: 3400.49, ..., 11488.8
\item \strong{R196} \emph{numeric}   Range value: 1151.71, ..., 5407.73
\item \strong{R197} \emph{numeric}   Range value: 616.399, ..., 5029.75
\item \strong{R198} \emph{numeric}   Range value: 4740.69, ..., 13659.2
\item \strong{R199} \emph{numeric}   Range value: 4794.5, ..., 13741.6
\item \strong{R200} \emph{numeric}   Range value: 2688.7, ..., 9776.41
\item \strong{R201} \emph{numeric}   Range value: 2483.58, ..., 9655.74
\item \strong{R202} \emph{numeric}   Range value: 2934.67, ..., 11795.2
\item \strong{R203} \emph{numeric}   Range value: 2932.11, ..., 11093.5
\item \strong{R204} \emph{numeric}   Range value: 1284.44, ..., 4640.22
\item \strong{R205} \emph{numeric}   Range value: 1239.75, ..., 4875.16
\item \strong{R206} \emph{numeric}   Range value: 278.924, ..., 2262.61
\item \strong{R207} \emph{numeric}   Range value: 405.021, ..., 2806.12
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
