% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENNSPARE_AD}
\alias{UPENNSPARE_AD}
\title{Spatial Pattern of Abnormalities for Recognition of Early AD (SPARE-AD). C. Davatzikos. U. Penn.: 2016-07-27}
\format{
A data frame with 840 observations and 13 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENNSPARE_AD)
}
\description{
Spatial Pattern of Abnormalities for Recognition of Early AD (SPARE-AD). C. Davatzikos. U. Penn.: 2016-07-27 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 1435
\item \strong{VISCODE} \emph{character} Visit code , with options: sc
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{ } Version Date Date: YYY-MM-DD
\item \strong{IMAGE_UID} \emph{character}    
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   , with options: COMPLETE
\item \strong{CBICA_ID} \emph{character} Subject ID used at CBICA  
\item \strong{DATE} \emph{ }   Date: YYY-MM-DD
\item \strong{TrueLabel} \emph{character} Diagnostic Group , with options: CN, AD, MCI
\item \strong{SPARE_AD} \emph{numeric} Spatial Pattern of Abnormalities for Recognition of Early AD (SPARE-AD) Range value: -4.57946, ..., 6.24647
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
