% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPENNSPARE_MCI}
\alias{UPENNSPARE_MCI}
\title{Spatial Pattern of Abnormalities for Recognition of Early MCI to AD conversion (SPARE-MCI). C. Davatzikos. U. Penn.: 2016-08-12}
\format{
A data frame with 330 observations and 14 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UPENNSPARE_MCI)
}
\description{
Spatial Pattern of Abnormalities for Recognition of Early MCI to AD conversion (SPARE-MCI). C. Davatzikos. U. Penn.: 2016-08-12 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 4, ..., 1427
\item \strong{VISCODE} \emph{character} Visit code , with options: sc
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{VERSION} \emph{ } Version Date Date: YYY-MM-DD
\item \strong{IMAGE_UID} \emph{character}    
\item \strong{RUNDATE} \emph{ }   Date: YYY-MM-DD
\item \strong{STATUS} \emph{character}   , with options: COMPLETE
\item \strong{CBICA_ID} \emph{character} Subject ID used at CBICA  
\item \strong{DATE} \emph{ }   Date: YYY-MM-DD
\item \strong{TimeToConversion} \emph{numeric} Time to Conversion from MCI to AD Range value: 6, ..., 48
\item \strong{TrueLabel} \emph{character} Diagnostic Group (LS - Long term Stable, SC - Short Converter) , with options: LS, SC
\item \strong{SPARE_MCI} \emph{numeric} Spatial Pattern of Abnormalities for Recognition of Early MCI to AD conversion (SPARE-MCI) Range value: -3.85355, ..., 3.51882
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
