% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{URMC_LABDATA}
\alias{URMC_LABDATA}
\title{ADNI3 URMC Lab Data}
\format{
A data frame with 140434 observations and 29 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(URMC_LABDATA)
}
\description{
ADNI3 URMC Lab Data data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{character}   , with options: ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{SampleDate} \emph{ }   Date: YYY-MM-DD
\item \strong{SampleTime} \emph{ }   hms: HH-MM-SS
\item \strong{RecDate} \emph{ }   Date: YYY-MM-DD
\item \strong{RecTime} \emph{ }   hms: HH-MM-SS
\item \strong{KitID} \emph{numeric}   Range value: 6174088, ..., 6528080
\item \strong{TestID} \emph{character}    
\item \strong{TestName} \emph{character}    
\item \strong{ResultValue} \emph{character}    
\item \strong{Flags} \emph{character}   , with options: H, L, VH, VL
\item \strong{Comments} \emph{character}    
\item \strong{LowerRange} \emph{character}    
\item \strong{UpperRange} \emph{character}    
\item \strong{Units} \emph{character}    
\item \strong{ResultValueConv} \emph{character}    
\item \strong{LowerRangeConv} \emph{character}    
\item \strong{UpperRangeConv} \emph{character}    
\item \strong{UnitsConv} \emph{character}    
\item \strong{ResultValueSI} \emph{character}    
\item \strong{LowerRangeSI} \emph{character}    
\item \strong{UpperRangeSI} \emph{character}    
\item \strong{UnitsSI} \emph{character}    
\item \strong{Fasting} \emph{character}   , with options: Y, N, U
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni3}
\keyword{adni4}
\keyword{biospecimen}
