% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UUCACIR}
\alias{UUCACIR}
\title{UU PET Analysis: 2016-07-18}
\format{
A data frame with 1822 observations and 17 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UUCACIR)
}
\description{
UU PET Analysis: 2016-07-18 data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 5296
\item \strong{VISCODE} \emph{character} Visit code , with options: v11, v41, v21, v06, v31, v51, v03, v04, v02
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Scan Date Date: YYY-MM-DD
\item \strong{IMAGEUID} \emph{numeric} ImageUID Range value: 226554, ..., 740360
\item \strong{PETTYPE} \emph{character} Type of PET scan , with options: AV45
\item \strong{REFCTRLPOP} \emph{character} Control population used for comparison , with options: ADNI2_158NL_AV45_cbl
\item \strong{REFREGION} \emph{character} Region used for normalization , with options: CBL
\item \strong{AVEREF} \emph{numeric} Average value of reference region Range value: 0.938881, ..., 2.14705
\item \strong{AVEASSOC} \emph{numeric} Average frontal, parietal and temporal cortices, norm to reference region Range value: 0.872387, ..., 2.34006
\item \strong{AVEFRONT} \emph{numeric} Average frontal cortex, norm to reference region Range value: 0.859352, ..., 2.56467
\item \strong{X2SDSIGPXL} \emph{numeric} Number of pixels with Zscores in the range [2, 3) SD Range value: 22, ..., 3194
\item \strong{X3SDSIGPXL} \emph{numeric} Number of pixels with Zscores  in the range  [3, inf) SD Range value: 0, ..., 13585
\item \strong{SUMZ2} \emph{numeric} Sum of pixels with Zscores in the range [2, inf) SD Range value: 48.361, ..., 186158.3291
\item \strong{SUMZ3} \emph{numeric} Sum of pixels with Zscores in the range [3, inf) SD Range value: 0, ..., 185016.4857
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{imaging}
\keyword{undefined_phase}
