% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UWNPSYCHSUM}
\alias{UWNPSYCHSUM}
\title{UWNPSYCHSUM}
\format{
A data frame with 11519 observations and 12 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(UWNPSYCHSUM)
}
\description{
UWNPSYCHSUM data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character} Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{COLPROT} \emph{character} Study protocol of data collection , with options: ADNI1, ADNIGO, ADNI2, ADNI3
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 7123
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{ADNI_MEM} \emph{numeric} Composite score for memory Range value: -3.871, ..., 4.703
\item \strong{ADNI_EF} \emph{numeric} Composite score for executive functioning Range value: -3.506, ..., 3.389
\item \strong{ADNI_LAN} \emph{numeric} Composite score for language cognitive items Range value: -4.874, ..., 3.151
\item \strong{ADNI_VS} \emph{numeric} Composite score for visuo-spatial skills Range value: -3.498, ..., 0.739
\item \strong{ADNI_EF2} \emph{numeric} Composite score for executive functioning, version 2 Range value: -3.988, ..., 2.78
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adnigo}
\keyword{assessments}
