% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VISITS}
\alias{VISITS}
\title{VISITS}
\format{
A data frame with 69 observations and 5 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(VISITS)
}
\description{
VISITS . More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{PHASE} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{ID} \emph{numeric}   Range value: 1, ..., 19
\item \strong{VISCODE} \emph{character}    
\item \strong{VISNAME} \emph{character}    
\item \strong{VISORDER} \emph{numeric}   Range value: 1, ..., 17
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{enrollment}
