% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VITALS}
\alias{VITALS}
\title{Vital Signs}
\format{
A data frame with 17259 observations and 27 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(VITALS)
}
\description{
Vital Signs data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 2, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code  
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{VISDATE} \emph{ } Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD
\item \strong{VSWEIGHT} \emph{numeric} 1a. Weight Range value: -1, ..., 493
\item \strong{VSWTUNIT} \emph{character} 1b. Weight Units 1=pounds ; 2=kilograms
\emph{Decoded Value: }\strong{Yes}
\item \strong{VSHEIGHT} \emph{numeric} 2a. Height Range value: -1, ..., 197
\item \strong{VSHTUNIT} \emph{character} 2b. Height Units 1=inches ; 2=centimeters
\emph{Decoded Value: }\strong{Yes}
\item \strong{VSBPSYS} \emph{numeric} Seated Blood Pressure: Systolic 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 76..250
\item \emph{ADNI3, ADNI4} 
}
\item \strong{VSBPDIA} \emph{numeric} Seated Blood Pressure: Diastolic 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 36..130
\item \emph{ADNI3, ADNI4} 
}
\item \strong{VSPULSE} \emph{numeric} 4. Seated Pulse Rate (per minute) 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 40..130
\item \emph{ADNI3, ADNI4} 
}
\item \strong{VSRESP} \emph{numeric} 5. Respirations (per minute) 
\itemize{
\item \emph{ADNI1, ADNIGO, ADNI2:} 6..40
\item \emph{ADNI3, ADNI4} 
}
\item \strong{VSTEMP} \emph{numeric} 6a. Temperature Range value: -1, ..., 101.9
\item \strong{VSTMPSRC} \emph{character} 6b. Temperature Source 1=Oral ; 2=Tympanic ; 3=Other
\emph{Decoded Value: }\strong{Yes}
\item \strong{VSTMPUNT} \emph{character} 6c. Temperature Units 1=Fahrenheit ; 2=Celsius
\emph{Decoded Value: }\strong{Yes}
\item \strong{VSHGTSC} \emph{character} Was height collected at screening or initial visit? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{ID} \emph{numeric} Record ID Range value: 7, ..., 187104
\item \strong{SITEID} \emph{numeric} Site ID Range value: 1, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: sc, annual
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni1}
\keyword{adni2}
\keyword{adni3}
\keyword{adni4}
\keyword{adnigo}
\keyword{medical_history}
