% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VS}
\alias{VS}
\title{[DERIVED]  Derived  Vital Signs}
\format{
A data frame with 120813 observations and 20 variables.
}
\source{
For more details see the help vignette: 
\code{vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')}
}
\usage{
data(VS)
}
\description{
Vital signs derived dataset.
}
\details{
\itemize{
\item \strong{STUDYID} \emph{character} Study Identifier ; , with character options: ADNI
\item \strong{DOMAIN} \emph{character} Domain Abbrevation , with character options: VS
\item \strong{USUBJID} \emph{character} Unique Subject Identifier  
\item \strong{VSSEQ} \emph{integer} Sequence Number Range value: 1, ..., 147
\item \strong{VSGRPID} \emph{factor} Group ID Study Protocol of Data Collection; , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{VSTESTCD} \emph{character} Vital Signs Test Short Name , with character options: DIABP, HEIGHT, PLUSE, RESP, SYSBP, TEMP, WEIGHT
\item \strong{VSTEST} \emph{character} Vital Signs Test Name , with character options: Diastolic Blood Pressure, Height, Pluse Rate, Respiratory Rate, Temperature, Weight
\item \strong{VSORRES} \emph{character} Result or Finding in Original Units  
\item \strong{VSORRESU} \emph{character} Original Units , with character options: mmHg, inch, NA, beats/min, breaths/min, F, LB, cm, kg, C
\item \strong{VSSTRESC} \emph{character} Character Result/Finding in Std Format  
\item \strong{VSSTRESN} \emph{numeric} Numeric Result/Finding in Standard Units Range value: 1.7, ..., 493
\item \strong{VSSTRESU} \emph{character} Standard Units , with character options: mmHg, cm, beats/min, breaths/min, C, kg
\item \strong{VSLOC} \emph{character} Location of Vital Signs Measurment , with character options: TYMPANIC, ORAL, OTHER, NA
\item \strong{VSBLFL} \emph{character} Baseline Flag , with character options: Y
\item \strong{VSDRVFL} \emph{character} Derived Flag , with character options: Yes
\item \strong{VISITNUM} \emph{numeric} Visit Number Range value: -2, ..., 49
\item \strong{VISIT} \emph{character} Visit Name  
\item \strong{EPOCH} \emph{character} Epoch , with character options: Screening, Baseline, Follow-up
\item \strong{VSDTC} \emph{Date} Date/Time of Measurements Date: YYY-MM-DD
\item \strong{VSDY} \emph{integer} Study Day of Vital Signs Range value: -363, ..., 7175
}
}
\keyword{derived_dataset}
