% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WATC}
\alias{WATC}
\title{Word Accentuation Test - Chicago}
\format{
A data frame with 1001 observations and 59 variables.
}
\source{
\href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\usage{
data(WATC)
}
\description{
Word Accentuation Test - Chicago data. More information is available at \href{https://adni.loni.usc.edu/data-samples/adni-data/}{https://adni.loni.usc.edu/data-samples/adni-data/}
}
\details{
\itemize{
\item \strong{ORIGPROT} \emph{character}   , with options: ADNI2, ADNI3, ADNI4, ADNI1, ADNIGO
\item \strong{COLPROT} \emph{factor}   , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4
\item \strong{PTID} \emph{character} Participant ID  
\item \strong{RID} \emph{numeric} Participant roster ID Range value: 21, ..., 10898
\item \strong{VISCODE} \emph{character} Visit code , with options: 4_init, 4_bl, 4_m12
\item \strong{VISCODE2} \emph{character} Translated visit code  
\item \strong{EXAMDATE} \emph{ } Examination Date Date: YYY-MM-DD
\item \strong{DONE} \emph{character} Was assessment/procedure done? 0=No ; 1=Yes
\emph{Decoded Value: }\strong{Yes}
\item \strong{NDREASON} \emph{character} If No, reason not done 1=Participant Unable for Cognitive Reasons ; 10=Not called for by protocol ; 11=Not a Spanish-speaking participant ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 7=Site Error ; 8=Other Reason ; 9=Unknown
\emph{Decoded Value: }\strong{Yes}
\item \strong{SOURCE} \emph{character} Information Source 1=Participant Visit ; 2=Telephone Call/Video Call
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC1} \emph{character} 1. polígamo 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC2} \emph{character} 2. búlgaro 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC3} \emph{character} 3. cúpula 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC4} \emph{character} 4. sándalo 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC5} \emph{character} 5. célibe 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC6} \emph{character} 6. metrópoli 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC7} \emph{character} 7. albedrío 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC8} \emph{character} 8. parónimo 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC9} \emph{character} 9. procónsul 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC10} \emph{character} 10. abogacía 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC11} \emph{character} 11. acólito 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC12} \emph{character} 12. rabí 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC13} \emph{character} 13. volátil 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC14} \emph{character} 14. diámetro 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC15} \emph{character} 15. cánon 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC16} \emph{character} 16. inverosímil 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC17} \emph{character} 17. ámbar 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC18} \emph{character} 18. alegoría 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC19} \emph{character} 19. isótopo 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC20} \emph{character} 20. pífano 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC21} \emph{character} 21. sístole 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC22} \emph{character} 22. ínfulas 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC23} \emph{character} 23. gélido 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC24} \emph{character} 24. hipérbole 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC25} \emph{character} 25. peroné 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC26} \emph{character} 26. retruécano 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC27} \emph{character} 27. tórrido 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC28} \emph{character} 28. táctil 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC29} \emph{character} 29. anómalo 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC30} \emph{character} 30. díscolo 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC31} \emph{character} 31. pécora 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC32} \emph{character} 32. púgil 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC33} \emph{character} 33. cóncavo 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC34} \emph{character} 34. neófito 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC35} \emph{character} 35. lóbrego 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC36} \emph{character} 36. sílice 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC37} \emph{character} 37. apátrida 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC38} \emph{character} 38. láudano 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC39} \emph{character} 39. acullá 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{WATC40} \emph{character} 40. supérstite 0=Incorrect ; 1=Correct
\emph{Decoded Value: }\strong{Yes}
\item \strong{RAWSCORE} \emph{numeric} Raw Score: Sum of scores Range value: 14, ..., 40
\item \strong{ID} \emph{numeric} Record ID Range value: 16546, ..., 126389
\item \strong{SITEID} \emph{numeric} Site ID Range value: 2, ..., 941
\item \strong{USERDATE} \emph{ } User Create Date Date: YYY-MM-DD
\item \strong{USERDATE2} \emph{ } User Last Modify Date Date: YYY-MM-DD
\item \strong{DD_CRF_VERSION_LABEL} \emph{character} CRF Version Label , with options: v1
\item \strong{LANGUAGE_CODE} \emph{character} Language Code e=English
\emph{Decoded Value: }\strong{Yes}
\item \strong{HAS_QC_ERROR} \emph{character} Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error
\emph{Decoded Value: }\strong{Yes}
\item \strong{update_stamp} \emph{ }   POSIXct: YYY-MM-DD
}
}
\keyword{adni4}
\keyword{assessments}
